Constants

Public

Private

PortfolioOptimiser.AllocMethods — Constant

AllocMethods = (:LP, :Greedy)

Methods for allocating assets to an AbstractPortfolio according to the optimised weights and latest asset prices.

:LP: uses Mixed-Integer Linear Programming optimisation to allocate the portfolio.

:Greedy: uses a greedy iterative algorithm.

PortfolioOptimiser.BLFMMethods — Constant

BLFMMethods = (:A, :B)

Versions of the factor Black-Litterman Model.

:A – Augmented Black-Litterman

Estimates the covariance matrix and expected returns vector based on a modified Black-Litterman model which augments its asset views with those of the factors [1].

\[\begin{align*} \bm{\Pi}_{a} &= \begin{cases} \delta\begin{bmatrix} \mathbf{\Sigma}\\ \mathbf{\Sigma}_{F} \mathbf{B}^{\intercal} \end{bmatrix} \bm{w} &\quad \mathrm{if~ flag = true}\\ \bm{\mu} - r &\quad \mathrm{if~ flag = false} \end{cases}\\ \mathbf{P}_{a} &= \begin{bmatrix} \mathbf{P} & \mathbf{0}\\ \mathbf{0} & \mathbf{P}_{F} \end{bmatrix}\\ \bm{Q}_{a} &= \begin{bmatrix} \bm{Q}\\ \bm{Q}_{F} \end{bmatrix}\\ \mathbf{\Sigma}_{a} &= \begin{bmatrix} \mathbf{\Sigma} & \mathbf{B} \mathbf{\Sigma}_{F}\\ \mathbf{\Sigma}_{F} \mathbf{B}^{\intercal} & \mathbf{\Sigma}_{F} \end{bmatrix}\\ \mathbf{\Omega}_{a} &= \begin{bmatrix} \mathbf{\Omega} & \mathbf{0}\\ \mathbf{0} & \mathbf{\Omega}_{F} \end{bmatrix}\\ \tau &= \dfrac{1}{T}\\ \mathbf{\Omega} &= \tau \mathrm{Diagonal}\left(\mathbf{P} \mathbf{\Sigma} \mathbf{P}^{\intercal}\right)\\ \mathbf{\Omega}_{F} &= \tau \mathrm{Diagonal}\left(\mathbf{P}_{F} \mathbf{\Sigma}_{F} \mathbf{P}_{F}^{\intercal}\right)\\ \mathbf{M}_{a} &= \left[ \left(\tau \mathbf{\Sigma}_{a} \right)^{-1} + \mathbf{P}_{a}^{\intercal} \mathbf{\Omega}_{a}^{-1} \mathbf{P}_{a}\right]^{-1}\\ \bm{\Pi}_{\mathrm{ABL}} &= \mathbf{M}_{a} \left[\left(\tau \mathbf{\Sigma}_{a}\right)^{-1} \bm{\Pi}_{a} + \mathbf{P}_{a}^{\intercal} \mathbf{\Omega}_{a}^{-1} \mathbf{Q}_{a} \right]\\ \bm{\mu}_{\mathrm{ABL}} &= \bm{\Pi}_{\mathrm{ABL}} + r\\ \mathbf{\Sigma}_{\mathrm{ABL}} &= \mathbf{\Sigma}_{a} + \mathbf{M}_{a} \end{align*}\]

Where:

$\bm{\Pi}_{a}$ is the augmented equilibrium excess returns.

$\delta$ is the risk aversion parameter.

$\mathbf{\Sigma}$ is the asset covariance matrix.

$\mathbf{\Sigma}_{F}$ is the factor covariance matrix.

$\bm{w}$ is the vector of benchmark asset weights.

$\mathbf{P}_{a}$ is the augmented views matrix. The zeros pad the matrix so all columns and rows are of equal length.

$\mathbf{P}$ is the asset views matrix.

$\mathbf{P}_{F}$ is the factor views matrix.

$\bm{Q}_{a}$ is the augmented views returns vector.

$\bm{Q}$ is the asset views returns vector.

$\bm{Q}_{F}$ is the factor views returns vector.

$\mathbf{\Sigma}_{a}$ is the augmented covariance matrix.

$\mathbf{B}$ is the loadings matrix.

$\mathbf{\Omega}_{a}$ is the covariance matrix of the augmented views.

$\mathbf{\Omega}$ is the covariance matrix of the asset views.

$\mathbf{\Omega}_{F}$ is the covariance matrix of the factor views.

$\mathbf{M}_{a}$ is an intermediate covariance matrix.

$\bm{\Pi}_{\mathbf{ABL}}$ is the equilibrium excess returns after being adjusted by the augmented views.

$\bm{\mu}_{\mathbf{ABL}}$ is the vector of asset expected returns obtained via the Augmented Black Litterman model.

$\mathbf{\Sigma}_{\mathrm{ABL}}$ is the asset covariance matrix obtained via the Augmented Black Litterman model.

:B– Bayesian Black-Litterman

Estimates the covariance matrix and expected returns vector based on a modified Black-Litterman model which updates its views with Bayesian statistics [2].

\[\begin{align*} \mathbf{\Sigma} &= \mathbf{B} \mathbf{\Sigma}_{F} \mathbf{B}^{\intercal} + \mathbf{D} \\ \mathbf{D} &= \begin{cases}\mathrm{Diagonal}\left(\mathrm{var}\left(\mathbf{X} - \mathbf{F} \mathbf{B}^{\intercal},\, \mathrm{dims} = 1\right)\right) &\quad \mathrm{if~ flag = true}\\ \mathbf{0} &\quad \mathrm{if~ flag = false} \end{cases}\\ \overline{\mathbf{\Sigma}}_{F} &= \left(\mathbf{\Sigma}_{F}^{-1} + \mathbf{P}_{F}^{\intercal} \mathbf{\Omega}_{F}^{-1} \mathbf{P}_{F}\right)^{-1}\\ \tau &= \dfrac{1}{T}\\ \mathbf{\Omega}_{F} &= \tau \mathrm{Diagonal}\left(\mathbf{P}_{F} \mathbf{\Sigma}_{F} \mathbf{P}_{F}^{\intercal}\right)\\ \overline{\bm{\Pi}}_{F} &= \overline{\mathbf{\Sigma}}_{F}^{-1} \left(\mathbf{\Sigma}_{F}^{-1} \bm{\Pi}_{F} + \mathbf{P}_{F}^{\intercal} \mathbf{\Omega}_{F}^{-1} \bm{Q}_{F}\right)\\ \bm{\Pi}_{F} &= \bm{\mu}_{F} - r\\ \mathbf{\Sigma}_{\mathrm{BF}} &= \mathbf{\Sigma}^{-1} \mathbf{B} \left( \overline{\mathbf{\Sigma}}_{F}^{-1} + \mathbf{B}^{\intercal} \mathbf{\Sigma}^{-1} \mathbf{B} \right)^{-1}\\ \mathbf{\Sigma}_{\mathrm{BLB}} &= \left(\mathbf{\Sigma}^{-1} - \mathbf{\Sigma}_{\mathrm{BF}} \mathbf{B}^{\intercal} \mathbf{\Sigma}^{-1}\right)^{-1}\\ \bm{\mu}_{\mathrm{BLB}} &= \mathbf{\Sigma}_{\mathrm{BLB}} \mathbf{\Sigma}_{\mathrm{BF}} \overline{\mathbf{\Sigma}}_{F}^{-1} \overline{\bm{\Pi}}_{F} + r \end{align*}\]

Where:

$\mathbf{\Sigma}$ is the estimated asset covariance matrix computed using the factor model.

$\mathbf{B}$ is the loadings matrix.

$\mathbf{\Sigma}_{F}$ is the covariance matrix of the factors.

$\mathbf{D}$ is a diagonal matrix constructed from the variances of the errors between the asset and factor returns corrected by the loadings matrix i.e. the variance is taken for all $T$ timestamps of $N_{a}$ assets. If a flag is false this matrix can be set to $\mathbf{0}$.

$\mathbf{X}$ is the T×Na matrix of asset returns, where T is the number of returns observations and Na the number of assets.

$\mathbf{F}$ is the T×Nf matrix of factor returns, where T is the number of returns observations and Nf the number of factors.

$\overline{\mathbf{\Sigma}}_{F}$ is the posterior covariance matrix of the factors after adjusting by the factor views.

$\mathbf{P}_{F}$ is the factor views matrix.

$\mathbf{\Omega}_{F}$ is the covariance matrix of the factor views.

$\overline{\bm{\Pi}}_{F}$ is the posterior equilibrium excess returns of the factors after adjusting by the factor views.

$\bm{\Pi}_{F}$ is the equilibrium excess returns vector of the factors.

$\bm{\mu}_{F}$ is the expected returns vector of the factors.

$r$ is the risk-free rate.

$\bm{Q}_{F}$ is the factor views returns vector.

$\mathbf{\Sigma}_{\mathrm{BF}}$ is an intermediate covariance matrix.

$\mathbf{\Sigma}_{\mathrm{BLB}}$ is the posterior asset covariance matrix, aka the asset covariance matrix obtained via the Bayesian Black-Litterman model.

$\bm{\mu}_{\mathrm{BLB}}$ is the posterior asset expected returns vector, aka the asset returns vector obtained via the Bayesian Black-Litterman model.

PortfolioOptimiser.BLHist — Constant

BLHist = (1, 2, 3)

Choice of estimate of the covariance matrix $\mathbf{\Sigma}$, means vector $\bm{\mu}$, and returns matrix $\mathbf{X}$ for optimising with different PortClasses. Each different estimate is appended by the suffixes:

No suffix: standard, computed by asset_statistics!.

$\mathrm{fm}$: factor model, computed by factor_statistics!.

$\mathrm{bl}$: Black-Litterman model, computed by black_litterman_statistics!.

$\mathrm{bl, fm}$: Black-Litterman factor model, computed by black_litterman_factor_satistics!.

The choices are:

:Classic: BLHist has no effect, always use $\bm{\mu}$, $\mathbf{\Sigma}$, $\mathbf{X}$.

:FM: $\bm{\mu}_{\mathrm{fm}}$;

1: $\mathbf{\Sigma}_\mathrm{fm}$, $\mathbf{X}_\mathrm{fm}$;

2: $\mathbf{\Sigma}$, $\mathbf{X}$;

3: throws error.

:BL: $\bm{\mu}_\mathrm{bl}$, $\mathbf{X}$;

1: $\mathbf{\Sigma}_\mathrm{bl}$;

2: $\mathbf{\Sigma}$;

3: throws error.

:BLFM: $\bm{\mu}_\mathrm{bl, fm}$;

1: $\mathbf{\Sigma}_\mathrm{bl, fm}$, $\mathbf{X}_\mathrm{fm}$;

2: $\mathbf{\Sigma}$, $\mathbf{X}$;

3: $\mathbf{\Sigma}_\mathrm{fm}$, $\mathbf{X}_\mathrm{fm}$.

PortfolioOptimiser.BinMethods — Constant

BinMethods = (:KN, :FD, :SC, :HGR)

Methods for calculating optimal bin widths for the mutual and variational information matrices computed by mut_var_info_mtx.

:KN: Knuth's.

:FD: Freedman-Diaconis'.

:SC: Scotts'.

:HGR: Hacine-Gharbi and Ravier's.

PortfolioOptimiser.BootstrapMethods — Constant

BootstrapMethods = (:Stationary, :Circular, :Moving)

Bootstrapping method for computing the uncertainty sets with wc_statistics!, which are used by Worst Case Mean Variance Optimisations (see PortTypes).

:Stationary: stationary.

:Circular: circular block.

:Moving: moving block.

PortfolioOptimiser.BoxMethods — Constant

BoxMethods = (:Stationary, :Circular, :Moving, :Normal, :Delta)

Available types of box sets that can be computed with wc_statistics!, which are used by Worst Case Mean Variance Optimisations (see PortTypes).

:Stationary: stationary.

:Circular: circular block.

:Moving: moving block.

:Normal: sets generated by assuming the returns are normally distributed.

:Delta: the set limits are assumed to fall in the extrema of a well-defined interval parametrised by a percentage of the covariance matrix and expected returns vector.

PortfolioOptimiser.CorMethods — Constant

CorMethods = (:Pearson, :Spearman, :Kendall, :Gerber0, :Gerber1, :Gerber2, :SB0, :SB1,
              :Gerber_SB0, :Gerber_SB1, :Abs_Pearson, :Abs_Spearman, :Abs_Kendall,
              :Distance, :Mutual_Info, :Tail, :Cov_to_Cor, :Custom_Func, :Custom_Val)

Methods for estimating the correlation and distance matrices, $\mathbf{C}$ and $\mathbf{D}$ respectively, in cor_dist_mtx.

:Pearson: Pearson correlation, $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{C}_{i,\,j} \right)}$.

:Spearman: Spearman correlation, $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{C}_{i,\,j} \right)}$.

:Kendall: Kendall correlation, $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{C}_{i,\,j} \right)}$.

:Gerber0: Gerber statistic 0 [3], $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{C}_{i,\,j} \right)}$.

:Gerber1: Gerber statistic 1 [3], $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{C}_{i,\,j} \right)}$.

:Gerber2: Gerber statistic 2 [3], $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{C}_{i,\,j} \right)}$.

:SB0: Smyth-Broby modification of the Gerber statistic 0 [4], $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{C}_{i,\,j} \right)}$.

:SB1: Smyth-Broby modification of the Gerber statistic 1 [4], $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{C}_{i,\,j} \right)}$.

:Gerber_SB0: Smyth-Broby modification of the Gerber statistic 0 that scales the contributions of significant co-movements by the vanilla Gerber count, as described in the conclusion of [4], $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{C}_{i,\,j} \right)}$.

:Gerber_SB1: Smyth-Broby modification of the Gerber statistic 1 that scales the contributions of significant co-movements by the vanilla Gerber count, as described in the conclusion of [4], $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{C}_{i,\,j} \right)}$.

:Abs_Pearson: absolute value of the Pearson correlation, $\mathbf{D}_{i,\,j} = \sqrt{1 - \left\lvert\mathbf{C}_{i,\,j}\right\rvert}$.

:Abs_Spearman: absolute value of the Spearman correlation, $\mathbf{D}_{i,\,j} = \sqrt{1 - \left\lvert\mathbf{C}_{i,\,j}\right\rvert}$.

:Abs_Kendall: absolute value of the Kendall correlation, $\mathbf{D}_{i,\,j} = \sqrt{1 - \left\lvert\mathbf{C}_{i,\,j}\right\rvert}$.

:Distance: distance correlation matrix, $\mathbf{D}_{i,\,j} = \sqrt{1 - \mathbf{C}_{i,\,j}}$

:Mutual_Info: mutual information matrix, $\mathbf{D}_{i,\,j}$ is the variation information matrix.

:Tail: lower tail dependence index matrix, $\mathbf{D}_{i,\,j} = -\log\left(\mathbf{C}_{i,\,j}\right)$

:Cov_to_Cor: the covariance matrix is converted to a correlation matrix, the distance matrix is computed by a distance function which defaults to $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{C}_{i,\,j} \right)}$.

:Custom_Func: custom function provided, the distance matrix is computed by a distance function which defaults to $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{C}_{i,\,j} \right)}$.

:Custom_Val: custom value provided, the distance matrix is computed by a distance function which defaults to $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{C}_{i,\,j} \right)}$.

PortfolioOptimiser.CovMethods — Constant

CovMethods = (:Full, :Semi, :Gerber0, :Gerber1, :Gerber2, :SB0, :SB1, :Gerber_SB0,
              :Gerber_SB1, :Custom_Func, :Custom_Val)

Methods for estimating the covariance matrix in covar_mtx.

:Full: full covariance matrix. Uses the StatsBase.CovarianceEstimator interface and StatsBase.cov method. Letting the user take advantage of weights and covariance estimation packages.

:Semi: lower semi-covariance matrix. Uses the StatsBase.CovarianceEstimator interface and StatsBase.cov method. Letting the user take advantage of weights and covariance estimation packages. Also lets the user provide a threshold value below which returns are considered bad enough to be included in the lower semi-covariance.

:Gerber0: Gerber statistic 0 [3].

:Gerber1: Gerber statistic 1 [3].

:Gerber2: Gerber statistic 2 [3].

:SB0: Smyth-Broby modification of the Gerber statistic 0 [4].

:SB1: Smyth-Broby modification of the Gerber statistic 1 [4].

:Gerber_SB0: Smyth-Broby modification of the Gerber statistic 0 that scales the contributions of significant co-movements by the vanilla Gerber count, as described in the conclusion of [4].

:Gerber_SB1: Smyth-Broby modification of the Gerber statistic 1 that scales the contributions of significant co-movements by the vanilla Gerber count, as described in the conclusion of [4].

:Custom_Func: custom function provided.

:Custom_Val: custom value provided.

PortfolioOptimiser.DenoiseMethods — Constant

DenoiseMethods = (:Fixed, :Spectral, :Shrink)

Methods for matrix denoising. See [5, Ch. 2] for details.

:Fixed: fixed.

:Spectral: spectral.

:Shrink: shrink.

PortfolioOptimiser.EllipseMethods — Constant

EllipseMethods = (:Stationary, :Circular, :Moving, :Normal)

Available types of elliptical sets that can be computed with wc_statistics!, which are used by Worst Case Mean Variance Optimisations (see PortTypes).

:Stationary: stationary.

:Circular: circular block.

:Moving: moving block.

:Normal: sets generated by assuming the returns are normally distributed.

PortfolioOptimiser.FSMethods — Constant

FSMethods = (:FReg, :BReg, :PCR)

Methods for feature selection when creating the loadings matrix.

:FReg: forward regression.

:Breg: backward regression.

:PCR: Principal Component Regression using PCA.

PortfolioOptimiser.HCObjFuncs — Constant

HCObjFuncs = (:Min_Risk, :Utility, :Sharpe, :Max_Ret, :Equal)

Objective funcions for :NCO HCPortTypes of HCPortfolio.

:Min_Risk, :Utility, :Sharpe and :Max_Ret: optimise the sub-portfolios as :Trad PortTypes of Portfolio according to their respective definitions in ObjFuncs.

:Equal: optimises the sub-portfolios as :RP PortTypes optimisations with equal risk contribution per asset/cluster. We can't offer customiseable risk contributions because the size and composition of the clusters is initially unknown and depends on the chosen linkage method.

PortfolioOptimiser.HCPortTypes — Constant

HCPortTypes = (:HRP, :HERC, :NCO)

Available optimisation types for HCPortfolio.

:HRP: Hierarchical Risk Parity [6–8].

:HERC: Hierarhical Equal Risk Contribution [7, 9, 10].

:NCO: Nested Clustered Optimisation [7, 11].

Both :HERC and :NCO types classify assets into k clusters and split the optimisation into two sub-optimisations.

Intra-cluster optimisation, where each cluster is optimised as its own independent portfolio.

Inter-cluster optimisation, each cluster is treated as a synthetic asset, for which relevant statistics are computed. These are then optimised like a regular portfolio.

Both optimisations are combined to produce the final answer.

PortfolioOptimiser.HCRiskMeasures — Constant

HCRiskMeasures = (:SD, :MAD, :SSD, :FLPM, :SLPM, :WR, :CVaR, :EVaR, :RVaR, :MDD, :ADD,
                  :CDaR, :UCI, :EDaR, :RDaR, :Kurt, :SKurt, :GMD, :RG, :RCVaR, :TG,
                  :RTG, :OWA, :Variance, :Equal, :VaR, :DaR, :DaR_r, :MDD_r, :ADD_r,
                  :CDaR_r, :UCI_r, :EDaR_r, :RDaR_r)

Available risk measures for optimisations of HCPortfolio.

These risk measures are available for all optimisation types:

:SD: standard deviation (SD). Solver must support MOI.SecondOrderCone, or JuMP must be able to transform it/them into a supported form.

:MAD: max absolute deviation (MAD).

:SSD: semi standard deviation (SSD). Solver must support MOI.SecondOrderCone, or JuMP must be able to transform it/them into a supported form.

:FLPM: first lower partial moment (omega ratio) (FLPM).

:SLPM: second lower partial moment (sortino ratio) (SLPM). Solver must support MOI.SecondOrderCone, or JuMP must be able to transform it/them into a supported form.

:WR: worst realisation (WR).

:CVaR: conditional value at risk (CVaR).

:EVaR: entropic value at risk (EVaR). Solver must support MOI.ExponentialCone, or JuMP must be able to transform it/them into a supported form.

:RVaR: relativistic value at risk (RVaR). Solver must support MOI.PowerCone, or JuMP must be able to transform it/them into a supported form.

:MDD: maximum drawdown of uncompounded cumulative returns (MDD_abs).

:ADD: average drawdown of uncompounded cumulative returns (ADD_abs).

:CDaR: conditional drawdown at risk of uncompounded cumulative returns (CDaR_abs).

:UCI: ulcer index of uncompounded cumulative returns (UCI_abs). Solver must support MOI.SecondOrderCone, or JuMP must be able to transform it/them into a supported form.

:EDaR: entropic drawdown at risk of uncompounded cumulative returns (EDaR_abs). Solver must support MOI.ExponentialCone, or JuMP must be able to transform it/them into a supported form.

:RDaR: relativistic drawdown at risk of uncompounded cumulative returns (RDaR_abs). Solver must support MOI.PowerCone, or JuMP must be able to transform it/them into a supported form.

:Kurt: square root kurtosis (Kurt). Solver must support MOI.PSDCone and MOI.SecondOrderCone, or JuMP must be able to transform it/them into a supported form.

:SKurt: square root semi-kurtosis (SKurt). Solver must support MOI.PSDCone and MOI.SecondOrderCone, or JuMP must be able to transform it/them into a supported form.

:GMD: gini mean difference (GMD).

:RG: range of returns (RG).

:RCVaR: range of conditional value at risk (RCVaR).

:TG: tail gini (TG).

:RTG: range of tail gini (RTG).

:OWA: ordered weight array (generic OWA weights) (OWA).

These risk measures are not available with :NCO optimisations.

:Variance: variance (Variance).

:Equal: equal risk contribution, 1/N where N is the number of assets.

:VaR: value at risk (VaR).

:DaR: drawdown at risk of uncompounded cumulative returns (DaR_abs).

:DaR_r: drawdown at risk of compounded cumulative returns (DaR_rel).

:MDD_r: maximum drawdown of compounded cumulative returns (MDD_rel).

:ADD_r: average drawdown of compounded cumulative returns (ADD_rel).

:CDaR_r: conditional drawdown at risk of compounded cumulative returns (CDaR_rel).

:UCI_r: ulcer index of compounded cumulative returns (UCI_rel). Solver must support MOI.SecondOrderCone, or JuMP must be able to transform it/them into a supported form.

:EDaR_r: entropic drawdown at risk of compounded cumulative returns (EDaR_rel). Solver must support MOI.ExponentialCone, or JuMP must be able to transform it/them into a supported form.

:RDaR_r: relativistic drawdown at risk of compounded cumulative returns (RDaR_rel). Solver must support MOI.PowerCone, or JuMP must be able to transform it/them into a supported form.

PortfolioOptimiser.KellyRet — Constant

KellyRet = (:None, :Approx, :Exact)

Available types of Kelly returns for Portfolio.

:None: arithmetic expected returns, $R(\bm{w}) = \bm{\mu} \cdot \bm{w}$.

:Approx: first moment approximation of the logarithmic returns, $R(\bm{w}) = \bm{\mu} \cdot \bm{w} - \dfrac{1}{2} \bm{w}^{\intercal} \mathbf{\Sigma} \bm{w}$.

:Exact: exact logarithmic returns, $R(\bm{w}) = \dfrac{1}{T}\sum\limits_{t=1}^{T}\ln\left(1 + \bm{x}_t \cdot \bm{w}\right)$. Where:

$\mathbf{\Sigma}$ is the covariance matrix of the asset returns.

$\bm{x}_t$ is the vector of asset returns at timestep $t$.

$\bm{\mu}$ is the vector of expected returns for each asset.

and $\bm{w}$ is the asset weights vector.

PortfolioOptimiser.LinkageTypes — Constant

LinkageTypes = (:single, :complete, :average, :ward, :ward_presquared, :DBHT)

Linkage types available when optimising a HCPortfolio.

:DBHT: is Direct Bubble Hierarchical Tree clustering.

The rest are linkage types supported by Clustering.hclust.

PortfolioOptimiser.MuMethods — Constant

MuMethods = (:Default, :JS, :BS, :BOP, :CAPM, :Custom_Func, :Custom_Val)

Methods for estimating the expected returns vector in mean_vec.

:Default: is the standard historical.

:JS: James-Stein.

:BS: Bayes-Stein.

:BOP: Bodnar-Okhrin-Parolya.

:CAPM: Capital Asset Pricing Model.

:Custom_Func: custom function provided.

:Custom_Val: custom value provided.

PortfolioOptimiser.MuTargets — Constant

MuTargets = (:GM, :VW, :SE)

Targets for the :JS, :BS and :BOP estimators in mean_vec and mu_estimator.

:GM: grand mean.

:VW: volatility-weighted grand mean.

:SE: mean square error of sample mean.

PortfolioOptimiser.NetworkMethods — Constant

NetworkMethods = (:None, :SDP, :IP)

Methods for enforcing network constraints for optimising Portfolio.

:None: No network constraint is used.

:SDP: Semi-definite programming constraint.

:IP: Integer programming constraint.

PortfolioOptimiser.ObjFuncs — Constant

ObjFuncs = (:Min_Risk, :Utility, :Sharpe, :Max_Ret)

Objective functions available for use in :Trad and :WC optimisations of Portfolio (see PortTypes).

PortfolioOptimiser.PortClasses — Constant

PortClasses = (:Classic, :FM, :BL, :BLFM)

Available choicees of summary parameters $\bm{\mu}$ and $\bm{\Sigma}$ Portfolio.

:Classic: when optimising with this option, $\bm{\mu}$ and $\bm{\Sigma}$ take their values from historical estimates computed by asset_statistics!.

:FM: when optimising with this option, $\bm{\mu}$ and $\bm{\Sigma}$ take their values from the factor model computed by factor_statistics!.

:BL: when optimising with this option, $\bm{\mu}$ and $\bm{\Sigma}$ take their values from the Black-Litterman model computed by black_litterman_statistics!.

:BLFM: when optimising with this option, $\bm{\mu}$ and $\bm{\Sigma}$ take their values from the factor Black-Litterman model computed by black_litterman_factor_satistics!. This model has two versions defined in BLFMMethods.

PortfolioOptimiser.PortTypes — Constant

PortTypes = (:Trad, :RP, :RRP, :WC)

Available optimisation types for Portfolio.

:Trad – Traditional Optimisations

Available objective functions for :Trad optimisations. We can chose any of the objective functions in ObjFuncs and risk measures in RiskMeasures.

:Min_Risk: minimum risk portfolio,

\[ \begin{align*} \underset{\bm{w}}{\min} &\qquad \phi_{j}(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\ &\qquad R(\bm{w}) \geq \overline{\mu} \end{align*}\,.\]

:Utility: maximum utility portfolio,

\[ \begin{align*} \underset{\bm{w}}{\max} &\qquad R(\bm{w}) - \lambda \phi_{j}(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\ &\qquad R(\bm{w}) \geq \overline{\mu} \end{align*}\,.\]

:Sharpe: maximum risk-adjusted return ratio portfolio,

\[ \begin{align*} \underset{\bm{w}}{\max} &\qquad \dfrac{R(\bm{w}) - r}{\phi_{j}(\bm{w})} \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\ &\qquad R(\bm{w}) \geq \overline{\mu} \end{align*}\,.\]

:Max_Ret: maximum return portfolio,

\[ \begin{align*} \underset{\bm{w}}{\max} &\qquad R(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\ &\qquad R(\bm{w}) \geq \overline{\mu} \end{align*}\,.\]

Where:

$\bm{w}$ are the asset weights.

$\phi_{i}$ is risk measure $i$ from the set of available risk measures $\left\{\Phi\right\}$ (see RiskMeasures).

$\mathbf{A} \bm{w} \geq \bm{B}$ is a set of linear constraints.

$c_{i}$ is the maximum acceptable value for risk measure $\phi_{i}$ of the optimised portfolio.

$R(\bm{w})$ is the return function from KellyRet.

$\overline{\mu}$ is the minimum acceptable return of the optimised portfolio.

$\lambda$ is the risk aversion coefficient.

and $r$ is the risk-free rate.

:RP – Risk Parity Optimisations

Optimises portfolios based on a vector of risk contributions per asset. We can chose any of the risk measures in RiskMeasures.

\[\begin{align*} \underset{\bm{w}}{\min} &\qquad \phi(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B} \\ &\qquad \bm{b} \cdot \ln(\bm{w}) \geq c \\ &\qquad R(\bm{w}) \geq \overline{\mu}\\ &\qquad \bm{w} \geq \bm{0} \end{align*}\,.\]

Where:

$\bm{w}$ are the asset weights.

$\phi$ a risk measure from the set of available risk measures (see RiskMeasures).

$\mathbf{A} \bm{w} \geq \bm{B}$ is a set of linear constraints.

$\bm{b}$ is the vector of maximum allowable risk contribution per asset to the optimised portfolio.

$c$ is an auxiliary variable.

$R(\bm{w})$ is the return function from KellyRet.

and $\overline{\mu}$ is the minimum acceptable return of the optimised portfolio.

:RRP – Relaxed Risk Parity Optimisations

Optimises portfolios based on a vector of risk contributions per asset. Defines its own risk measure using the portfolio returns covariance.

\[\begin{align*} \underset{\bm{w}}{\min} &\qquad \psi - \gamma \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B} \\ &\qquad \bm{w}^\intercal \, \mathbf{\Sigma}\, \bm{w} \leq (\psi^{2} - \rho^{2})\\ &\qquad \lambda \bm{w}^\intercal \, \mathbf{\Theta}\, \bm{w} \leq \rho^{2}\\ &\qquad \bm{\zeta} = \mathbf{\Sigma} \bm{w}\\ &\qquad w_{i} \zeta_{i} \geq \gamma^{2} b_{i} \qquad \forall \, i = 1,\, \ldots{},\, N\\ &\qquad R(\bm{w}) \geq \overline{\mu}\\ &\qquad \sum\limits_{i=1}^{N} w_{i} = 1\\ &\qquad \bm{w} \geq \bm{0}\\ &\qquad \psi,\, \gamma,\, \rho \geq 0 \end{align*}\,.\]

Where:

$\bm{w}$ are the asset weights.

$\psi$ is the average risk of the portfolio.

$\gamma$ is the lower bound of the risk contribution for each asset.

$\mathbf{A} \bm{w} \geq \bm{B}$ is a set of linear constraints.

$\mathbf{\Sigma}$ is the portfolio covariance.

$\rho$ is a regularisation variable.

$\mathbf{\Theta} = \mathrm{diag}\left(\mathbf{\Sigma}\right)$ .

$\lambda$ is a penalty parameter for $\rho$, taken from the available choices in RRPVersions.

$\bm{\zeta}$ is the vector of marginal risk for each asset.

$b_{i}$ is the maximum allowable risk contribution for asset $i$.

$N$ is the number of assets.

$R(\bm{w})$ is the return function from KellyRet.

and $\overline{\mu}$ is the minimum acceptable return of the optimised portfolio.

:WC – Worst Case Mean Variance Optimisations

Computes the worst case mean variance portfolio according to user-selected uncertainty sets (see UncertaintyTypes) for the portfolio return and covariance. We can chose any of the objective functions in ObjFuncs.

:Min_Risk: worst case minimum risk mean-variance portfolio,

\[ \begin{align*} \underset{\bm{w}}{\min} &\qquad \underset{\mathbf{\Sigma}\, \in\, U_{\mathbf{\Sigma}}}{\max}\, \bm{w}^{\intercal}\, \mathbf{\Sigma}\, \bm{w}\\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B}\,. \end{align*}\]

:Utility: worst case maximum utility mean-variance portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad \underset{\bm{\mu}\, \in\, U_{\bm{\mu}}}{\min} R(\bm{w})\, -\, \underset{\mathbf{\Sigma}\, \in\, U_{\mathbf{\Sigma}}}{\max}\, \lambda \bm{w}^{\intercal}\, \mathbf{\Sigma}\, \bm{w}\\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B}\,. \end{align*}\]

:Sharpe: worst case maximum risk-adjusted return ratio mean-variance portfolio,

\[ \begin{align*} \underset{\bm{w}}{\max} &\qquad \dfrac{\underset{\bm{\mu}\, \in\, U_{\bm{\mu}}}{\min} R(\bm{w}) - r}{\underset{\mathbf{\Sigma}\, \in\, U_{\mathbf{\Sigma}}}{\max}\, \left(\bm{w}^{\intercal}\, \mathbf{\Sigma}\, \bm{w}\right)^{1/2}} \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B}\,. \end{align*}\]

:Max_Ret: worst case maximum return mean-variance portfolio,

\[ \begin{align*} \underset{\bm{w}}{\max} &\qquad \underset{\bm{\mu}\, \in\, U_{\bm{\mu}}}{\min} R(\bm{w})\\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B}\,. \end{align*}\]

Where:

$\bm{w}$ are the asset weights.

$\mathbf{\Sigma}$ is the covariance matrix of asset returns.

$U_{\Sigma}$ is the uncertainty set for the covariance matrix, they can be:

\[\begin{align*} U_{\Sigma}^{\mathrm{box}} &= \left\{\mathbf{\Sigma}\, \vert\, \mathbf{\Sigma}_{l} \leq \mathbf{\Sigma} \leq \mathbf{\Sigma}_{u},\, \mathbf{\Sigma} \succeq 0\right\}\\ U_{\Sigma}^{\mathrm{ellipse}} &= \left\{\mathbf{\Sigma}\, \vert\, \left[\mathrm{vec}\left(\mathbf{\Sigma}\right) - \mathrm{vec}\left(\hat{\mathbf{\Sigma}}\right)\right] \mathbf{\Sigma}_{\mathbf{\Sigma}}^{-1} \left[\mathrm{vec}\left(\mathbf{\Sigma}\right) - \mathrm{vec}\left(\hat{\mathbf{\Sigma}}\right)\right]^{\intercal} \leq k_{\mathbf{\Sigma}}^2 ,\, \mathbf{\Sigma} \succeq 0\right\}\,. \end{align*}\]

Where the following variables are estimated by assuming that the portfolio's asset return covariance can be generated by some matrix distribution. The distribution is sampled, and the estimates are calculated from them. Available choices can be found in BoxMethods and EllipseMethods for the box and ellipse sets respectively:

the $\mathrm{l}$ and $\mathrm{u}$ subscripts denote lower and upper bounds for the covariance matrix given the samples.

$\mathbf{\Sigma}_{\mathbf{\Sigma}}$ is the covariance of the samples.

$\hat{\mathbf{\Sigma}}$ the expected covariance given the samples.

and $k_{\mathbf{\Sigma}}$ is a significance parameter of the matrix distribution.

$\mathbf{A} \bm{w} \geq \bm{B}$ is a set of linear constraints.

$\bm{\mu}$ is the vector of expected returns for each asset.

$R(\bm{w})$ is the return function from KellyRet.

$U_{\bm{\mu}}$ is the uncertainty set for the asset returns, they can be:

\[\begin{align*} U_{\bm{\mu}}^{\mathrm{box}} &= \left\{\bm{\mu}\, \vert\, \vert \bm{\mu} - \bm{\hat{\mu}} \vert \leq \delta \right\}\\ U_{\bm{\mu}}^{\mathrm{ellipse}} &= \left\{\bm{\mu}\, \vert\, \left(\bm{\mu} - \bm{\hat{\mu}}\right) \mathbf{\Sigma}_{\bm{\mu}}^{-1} \left(\bm{\mu} - \bm{\hat{\mu}}\right)^{\intercal} \leq k_{\bm{\mu}}^{2}\right\}\,. \end{align*}\]

Where the following variables are estimated by assuming that the portfolio's asset expected returns can be generated by some distribution. The distribution is sampled, and the estimates are calculated from them. Available choices can be found in BoxMethods and EllipseMethods for the box and ellipse sets respectively:

$\hat{\bm{\mu}}$ the expected portfolio asset expected returns given the samples.

$\mathbf{\Sigma}_{\bm{\mu}}$ is the covariance of the samples.

and $k_{\bm{\mu}}$ is a significance parameter of the distribution.

$\lambda$ is the risk aversion coefficient.

and $r$ is the risk-free rate.

The worst case uncertainty sets are computed by wc_statistics!.

PortfolioOptimiser.PosdefFixMethods — Constant

PosdefFixMethods = (:None, :Nearest, :Custom_Func)

Methods for fixing non-positive definite matrices.

:None: no fix is applied.

:Nearest: nearest correlation matrix.

Custom_Func: custom function provided.

PortfolioOptimiser.RRPVersions — Constant

RRPVersions = (:None, :Reg, :Reg_Pen)

Available versions of Relaxed Risk Parity Optimisations (see PortTypes).

:None: no penalty.

:Reg: regularisation constraint, $\rho$.

:Reg_Pen: regularisation and penalisation constraints, $\lambda$ and $\rho$.

PortfolioOptimiser.RegCriteria — Constant

RegCriteria = (:pval, :aic, :aicc, :bic, :r2, :adjr2)

Criteria for feature selection in regression functions.

:pval: p-value feature selection.

The rest are methods applied to a fitted General Linear Model from GLM.jl.

:aic, :aicc, :bic: selects based on lowest value.

:r2, :adjr2: selects based on highest value.

PortfolioOptimiser.RiskMeasureNames — Constant

RiskMeasureNames = (SD = "Standard Deviation", MAD = "Mean Absolute Deviation",
                    SSD = "Semi Standard Deviation",
                    FLPM = "First Lower Partial Moment",
                    SLPM = "Second Lower Partial Moment", WR = "Worst Realisation",
                    CVaR = "Conditional Value at Risk",
                    EVaR = "Entropic Value at Risk",
                    RVaR = "Relativistic Value at Risk",
                    MDD = "Uncompounded Max Drawdown",
                    ADD = "Uncompounded Average Drawdown",
                    CDaR = "Conditional Uncompounded Drawdown at Risk",
                    UCI = "Uncompounded Ulcer Index",
                    EDaR = "Entropic Uncompounded Drawdown at Risk",
                    RDaR = "Relativistic Uncompounded Drawdown at Risk",
                    Kurt = "Square Root Kurtosis",
                    SKurt = "Square Root Semi Kurtosis", GMD = "Gini Mean Difference",
                    RG = "Range", RCVaR = "Conditional Value at Risk Range",
                    TG = "Tail Gini", RTG = "Tail Gini Range",
                    OWA = "Ordered Weight Average", Variance = "Variance",
                    Equal = "Equal Risk", VaR = "Value at Risk",
                    DaR = "Uncompounded Drawdown at Risk",
                    DaR_r = "Compounded Drawdown at Risk",
                    MDD_r = "Compounded Max Drawdown",
                    ADD_r = "Compounded Average Drawdown",
                    CDaR_r = "Conditional Uncompounded Drawdown at Risk",
                    UCI_r = "Compounded Ulcer Index",
                    EDaR_r = "Entropic Compounded Drawdown at Risk",
                    RDaR_r = "Relativistic Compounded Drawdown at Risk")

Names of risk measures.

PortfolioOptimiser.RiskMeasures — Constant

RiskMeasures = (:SD, :MAD, :SSD, :FLPM, :SLPM, :WR, :CVaR, :EVaR, :RVaR, :MDD, :ADD,
                :CDaR, :UCI, :EDaR, :RDaR, :Kurt, :SKurt, :GMD, :RG, :RCVaR, :TG,
                :RTG, :OWA)

Available risk measures for :Trad and :RP type (see PortTypes) of Portfolio.

:SD: standard deviation (SD). Solver must support MOI.SecondOrderCone, or JuMP must be able to transform it/them into a supported form.

:MAD: max absolute deviation (MAD).

:SSD: semi standard deviation (SSD). Solver must support MOI.SecondOrderCone, or JuMP must be able to transform it/them into a supported form.

:FLPM: first lower partial moment (omega ratio) (FLPM).

:SLPM: second lower partial moment (sortino ratio) (SLPM). Solver must support MOI.SecondOrderCone, or JuMP must be able to transform it/them into a supported form.

:WR: worst realisation (WR).

:CVaR: conditional value at risk (CVaR).

:EVaR: entropic value at risk (EVaR). Solver must support MOI.ExponentialCone, or JuMP must be able to transform it/them into a supported form.

:RVaR: relativistic value at risk (RVaR). Solver must support MOI.PowerCone, or JuMP must be able to transform it/them into a supported form.

:MDD: maximum drawdown of uncompounded cumulative returns (MDD_abs).

:ADD: average drawdown of uncompounded cumulative returns (ADD_abs).

:CDaR: conditional drawdown at risk of uncompounded cumulative returns (CDaR_abs).

:UCI: ulcer index of uncompounded cumulative returns (UCI_abs). Solver must support MOI.SecondOrderCone, or JuMP must be able to transform it/them into a supported form.

:EDaR: entropic drawdown at risk of uncompounded cumulative returns (EDaR_abs). Solver must support MOI.ExponentialCone, or JuMP must be able to transform it/them into a supported form.

:RDaR: relativistic drawdown at risk of uncompounded cumulative returns (RDaR_abs). Solver must support MOI.PowerCone, or JuMP must be able to transform it/them into a supported form.

:Kurt: square root kurtosis (Kurt). Solver must support MOI.PSDCone and MOI.SecondOrderCone, or JuMP must be able to transform it/them into a supported form.

:SKurt: square root semi-kurtosis (SKurt). Solver must support MOI.PSDCone and MOI.SecondOrderCone, or JuMP must be able to transform it/them into a supported form.

:GMD: gini mean difference (GMD).

:RG: range of returns (RG).

:RCVaR: range of conditional value at risk (RCVaR).

:TG: tail gini (TG).

:RTG: range of tail gini (RTG).

:OWA: ordered weight array (generic OWA weights) (OWA).

PortfolioOptimiser.TrackingErrKinds — Constant

TrackingErrKinds = (:Weights, :Returns)

Available kinds of tracking errors for Portfolio.

:Weights: provide a vector of asset weights which is used to compute the vector of benchmark returns, - $\bm{b} = \mathbf{X} \bm{w}$,where $\bm{b}$ is the benchmark returns vector, $\mathbf{X}$ the $T \times{} N$ asset returns matrix, and $\bm{w}$ the asset weights vector.

:Returns: directly provide the vector of benchmark returns. The benchmark is then used as a reference to optimise a portfolio that tracks it up to a given error.

PortfolioOptimiser.UncertaintyTypes — Constant

UncertaintyTypes = (:None, :Box, :Ellipse)

Available types of uncertainty sets that can be computed with wc_statistics!, which are used by Worst Case Mean Variance Optimisations (see PortTypes, EllipseMethods, and BoxMethods).

:Box: are box uncertainty sets, ie the sets are full matrices.

:Ellipse: are elliptical uncertainty sets, ie the sets are diagonal matrices.

PortfolioOptimiser.ValidTermination — Constant

ValidTermination = (MOI.OPTIMAL, MOI.ALMOST_OPTIMAL, MOI.LOCALLY_SOLVED,
                    MOI.ALMOST_LOCALLY_SOLVED, MOI.SOLUTION_LIMIT, MOI.OBJECTIVE_LIMIT)

Valid JuMP termination codes after optimising an instance of Portfolio. If the termination code is different to these, then the failures are logged in the .fail field of HCPortfolio and Portfolio.