Constraint Functions

Public

PortfolioOptimiser.asset_constraints — Method

asset_constraints(constraints::DataFrame, asset_sets::DataFrame)

Create the linear constraint matrix A and vector B:

$\mathbf{A} \bm{w} \geq \bm{B}$.

Inputs

constraints: Nc×10 Dataframe, where Nc is the number of constraints. The required columns are:

Enabled: (Bool) indicates if the constraint is enabled.

Type: (String) specifies the object(s) to which a constraint applies:

Asset: specific asset.

Subset: whole class.

All Assets: all assets.

All Subsets: all asset classes.

Each Asset in Subset: specific assets in a class.

Set: (String) if Type is Subset, All Subsets or Each Asset in Subset, specifies the asset class set.

Position: (String) name of the asset or asset class to which the constraint applies.

Sign: (String) specifies whether the constraint is a lower or upper bound:

>=: lower bound.

<=: upper bound.

Weight: (<:Real) value of the constraint.

Relative_Type: (String) specifies to what the constraint is relative:

Empty string: nothing.

Asset: other asset.

Subset: other class.

Relative_Set: (String) if Relative_Type is Subset, specifies the name of the set of asset classes.

Relative_Position: (String) name of the asset or asset class of the relative constraint.

Factor: (<:Real) the factor of the relative constraint.

asset_sets: Na×D DataFrame where Na is the number of assets and D the number of columns.

Asset: list of assets, this is the only mandatory column.

Subsequent columns specify the asset class sets.

Outputs

A: Nc×Na matrix of constraints where Nc is the number of constraints and Na is the number of assets.

B: Nc×1 vector of constraints where Nc is the number of constraints.

Examples

asset_sets = DataFrame(
    "Asset" => ["FB", "GOOGL", "NTFX", "BAC", "WFC", "TLT", "SHV", "FCN", "TKO", "ZOO", "ZVO", "ZX", "ZZA", "ZZB", "ZZC"],    "Class 1" => ["Equity", "Equity", "Equity", "Equity", "Equity", "Fixed Income", "Fixed Income", "Equity", "Equity", "Equity", "Fixed Income", "Fixed Income", "Equity", "Fixed Income", "Equity"],    "Class 2" => ["Technology", "Technology", "Technology", "Financial", "Financial", "Treasury", "Treasury", "Financial", "Entertainment", "Treasury", "Financial", "Financial", "Entertainment", "Technology", "Treasury"])
constraints = DataFrame(
    "Enabled" => [true, true, true, true, true, true, true, true, true, true, true, true, true, true, true],    "Type" => ["Subset", "All Subsets", "Asset", "Asset", "Subset", "All Assets", "Each Asset in Subset", "Asset", "All Assets", "All Assets", "Subset", "All Subsets", "All Subsets", "Each Asset in Subset", "Each Asset in Subset"],    "Set" => ["Class 1", "Class 1", "", "", "Class 2", "", "Class 1", "Class 1", "Class 2", "", "Class 1", "Class 2", "Class 2", "Class 2", "Class 1"],    "Position" => ["Equity", "Fixed Income", "BAC", "WFC", "Financial", "", "Equity", "FCN", "TKO", "ZOO", "Fixed Income", "Treasury", "Entertainment", "Treasury", "Equity"],    "Sign" => ["<=", "<=", "<=", "<=", ">=", ">=", ">=", "<=", ">=", "<=", ">=", "<=", ">=", "<=", ">="],    "Weight" => [0.6, 0.5, 0.1, "", "", 0.02, "", "", "", "", "", "", "", 0.27, ""],    "Relative_Type" => ["", "", "", "Asset", "Subset", "", "Asset", "Subset", "Asset", "Subset", "Asset", "Asset", "Subset", "", "Subset"],    "Relative_Set" => ["", "", "", "", "Class 1", "", "", "Class 1", "", "Class 2", "", "Class 2", "Class 2", "", "Class 2"],    "Relative_Position" => ["", "", "", "FB", "Fixed Income", "", "TLT", "Equity", "NTFX", "Financial", "WFC", "ZOO", "Entertainment", "", "Entertainment"],    "Factor" => ["", "", "", 1.2, 0.5, "", 0.4, 0.7, 0.21, 0.11, 0.13, -0.17, 0.23, "", -0.31])
A, B = asset_constraints(constraints, asset_sets)

PortfolioOptimiser.asset_views — Method

asset_views(views::DataFrame, asset_sets::DataFrame)

Create the asset views matrix P and vector Q:

$\mathbf{P} \bm{w} \geq \bm{Q}$.

Inputs

views: Nv×9 DataFrame, where Nv is the number of views. The required columns are:

Enabled: (Bool) indicates if the view is enabled.

Type: (String) specifies the object(s) to which a view applies:

Asset: specific asset.

Subset: whole class.

Set: (String) if Type is Subset, specifies the asset class set.

Position: (String) name of the asset or asset class to which the view applies.

Sign: (String) specifies whether the view is a lower or upper bound:

>=: lower bound.

<=: upper bound.

Return: (<:Real) the view's return.

Relative_Type: (String) specifies to what the view is relative:

Empty string: nothing.

Asset: other asset.

Subset: other class.

Relative_Set: (String) if Relative_Type is Subset, specifies the name of the set of asset classes.

Relative_Position: (String) name of the asset or asset class of the relative view.

asset_sets: Na×D DataFrame where Na is the number of assets and D the number of columns.

Asset: list of assets, this is the only mandatory column.

Subsequent columns specify the asset class sets.

Outputs

P: Nv×Na matrix of views where Nv is the number of views and Na is the number of assets.

Q: Nv×1 vector of views where Nv is the number of views.

Examples

asset_sets = DataFrame(
        "Asset" => ["FB", "GOOGL", "NTFX", "BAC", "WFC", "TLT", "SHV"],        "Class 1" => ["Equity", "Equity", "Equity", "Equity", "Equity", "Fixed Income", "Fixed Income"],        "Class 2" => ["Technology", "Technology", "Technology", "Financial", "Financial", "Treasury", "Treasury"],    )
views = DataFrame(
        "Enabled" => [true, true, true, true, true],        "Type" => ["Asset", "Subset", "Subset", "Asset", "Subset"],        "Set" => ["", "Class 2", "Class 1", "", "Class 1"],        "Position" => ["WFC", "Financial", "Equity", "FB", "Fixed Income"],        "Sign" => ["<=", ">=", ">=", ">=", "<="],        "Return" => [0.3, 0.1, 0.05, 0.03, 0.017],        "Relative_Type" => ["Asset", "Subset", "Asset", "", ""],        "Relative_Set" => ["", "Class 1", "", "", ""],        "Relative_Position" => ["FB", "Fixed Income", "TLT", "", ""],    )
P, Q = asset_views(views, asset_sets)

PortfolioOptimiser.cluster_matrix — Method

cluster_matrix

PortfolioOptimiser.cluster_matrix — Method

cluster_matrix

PortfolioOptimiser.factor_constraints — Method

factor_constraints(constraints::DataFrame, loadings::DataFrame)

Create the factor constraints matrix C and vector D:

$\mathbf{C} \bm{w} \geq \bm{D}$.

Inputs

constraints: Nc×4 Dataframe, where Nc is the number of constraints. The required columns are:

Enabled: (Bool) indicates if the constraint is enabled.

Factor: (String) name of the constraint's factor.

Sign: (String) specifies whether the constraint is a lower or upper bound:

>=: lower bound.

<=: upper bound.

Value: (<:Real) the upper or lower bound of the factor's value.

Relative_Factor: (String) factor to which the constraint is relative.

loadings: Nl×Nf loadings DataFrame, where Nl is the number of data points, and Nf is the number of factors.

Outputs

C: Nc×Nf matrix of constraints where Nc is the number of constraints and Nf is the number of factors.

D: Nc×1 vector of constraints where Nc is the number of constraints.

Examples

loadings = DataFrame(
        "const" => [0.0004, 0.0002, 0.0000, 0.0006, 0.0001, 0.0003, -0.0003],        "MTUM" => [0.1916, 1.0061, 0.8695, 1.9996, 0.0000, 0.0000, 0.0000],        "QUAL" => [0.0000, 2.0129, 1.4301, 0.0000, 0.0000, 0.0000, 0.0000],        "SIZE" => [0.0000, 0.0000, 0.0000, 0.4717, 0.0000, -0.1857, 0.0000],        "USMV" => [-0.7838, -1.6439, -1.0176, -1.4407, 0.0055, 0.5781, 0.0000],        "VLUE" => [1.4772, -0.7590, -0.4090, 0.0000, -0.0054, -0.4844, 0.9435],    )
constraints = DataFrame(
        "Enabled" => [true, true, true, true],        "Factor" => ["MTUM", "USMV", "VLUE", "const"],        "Sign" => ["<=", "<=", ">=", ">="],        "Value" => [0.9, -1.2, 0.3, -0.1],        "Relative_Factor" => ["USMV", "", "", "SIZE"],    )
C, D = factor_constraints(constraints, loadings)

PortfolioOptimiser.factor_views — Method

factor_views(views::DataFrame, loadings::DataFrame)

Create the factor views matrix P and vector Q:

$\mathbf{P} \bm{w} \geq \bm{Q}$.

Inputs

views: Nv×4 DataFrame, where Nv is the number of views. The required columns are:

Enabled: (Bool) indicates if the view is enabled.

Factor: (String) name of the view's factor.

Sign: (String) specifies whether the view is a lower or upper bound:

>=: lower bound.

<=: upper bound.

Value: (<:Real) the upper or lower bound of the factor's value.

Relative_Factor: (String) factor to which the view is relative.

loadings: Nl×Nf loadings DataFrame, where Nl is the number of data points, and Nf is the number of factors.

Outputs

P: Nv×Nf matrix of views where Nv is the number of views and Nf is the number of factors.

Q: Nv×1 vector of views where Nv is the number of views.

Examples

loadings = DataFrame(
        "const" => [0.0004, 0.0002, 0.0000, 0.0006, 0.0001, 0.0003, -0.0003],        "MTUM" => [0.1916, 1.0061, 0.8695, 1.9996, 0.0000, 0.0000, 0.0000],        "QUAL" => [0.0000, 2.0129, 1.4301, 0.0000, 0.0000, 0.0000, 0.0000],        "SIZE" => [0.0000, 0.0000, 0.0000, 0.4717, 0.0000, -0.1857, 0.0000],        "USMV" => [-0.7838, -1.6439, -1.0176, -1.4407, 0.0055, 0.5781, 0.0000],        "VLUE" => [1.4772, -0.7590, -0.4090, 0.0000, -0.0054, -0.4844, 0.9435],    )
views = DataFrame(
    "Enabled" => [true, true, true],    "Factor" => ["MTUM", "USMV", "VLUE"],    "Sign" => ["<=", "<=", ">="],    "Value" => [0.9, -1.2, 0.3],    "Relative_Factor" => ["USMV", "", ""])
P, Q = factor_views(views, loadings)

PortfolioOptimiser.hrp_constraints — Method

hrp_constraints(constraints::DataFrame, asset_sets::DataFrame)

Create the upper and lower bounds constraints for hierarchical risk parity portfolios.

Inputs

constraints: Nc×4 Dataframe, where Nc is the number of constraints. The required columns are:

Enabled: (Bool) indicates if the constraint is enabled.

Type: (String) specifies the object(s) to which a constraint applies:

Asset: specific asset.

All Assets: all assets.

Each Asset in Subset: specific assets in a class.

Position: (String) name of the asset or asset class to which the constraint applies.

Sign: (String) specifies whether the constraint is a lower or upper bound:

>=: lower bound.

<=: upper bound.

Weight: (<:Real) value of the constraint.

asset_sets: Na×D DataFrame where Na is the number of assets and D the number of columns.

Asset: list of assets, this is the only mandatory column.

Subsequent columns specify the asset class sets.

Outputs

w_min: Na×1 vector of the lower bounds for asset weights.

w_max: Na×1 vector of the upper bounds for asset weights.

Examples

asset_sets = DataFrame(
        "Asset" => ["FB", "GOOGL", "NTFX", "BAC", "WFC", "TLT", "SHV"],        "Class 1" => ["Equity", "Equity", "Equity", "Equity", "Equity", "Fixed Income", "Fixed Income"],        "Class 2" => ["Technology", "Technology", "Technology", "Financial", "Financial", "Treasury", "Treasury"],    )
constraints = DataFrame(
    "Enabled" => [true, true, true, true, true, true],    "Type" => ["Asset", "Asset", "All Assets", "All Assets", "Each Asset in Subset", "Each Asset in Subset"],    "Set" => ["", "", "", "", "Class 1", "Class 2"],    "Position" => ["BAC", "FB", "", "", "Fixed Income", "Financial"],    "Sign" => [">=", "<=", "<=", ">=", "<=", "<="],    "Weight" => [0.02, 0.085, 0.09, 0.01, 0.07, 0.06])
w_min, w_max = hrp_constraints(constraints, asset_sets)

PortfolioOptimiser.related_assets — Method

related_assets

PortfolioOptimiser.related_assets — Method

related_assets

PortfolioOptimiser.rp_constraints — Method

rp_constraints(
    asset_sets::DataFrame,    type::Symbol = :Asset,    class_col::Union{String, Symbol, Nothing} = nothing)

Constructs risk contribution constraint vector for the risk parity optimisation (:RP and :RRP types of PortTypes).

Inputs

asset_sets: Na×D DataFrame where Na is the number of assets and D the number of columns.

Asset: list of assets, this is the only mandatory column.

Subsequent columns specify the asset class sets. They are only used if type == :Subset.

class_col: index of set of classes from asset_sets to use in when type == :Subset.

Outputs

rw: risk contribution constraint vector.

Examples

asset_sets = DataFrame(
        "Asset" => ["FB", "GOOGL", "NTFX", "BAC", "WFC", "TLT", "SHV"],        "Class 1" => ["Equity", "Equity", "Equity", "Equity", "Equity", "Fixed Income", "Fixed Income"],        "Class 2" => ["Technology", "Technology", "Technology", "Financial", "Financial", "Treasury", "Treasury"],    )

rw_a = rp_constraints(asset_sets, :Asset)
rw_c = rp_constraints(asset_sets, :Subset, "Class 2")

PortfolioOptimiser.turnover_constraints — Method

turnover_constraints(constraints::DataFrame, asset_sets::DataFrame)

"Asset"

"All Assets"

"Each Asset in Subset"

Private