Direct Bubble Hierarchical Tree Clustering

PortfolioOptimiser.PMFG_T2s — Function

PMFG_T2s(W::AbstractMatrix{<:Number}, nargout::Integer = 3)

Constructs a Triangulated Maximally Filtered Graph (TMFG) starting from a tetrahedron and recursively inserting vertices inside existing triangles (T2 move) in order to approximate a Maximal Planar Graph with the largest total weight, aka Planar Maximally Filtered Graph (PMFG). All weights are non-negative [PMFG].

Arguments

W: N×N matrix of non-negative weights.

nargout: number of output arguments, the same arguments are always returne, this only controls whether some arguments are empty or not.

Outputs

A: adjacency matrix of the PMFG with weights.

tri: list of triangles (triangular faces).

clique3: list of 3-cliques taht are not triangular faces, all 3-cliques are given by [tri; clique3].

cliques: list of all 4-cliques, if nargout <= 3, this will be returned as an empty array.

cliqueTree: 4-cliques tree structure (adjacency matrix), if nargout <= 4, it is returned as an empty array.

PortfolioOptimiser.distance_wei — Function

distance_wei(L::AbstractMatrix{<:Number})

The distance matrix contains lengths of shortest paths between all node pairs. An entry [u, v] represents the length of the shortest path from node u to node v. The average shortest path length is the characteristic path length of the network. The function uses Dijkstra's algorithm.

Inputs

L: Directed/undirected connection-length matrix.

Lengths between disconnected nodes are set to Inf.

Lengths on the main diagonal are set to 0.

Note

The input matrix must be a connection-length matrix typically obtained via a mapping from weight to length. For instance, in a weighted correlation network, higher correlations are more naturally interpreted as shorter distances, and the input matrix should therefore be some inverse of the connectivity matrix, i.e. a distance matrix.

The number of edges in the shortest weighted path may in general exceed the number of edges in the shortest binary paths (i.e. the shortest weighted paths computed on the binarised connectivity matrix), because the shortest weighted paths have the minimal weighted distance, not necessarily the minimal number of edges.

Outputs

D: distance (shortest weighted path) matrix.

B: number of edged in the shortest weigthed path matrix.

PortfolioOptimiser.clique3 — Function

clique3(A::AbstractMatrix{<:Number})

Computes the list of 3-cliques.

Inputs

A: N×N adjacency matrix of a Maximal Planar Graph (MPG).

Outputs

K3: vector of vectors with the corresponding indices of candidate cliques.

E: matrix with non-zero indices and entries of candidate cliques.

CliqList: Nc×3 matrix. Each row vector lists the three vertices consisting of a 3-clique in the MPG.

PortfolioOptimiser.breadth — Function

breadth(CIJ::AbstractMatrix{<:Number}, source::Integer)

Breadth-first search.

Inputs

CIJ: binary (directed/undirected) connection matrix.

source: source vertex.

Outputs

distance: distance between source and i'th vertex (0 for source vertex).

branch: vertex that precedes i in teh breadth-first search treee (-1 for source vertex).

Note

Breadth-first search tree does not contain all paths (or all shortest paths), but allows the determination of at least one path with minimum distace. The entire graph is explored, starting from source vertex source.

PortfolioOptimiser.BubbleCluster8s — Function

BubbleCluster8s(Rpm, Dpm, Hb, Mb, Mv, CliqList)

Obtains non-discrete and discrete clusterings from the bubble topology of the Planar Maximally Filtered Graph (PMFG).

Inputs

Rpm: N×N sparse weighted adjacency matrix of the PMFG.

Dpm: N×N shortest path lengths matrix of the PMFG.

Hb: undirected bubble tree of the PMFG.

Mb: Nc×Nb bubble membership matrix for 3-cliques. Mb[n, bi] = 1 indicated that 3-clique n belongs to bubble bi.

Mv: N×Nb bubble membership matrix for vertices.

CliqList: Nc×3 matrix of 3-cliques. Each row vector contains the list of vertices for a particular 3-clique.

Outputs

Adjv: N×Nk cluster membership matrix for vertices for non-discrete clustering via the bubble topology. Adjv[n, k] = 1 indicates cluster membership of vertex n to the k'th non-discrete cluster.

Tc: N×1 cluster membership vector. Tc[n] = k indicates cluster membership of vertex n to the k'th discrete cluster.

PortfolioOptimiser.DBHTRootMethods — Constant

DBHTRootMethods = (:Unique, :Equal)

Methods for finding the root of a Direct Bubble Hierarchical Clustering Tree in DBHTs, in case there is more than one candidate.

:Unique: create a unique root.

:Equal: the root is created from the candidate's adjacency tree.

PortfolioOptimiser.CliqHierarchyTree2s — Function

CliqHierarchyTree2s(Apm::AbstractMatrix{<:Number}, method::Symbol = :Unique)

Looks for 3-cliques of a Maximal Planar Graph (MPG), then construct a hierarchy of the cliques with the definition of "inside" a clique being a subgraph of smaller size when the entire graph is made disjoint by removing the clique [NHPG].

Inputs

Apm: N×N adjacency matrix of an MPG.

method: method for finding the root of the graph DBHTRootMethods. Uses Voronoi tesselation between tiling triangles.

Outputs

H1: Nc×Nc adjacency matrix for 3-clique hierarchical tree where Nc is the number of 3-cliques.

H2: Nb×Nb adjacency matrix for the bubble hierarchical tree where Nb is the number of bubbles.

Mb: Nc×Nb bubble membership matrix. Mb[n, bi] = 1 indicates that 3-clique n, belongs to bubble bi.

CliqList: Nc×3 matrix. Each row vector lists the three vertices consisting of a 3-clique in the MPG.

Sb: Nc×1 vector. Sb[n] = 1 indicates 3-clique n is separating.

PortfolioOptimiser.DBHTs — Function

DBHTs(
    D::AbstractMatrix{<:Number},
    S::AbstractMatrix{<:Number};
    branchorder::Symbol = :optimal,
    method::Symbol = :Unique,
)

Perform Direct Bubble Hierarchical Tree clustering, a deterministic clustering algorithm [DBHTs]. This version uses a graph-theoretic filtering technique called Triangulated Maximally Filtered Graph (TMFG).

Arguments

D: N×N dissimilarity matrix, e.g. a distance matrix.

S: N×N non-negative similarity matrix, e.g. a correlation matrix + 1, S = 2 .- D .^2 / 2; or another choice could be S = exp.(-D).

branchorder: is a parameter for ordering the final dendrogram's branches. The choices are defined by BranchOrderTypes.

method: method for finding the root of a Direct Bubble Hierarchical Clustering Tree in case there is more than one candidate DBHTRootMethods.

Outputs

T8: N×1 cluster membership vector.

Rpm: N×N adjacency matrix of the Planar Maximally Filtered Graph (PMFG).

Adjv: Bubble cluster membership matrix from BubbleCluster8s.

Dpm: N×N shortest path length matrix of the PMFG.

Mv: N×Nb bubble membership matrix. Mv[n, bi] = 1 means vertex n is a vertex of bubble bi.

Z: DBHT hierarchy matrix.

Z_hclust: Z matrix in Clustering.Hclust format.

PMFGGuido Previde Massara, T. Di Matteo, Tomaso Aste, Network Filtering for Big Data: Triangulated Maximally Filtered Graph, Journal of Complex Networks, Volume 5, Issue 2, June 2017, Pages 161–178, https://doi.org/10.1093/comnet/cnw015.

NHPGSong, W. M., Di Matteo, T., & Aste, T. (2011). Nested hierarchies in planar graphs. Discrete Applied Mathematics, 159(17), 2135-2146.

DBHTsSong, Won-Min, T. Di Matteo, and Tomaso Aste. "Hierarchical information clustering by means of topologically embedded graphs." PloS one 7.3 (2012): e31929.