Definitions

PortfolioOptimiser.AbstractPortfolio — Type

Abstract type for portfolios. Concrete portfolios subtype this see Portfolio and HCPortfolio.

abstract type AbstractPortfolio end

PortfolioOptimiser.RiskMeasures — Constant

const RiskMeasures = (
    :SD,
    :MAD,
    :SSD,
    :FLPM,
    :SLPM,
    :WR,
    :CVaR,
    :EVaR,
    :RVaR,
    :MDD,
    :ADD,
    :CDaR,
    :UCI,
    :EDaR,
    :RDaR,
    :Kurt,
    :SKurt,
    :GMD,
    :RG,
    :RCVaR,
    :TG,
    :RTG,
    :OWA,
)

Available risk measures for type = :trad optimisations of Portfolio. - :SD = standard deviation SD - :MAD = max absolute deviation MAD - :SSD = semi standard deviation SSD - :FLPM = first lower partial moment (omega ratio) FLPM - :SLPM = second lower partial moment (sortino ratio) SLPM - :WR = worst realisation WR - :CVaR = conditional value at risk CVaR - :EVaR = entropic value at risk EVaR - :RVaR = relativistic value at risk RVaR - :MDD = maximum drawdown for uncompounded cummulative returns MDD_abs - :ADD = average drawdown for uncompounded cummulative returns ADD_abs - :CDaR = conditional drawdown at risk for uncompounded cummulative returns CDaR_abs - :UCI = ulcer index for uncompounded cummulative returns UCI_abs - :EDaR = entropic drawdown at risk for uncompounded cummulative returns EDaR_abs - :RDaR = relativistic drawdown at risk for uncompounded cummulative returns RDaR_abs - :Kurt = square root kurtosis Kurt - :SKurt = square root semi-kurtosis SKurt - :GMD = gini mean difference GMD - :RG = range of returns RG - :RCVaR = range of conditional value at risk RCVaR - :TG = tail gini TG - :RTG = range of tail gini RTG - :OWA = ordered weight array (generic OWA weights) OWA

PortfolioOptimiser.KellyRet — Constant

const KellyRet = (:none, :approx, :exact)

Available types of Kelly returns for Portfolio. - :none: arithmetic mean return - :approx: first moment approximation of logarithmic returns, $\mu_{\ln_{\mathrm{a}}} = \mathbf{X} \bm{w} - \dfrac{1}{2} \bm{w}^{\intercal} \mathbf{\Sigma} \bm{w})$ - :exact: exact logarithmic returns, $\mu_{\ln} = \ln\left(1+ \mathbf{X} \bm{w}\right)$

PortfolioOptimiser.TrackingErrKinds — Constant

const TrackingErrKinds = (:weights, :returns)

Available kinds of tracking errors for Portfolio.

PortfolioOptimiser.ObjFuncs — Constant

Available objective functions for Portfolio.

const ObjFuncs = (:min_risk, :utility, :sharpe, :max_ret)

PortfolioOptimiser.ValidTermination — Constant

Valid JuMP terminations after optimising an instance of Portfolio.

const ValidTermination =
    (MOI.OPTIMAL, MOI.ALMOST_OPTIMAL, MOI.LOCALLY_SOLVED, MOI.ALMOST_LOCALLY_SOLVED)

PortfolioOptimiser.PortClasses — Constant

Available classes for Portfolio.

const PortClasses = (:classic,)

PortfolioOptimiser.PortTypes — Constant

Available optimisation types for Portfolio.

const PortTypes = (:trad, :rp, :rrp, :OWA, :wc)

PortfolioOptimiser.RRPVersions — Constant

Available versions of relaxed risk parity optimisations of Portfolio.

const RRPVersions = (:none, :reg, :reg_pen)

PortfolioOptimiser.RPConstraintTypes — Constant

Types of risk parity constraints for rp_constraints.

const RPConstraintTypes = (:assets, :classes)

PortfolioOptimiser.UncertaintyTypes — Constant

Types of uncertainty sets for worst case optimisations of Portfolio.

const UncertaintyTypes = (:none, :box, :ellipse)

PortfolioOptimiser.KindBootstrap — Constant

Bootstrap for worst case optimisations.

const KindBootstrap = (:stationary, :circular, :moving)

PortfolioOptimiser.EllipseTypes — Constant

Ellipse and box types for worst case optimisations.

const EllipseTypes = (:stationary, :circular, :moving, :normal)
const BoxTypes = (EllipseTypes..., :delta)