Ordered Weight Arrays

Public API

PortfolioOptimiser.owa_gmd — Function

owa_gmd(T::Integer)

Computes the Gini Mean Difference (GMD) of a returns series [OWA].

Inputs

T: number of observations of the returns series.

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_cvar — Function

owa_cvar(T::Integer, α::Real = 0.05)

Calculate the OWA weights corresponding to the Critical Value at Risk (CVaR) of a returns series [OWA].

Inputs

T: number of observations of the returns series.

α: significance level of CVaR, α ∈ (0, 1).

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_wcvar — Function

owa_wcvar(
    T::Integer,
    αs::AbstractVector{<:Real},
    ws::AbstractVector{<:Real},
)

Compute the OWA weights for the Weighted Conditional Value at Risk (WCVaR) of a returns series [OWA].

Inputs

T: number of observations of the returns series.

αs: N×1 vector of significance levels of each CVaR model, where N is the number of models, each α_n ∈ (0, 1).

ws: N×1 vector of weights of each CVaR model, where N is the number of models.

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_tg — Function

owa_tg(
    T::Integer;
    alpha_i::Real = 0.0001,
    alpha::Real = 0.05,
    a_sim::Integer = 100,
)

Compute the OWA weights for the Tail Gini of a returns series [OWA].

Inputs

T: number of observations of the returns series.

alpha_i: initial significance level of Tail Gini.

alpha: significance level of Tail Gini.

a_sim: number of CVaRs to approximate the Tail Gini.

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_wr — Function

owa_wr(T::Integer)

Compute the OWA weights for the Worst Realisation (WR) of a returns series [OWA].

Inputs

T: number of observations of the returns series.

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_rcvar — Function

owa_rcvar(T::Integer; alpha::Real = 0.05, beta::Union{<:Real, Nothing} = nothing)

Compute the OWA weights for the CVaR Range of a returns series [OWA].

Inputs

T: number of observations of the returns series.

alpha: significance level of CVaR losses, alpha ∈ (0, 1).

beta: significance level of CVaR gains, beta ∈ (0, 1).

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_rwcvar — Function

owa_rwcvar(
    T::Integer,
    αs::AbstractVector{<:Real},
    wα::AbstractVector{<:Real},
    βs::Union{AbstractVector{<:Real}, Nothing} = nothing,
    wβ::Union{AbstractVector{<:Real}, Nothing} = nothing,
)

Compute the OWA weights for the Weighted Conditional Value at Risk (WCVaR) of a returns series [OWA].

Inputs

T: number of observations of the returns series.

αs: N×1 vector of significance levels of the losses for each CVaR model, where N is the number of losses models, each α_n ∈ (0, 1).

wα: N×1 vector of weights of the losses for each CVaR model, where N is the number of losses models.

βs: M×1 vector of significance levels of the gains for each CVaR model, where M is the number of gains models, each β_m ∈ (0, 1).

wβ: M×1 vector of weights of the gains for each CVaR model, where M is the number of gains models.

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_rtg — Function

owa_rtg(
    T::Integer;
    alpha_i::Real = 0.0001,
    alpha::Real = 0.05,
    a_sim::Integer = 100,
    beta_i::Union{<:Real, Nothing} = nothing,
    beta::Union{<:Real, Nothing} = nothing,
    b_sim::Union{<:Integer, Nothing} = nothing,
)

Compute the OWA weights for the Tail Gini Range of a returns series [OWA].

Inputs

T: number of observations of the returns series.

alpha_i: initial significance level of Tail Gini losses.

alpha: significance level of Tail Gini losses.

a_sim: number of CVaRs to approximate the Tail Gini losses.

beta_i: initial significance level of Tail Gini gains.

beta: significance level of Tail Gini gains.

b_sim: number of CVaRs to approximate the Tail Gini gains.

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_rg — Function

owa_rg(T::Integer)

Compute the OWA weights for the Range of a returns series [OWA].

Inputs

T: number of observations of the returns series.

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_l_moment — Function

owa_l_moment(T::Integer, k::Integer = 2)

Calculates the OWA weights of the k'th linear moment (l-moment) of a returns series [OWAL].

Inputs

T: number of observations of the returns series.

k: order of the l-moment.

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_l_moment_crm — Function

owa_l_moment_crm(
    T::Integer;
    k::Integer = 2,
    method::Symbol = :SD,
    g::Real = 0.5,
    max_phi::Real = 0.5,
    solvers = Dict(),
)

Compute the OWA weights for the convex risk measure considering higher order L-moments [OWAL].

Inputs

T: number of observations of the returns series.

k: order of the l-moment, k ≥ 2.

method: method for computing the weights used to combine L-moments higher than 2, used in OWAMethods.

:CRRA: Normalised Constant Relative Risk Aversion Coefficients.

:E: Maximum Entropy.

:SS: Minimum Sum of Squares.

:SD: Minimum Square Distance.

g: the risk aversion coefficient.

max_phi: maximum weight constraint of the L-moments.

solvers: Dict or NamedTuple with the keys:

:solver: which contains the JuMP optimiser.

:params: for the solver-specific parameters.

Outputs

w: T×1 ordered weight vector.

Private API

PortfolioOptimiser.OWAMethods — Constant

OWAMethods = (:CRRA, :E, :SS, :SD)

Methods for computing the weights used to combine L-moments higher than 2, used in owa_l_moment_crm.

:CRRA: Normalised Constant Relative Risk Aversion Coefficients.

:E: Maximum Entropy.

:SS: Minimum Sum of Squares.

:SD: Minimum Square Distance.

PortfolioOptimiser._optimize_owa — Function

_optimize_owa(model, solvers)

Internal function to optimise an OWA JuMP model.

Inputs

model: JuMP model.

solvers: Dict or NamedTuple with the keys:

:solver: which contains the JuMP optimiser.

:params: for the solver-specific parameters.

Outputs

term_status: JuMP termination status.

solvers_tried: Dictionary that contains a dictionary of failed optimisations. Dict(key => Dict(...)), where key is the solver key used for the iteration of solver that failed.

If an MOI call fails on a model:

Dict(:jump_error => jump_error): JuMP error code.

If the optimiser fails to optimise the model satisfactorily:

Dict(:objective_val => JuMP.objective_value(model), :term_status => term_status, :params => haskey(val, :params) ? val[:params] : missing), where val is the value of the dictionary corresponding to key.

PortfolioOptimiser._crra_method — Function

_crra_method(ws::AbstractMatrix, k::Integer, g::Real)

Internal function for computing the Normalized Constant Relative Risk Aversion coefficients.

Inputs

ws: T×(k-1) matrix where T is the number of observations and k the order of the L-moments to combine, the i'th column contains the weights for the (i+1)'th l-moment.

k: the maximum order of the L-moments.

g: the risk aversion coefficient.

Outputs

w: T×1 ordered weight vector of the combined L-moments.

OWACajas, Dany, OWA Portfolio Optimization: A Disciplined Convex Programming Framework (December 18, 2021). Available at SSRN: https://ssrn.com/abstract=3988927 or http://dx.doi.org/10.2139/ssrn.3988927

OWALCajas, Dany, Higher Order Moment Portfolio Optimization with L-Moments (March 19, 2023). Available at SSRN: https://ssrn.com/abstract=4393155 or http://dx.doi.org/10.2139/ssrn.4393155