Constraints

Public

PortfolioOptimiser.AKelly — Type

@kwdef mutable struct AKelly <: RetType
    formulation::SDSquaredFormulation = SOCSD()
end

PortfolioOptimiser.EKelly — Type

struct EKelly <: RetType end

PortfolioOptimiser.IP — Type

@kwdef struct IP{T1 <: AbstractMatrix{<:Real},
                 T2 <: Union{<:Integer, <:AbstractVector{<:Integer}}, T3 <: Real} <: AdjacencyConstraint
    A::T1 = Matrix{Float64}(undef, 0, 0)
    k::T2 = 1
    scale::T3 = 100_000.0
end

Defines the exact network constraint using mixed-integer programming.

\[\begin{align} \left(\mathbf{A} + \mathbf{I}\right)\bm{y} &\leq \bm{k}\\ w_{i} &\leq b_{u} y_{i} \quad \forall i = 1,\,\ldots,\,N \\ w_{i} &\geq b_{l} y_{i} \quad \forall i = 1,\,\ldots,\,N\,. \end{align}\]

Where:

$\mathbf{A}$ is the $N \times N$ adjacency matrix.

$\mathbf{I}$ is the identity matrix.

Since each row of $\left(\mathbf{A} + \mathbf{I}\right)$ corresponds to a path, duplicate rows add no new information whilst increasing the problem's size, therefore we only store unique rows.

$\bm{y}$ is an $N \times 1$ vector of binary $\{0,\,1\}$ decision variables, which decide whether or not the asset should be included in the portfolio.

$\bm{k}$:

if is a vector: $M\times 1$ vector defining the maximum number of assets allowed per unique path, where $M$ is the number of unique paths.

if is a scalar: defines the maximum number of assets allowed for all unique paths.

$w_{i}$ is the $i$-th asset weight.

$b_{u},\,b_{l}$ are the upper and lower bounds of the sum of the long and sum of the short asset weights, respectively.

Thus the constraint means we will invest in at most $\bm{k}$ assets per corresponding unique path.

This approach can be appied to any risk measure without work arounds like SDP. However it is more computationally costly to optimise, and may fail when $\mathbf{A}$ is close to the all ones matrix.

Parameters

A: adjacency matrix, only stores unique(A + I, dims = 1).

k:

if is a vector: maximum number of assets per unique path.

if A is not empty, checks that the length of k is equal to the size of unique(A + I, dims = 1).

if is a scalar: maximum number of assets for all unique paths.

scale: scaling variable for an auxiliary binary decision variable when optimising the Sharpe objective function.

PortfolioOptimiser.NoAdj — Type

struct NoAdj <: AdjacencyConstraint end

PortfolioOptimiser.NoKelly — Type

struct NoKelly <: RetType end

PortfolioOptimiser.NoTR — Type

struct NoTR <: AbstractTR end

PortfolioOptimiser.NoTracking — Type

struct NoTracking <: TrackingErr end

PortfolioOptimiser.SDP — Type

@kwdef mutable struct SDP{T1 <: AbstractMatrix{<:Real}, T2 <: Real} <: AdjacencyConstraint
    A::T1 = Matrix{Float64}(undef, 0, 0)
    penalty::T2 = 0.05
end

Defines an approximate network constraint using semi-definite programming.

\[\begin{align} \begin{bmatrix} \mathbf{W} & \bm{w}\\ \bm{w}^{\intercal} & 1 \end{bmatrix} &\succeq 0\\ \mathbf{W} &= \mathbf{W}^{\intercal}\\ \mathbf{A} \odot \mathbf{W} &= \bm{0}\,. \end{align}\]

Where:

$\mathbf{W}$ is an auxiliary variable that approximates the outer product of asset weights $\bm{w} \otimes \bm{w}$.

$\mathbf{A}$ is the $N\times N$ adjacency matrix. It tells us which assets are connected. The matrix can only take values of 1 or 0. If entry $(i,\,j)$ is equal to 1, assets $i$ and $j$ are connected.

$\odot$ is the Hadamard (element-wise) product.

When the variance risk measure SD is being used, whether in the objective function or as one of the risk constraints. Its definition will change when this constraint is active. The new definition is this.

\[\begin{align} \phi_{\mathrm{var}}(\bm{w}) &= \mathrm{Tr}\left(\mathbf{\Sigma}\mathbf{W}\right)\,. \end{align}\]

Where:

$\mathrm{Tr}(\cdot)$ is the trace operator.

$\mathbf{\Sigma}$ is the covariance matrix.

However, this will not work if the variance is not being constrained, or if it is not in the objective function. For that we add/subtract the following penalty factor to the objective function.

\[\begin{align} \underset{\bm{w}}{\mathrm{opt}} &\quad \phi(\bm{w}) \pm \lambda \mathrm{Tr}\left(\mathbf{X}\right)\,. \end{align}\]

Where:

$\mathrm{opt}$ is $\mathrm{min}$ when the objective is convex and $\mathrm{max}$ when it is concave.

$\pm$ is $+$ when the objective is convex and $-$ when it is concave.

$\lambda$ is a penalty factor.

This approach works better than IP when $\mathbf{A}$ is close to the all ones matrix, even though it's an approximation.

Parameters

A: N×N adjacency matrix.

penalty: penalty factor when the variance SD risk measure isn't being used, either in a constraint or in the objective function.

PortfolioOptimiser.TR — Type

@kwdef mutable struct TR{T1 <: Union{<:Real, <:AbstractVector{<:Real}},
                         T2 <: AbstractVector{<:Real}} <: AbstractTR
    val::T1 = 0.0
    w::T2 = Vector{Float64}(undef, 0)
end

PortfolioOptimiser.TrackRet — Type

@kwdef mutable struct TrackRet{T1 <: Real, T2 <: AbstractVector{<:Real}} <: TrackingErr
    err::T1 = 0.0
    w::T2 = Vector{Float64}(undef, 0)
end

PortfolioOptimiser.TrackWeight — Type

@kwdef mutable struct TrackWeight{T1 <: Real, T2 <: AbstractVector{<:Real}} <: TrackingErr
    err::T1 = 0.0
    w::T2 = Vector{Float64}(undef, 0)
end

Private

PortfolioOptimiser.AbstractTR — Type

abstract type AbstractTR end

PortfolioOptimiser.AdjacencyConstraint — Type

abstract type AdjacencyConstraint end

PortfolioOptimiser.RetType — Type

abstract type RetType end

PortfolioOptimiser.TrackingErr — Type

abstract type TrackingErr end