Optimisation kinds

Public

PortfolioOptimiser.BasicRRB — Type

struct BasicRRB <: RRBVersion end

Basic relaxed risk budget optimisation version.

See also: RRBVersion, RRB.

PortfolioOptimiser.HERC — Type

struct HERC <: HCOptimType end

PortfolioOptimiser.HRP — Type

struct HRP <: HCOptimType end

PortfolioOptimiser.NCO — Type

mutable struct NCO <: HCOptimType
    internal::NCOArgs
    external::NCOArgs
    finaliser::HCOptWeightFinaliser
end

PortfolioOptimiser.NOC — Type

mutable struct NOC{T1, T2, T3, T4, T5, T6, T7, T8} <: OptimType

Near optimal centering optimisation type. This type of optimisation defines a near-optimal convex region around a point in the efficient frontier, and finds the portfolio which best fits the analytic centre of the region.

See also: OptimType, RiskMeasure, RetType, PortClass, CustomConstraint, CustomObjective, AbstractScalarisation.

Keyword Parameters

flag::Bool = true:

bins::T1 = 20.0 where T1 <: Real:

w_opt::T2 = Vector{Float64}(undef, 0) where T2 <: AbstractVector{<:Real}: Vector of weights of the efficient frontier portfolio.

w_min::T3 = Vector{Float64}(undef, 0) where T3 <: AbstractVector{<:Real}: Vector of weights of the minimal risk portfolio.

w_max::T4 = Vector{Float64}(undef, 0) where T4 <: AbstractVector{<:Real}: Vector of weights of the maxumal return portfolio.

w_opt_ini::T6 = Vector{Float64}(undef, 0) where T6 <: AbstractVector{<:Real}: The initial weights of the efficient frontier portfolio optimisation.

Irrelevant if the solver does not support them.

w_min_ini::T6 = Vector{Float64}(undef, 0) where T6 <: AbstractVector{<:Real}: The initial weights of the minimum risk optimisation.

Irrelevant if the solver does not support them.

w_max_ini::T7 = Vector{Float64}(undef, 0) where T7 <: AbstractVector{<:Real}: The initial weights of the maximum return optimisation.

Irrelevant if the solver does not support them.

rm::Union{AbstractVector, <:RiskMeasure} = Variance(): The risk measure(s) to be used.

If multiple instances of the same risk measure are used, they must be grouped in a single vector wrapped in another vector, see examples.

obj::ObjectiveFunction = MinRisk():

kelly::RetType = NoKelly():

class::PortClass = Classic():

w_ini::T8 = Vector{Float64}(undef, 0) where T8 <: AbstractVector{<:Real}: The initial weights for the optimisation of the near optimal centering portfolio.

Irrelevant if the solver does not support them.

custom_constr::CustomConstraint = NoCustomConstraint(): Add custom constraints to the optimisation problem.

custom_obj::CustomObjective = NoCustomObjective(): Add custom terms to the objective function.

scalarisation::AbstractScalarisation = ScalarSum(): The scalarisation function to be used.

Only relevant when multiple risk measures are used.

str_names::Bool = false: Whether to use string names in the JuMP model.

PortfolioOptimiser.RB — Type

mutable struct RB{T1} <: OptimType

Risk budget optimisation type. Allows the user to specify a risk budget vector specifying the maximum risk contribution per asset or factor. The asset weights are then optimised to meet the risk budget per asset/factor as optimally as possible.

See also: OptimType, RiskMeasure, RetType, PortClass, CustomConstraint, CustomObjective, AbstractScalarisation.

Keyword Parameters

rm::Union{AbstractVector, <:RiskMeasure} = Variance(): The risk measure(s) to be used.

If multiple instances of the same risk measure are used, they must be grouped in a single vector wrapped in another vector, see examples.

kelly::RetType = NoKelly(): The Kelly criterion to be used.

class::PortClass = Classic(): The portfolio class to be used.

w_ini::T1 = Vector{Float64}(undef, 0) where T1 <: AbstractVector: The initial weights for the optimisation.

Irrelevant if the solver does not support them.

custom_constr::CustomConstraint = NoCustomConstraint(): Add custom constraints to the optimisation problem.

custom_obj::CustomObjective = NoCustomObjective(): Add custom terms to the objective function.

scalarisation::AbstractScalarisation = ScalarSum(): The scalarisation function to be used.

Only relevant when multiple risk measures are used.

str_names::Bool = false: Whether to use string names in the JuMP model.

Examples

# Default constructor.
opt_type = RB()

# Single risk measure.
opt_type = RB(; rm = SD())

# Multiple risk measures.
opt_type = RB(; rm = [Variance(), CVaR(; alpha = 0.15)])

# Incorrect use of multiple risk measures of the same type.
# This will produce a JuMP registration error when optimise! is called.
opt_type = RB(; rm = [CVaR(), CVaR(; alpha = 0.2)])

# Correct use of multiple risk measures of the same type.
opt_type = RB(; rm = [[CVaR(), CVaR(; alpha = 0.2)]])

# Incorrect use of multiple risk measures, some of the same type.
# This will produce a JuMP registration error regarding the CVaR
# risk measure when optimise! is called.
opt_type = RB(; rm = [MAD(), CVaR(), CVaR(; alpha = 0.2)])

# Correct use of multiple risk measures, some of the same type.
opt_type = RB(; rm = [MAD(), [CVaR(), CVaR(; alpha = 0.2)]])

PortfolioOptimiser.RRB — Type

mutable struct RRB{T1} <: OptimType

The relaxed risk budget optimisation only applies to the variance risk measure.

See also: OptimType, RRBVersion, RetType, PortClass, CustomConstraint, CustomObjective, AbstractScalarisation.

Keyword Parameters

version::RRBVersion = BasicRRB(): Relaxed risk budget optimisation version.

kelly::RetType = NoKelly(): The Kelly criterion to be used.

class::PortClass = Classic(): The portfolio class to be used.

w_ini::T1 = Vector{Float64}(undef, 0) where T1 <: AbstractVector: The initial weights for the optimisation.

Irrelevant if the solver does not support them.

custom_constr::CustomConstraint = NoCustomConstraint(): Add custom constraints to the optimisation problem.

custom_obj::CustomObjective = NoCustomObjective(): Add custom terms to the objective function.

str_names::Bool = false: Whether to use string names in the JuMP model.

PortfolioOptimiser.RegPenRRB — Type

mutable struct RegPenRRB{T1} <: RRBVersion

Relaxed risk budget optimisation version with regularisation and penalty.

See also: RRBVersion, RRB.

Keyword Parameters

penalty::Real = 1.0: The penalty to be used.

Examples

# Default constructor.
rrb_ver = RegPenRRB()

# Custom penalty.
rrb_ver = RegPenRRB(; penalty = 0.5)

PortfolioOptimiser.RegRRB — Type

struct RegRRB <: RRBVersion end

Relaxed risk budget optimisation version with regularisation.

See also: RRBVersion, RRB.

PortfolioOptimiser.ScalarLogSumExp — Type

mutable struct ScalarLogSumExp{T1 <: Real} <: AbstractScalarisation end

Scalarises the risk measures as the logsumexp of the weighted risk measures.

See also: AbstractScalarisation.

\[\begin{align} r &= \frac{1}{\gamma} \log \left( \sum_{i = 1}^{N} \exp(\gamma r_i w_i) \right) \end{align}\]

Where:

$r$ is the scalarised risk.

$r_i$ is the $i$-th risk measure.

$w_i$ is the weight of the $i$-th risk measure.

$\gamma$ is a parameter that controls the shape of the scalarisation.

Keyword Parameters

gamma::Real = 1.0: gamma > 0. As gamma → 0, the scalarisation approaches ScalarSum. As gamma → Inf, the scalarisation approaches ScalarMax.

Examples

# Default constructor.
scalariser = ScalarLogSumExp()

# Approximate ScalarSum()
scalariser = ScalarLogSumExp(; gamma = 1e-6)

# Approximate ScalarMax()
scalariser = ScalarLogSumExp(; gamma = 1e6)

PortfolioOptimiser.ScalarMax — Type

struct ScalarMax <: AbstractScalarisation end

Scalarises the risk measures by taking the maximum of them.

See also: AbstractScalarisation.

\[\begin{align} r &= \max \left( \bm{r} \odot \bm{w} \right) \end{align}\]

Where:

$r$ is the scalarised risk.

$\bm{r}$ is the vector of risk measures.

$\bm{w}$ is the corresponding vector of risk measure weights.

$\odot$ is the Hadamard (element-wise) multiplication.

Examples

scalariser = ScalarMax()

PortfolioOptimiser.ScalarSum — Type

struct ScalarSum <: AbstractScalarisation end

Scalarises the risk measures as a weighted sum.

See also: AbstractScalarisation.

\[\begin{align} r &= \bm{r} \cdot \bm{w} \end{align}\]

Where:

$r$ is the scalarised risk.

$\bm{r}$ is the vector of risk measures.

$\bm{w}$ is the corresponding vector of risk measure weights

$\cdot$ is the dot product.

Examples

scalariser = ScalarSum()

PortfolioOptimiser.SchurHRP — Type

mutable struct SchurHRP <: HCOptimType

PortfolioOptimiser.Trad — Type

mutable struct Trad{T1} <: OptimType

Traditional optimisation type.

See also: OptimType, RiskMeasure, ObjectiveFunction, RetType, PortClass, CustomConstraint, CustomObjective, AbstractScalarisation.

Keyword Parameters

rm::Union{AbstractVector, <:RiskMeasure} = Variance(): The risk measure(s) to be used.

If multiple instances of the same risk measure are used, they must be grouped in a single vector wrapped in another vector, see examples.

obj::ObjectiveFunction = MinRisk(): The objective function to be used.

kelly::RetType = NoKelly(): The Kelly criterion to be used.

class::PortClass = Classic(): The portfolio class to be used.

w_ini::T1 = Vector{Float64}(undef, 0) where T1 <: AbstractVector: The initial weights for the optimisation.

Irrelevant if the solver does not support them.

custom_constr::CustomConstraint = NoCustomConstraint(): Add custom constraints to the optimisation problem.

custom_obj::CustomObjective = NoCustomObjective(): Add custom terms to the objective function.

scalarisation::AbstractScalarisation = ScalarSum(): The scalarisation function to be used.

Only relevant when multiple risk measures are used.

str_names::Bool = false: Whether to use string names in the JuMP model.

Examples

# Default constructor.
opt_type = Trad()

# Single risk measure.
opt_type = Trad(; rm = SD())

# Multiple risk measures.
opt_type = Trad(; rm = [Variance(), CVaR(; alpha = 0.15)])

# Incorrect use of multiple risk measures of the same type.
# This will produce a JuMP registration error when optimise! is called.
opt_type = Trad(; rm = [CVaR(), CVaR(; alpha = 0.2)])

# Correct use of multiple risk measures of the same type.
opt_type = Trad(; rm = [[CVaR(), CVaR(; alpha = 0.2)]])

# Incorrect use of multiple risk measures, some of the same type.
# This will produce a JuMP registration error regarding the CVaR
# risk measure when optimise! is called.
opt_type = Trad(; rm = [MAD(), CVaR(), CVaR(; alpha = 0.2)])

# Correct use of multiple risk measures, some of the same type.
opt_type = Trad(; rm = [MAD(), [CVaR(), CVaR(; alpha = 0.2)]])

PortfolioOptimiser.efficient_frontier! — Function

efficient_frontier!(port::Portfolio, type::Union{Trad, NOC, NCO} = Trad();
                             w_min_ini::AbstractVector = Vector{Float64}(undef, 0),
                             w_max_ini::AbstractVector = Vector{Float64}(undef, 0),
                             points::Integer = 20, rf::Real = 0.0)

PortfolioOptimiser.optimise! — Method

optimise!(port::Portfolio, type::OptimType)

Private

PortfolioOptimiser.AbstractOptimType — Type

abstract type AbstractOptimType end

Abstract type for the different types of optimisations.

See also: OptimType and HCOptimType.

PortfolioOptimiser.AbstractScalarisation — Type

abstract type AbstractScalarisation end

Abstract type for scalarisation functions used when simultaneously optimising for multiple risk measures.

See also: ScalarSum, ScalarMax, ScalarLogSumExp.

PortfolioOptimiser.HCOptimType — Type

abstract type HCOptimType <: AbstractOptimType end

Abstract type for hierarchical optimisations.

See also: AbstractOptimType, OptimType, HRP, SchurHRP, HERC, NCO.

PortfolioOptimiser.OptimType — Type

abstract type OptimType <: AbstractOptimType end

Abstract type for optimisations that are not hierarchical.

See also: AbstractOptimType, HCOptimType, Trad, RB, RRB, NOC.

PortfolioOptimiser.RRBVersion — Type

abstract type RRBVersion end

Abstract type for the different versions of the relaxed risk budget optimisation.

See also: RRB, BasicRRB, RegRRB, RegPenRRB.