Parameter estimation

Public

PortfolioOptimiser.NoPosdef — Type

struct NoPosdef <: PosdefFix end

Non positive definite matrices will not be fixed in in posdef_fix!.

PortfolioOptimiser.PosdefNearest — Type

@kwdef mutable struct PosdefNearest <: PosdefFix
    method::NearestCorrelationMatrix.NCMAlgorithm = NearestCorrelationMatrix.Newton(;
                                                                                    tau = 1e-12)
end

Defines which method from NearestCorrelationMatrix to use in posdef_fix!.

PortfolioOptimiser.DenoiseFixed — Type

@kwdef mutable struct DenoiseFixed{T1, T2, T3, T4} <: Denoise
    detone::Bool = false
    mkt_comp::Integer = 1
    kernel = AverageShiftedHistograms.Kernels.gaussian
    m::Integer = 10
    n::Integer = 1000
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the parameters for using the fixed method in denoise! [1, Chapter 2].

Parameters

detone:

true: remove the largest mkt_comp eigenvalues from the correlation matrix.

Warning

Removing eigenvalues from the matrix may make it singular.

mkt_comp: the number of largest eigenvalues to remove from the correlation matrix.

kernel: kernel for fitting the average shifted histograms from AverageShiftedHistograms.jl Kernel Functions.

m: number of adjacent histograms to smooth over AverageShiftedHistograms.ash.

n: number of points used when creating the range of values to which the average shifted histogram is to be fitted AverageShiftedHistograms.ash.

args: arguments for Optim.optimize

kwargs: keyword arguments for Optim.optimize

PortfolioOptimiser.DenoiseShrink — Type

@kwdef mutable struct DenoiseShrink{T1, T2, T3, T4, T5} <: Denoise
    alpha::Real = 0.0
    detone::Bool = false
    mkt_comp::Integer = 1
    kernel = AverageShiftedHistograms.Kernels.gaussian
    m::Integer = 10
    n::Integer = 1000
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the parameters for using the shrink method in denoise! [1, Chapter 2].

Parameters

alpha: tuning parameter for how much the matrix should be shrunk, alpha ∈ [0, 1].

detone:

true: take only the largest mkt_comp eigenvalues from the correlation matrix.

Warning

Removing eigenvalues from the matrix may make it singular.

mkt_comp: the number of largest eigenvalues to keep from the correlation matrix.

kernel: kernel for fitting the average shifted histograms from AverageShiftedHistograms.jl Kernel Functions.

m: number of adjacent histograms to smooth over AverageShiftedHistograms.ash.

n: number of points used when creating the range of values to which the average shifted histogram is to be fitted AverageShiftedHistograms.ash.

args: arguments for Optim.optimize

kwargs: keyword arguments for Optim.optimize

PortfolioOptimiser.DenoiseSpectral — Type

@kwdef mutable struct DenoiseSpectral{T1, T2, T3, T4} <: Denoise
    detone::Bool = false
    mkt_comp::Integer = 1
    kernel = AverageShiftedHistograms.Kernels.gaussian
    m::Integer = 10
    n::Integer = 1000
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the parameters for using the spectral method in denoise! [1, Chapter 2].

Parameters

detone:

true: take only the largest mkt_comp eigenvalues from the correlation matrix.

Warning

Removing eigenvalues from the matrix may make it singular.

mkt_comp: the number of largest eigenvalues to keep from the correlation matrix.

kernel: kernel for fitting the average shifted histograms from AverageShiftedHistograms.jl Kernel Functions.

m: number of adjacent histograms to smooth over AverageShiftedHistograms.ash.

n: number of points used when creating the range of values to which the average shifted histogram is to be fitted AverageShiftedHistograms.ash.

args: arguments for Optim.optimize

kwargs: keyword arguments for Optim.optimize

PortfolioOptimiser.NoDenoise — Type

struct NoDenoise <: Denoise end

No denoising is performed in denoise!.

PortfolioOptimiser.DistanceCanonical — Type

struct DistanceCanonical <: DistanceMethod end

Struct for computing the canonical distance for a given correlation subtype of PortfolioOptimiserCovCor.

PortfolioOptimiser.DistanceLog — Type

struct DistanceLog <: DistanceMethod end

Defines the log-distance matrix from the correlation matrix.

\[D_{i,\,j} = -\log\left(C_{i,\,j}\right)\,.\]

Where:

$D_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N log-distance matrix.

$C_{i,\,j}$: is the $(i,\,j)$-th entry of an absolute correlation matrix.

PortfolioOptimiser.DistanceMLP — Type

@kwdef mutable struct DistanceMLP <: DistanceMethod
    absolute::Bool = false
end

Defines the distance matrix from a correlation matrix [2] in dist.

\[D_{i,\,j} = \begin{cases} \sqrt{\dfrac{1}{2} \left(\mathbf{1} - C_{i,\,j}\right)} &\quad \mathrm{if~ absolute = false}\\ \sqrt{1 - \lvert C_{i,\,j} \rvert} &\quad \mathrm{if~ absolute = true}\nonumber\,. \end{cases}\]

Where:

$D_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N distance matrix $\mathbf{C}$.

$C_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N correlation matrix $\mathbf{D}$.

absolute: flag for stating whether or not an absolute correlation is being used.

Parameters

absolute: flag for stating whether or not an absolute correlation is being used.

PortfolioOptimiser.DistanceSqMLP — Type

@kwdef mutable struct DistanceSqMLP <: DistanceMethod
    absolute::Bool = false
    distance::Distances.UnionMetric = Distances.Euclidean()
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the distance of distances matrix from a correlation matrix [2] in dist.

\[\tilde{D}_{i,\,j} = f_{m}\left(\bm{D}_{i},\, \bm{D}_j\right)\,.\]

Where:

$\bm{D}_{i}$: is the $i$-th column/row of the N×N distance matrix defined in DistanceMLP.

$f_{m}$: is the pairwise distance function for metric $m$. We use the Distances.pairwise function which computes the entire matrix at once, the output is a vector so it gets reshaped into an N×N matrix.

$\tilde{D}_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N distances of distances matrix.

absolute: is a flag whether the correlation is absolute.

Parameters

absolute: flag for stating whether or not an absolute correlation is being used.

distance: distance metric from Distances.jl.

args: args for the Distances.pairwise function.

kwargs: key word args for the Distances.pairwise function.

PortfolioOptimiser.DistanceVarInfo — Type

@kwdef mutable struct DistanceVarInfo <: DistanceMethod
    bins::Union{<:Integer, <:AbstractBins} = HGR()
    normalise::Bool = true
end

Defines the variation of information distance matrix.

Parameters

bins: defines the bin function, or bin width directly and if so bins > 0.

normalise: whether or not to normalise the variation of information.

PortfolioOptimiser.Freedman — Type

struct Freedman <: AstroBins end

Freedman's bin width algorithm from astropy.

PortfolioOptimiser.HGR — Type

struct HGR <: AbstractBins end

Hacine-Gharbi and Ravier's bin width algorithm [3].

PortfolioOptimiser.Knuth — Type

struct Knuth <: AstroBins end

Knuth's bin width algorithm from astropy.

PortfolioOptimiser.Scott — Type

struct Scott <: AstroBins end

Scott's bin width algorithm from astropy.

PortfolioOptimiser.ClusterNode — Type

struct ClusterNode{tid, tl, tr, td, tcnt}
    id::tid
    left::tl
    right::tr
    height::td
    level::tcnt

    function ClusterNode(id, left::Union{ClusterNode, Nothing} = nothing,
                         right::Union{ClusterNode, Nothing} = nothing, height::Real = 0.0,
                         level::Int = 1)
        ilevel = isnothing(left) ? level : (left.level + right.level)

        return new{typeof(id), typeof(left), typeof(right), typeof(height), typeof(level)}(id,
                                                                                         left,
                                                                                         right,
                                                                                         height,
                                                                                         ilevel)
    end
end

Structure for definining a cluster node. This is used for turning a clustering result into a tree, is_leaf, pre_order, to_tree.

Parameters

id: node ID.

left: node to the left.

right: node to the right.

height: node height.

level: node level.

PortfolioOptimiser.DBHT — Type

mutable struct DBHT <: HClustAlg
    distance::DistanceMethod
    similarity::DBHTSimilarity
    root_method::DBHTRootMethod
end

Defines the parameters for computing DBHTs [4].

PortfolioOptimiser.DBHTExp — Type

struct DBHTExp <: DBHTSimilarity end

Defines the similarity matrix for use in PMFG_T2s as the element-wise exponential decay of the dissimilarity matrix in dbht_similarity.

PortfolioOptimiser.DBHTMaxDist — Type

struct DBHTMaxDist <: DBHTSimilarity end

Defines the similarity matrix for use in PMFG_T2s as the element-wise distance from the maximum value of the dissimilarity matrix dbht_similarity.

PortfolioOptimiser.EqualDBHT — Type

struct EqualDBHT <: DBHTRootMethod end

Create a clique's root from its adjacency tree in CliqueRoot.

PortfolioOptimiser.HAC — Type

@kwdef mutable struct HAC <: HClustAlg
    linkage::Symbol = :ward
end

Use a hierarchical clustering algorithm from Clustering.jl.

Parameters

linkage: linkage type supported by hclust.

PortfolioOptimiser.HCOpt — Type

@kwdef mutable struct HCOpt{T1 <: Integer, T2 <: Integer}
    branchorder::Symbol = :optimal
    k_method::NumClusterMethod = TwoDiff()
    k::T1 = 0
    max_k::T2 = 0
end

Defines the options for processing clustering results in an instance of Clustering.Hclust.

Parameters

branchorder: parameter for ordering a dendrogram's branches accepted by Clustering.jl.

k_method: method subtyping NumClusterMethod for computing the number of clusters.

k: directly provide the number of clusters, if 0 use k_method for computing the number of clusters.

max_k: maximum number of clusters, if 0 defaults to ⌈sqrt(N)⌉, where N is the number of assets.

PortfolioOptimiser.StdSilhouette — Type

@kwdef mutable struct StdSilhouette <: NumClusterMethod
    metric::Union{Distances.SemiMetric, Nothing} = nothing
end

Use the standardised silhouette score for computing the number of clusters in calc_k_clusters.

PortfolioOptimiser.TwoDiff — Type

struct TwoDiff <: NumClusterMethod end

Use the two difference gap statistic for computing the number of clusters in calc_k_clusters.

PortfolioOptimiser.UniqueDBHT — Type

struct UniqueDBHT <: DBHTRootMethod end

Create a unique root for a clique in CliqueRoot [5].

PortfolioOptimiser.CorDistance — Type

mutable struct CorDistance <: PortfolioOptimiserCovCor
    distance::Distances.UnionMetric
    dist_args::Tuple
    dist_kwargs::NamedTuple
    mean_w1::Union{<:AbstractWeights, Nothing}
    mean_w2::Union{<:AbstractWeights, Nothing}
    mean_w3::Union{<:AbstractWeights, Nothing}
end

PortfolioOptimiser.CorGerberSB1 — Type

mutable struct CorGerberSB1{T1, T2, T3, T4, T5} <: CorSB
    normalise::Bool
    threshold::T1
    c1::T2
    c2::T3
    c3::T4
    n::T5
    ve::StatsBase.CovarianceEstimator
    std_w::Union{<:AbstractWeights, Nothing}
    mean_w::Union{<:AbstractWeights, Nothing}
    posdef::PosdefFix
end

PortfolioOptimiser.CorKendall — Type

@kwdef mutable struct CorKendall <: CorRank
    absolute::Bool = false
end

Kendall type correlation estimator.

Parameters

absolute: whether or not to compute an absolute correlation, abs.(corkendall(X)).

PortfolioOptimiser.CorLTD — Type

mutable struct CorLTD <: PortfolioOptimiserCovCor
    alpha::Real
    ve::StatsBase.CovarianceEstimator
    w::Union{<:AbstractWeights, Nothing}
end

PortfolioOptimiser.CorMutualInfo — Type

@kwdef mutable struct CorMutualInfo <: PortfolioOptimiserCovCor
    bins::Union{<:Integer, <:AbstractBins} = HGR()
    normalise::Bool = true
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    w::Union{<:AbstractWeights, Nothing} = nothing
end

Define the mutual information correlation matrix.

Parameters

bins: defines the bin function, or bin width directly and if so bins > 0.

normalise: whether or not to normalise the mutual information.

ve: variance estimator.

w: variance weights, if nothing

PortfolioOptimiser.CorSpearman — Type

@kwdef mutable struct CorSpearman <: CorRank
    absolute::Bool = false
end

Spearman type correlation estimator.

Parameters

absolute: whether or not to compute an absolute correlation, abs.(corspearman(X)).

PortfolioOptimiser.CovFull — Type

@kwdef mutable struct CovFull <: CorPearson
    absolute::Bool = false
    ce::StatsBase.CovarianceEstimator = StatsBase.SimpleCovariance(; corrected = true)
    w::Union{<:AbstractWeights, Nothing} = nothing
end

Full Pearson-type covariance and correlation estimator.

Parameters

absolute: whether or not to compute an absolute correlation.

ce: covariance estimator.

w: weights for computing the covariance, if nothing apply no weights.

PortfolioOptimiser.CovSemi — Type

@kwdef mutable struct CovSemi <: CorPearson
    absolute::Bool = false
    ce::StatsBase.CovarianceEstimator = StatsBase.SimpleCovariance(; corrected = true)
    target::Union{<:Real, AbstractVector{<:Real}} = 0.0
    w::Union{<:AbstractWeights, Nothing} = nothing
end

Semi Pearson-type covariance and correlation estimator.

Parameters

absolute: whether or not to compute an absolute correlation, abs.(cor(X)).

ce: covariance estimator.

target: minimum acceptable return target.

if isa(target, Real): apply the same target to all assets.

if isa(target, AbstractVector): apply individual target to each asset.

w: weights for computing the covariance, if nothing apply no weights.

PortfolioOptimiser.LoGo — Type

@kwdef mutable struct LoGo <: AbstractLoGo
    distance::DistanceMethod = DistanceMLP()
    similarity::DBHTSimilarity = DBHTMaxDist()
end

PortfolioOptimiser.PortCovCor — Type

@kwdef mutable struct PortCovCor <: PortfolioOptimiserCovCor
    ce::CovarianceEstimator = CovFull(;)
    posdef::PosdefFix = PosdefNearest(;)
    denoise::Denoise = NoDenoise(;)
    logo::AbstractLoGo = NoLoGo(;)
end

PortfolioOptimiser.SimpleVariance — Type

@kwdef mutable struct SimpleVariance <: StatsBase.CovarianceEstimator
    corrected::Bool = true
end

Simple variance estimator.

Parameters

corrected: if true correct the bias by dividing by N-1, if false the bias is not corrected and the division is by N.

PortfolioOptimiser.GM — Type

struct GM <: MeanTarget end

PortfolioOptimiser.MuBOP — Type

@kwdef mutable struct MuBOP{T1} <: MeanEstimator
    target::MeanTarget = GM()
    w::Union{<:AbstractWeights, Nothing} = nothing
    sigma::T1 = Matrix{Float64}(undef, 0, 0)
end

PortfolioOptimiser.MuBS — Type

@kwdef mutable struct MuBS{T1} <: MeanEstimator
    target::MeanTarget = GM()
    w::Union{<:AbstractWeights, Nothing} = nothing
    sigma::T1 = Matrix{Float64}(undef, 0, 0)
end

PortfolioOptimiser.MuJS — Type

@kwdef mutable struct MuJS{T1} <: MeanEstimator
    target::MeanTarget = GM()
    w::Union{<:AbstractWeights, Nothing} = nothing
    sigma::T1 = Matrix{Float64}(undef, 0, 0)
end

PortfolioOptimiser.MuSimple — Type

@kwdef mutable struct MuSimple <: MeanEstimator
    w::Union{<:AbstractWeights, Nothing} = nothing
end

PortfolioOptimiser.SE — Type

struct SE <: MeanTarget end

PortfolioOptimiser.VW — Type

struct VW <: MeanTarget end

PortfolioOptimiser.posdef_fix! — Method

posdef_fix!(method::PosdefFix, X::AbstractMatrix)

PortfolioOptimiser.denoise! — Method

denoise!(ce::Denoise, posdef::PosdefFix, X::AbstractMatrix, q::Real)

PortfolioOptimiser.dist — Method

dist(de::DistanceMethod, X, Y)

PortfolioOptimiser.calc_k_clusters — Method

calc_k_clusters(hclust_opt::HCOpt, dist::AbstractMatrix, clustering)

PortfolioOptimiser.dbht_similarity — Method

dbht_similarity(::DBHTExp, S, D)

PortfolioOptimiser.is_leaf — Method

is_leaf(a::ClusterNode)

PortfolioOptimiser.pre_order — Function

pre_order(a::ClusterNode, func::Function = x -> x.id)

PortfolioOptimiser.to_tree — Method

to_tree(a::Hclust)

PortfolioOptimiser.coskew — Method

coskew(::SkewFull, X::AbstractMatrix, mu::AbstractVector)

PortfolioOptimiser.coskew — Method

coskew(se::SkewSemi, X::AbstractMatrix, mu::AbstractVector)

PortfolioOptimiser.target_mean — Method

target_mean(::GM, mu::AbstractVector, sigma::AbstractMatrix, inv_sigma, T::Integer,
                     N::Integer)

PortfolioOptimiser.black_litterman — Method

black_litterman(bl::BLType, X::AbstractMatrix, P::AbstractMatrix,
                         Q::AbstractVector, w::AbstractVector;
                         cov_type::PortfolioOptimiserCovCor = PortCovCor(),
                         mu_type::MeanEstimator = MuSimple())

PortfolioOptimiser.owa_cvar — Function

owa_cvar(T::Integer; alpha::Real = 0.05)

Calculate the OWA weights corresponding to the Critical Value at Risk (CVaR) of a returns series [OWA].

Inputs

Outputs

PortfolioOptimiser.owa_gmd — Method

owa_gmd(T::Integer)

Computes the Gini Mean Difference (GMD) of a returns series [OWA].

Inputs

Outputs

PortfolioOptimiser.owa_l_moment — Function

owa_l_moment(T::Integer; k::Integer = 2)

Calculates the OWA weights of the k'th linear moment (L-moment) of a returns series [6].

Inputs

k: order of the L-moment.

Outputs

PortfolioOptimiser.owa_l_moment_crm — Method

owa_l_moment_crm(T::Integer; k::Integer = 2, method::Symbol = :SD, g::Real = 0.5,
                 max_phi::Real = 0.5, solvers = Dict())

Compute the OWA weights for the convex risk measure considering higher order L-moments [6].

Inputs

k: order of the L-moment, k ≥ 2.

method: method for computing the weights used to combine L-moments higher than 2, used in OWAMethods.

:CRRA: Normalised Constant Relative Risk Aversion Coefficients.

:E: Maximum Entropy. Solver must support MOI.RelativeEntropyCone and MOI.NormOneCone.

:SS: Minimum Sum of Squares. Solver must support MOI.SecondOrderCone.

:SD: Minimum Square Distance. Solver must support MOI.SecondOrderCone.

g: the risk aversion coefficient.

max_phi: maximum weight constraint of the L-moments.

Outputs

PortfolioOptimiser.owa_rcvar — Method

owa_rcvar(T::Integer; alpha::Real = 0.05, beta::Real = alpha)

Compute the OWA weights for the CVaR Range of a returns series [OWA].

Inputs

Outputs

PortfolioOptimiser.owa_rg — Method

owa_rg(T::Integer)

Compute the OWA weights for the Range of a returns series [OWA].

Inputs

Outputs

PortfolioOptimiser.owa_rtg — Method

owa_rtg(T::Integer; alpha_i::Real = 0.0001, alpha::Real = 0.05, a_sim::Integer = 100,
        beta_i::Real = alpha_i, beta::Real = alpha, b_sim::Integer = a_sim)

Compute the OWA weights for the Tail Gini Range of a returns series [OWA].

Inputs

Outputs

PortfolioOptimiser.owa_rwcvar — Function

owa_rwcvar(T::Integer, alphas::AbstractVector{<:Real}, weights_a::AbstractVector{<:Real};
           betas::AbstractVector{<:Real} = alphas,
           weights_b::AbstractVector{<:Real} = weights_b)

Compute the OWA weights for the Weighted Conditional Value at Risk (WCVaR) of a returns series [OWA].

Inputs

alphas: N×1 vector of significance levels of the losses for each CVaR model, where N is the number of losses models, each .

weights_a: N×1 vector of weights of the losses for each CVaR model, where N is the number of losses models.

betas: M×1 vector of significance levels of the gains for each CVaR model, where M is the number of gains models, each .

weights_b: M×1 vector of weights of the gains for each CVaR model, where M is the number of gains models.

Outputs

PortfolioOptimiser.owa_tg — Method

owa_tg(T::Integer; alpha_i::Real = 0.0001, alpha::Real = 0.05, a_sim::Integer = 100)

Compute the OWA weights for the Tail Gini of a returns series [OWA].

Inputs

Outputs

PortfolioOptimiser.owa_wcvar — Method

owa_wcvar(T::Integer, alphas::AbstractVector{<:Real}, weights::AbstractVector{<:Real})

Compute the OWA weights for the Weighted Conditional Value at Risk (WCVaR) of a returns series [OWA].

Inputs

alphas: N×1 vector of significance levels of each CVaR model, where N is the number of models, each .

weights: N×1 vector of weights of each CVaR model, where N is the number of models.

Outputs

PortfolioOptimiser.owa_wr — Method

owa_wr(T::Integer)

Compute the OWA weights for the Worst Realisation (WR) of a returns series [OWA].

Inputs

Outputs

PortfolioOptimiser.DBHTs — Method

DBHTs(D::AbstractMatrix{<:Real}, S::AbstractMatrix{<:Real}; branchorder::Symbol = :optimal,
      method::Symbol = :Unique)

Perform Direct Bubble Hierarchical Tree clustering, a deterministic clustering algorithm [4]. This version uses a graph-theoretic filtering technique called Triangulated Maximally Filtered Graph (TMFG).

Arguments

D: N×N dissimilarity matrix, e.g. a distance matrix.

S: N×N non-negative similarity matrix, examples include:

$\mathbf{S} = \mathbf{C} + \lvert \min \mathbf{C} \rvert$.

$\mathbf{S} = \lceil\max \left(\mathbf{D}^{\circ 2}\right)\rceil - \mathbf{D}^{\circ 2}$.

$\mathbf{S} = \exp \circ (-\mathbf{D})$.

Where $\mathbf{C}$ is the correlation matrix, $\mathbf{D}$ the dissimilarity matrix D, and $\circ$ the Hadamard (elementwise) operator.

branchorder: parameter for ordering the final dendrogram's branches accepted by Clustering.jl.

method: method for finding the root of a Direct Bubble Hierarchical Clustering Tree in case there is more than one candidate DBHTRootMethod.

:Unique: create a unique root.

:Equal: the root is created from the candidate's adjacency tree.

Outputs

T8: N×1 cluster membership vector.

Rpm: N×N adjacency matrix of the Planar Maximally Filtered Graph (PMFG).

Adjv: Bubble cluster membership matrix from BubbleCluster8s.

Dpm: N×N shortest path length matrix of the PMFG.

Mv: N×Nb bubble membership matrix. Mv[n, bi] = 1 means vertex n is a vertex of bubble bi.

Z: (N-1)×3 linkage matrix in the same format as the output from Matlab.

Z_hclust: Z matrix in Clustering.Hclust format.

PortfolioOptimiser.J_LoGo — Method

J_LoGo(sigma, separators, cliques)

Compute the sparse inverse covariance from a clique tree and separators [7].

Inputs

sigma: N×N covariance matrix.

separators: list of 3-cliques that are not triangular faces.

cliques: list of all 4-cliques.

Outputs

jlogo: J_LoGo covariance matrix.

PortfolioOptimiser.PMFG_T2s — Function

PMFG_T2s(W::AbstractMatrix{<:Real}, nargout::Integer = 3)

Constructs a Triangulated Maximally Filtered Graph (TMFG) starting from a tetrahedron and recursively inserting vertices inside existing triangles (T2 move) in order to approximate a Maximal Planar Graph with the largest total weight, aka Planar Maximally Filtered Graph (PMFG). All weights are non-negative [8].

Arguments

W: N×N matrix of non-negative weights.

nargout: number of output arguments, the same arguments are always returne, this only controls whether some arguments are empty or not.

Outputs

A: adjacency matrix of the PMFG with weights.

tri: list of triangles (triangular faces).

clique3: list of 3-cliques taht are not triangular faces, all 3-cliques are given by [tri; clique3].

cliques: list of all 4-cliques, if nargout <= 3, this will be returned as an empty array.

cliqueTree: 4-cliques tree structure (adjacency matrix), if nargout <= 4, it is returned as an empty array.

Private

PortfolioOptimiser.PosdefFix — Type

abstract type PosdefFix end

Abstract type for subtyping methods for fixing non positive definite matrices.

PortfolioOptimiser.Denoise — Type

abstract type Denoise end

Abstract type for subtyping denoising methods.

PortfolioOptimiser.AbstractBins — Type

abstract type AbstractBins end

Abstract type for defining bin width estimation functions when computing DistanceVarInfo and CorMutualInfo distance and correlation matrices respectively.

PortfolioOptimiser.AstroBins — Type

abstract type AstroBins <: AbstractBins end

Abstract type for defining which bin width function to use from astropy.

PortfolioOptimiser.DistanceMethod — Type

abstract type DistanceMethod end

Abstract type for subtyping methods for computing distance matrices from correlation ones.

PortfolioOptimiser.DBHTRootMethod — Type

abstract type DBHTRootMethod end

Abstract type for subtyping methods creating roots of cliques in CliqueRoot [5].

PortfolioOptimiser.DBHTSimilarity — Type

abstract type DBHTSimilarity end

Abstract type for subtyping methods for defining functions for computing similarity matrices used in DBHT clustering PMFG_T2s [4, 8].

PortfolioOptimiser.HClustAlg — Type

abstract type HClustAlg end

Abstract type for subtyping hierarchical clustering methods.

PortfolioOptimiser.NumClusterMethod — Type

abstract type NumClusterMethod end

Abstract type for subtyping methods for determining the number of clusters in a Clustering.Hclust when calling calc_k_clusters.

PortfolioOptimiser.CorPearson — Type

abstract type CorPearson <: PortfolioOptimiserCovCor end

Abstract type for subtyping Pearson type covariance estimators.

PortfolioOptimiser.CorRank — Type

abstract type CorRank <: PortfolioOptimiserCovCor end

Abstract type for subtyping rank based covariance estimators.

PortfolioOptimiser.PortfolioOptimiserCovCor — Type

abstract type PortfolioOptimiserCovCor <: StatsBase.CovarianceEstimator end

Abstract type for subtyping portfolio optimiser covaraince and correlation estimators.

PortfolioOptimiser.MeanEstimator — Type

abstract type MeanEstimator end

PortfolioOptimiser.MeanTarget — Type

abstract type MeanTarget end

PortfolioOptimiser.OWAMethods — Type

OWAMethods = (:CRRA, :E, :SS, :SD)

Methods for computing the weights used to combine L-moments higher than 2 [6].

:CRRA: Normalised Constant Relative Risk Aversion Coefficients.

:E: Maximum Entropy. Solver must support MOI.RelativeEntropyCone and MOI.NormOneCone.

:SS: Minimum Sum of Squares. Solver must support MOI.SecondOrderCone.

:SD: Minimum Square Distance. Solver must support MOI.SecondOrderCone.

PortfolioOptimiser._posdef_fix! — Method

_posdef_fix!(method::PosdefNearest, X::AbstractMatrix)

Overload this for other posdef fix methods.

PortfolioOptimiser._bl_mu_cov_w — Method

_mu_cov_w(tau, omega, P, Pi, Q, rf, sigma, delta, T, N, opt, cov_type, cov_flag = true)

Internal function for computing the Black Litterman statistics as defined in black_litterman. See .

Inputs

tau: variable of the same name in the Black-Litterman model.

omega: variable of the same name in the Black-Litterman model.

P: variable of the same name in the Black-Litterman model.

Pi: variable of the same name in the Black-Litterman model.

Q: variable of the same name in the Black-Litterman model.

rf: variable of the same name in the Black-Litterman model.

sigma: variable of the same name in the Black-Litterman model.

delta: variable of the same name in the Black-Litterman model.

T: variable of the same name in the Black-Litterman model.

N: variable of the same name in the Black-Litterman model.

opt: any valid instance of opt for .

cov_type: any valid value from .

cov_flag: whether the matrix is a covariance matrix or not.

Outputs

mu: asset expected returns vector obtained via the Black-Litterman model.

cov_mtx: asset covariance matrix obtained via the Black-Litterman model.

w: asset weights obtained via the Black-Litterman model.

Pi_: equilibrium excess returns after being adjusted by the views.

PortfolioOptimiser._crra_method — Method

_crra_method(weights::AbstractMatrix{<:Real}, k::Integer, g::Real)

Internal function for computing the Normalized Constant Relative Risk Aversion coefficients.

Inputs

weights: T×(k-1) matrix where T is the number of observations and k the order of the L-moments to combine, the i'th column contains the weights for the (i+1)'th L-moment.

k: the maximum order of the L-moments.

g: the risk aversion coefficient.

Outputs

w: T×1 ordered weight vector of the combined L-moments.

PortfolioOptimiser._optimise_JuMP_model — Method

_optimise_JuMP_model(model, solvers)

Internal function to optimise an OWA JuMP model.

Inputs

model: JuMP model.

Outputs

term_status: JuMP termination status.

solvers_tried: Dictionary that contains a dictionary of failed optimisations. Dict(key => Dict(...)), where key is the solver key used for the iteration of solver that failed.

If an MOI call fails on a model:

Dict(:jump_error => jump_error): JuMP error code.

If the optimiser fails to optimise the model satisfactorily:

Dict(:objective_val => JuMP.objective_value(model), :term_status => term_status, :params => haskey(val, :params) ? val[:params] : missing), where val is the value of the dictionary corresponding to key.

PortfolioOptimiser.AdjCliq — Method

AdjCliq(A::AbstractMatrix{<:Real}, CliqList::AbstractMatrix{<:Real},
        CliqRoot::AbstractVector{<:Real})

Find adjacent clique to the root candidates.

Inputs

A: N×N adjacency matrix.

CliqList: Nc×3 matrix. Each row vector lists the three vertices consisting of a 3-clique in the MPG.

CliqRoot: Nc×1 vector of root cliques.

Outputs

Adj: Nc×Nc adjacency matrix of the cliques with the root cliques.

PortfolioOptimiser.BubbleCluster8s — Method

BubbleCluster8s(Rpm::AbstractMatrix{<:Real}, Dpm::AbstractMatrix{<:Real},
                Hb::AbstractMatrix{<:Real}, Mb::AbstractMatrix{<:Real},
                Mv::AbstractMatrix{<:Real}, CliqList::AbstractMatrix{<:Real})

Obtains non-discrete and discrete clusterings from the bubble topology of the Planar Maximally Filtered Graph (PMFG).

Inputs

Rpm: N×N sparse weighted adjacency matrix of the PMFG.

Dpm: N×N shortest path lengths matrix of the PMFG.

Hb: undirected bubble tree of the PMFG.

Mb: Nc×Nb bubble membership matrix for 3-cliques. Mb[n, bi] = 1 indicated that 3-clique n belongs to bubble bi.

Mv: N×Nb bubble membership matrix for vertices.

CliqList: Nc×3 matrix. Each row vector lists the three vertices consisting of a 3-clique in the MPG.

Outputs

Adjv: N×Nk cluster membership matrix for vertices for non-discrete clustering via the bubble topology. Adjv[n, k] = 1 indicates cluster membership of vertex n to the k'th non-discrete cluster.

Tc: N×1 cluster membership vector. Tc[n] = k indicates cluster membership of vertex n to the k'th discrete cluster.

PortfolioOptimiser.BubbleHierarchy — Method

BubbleHierarchy(Pred::AbstractVector{<:Real}, Sb::AbstractVector{<:Real})

Build the bubble hierarchy.

Inputs

Pred: Nc×1 vector of predicted hierarchies.

Sb: Nc×1 vector. Sb[n] = 1 indicates 3-clique n is separating.

Outputs

Mb: Nc×Nb bubble membership matrix for 3-cliques. Mb[n, bi] = 1 indicated that 3-clique n belongs to bubble bi.

H2: Nb×Nb adjacency matrix for the bubble hierarchical tree where Nb is the number of bubbles.

PortfolioOptimiser.BubbleMember — Method

BubbleMember(Rpm::AbstractMatrix{<:Real}, Mv::AbstractMatrix{<:Real},
             Mc::AbstractMatrix{<:Real})

Assigns each vertex in the to a specific bubble.

Inputs

Rpm: N×N sparse weighted adjacency matrix of the PMFG.

Mv: N×Nb bubble membership matrix. Mv[n, bi] = 1 means vertex n is a vertex of bubble bi.

Mc: Matrix of the bubbles which coincide with the cluster.

Outputs

Mvv: Matrix of the vertices belonging to the bubble.

PortfolioOptimiser.BuildHierarchy — Method

BuildHierarchy(M::AbstractMatrix{<:Real})

Builds the predicted hierarchy.

Inputs

M: N×Nc matrix of nodes and 3-cliques.

Outputs

Pred: Nc×1 vector of predicted hierarchies.

PortfolioOptimiser.CliqHierarchyTree2s — Function

CliqHierarchyTree2s(Apm::AbstractMatrix{<:Real}, method::Symbol = :Unique)

Looks for 3-cliques of a Maximal Planar Graph (MPG), then construct a hierarchy of the cliques with the definition of "inside" a clique being a subgraph of smaller size when the entire graph is made disjoint by removing the clique [5].

Inputs

Apm: N×N adjacency matrix of an MPG.

method: method for finding the root of the graph DBHTRootMethod. Uses Voronoi tesselation between tiling triangles.

UniqueDBHT(): create a unique root.

EqualDBHT(): the root is created from the candidate's adjacency tree.

Outputs

H1: Nc×Nc adjacency matrix for 3-clique hierarchical tree where Nc is the number of 3-cliques.

H2: Nb×Nb adjacency matrix for the bubble hierarchical tree where Nb is the number of bubbles.

Mb: Nc×Nb bubble membership matrix for 3-cliques. Mb[n, bi] = 1 indicated that 3-clique n belongs to bubble bi.

CliqList: Nc×3 matrix. Each row vector lists the three vertices consisting of a 3-clique in the MPG.

Sb: Nc×1 vector. Sb[n] = 1 indicates 3-clique n is separating.

PortfolioOptimiser.CliqueRoot — Method

CliqueRoot(::UniqueDBHT, Root, Pred, Nc, args...)

PortfolioOptimiser.DendroConstruct — Method

DendroConstruct(Zi::AbstractMatrix{<:Real}, LabelVec1::AbstractVector{<:Real},
                LabelVec2::AbstractVector{<:Real},
                LinkageDist::Union{<:Real, AbstractVector{<:Real}})

Construct the linkage matrix by continuially adding rows to the matrix.

Inputs

Zi: Linkage matrix at iteration i in the same format as the output from Matlab.

LabelVec1: label vector for the vertices in the bubble for the previous valid iteration.

LabelVec2: label vector for the vertices in the bubble for the trial iteration.

Outputs

Z: Linkage matrix at iteration i + 1 in the same format as the output from Matlab.

PortfolioOptimiser.DirectHb — Method

DirectHb(Rpm::AbstractMatrix{<:Real}, Hb::AbstractMatrix{<:Real},
         Mb::AbstractMatrix{<:Real}, Mv::AbstractMatrix{<:Real},
         CliqList::AbstractMatrix{<:Real})

Computes the directions on each separating 3-clique of a Maximal Planar Graph (MPH), hence computes the Directed Bubble Hierarchy Tree (DBHT).

Inputs

Rpm: N×N sparse weighted adjacency matrix of the Planar Maximally Filtered Graph (MPFG).

Hb: Undirected bubble tree of the PMFG.

Mb: Nc×Nb bubble membership matrix for 3-cliques. Mb[n, bi] = 1 indicated that 3-clique n belongs to bubble bi.

Mv: N×Nb bubble membership matrix for vertices.

CliqList: Nc×3 matrix. Each row vector lists the three vertices consisting of a 3-clique in the MPG.

Outputs

Hc: Nb×Nb unweighted directed adjacency matrix of the DBHT. Hc[i, j]=1 indicates a directed edge from bubble i to bubble j.

PortfolioOptimiser.FindDisjoint — Method

FindDisjoint(Adj::AbstractMatrix{<:Real}, Cliq::AbstractVector{<:Real})

Finds disjointed cliques in adjacency matrix.

Inputs

Adj: N×N adjacency matrix.

Cliq: 3×1 vector of 3-cliques.

Outputs

T: N×1 vector containing the adjacency number of each node.

IndxNot: N×1 vector of nodes with no adjacencies.

PortfolioOptimiser.HierarchyConstruct4s — Method

HierarchyConstruct4s(Rpm::AbstractMatrix{<:Real}, Dpm::AbstractMatrix{<:Real},
                     Tc::AbstractVector{<:Real}, Mv::AbstractMatrix{<:Real})

Constructs the intra- and inter-cluster hierarchy by utilizing Bubble Hierarchy structure of a Maximal Planar graph, in this a Planar Maximally Filtered Graph (PMFG).

Inputs

Rpm: N×N sparse weighted adjacency matrix of the PMFG.

Dpm: N×N shortest path lengths matrix of the PMFG.

Tc: N×1 cluster membership vector. Tc[n] = k indicates cluster membership of vertex n to the k'th discrete cluster.

Mv: N×Nb bubble membership matrix. Mv[n, bi] = 1 means vertex n is a vertex of bubble bi.

Outputs

Z: (N-1)×3 linkage matrix in the same format as the output from Matlab.

PortfolioOptimiser.LinkageFunction — Method

LinkageFunction(d::AbstractMatrix{<:Real}, labelvec::AbstractVector{<:Real})

Looks for the pair of clusters with the best linkage.

Inputs

d: Nv×Nv distance matrix for a list of vertices assigned to a bubble.

labelvec: label vector for the vertices in the bubble.

Outputs

PairLink: pair of links with the best linkage.

dvu: value of the best linkage.

PortfolioOptimiser._build_link_and_dendro — Method

_build_link_and_dendro(rg::AbstractRange, dpm::AbstractMatrix{<:Real},
                       LabelVec::AbstractVector{<:Real}, LabelVec1::AbstractVector{<:Real},
                       LabelVec2::AbstractVector{<:Real}, V::AbstractVector{<:Real},
                       nc::Real, Z::AbstractMatrix{<:Real})

Computes iterates over the vertices to construct the linkage matrix iteration by iteration.

Inputs

rg: range of indices of the vertices in a bubble.

dpm: Nv×Nv distance matrix for a list of vertices assigned to a bubble.

LabelVec: vector labels of all vertices.

LabelVec1: label vector for the vertices in the bubble for the previous valid iteration.

LabelVec2: label vector for the vertices in the bubble for the trial iteration.

Outputs

Z: updated linkage matrix in the same format as the output from Matlab.

nc: updated inverse of the linkage distance.

LabelVec1: updated LabelVec1 for the next iteration.

PortfolioOptimiser.breadth — Method

breadth(CIJ::AbstractMatrix{<:Real}, source::Integer)

Breadth-first search.

Inputs

CIJ: binary (directed/undirected) connection matrix.

source: source vertex.

Outputs

distance: distance between source and i'th vertex (0 for source vertex).

branch: vertex that precedes i in the breadth-first search tree (-1 for source vertex).

Note

Breadth-first search tree does not contain all paths (or all shortest paths), but allows the determination of at least one path with minimum distace. The entire graph is explored, starting from source vertex source. # Colours

Original written by: Olaf Sporns, Indiana University, 2002/2007/2008

PortfolioOptimiser.clique3 — Method

clique3(A::AbstractMatrix{<:Real})

Computes the list of 3-cliques.

Inputs

A: N×N adjacency matrix of a Maximal Planar Graph (MPG).

Outputs

K3: vector of vectors with the corresponding indices of candidate cliques.

E: matrix with non-zero indices and entries of candidate cliques.

CliqList: Nc×3 matrix. Each row vector lists the three vertices consisting of a 3-clique in the MPG.

PortfolioOptimiser.distance_wei — Method

distance_wei(L::AbstractMatrix{<:Real})

The distance matrix contains lengths of shortest paths between all node pairs. An entry [u, v] represents the length of the shortest path from node u to node v. The average shortest path length is the characteristic path length of the network. The function uses Dijkstra's algorithm.

Inputs

L: Directed/undirected connection-length matrix.

Lengths between disconnected nodes are set to Inf.

Lengths on the main diagonal are set to 0.

Note

The input matrix must be a connection-length matrix typically obtained via a mapping from weight to length. For instance, in a weighted correlation network, higher correlations are more naturally interpreted as shorter distances, and the input matrix should therefore be some inverse of the connectivity matrix, i.e. a distance matrix.

The number of edges in the shortest weighted path may in general exceed the number of edges in the shortest binary paths (i.e. the shortest weighted paths computed on the binarised connectivity matrix), because the shortest weighted paths have the minimal weighted distance, not necessarily the minimal number of edges.

Outputs

D: distance (shortest weighted path) matrix.

B: number of edged in the shortest weigthed path matrix.

Note

Based on a Matlab implementation by:

Mika Rubinov, UNSW/U Cambridge, 2007-2012.

Rick Betzel and Andrea Avena, IU, 2012

PortfolioOptimiser.turn_into_Hclust_merges — Method

turn_into_Hclust_merges(Z::AbstractMatrix{<:Real})

Turns a Matlab-style linkage matrix to a useable format for Hclust.

Inputs

Z: Matlab-style linkage matrix.

Outputs

Z: Hclust-style linkage matrix.

OWACajas, Dany, OWA Portfolio Optimization: A Disciplined Convex Programming Framework (December 18, 2021). Available at SSRN: https://ssrn.com/abstract=3988927 or http://dx.doi.org/10.2139/ssrn.3988927