Covariance, correlation, cokurtosis and coskewness

Public

PortfolioOptimiser.CorGerber0 — Type

@kwdef mutable struct CorGerber0{T1 <: Real} <: CorGerberBasic
    normalise::Bool = false
    threshold::T1 = 0.5
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    std_w::Union{<:AbstractWeights, Nothing} = nothing
    mean_w::Union{<:AbstractWeights, Nothing} = nothing
    posdef::PosdefFix = PosdefNearest()
end

Gerber type 0 covariance and correlation matrices.

Parameters

normalise:

if true: Z-normalise the data before applying the Gerber criteria.

threshold: Gerber significance threshold, threshold ∈ (0, 1).

ve: variance estimator StatsBase.CovarianceEstimator.

std_w: optional T×1 vector of weights for computing the variance.

Only used when normalise == true:

mean_w: optional T×1 vector of weights for computing the mean.

posdef: method for fixing the Gerber type 0 correaltion matrix PosdefFix.

PortfolioOptimiser.CorGerber1 — Type

@kwdef mutable struct CorGerber1{T1 <: Real} <: CorGerberBasic
    normalise::Bool = false
    threshold::T1 = 0.5
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    std_w::Union{<:AbstractWeights, Nothing} = nothing
    mean_w::Union{<:AbstractWeights, Nothing} = nothing
    posdef::PosdefFix = PosdefNearest()
end

Gerber type 1 covariance and correlation matrices.

Parameters

normalise:

if true: Z-normalise the data before applying the Gerber criteria.

threshold: Gerber significance threshold, threshold ∈ (0, 1).

ve: variance estimator StatsBase.CovarianceEstimator.

std_w: optional T×1 vector of weights for computing the variance.

Only used when normalise == true:

mean_w: optional T×1 vector of weights for computing the mean.

posdef: method for fixing the Gerber type 1 correaltion matrix PosdefFix.

PortfolioOptimiser.CorGerber2 — Type

@kwdef mutable struct CorGerber2{T1 <: Real} <: CorGerberBasic
    normalise::Bool = false
    threshold::T1 = 0.5
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    std_w::Union{<:AbstractWeights, Nothing} = nothing
    mean_w::Union{<:AbstractWeights, Nothing} = nothing
    posdef::PosdefFix = PosdefNearest()
end

Gerber type 2 covariance and correlation matrices.

Parameters

normalise:

if true: Z-normalise the data before applying the Gerber criteria.

threshold: Gerber significance threshold, threshold ∈ (0, 1).

ve: variance estimator StatsBase.CovarianceEstimator.

std_w: optional T×1 vector of weights for computing the variance.

Only used when normalise == true:

mean_w: optional T×1 vector of weights for computing the mean.

posdef: method for fixing the Gerber type 2 correaltion matrix PosdefFix.

PortfolioOptimiser.CorGerberSB0 — Type

@kwdef mutable struct CorGerberSB0{T1, T2, T3, T4, T5} <: CorSB
    normalise::Bool = false
    threshold::T1 = 0.5
    c1::T2 = 0.5
    c2::T3 = 0.5
    c3::T4 = 4.0
    n::T5 = 2.0
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    std_w::Union{<:AbstractWeights, Nothing} = nothing
    mean_w::Union{<:AbstractWeights, Nothing} = nothing
    posdef::PosdefFix = PosdefNearest()
end

Smyth-Broby modification with vote counting of the Gerber type 0 covariance and correlation matrices.

Parameters

normalise:

if true: Z-normalise the data before applying the Gerber criteria.

threshold: Gerber significance threshold, threshold ∈ (0, 1).

c1: confusion zone threshold ($c_1$ in the paper), c1 ∈ (0, 1].

c2: indecision zone threshold ($c_2$ in the paper), c2 ∈ (0, 1].

c3: large co-movement threshold (4 in the paper).

n: exponent of the regularisation term ($n = 2$ in the paper).

ve: variance estimator StatsBase.CovarianceEstimator.

std_w: optional T×1 vector of weights for computing the variance.

mean_w: optional T×1 vector of weights for computing the mean.

posdef: method for fixing the Smyth-Broby modification with vote counting of the Gerber type 0 correaltion matrix PosdefFix.

PortfolioOptimiser.CorGerberSB1 — Type

@kwdef mutable struct CorGerberSB1{T1, T2, T3, T4, T5} <: CorSB
    normalise::Bool = false
    threshold::T1 = 0.5
    c1::T2 = 0.5
    c2::T3 = 0.5
    c3::T4 = 4.0
    n::T5 = 2.0
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    std_w::Union{<:AbstractWeights, Nothing} = nothing
    mean_w::Union{<:AbstractWeights, Nothing} = nothing
    posdef::PosdefFix = PosdefNearest()
end

Smyth-Broby modification with vote counting of the Gerber type 1 covariance and correlation matrices.

Parameters

normalise:

if true: Z-normalise the data before applying the Gerber criteria.

threshold: Gerber significance threshold, threshold ∈ (0, 1).

c1: confusion zone threshold ($c_1$ in the paper), c1 ∈ (0, 1].

c2: indecision zone threshold ($c_2$ in the paper), c2 ∈ (0, 1].

c3: large co-movement threshold (4 in the paper).

n: exponent of the regularisation term ($n = 2$ in the paper).

ve: variance estimator StatsBase.CovarianceEstimator.

std_w: optional T×1 vector of weights for computing the variance.

mean_w: optional T×1 vector of weights for computing the mean.

posdef: method for fixing the Smyth-Broby modification with vote counting of the Gerber type 1 correaltion matrix PosdefFix.

PortfolioOptimiser.CorKendall — Type

@kwdef mutable struct CorKendall <: CorRank
    absolute::Bool = false
end

Kendall type correlation estimator.

Parameters

absolute:

if true: compute an absolute correlation, abs.(corkendall(X)).

PortfolioOptimiser.CorLTD — Type

mutable struct CorLTD <: PortfolioOptimiserCovCor
    alpha::Real
    ve::StatsBase.CovarianceEstimator
    w::Union{<:AbstractWeights, Nothing}
end

Lower tail dependence correlation and covariance matrix estimator.

Parameters

alpha: significance level of the lower tail dependence, alpha ∈ (0, 1).

Only used when computing covariance matrices:

ve: variance estimator StatsBase.CovarianceEstimator.

w: optional T×1 vector of weights for computing the variance.

PortfolioOptimiser.CorMutualInfo — Type

@kwdef mutable struct CorMutualInfo <: PortfolioOptimiserCovCor
    bins::Union{<:Integer, <:AbstractBins} = HGR()
    normalise::Bool = true
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    w::Union{<:AbstractWeights, Nothing} = nothing
end

Mutual information correlation matrix estimator.

Parameters

bins:

if isa(bins, AbstractBins): defines the function for computing bin widths.

if isa(bins, Integer) and bins > 0: directly provide the number of bins.

normalise:

if true: normalise the mutual information.

Only used when computing covariance matrices:

ve: variance estimator StatsBase.CovarianceEstimator.

w: optional T×1 vector of weights for computing the variance.

PortfolioOptimiser.CorSB0 — Type

@kwdef mutable struct CorSB0{T1, T2, T3, T4, T5} <: CorSB
    normalise::Bool = false
    threshold::T1 = 0.5
    c1::T2 = 0.5
    c2::T3 = 0.5
    c3::T4 = 4.0
    n::T5 = 2.0
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    std_w::Union{<:AbstractWeights, Nothing} = nothing
    mean_w::Union{<:AbstractWeights, Nothing} = nothing
    posdef::PosdefFix = PosdefNearest()
end

Smyth-Broby modification of the Gerber type 0 covariance and correlation matrices.

Parameters

normalise:

if true: Z-normalise the data before applying the Gerber criteria.

threshold: Gerber significance threshold, threshold ∈ (0, 1).

c1: confusion zone threshold ($c_1$ in the paper), c1 ∈ (0, 1].

c2: indecision zone threshold ($c_2$ in the paper), c2 ∈ (0, 1].

c3: large co-movement threshold (4 in the paper).

n: exponent of the regularisation term ($n = 2$ in the paper).

ve: variance estimator StatsBase.CovarianceEstimator.

std_w: optional T×1 vector of weights for computing the variance.

mean_w: optional T×1 vector of weights for computing the mean.

posdef: method for fixing the Smyth-Broby modification of the Gerber type 0 correaltion matrix PosdefFix.

PortfolioOptimiser.CorSB1 — Type

@kwdef mutable struct CorSB1{T1, T2, T3, T4, T5} <: CorSB
    normalise::Bool = false
    threshold::T1 = 0.5
    c1::T2 = 0.5
    c2::T3 = 0.5
    c3::T4 = 4.0
    n::T5 = 2.0
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    std_w::Union{<:AbstractWeights, Nothing} = nothing
    mean_w::Union{<:AbstractWeights, Nothing} = nothing
    posdef::PosdefFix = PosdefNearest()
end

Smyth-Broby modification of the Gerber type 1 covariance and correlation matrices.

Parameters

normalise:

if true: Z-normalise the data before applying the Gerber criteria.

threshold: Gerber significance threshold, threshold ∈ (0, 1).

c1: confusion zone threshold ($c_1$ in the paper), c1 ∈ (0, 1].

c2: indecision zone threshold ($c_2$ in the paper), c2 ∈ (0, 1].

c3: large co-movement threshold (4 in the paper).

n: exponent of the regularisation term ($n = 2$ in the paper).

ve: variance estimator StatsBase.CovarianceEstimator.

std_w: optional T×1 vector of weights for computing the variance.

mean_w: optional T×1 vector of weights for computing the mean.

posdef: method for fixing the Smyth-Broby modification of the Gerber type 1 correaltion matrix PosdefFix.

PortfolioOptimiser.CorSpearman — Type

@kwdef mutable struct CorSpearman <: CorRank
    absolute::Bool = false
end

Spearman type correlation estimator.

Parameters

absolute:

if true: compute an absolute correlation, abs.(corspearman(X)).

PortfolioOptimiser.CovDistance — Type

mutable struct CovDistance <: PortfolioOptimiserCovCor
    distance::Distances.Metric
    dist_args::Tuple
    dist_kwargs::NamedTuple
    mean_w1::Union{<:AbstractWeights, Nothing}
    mean_w2::Union{<:AbstractWeights, Nothing}
    mean_w3::Union{<:AbstractWeights, Nothing}
end

Distance covariance and correlation matrix estimator.

Parameters

distance: distance metric from Distances.jl.

dist_args: args for the Distances.pairwise function of Distances.jl.

dist_kwargs: kwargs for the Distances.pairwise function of Distances.jl.

mean_w1: optional T×1 vector of weights for computing the mean of the pairwise distance matrices along its rows (dims = 1).

mean_w2: optional T×1 vector of weights for computing the mean of the pairwise distance matrices along its columns (dims = 2).

mean_w3: optional T×1 vector of weights for computing the mean of the entirety of the pairwise distance matrices.

PortfolioOptimiser.CovFull — Type

@kwdef mutable struct CovFull <: CorPearson
    absolute::Bool = false
    ce::StatsBase.CovarianceEstimator = StatsBase.SimpleCovariance(; corrected = true)
    w::Union{<:AbstractWeights, Nothing} = nothing
end

Full Pearson-type covariance and correlation estimator.

Parameters

absolute:

if true: compute an absolute correlation, abs.(cor(X)).

ce: covariance estimator.

w: optional T×1 vector of weights for computing the covariance.

PortfolioOptimiser.CovSemi — Type

@kwdef mutable struct CovSemi <: CorPearson
    absolute::Bool = false
    ce::StatsBase.CovarianceEstimator = StatsBase.SimpleCovariance(; corrected = true)
    target::Union{<:Real, AbstractVector{<:Real}} = 0.0
    w::Union{<:AbstractWeights, Nothing} = nothing
end

Semi Pearson-type covariance and correlation estimator.

Parameters

absolute:

if true: compute an absolute correlation, abs.(cor(X)).

ce: covariance estimator.

target: minimum return threshold for classifying downside returns.

if isa(target, Real): apply the same target to all assets.

if isa(target, AbstractVector): apply individual target to each asset.

w: optional T×1 vector of weights for computing the covariance.

PortfolioOptimiser.KurtFull — Type

@kwdef mutable struct KurtFull <: KurtEstimator
    posdef::PosdefFix = PosdefNearest(;)
    denoise::Denoise = NoDenoise(;)
    logo::AbstractLoGo = NoLoGo(;)
end

Full cokurtosis estimator.

Parameters

posdef: method for fixing non a positive definite cokurtosis matrix PosdefFix.

denoise: method for denoising the cokurtosis matrix Denoise.

logo: method for computing the LoGo cokurtosis matrix AbstractLoGo.

PortfolioOptimiser.KurtSemi — Type

@kwdef mutable struct KurtSemi <: KurtEstimator
    target::Union{<:Real, AbstractVector{<:Real}} = 0.0
    posdef::PosdefFix = PosdefNearest(;)
    denoise::Denoise = NoDenoise(;)
    logo::AbstractLoGo = NoLoGo(;)
end

Semi cokurtosis estimator.

Parameters

target: minimum return threshold for classifying downside returns.

if isa(target, Real): apply the same target to all assets.

if isa(target, AbstractVector): apply individual target to each asset.

posdef: method for fixing non a positive definite semi cokurtosis matrix PosdefFix.

denoise: method for denoising the semi cokurtosis matrix Denoise.

logo: method for computing the LoGo semi cokurtosis matrix AbstractLoGo.

PortfolioOptimiser.LoGo — Type

@kwdef mutable struct LoGo <: AbstractLoGo
    distance::DistMethod = DistMLP()
    similarity::DBHTSimilarity = DBHTMaxDist()
end

Compute the LoGo covariance and correlation matrix estimator.

Parameters

distance: method for computing the distance (disimilarity) matrix from the correlation matrix if the distance matrix is not provided to logo!.

similarity: method for computing the similarity matrix from the correlation and distance matrices. The distance matrix is used to compute sparsity pattern of the inverse of the LoGo covariance and correlation matrices.

PortfolioOptimiser.NoLoGo — Type

struct NoLoGo <: AbstractLoGo end

Leave the matrix as is.

PortfolioOptimiser.PortCovCor — Type

@kwdef mutable struct PortCovCor <: PortfolioOptimiserCovCor
    ce::CovarianceEstimator = CovFull(;)
    posdef::PosdefFix = PosdefNearest(;)
    denoise::Denoise = NoDenoise(;)
    logo::AbstractLoGo = NoLoGo(;)
end

PortfolioOptimiser covariance and correlation estimator.

Parameters

ce: covariance estimator.

posdef: method for fixing the portfolio covariance or correlation matrix PosdefFix.

denoise: method for denoising the portfolio covariance or correlation matrix Denoise.

logo: method for computing the LoGo portfolio covariance or correlation matrix AbstractLoGo.

PortfolioOptimiser.SimpleVariance — Type

@kwdef mutable struct SimpleVariance <: StatsBase.CovarianceEstimator
    corrected::Bool = true
end

Simple variance estimator.

Parameters

corrected:

if true: correct the bias dividing by N-1 instead of N.

PortfolioOptimiser.SkewFull — Type

struct SkewFull <: SkewEstimator end

Full cokurtosis estimator.

PortfolioOptimiser.SkewSemi — Type

@kwdef mutable struct SkewSemi <: SkewEstimator
    target::Union{<:Real, AbstractVector{<:Real}} = 0.0
end

Semi cokurtosis estimator.

Parameters

target: minimum return threshold for classifying downside returns.

if isa(target, Real): apply the same target to all assets.

if isa(target, AbstractVector): apply individual target to each asset.

PortfolioOptimiser.coskew — Method

coskew(::SkewFull, X::AbstractMatrix, mu::AbstractVector)

PortfolioOptimiser.coskew — Method

coskew(se::SkewSemi, X::AbstractMatrix, mu::AbstractVector)

PortfolioOptimiser.logo! — Function

logo!(je::LoGo, posdef::PosdefFix, X::AbstractMatrix, D = nothing)

PortfolioOptimiser.logo! — Function

logo!(::NoLoGo, ::PosdefFix, ::AbstractMatrix, D = nothing)

Private

PortfolioOptimiser.PosdefFixCovCor — Type

const PosdefFixCovCor = Union{<:CorGerber, PortCovCor}

Covariance and correlation estimators that support positive definite fixes.

PortfolioOptimiser.AbstractLoGo — Type

abstract type AbstractLoGo end

Abstract type for subtyping LoGo covariance and correlation matrix estimators.

PortfolioOptimiser.CorGerber — Type

abstract type CorGerber <: PortfolioOptimiserCovCor end

Abstract type for subtyping Gerber type covariance and correlation estimators.

PortfolioOptimiser.CorGerberBasic — Type

abstract type CorGerberBasic <: CorGerber end

Abstract type for subtyping the original Gerber type covariance and correlation estimators.

PortfolioOptimiser.CorGerberSB — Type

abstract type CorSB <: CorGerber end

Abstract type for subtyping the Smyth-Broby modifications with vote counting of Gerber type covariance and correlation estimators.

PortfolioOptimiser.CorPearson — Type

abstract type CorPearson <: PortfolioOptimiserCovCor end

Abstract type for subtyping Pearson type covariance estimators.

PortfolioOptimiser.CorRank — Type

abstract type CorRank <: PortfolioOptimiserCovCor end

Abstract type for subtyping rank based covariance estimators.

PortfolioOptimiser.CorSB — Type

abstract type CorSB <: CorGerber end

Abstract type for subtyping the Smyth-Broby modifications of Gerber type covariance and correlation estimators.

PortfolioOptimiser.KurtEstimator — Type

abstract type KurtEstimator end

Abstract type for subtyping cokurtosis estimators.

PortfolioOptimiser.PortfolioOptimiserCovCor — Type

abstract type PortfolioOptimiserCovCor <: StatsBase.CovarianceEstimator end

Abstract type for subtyping portfolio covariance and correlation estimators.

PortfolioOptimiser.SkewEstimator — Type

abstract type SkewEstimator end

Abstract type for subtyping coskew estimators.