Distance matrices

Public

PortfolioOptimiser.DistCanonical — Type

struct DistCanonical <: DistMethod end

Struct for computing the canonical distance for a given correlation estimator in get_default_dist.

Correlation estimator	Canonical distance

CorMutualInfo	DistVarInfo

CorLTD	DistLog

CovDistance	DistCor

Any other estimator	DistMLP

PortfolioOptimiser.DistCor — Type

struct DistCor <: DistMethod end

Defines the distance matrix from the correlation matrix.

\[\begin{align} D_{i,\,j} &= \sqrt{1 - C_{i,\,j}}\,. \end{align}\]

Where:

$D_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N distance matrix.

$C_{i,\,j}$: is the $(i,\,j)$-th entry of a distance correlation matrix.

PortfolioOptimiser.DistDistCanonical — Type

struct DistDistCanonical <: DistMethod end

Struct for computing the canonical distance for a given correlation estimator in get_default_dist.

Correlation estimator	Canonical distance

CorMutualInfo	DistDistVarInfo

CorLTD	DistDistLog

CovDistance	DistDistCor

Any other estimator	DistDistMLP

PortfolioOptimiser.DistDistLog — Type

@kwdef mutable struct DistDistLog <: DistMethod
    distance::Distances.Metric
    args::Tuple
    kwargs::NamedTuple
end

Defines the distance of distances matrix from the correlation matrix.

\[\begin{align} D_{i,\,j} &= -\log\left(C_{i,\,j}\right)\,. \end{align}\]

Where:

$D_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N log-distance matrix.

$C_{i,\,j}$: is the $(i,\,j)$-th entry of an absolute correlation matrix.

Parameters

absolute:

if true: the correlation being used is absolute.

distance: distance metric from Distances.jl.

args: args for the Distances.pairwise function.

kwargs: key word args for the Distances.pairwise function.

PortfolioOptimiser.DistDistMLP — Type

@kwdef mutable struct DistDistMLP <: DistMethod
    absolute::Bool = false
    distance::Distances.Metric = Distances.Euclidean()
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the distance of distances matrix from a correlation matrix [2] in dist.

\[\begin{align} \tilde{D}_{i,\,j} &= f_{m}\left(\bm{D}_{i},\, \bm{D}_j\right)\,. \end{align}\]

Where:

$\bm{D}_{i}$: is the $i$-th column/row of the N×N distance matrix defined in DistMLP.

$f_{m}$: is the pairwise distance function for metric $m$. We use the Distances.pairwise function which computes the entire matrix at once.

$\tilde{D}_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N distances of distances matrix.

absolute:

if true: the correlation being used is absolute.

Parameters

absolute:

if true: the correlation being used is absolute.

distance: distance metric from Distances.jl.

args: args for the Distances.pairwise function.

kwargs: key word args for the Distances.pairwise function.

PortfolioOptimiser.DistDistVarInfo — Type

@kwdef mutable struct DistDistVarInfo <: DistMethod
    bins::Union{<:Integer, <:AbstractBins} = HGR()
    normalise::Bool = true
end

Defines the variation of information distance of distances matrix.

Parameters

bins:

if isa(bins, AbstractBins): defines the function for computing bin widths.

if isa(bins, Integer) and bins > 0: directly provide the number of bins.

normalise:

if true: normalise the mutual information.

PortfolioOptimiser.DistLog — Type

struct DistLog <: DistMethod end

Defines the log-distance matrix from the correlation matrix.

\[\begin{align} D_{i,\,j} &= -\log\left(C_{i,\,j}\right)\,. \end{align}\]

Where:

$D_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N log-distance matrix.

$C_{i,\,j}$: is the $(i,\,j)$-th entry of an absolute correlation matrix.

PortfolioOptimiser.DistMLP — Type

@kwdef mutable struct DistMLP <: DistMethod
    absolute::Bool = false
end

Defines the distance matrix from a correlation matrix [2] in dist.

\[\begin{align} D_{i,\,j} &= \begin{cases} \sqrt{\dfrac{1}{2} \left(1 - C_{i,\,j}\right)} &\quad \mathrm{if~ absolute = false}\\ \sqrt{1 - \lvert C_{i,\,j} \rvert} &\quad \mathrm{if~ absolute = true}\,. \end{cases} \end{align}\]

Where:

$D_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N distance matrix $\mathbf{C}$.

$C_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N correlation matrix $\mathbf{D}$.

absolute:

if true: the correlation being used is absolute.

Parameters

absolute:

if true: the correlation being used is absolute.

PortfolioOptimiser.DistVarInfo — Type

@kwdef mutable struct DistVarInfo <: DistMethod
    bins::Union{<:Integer, <:AbstractBins} = HGR()
    normalise::Bool = true
end

Defines the variation of information distance matrix.

Parameters

bins:

if isa(bins, AbstractBins): defines the function for computing bin widths.

if isa(bins, Integer) and bins > 0: directly provide the number of bins.

normalise:

if true: normalise the mutual information.

PortfolioOptimiser.Freedman — Type

struct Freedman <: AstroBins end

Freedman's bin width algorithm from astropy.

PortfolioOptimiser.HGR — Type

struct HGR <: AbstractBins end

Hacine-Gharbi and Ravier's bin width algorithm [3].

PortfolioOptimiser.Knuth — Type

struct Knuth <: AstroBins end

Knuth's bin width algorithm from astropy.

PortfolioOptimiser.Scott — Type

struct Scott <: AstroBins end

Scott's bin width algorithm from astropy.

PortfolioOptimiser.dist — Method

dist(de::DistMethod, X, Y)

Private

PortfolioOptimiser.AbstractBins — Type

abstract type AbstractBins end

Abstract type for defining the bin width estimation functions when computing DistVarInfo and CorMutualInfo distance and correlation matrices respectively.

PortfolioOptimiser.AstroBins — Type

abstract type AstroBins <: AbstractBins end

Abstract type for defining which bin width function to use from astropy.

PortfolioOptimiser.DistDistCor — Type

struct DistdistCor <: DistMethod end

Defines the distance of distances matrix from the correlation matrix.

\[\begin{align} D_{i,\,j} &= \sqrt{1 - C_{i,\,j}}\,. \end{align}\]

Where:

$D_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N distance matrix.

$C_{i,\,j}$: is the $(i,\,j)$-th entry of a distance correlation matrix.

PortfolioOptimiser.DistMethod — Type

abstract type DistMethod end

Abstract type for subtyping methods for computing distance matrices from correlation ones.

PortfolioOptimiser.get_default_dist — Method

get_default_dist(dist_type::DistMethod, cor_type::PortfolioOptimiserCovCor)

Inputs

if isa(cor_type, PortCovCor): operates on the internal correlation estimator cor_type.ce.

else: directly operates on the correlation estimator cor_type.