Matrix denoising

Public

PortfolioOptimiser.DenoiseFixed — Type

@kwdef mutable struct DenoiseFixed{T1, T2, T3, T4} <: AbstractDenoise
    kernel = AverageShiftedHistograms.Kernels.gaussian
    m::Integer = 10
    n::Integer = 1000
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the parameters for using the fixed type in denoise! [1, Chapter 2]. This type performs an eigendecomposition of the original correlation matrix, sets the eigenvalues that are below the noise significance threshold to their average, and reconstructs the correlation matrix using the modified values.

Parameters

detone:

true: remove the largest mkt_comp eigenvalues from the correlation matrix.

Warning

Removing eigenvalues from the matrix may make it singular.

mkt_comp: the number of largest eigenvalues to remove from the correlation matrix.

kernel: kernel for fitting the average shifted histograms from AverageShiftedHistograms.jl Kernel Functions.

m: number of adjacent histograms to smooth over AverageShiftedHistograms.ash.

n: number of points used when creating the range of values to which the average shifted histogram is to be fitted AverageShiftedHistograms.ash.

args: arguments for Optim.optimize

kwargs: keyword arguments for Optim.optimize

PortfolioOptimiser.DenoiseShrink — Type

@kwdef mutable struct DenoiseShrink{T1, T2, T3, T4, T5} <: AbstractDenoise
    alpha::Real = 0.0
    detone::Bool = false
    mkt_comp::Integer = 1
    kernel = AverageShiftedHistograms.Kernels.gaussian
    m::Integer = 10
    n::Integer = 1000
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the parameters for using the shrink type in denoise! [1, Chapter 2]. This type shrinks the covariance to a target of unequal variance of the matrix reconstructed with the eigenvalues that are below the noise significance value. The degree of shrinkage is parametrised by alpha.

\[\begin{align} \mathbf{C} &= \mathbf{C}_{u} + \alpha \mathbf{C}_{l} + (1 - \alpha) \mathrm{Diag}(\mathbf{C})\\ \mathbf{C}_{u} &= \mathbf{E}_{u} \mathbf{\lambda}_{u} \mathbf{E}_{u}^{\intercal}\\ \mathbf{C}_{l} &= \mathbf{E}_{l} \mathbf{\lambda}_{l} \mathbf{E}_{l}^{\intercal}\,. \end{align}\]

Where:

$\mathbf{C}$ is the denoised correlation matrix.

$\mathbf{C}_{u}$ is the reconstructed correlation matrix out of the eigenvalues that are above the noise significance threshold, and their corresponding eigenvectors.

$\mathbf{C}_{l}$ is the reconstructed correlation matrix out of the eigenvalues that are below the noise significance threshold, and their corresponding eigenvectors.

$\mathbf{E}_{u}$ are the eigenvectors corresponding to the eigenvalues that are above the noise significance threshold.

$\mathbf{\lambda}_{u}$ is the diagonal matrix of eigenvalues that are above the noise significance threshold.

$\mathbf{E}_{l}$ are the eigenvectors corresponding to the eigenvalues that are below the noise significance threshold.

$\mathbf{\lambda}_{l}$ is the diagonal matrix of eigenvalues that are below the noise significance threshold.

Parameters

alpha: tuning parameter for how much the matrix should be shrunk, alpha ∈ [0, 1].

detone:

true: take only the largest mkt_comp eigenvalues from the correlation matrix.

Warning

Removing eigenvalues from the matrix may make it singular.

mkt_comp: the number of largest eigenvalues to keep from the correlation matrix.

kernel: kernel for fitting the average shifted histograms from AverageShiftedHistograms.jl Kernel Functions.

m: number of adjacent histograms to smooth over AverageShiftedHistograms.ash.

n: number of points used when creating the range of values to which the average shifted histogram is to be fitted AverageShiftedHistograms.ash.

args: arguments for Optim.optimize

kwargs: keyword arguments for Optim.optimize

PortfolioOptimiser.DenoiseSpectral — Type

@kwdef mutable struct DenoiseSpectral{T1, T2, T3, T4} <: AbstractDenoise
    detone::Bool = false
    mkt_comp::Integer = 1
    kernel = AverageShiftedHistograms.Kernels.gaussian
    m::Integer = 10
    n::Integer = 1000
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the parameters for using the spectral type in denoise! [1, Chapter 2]. This type performs an eigendecomposition of the original correlation matrix, sets the eigenvalues that are below the noise significance threshold to zero, and reconstructs the correlation matrix using the modified values.

Parameters

detone:

true: take only the largest mkt_comp eigenvalues from the correlation matrix.

Warning

Removing eigenvalues from the matrix may make it singular.

mkt_comp: the number of largest eigenvalues to keep from the correlation matrix.

kernel: kernel for fitting the average shifted histograms from AverageShiftedHistograms.jl Kernel Functions.

m: number of adjacent histograms to smooth over AverageShiftedHistograms.ash.

n: number of points used when creating the range of values to which the average shifted histogram is to be fitted AverageShiftedHistograms.ash.

args: arguments for Optim.optimize

kwargs: keyword arguments for Optim.optimize

PortfolioOptimiser.NoDenoise — Type

struct NoDenoise <: AbstractDenoise end

Apply no denoising in denoise!.

PortfolioOptimiser.denoise! — Method

denoise!(ce::AbstractDenoise, posdef::AbstractPosdefFix, X::AbstractMatrix, q::Real)

Private

PortfolioOptimiser.AbstractDenoise — Type

abstract type AbstractDenoise end

Abstract type for subtyping denoising types.