Expected returns

Public

PortfolioOptimiser.GM — Type

struct GM <: MeanTarget end

Grand mean target.

PortfolioOptimiser.MuBOP — Type

@kwdef mutable struct MuBOP <: MeanSigmaEstimator
    target::MeanTarget = GM()
    w::Union{<:AbstractWeights, Nothing} = nothing
    sigma::Union{<:AbstractMatrix, Nothing} = nothing
end

Bodnar-Okhrin-Parolya [8] expected returns vector estimator.

Parameters

target: correction target for the estimator.

w: optional T×1 vector of weights for computing the simple expected returns vector.

sigma: value of the covariance matrix used for correcting the simple expected returns vector.

if computing with asset_statistics! and isnothing(sigma): use the covariance matrix computed by .

PortfolioOptimiser.MuBS — Type

@kwdef mutable struct MuBS <: MeanSigmaEstimator
    target::MeanTarget = GM()
    w::Union{<:AbstractWeights, Nothing} = nothing
    sigma::Union{<:AbstractMatrix, Nothing} = nothing
end

Bayes-Stein [9] expected returns vector estimator.

Parameters

target: correction target for the estimator.

w: optional T×1 vector of weights for computing the simple expected returns vector.

sigma: value of the covariance matrix used for correcting the simple expected returns vector.

if computing with asset_statistics! and isnothing(sigma): use the covariance matrix computed by .

PortfolioOptimiser.MuJS — Type

@kwdef mutable struct MuJS <: MeanSigmaEstimator
    target::MeanTarget = GM()
    w::Union{<:AbstractWeights, Nothing} = nothing
    sigma::Union{<:AbstractMatrix, Nothing} = nothing
end

James-Stein [10, 11] expected returns vector estimator.

Parameters

target: correction target for the estimator.

w: optional T×1 vector of weights for computing the simple expected returns vector.

sigma: value of the covariance matrix used for correcting the simple expected returns vector.

if computing with asset_statistics! and isnothing(sigma): use the covariance matrix computed by .

PortfolioOptimiser.MuSimple — Type

@kwdef mutable struct MuSimple <: MeanEstimator
    w::Union{<:AbstractWeights, Nothing} = nothing
end

Simple expected returns vector estimator.

Parameters

w: optional T×1 vector of weights for computing the expected returns vector.

PortfolioOptimiser.SE — Type

struct SE <: MeanTarget end

Mean square error of sample mean.

PortfolioOptimiser.VW — Type

struct VW <: MeanTarget end

Volatility-weighted grand mean.

PortfolioOptimiser.target_mean — Method

target_mean(::GM, mu::AbstractVector, sigma::AbstractMatrix, inv_sigma, T::Integer,
                     N::Integer)

Private

PortfolioOptimiser.MeanEstimator — Type

abstract type MeanEstimator end

Abstract type for subtyping expected returns vectors estimators.

PortfolioOptimiser.MeanSigmaEstimator — Type

abstract type MeanSigmaEstimator <: MeanEstimator end

Abstract type for subtyping expected returns vectors estimators that use covariance matrices for their corrections.

PortfolioOptimiser.MeanTarget — Type

abstract type MeanTarget end

Abstract type for subtyping correction targets of expected returns estimators that use covariance matrices for their corrections.