Ordered Weight Arrays

Public

PortfolioOptimiser.CRRA — Type

@kwdef mutable struct CRRA{T1 <: Real} <: OWAMethods
    g::T1 = 0.5
end

Normalised Constant Relative Risk Aversion Coefficients.

Parameters

g: Risk aversion coefficient.

PortfolioOptimiser.MaxEntropy — Type

@kwdef mutable struct MaxEntropy{T1 <: Real, T2 <: AbstractDict} <: OWAMethods
    max_phi::T1 = 0.5
    solvers::T2 = Dict()
end

Maximum Entropy. Solver must support MOI.RelativeEntropyCone and MOI.NormOneCone.

Parameters

max_phi: Maximum weight constraint of the L-moments.

PortfolioOptimiser.MinSqDist — Type

@kwdef mutable struct MinSqDist{T1 <: Real} <: OWAMethods
    max_phi::T1 = 0.5
end

Minimum Square Distance. Solver must support MOI.SecondOrderCone.

Parameters

max_phi: Maximum weight constraint of the L-moments.

PortfolioOptimiser.MinSumSq — Type

@kwdef mutable struct MinSumSq{T1 <: Real} <: OWAMethods
    max_phi::T1 = 0.5
end

Minimum Sum of Squares. Solver must support MOI.SecondOrderCone.

Parameters

max_phi: Maximum weight constraint of the L-moments.

PortfolioOptimiser.owa_cvar — Function

owa_cvar(T::Integer; alpha::Real = 0.05)

Calculate the OWA weights corresponding to the Critical Value at Risk (CVaR) of a returns series [OWA].

Inputs

Outputs

PortfolioOptimiser.owa_gmd — Method

owa_gmd(T::Integer)

Computes the Gini Mean Difference (GMD) of a returns series [OWA].

Inputs

Outputs

PortfolioOptimiser.owa_l_moment — Function

owa_l_moment(T::Integer; k::Integer = 2)

Calculates the OWA weights of the k'th linear moment (L-moment) of a returns series [12].

Inputs

k: order of the L-moment.

Outputs

PortfolioOptimiser.owa_l_moment_crm — Method

owa_l_moment_crm(T::Integer; k::Integer = 2, method::Symbol = :SD, g::Real = 0.5,
                 max_phi::Real = 0.5, solvers = Dict())

Compute the OWA weights for the convex risk measure considering higher order L-moments [12].

Inputs

k: order of the L-moment, k ≥ 2.

method: method for computing the weights used to combine L-moments higher than 2, used in OWAMethods.

:CRRA: Normalised Constant Relative Risk Aversion Coefficients.

:E: Maximum Entropy. Solver must support MOI.RelativeEntropyCone and MOI.NormOneCone.

:SS: Minimum Sum of Squares. Solver must support MOI.SecondOrderCone.

:SD: Minimum Square Distance. Solver must support MOI.SecondOrderCone.

g: the risk aversion coefficient.

max_phi: maximum weight constraint of the L-moments.

Outputs

PortfolioOptimiser.owa_rcvar — Method

owa_rcvar(T::Integer; alpha::Real = 0.05, beta::Real = alpha)

Compute the OWA weights for the CVaR Range of a returns series [OWA].

Inputs

Outputs

PortfolioOptimiser.owa_rg — Method

owa_rg(T::Integer)

Compute the OWA weights for the Range of a returns series [OWA].

Inputs

Outputs

PortfolioOptimiser.owa_rtg — Method

owa_rtg(T::Integer; alpha_i::Real = 0.0001, alpha::Real = 0.05, a_sim::Integer = 100,
        beta_i::Real = alpha_i, beta::Real = alpha, b_sim::Integer = a_sim)

Compute the OWA weights for the Tail Gini Range of a returns series [OWA].

Inputs

Outputs

PortfolioOptimiser.owa_rwcvar — Function

owa_rwcvar(T::Integer, alphas::AbstractVector{<:Real}, weights_a::AbstractVector{<:Real};
           betas::AbstractVector{<:Real} = alphas,
           weights_b::AbstractVector{<:Real} = weights_b)

Compute the OWA weights for the Weighted Conditional Value at Risk (WCVaR) of a returns series [OWA].

Inputs

alphas: N×1 vector of significance levels of the losses for each CVaR model, where N is the number of losses models, each .

weights_a: N×1 vector of weights of the losses for each CVaR model, where N is the number of losses models.

betas: M×1 vector of significance levels of the gains for each CVaR model, where M is the number of gains models, each .

weights_b: M×1 vector of weights of the gains for each CVaR model, where M is the number of gains models.

Outputs

PortfolioOptimiser.owa_tg — Method

owa_tg(T::Integer; alpha_i::Real = 0.0001, alpha::Real = 0.05, a_sim::Integer = 100)

Compute the OWA weights for the Tail Gini of a returns series [OWA].

Inputs

Outputs

PortfolioOptimiser.owa_wcvar — Method

owa_wcvar(T::Integer, alphas::AbstractVector{<:Real}, weights::AbstractVector{<:Real})

Compute the OWA weights for the Weighted Conditional Value at Risk (WCVaR) of a returns series [OWA].

Inputs

alphas: N×1 vector of significance levels of each CVaR model, where N is the number of models, each .

weights: N×1 vector of weights of each CVaR model, where N is the number of models.

Outputs

PortfolioOptimiser.owa_wr — Method

owa_wr(T::Integer)

Compute the OWA weights for the Worst Realisation (WR) of a returns series [OWA].

Inputs

Outputs

Private

PortfolioOptimiser.OWAMethods — Type

abstract type OWAMethods end

Abstract type for subtyping Ordered Weight Array (OWA) methods for computing the weights used to combine L-moments higher than 2 [12] in owa_l_moment_crm.

PortfolioOptimiser._crra_method — Method

_crra_method(weights::AbstractMatrix{<:Real}, k::Integer, g::Real)

Internal function for computing the Normalized Constant Relative Risk Aversion coefficients.

Inputs

weights: T×(k-1) matrix where T is the number of observations and k the order of the L-moments to combine, the i'th column contains the weights for the (i+1)'th L-moment.

k: the maximum order of the L-moments.

g: the risk aversion coefficient.

Outputs

w: T×1 ordered weight vector of the combined L-moments.

PortfolioOptimiser._optimise_JuMP_model — Method

_optimise_JuMP_model(model, solvers)

Internal function to optimise an OWA JuMP model.

Inputs

model: JuMP model.

Outputs

term_status: JuMP termination status.

solvers_tried: Dictionary that contains a dictionary of failed optimisations. Dict(key => Dict(...)), where key is the solver key used for the iteration of solver that failed.

If an MOI call fails on a model:

Dict(:jump_error => jump_error): JuMP error code.

If the optimiser fails to optimise the model satisfactorily:

Dict(:objective_val => JuMP.objective_value(model), :term_status => term_status, :params => haskey(val, :params) ? val[:params] : missing), where val is the value of the dictionary corresponding to key.

OWACajas, Dany, OWA Portfolio Optimization: A Disciplined Convex Programming Framework (December 18, 2021). Available at SSRN: https://ssrn.com/abstract=3988927 or http://dx.doi.org/10.2139/ssrn.3988927