Regression

Public

PortfolioOptimiser.AIC — Type

struct AIC <: MinValStepwiseRegressionCriteria end

Akaike's Information Criterion.

PortfolioOptimiser.AICC — Type

struct AICC <: MinValStepwiseRegressionCriteria end

Corrected Akaike's Information Criterion.

PortfolioOptimiser.AdjRSq — Type

struct AdjRSq <: MaxValStepwiseRegressionCriteria end

Adjusted R² for a linear model criterion.

PortfolioOptimiser.BIC — Type

struct BIC <: MinValStepwiseRegressionCriteria end

Bayesian Information Criterion.

PortfolioOptimiser.BReg — Type

@kwdef mutable struct BReg <: StepwiseRegression
    criterion::StepwiseRegressionCriteria = PVal(;)
end

Backward stepwise regression. Starts by assuming all features are significant and uses criterion to remove the worst performing one each iteration.

Parameters

criterion: criterion for feature selection.

isa(criterion, PVal): when no asset meets the selecion criterion, the list of significant features can be empty, in such cases the best factor is added to the list.

PortfolioOptimiser.FReg — Type

@kwdef mutable struct FReg <: StepwiseRegression
    criterion::StepwiseRegressionCriteria = PVal(;)
end

Forward stepwise regression. Starts by assuming no factor is significant and uses criterion to add the best performing one each iteration.

Parameters

criterion: criterion for feature selection.

isa(criterion, PVal): when no asset meets the selecion criterion, the list of significant features can be empty, in such cases the best factor is added to the list.

PortfolioOptimiser.FactorType — Type

@kwdef mutable struct FactorType
    error::Bool = true
    B::Union{Nothing, DataFrame} = nothing
    method::RegressionType = FReg(;)
    ve::StatsBase.CovarianceEstimator = SimpleVariance(;)
    var_w::Union{<:AbstractWeights, Nothing} = nothing
end

PortfolioOptimiser.PCAReg — Type

@kwdef mutable struct PCAReg <: DimensionReductionRegression
    ve::StatsBase.CovarianceEstimator = SimpleVariance(;)
    std_w::Union{<:AbstractWeights, Nothing} = nothing
    mean_w::Union{<:AbstractWeights, Nothing} = nothing
    target::DimensionReductionTarget = PCATarget(;)
end

PortfolioOptimiser.PCATarget — Type

@kwdef mutable struct PCATarget <: DimensionReductionTarget
    kwargs::NamedTuple = (;)
end

PortfolioOptimiser.PPCATarget — Type

@kwdef mutable struct PPCATarget <: DimensionReductionTarget
    kwargs::NamedTuple = (;)
end

PortfolioOptimiser.PVal — Type

@kwdef mutable struct PVal{T1 <: Real} <: StepwiseRegressionCriteria
    threshold::T1 = 0.05
end

P-value as feature selection criterion.

Parameters

threshold: threshold for classifying significant p-values. Only features whose p-values are lower than threshold are considered significant.

PortfolioOptimiser.RSq — Type

struct RSq <: MaxValStepwiseRegressionCriteria end

R² of a linear model criterion.

PortfolioOptimiser.loadings_matrix — Function

loadings_matrix(x::DataFrame, y::DataFrame, method::RegressionType = FReg())

Private

PortfolioOptimiser.DimensionReductionRegression — Type

abstract type DimensionReductionRegression <: RegressionType end

Abstract type for subtyping dimensionality reduction regression methods for computing the loadings matrix in loadings_matrix.

PortfolioOptimiser.DimensionReductionTarget — Type

abstract type DimensionReductionTarget end

PortfolioOptimiser.MaxValStepwiseRegressionCriteria — Type

abstract type MinValStepwiseRegressionCriteria <: StepwiseRegressionCriteria end

Abstract type for subtyping selection criteria where larger values are more significant.

PortfolioOptimiser.MinValStepwiseRegressionCriteria — Type

abstract type MinValStepwiseRegressionCriteria <: StepwiseRegressionCriteria end

Abstract type for subtyping selection criteria where smaller values are more significant.

PortfolioOptimiser.RegressionType — Type

abstract type RegressionType end

Abstract type for subtyping regression methods for computing the loadings matrix in loadings_matrix.

PortfolioOptimiser.StepwiseRegression — Type

abstract type StepwiseRegression <: RegressionType end

Abstract type for subtyping stepwise regression methods for computing the loadings matrix in loadings_matrix.

PortfolioOptimiser.StepwiseRegressionCriteria — Type

abstract type StepwiseRegressionCriteria end

Abstract type for subtyping selection criteria for selecting significant features when using StepwiseRegression methods.