Worst case mean variance sets

Public

PortfolioOptimiser.ArchWC — Type

@kwdef mutable struct ArchWC{T1 <: Integer, T2 <: Integer, T3 <: Real} <: WorstCaseMethod
    bootstrap::WorstCaseArchMethod = StationaryBS()
    n_sim::T1 = 3_000
    block_size::T2 = 3
    q::T3 = 0.05
    seed::Union{<:Integer, Nothing} = nothing
end

PortfolioOptimiser.Box — Type

struct Box <: WorstCaseSet end

Box sets for worst case mean variance optimisation.

PortfolioOptimiser.CircularBS — Type

struct CircularBS <: WorstCaseArchMethod end

Circular block bootstrap.

PortfolioOptimiser.DeltaWC — Type

@kwdef mutable struct DeltaWC{T1 <: Real, T2 <: Real} <: WorstCaseMethod
    dcov::T1 = 0.1
    dmu::T2 = 0.1
end

PortfolioOptimiser.Ellipse — Type

struct Ellipse <: WorstCaseSet end

Elliptical sets for worst case mean variance optimisation.

PortfolioOptimiser.KGeneralWC — Type

struct KGeneralWC <: WorstCaseKMethod end

PortfolioOptimiser.KNormalWC — Type

struct KNormalWC <: WorstCaseKMethod end

PortfolioOptimiser.MovingBS — Type

struct MovingBS <: WorstCaseArchMethod end

Moving block bootstrap.

PortfolioOptimiser.NoWC — Type

@kwdef mutable struct NoWC <: WorstCaseSet
    formulation::SDSquaredFormulation = SOCSD()
end

Use no set for worst case mean variance optimisation.

Parameters

formulation: quadratic expression formulation of SD risk measure to use SDSquaredFormulation.

PortfolioOptimiser.NormalWC — Type

@kwdef mutable struct NormalWC{T1 <: Integer, T2 <: Real} <: WorstCaseMethod
    n_sim::T1 = 3_000
    q::T2 = 0.05
    rng::AbstractRNG = Random.default_rng()
    seed::Union{<:Integer, Nothing} = nothing
end

PortfolioOptimiser.StationaryBS — Type

struct StationaryBS <: WorstCaseArchMethod end

Stationary bootstrap.

PortfolioOptimiser.WCType — Type

@kwdef mutable struct WCType
    cov_type::PortfolioOptimiserCovCor = PortCovCor(;)
    mu_type::MeanEstimator = MuSimple(;)
    box::WorstCaseMethod = NormalWC(;)
    ellipse::WorstCaseMethod = NormalWC(;)
    k_sigma::Union{<:Real, WorstCaseKMethod} = KNormalWC(;)
    k_mu::Union{<:Real, WorstCaseKMethod} = KNormalWC(;)
    posdef::PosdefFix = PosdefNearest(;)
    diagonal::Bool = false
end

Private

PortfolioOptimiser.WorstCaseArchMethod — Type

abstract type WorstCaseArchMethod <: WorstCaseMethod end

Abstract type for subtyping methods for computing worst case mean variance for bootstrapping with arch.

PortfolioOptimiser.WorstCaseKMethod — Type

abstract type WorstCaseKMethod end

PortfolioOptimiser.WorstCaseMethod — Type

abstract type WorstCaseMethod end

Abstract type for subtyping methods for computing worst case mean variance sets.

PortfolioOptimiser.WorstCaseSet — Type

abstract type WorstCaseSet end

Abstract type for subtyping worst case mean variance set types.