Parameter Estimation Types

Public

PortfolioOptimiser.CorDistance — Type

mutable struct CorDistance <: PortfolioOptimiserCovCor
    distance::Distances.UnionMetric
    dist_args::Tuple
    dist_kwargs::NamedTuple
    mean_w1::Union{<:AbstractWeights, Nothing}
    mean_w2::Union{<:AbstractWeights, Nothing}
    mean_w3::Union{<:AbstractWeights, Nothing}
end

PortfolioOptimiser.CorGerberSB1 — Type

mutable struct CorGerberSB1{T1, T2, T3, T4, T5} <: CorSB
    normalise::Bool
    threshold::T1
    c1::T2
    c2::T3
    c3::T4
    n::T5
    ve::StatsBase.CovarianceEstimator
    std_w::Union{<:AbstractWeights, Nothing}
    mean_w::Union{<:AbstractWeights, Nothing}
    posdef::PosdefFix
end

PortfolioOptimiser.CorLTD — Type

mutable struct CorLTD <: PortfolioOptimiserCovCor
    alpha::Real
    ve::StatsBase.CovarianceEstimator
    std_w::Union{<:AbstractWeights, Nothing}
end

PortfolioOptimiser.CorMutualInfo — Type

mutable struct CorMutualInfo <: PortfolioOptimiserCovCor
    bins::Union{<:Integer, <:AbstractBins}
    normalise::Bool
    ve::StatsBase.CovarianceEstimator
    std_w::Union{<:AbstractWeights, Nothing}
end

PortfolioOptimiser.DistanceCanonical — Type

struct DistanceCanonical <: DistanceMethod end

Struct for computing the canonical distance for a given correlation subtype of PortfolioOptimiserCovCor.

PortfolioOptimiser.DistanceLog — Type

struct DistanceLog <: DistanceMethod end

Defines the log-distance matrix from the correlation matrix.

\[D_{i,\,j} = -\log\left(C_{i,\,j}\right)\,.\]

Where:

$D_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N log-distance matrix.

$C_{i,\,j}$: is the $(i,\,j)$-th entry of an absolute correlation matrix.

PortfolioOptimiser.DistanceMLP — Type

@kwdef mutable struct DistanceMLP <: DistanceMethod
    absolute::Bool = false
end

Defines the distance matrix from a correlation matrix [1] in dist.

\[D_{i,\,j} = \begin{cases} \sqrt{\dfrac{1}{2} \left(\mathbf{1} - C_{i,\,j}\right)} &\quad \mathrm{if~ absolute = false}\\ \sqrt{1 - \lvert C_{i,\,j} \rvert} &\quad \mathrm{if~ absolute = true}\nonumber\,. \end{cases}\]

Where:

$D_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N distance matrix $\mathbf{C}$.

$C_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N correlation matrix $\mathbf{D}$.

absolute: flag for stating whether or not an absolute correlation is being used.

Parameters

absolute: flag for stating whether or not an absolute correlation is being used.

PortfolioOptimiser.DistanceSqMLP — Type

@kwdef mutable struct DistanceSqMLP <: DistanceMethod
    absolute::Bool = false
    distance::Distances.UnionMetric = Distances.Euclidean()
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the distance of distances matrix from a correlation matrix [1] in dist.

\[\tilde{D}_{i,\,j} = f_{m}\left(\bm{D}_{i},\, \bm{D}_j\right)\,.\]

Where:

$\bm{D}_{i}$: is the $i$-th column/row of the N×N distance matrix defined in DistanceMLP.

$f_{m}$: is the pairwise distance function for metric $m$. We use the Distances.pairwise function which computes the entire matrix at once, the output is a vector so it gets reshaped into an N×N matrix.

$\tilde{D}_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N distances of distances matrix.

absolute: is a flag whether the correlation is absolute.

Parameters

absolute: flag for stating whether or not an absolute correlation is being used.

distance: distance metric from Distances.jl.

args: args for the Distances.pairwise function.

kwargs: key word args for the Distances.pairwise function.

PortfolioOptimiser.Fixed — Type

@kwdef mutable struct Fixed{T1, T2, T3, T4} <: Denoise
    detone::Bool = false
    mkt_comp::Integer = 1
    kernel = AverageShiftedHistograms.Kernels.gaussian
    m::Integer = 10
    n::Integer = 1000
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the parameters for using the fixed method in denoise! [2, Chapter 2].

Parameters

detone:

true: remove the largest mkt_comp eigenvalues from the correlation matrix.

Warning

Removing eigenvalues from the matrix may make it singular.

mkt_comp: the number of largest eigenvalues to remove from the correlation matrix.

kernel: kernel for fitting the average shifted histograms from AverageShiftedHistograms.jl Kernel Functions.

m: number of adjacent histograms to smooth over AverageShiftedHistograms.jl Usage.

n: number of points used when creating the range of values to which the average shifted histogram is to be fitted AverageShiftedHistograms.jl Usage.

args: arguments for Optim.optimize

kwargs: keyword arguments for Optim.optimize

PortfolioOptimiser.GM — Type

struct GM <: MeanTarget end

PortfolioOptimiser.MuBOP — Type

@kwdef mutable struct MuBOP{T1} <: MeanEstimator
    target::MeanTarget = GM()
    w::Union{<:AbstractWeights, Nothing} = nothing
    sigma::T1 = Matrix{Float64}(undef, 0, 0)
end

PortfolioOptimiser.MuBS — Type

@kwdef mutable struct MuBS{T1} <: MeanEstimator
    target::MeanTarget = GM()
    w::Union{<:AbstractWeights, Nothing} = nothing
    sigma::T1 = Matrix{Float64}(undef, 0, 0)
end

PortfolioOptimiser.MuJS — Type

@kwdef mutable struct MuJS{T1} <: MeanEstimator
    target::MeanTarget = GM()
    w::Union{<:AbstractWeights, Nothing} = nothing
    sigma::T1 = Matrix{Float64}(undef, 0, 0)
end

PortfolioOptimiser.MuSimple — Type

@kwdef mutable struct MuSimple <: MeanEstimator
    w::Union{<:AbstractWeights, Nothing} = nothing
end

PortfolioOptimiser.NoDenoise — Type

struct NoDenoise <: Denoise end

No denoising is performed in denoise!.

PortfolioOptimiser.NoPosdef — Type

struct NoPosdef <: PosdefFix end

Non positive definite matrices will not be fixed.

PortfolioOptimiser.PortCovCor — Type

@kwdef mutable struct PortCovCor <: PortfolioOptimiserCovCor
    ce::CovarianceEstimator = CovFull(;)
    posdef::PosdefFix = PosdefNearest(;)
    denoise::Denoise = NoDenoise(;)
    jlogo::AbstractJLoGo = NoJLoGo(;)
end

PortfolioOptimiser.PosdefNearest — Type

@kwdef mutable struct PosdefNearest <: PosdefFix
    method::NearestCorrelationMatrix.NCMAlgorithm = NearestCorrelationMatrix.Newton(;
                                                                                    tau = 1e-12)
end

Defines which method from NearestCorrelationMatrix to use in posdef_fix!.

PortfolioOptimiser.SE — Type

struct SE <: MeanTarget end

PortfolioOptimiser.Shrink — Type

mutable struct Shrink{T1, T2, T3, T4, T5} <: Denoise
    detone::Bool = false
    mkt_comp::Integer = 1
    kernel = AverageShiftedHistograms.Kernels.gaussian
    m::Integer = 10
    n::Integer = 1000
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the parameters for using the shrink method in denoise! [2, Chapter 2].

Parameters

detone:

true: take only the largest mkt_comp eigenvalues from the correlation matrix.

Warning

Removing eigenvalues from the matrix may make it singular.

mkt_comp: the number of largest eigenvalues to keep from the correlation matrix.

kernel: kernel for fitting the average shifted histograms from AverageShiftedHistograms.jl Kernel Functions.

m: number of adjacent histograms to smooth over AverageShiftedHistograms.jl Usage.

n: number of points used when creating the range of values to which the average shifted histogram is to be fitted AverageShiftedHistograms.jl Usage.

args: arguments for Optim.optimize

kwargs: keyword arguments for Optim.optimize

PortfolioOptimiser.Spectral — Type

@kwdef mutable struct Spectral{T1, T2, T3, T4} <: Denoise
    detone::Bool = false
    mkt_comp::Integer = 1
    kernel = AverageShiftedHistograms.Kernels.gaussian
    m::Integer = 10
    n::Integer = 1000
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the parameters for using the spectral method in denoise! [2, Chapter 2].

Parameters

detone:

true: take only the largest mkt_comp eigenvalues from the correlation matrix.

Warning

Removing eigenvalues from the matrix may make it singular.

mkt_comp: the number of largest eigenvalues to keep from the correlation matrix.

kernel: kernel for fitting the average shifted histograms from AverageShiftedHistograms.jl Kernel Functions.

m: number of adjacent histograms to smooth over AverageShiftedHistograms.jl Usage.

n: number of points used when creating the range of values to which the average shifted histogram is to be fitted AverageShiftedHistograms.jl Usage.

args: arguments for Optim.optimize

kwargs: keyword arguments for Optim.optimize

PortfolioOptimiser.VW — Type

struct VW <: MeanTarget end

Private

PortfolioOptimiser.Denoise — Type

abstract type Denoise end

Abstract type for subtyping denoising methods.

PortfolioOptimiser.DistanceMethod — Type

abstract type DistanceMethod end

Abstract type for subtyping methods for computing distance matrices from correlation ones.

PortfolioOptimiser.MeanEstimator — Type

abstract type MeanEstimator end

PortfolioOptimiser.MeanTarget — Type

abstract type MeanTarget end

PortfolioOptimiser.OWAMethods — Type

OWAMethods = (:CRRA, :E, :SS, :SD)

Methods for computing the weights used to combine L-moments higher than 2 [3].

:CRRA: Normalised Constant Relative Risk Aversion Coefficients.

:E: Maximum Entropy. Solver must support MOI.RelativeEntropyCone and MOI.NormOneCone.

:SS: Minimum Sum of Squares. Solver must support MOI.SecondOrderCone.

:SD: Minimum Square Distance. Solver must support MOI.SecondOrderCone.

PortfolioOptimiser.PortfolioOptimiserCovCor — Type

abstract type PortfolioOptimiserCovCor <: StatsBase.CovarianceEstimator end

PortfolioOptimiser.PosdefFix — Type

abstract type PosdefFix end

Abstract type for subtyping methods for fixing non positive definite matrices.