Portfolio Optimisation

Public API

PortfolioOptimiser.Portfolio — Method

Portfolio(;
    # Portfolio characteristics.
    prices::TimeArray = TimeArray(TimeType[], []),
    returns::DataFrame = DataFrame(),
    ret::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    timestamps::AbstractVector = Vector{Date}(undef, 0),
    assets::AbstractVector = Vector{String}(undef, 0),
    short::Bool = false,
    short_u::Real = 0.2,
    long_u::Real = 1.0,
    min_number_effective_assets::Integer = 0,
    max_number_assets::Integer = 0,
    max_number_assets_factor::Real = 100_000.0,
    f_prices::TimeArray = TimeArray(TimeType[], []),
    f_returns::DataFrame = DataFrame(),
    f_ret::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    f_timestamps::AbstractVector = Vector{Date}(undef, 0),
    f_assets::AbstractVector = Vector{String}(undef, 0),
    loadings::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    # Risk parameters.
    msv_target::Union{<:Real, AbstractVector{<:Real}} = Inf,
    lpm_target::Union{<:Real, AbstractVector{<:Real}} = Inf,
    alpha_i::Real = 0.0001,
    alpha::Real = 0.05,
    a_sim::Integer = 100,
    beta_i::Real = Inf,
    beta::Real = Inf,
    b_sim::Integer = 0,
    kappa::Real = 0.3,
    gs_threshold::Real = 0.5,
    max_num_assets_kurt::Integer = 0,
    # Benchmark constraints.
    turnover::Real = Inf,
    turnover_weights::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
    kind_tracking_err::Symbol = :Weights,
    tracking_err::Real = Inf,
    tracking_err_returns::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
    tracking_err_weights::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
    bl_bench_weights::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
    # Risk and return constraints.
    a_mtx_ineq::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    b_vec_ineq::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
    risk_budget::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
    mu_l::Real = Inf,
    dev_u::Real = Inf,
    mad_u::Real = Inf,
    sdev_u::Real = Inf,
    cvar_u::Real = Inf,
    wr_u::Real = Inf,
    flpm_u::Real = Inf,
    slpm_u::Real = Inf,
    mdd_u::Real = Inf,
    add_u::Real = Inf,
    cdar_u::Real = Inf,
    uci_u::Real = Inf,
    evar_u::Real = Inf,
    edar_u::Real = Inf,
    gmd_u::Real = Inf,
    rg_u::Real = Inf,
    rcvar_u::Real = Inf,
    tg_u::Real = Inf,
    rtg_u::Real = Inf,
    owa_u::Real = Inf,
    owa_w::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
    krt_u::Real = Inf,
    skrt_u::Real = Inf,
    rvar_u::Real = Inf,
    rdar_u::Real = Inf,
    # Optimisation model inputs.
    mu_type::Symbol = :Default,
    mu::AbstractVector = Vector{Float64}(undef, 0),
    cov_type::Symbol = :Full,
    jlogo::Bool = false,
    cov::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    kurt::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    skurt::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    posdef_fix::Symbol = :None,
    L_2::AbstractMatrix = SparseMatrixCSC{Float64, Int}(undef, 0, 0),
    S_2::AbstractMatrix = SparseMatrixCSC{Float64, Int}(undef, 0, 0),
    mu_f::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
    cov_f::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    mu_fm::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
    cov_fm::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    mu_bl::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
    cov_bl::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    mu_bl_fm::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
    cov_bl_fm::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    returns_fm::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    z::AbstractDict = Dict(),
    # Inputs of Worst Case Optimization Models.
    cov_l::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    cov_u::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    cov_mu::AbstractMatrix{<:Real} = Diagonal{Float64}(undef, 0),
    cov_sigma::AbstractMatrix{<:Real} = Diagonal{Float64}(undef, 0),
    d_mu::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
    k_mu::Real = Inf,
    k_sigma::Real = Inf,
    # Optimal portfolios.
    optimal::AbstractDict = Dict(),
    limits::DataFrame = DataFrame(),
    frontier::AbstractDict = Dict(),
    # Solutions.
    solvers::AbstractDict = Dict(),
    opt_params::AbstractDict = Dict(),
    fail::AbstractDict = Dict(),
    model = JuMP.Model(),
    # Allocation.
    latest_prices::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
    alloc_optimal::AbstractDict = Dict(),
    alloc_solvers::AbstractDict = Dict(),
    alloc_params::AbstractDict = Dict(),
    alloc_fail::AbstractDict = Dict(),
    alloc_model::AbstractDict = Dict(),
)

Creates an instance of Portfolio containing all internal data necessary for convex portfolio optimisations as well as failed and successful results.

Inputs

Portfolio characteristics.

prices: (T+1)×Na TimeArray with asset pricing information where the time stamp field is timestamp, T is the number of returns observations and Na the number of assets. If prices is not empty, then returns, ret, timestamps, assets and latest_prices are ignored because their respective fields are obtained from prices.

returns: T×(Na+1) DataFrame where T is the number of returns observations, Na is the number of assets, the extra column is timestamp, which contains the timestamps of the returns. If prices is empty and returns is not empty, ret, timestamps and assets are ignored because their respective fields are obtained from returns.

ret: T×Na matrix of returns. Its value is saved in the returns field of Portfolio. If prices or returns are not empty, this value is obtained from within the function.

timestamps: T×1 vector of timestamps. Its value is saved in the timestamps field of Portfolio. If prices or returns are not empty, this value is obtained from within the function.

assets: Na×1 vector of assets. Its value is saved in the assets field of Portfolio. If prices or returns are not empty, this value is obtained from within the function.

short: whether or not to allow negative weights, i.e. whether a portfolio accepts shorting.

short_u: absolute value of the sum of all short (negative) weights.

long_u: sum of all long (positive) weights.

min_number_effective_assets: if non-zero, guarantees that at least number of assets make significant contributions to the final portfolio weights.

max_number_assets: if non-zero, guarantees at most this number of assets make non-zero contributions to the final portfolio weights. Requires an optimiser that supports binary variables.

max_number_assets_factor: scaling factor needed to create a decision variable when max_number_assets is non-zero.

f_prices: (T+1)×Nf TimeArray with factor pricing information where the time stamp field is f_timestamp, T is the number of factors returns observations and Nf the number of factors. If f_prices is not empty, then f_returns, f_ret, f_timestamps, f_assets and latest_prices are ignored because their respective fields are obtained from f_prices.

f_returns: T×(Nf+1) DataFrame where T is the number of factor returns observations, Nf is the number of factors, the extra column is f_timestamp, which contains the timestamps of the factor returns. If f_prices is empty and f_returns is not empty, f_ret, f_timestamps and f_assets are ignored because their respective fields are obtained from f_returns.

f_ret: T×Nf matrix of factor returns. Its value is saved in the f_returns field of Portfolio. If f_prices or f_returns are not empty, this value is obtained from within the function.

f_timestamps: T×1 vector of factor timestamps. Its value is saved in the f_timestamps field of Portfolio. If f_prices or f_returns are not empty, this value is obtained from within the function.

f_assets: Nf×1 vector of assets. Its value is saved in the f_assets field of Portfolio. If f_prices or f_returns are not empty, this value is obtained from within the function.

loadings: loadings matrix for black litterman models.

Risk parameters.

msv_target: target value for for mean absolute deviation and mean semivariance risk measures. It can have two meanings depending on its type and value.

If it's a Real number and infinite, or an empty vector. The target will be the mean returns vector mu.

Else the target is the value of msv_target. If msv_target is a vector, its length should be Na, where Na is the number of assets.

lpm_target: target value for the first and second lower partial moment risk measures. It can have two meanings depending on its type and value.

If it's a Real number and infinite, or an empty vector. The target will be the value of the risk free rate rf, provided in the opt_port! function.

Else the target is the value of lpm_target. If lpm_target is a vector, its length should be Na, where Na is the number of assets.

alpha_i: initial significance level of Tail Gini losses, alpha_i < alpha.

alpha: significance level of CVaR losses, alpha ∈ (0, 1).

a_sim: number of CVaRs to approximate the Tail Gini losses.

beta_i: initial significance level of Tail Gini gains.

beta:

b_sim: number of CVaRs to approximate the Tail Gini gains.

kappa:

gs_threshold:

max_num_assets_kurt:

Benchmark constraints.

turnover:

turnover_weights:

kind_tracking_err:

tracking_err:

tracking_err_returns:

tracking_err_weights:

bl_bench_weights:

Risk and return constraints.

a_mtx_ineq:

b_vec_ineq:

risk_budget:

mu_l:

dev_u:

mad_u:

sdev_u:

cvar_u:

wr_u:

flpm_u:

slpm_u:

mdd_u:

add_u:

cdar_u:

uci_u:

evar_u:

edar_u:

gmd_u:

rg_u:

rcvar_u:

tg_u:

rtg_u:

owa_u:

owa_w:

krt_u:

skrt_u:

rvar_u:

rdar_u:

Optimisation model inputs.

mu_type:

mu:

cov_type:

jlogo:

cov:

kurt:

skurt:

posdef_fix:

L_2:

S_2:

mu_f:

cov_f:

mu_fm:

cov_fm:

mu_bl:

cov_bl:

mu_bl_fm:

cov_bl_fm:

returns_fm:

z:

Inputs of Worst Case Optimization Models.

cov_l:

cov_u:

cov_mu:

cov_sigma:

d_mu:

k_mu:

k_sigma:

Optimal portfolios.

optimal:

limits:

frontier:

Solutions.

solvers:

opt_params:

fail:

model:

Allocation.

latest_prices:

alloc_optimal:

alloc_solvers:

alloc_params:

alloc_fail:

alloc_model:

PortfolioOptimiser.opt_port! — Function

opt_port!(
    portfolio::Portfolio;
    class::Symbol = :Classic,
    hist::Integer = 1,
    kelly::Symbol = :None,
    l::Real = 2.0,
    obj::Symbol = :Sharpe,
    rf::Real = 0.0,#1.0329^(1 / 252) - 1
    rm::Symbol = :SD,
    rrp_penalty::Real = 1.0,
    rrp_ver::Symbol = :None,
    save_opt_params::Bool = true,
    string_names::Bool = false,
    type::Symbol = :Trad,
    u_cov::Symbol = :Box,
    u_mu::Symbol = :Box,
)

PortfolioOptimiser.HCPortfolio — Method

HCPortfolio(;
    # Portfolio characteristics.
    prices::TimeArray = TimeArray(TimeType[], []),
    returns::DataFrame = DataFrame(),
    ret::Matrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    timestamps::Vector{<:Dates.AbstractTime} = Vector{Date}(undef, 0),
    assets::AbstractVector = Vector{String}(undef, 0),
    # Risk parmeters.
    alpha_i::Real = 0.0001,
    alpha::Real = 0.05,
    a_sim::Integer = 100,
    beta_i::Real = Inf,
    beta::Real = Inf,
    b_sim::Integer = 0,
    kappa::Real = 0.3,
    alpha_tail::Real = 0.05,
    gs_threshold::Real = 0.5,
    owa_w::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
    # Optimisation parameters.
    mu_type::Symbol = :Default,
    mu::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
    cov_type::Symbol = :Full,
    jlogo::Bool = false,
    cov::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    posdef_fix::Symbol = :None,
    bins_info::Union{Symbol, <:Integer} = :KN,
    w_min::Union{<:Real, AbstractVector{<:Real}} = 0.0,
    w_max::Union{<:Real, AbstractVector{<:Real}} = 1.0,
    codep_type::Symbol = :Pearson,
    codep::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    dist::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    clusters = Hclust{Float64}(Matrix{Int64}(undef, 0, 2), Float64[], Int64[], :nothing),
    k::Integer = 0,
    # Optimal portfolios.
    optimal::AbstractDict = Dict(),
    # Solutions.
    solvers::AbstractDict = Dict(),
    opt_params::AbstractDict = Dict(),
    fail::AbstractDict = Dict(),
    # Allocation.
    latest_prices::AbstractVector = Vector{Float64}(undef, 0),
    alloc_optimal::AbstractDict = Dict(),
    alloc_solvers::AbstractDict = Dict(),
    alloc_params::AbstractDict = Dict(),
    alloc_fail::AbstractDict = Dict(),
    alloc_model::AbstractDict = Dict(),
)

Private API

PortfolioOptimiser.AbstractPortfolio — Type

AbstractPortfolio

Abstract type for portfolios. Concrete portfolios subtype this see Portfolio and HCPortfolio.

PortfolioOptimiser.Portfolio — Type

mutable struct Portfolio{
    # Portfolio characteristics
    ast,
    dat,
    r,
    s,
    us,
    ul,
    ssl,
    mnea,
    mna,
    mnaf,
    tfa,
    tfdat,
    tretf,
    l,
    # Risk parameters
    msvt,
    lpmt,
    ai,
    a,
    as,
    tat,
    bi,
    b,
    bs,
    k,
    tiat,
    lnk,
    topk,
    tomk,
    tk2,
    tkinv,
    tinvopk,
    tinvomk,
    gst,
    mnak,
    # Benchmark constraints
    to,
    tobw,
    kte,
    te,
    rbi,
    bw,
    blbw,
    # Risk and return constraints
    ami,
    bvi,
    rbv,
    ler,
    ud,
    umad,
    usd,
    ucvar,
    uwr,
    uflpm,
    uslpm,
    umd,
    uad,
    ucdar,
    uuci,
    uevar,
    uedar,
    ugmd,
    ur,
    urcvar,
    utg,
    urtg,
    uowa,
    wowa,
    uk,
    usk,
    urvar,
    urdar,
    # Optimisation model inputs
    ttmu,
    tmu,
    ttcov,
    tjlogo,
    tcov,
    tkurt,
    tskurt,
    tpdf,
    tl2,
    ts2,
    tmuf,
    tcovf,
    tmufm,
    tcovfm,
    tmubl,
    tcovbl,
    tmublf,
    tcovblf,
    trfm,
    tz,
    tcovl,
    tcovu,
    tcovmu,
    tcovs,
    tdmu,
    tkmu,
    tks,
    topt,
    tlim,
    tfront,
    tsolv,
    tf,
    toptpar,
    tmod,
    # Allocation
    tlp,
    taopt,
    tasolv,
    taoptpar,
    taf,
    tamod,
} <: AbstractPortfolio
    # Portfolio characteristics.
    assets::ast
    timestamps::dat
    returns::r
    short::s
    short_u::us
    long_u::ul
    sum_short_long::ssl
    min_number_effective_assets::mnea
    max_number_assets::mna
    max_number_assets_factor::mnaf
    f_assets::tfa
    f_timestamps::tfdat
    f_returns::tretf
    loadings::l
    # Risk parameters.
    msv_target::msvt
    lpm_target::lpmt
    alpha_i::ai
    alpha::a
    a_sim::as
    at::tat
    beta_i::bi
    beta::b
    b_sim::bs
    kappa::k
    invat::tiat
    ln_k::lnk
    opk::topk
    omk::tomk
    invk2::tk2
    invk::tkinv
    invopk::tinvopk
    invomk::tinvomk
    gs_threshold::gst
    max_num_assets_kurt::mnak
    # Benchmark constraints.
    turnover::to
    turnover_weights::tobw
    kind_tracking_err::kte
    tracking_err::te
    tracking_err_returns::rbi
    tracking_err_weights::bw
    bl_bench_weights::blbw
    # Risk and return constraints.
    a_mtx_ineq::ami
    b_vec_ineq::bvi
    risk_budget::rbv
    mu_l::ler
    dev_u::ud
    mad_u::umad
    sdev_u::usd
    cvar_u::ucvar
    wr_u::uwr
    flpm_u::uflpm
    slpm_u::uslpm
    mdd_u::umd
    add_u::uad
    cdar_u::ucdar
    uci_u::uuci
    evar_u::uevar
    edar_u::uedar
    gmd_u::ugmd
    rg_u::ur
    rcvar_u::urcvar
    tg_u::utg
    rtg_u::urtg
    owa_u::uowa
    owa_w::wowa
    krt_u::uk
    skrt_u::usk
    rvar_u::urvar
    rdar_u::urdar
    # Optimisation model inputs.
    mu_type::ttmu
    mu::tmu
    cov_type::ttcov
    jlogo::tjlogo
    cov::tcov
    kurt::tkurt
    skurt::tskurt
    posdef_fix::tpdf
    L_2::tl2
    S_2::ts2
    mu_f::tmuf
    cov_f::tcovf
    mu_fm::tmufm
    cov_fm::tcovfm
    mu_bl::tmubl
    cov_bl::tcovbl
    mu_bl_fm::tmublf
    cov_bl_fm::tcovblf
    returns_fm::trfm
    z::tz
    # Inputs of Worst Case Optimization Models.
    cov_l::tcovl
    cov_u::tcovu
    cov_mu::tcovmu
    cov_sigma::tcovs
    d_mu::tdmu
    k_mu::tkmu
    k_sigma::tks
    # Optimal portfolios.
    optimal::topt
    limits::tlim
    frontier::tfront
    # Solver params.
    solvers::tsolv
    opt_params::toptpar
    fail::tf
    model::tmod
    # Allocation
    latest_prices::tlp
    alloc_optimal::taopt
    alloc_solvers::tasolv
    alloc_params::taoptpar
    alloc_fail::taf
    alloc_model::tamod
end

Structure for optimising portfolios.

Fieldnames

Portfolio characteristics.

PortfolioOptimiser.HCPortfolio — Type

mutable struct HCPortfolio{
    ast,
    dat,
    r,
    # Risk parmeters.
    ai,
    a,
    as,
    bi,
    b,
    bs,
    k,
    ata,
    gst,
    # Optimisation parameters.
    tymu,
    tmu,
    tycov,
    tcov,
    tbin,
    wmi,
    wma,
    ttco,
    tco,
    tdist,
    tcl,
    tk,
    # Optimal portfolios.
    topt,
    # Solutions.
    tsolv,
    toptpar,
    tf,
} <: AbstractPortfolio
    # Portfolio characteristics.
    assets::ast
    timestamps::dat
    returns::r
    # Risk parmeters.
    alpha_i::ai
    alpha::a
    a_sim::as
    beta_i::bi
    beta::b
    b_sim::bs
    kappa::k
    alpha_tail::ata
    gs_threshold::gst
    # Optimisation parameters.
    mu_type::tymu
    mu::tmu
    cov_type::tycov
    jlogo::tjlogo
    cov::tcov
    bins_info::tbin
    w_min::wmi
    w_max::wma
    codep_type::ttco
    codep::tco
    dist::tdist
    clusters::tcl
    k::tk
    # Optimal portfolios.
    optimal::topt
    # Solutions.
    solvers::tsolv
    opt_params::toptpar
    fail::tf
end

HCPortfolio(;
    # Portfolio characteristics.
    returns = DataFrame(),
    ret::Matrix{<:Real} = Matrix{Float64}(undef, 0, 0),
    timestamps::Vector{<:Dates.AbstractTime} = Vector{Date}(undef, 0),
    assets = Vector{String}(undef, 0),
    # Risk parmeters.
    alpha_i::Real = 0.0001,
    alpha::Real = 0.05,
    a_sim::Integer = 100,
    beta_i::Real = Inf,
    beta::Real = Inf,
    b_sim::Integer = 0,
    kappa::Real = 0.3,
    alpha_tail::Real = 0.05,
    gs_threshold::Real = 0.5,
    # Optimisation parameters.
    mu_type::Symbol = :Default,
    mu = Vector{Float64}(undef, 0),
    cov_type::Symbol = :Full,
    tjlogo::Bool = false,
    cov = Matrix{Float64}(undef, 0, 0),
    bins_info::Union{Symbol, Int} = :KN,
    w_min::Union{<:Real, AbstractVector{<:Real}} = 0.0,
    w_max::Union{<:Real, AbstractVector{<:Real}} = 1.0,
    codep_type::Symbol = :Pearson,
    codep = Matrix{Float64}(undef, 0, 0),
    dist = Matrix{Float64}(undef, 0, 0),
    clusters = Hclust{Float64}(Matrix{Int64}(undef, 0, 2), Float64[], Int64[], :nothing),
    k::Union{Int, Nothing} = nothing,
    # Solutions.
    solvers::AbstractDict = Dict(),
    opt_params::AbstractDict = Dict(),
    fail::AbstractDict = Dict(),
)

PortfolioOptimiser.RiskMeasures — Constant

RiskMeasures = (
    :SD,
    :MAD,
    :SSD,
    :FLPM,
    :SLPM,
    :WR,
    :CVaR,
    :EVaR,
    :RVaR,
    :MDD,
    :ADD,
    :CDaR,
    :UCI,
    :EDaR,
    :RDaR,
    :Kurt,
    :SKurt,
    :GMD,
    :RG,
    :RCVaR,
    :TG,
    :RTG,
    :OWA,
)

Available risk measures for :Trad and :RP type (see PortTypes) of Portfolio.

:SD = standard deviation (SD).

:MAD = max absolute deviation (MAD).

:SSD = semi standard deviation (SSD).

:FLPM = first lower partial moment (omega ratio) (FLPM).

:SLPM = second lower partial moment (sortino ratio) (SLPM).

:WR = worst realisation (WR).

:CVaR = conditional value at risk (CVaR).

:EVaR = entropic value at risk (EVaR).

:RVaR = relativistic value at risk (RVaR).

:MDD = maximum drawdown of uncompounded cumulative returns (MDD_abs).

:ADD = average drawdown of uncompounded cumulative returns (ADD_abs).

:CDaR = conditional drawdown at risk of uncompounded cumulative returns (CDaR_abs).

:UCI = ulcer index of uncompounded cumulative returns (UCI_abs).

:EDaR = entropic drawdown at risk of uncompounded cumulative returns (EDaR_abs).

:RDaR = relativistic drawdown at risk of uncompounded cumulative returns (RDaR_abs).

:Kurt = square root kurtosis (Kurt).

:SKurt = square root semi-kurtosis (SKurt).

:GMD = gini mean difference (GMD).

:RG = range of returns (RG).

:RCVaR = range of conditional value at risk (RCVaR).

:TG = tail gini (TG).

:RTG = range of tail gini (RTG).

:OWA = ordered weight array (generic OWA weights) (OWA).

PortfolioOptimiser.HRRiskMeasures — Constant

HRRiskMeasures = (
    :SD,
    :MAD,
    :SSD,
    :FLPM,
    :SLPM,
    :WR,
    :CVaR,
    :EVaR,
    :RVaR,
    :MDD,
    :ADD,
    :CDaR,
    :UCI,
    :EDaR,
    :RDaR,
    :Kurt,
    :SKurt,
    :GMD,
    :RG,
    :RCVaR,
    :TG,
    :RTG,
    :OWA,
    :Variance,
    :Equal,
    :VaR,
    :DaR,
    :DaR_r,
    :MDD_r,
    :ADD_r,
    :CDaR_r,
    :EDaR_r,
    :RDaR_r,
)

Available risk measures for optimisations of HCPortfolio.

:SD = standard deviation (SD).

:MAD = max absolute deviation (MAD).

:SSD = semi standard deviation (SSD).

:FLPM = first lower partial moment (Omega ratio) (FLPM).

:SLPM = second lower partial moment (Sortino ratio) (SLPM).

:WR = worst realisation (WR).

:CVaR = conditional value at risk (CVaR).

:EVaR = entropic value at risk (EVaR).

:RVaR = relativistic value at risk (RVaR).

:MDD = maximum drawdown of uncompounded cumulative returns (Calmar ratio) (MDD_abs).

:ADD = average drawdown of uncompounded cumulative returns (ADD_abs).

:CDaR = conditional drawdown at risk of uncompounded cumulative returns (CDaR_abs).

:UCI = ulcer index of uncompounded cumulative returns (UCI_abs).

:EDaR = entropic drawdown at risk of uncompounded cumulative returns (EDaR_abs).

:RDaR = relativistic drawdown at risk of uncompounded cumulative returns (RDaR_abs).

:Kurt = square root kurtosis (Kurt).

:SKurt = square root semi-kurtosis (SKurt).

:GMD = gini mean difference (GMD).

:RG = range of returns (RG).

:RCVaR = range of conditional value at risk (RCVaR).

:TG = tail gini (TG).

:RTG = range of tail gini (RTG).

:OWA = ordered weight array (generic OWA weights) (OWA).

:Variance = variance (Variance).

:Equal = equal risk contribution, 1/N where N is the number of assets.

:VaR = value at risk (VaR).

:DaR = drawdown at risk of uncompounded cumulative returns (DaR_abs).

:DaR_r = drawdown at risk of compounded cumulative returns (DaR_rel).

:MDD_r = maximum drawdown of compounded cumulative returns (MDD_rel).

:ADD_r = average drawdown of compounded cumulative returns (ADD_rel).

:CDaR_r = conditional drawdown at risk of compounded cumulative returns (CDaR_rel).

:UCI_r = ulcer index of compounded cumulative returns (UCI_rel).

:EDaR_r = entropic drawdown at risk of compounded cumulative returns (EDaR_rel).

:RDaR_r = relativistic drawdown at risk of compounded cumulative returns (RDaR_rel).