Clustering

Public

PortfolioOptimiser.cluster_assets! — Method

cluster_assets!(port::HCPortfolio; hclust_alg::HClustAlg = HAC(),
                hclust_opt::HCOpt = HCOpt())

Hierarchically cluster the assets in a hierarchical portfolio HCPortfolio using the covariance and distance matrices stored in the portfolio. Save the results in the portfolio. See the arguments types' docs for details.

Inputs

port: hierarchical clustering portfolio HCPortfolio.

hclust_alg: hierarchical clustering algorithm HClustAlg.

hclust_opt: options for determining the number of clusters HCOpt.

PortfolioOptimiser.cluster_assets — Method

cluster_assets(port::HCPortfolio; hclust_alg::HClustAlg = HAC(),
               hclust_opt::HCOpt = HCOpt())

Hierarchically cluster the assets in a hierarchical portfolio HCPortfolio using the covariance and distance matrices stored in the portfolio. See the arguments types' docs for details.

Inputs

port: hierarchical clustering portfolio HCPortfolio.

hclust_alg: hierarchical clustering algorithm HClustAlg.

hclust_opt: options for determining the number of clusters HCOpt.

Outputs

idx: clustering assignments after cutting the tree into k levels Clustering.cutree.

clustering: Clustering.Hclust of the portfolio assets.

k: optimum number of clusters.

PortfolioOptimiser.cluster_assets — Method

cluster_assets(port::Portfolio; cor_type::PortfolioOptimiserCovCor = PortCovCor(),
               dist_type::DistanceMethod = DistanceCanonical(),
               hclust_alg::HClustAlg = HAC(), hclust_opt::HCOpt = HCOpt())

Hierarchically cluster the assets in a hierarchical portfolio HCPortfolio using the covariance and distance matrices stored in the portfolio. See the arguments types' docs for details.

Inputs

cor_type: correlation matrix estimator PortfolioOptimiserCovCor.

dist_type: method for computing the distance matrix DistanceMethod.

port: hierarchical clustering portfolio HCPortfolio.

hclust_alg: hierarchical clustering algorithm HClustAlg.

hclust_opt: options for determining the number of clusters HCOpt.

Outputs

idx: clustering assignments after cutting the tree into k levels Clustering.cutree.

clustering: Clustering.Hclust of the portfolio assets.

k: optimum number of clusters.

S: N×N asset correlation matrix.

D: N×N asset distance matrix.

Private

PortfolioOptimiser._hcluster — Function

_hcluster(ca::DBHT, port::HCPortfolio, hclust_opt::HCOpt = HCOpt())

Use DBHTs to hierarchically cluster the assets in a hierarchical portfolio HCPortfolio using the covariance and distance matrices stored in the portfolio. See the arguments types' docs for details.

Inputs

ca: [DBHT] options for clustering with DBHTs.

port: hierarchical clustering portfolio HCPortfolio.

hclust_opt: options for determining the number of clusters HCOpt.

Outputs

clustering: Clustering.Hclust of the portfolio assets.

k: optimum number of clusters.

PortfolioOptimiser._hcluster — Function

_hcluster(ca::HAC, port::HCPortfolio, hclust_opt::HCOpt = HCOpt())

Use Clustering.hclust to hierarchically cluster the assets in a hierarchical portfolio HCPortfolio using the covariance and distance matrices stored in the portfolio. See the arguments types' docs for details.

Inputs

ca: linkage for Clustering.hclust.

port: hierarchical clustering portfolio HCPortfolio.

hclust_opt: options for determining the number of clusters HCOpt.

Outputs

clustering: Clustering.Hclust of the portfolio assets.

k: optimum number of clusters.