Network

Public

PortfolioOptimiser.centrality_vector — Method

centrality_vector(port::AbstractPortfolio; X::AbstractMatrix = port.returns,
                  cor_type::PortfolioOptimiserCovCor = PortCovCor(),
                  dist_type::DistMethod = DistCanonical(),
                  network_type::NetworkType = MST())

Compute the centrality vector centrality_vector. See the argument types' docs for details.

Inputs

port: portfolio AbstractPortfolio.

X: T×N returns matrix.

cor_type: correlation matrix estimator PortfolioOptimiserCovCor.

dist_type: method for computing the distance matrix DistMethod.

network_type: method for computing the asset network NetworkType.

Outputs

C: N×1 centrality vector.

PortfolioOptimiser.cluster_matrix — Method

cluster_matrix(port::AbstractPortfolio; X::AbstractMatrix = port.returns,
               cor_type::PortfolioOptimiserCovCor = PortCovCor(),
               dist_type::DistMethod = DistCanonical(),
               hclust_alg::HClustAlg = HAC(), hclust_opt::HCOpt = HCOpt())

Compute the centrality vector cluster_matrix. See the argument types' docs for details.

Inputs

port: portfolio AbstractPortfolio.

X: T×N returns matrix.

cor_type: correlation matrix estimator PortfolioOptimiserCovCor.

dist_type: method for computing the distance matrix DistMethod.

hclust_alg: method for hierarhically clustering assets HClustAlg.

hclust_opt: options for determining the number of clusters HCOpt.

Outputs

C: N×N cluster-based adjacency matrix.

PortfolioOptimiser.connected_assets — Method

connected_assets(port::AbstractPortfolio; X::AbstractMatrix = port.returns,
                 cor_type::PortfolioOptimiserCovCor = PortCovCor(),
                 dist_type::DistMethod = DistCanonical(),
                 network_type::NetworkType = MST())

Compute the percentage of the portfolio comprised of connected assets connected_assets via a connection-based adjacency matrix. See the argument types' docs for details.

Inputs

port: portfolio AbstractPortfolio.

X: T×N returns matrix.

cor_type: correlation matrix estimator PortfolioOptimiserCovCor.

dist_type: method for computing the distance matrix DistMethod.

network_type: method for computing the asset network NetworkType.

Outputs

c: percentage of the portfolio comprised of assets connected via a connection-based adjacency matrix.

PortfolioOptimiser.connection_matrix — Method

connection_matrix(port::AbstractPortfolio; X::AbstractMatrix = port.returns,
                  cor_type::PortfolioOptimiserCovCor = PortCovCor(),
                  dist_type::DistMethod = DistCanonical(),
                  network_type::NetworkType = MST())

Compute the connection matrix connection_matrix. See the argument types' docs for details.

Inputs

port: portfolio AbstractPortfolio.

X: T×N returns matrix.

cor_type: correlation matrix estimator PortfolioOptimiserCovCor.

dist_type: method for computing the distance matrix DistMethod.

network_type: method for computing the asset network NetworkType.

Outputs

C: N×N connection-based adjacency matrix.

PortfolioOptimiser.related_assets — Method

related_assets(port::AbstractPortfolio; X::AbstractMatrix = port.returns,
               type::Symbol = isa(port, Portfolio) ? :Trad : :HRP,
               cor_type::PortfolioOptimiserCovCor = PortCovCor(),
               dist_type::DistMethod = DistCanonical(),
               hclust_alg::HClustAlg = HAC(), hclust_opt::HCOpt = HCOpt())

Compute the percentage of the portfolio comprised of related assets related_assets via a cluster-based adjacency matrix. See the argument types' docs for details.

Inputs

port: portfolio AbstractPortfolio.

X: T×N returns matrix.

cor_type: correlation matrix estimator PortfolioOptimiserCovCor.

dist_type: method for computing the distance matrix DistMethod.

hclust_alg: method for hierarhically clustering assets HClustAlg.

hclust_opt: options for determining the number of clusters HCOpt.

Outputs

c: percentage of the portfolio comprised of related assets via a connection-based adjacency matrix.

Private