Risk

Public

PortfolioOptimiser.calc_risk — Method

calc_risk(port::AbstractPortfolio; X::AbstractMatrix = port.returns,
          type::Symbol = isa(port, Portfolio) ? :Trad : :HRP,
          rm::AbstractRiskMeasure = SD())

Compute the risk for an AbstractRiskMeasure for a portfolio.

Inputs

port: portfolio.

X: T×N returns matrix.

type: optimisation type used to retrieve the weights vector from port.optimal[type].

rm: rism measure.

Outputs

r: risk.

PortfolioOptimiser.factor_risk_contribution — Method

factor_risk_contribution(port::AbstractPortfolio; X::AbstractMatrix = port.returns,
                         F::AbstractMatrix = port.f_returns,
                         type::Symbol = isa(port, Portfolio) ? :Trad : :HRP,
                         rm::AbstractRiskMeasure = SD(), delta::Real = 1e-6)

Compute the factor risk contribution for an AbstractRiskMeasure for a portfolio.

Inputs

port: portfolio.

w: Na×1 vector of asset weights.

X: T×Na matrix of asset returns.

F: T×Nf matrix of factor returns.

type: optimisation type used to retrieve the weights vector from port.optimal[type].

rm: risk measure AbstractRiskMeasure.

delta: small displacement used for computing the marginal risk and equal risk measure Equal.

Outputs

rc_f: Nf×1 vector of risk contribution per factor.

PortfolioOptimiser.number_effective_assets — Method

number_effective_assets(port; type::Symbol = isa(port, Portfolio) ? :Trad : :HRP)

Compute the number of effective assets.

PortfolioOptimiser.risk_contribution — Method

risk_contribution(port::AbstractPortfolio; X::AbstractMatrix = port.returns,
                  type::Symbol = isa(port, Portfolio) ? :Trad : :HRP,
                  rm::AbstractRiskMeasure = SD(), delta::Real = 1e-6,
                  marginal::Bool = false)

Compute the asset risk contribution for an AbstractRiskMeasure for a portfolio.

Inputs

port: portfolio.

X: T×N returns matrix.

type: optimisation type used to retrieve the weights vector from port.optimal[type].

rm: risk measure AbstractRiskMeasure.

delta: small displacement used for computing the marginal risk and equal risk measure Equal.

marginal:

if true: compute the marginal risk contribution.

else: compute the risk by contribution by multiplying the marginal risk by the asset weight.

Outputs

rc: Na×1 vector of risk contribution per asset.

PortfolioOptimiser.sharpe_ratio — Method

sharpe_ratio(port::AbstractPortfolio; X::AbstractMatrix = port.returns,
             mu::AbstractVector = port.mu,
             type::Symbol = isa(port, Portfolio) ? :Trad : :HRP,
             rm::AbstractRiskMeasure = SD(), delta::Real = 1e-6, rf::Real = 0.0,
             kelly::Bool = false)

Compute the risk-adjusted return ratio for an AbstractRiskMeasure for a portfolio.

Inputs

port: portfolio.

X: T×N matrix of asset returns.

mu: N×1 vector of expected returns.

type: optimisation type used to retrieve the weights vector from port.optimal[type].

rm: risk measure AbstractRiskMeasure.

delta: small displacement used for computing the Equal risk measure.

rf: risk free rate.

kelly:

if true: use the kelly return.

else: use the arithmetic return.

Outputs

sr: risk adjusted return ratio.

Private