Risk Measures

Public

PortfolioOptimiser.ERM — Function

ERM(x::AbstractVector, z::Real = 1.0, α::Real = 0.05)

Compute the Entropic Risk Measure.

\[\mathrm{ERM}(\bm{X},\, z, \,\alpha) = z \ln \left(\dfrac{M_{\bm{X}}\left(z^{-1}\right)}{\alpha} \right)\,,\]

where $M_{\bm{X}}\left(z^{-1}\right)$ is the moment generating function of $\bm{X}$.

Inputs

x: vector.

alpha: significance level, α ∈ (0, 1).

z: entropic moment, can be obtained from get_z_from_model and get_z after optimising a Portfolio.

ERM(x::AbstractVector, solvers::Union{NamedTuple, AbstractDict}, α::Real = 0.05)

Compute the Entropic Risk Measure by minimising the function with respect to z. Used in _EVaR, _EDaR and _EDaR_r.

\[\mathrm{ERM} = \begin{cases} \underset{z,\, t,\, u}{\min} & t + z \ln\left(\dfrac{1}{\alpha T}\right)\\ \mathrm{s.t.} & z \geq \sum\limits_{i=1}^{T} u_{i}\\ & (-x_{i}-t,\, z,\, u_{i}) \in \mathcal{K}_{\exp} \, \forall \, i=1,\,\dots{},\, T \end{cases}\,,\]

where $\mathcal{K}_{\exp}$ is the exponential cone.

Inputs

x: vector of portfolio returns.

solvers: named tuple or abstract dict containing the a JuMP-compatible solver capable of solving exponential conic problems, this argument can be formulated in various ways depending on the user's needs.

solvers = Dict(
    # Key-value pair for the solver, solution acceptance criteria, and solver attributes.
    :Clarabel => Dict(
        # Solver we wish to use.
        :solver => Clarabel.Optimizer, 
        # (Optional) Solution acceptance criteria.
        :check_sol => (allow_local = true, allow_almost = true), 
        # (Optional) Solver-specific attributes.
        :params => Dict("verbose" => false)
    )
)

The dictionary contains a key value pair for each solver (plus optional solution acceptance criteria and optional attributes) we want to use.

:solver: defines the solver to use. One can also use JuMP.optimizer_with_attributes to direcly provide a solver with attributes already attached.

:check_sol: (optional) defines the keyword arguments passed on to JuMP.is_solved_and_feasible for accepting/rejecting solutions.

:params: (optional) defines solver-specific parameters.

As long as the same structure is followed, one can use named tuples or even combinations of named tuples and abstract dictionaries.

solvers = (
    Clarabel = (
        # Solver with attributes
        solver = optimizer_with_attributes(Clarabel.Optimizer, "verbose" => false), 
        check_sol = (allow_local = true, allow_almost=true)
    )
)

Users are also able to provide multiple solvers by adding additional key-value pairs to the top-level dictionary/tuple as in the following snippet.

solvers = Dict(
    :Clarabel => Dict(
        :solver => Clarabel.Optimizer, 
        :check_sol => (allow_local = true, allow_almost = true), 
        :params => Dict("verbose" => false)
    ),
    :COSMO => Dict(
        :solver => COSMO.Optimizer,
        ...
    ), ...
)

PortfolioOptimiser will iterate over the solvers until it finds the first one to successfully solve the problem.

α: significance level, α ∈ (0, 1).

If no valid solution is found then NaN will be returned.

Private

PortfolioOptimiser._ADD — Method

_ADD(x::AbstractVector)

Compute the Average Drawdown of uncompounded cumulative returns.

\[\mathrm{ADD_{a}}(\bm{X}) = \dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{a}}(\bm{X},\, j)\,,\]

where $\mathrm{DD_{a}}(\bm{X},\, j)$ is the Drawdown of uncompounded cumulative returns as defined in _DaR.

Inputs

x: vector of portfolio returns.

PortfolioOptimiser._ADD_r — Method

_ADD_r(x::AbstractVector)

Compute the Average Drawdown of compounded cumulative returns.

\[\mathrm{ADD_{r}}(\bm{r}) = \dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{r}}(\bm{X},\, j)\,,\]

where $\mathrm{DD_{a}}(\bm{X},\, j)$ is the Drawdown of compounded cumulative returns as defined in _DaR_r.

Inputs

x: vector of portfolio returns.

PortfolioOptimiser._CDaR — Function

_CDaR(x::AbstractVector, alpha::Real = 0.05)

Compute the Conditional Drawdown at Risk of uncompounded cumulative returns.

\[\mathrm{CDaR_{a}}(\bm{X},\, \alpha) = \mathrm{DaR_{a}}(\bm{X},\, \alpha) + \dfrac{1}{\alpha T} \sum\limits_{j=0}^{T} \max\left[\mathrm{DD_{a}}(\bm{X},\, j) - \mathrm{DaR_{a}}(\bm{X},\, \alpha),\, 0 \right] \,,\]

where $\mathrm{DD_{a}}(\bm{X},\, j)$ is the Drawdown of uncompounded cumulative returns as defined in _DaR, and $\mathrm{DaR_{a}}(\bm{X},\, \alpha)$ the Drawdown at Risk of uncompounded cumulative returns as defined in _DaR.

Inputs

x: vector of portfolio returns.

alpha: significance level, α ∈ (0, 1).

PortfolioOptimiser._CDaR_r — Function

_CDaR_r(x::AbstractVector, alpha::Real = 0.05)

Compute the Conditional Drawdown at Risk of compounded cumulative returns.

\[\mathrm{CDaR_{r}}(\bm{X},\, \alpha) = \mathrm{DaR_{r}}(\bm{X},\, \alpha) + \dfrac{1}{\alpha T} \sum\limits_{j=0}^{T} \max\left[\mathrm{DD_{r}}(\bm{X},\, j) - \mathrm{DaR_{r}}(\bm{X},\, \alpha),\, 0 \right] \,,\]

where $\mathrm{DD_{r}}(\bm{X},\, j)$ is the Drawdown of compounded cumulative returns as defined in _DaR_r, and $\mathrm{DaR_{r}}(\bm{X},\, \alpha)$ the Drawdown at Risk of compounded cumulative returns as defined in _DaR_r.

Inputs

x: vector of portfolio returns.

alpha: significance level, α ∈ (0, 1).

PortfolioOptimiser._CVaR — Function

_CVaR(x::AbstractVector, α::Real = 0.05)

Compute the Conditional Value at Risk.

\[\mathrm{CVaR}(\bm{X},\, \alpha) = \mathrm{VaR}(\bm{X},\, \alpha) - \dfrac{1}{\alpha T} \sum\limits_{t=1}^{T} \min\left( X_{t} + \mathrm{VaR}(\bm{X},\, \alpha),\, 0\right)\,,\]

where $\mathrm{VaR}(\bm{X},\, \alpha)$ is the value at risk as defined in _VaR.

Inputs

x: vector of portfolio returns.

α: significance level, α ∈ (0, 1).

Warning

In-place sorts the input vector.

PortfolioOptimiser._DVar — Method

_DVar(x::AbstractVector)

Compute the Brownian distance variance.

\[\begin{align*} \mathrm{dVar}(\bm{X}) &= \mathrm{dCov}(\bm{X},\, \bm{X}) = \dfrac{1}{T^{2}} \sum\limits_{i=1}^{T}\sum\limits_{j=1}^{T} A_{i,\,j}^2\\ \mathrm{dCov}(\bm{X},\, \bm{Y}) &= \dfrac{1}{T^{2}} \sum\limits_{i=1}^{T} \sum\limits_{j=1}^{T} A_{i,\,j} B_{i,\,j}\\ A_{i,\,j} &= a_{i,\,j} - \bar{a}_{i\,.} - \bar{a}_{.\,j} + \bar{a}_{.\,.}\\ B_{i,\,j} &= b_{i,\,j} - \bar{b}_{i\,.} - \bar{b}_{.\,j} + \bar{b}_{.\,.}\\ a_{i,\,j} &= \lVert X_{i} - X_{j} \rVert_{2}, \quad \forall i,\, j = 1,\, \ldots ,\, T\\ b_{i,\,j} &= \lVert Y_{i} - Y_{j} \rVert_{2}, \quad \forall i,\, j = 1,\, \ldots ,\, T \end{align*}\,,\]

where:

$\bm{X}$ and $\bm{Y}$ are random variables, they are equal in this case as they are the portfolio returns.

$a_{i,\,j}$ and $b_{i,\,j}$ are entries of a distance matrix where $i$ and $j$ are points in time. Each entry is defined as the Euclidean distance $\lVert . \rVert_{2}$ between the value of the random variable at time $i$ and its value at time $j$.

$\bar{a}_{i\,.}$ and $\bar{b}_{i\,.}$ are the $i$-th row means of their respective matrices.

$\bar{a}_{.\,j}$ and $\bar{b}_{.\,j}$ are the $j$-th column means of their respective matrices.

$\bar{a}_{.\,.}$ and $\bar{b}_{.\,.}$ are the grand means of their respective matrices.

$A_{i,\,j}$ and $B_{i,\,j}$ are the doubly centered distances.

PortfolioOptimiser._DaR — Function

_DaR(x::AbstractArray, alpha::Real = 0.05)

Compute the Drawdown at Risk of uncompounded cumulative returns.

\[\begin{align*} \mathrm{DaR_{a}}(\bm{X},\, \alpha) &= \underset{j \in (0,\, T)}{\max} \left\{ \mathrm{DD_{a}}(\bm{X},\, j) \in \mathbb{R} : F_{\mathrm{DD}}\left(\mathrm{DD_{a}}(\bm{X},\, j)\right) < 1 - \alpha \right\}\\ \mathrm{DD_{a}}(\bm{X},\, j) &= \underset{t \in (0,\, j)}{\max}\left( \sum\limits_{i=0}^{t} x_{i} \right) - \sum\limits_{i=0}^{j} x_{i} \end{align*}\,.\]

Inputs

x: vector of portfolio returns.

alpha: significance level, α ∈ (0, 1).

PortfolioOptimiser._DaR_r — Function

_DaR_r(x::AbstractArray; alpha::Real = 0.05)

Compute the Drawdown at Risk of compounded cumulative returns.

\[\begin{align*} \mathrm{DaR_{r}}(\bm{X},\, \alpha) &= \underset{j \in (0,\, T)}{\max} \left\{ \mathrm{DD_{r}}(\bm{X},\, j) \in \mathbb{R} : F_{\mathrm{DD}}\left(\mathrm{DD_{r}}(\bm{X},\, j)\right) < 1 - \alpha \right\}\\ \mathrm{DD_{r}}(\bm{X},\, j) &= \underset{t \in (0,\, j)}{\max}\left( \prod\limits_{i=0}^{t} \left(1+x_{i}\right) \right) - \prod\limits_{i=0}^{j} \left(1+x_{i}\right) \end{align*}\,.\]

Inputs

x: vector of portfolio returns.

alpha: significance level, α ∈ (0, 1).

PortfolioOptimiser._EDaR — Function

_EDaR(x::AbstractVector, solvers::AbstractDict; alpha::Real = 0.05)

Compute the Entropic Drawdown at Risk of uncompounded cumulative returns.

\[\begin{align*} \mathrm{EDaR_{a}}(\bm{X},\alpha) &= \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\mathrm{DD_{a}}(\bm{X}),\, z, \,\alpha)\right\}\\ \mathrm{DD_{a}}(\bm{X}) &= \left\{j \in (0,\, T) : \mathrm{DD_{a}}(\bm{X},\, j) \right\}\,,\end{align*}\]

where $\mathrm{ERM}(\bm{X},\, z, \,\alpha)$ is the entropic risk measure as defined in ERM and $\mathrm{DD_{a}}(\bm{X},\, j)$ the drawdown of uncompounded cumulative returns as defined in _DaR.

Inputs

x: vector of portfolio returns.

alpha: significance level, α ∈ (0, 1).

PortfolioOptimiser._EDaR_r — Function

_EDaR_r(x::AbstractVector, solvers::AbstractDict; alpha::Real = 0.05)

Compute the Entropic Drawdown at Risk of compounded cumulative returns.

\[\begin{align*} \mathrm{EDaR_{r}}(\bm{X},\alpha) &= \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\mathrm{DD_{r}}(\bm{X}),\, z, \,\alpha)\right\}\\ \mathrm{DD_{r}}(\bm{X}) &= \left\{j \in (0,\, T) : \mathrm{DD_{r}}(\bm{X},\, j) \right\}\,, \end{align*}\]

where $\mathrm{ERM}(\bm{X},\, z, \,\alpha)$ is the entropic risk measure as defined in ERM and $\mathrm{DD_{r}}(\bm{X},\, j)$ the drawdown of compounded cumulative returns as defined in _DaR_r.

Inputs

x: vector of portfolio returns.

alpha: significance level, α ∈ (0, 1).

κ: relativistic deformation parameter.

PortfolioOptimiser._EVaR — Function

_EVaR(x::AbstractVector, solvers::AbstractDict, alpha::Real = 0.05)

Compute the Entropic Value at Risk.

\[\mathrm{EVaR}(\bm{X},\alpha) = \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\bm{X},\, z, \,\alpha)\right\}\,,\]

where $\mathrm{ERM}(\bm{X},\, z, \,\alpha)$ is the entropic risk measure as defined in ERM.

Inputs

x: vector of portfolio returns.

alpha: significance level, α ∈ (0, 1).

PortfolioOptimiser._FLPM — Function

_FLPM(x::AbstractVector, r::Real = 0.0)

Compute the First Lower Partial Moment (Omega ratio).

\[\mathrm{FLPM}(\bm{X},\, r) = \dfrac{1}{T} \sum\limits_{t=1}^{T}\max\left(r - \bm{X}_{t},\, 0\right)\,.\]

Inputs

x: vector of portfolio returns.

r: minimum return target.

PortfolioOptimiser._GMD — Method

_GMD(x::AbstractVector)

Compute the Gini Mean Difference.

Inputs

x: vector of portfolio returns.

PortfolioOptimiser._Kurt — Function

_Kurt(x::AbstractVector)

Compute the square root kurtosis.

\[\mathrm{Kurt}(\bm{X}) = \left[\dfrac{1}{T} \sum\limits_{t=1}^{T} \left( X_{t} - \mathbb{E}(\bm{X}) \right)^{4} \right]^{1/2}\,.\]

Inputs

x: vector of portfolio returns.

PortfolioOptimiser._MAD — Function

_MAD(x::AbstractVector, w::Union{AbstractWeights, Nothing} = nothing)

Compute the Mean Absolute Deviation.

\[\mathrm{MAD}(\bm{X}) = \dfrac{1}{T} \sum\limits_{t=1}^T \left\lvert X_{t} - \mathbb{E}(\bm{X}) \right\rvert\,.\]

Inputs

x: vector of portfolio returns.

w: optional vector of weights for computing the mean.

PortfolioOptimiser._MDD — Method

_MDD(x::AbstractVector)

Compute the Maximum Drawdown of uncompounded cumulative returns.

\[\mathrm{MDD_{a}}(\bm{X}) = \underset{j \in (0,\, T)}{\max} \mathrm{DD_{a}}(\bm{X},\, j)\,,``` where ``\mathrm{DD_{a}}(\bm{X},\, j)`` is the Drawdown of uncompounded cumulative returns as defined in [`_DaR`](@ref). # Inputs - `x`: vector of portfolio returns.\]

PortfolioOptimiser._MDD_r — Method

_MDD_r(x::AbstractVector)

Compute the Maximum Drawdown of compounded cumulative returns.

\[\mathrm{MDD_{r}}(\bm{X}) = \underset{j \in (0,\, T)}{\max} \mathrm{DD_{r}}(\bm{X},\, j)\,,``` where ``\mathrm{DD_{a}}(\bm{X},\, j)`` is the Drawdown of compounded cumulative returns as defined in [`_DaR_r`](@ref). # Inputs - `x`: vector of portfolio returns.\]

PortfolioOptimiser._OWA — Method

_OWA(x::AbstractVector, w::AbstractVector)

Compute the Ordered Weight Array risk measure.

Inputs

w: vector of asset weights.

x: vector of portfolio returns.

PortfolioOptimiser._RCVaR — Method

_RCVaR(x::AbstractVector; alpha::Real = 0.05, beta::Real = alpha)

Compute the _CVaR Range.

Inputs

x: vector of portfolio returns.

alpha: significance level of _CVaR losses, $α ∈ (0, 1)$.

beta: significance level of _CVaR gains, beta in (0, 1).

Warning

In-place sorts the input vector.

PortfolioOptimiser._RDaR — Function

_RDaR(x::AbstractVector, solvers::AbstractDict; alpha::Real = 0.05, kappa::Real = 0.3)

Compute the Relativistic Drawdown at Risk of uncompounded cumulative returns.

\[\mathrm{RDaR_{a}}(\bm{X},\, \alpha,\, \kappa) = \mathrm{RRM}(\mathrm{DD_{a}}(\bm{X}),\, \alpha,\, \kappa)\,,``` where ``\mathrm{RRM}(\mathrm{DD_{a}}(\bm{X}),\, \alpha,\, \kappa)`` is the relativistic risk measure as defined in [`RRM`](@ref), and ``\mathrm{DD_{a}}(\bm{X})`` the drawdown of uncompounded cumulative returns as defined in [`_DaR`](@ref). # Inputs - `x`: vector of portfolio returns. - `alpha`: significance level, `α ∈ (0, 1)`. - `κ`: relativistic deformation parameter.\]

PortfolioOptimiser._RDaR_r — Function

_RDaR_r(x::AbstractVector, solvers::AbstractDict; alpha::Real = 0.05, kappa::Real = 0.3)

Compute the Relativistic Drawdown at Risk of compounded cumulative returns.

\[\mathrm{RDaR_{r}}(\bm{X},\, \alpha,\, \kappa) = \mathrm{RRM}(\mathrm{DD_{r}}(\bm{X}),\, \alpha,\, \kappa)\,,``` where ``\mathrm{RRM}(\mathrm{DD_{r}}(\bm{X}),\, \alpha,\, \kappa)`` is the Relativistic Risk Measure as defined in [`RRM`](@ref) where the returns vector, and ``\mathrm{DD_{r}}(\bm{X})`` the drawdown of compounded cumulative returns as defined in [`_DaR_r`](@ref). # Inputs - `x`: vector of portfolio returns. - `alpha`: significance level, `α ∈ (0, 1)`. - `κ`: relativistic deformation parameter.\]

PortfolioOptimiser._RG — Method

_RG(x::AbstractVector)

Compute the Range.

Inputs

x: vector of portfolio returns.

PortfolioOptimiser._RTG — Method

_RTG(x::AbstractVector; alpha_i::Real = 0.0001, alpha::Real = 0.05, a_sim::Real = 100,
     beta_i::Real = alpha_i, beta::Real = alpha, b_sim::Integer = a_sim)

Compute the Tail Gini Range.

Inputs

x: vector of portfolio returns.

alpha_i: start value of the significance level of CVaR losses, `0 <alphai < alpha < 1`.

alpha: end value of the significance level of _CVaR losses, $α ∈ (0, 1)$.

a_sim: number of CVaRs to approximate the Tail Gini losses, a_sim > 0.

beta_i: start value of the significance level of CVaR gains, `0 < betai < beta < 1`.

beta: end value of the significance level of _CVaR gains, beta in (0, 1).

b_sim: number of CVaRs to approximate the Tail Gini gains, b_sim > 0.

Warning

In-place sorts the input vector.

PortfolioOptimiser._RVaR — Function

_RVaR(x::AbstractVector, solvers::AbstractDict, alpha::Real = 0.05, κ::Real = 0.3)

Compute the Relativistic Value at Risk.

\[\mathrm{RVaR}(\bm{X},\, \alpha,\, \kappa) = \mathrm{RRM}(\bm{X},\, \alpha,\, \kappa)\,,``` where ``\mathrm{RRM}(\bm{X},\, \alpha,\, \kappa)`` is the Relativistic Risk Measure as defined in [`RRM`](@ref). # Inputs - `x`: vector of portfolio returns. - `alpha`: significance level, `α ∈ (0, 1)`. - `κ`: relativistic deformation parameter.\]

PortfolioOptimiser._SD — Method

_SD(w::AbstractVector, Σ::AbstractMatrix)

Compute the Standard Deviation. Square root of _Variance.

\[\mathrm{SD}(\bm{w},\, \mathbf{\Sigma}) = \left[\bm{w}^\intercal \, \mathbf{\Sigma} \, \bm{w}\right]^{1/2}\,.\]

Inputs

w: vector of asset weights.

Σ: covariance matrix of asset returns.

PortfolioOptimiser._SKurt — Function

_SKurt(x::AbstractVector)

Compute the square root semi-kurtosis.

\[\mathrm{SKurt}(\bm{X}) = \left[\dfrac{1}{T} \sum\limits_{t=1}^{T} \min\left( X_{t} - \mathbb{E}(\bm{X}),\, 0 \right)^{4} \right]^{1/2}\,.\]

Inputs

x: vector of portfolio returns.

PortfolioOptimiser._SLPM — Function

_SLPM(x::AbstractVector, r::Real = 0.0)

Compute the Second Lower Partial Moment (Sortino Ratio).

\[\mathrm{SLPM}(\bm{X},\, r) = \left[\dfrac{1}{T-1} \sum\limits_{t=1}^{T}\max\left(r - \bm{X}_{t},\, 0\right)^{2}\right]^{1/2}\,.\]

Inputs

x: vector of portfolio returns.

r: minimum return target.

PortfolioOptimiser._SSD — Function

_SSD(x::AbstractVector, r::Real = 0.0, w::Union{AbstractWeights, Nothing} = nothing)

Compute the mean Semi-Standard Deviation.

\[\mathrm{SSD}(\bm{X}) = \left[\dfrac{1}{T-1} \sum\limits_{t=1}^{T}\min\left(\bm{X}_{t} - \mathbb{E}(\bm{X}),\, r\right)^{2}\right]^{1/2}\,.\]

Inputs

x: vector of portfolio returns.

r: minimum return target.

w: optional vector of weights for computing the mean.

PortfolioOptimiser._Skew — Method

_Skew(w::AbstractVector, V::AbstractArray)

PortfolioOptimiser._TG — Method

_TG(x::AbstractVector; alpha_i::Real = 0.0001, alpha::Real = 0.05, a_sim::Int = 100)

Compute the Tail Gini.

Inputs

x: vector of portfolio returns.

alpha_i: start value of the significance level of CVaR losses, `0 <alphai < alpha < 1`.

alpha: end value of the significance level of _CVaR losses, $α ∈ (0, 1)$.

a_sim: number of CVaRs to approximate the Tail Gini losses, a_sim > 0.

Warning

In-place sorts the input vector.

PortfolioOptimiser._UCI — Method

_UCI(x::AbstractVector)

Compute the Ulcer Index of uncompounded cumulative returns.

\[\mathrm{UCI_{a}}(\bm{X}) = \left[\dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{a}}(\bm{X},\, j)^{2}\right]^{1/2}\,,``` where ``\mathrm{DD_{a}}(\bm{X},\, j)`` is the Drawdown of uncompounded cumulative returns as defined in [`_DaR`](@ref). # Inputs - `x`: vector of portfolio returns.\]

PortfolioOptimiser._UCI_r — Method

_UCI_r(x::AbstractVector)

Compute the Ulcer Index of compounded cumulative returns.

\[\mathrm{UCI_{r}}(\bm{X}) = \left[\dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{r}}(\bm{X},\, j)^{2}\right]^{1/2}\,,``` where ``\mathrm{DD_{r}}(\bm{X},\, j)`` is the Drawdown of compounded cumulative returns as defined in [`_DaR_r`](@ref). # Inputs - `x`: vector of portfolio returns.\]

PortfolioOptimiser._VaR — Function

_VaR(x::AbstractVector, α::Real = 0.05)

Compute the Value at Risk, used in _CVaR.

\[\mathrm{VaR}(\bm{X},\, \alpha) = -\underset{t \in (0,\, T)}{\inf} \left\{ X_{t} \in \mathbb{R} : F_{\bm{X}}(X_{t}) > \alpha \right\}\,.\]

Inputs

x: vector of portfolio returns.

α: significance level, α ∈ (0, 1).

Warning

In-place sorts the input vector.

PortfolioOptimiser._Variance — Method

_Variance(w::AbstractVector, Σ::AbstractMatrix)

Compute the Variance. Square of [`SD`](@ref).

\[\mathrm{Variance}(\bm{w},\, \mathbf{\Sigma}) = \bm{w}^\intercal \, \mathbf{\Sigma}\, \bm{w}\,.\]

Inputs

w: vector of asset weights.

Σ: covariance matrix of asset returns.

PortfolioOptimiser._WR — Method

_WR(x::AbstractVector)

Compute the Worst Realisation or Worst Case Scenario.

\[\mathrm{WR}(\bm{X}) = -\min(\bm{X})\,.\]

Inputs

x: vector of portfolio returns.