Miscelaneous

Public

Private

PortfolioOptimiser.get_first_rm — Method

get_first_rm(rm::Union{AbstractVector, <:AbstractRiskMeasure})

Get the first risk measure, used in efficient_frontier!.

Inputs

rm: risk measure or vector of risk measures AbstractRiskMeasure.

PortfolioOptimiser.get_rm_symbol — Method

get_rm_symbol(rm::Union{AbstractVector, <:AbstractRiskMeasure})

Get a symbol for the risk measure(s). If multiple measures are given, they are concatenated by underscores.

Inputs

rm: risk measure or vector of risk measures AbstractRiskMeasure.

PortfolioOptimiser.set_rm_properties! — Method

set_rm_properties!(rm::AbstractRiskMeasure, solvers::AbstractDict,
                   sigma::Union{Nothing, <:AbstractMatrix{<:Real}})

Set properties for risk measures that use solvers or covariance matrices.

Inputs

rm: risk measure AbstractRiskMeasure.

solvers: solvers.

sigma: covariance matrix.

PortfolioOptimiser.unset_set_rm_properties! — Method

unset_set_rm_properties!(rm::AbstractRiskMeasure, solver_flag::Bool, sigma_flag::Bool)

Unset properties for risk measures that use solvers or covariance matrices.

Inputs

rm: risk measure AbstractRiskMeasure.

solvers: solvers.

sigma: covariance matrix.