Statistics

Public

PortfolioOptimiser.factor_risk_contribution — Method

factor_risk_contribution(rm::AbstractRiskMeasure, w::AbstractVector;
                         X::AbstractMatrix = Matrix{Float64}(undef, 0, 0),
                         assets::AbstractVector = Vector{String}(undef, 0),
                         F::AbstractMatrix = Matrix{Float64}(undef, 0, 0),
                         f_assets::AbstractVector = Vector{String}(undef, 0),
                         B::DataFrame = DataFrame(),
                         regression_type::RegressionType = FReg(),
                         V::AbstractMatrix = Matrix{Float64}(undef, 0, 0),
                         SV::AbstractMatrix = Matrix{Float64}(undef, 0, 0),
                         delta::Real = 1e-6, kwargs...)

Compute the factor risk contribution for an AbstractRiskMeasure and asset weight vector.

Inputs

rm: risk measure AbstractRiskMeasure.

w: Na×1 vector of asset weights.

X: T×Na matrix of asset returns.

assets: vector of asset names.

F: T×Nf matrix of factor returns.

f_assets: vector of factor names.

B: loadings matrix.

regression_type: regression type used for computing the loadings matrix.

V: Na×Na matrix of sum of negative spectral slices of the coskewness.

SV: Na×Na matrix of sum of negative spectral slices of the semi coskewness.

delta: small displacement used for computing the marginal risk and equal risk measure Equal.

kwargs: catch-all for any missing keyword arguments for calc_risk.

Outputs

rc_f: Nf×1 vector of risk contribution per factor.

PortfolioOptimiser.risk_bounds — Method

risk_bounds(rm::AbstractRiskMeasure, w1::AbstractVector, w2::AbstractVector;
            X::AbstractMatrix = Matrix{Float64}(undef, 0, 0),
            V::AbstractMatrix = Matrix{Float64}(undef, 0, 0),
            SV::AbstractMatrix = Matrix{Float64}(undef, 0, 0), delta::Real = 1e-6,
            scale::Bool = false, kwargs...)

Compute the risk bounds for an AbstractRiskMeasure and pair of asset weight vectors.

Inputs

rm: risk measure AbstractRiskMeasure.

w1: N×1 vector of asset weights for lower bound.

w2: N×1 vector of asset weights for upper bound.

X: T×N matrix of asset returns.

V: N×N matrix of sum of negative spectral slices of the coskewness.

SV: N×N matrix of sum of negative spectral slices of the semi coskewness.

delta: small displacement used for computing the Equal risk measure.

scale:

if true: divides the kurtosis and semi kurtosis by 2, used in risk_contribution.

kwargs: catch-all for any missing keyword arguments for calc_risk.

Outputs

r1: lower risk bound.

r2: upper risk bound.

PortfolioOptimiser.risk_contribution — Method

risk_contribution(rm::AbstractRiskMeasure, w::AbstractVector;
                  X::AbstractMatrix = Matrix{Float64}(undef, 0, 0),
                  V::AbstractMatrix = Matrix{Float64}(undef, 0, 0),
                  SV::AbstractMatrix = Matrix{Float64}(undef, 0, 0),
                  delta::Real = 1e-6, marginal::Bool = false, kwargs...)

Compute the asset risk contribution for an AbstractRiskMeasure and asset weight vector.

Inputs

rm: risk measure AbstractRiskMeasure.

w: N×1 vector of asset weights.

X: T×N matrix of asset returns.

V: N×N matrix of sum of negative spectral slices of the coskewness.

SV: N×N matrix of sum of negative spectral slices of the semi coskewness.

delta: small displacement used for computing the marginal risk and equal risk measure Equal.

marginal:

if true: compute the marginal risk contribution.

else: compute the risk by contribution by multiplying the marginal risk by the asset weight.

kwargs: catch-all for any missing keyword arguments for calc_risk.

Outputs

rc: N×1 vector of risk contribution per asset.

PortfolioOptimiser.sharpe_ratio — Method

sharpe_ratio(rm::AbstractRiskMeasure, w::AbstractVector;
             mu::AbstractVector = Vector{Float64}(undef, 0),
             X::AbstractMatrix = Matrix{Float64}(undef, 0, 0),
             V::AbstractMatrix = Matrix{Float64}(undef, 0, 0),
             SV::AbstractMatrix = Matrix{Float64}(undef, 0, 0), delta::Real = 1e-6,
             rf::Real = 0.0, kelly::Bool = false)

Compute the risk-adjusted return ratio for an AbstractRiskMeasure and asset weights vector.

Inputs

rm: risk measure AbstractRiskMeasure.

w: N×1 vector of asset weights.

mu: N×1 vector of expected returns.

X: T×N matrix of asset returns.

V: N×N matrix of sum of negative spectral slices of the coskewness.

SV: N×N matrix of sum of negative spectral slices of the semi coskewness.

delta: small displacement used for computing the Equal risk measure.

rf: risk free rate.

kelly:

if true: use the kelly return.

else: use the arithmetic return.

Outputs

sr: risk adjusted return ratio.

Private