Risk measaures

Some risk measures require solvers to compute.

solvers = Dict(
               # Key-value pair for the solver, solution acceptance 
               # criteria, and solver attributes.
               :Clarabel => Dict(
                                 # Solver we wish to use.
                                 :solver => Clarabel.Optimizer,
                                 # (Optional) Solution acceptance criteria.
                                 :check_sol => (allow_local = true, allow_almost = true),
                                 # (Optional) Solver-specific attributes.
                                 :params => Dict("verbose" => false)))

The dictionary contains a key value pair for each solver (plus optional solution acceptance criteria and optional attributes) we want to use.

:solver: defines the solver to use. One can also use JuMP.optimizer_with_attributes to direcly provide a solver with attributes already attached.

:check_sol: (optional) defines the keyword arguments passed on to JuMP.is_solved_and_feasible for accepting/rejecting solutions.

:params: (optional) defines solver-specific parameters.

Users are also able to provide multiple solvers by adding additional key-value pairs to the top-level dictionary as in the following snippet.

using JuMP
solvers = Dict(:Clarabel => Dict(:solver => Clarabel.Optimizer,
                                 :check_sol => (allow_local = true, allow_almost = true),
                                 :params => Dict("verbose" => false)),
               # Provide solver with pre-attached attributes and no arguments 
               # for the `JuMP.is_solved_and_feasible` function.
               :COSMO => Dict(:solver => JuMP.optimizer_with_attributes(COSMO.Optimizer,
                                                                        "maxiter" => 5000)))

PortfolioOptimiser will iterate over the solvers until it finds the first one to successfully solve the problem.

Public

PortfolioOptimiser.ADD — Type

@kwdef mutable struct ADD{T1 <: Real} <: RiskMeasure
    settings::RMSettings = RMSettings()
end

Defines the Average Drawdown of uncompounded returns _ADD risk measure.

Parameters

settings: risk measure settings RMSettings.

PortfolioOptimiser.ADD_r — Type

@kwdef mutable struct ADD_r{T1 <: Real} <: HCRiskMeasure
    settings::HCRMSettings = HCRMSettings()
    alpha::T1 = 0.05
end

Defines the Average Drawdown of compounded cumulative returns _ADD_r risk measure.

Parameters

settings: risk measure settings HCRMSettings.

PortfolioOptimiser.CDaR — Type

@kwdef mutable struct CDaR{T1 <: Real} <: RiskMeasure
    settings::RMSettings = RMSettings()
    alpha::T1 = 0.05
end

Defines the Conditional Drawdown at Risk of uncompounded returns at Risk _CDaR risk measure.

Parameters

settings: risk measure settings RMSettings.

alpha: significance level, alpha ∈ (0, 1).

PortfolioOptimiser.CDaR_r — Type

@kwdef mutable struct CDaR_r{T1 <: Real} <: HCRiskMeasure
    settings::HCRMSettings = HCRMSettings()
    alpha::T1 = 0.05
end

Defines the Conditional Drawdown at Risk of compounded cumulative returns _CDaR_r risk measure.

Parameters

settings: risk measure settings HCRMSettings.

alpha: significance level, alpha ∈ (0, 1).

PortfolioOptimiser.CVaR — Type

@kwdef mutable struct CVaR{T1 <: Real} <: RiskMeasure
    settings::RMSettings = RMSettings()
    alpha::T1 = 0.05
end

Defines the Conditional Value at Risk _CVaR risk measure.

Parameters

settings: risk measure settings RMSettings.

alpha: significance level, alpha ∈ (0, 1).

PortfolioOptimiser.CVaRRG — Type

@kwdef mutable struct CVaRRG{T1 <: Real} <: RiskMeasure
    settings::RMSettings = RMSettings()
end

Defines the Conditional Value at Risk Range _CVaRRG risk measure.

Parameters

settings: risk measure settings RMSettings.

alpha: significance level of CVaR losses, alpha ∈ (0, 1).

beta: significance level of CVaR gains, beta ∈ (0, 1).

PortfolioOptimiser.DaR — Type

@kwdef mutable struct DaR{T1 <: Real} <: HCRiskMeasure
    settings::HCRMSettings = HCRMSettings()
    alpha::T1 = 0.05
end

Defines the Drawdown at Risk of uncompounded cumulative returns _DaR risk measure.

Parameters

settings: risk measure settings HCRMSettings.

alpha: significance level, alpha ∈ (0, 1).

PortfolioOptimiser.DaR_r — Type

@kwdef mutable struct DaR_r{T1 <: Real} <: HCRiskMeasure
    settings::HCRMSettings = HCRMSettings()
    alpha::T1 = 0.05
end

Defines the Drawdown at Risk of compounded cumulative returns _DaR_r risk measure.

Parameters

settings: risk measure settings HCRMSettings.

alpha: significance level, alpha ∈ (0, 1).

PortfolioOptimiser.EDaR — Type

@kwdef mutable struct EDaR{T1 <: Real} <: RiskMeasure
    settings::RMSettings = RMSettings()
    alpha::T1 = 0.05
    solvers::Union{<:AbstractDict, Nothing} = nothing
end

Defines the Entropic Drawdown at Risk of uncompounded cumulative returns _EDaR risk measure.

Parameters

settings: risk measure settings RMSettings.

alpha: significance level, alpha ∈ (0, 1).

solvers: optional abstract dict containing the a JuMP-compatible solver capable of solving 3D power cone problems.

if nothing: use the solvers stored in the instance of Portfolio or HCPortfolio.

else: use these ones.

PortfolioOptimiser.EDaR_r — Type

@kwdef mutable struct EDaR_r{T1 <: Real} <: HCRiskMeasure
    settings::HCRMSettings = HCRMSettings()
    alpha::T1 = 0.05
    solvers::Union{<:AbstractDict, Nothing} = nothing
end

Defines the Entropic Drawdown at Risk of compounded cumulative returns _EDaR_r risk measure.

Parameters

settings: risk measure settings HCRMSettings.

alpha: significance level, alpha ∈ (0, 1).

solvers: optional abstract dict containing the a JuMP-compatible solver capable of solving 3D power cone problems.

if nothing: use the solvers stored in the instance of Portfolio or HCPortfolio.

else: use these ones.

PortfolioOptimiser.EVaR — Type

@kwdef mutable struct EVaR{T1 <: Real} <: RiskMeasure
    settings::RMSettings = RMSettings()
    alpha::T1 = 0.05
    solvers::Union{<:AbstractDict, Nothing} = nothing
end

Defines the Entropic Value at Risk _EVaR risk measure.

Parameters

settings: risk measure settings RMSettings.

alpha: significance level, alpha ∈ (0, 1).

solvers: optional abstract dict containing the a JuMP-compatible solver capable of solving 3D power cone problems.

if nothing: use the solvers stored in the instance of Portfolio or HCPortfolio.

else: use these ones.

PortfolioOptimiser.Equal — Type

@kwdef struct Equal <: HCRiskMeasure
    settings::HCRMSettings = HCRMSettings()
end

Defines the Equal risk measure, where risk is allocated evenly among a group of assets.

Parameters

settings: risk measure settings HCRMSettings.

PortfolioOptimiser.FLPM — Type

@kwdef mutable struct FLPM{T1 <: Real} <: RiskMeasure
    settings::RMSettings = RMSettings()
    target::T1 = 0.0
end

Defines the First Lower Partial Moment (Omega ratio) _FLPM risk measure.

Parameters

settings: risk measure settings RMSettings.

target: minimum return threshold for classifying downside returns.

PortfolioOptimiser.GMD — Type

@kwdef struct GMD <: RiskMeasure
    settings::RMSettings = RMSettings()
    owa::OWASettings = OWASettings()
end

Defines the Gini Mean Difference _GMD risk measure.

Parameters

settings: risk measure settings RMSettings.

owa: OWA risk measure settings OWASettings.

PortfolioOptimiser.HCRMSettings — Type

@kwdef mutable struct HCRMSettings{T1 <: Real}
    scale::T1 = 1.0
end

Risk measure settings for concrete subtypes of HCRiskMeasure.

Parameters

scale: factor for scaling the risk when adding it to the risk being minimised.

PortfolioOptimiser.Kurt — Type

@kwdef mutable struct Kurt{T1 <: Real} <: RiskMeasure
    settings::RMSettings = RMSettings()
    w::Union{<:AbstractWeights, Nothing} = nothing
    mu::Union{<:AbstractVector, Nothing} = nothing
end

Defines the Square Root Kurtosis _Kurt risk measure.

Parameters

settings: risk measure settings RMSettings.

w: optional T×1 vector of weights for computing the expected return in _Kurt.

kt: optional N^2×N^2 cokurtosis matrix.

If nothing: use the cokurtosis matrix stored in the instance of Portfolio.

else: use this one.

PortfolioOptimiser.MAD — Type

@kwdef mutable struct MAD <: RiskMeasure
    settings::RMSettings = RMSettings()
    w::Union{<:AbstractWeights, Nothing} = nothing
    mu::Union{<:AbstractVector, Nothing} = nothing
end

Defines the Mean Absolute Deviation _MAD risk measure.

Parameters

settings: risk measure settings RMSettings.

w: optional T×1 vector of weights for computing the expected return in _MAD.

mu: optional N×1 vector of expected asset returns.

If nothing: use the expected asset returns stored in the instance of Portfolio.

else: use this one.

PortfolioOptimiser.MDD — Type

@kwdef mutable struct MDD{T1 <: Real} <: RiskMeasure
    settings::RMSettings = RMSettings()
end

Defines the Maximum Drawdown (Calmar ratio) of uncompounded returns _MDD risk measure.

Parameters

settings: risk measure settings RMSettings.

PortfolioOptimiser.MDD_r — Type

@kwdef mutable struct MDD_r{T1 <: Real} <: HCRiskMeasure
    settings::HCRMSettings = HCRMSettings()
    alpha::T1 = 0.05
end

Defines the Maximum Drawdown (Calmar ratio) of compounded cumulative returns _MDD_r risk measure.

Parameters

settings: risk measure settings HCRMSettings.

PortfolioOptimiser.OWA — Type

@kwdef struct OWA <: RiskMeasure
    settings::RMSettings = RMSettings()
    owa::OWASettings = OWASettings()
    w::Union{<:AbstractVector, Nothing} = nothing
end

Defines the generic Ordered Weight Array _OWA risk measure.

Parameters

settings: risk measure settings RMSettings.

owa: OWA risk measure settings OWASettings.

w: optional T×1 vector of ordered weights.

if nothing: use owa_gmd to compute the weights.

else: use this value.

PortfolioOptimiser.OWASettings — Type

@kwdef mutable struct OWASettings{T1<:AbstractVector{<:Real}}
    approx::Bool = true
    p::T1 = Float64[2, 3, 4, 10, 50]
end

Defines the settings for Ordered Weight Array (OWA) risk measures.

Parameters

approx: whether or not to use the approximate formulation based on power cone norms.

p:

if approx == true: vector of the order of p-norms to use in the approximation.

else: does nothing.

PortfolioOptimiser.QuadSD — Type

struct QuadSD <: SDSquaredFormulation end

The risk expression will be the explicit quadratic form of the variance, dot(w, sigma, w). Where w is the N×1 vector of asset weights and sigma the covariance matrix.

PortfolioOptimiser.RG — Type

@kwdef mutable struct RG{T1 <: Real} <: RiskMeasure
    settings::RMSettings = RMSettings()
end

Defines the Range _RG risk measure.

Parameters

settings: risk measure settings RMSettings.

PortfolioOptimiser.RLDaR — Type

@kwdef mutable struct RLDaR{T1 <: Real} <: RiskMeasure
    settings::RMSettings = RMSettings()
    alpha::T1 = 0.05
    solvers::Union{<:AbstractDict, Nothing} = nothing
end

Defines the Relativistic Drawdown at Risk of uncompounded cumulative returns _RLDaR risk measure.

Parameters

settings: risk measure settings RMSettings.

alpha: significance level, alpha ∈ (0, 1).

kappa: relativistic deformation parameter, κ ∈ (0, 1).

solvers: optional abstract dict containing the a JuMP-compatible solver capable of solving 3D power cone problems.

if nothing: use the solvers stored in the instance of Portfolio or HCPortfolio.

else: use these ones.

PortfolioOptimiser.RLDaR_r — Type

@kwdef mutable struct RLDaR_r{T1 <: Real} <: RiskMeasure
    settings::RMSettings = RMSettings()
    alpha::T1 = 0.05
    solvers::Union{<:AbstractDict, Nothing} = nothing
end

Defines the Relativistic Drawdown at Risk of compounded cumulative returns _RLDaR_r risk measure.

Parameters

settings: risk measure settings RMSettings.

alpha: significance level, alpha ∈ (0, 1).

kappa: relativistic deformation parameter, κ ∈ (0, 1).

solvers: optional abstract dict containing the a JuMP-compatible solver capable of solving 3D power cone problems.

if nothing: use the solvers stored in the instance of Portfolio or HCPortfolio.

else: use these ones.

PortfolioOptimiser.RLVaR — Type

@kwdef mutable struct RLVaR{T1 <: Real} <: RiskMeasure
    settings::RMSettings = RMSettings()
    alpha::T1 = 0.05
    solvers::Union{<:AbstractDict, Nothing} = nothing
end

Defines the Relativistic Value at Risk _RLVaR risk measure.

Parameters

settings: risk measure settings RMSettings.

alpha: significance level, alpha ∈ (0, 1).

kappa: relativistic deformation parameter, κ ∈ (0, 1).

solvers: optional abstract dict containing the a JuMP-compatible solver capable of solving 3D power cone problems.

if nothing: use the solvers stored in the instance of Portfolio or HCPortfolio.

else: use these ones.

PortfolioOptimiser.RMSettings — Type

mutable struct RMSettings{T1 <: Real, T2 <: Real}
    flag::Bool = true
    scale::T1 = 1.0
    ub::T2 = Inf
end

Risk measure settings for concrete subtypes of RiskMeasure.

Parameters

When optimising a Portfolio.

flag: if true the risk will contribute to the JuMP model's risk expression.

scale: factor for scaling the risk when adding it to the JuMP model's risk expression.

ub: if is finite, sets the upper bound for the risk.

When optimising a HCPortfolio.

flag: does nothing.

scale: factor for scaling the risk when adding it to the risk being minimised.

ub: does nothing.

PortfolioOptimiser.SD — Type

@kwdef mutable struct SD{T1 <: Union{AbstractMatrix, Nothing}} <: RiskMeasure
    settings::RMSettings = RMSettings()
    formulation::SDFormulation = SOCSD()
    sigma::Union{<:AbstractMatrix, Nothing} = nothing
end

Defines the Standard Deviation _SD risk measure.

Parameters

settings: risk measure settings RMSettings.

formulation: formulation of the standard deviation/variance SDFormulation.

sigma: optional N×N covariance matrix.

if nothing: use the covariance matrix stored in the instance of Portfolio or HCPortfolio.

else: use this one.

PortfolioOptimiser.SKurt — Type

@kwdef mutable struct SKurt{T1 <: Real} <: RiskMeasure
    settings::RMSettings = RMSettings()
    target::T1 = 0.0
    w::Union{<:AbstractWeights, Nothing} = nothing
    mu::Union{<:AbstractVector, Nothing} = nothing
end

Defines the Square Root Semi Kurtosis _SKurt risk measure.

Parameters

settings: risk measure settings RMSettings.

target: minimum return threshold for classifying downside returns.

w: optional T×1 vector of weights for computing the expected return in _SKurt.

kt: optional N^2×N^2 semi cokurtosis matrix.

If nothing: use the semi cokurtosis matrix stored in the instance of Portfolio.

else: use this one.

PortfolioOptimiser.SLPM — Type

@kwdef mutable struct SLPM{T1 <: Real} <: RiskMeasure
    settings::RMSettings = RMSettings()
    target::T1 = 0.0
end

Defines the Second Lower Partial Moment (Sortino ratio) _SLPM risk measure.

Parameters

settings: risk measure settings RMSettings.

target: minimum return threshold for classifying downside returns.

PortfolioOptimiser.SOCSD — Type

struct SOCSD <: SDSquaredFormulation end

The model will use a MOI.SecondOrderCone to define the standard deviation dev and make the risk expression sd_risk = dev^2.

PortfolioOptimiser.SSD — Type

@kwdef mutable struct SSD{T1 <: Real} <: RiskMeasure
    settings::RMSettings = RMSettings()
    target::T1 = 0.0
    w::Union{<:AbstractWeights, Nothing} = nothing
    mu::Union{<:AbstractVector, Nothing} = nothing
end

Defines the Semi Standard Deviation _SSD risk measure.

Parameters

settings: risk measure settings RMSettings.

target: minimum return threshold for classifying downside returns.

w: optional T×1 vector of weights for computing the expected return in _SSD.

mu: optional N×1 vector of expected asset returns.

If nothing: use the expected asset returns stored in the instance of Portfolio.

else: use this one.

PortfolioOptimiser.SSkew — Type

@kwdef struct SSkew <: RiskMeasure
    settings::RMSettings = RMSettings()
end

Define the Quadratic SSkewness _Skew risk measure.

Parameters

settings: risk measure settings RMSettings.

PortfolioOptimiser.SVariance — Type

@kwdef mutable struct SVariance{T1 <: Real} <: HCRiskMeasure
    settings::HCRMSettings = HCRMSettings()
    target::T1 = 0.0
    w::Union{<:AbstractWeights, Nothing} = nothing
    mu::Union{<:AbstractVector, Nothing} = nothing
end

Defines the Semi Variance _SVariance risk measure.

Parameters

settings: risk measure settings HCRMSettings.

target: minimum return threshold for classifying downside returns.

w: optional T×1 vector of weights for computing the expected return in _SVariance.

mu: optional N×1 vector of expected asset returns.

If nothing: use the expected asset returns stored in the instance of Portfolio.

else: use this one.

PortfolioOptimiser.SimpleSD — Type

struct SOCSD <: SDSquaredFormulation end

The model will use a MOI.SecondOrderCone to define the standard deviation dev and make the risk expression sd_risk = dev.

PortfolioOptimiser.Skew — Type

@kwdef struct Skew <: RiskMeasure
    settings::RMSettings = RMSettings()
end

Define the Quadratic Skewness _Skew risk measure.

Parameters

settings: risk measure settings RMSettings.

PortfolioOptimiser.TG — Type

@kwdef mutable struct TG{T1 <: Real, T2 <: Real, T3 <: Integer} <: RiskMeasure
    settings::RMSettings
    owa::OWASettings
    alpha_i::T1 = 0.0001
    alpha::T2 = 0.05
    a_sim::T3 = 100
end

Defines the Tail Gini Difference _TG risk measure.

Parameters

settings: risk measure settings RMSettings.

owa: OWA risk measure settings OWASettings.

alpha_i: start value of the significance level of CVaR losses, 0 < alpha_i < alpha < 1.

alpha: end value of the significance level of CVaR losses, alpha ∈ (0, 1).

a_sim: number of CVaRs to approximate the Tail Gini losses, a_sim > 0.

PortfolioOptimiser.TGRG — Type

mutable struct TGRG{T1 <: Real, T2 <: Real, T3 <: Integer, T4 <: Real, T5 <: Real, T6 <: Integer} <: RiskMeasure
    settings::RMSettings
    owa::OWASettings
    alpha_i::T1 = 0.0001
    alpha::T2 = 0.05
    a_sim::T3 = 100
    beta_i::T4 = 0.0001
    beta::T5 = 0.05
    b_sim::T6 = 100
end

Defines the Tail Gini Difference Range _TGRG risk measure.

Parameters

settings: risk measure settings RMSettings.

owa: OWA risk measure settings OWASettings.

alpha_i: start value of the significance level of CVaR losses, 0 < alpha_i < alpha < 1.

alpha: end value of the significance level of CVaR losses, alpha ∈ (0, 1).

a_sim: number of CVaRs to approximate the Tail Gini losses, a_sim > 0.

beta_i: start value of the significance level of CVaR gains, 0 < beta_i < beta < 1.

beta: end value of the significance level of CVaR gains, beta ∈ (0, 1).

b_sim: number of CVaRs to approximate the Tail Gini gains, b_sim > 0.

PortfolioOptimiser.UCI — Type

@kwdef mutable struct UCI{T1 <: Real} <: RiskMeasure
    settings::RMSettings = RMSettings()
end

Defines the Ulcer Index _UCI risk measure.

Parameters

settings: risk measure settings RMSettings.

PortfolioOptimiser.UCI_r — Type

@kwdef mutable struct UCI_r{T1 <: Real} <: HCRiskMeasure
    settings::HCRMSettings = HCRMSettings()
    alpha::T1 = 0.05
end

Defines the Ulcer Index of compounded cumulative returns _UCI_r risk measure.

Parameters

settings: risk measure settings HCRMSettings.

PortfolioOptimiser.VaR — Type

@kwdef mutable struct VaR{T1 <: Real} <: HCRiskMeasure
    settings::HCRMSettings = HCRMSettings()
    alpha::T1 = 0.05
end

Defines the Value at Risk _VaR risk measure.

Parameters

settings: risk measure settings HCRMSettings.

alpha: significance level, alpha ∈ (0, 1).

PortfolioOptimiser.Variance — Type

@kwdef mutable struct Variance{T1 <: Union{AbstractMatrix, Nothing}} <: HCRiskMeasure
    settings::HCRMSettings = HCRMSettings()
    sigma::Union{<:AbstractMatrix, Nothing} = nothing
end

Defines the Variance _Variance risk measure.

Parameters

settings: risk measure settings RMSettings.

sigma: optional N×N covariance matrix.

if nothing: use the covariance matrix stored in the instance of Portfolio or HCPortfolio.

else: use this one.

PortfolioOptimiser.WR — Type

@kwdef mutable struct WR{T1 <: Real} <: RiskMeasure
    settings::RMSettings = RMSettings()
end

Defines the Worst Realisation _WR risk measure.

Parameters

settings: risk measure settings RMSettings.

PortfolioOptimiser.dVar — Type

@kwdef struct dVar <: RiskMeasure
    settings::RMSettings = RMSettings()
end

Define the Brownian Distance Variance _dVar risk measure.

Parameters

settings: risk measure settings RMSettings.

PortfolioOptimiser.ERM — Function

ERM(x::AbstractVector, solvers:AbstractDict, α::Real = 0.05)

Compute the Entropic Risk Measure. Used in _EVaR, _EDaR and _EDaR_r.

\[\begin{align} \mathrm{ERM}(\bm{X},\, z, \,\alpha) &= \begin{cases} \underset{z,\, t,\, u}{\inf} & t + z \ln\left(\dfrac{1}{\alpha T}\right)\\ \mathrm{s.t.} & z \geq \sum\limits_{i=1}^{T} u_{i}\nonumber\\ & (-X_{i}-t,\, z,\, u_{i}) \in \mathcal{K}_{\exp} \, \forall \, i=1,\,\dots{},\, T \end{cases}\\ &= \underset{z>0}{\inf}\left\{ z \ln \left(\dfrac{M_{\bm{X}}\left(z^{-1}\right)}{\alpha} \right)\right\}\,. \end{align}\]

Where:

$M_{\bm{X}}\left(t\right)$ is the moment generating function of $\bm{X}$.

$\mathcal{K}_{\mathrm{exp}}$ is the exponential cone.

$\alpha \in (0,\,1)$ is the significance parameter.

Inputs

x: T×1 returns vector.

solvers: abstract dict containing the a JuMP-compatible solver capable of solving exponential conic problems.

α: significance level, α ∈ (0, 1).

If no valid solution is found then NaN will be returned.

Outputs

r: risk.

PortfolioOptimiser.ERM — Function

ERM(x::AbstractVector, z::Real = 1.0, α::Real = 0.05)

Compute the Entropic Risk Measure.

\[\begin{align} \mathrm{ERM}(\bm{X},\, z, \,\alpha) &= z \ln \left(\dfrac{M_{\bm{X}}\left(z^{-1}\right)}{\alpha} \right)\,. \end{align}\]

Where:

$M_{\bm{X}}\left(t\right)$ is the moment generating function of $\bm{X}$.

$\alpha \in (0,\,1)$ is the significance parameter.

Inputs

x: T×1 returns vector.

α: significance level, α ∈ (0, 1).

z: entropic moment, can be obtained from get_z_from_model and get_z after optimising a Portfolio.

Outputs

r: risk.

PortfolioOptimiser.RRM — Function

RRM(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05, κ::Real = 0.3)

Compute the Relativistic Risk Measure. Used in _RLVaR, _RLDaR and _RLDaR_r.

\[\begin{align} \mathrm{RRM}^{\kappa}_{\alpha}(X) &= \left\{ \begin{align} &\underset{z,\, t,\, \psi,\, \theta,\, \varepsilon,\, \omega}{\text{inf}} && t + z \ln_{\kappa} \left(\dfrac{1}{\alpha T}\right) + \sum\limits_{i=1}^T \left(\psi_{i} + \theta_{i} \right) \nonumber\\ &\mathrm{s.t.} && -X - t + \varepsilon + \omega \leq 0 \nonumber\\ &&&z \geq 0 \\ &&&\left( z\left(\dfrac{1+\kappa}{2\kappa}\right),\, \psi_{i}\left(\dfrac{1+\kappa}{\kappa}\right),\, \varepsilon_{i} \right) \in \mathcal{P}_{3}^{1/(1+\kappa),\, \kappa/(1+\kappa)}\nonumber\\ &&&\left( \omega_{i}\left(\dfrac{1}{1-\kappa}\right),\, \theta_{i}\left(\dfrac{1}{\kappa}\right),\, -z \left(\dfrac{1}{2\kappa}\right) \right) \in \mathcal{P}_{3}^{1-\kappa,\, \kappa}\nonumber\\ &&&\forall \, i=1,\,\dots{},\, T \nonumber \end{align} \right.\,. \end{align}\]

Where:

$\mathcal{P}_3^{\alpha,\, 1-\alpha}$ is the power cone 3D.

$\alpha \in (0,\,1)$ is the significance parameter.

$\kappa \in (0,\,1)$ is the relativistic deformation parameter.

Inputs

x: T×1 returns vector.

solvers: abstract dict containing the a JuMP-compatible solver capable of solving 3D power cone problems.

α: significance level, α ∈ (0, 1).

κ: relativistic deformation parameter, κ ∈ (0, 1).

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(::ADD, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the ADD via _ADD. Inputs correspond to those of _ADD.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(::ADD_r, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the ADD_r via _ADD_r. Inputs correspond to those of _ADD_r.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(cdar::CDaR, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the CDaR via _CDaR. Inputs correspond to those of _CDaR.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(cdar::CDaR_r, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the CDaR_r via _CDaR_r. Inputs correspond to those of _CDaR_r.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(cvar::CVaR, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the CVaR via _CVaR. Inputs correspond to those of _CVaR.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(rcvar::CVaRRG, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the CVaRRG via _CVaRRG. Inputs correspond to those of _CVaRRG.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(dar::DaR, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the DaR via _DaR. Inputs correspond to those of _DaR.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(dar::DaR_r, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the DaR_r via _DaR_r. Inputs correspond to those of _DaR_r.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(edar::EDaR, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the EDaR via _EDaR. Inputs correspond to those of _EDaR.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(edar::EDaR_r, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the EDaR_r via _EDaR_r. Inputs correspond to those of _EDaR_r.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(evar::EVaR, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the EVaR via _EVaR. Inputs correspond to those of _EVaR.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(::Equal, w::AbstractVector; delta::Real = 0, kwargs...)

Compute the risk as the inverse of the length of w.

Inputs

w: N×1 vector of weights.

delta: is a displacement, used in risk_contribution and factor_risk_contribution.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(svariance::FLPM, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the FLPM via _FLPM. Inputs correspond to those of _FLPM.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(::GMD, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the GMD via _GMD. Inputs correspond to those of _GMD.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(kt::Kurt, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the Kurt via _Kurt. Inputs correspond to those of _Kurt.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(mad::MAD, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the MAD via _MAD. Inputs correspond to those of _MAD.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(::MDD, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the MDD via _MDD. Inputs correspond to those of _MDD.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(::MDD_r, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the MDD_r via _MDD_r. Inputs correspond to those of _MDD_r.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(owa::OWA, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the OWA via _OWA. Inputs correspond to those of _OWA.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(::RG, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the RG via _RG. Inputs correspond to those of _RG.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(rdar::RLDaR, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the RLDaR via _RLDaR. Inputs correspond to those of _RLDaR.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(rdar::RLDaR_r, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the RLDaR_r via _RLDaR_r. Inputs correspond to those of _RLDaR_r.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(rvar::RLVaR, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the RLVaR via _RLVaR. Inputs correspond to those of _RLVaR.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(sd::SD, w::AbstractVector; kwargs...)

Compute the SD via _SD. Inputs correspond to those of _SD.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(skt::SKurt, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the SKurt via _SKurt. Inputs correspond to those of _SKurt.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(svariance::SLPM, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the SLPM via _SLPM. Inputs correspond to those of _SLPM.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(mad::SSD, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the SSD via _SSD. Inputs correspond to those of _SSD.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(::SSkew, w::AbstractVector; SV::AbstractMatrix, kwargs...)

Compute the SSkew via _Skew. Inputs correspond to those of _Skew.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(svariance::SVariance, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the SVariance via _SVariance. Inputs correspond to those of _SVariance.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(::Skew, w::AbstractVector; V::AbstractMatrix, kwargs...)

Compute the Skew via _Skew. Inputs correspond to those of _Skew.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(tg::TG, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the TG via _TG. Inputs correspond to those of _TG.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(rtg::TGRG, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the TGRG via _TGRG. Inputs correspond to those of _TGRG.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(::UCI, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the UCI via _UCI. Inputs correspond to those of _UCI.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(::UCI_r, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the UCI_r via _UCI_r. Inputs correspond to those of _UCI_r.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(var::VaR, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the VaR via _VaR. Inputs correspond to those of _VaR.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(sd::Variance, w::AbstractVector; kwargs...)

Compute the Variance via _Variance. Inputs correspond to those of _Variance.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(::WR, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the WR via _WR. Inputs correspond to those of _WR.

Outputs

r: risk.

PortfolioOptimiser.calc_risk — Method

calc_risk(::dVar, w::AbstractVector; X::AbstractMatrix, kwargs...)

Compute the dVar via _dVar. Inputs correspond to those of _dVar.

Outputs

r: risk.

Private

PortfolioOptimiser.AbstractRiskMeasure — Type

abstract type AbstractRiskMeasure end

Abstract type for subtyping risk measures.

PortfolioOptimiser.HCRiskMeasure — Type

abstract type HCRiskMeasure <: AbstractRiskMeasure end

Abstract type for subtyping risk meaasures that can only be used to optimise HCPortfolio.

PortfolioOptimiser.RiskMeasure — Type

abstract type RiskMeasure <: AbstractRiskMeasure end

Abstract type for subtyping risk measures that can be used to optimise Portfolio and HCPortfolio.

PortfolioOptimiser.SDFormulation — Type

abstract type SDFormulation end

Abstract type for Mean-Variance optimisation formulations.

PortfolioOptimiser.SDSquaredFormulation — Type

abstract type SDSquaredFormulation <: SDFormulation end

Mean variance formulation will produce a JuMP.QuadExpr for the JuMP model's sd_risk.

The reason we have these is because there are multiple ways of defining the mean-variance optimisation and because NOC optimisations are only compatible with strictly convex risk functions, so they can only be performed with SimpleSD.

PortfolioOptimiser._ADD — Method

_ADD(x::AbstractVector)

Compute the Average Drawdown of uncompounded cumulative returns.

\[\begin{align} \mathrm{ADD_{a}}(\bm{X}) = \dfrac{1}{T} \sum\limits_{j=1}^{T} \mathrm{DD_{a}}(\bm{X}, j)\,. \end{align}\]

Where:

$\mathrm{DD_{a}}(\bm{X}, j)$ is the Drawdown of uncompounded cumulative returns at time $j$ as defined in _DaR.

Inputs

x: T×1 returns vector.

Outputs

r: risk.

PortfolioOptimiser._ADD_r — Method

_ADD_r(x::AbstractVector)

Compute the Average Drawdown of compounded cumulative returns.

\[\begin{align} \mathrm{ADD_{r}}(\bm{X}) = \dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{r}}(\bm{X},\, j)\,. \end{align}\]

Where:

$\mathrm{DD_{r}}(\bm{X},\, j)$ is the Drawdown of compounded cumulative returns at time $j$ as defined in _DaR_r.

Inputs

x: T×1 returns vector.

Outputs

r: risk.

PortfolioOptimiser._CDaR — Function

_CDaR(x::AbstractVector, α::Real = 0.05)

Compute the Conditional Drawdown at Risk of uncompounded cumulative returns.

\[\begin{align} \mathrm{CDaR_{a}}(\bm{X},\, \alpha) = \mathrm{DaR_{a}}(\bm{X},\, \alpha) + \dfrac{1}{\alpha T} \sum\limits_{j=0}^{T} \max\left(\mathrm{DD_{a}}(\bm{X},\, j) - \mathrm{DaR_{a}}(\bm{X},\, \alpha),\, 0 \right)\,. \end{align}\]

Where:

$\mathrm{DD_{a}}(\bm{X},\, j)$ is the Drawdown of uncompounded cumulative returns at time $j$ as defined in _DaR, and $\mathrm{DaR_{a}}(\bm{X},\, \alpha)$ the Drawdown at Risk of uncompounded cumulative returns as defined in _DaR.

Inputs

x: T×1 returns vector.

α: significance level, α ∈ (0, 1).

Warning

In-place sorts the input vector.

Outputs

r: risk.

PortfolioOptimiser._CDaR_r — Function

_CDaR_r(x::AbstractVector, α::Real = 0.05)

Compute the Conditional Drawdown at Risk of compounded cumulative returns.

\[\begin{align} \mathrm{CDaR_{r}}(\bm{X},\, \alpha) = \mathrm{DaR_{r}}(\bm{X},\, \alpha) + \dfrac{1}{\alpha T} \sum\limits_{j=0}^{T} \max\left(\mathrm{DD_{r}}(\bm{X},\, j) - \mathrm{DaR_{r}}(\bm{X},\, \alpha),\, 0 \right)\,. \end{align}\]

Where:

$\mathrm{DD_{r}}(\bm{X},\, j)$ is the Drawdown of compounded cumulative returns at time $j$ as defined in _DaR_r.

$\mathrm{DaR_{r}}(\bm{X},\, \alpha)$ the Drawdown at Risk of compounded cumulative returns as defined in _DaR_r.

Inputs

x: T×1 returns vector.

α: significance level, α ∈ (0, 1).

Warning

In-place sorts the input vector.

Outputs

r: risk.

PortfolioOptimiser._CVaR — Function

_CVaR(x::AbstractVector, α::Real = 0.05)

Compute the Conditional Value at Risk.

\[\begin{align} \mathrm{CVaR}(\bm{X},\, \alpha) = \mathrm{VaR}(\bm{X},\, \alpha) + \dfrac{1}{\alpha T} \sum\limits_{t=1}^{T} \max\left( -X_{t} - \mathrm{VaR}(\bm{X},\, \alpha),\, 0\right)\,. \end{align}\]

Where:

$\mathrm{VaR}(\bm{X},\, \alpha)$ is the Value at Risk as defined in _VaR.

Inputs

x: T×1 returns vector.

α: significance level, α ∈ (0, 1).

Warning

In-place sorts the input vector.

Outputs

r: risk.

PortfolioOptimiser._CVaRRG — Method

_CVaRRG(x::AbstractVector; alpha::Real = 0.05, beta::Real = alpha)

Compute the _CVaR Range.

Inputs

x: T×1 returns vector.

alpha: significance level of CVaR losses, alpha ∈ (0, 1).

beta: significance level of CVaR gains, beta ∈ (0, 1).

Warning

In-place sorts the input vector.

Outputs

r: risk.

PortfolioOptimiser._DaR — Function

_DaR(x::AbstractArray, α::Real = 0.05)

Compute the Drawdown at Risk of uncompounded cumulative returns.

\[\begin{align} \mathrm{DaR_{a}}(\bm{X},\, \alpha) &= \underset{j \in (0,\, T)}{\max} \left\{ \mathrm{DD_{a}}(\bm{X},\, j) \in \mathbb{R} : F_{\mathrm{DD}}\left(\mathrm{DD_{a}}(\bm{X},\, j)\right) < 1 - \alpha \right\}\\ \mathrm{DD_{a}}(\bm{X},\, j) &= \underset{t \in (0,\, j)}{\max}\left( \sum\limits_{i=0}^{t} X_{i} \right) - \sum\limits_{i=0}^{j} X_{i}\\ \mathrm{DD_{a}}(\bm{X}) &= \mathrm{DD_{a}}(\bm{X},\, j) \quad \forall j = 1,\,\ldots,\,T\,. \end{align}\]

Where:

$\mathrm{DD_{a}}(\bm{X})$ is the Drawdown of uncompounded cumulative returns.

$\mathrm{DD_{a}}(\bm{X},\, j)$ is the Drawdown of uncompounded cumulative returns at time $j$.

$\mathrm{DaR_{a}}(\bm{X},\, \alpha)$ the Drawdown at Risk of uncompounded cumulative returns.

Inputs

x: T×1 returns vector.

α: significance level, α ∈ (0, 1).

Warning

In-place sorts the input vector.

Outputs

r: risk.

PortfolioOptimiser._DaR_r — Function

_DaR_r(x::AbstractArray, α::Real = 0.05)

Compute the Drawdown at Risk of compounded cumulative returns.

\[\begin{align} \mathrm{DaR_{r}}(\bm{X},\, \alpha) &= \underset{j \in (0,\, T)}{\max} \left\{ \mathrm{DD_{r}}(\bm{X},\, j) \in \mathbb{R} : F_{\mathrm{DD}}\left(\mathrm{DD_{r}}(\bm{X},\, j)\right) < 1 - \alpha \right\}\\ \mathrm{DD_{r}}(\bm{X},\, j) &= \underset{t \in (0,\, j)}{\max}\left( \prod\limits_{i=0}^{t} \left(1+X_{i}\right) \right) - \prod\limits_{i=0}^{j} \left(1+X_{i}\right)\\ \mathrm{DD_{r}}(\bm{X}) &= \mathrm{DD_{r}}(\bm{X},\, j) \quad \forall j = 1,\,\ldots,\,T\,. \end{align}\]

Where:

$\mathrm{DD_{r}}(\bm{X})$ is the Drawdown of compounded cumulative returns.

$\mathrm{DD_{r}}(\bm{X},\, j)$ is the Drawdown of compounded cumulative returns at time $j$.

$\mathrm{DaR_{r}}(\bm{X},\, \alpha)$ the Drawdown at Risk of compounded cumulative returns.

Inputs

x: T×1 returns vector.

α: significance level, α ∈ (0, 1).

Warning

In-place sorts the input vector.

Outputs

r: risk.

PortfolioOptimiser._EDaR — Function

_EDaR(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05)

Compute the Entropic Drawdown at Risk of uncompounded cumulative returns.

\[\begin{align} \mathrm{EDaR_{a}}(\bm{X},\alpha) &= \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\mathrm{DD_{a}}(\bm{X}),\, z, \,\alpha)\right\}\,. \end{align}\]

Where:

$\mathrm{ERM}(\mathrm{DD_{a}}(\bm{X}),\, z, \,\alpha)$ is the Entropic Risk Measure as defined in ERM, using the Drawdown of uncompounded cumulative returns as defined in _DaR.

Inputs

x: T×1 returns vector.

solvers: abstract dict containing the a JuMP-compatible solver capable of solving 3D power cone problems.

α: significance level, α ∈ (0, 1).

Outputs

r: risk.

PortfolioOptimiser._EDaR_r — Function

_EDaR_r(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05)

Compute the Entropic Drawdown at Risk of compounded cumulative returns.

\[\begin{align} \mathrm{EDaR_{r}}(\bm{X},\alpha) &= \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\mathrm{DD_{r}}(\bm{X}),\, z, \,\alpha)\right\}\,. \end{align}\]

Where:

$\mathrm{ERM}(\mathrm{DD_{r}}(\bm{X}),\, z, \,\alpha)$ is the Entropic Risk Measure as defined in ERM, using the Drawdown of compounded cumulative returns as defined in _DaR_r.

Inputs

x: T×1 returns vector.

solvers: abstract dict containing the a JuMP-compatible solver capable of solving 3D power cone problems.

α: significance level, α ∈ (0, 1).

Outputs

r: risk.

PortfolioOptimiser._EVaR — Function

_EVaR(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05)

Compute the Entropic Value at Risk.

\[\begin{align} \mathrm{EVaR}(\bm{X},\alpha) = \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\bm{X},\, z, \,\alpha)\right\}\,. \end{align}\]

Where:

$\mathrm{ERM}(\bm{X},\, z, \,\alpha)$ is the entropic risk measure as defined in ERM.

Inputs

x: T×1 returns vector.

solvers: abstract dict containing the a JuMP-compatible solver capable of solving 3D power cone problems.

α: significance level, α ∈ (0, 1).

Outputs

r: risk.

PortfolioOptimiser._FLPM — Function

_FLPM(x::AbstractVector, r::Real = 0.0)

Compute the First Lower Partial Moment (Omega ratio).

\[\begin{align} \mathrm{FLPM}(\bm{X},\, r) = \dfrac{1}{T} \sum\limits_{t=1}^{T}\max\left(r - X_{t},\, 0\right)\,. \end{align}\]

Inputs

x: T×1 returns vector.

r: minimum return target.

Outputs

r: risk.

PortfolioOptimiser._GMD — Method

_GMD(x::AbstractVector)

Compute the Gini Mean Difference.

Inputs

x: T×1 returns vector.

Warning

In-place sorts the input vector.

Outputs

r: risk.

PortfolioOptimiser._Kurt — Function

_Kurt(x::AbstractVector, w::Union{AbstractWeights, Nothing} = nothing;
               scale::Bool = false)

Compute the Square Root Kurtosis.

\[\begin{align} \mathrm{Kurt}(\bm{X}) = \left(\dfrac{1}{T} \sum\limits_{t=1}^{T} \left( X_{t} - \mathbb{E}(\bm{X}) \right)^{4} \right)^{1/2}\,. \end{align}\]

Inputs

x: T×1 returns vector.

w: T×1 optional vector of weights for computing the expected return.

scale: if true divides by 2, used in risk_contribution.

Outputs

r: risk.

PortfolioOptimiser._MAD — Function

_MAD(x::AbstractVector, w::Union{AbstractWeights, Nothing} = nothing)

Compute the Mean Absolute Deviation.

\[\begin{align} \mathrm{MAD}(\bm{X}) = \dfrac{1}{T} \sum\limits_{t=1}^T \left\lvert X_{t} - \mathbb{E}(\bm{X}) \right\rvert\,. \end{align}\]

Inputs

x: T×1 returns vector.

w: T×1 optional vector of weights for computing the expected return.

Outputs

r: risk.

PortfolioOptimiser._MDD — Method

_MDD(x::AbstractVector)

Compute the Maximum Drawdown (Calmar ratio) of uncompounded cumulative returns.

\[\begin{align} \mathrm{MDD_{a}}(\bm{X}) = \max\mathrm{DD_{a}}(\bm{X})\,. \end{align}\]

Where:

$\mathrm{DD_{a}}(\bm{X})$ is the Drawdown of uncompounded cumulative returns as defined in _DaR.

Inputs

x: T×1 returns vector.

Outputs

r: risk.

PortfolioOptimiser._MDD_r — Method

_MDD_r(x::AbstractVector)

Compute the Maximum Drawdown of compounded cumulative returns.

\[\begin{align} \mathrm{MDD_{r}}(\bm{X}) = \max \mathrm{DD_{r}}(\bm{X})\,. \end{align}\]

Where:

$\mathrm{DD_{r}}(\bm{X})$ is the Drawdown of compounded cumulative returns as defined in _DaR_r.

Inputs

x: T×1 returns vector.

Outputs

r: risk.

PortfolioOptimiser._OWA — Method

_OWA(x::AbstractVector, w::AbstractVector)

Compute the Ordered Weight Array risk measure.

Inputs

w: T×1 precomputed vector of OWA weights. Can be computed with owa_gmd, owa_rg, owa_rcvar, owa_tg, owa_rtg, owa_l_moment or owa_l_moment_crm.

x: T×1 returns vector.

Warning

In-place sorts the input vector.

Outputs

r: risk.

PortfolioOptimiser._RG — Method

_RG(x::AbstractVector)

Compute the Range.

Inputs

x: T×1 returns vector.

Warning

In-place sorts the input vector.

Outputs

r: risk.

PortfolioOptimiser._RLDaR — Function

_RLDaR(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05, κ::Real = 0.3)

Compute the Relativistic Drawdown at Risk of uncompounded cumulative returns.

\[\begin{align} \mathrm{RLDaR_{a}}(\bm{X},\, \alpha,\, \kappa) = \mathrm{RRM}(\mathrm{DD_{a}}(\bm{X}),\, \alpha,\, \kappa)\,. \end{align}\]

Where:

$\mathrm{RRM}(\mathrm{DD_{a}}(\bm{X}),\, \alpha,\, \kappa)$ is the Relativistic Risk Measure as defined in RRM, using the Drawdown of uncompounded cumulative returns as defined in _DaR.

Inputs

x: T×1 returns vector.

solvers: abstract dict containing the a JuMP-compatible solver capable of solving 3D power cone problems.

α: significance level, α ∈ (0, 1).

κ: relativistic deformation parameter, κ ∈ (0, 1).

Outputs

r: risk.

PortfolioOptimiser._RLDaR_r — Function

_RLDaR_r(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05, κ::Real = 0.3)

Compute the Relativistic Drawdown at Risk of compounded cumulative returns.

\[\begin{align} \mathrm{RLDaR_{r}}(\bm{X},\, \alpha,\, \kappa) = \mathrm{RRM}(\mathrm{DD_{r}}(\bm{X}),\, \alpha,\, \kappa)\,. \end{align}\]

Where:

$\mathrm{RRM}(\mathrm{DD_{r}}(\bm{X}),\, \alpha,\, \kappa)$ is the Relativistic Risk Measure as defined in RRM, using the Drawdown of compounded cumulative returns as defined in _DaR_r.

Inputs

x: T×1 returns vector.

solvers: abstract dict containing the a JuMP-compatible solver capable of solving 3D power cone problems.

α: significance level, α ∈ (0, 1).

κ: relativistic deformation parameter, κ ∈ (0, 1).

Outputs

r: risk.

PortfolioOptimiser._RLVaR — Function

_RLVaR(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05, κ::Real = 0.3)

Compute the Relativistic Value at Risk.

\[\begin{align} \mathrm{RLVaR}(\bm{X},\, \alpha,\, \kappa) = \mathrm{RRM}(\bm{X},\, \alpha,\, \kappa)\,. \end{align}\]

Where:

$\mathrm{RRM}(\bm{X},\, \alpha,\, \kappa)$ is the Relativistic Risk Measure as defined in RRM.

Inputs

x: T×1 returns vector.

α: significance level, α ∈ (0, 1).

κ: relativistic deformation parameter, κ ∈ (0, 1).

Outputs

r: risk.

PortfolioOptimiser._SD — Method

_SD(w::AbstractVector, Σ::AbstractMatrix)

Compute the Standard Deviation. Square root of _Variance.

\[\begin{align} \mathrm{SD}(\bm{w},\, \mathbf{\Sigma}) = \left(\bm{w}^\intercal \, \mathbf{\Sigma} \, \bm{w}\right)^{1/2}\,. \end{align}\]

Inputs

w: vector of asset weights.

Σ: covariance matrix of asset returns.

Outputs

r: risk.

PortfolioOptimiser._SKurt — Function

_SKurt(x::AbstractVector, r::Real = 0.0,
        w::Union{AbstractWeights, Nothing} = nothing; scale::Bool = false)

Compute the Square Root Semi Kurtosis.

\[\begin{align} \mathrm{SKurt}(\bm{X}) = \left(\dfrac{1}{T} \sum\limits_{t=1}^{T} \min\left( X_{t} - \mathbb{E}(\bm{X}),\, r \right)^{4} \right)^{1/2}\,. \end{align}\]

Inputs

x: T×1 returns vector.

r: minimum return target.

w: T×1 optional vector of weights for computing the expected return.

scale: if true divides by 2, used in risk_contribution.

Outputs

r: risk.

PortfolioOptimiser._SLPM — Function

_SLPM(x::AbstractVector, r::Real = 0.0)

Compute the Second Lower Partial Moment (Sortino Ratio).

\[\begin{align} \mathrm{SLPM}(\bm{X},\, r) = \left(\dfrac{1}{T-1} \sum\limits_{t=1}^{T}\max\left(r - X_{t},\, 0\right)^{2}\right)^{1/2}\,. \end{align}\]

Inputs

x: T×1 returns vector.

r: minimum return target.

Outputs

r: risk.

PortfolioOptimiser._SSD — Function

_SSD(x::AbstractVector, r::Real = 0.0, w::Union{AbstractWeights, Nothing} = nothing)

Compute the Semi-Standard Deviation.

\[\begin{align} \mathrm{SSD}(\bm{X}) = \left(\dfrac{1}{T-1} \sum\limits_{t=1}^{T}\min\left(X_{t} - \mathbb{E}(\bm{X}),\, r\right)^{2}\right)^{1/2}\,. \end{align}\]

Inputs

x: T×1 returns vector.

r: minimum return target.

w: T×1 optional vector of weights for computing the expected return.

Outputs

r: risk.

PortfolioOptimiser._SVariance — Function

_SVariance(x::AbstractVector, r::Real = 0.0, w::Union{AbstractWeights, Nothing} = nothing)

Compute the Semi-Variance, this is the square of _SSD.

\[\begin{align} \mathrm{SVariance}(\bm{X}) = \dfrac{1}{T-1} \sum\limits_{t=1}^{T}\min\left(X_{t} - \mathbb{E}(\bm{X}),\, r\right)^{2}\,. \end{align}\]

Inputs

x: T×1 returns vector.

r: minimum return target.

w: T×1 optional vector of weights for computing the expected return.

Outputs

r: risk.

PortfolioOptimiser._Skew — Method

_Skew(w::AbstractVector, V::AbstractArray)

Compute the Quadratic Skewness/Semi Skewness.

\[\begin{align} \nu &= \bm{w}^{\intercal} \mathbf{V} \bm{w}\\ \end{align}\]

Where:

$\bm{w}$ is the vector of asset weights.

$\mathbf{V}$ is the sum of the symmetric negative spectral slices of coskewness or semicoskewness.

Inputs

w: N×1 vector of weights.

V: N×N matrix of sum of negative spectral slices of the coskewness or semi coskewness.

Outputs

r: risk.

PortfolioOptimiser._TG — Method

_TG(x::AbstractVector; alpha_i::Real = 0.0001, alpha::Real = 0.05, a_sim::Int = 100)

Compute the Tail Gini.

Inputs

x: T×1 returns vector.

alpha_i: start value of the significance level of CVaR losses, 0 < alpha_i < alpha < 1.

alpha: end value of the significance level of CVaR losses, alpha ∈ (0, 1).

a_sim: number of CVaRs to approximate the Tail Gini losses, a_sim > 0.

Warning

In-place sorts the input vector.

Outputs

r: risk.

PortfolioOptimiser._TGRG — Method

_TGRG(x::AbstractVector; alpha_i::Real = 0.0001, alpha::Real = 0.05, a_sim::Real = 100,
     beta_i::Real = alpha_i, beta::Real = alpha, b_sim::Integer = a_sim)

Compute the Tail Gini Range.

Inputs

x: T×1 returns vector.

alpha_i: start value of the significance level of CVaR losses, 0 < alpha_i < alpha < 1.

alpha: end value of the significance level of CVaR losses, alpha ∈ (0, 1).

a_sim: number of CVaRs to approximate the Tail Gini losses, a_sim > 0.

beta_i: start value of the significance level of CVaR gains, 0 < beta_i < beta < 1.

beta: end value of the significance level of CVaR gains, beta ∈ (0, 1).

b_sim: number of CVaRs to approximate the Tail Gini gains, b_sim > 0.

Warning

In-place sorts the input vector.

Outputs

r: risk.

PortfolioOptimiser._UCI — Method

_UCI(x::AbstractVector)

Compute the Ulcer Index of uncompounded cumulative returns.

\[\begin{align} \mathrm{UCI_{a}}(\bm{X}) = \left(\dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{a}}(\bm{X},\, j)^{2}\right)^{1/2}\,. \end{align}\]

Where:

$\mathrm{DD_{a}}(\bm{X},\, j)$ is the Drawdown of uncompounded cumulative returns at time $j$ as defined in _DaR.

Inputs

x: T×1 returns vector.

Outputs

r: risk.

PortfolioOptimiser._UCI_r — Method

_UCI_r(x::AbstractVector)

Compute the Ulcer Index of compounded cumulative returns.

\[\begin{align} \mathrm{UCI_{r}}(\bm{X}) = \left(\dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{r}}(\bm{X},\, j)^{2}\right)^{1/2}\,. \end{align}\]

Where:

$\mathrm{DD_{r}}(\bm{X},\, j)$ is the Drawdown of compounded cumulative returns at time $j$ as defined in _DaR_r.

Inputs

x: T×1 returns vector.

Outputs

r: risk.

PortfolioOptimiser._VaR — Function

_VaR(x::AbstractVector, α::Real = 0.05)

Compute the Value at Risk, used in _CVaR.

\[\begin{align} \mathrm{VaR}(\bm{X},\, \alpha) = -\underset{t \in (0,\, T)}{\inf} \left\{ X_{t} \in \mathbb{R} : F_{\bm{X}}(X_{t}) > \alpha \right\}\,. \end{align}\]

Inputs

x: T×1 returns vector.

α: significance level, α ∈ (0, 1).

Warning

In-place sorts the input vector.

Outputs

r: risk.

PortfolioOptimiser._Variance — Method

_Variance(w::AbstractVector, Σ::AbstractMatrix)

Compute the Variance. Square of _SD.

\[\begin{align} \mathrm{Variance}(\bm{w},\, \mathbf{\Sigma}) = \bm{w}^\intercal \, \mathbf{\Sigma}\, \bm{w}\,. \end{align}\]

Inputs

w: vector of asset weights.

Σ: covariance matrix of asset returns.

Outputs

r: risk.

PortfolioOptimiser._WR — Method

_WR(x::AbstractVector)

Compute the Worst Realisation or Worst Case Scenario.

\[\begin{align} \mathrm{WR}(\bm{X}) = -\min(\bm{X})\,. \end{align}\]

Inputs

x: T×1 returns vector.

Outputs

r: risk.

PortfolioOptimiser._dVar — Method

_dVar(x::AbstractVector)

Compute the Brownian distance variance.

\[\begin{align} \mathrm{dVar}(\bm{X}) &= \mathrm{dCov}(\bm{X},\, \bm{X}) = \dfrac{1}{T^{2}} \sum\limits_{i=1}^{T}\sum\limits_{j=1}^{T} A_{i,\,j}^2\\ \mathrm{dCov}(\bm{X},\, \bm{Y}) &= \dfrac{1}{T^{2}} \sum\limits_{i=1}^{T} \sum\limits_{j=1}^{T} A_{i,\,j} B_{i,\,j}\\ A_{i,\,j} &= a_{i,\,j} - \bar{a}_{i\,.} - \bar{a}_{.\,j} + \bar{a}_{.\,.}\\ B_{i,\,j} &= b_{i,\,j} - \bar{b}_{i\,.} - \bar{b}_{.\,j} + \bar{b}_{.\,.}\\ a_{i,\,j} &= \lVert X_{i} - X_{j} \rVert_{2}, \quad \forall i,\, j = 1,\, \ldots ,\, T\\ b_{i,\,j} &= \lVert Y_{i} - Y_{j} \rVert_{2}, \quad \forall i,\, j = 1,\, \ldots ,\, T\,. \end{align}\]

where:

$\bm{X}$ and $\bm{Y}$ are random variables, they are equal in this case as they are the portfolio returns.

$a_{i,\,j}$ and $b_{i,\,j}$ are entries of a distance matrix where $i$ and $j$ are points in time. Each entry is defined as the Euclidean distance $\lVert \cdot \rVert_{2}$ between the value of the random variable at time $i$ and its value at time $j$.

$\bar{a}_{i,\,\cdot}$ and $\bar{b}_{i,\,\cdot}$ are the $i$-th row means of their respective matrices.

$\bar{a}_{\cdot,\,j}$ and $\bar{b}_{\cdot,\,j}$ are the $j$-th column means of their respective matrices.

$\bar{a}_{\cdot,\,\cdot}$ and $\bar{b}_{\cdot,\,\cdot}$ are the grand means of their respective matrices.

$A_{i,\,j}$ and $B_{i,\,j}$ are doubly centered distances.

Inputs

x: T×1 returns vector.

Outputs

r: risk.