Generalities

This section details general types, structures, and constants that play a supporting role in defining risk measures.

Abstract types

PortfolioOptimiser's risk measures are implemented using Julia's type hierarchy. This makes it easy to add new ones by implementing the relevant types and methods.

By properly subtyping and following the implementation instructions of each abstract type, it's possible to define risk measures that will be entirely compatible with current functionality, without the need for extending method definitions and/or subtyping any of the myriad of abstract types that allow users to define custom behaviour within the library.

PortfolioOptimiser.AbstractRiskMeasure — Type

abstract type AbstractRiskMeasure end

Supertype for all risk measaures.

See also: RiskMeasure, HCRiskMeasure, NoOptRiskMeasure.

PortfolioOptimiser.RiskMeasure — Type

abstract type RiskMeasure <: AbstractRiskMeasure end

Supertype for risk measures compatible with optimisations which accept risk measures.

See also: RiskMeasureSolvers, RiskMeasureSigma, RiskMeasureSkew, RiskMeasureOWA, RiskMeasureMu, RiskMeasureTarget, RMSettings, expected_risk, set_rm.

Implementation

To ensure a risk measure can be used any of the above optimisation types, it must abide by a few rules.

Implement: Base.iterate, Base.Symbol, Base.length, Base.getindex, and Base.view.

struct MyRiskMeasure <: RiskMeasure
    # Properties of MyRiskMeasure
end

Base.iterate(S::MyRiskMeasure, state = 1) = state > 1 ? nothing : (S, state + 1)
function Base.String(s::MyRiskMeasure)
    return "MyRiskMeasure"
end
function Base.Symbol(::MyRiskMeasure)
    return Symbol("MyRiskMeasure")
end
function Base.length(::MyRiskMeasure)
    return 1
end
function Base.getindex(S::MyRiskMeasure, ::Any)
    return S
end
function Base.view(S::MyRiskMeasure, ::Any)
    return S
end

Include a settings::RMSettings property, RMSettings.

struct MyRiskMeasure <: RiskMeasure
    # Properties of MyRiskMeasure
    settings::RMSettings
end

Implement your measure's risk calculation method, expected_risk. This will let the library use the risk function everywhere it needs to.

function expected_risk(my_risk::MyRiskMeasure, w::AbstractVector; kwargs...)
    # Risk measure calculation
end

A scalar JuMP model implementation of set_rm. If appropriate, a vector equivalent.

# The scalar function.
function PortfolioOptimiser.set_rm(port, rm::MyRiskMeasure, type::Union{Trad, RB, NOC};
                                   kwargs...)
    # Get optimisation model.
    model = port.model

    ###
    # Variables, constraints, expressions, etc.
    ###

    # Define the risk expression for MyRiskMeasure
    @expression(model, MyRiskMeasure_risk, ...)

    # Define the key name for the upper bound.
    # The upper bound key will be `Symbol("MyRiskMeasure_risk_ub")`
    ub_key = "MyRiskMeasure_risk"

    # Set the upper bound on MyRiskMeasure_risk. 
    # If isinf(rm.settings.ub), no upper bound will be set.
    set_rm_risk_upper_bound(type, model, MyRiskMeasure_risk, rm.settings.ub, ub_key)

    # Add the risk to the risk expression.
    # If rm.settings.flag == true, MyRiskMeasure_risk will be added to the vector of risks.
    # This means it will form part of the objective function when the objective function 
    # is MinRisk, Utility or Sharpe.
    # If rm.settings.flag == false, MyRiskMeasure_risk will only be used as a risk 
    # constraint.
    set_risk_expression(model, MyRiskMeasure_risk, rm.settings.scale, rm.settings.flag)

    return nothing
end

# Vector equivalent if it's possible to provide multiple instances of MyRiskMeasure.
function PortfolioOptimiser.set_rm(port, rms::AbstractVector{<:MyRiskMeasure},
                                   type::Union{Trad, RB, NOC}; kwargs...)

    # Get optimisation model.
    model = port.model
    count = length(rms)

    ###
    # Variables, constraints, expressions, that can be initialised beforehand etc.
    ###

    # Define the risk expression for MyRiskMeasure.
    # It can also be defined in the iteration using the 
    # `model[key] = @expression(model, ..., ...) `
    # construct, then reference the `model[key]` expression.
    # Or use its anonymous version
    # `key = @expression(model, ..., ...)` 
    # and reference the `key` expression.
    @expression(model, MyRiskMeasure_risk[1:count], ...)

    for (i, rm) ∈ pairs(rms)
        ###
        # Variables, constraints, expressions, that must be set during each iteration.
        # If you want to register them in the model use:
        # `model[constraint_key] = @constraint(model, ..., ...)`
        # `model[expression_key] = @expression(model, ..., ...)`
        # `model[variable_key] = @variable(model, ..., ...)`
        # If you want them to be anonymous (i.e. not registered in the model) use:
        # `constraint_key = @constraint(model, ..., ...)`
        # `expression_key = @expression(model, ..., ...)`
        # `variable_key = @variable(model, ..., ...)`
        # If they were defined outside of the loop like MyRiskMeasure_risk 
        # but need to be modified then use:
        # `add_to_expression!(MyRiskMeasure_risk[i], ...).`
        ###

        # Define the key name for the upper bound.
        # The upper bound key will be `Symbol("MyRiskMeasure_risk_$(i)_ub")`
        ub_key = "MyRiskMeasure_risk_$(i)"

        # Set the upper bound on MyRiskMeasure_risk[i]. 
        # If isinf(rm.settings.ub), no upper bound will be set.
        set_rm_risk_upper_bound(type, model, MyRiskMeasure_risk[i], rm.settings.ub, ub_key)

        # Add the risk to the risk expression.
        # If rm.settings.flag == true, MyRiskMeasure_risk[i] will be added to the vector of
        # risks. This means it will form part of the objective function when the objective 
        # function is MinRisk, Utility or Sharpe.
        # If rm.settings.flag == false, MyRiskMeasure_risk[i] will only be used as a risk 
        # constraint.
        set_risk_expression(model, MyRiskMeasure_risk[i], rm.settings.scale,
                            rm.settings.flag)
    end

    return nothing
end

If a risk measure is to be compatible with hierarchical optimisations that take risk measures as parameters, and it contains a properties which can/must be indexed/computed per asset, like a vector or matrix, it must implement set_custom_hc_rm! and unset_custom_hc_rm! which dispatches on the custom risk measure.

struct MyRiskMeasure{T1, T2, T3} <: RiskMeasure
    # Properties containing asset information (computable or indexable).
    indexable_vector::Vector{T1}
    indexable_matrix::Matrix{T1}
    computable_vector::Vector{T1}
    computable_matrix::Matrix{T1}
    computable_vector_args::T2
    computable_matrix_args::T3
end

# We have some computable properties, so we need to define the function to do so.
function compute_MyRiskMeasure_vec_mtx!(rm::MyRiskMeasure, args...)
    # Compute vector and matrix
    new_computable_vector = ...
    new_computable_matrix = ...

    rm.computable_vector = new_computable_vector
    rm.computable_matrix = new_computable_matrix

    return nothing
end

# port is the portfolio, sigma is the covariance matrix, cluster are the indices defining the cluster.
function set_custom_hc_rm!(rm::MyRiskMeasure, port, sigma, cluster)
    old_i_vector = rm.indexable_vector
    old_i_matrix = rm.indexable_matrix
    old_c_vector = rm.computable_vector
    old_c_matrix = rm.computable_matrix

    ###
    ###
    # These can be placed inside if statements that condition the indexing
    rm.indexable_vector = rm.indexable_vector[cluster]
    rm.indexable_matrix = rm.indexable_matrix[cluster, cluster]
    ###
    ###

    compute_MyRiskMeasure_vec_mtx!(rm, port, sigma, cluster)

    return Tuple(old_i_vector, old_i_matrix, old_c_vector, old_c_matrix)
end
function unset_custom_hc_rm!(rm::MyRiskMeasure, old_custom)
    rm.indexable_vector = old_custom[1]
    rm.indexable_matrix = old_custom[2]
    rm.computable_vector = old_custom[3]
    rm.computable_matrix = old_custom[4]

    return nothing
end

Similarly, if the risk measure is to be used in NCO optimisations, and it contains a properties which can/must be indexed/computed per asset, like a vector or matrix, it must implement pre_modify_intra_port!, post_modify_intra_port!, reset_intra_port!, pre_modify_inter_port!, post_modify_inter_port!, reset_inter_port!, which dispatch on custom structures that. The functions can then check for the custom risk measure and modify it as in the previous bullet point. See the function's docstrings for explanations on their arguments and use.

# Structures for dispatching on.
struct MyPreModify <: AbstractNCOModify
    # Custom properties.
end
struct MyPostModify <: AbstractNCOModify
    # Custom properties.
end

# Procedures for computing or modifying risk measures for the internal optimisations.
# Each cluster is treated as a single portfolio.
function pre_modify_intra_port!(pre_modify::MyPreModify, intra_port, internal_args, i,
                                cluster, cidx, idx_sq, Nc, special_rm_idx)
    # Modify intra-cluster portfolio pre computation of statistics.
    return pre_mod_output
end
function post_modify_intra_port!(post_modify::MyPostModify, intra_port, internal_args, i,
                                 cluster, cidx, idx_sq, Nc, special_rm_idx)
    # Modify intra-cluster portfolio post computation of statistics.
    return post_mod_output
end
function reset_intra_port!(pre_modify::MyPreModify, pre_mod_output,
                           post_modify::MyPostModify, post_mod_output, intra_port,
                           internal_args, i, cluster, cidx, idx_sq, Nc, special_rm_idx)
    # Reset any changes done to the optimisation arguments.
    return nothing
end

# Procedures for computing or modifying risk measures for the external optimisation.
# Each cluster is turned into a synthetic asset and a portfolio optimisation is
# performed on them.
function pre_modify_inter_port!(pre_modify::MyPreModify, inter_port, wi, external_args,
                                special_rm_idx)
    # Modify inter-cluster portfolio pre computation of statistics.
    return pre_mod_output
end
function post_modify_inter_port!(post_modify::MyPostModify, inter_port, wi, external_args,
                                 special_rm_idx)
    # Modify inter-cluster portfolio post computation of statistics.
    return post_mod_output
end
function reset_inter_port!(pre_modify::MyPreModify, pre_mod_output,
                           post_modify::MyPostModify, post_mod_output, inter_port, wi,
                           external_args, special_rm_idx)
    # Reset any changes done to the optimisation arguments.
    return nothing
end

# The NCO optimisation type would then be defined like so.
type = NCO(;
           internal = NCOArgs(; type = Trad(; rm = MyRiskMeasure()),
                              pre_modify = MyPreModify(), post_modify = MyPostModify()),
           # In case the external optimisation is to use something different.
           external = ...)

PortfolioOptimiser.HCRiskMeasure — Type

abstract type HCRiskMeasure <: AbstractRiskMeasure end

Supertype for risk measures compatible with optimisations which accept risk measures and do not use JuMP models.

See also: HCRiskMeasureSolvers, HCRiskMeasureMu, HCRiskMeasureTarget, HCRMSettings, expected_risk.

Implementation

To ensure a risk measure can be used by the above optimisation types, it must abide by a few rules.

Implement: Base.iterate, Base.Symbol, Base.length, Base.getindex, and Base.view.

struct MyHCRiskMeasure <: HCRiskMeasure
    # Properties of MyHCRiskMeasure
end

Base.iterate(S::MyHCRiskMeasure, state = 1) = state > 1 ? nothing : (S, state + 1)
function Base.String(s::MyHCRiskMeasure)
    return "MyHCRiskMeasure"
end
function Base.Symbol(::MyHCRiskMeasure)
    return Symbol("MyHCRiskMeasure")
end
function Base.length(::MyHCRiskMeasure)
    return 1
end
function Base.getindex(S::MyHCRiskMeasure, ::Any)
    return S
end
function Base.view(S::MyHCRiskMeasure, ::Any)
    return S
end

Include a settings::HCRMSettings property, HCRMSettings.

struct MyHCRiskMeasure <: HCRiskMeasure
    # Properties of MyHCRiskMeasure
    settings::HCRMSettings
end

Implement your measure's risk calculation method, expected_risk. This will let the library use the risk function everywhere it needs to.

function expected_risk(my_risk::MyHCRiskMeasure, w::AbstractVector; kwargs...)
    # Risk measure calculation
end

If a risk measure is to be compatible with hierarchical optimisations that take risk measures as parameters, and it contains a properties which can/must be indexed/computed per asset, like a vector or matrix, it must implement set_custom_hc_rm! and unset_custom_hc_rm! which dispatches on the custom risk measure.

struct MyHCRiskMeasure{T1, T2, T3} <: HCRiskMeasure
    # Properties containing asset information (computable or indexable).
    indexable_vector::Vector{T1}
    indexable_matrix::Matrix{T1}
    computable_vector::Vector{T1}
    computable_matrix::Matrix{T1}
    computable_vector_args::T2
    computable_matrix_args::T3
end

# We have some computable properties, so we need to define the function to do so.
function compute_MyRiskMeasure_vec_mtx!(rm::MyHCRiskMeasure, args...)
    # Compute vector and matrix
    new_computable_vector = ...
    new_computable_matrix = ...

    rm.computable_vector = new_computable_vector
    rm.computable_matrix = new_computable_matrix

    return nothing
end

# port is the portfolio, sigma is the covariance matrix, cluster are the indices defining the cluster.
function set_custom_hc_rm!(rm::MyHCRiskMeasure, port, sigma, cluster)
    old_i_vector = rm.indexable_vector
    old_i_matrix = rm.indexable_matrix
    old_c_vector = rm.computable_vector
    old_c_matrix = rm.computable_matrix

    ###
    ###
    # These can be placed inside if statements that condition the indexing
    rm.indexable_vector = rm.indexable_vector[cluster]
    rm.indexable_matrix = rm.indexable_matrix[cluster, cluster]
    ###
    ###

    compute_MyRiskMeasure_vec_mtx!(rm, port, sigma, cluster)

    return Tuple(old_i_vector, old_i_matrix, old_c_vector, old_c_matrix)
end
function unset_custom_hc_rm!(rm::MyHCRiskMeasure, old_custom)
    rm.indexable_vector = old_custom[1]
    rm.indexable_matrix = old_custom[2]
    rm.computable_vector = old_custom[3]
    rm.computable_matrix = old_custom[4]

    return nothing
end

PortfolioOptimiser.NoOptRiskMeasure — Type

abstract type NoOptRiskMeasure <: AbstractRiskMeasure end

Abstract type for risk measures that cannot be used in optimisations but can be used as performance measurements via expected_risk. This can be for two reasons:

They can be negative, therefore unsuitable for optimisations that accept risk measures and do not use JuMP models.

They have no known optimisation formulation, therefore unsuitable for optimisations that accept risk measures and use JuMP models.

See also: expected_risk.

Implementation

Implement: expected_risk for the type. This will let the library use the risk function everywhere it needs to.

struct MyNoOptRiskMeasure <: NoOptRiskMeasure
    # Properties of MyNoOptRiskMeasure
end

function expected_risk(my_risk::MyNoOptRiskMeasure, w::AbstractVector; kwargs...)
    # Risk measure calculation
end

PortfolioOptimiser.RiskMeasureSigma — Type

abstract type RiskMeasureSigma <: RiskMeasure end

Abstract type for subtyping RiskMeasure which can use an N×N covariance matrix.

See also: RiskMeasure, Variance, SD, WCVariance.

Implementation

Concrete subtypes must contain the following properties:

sigma::Union{<:AbstractMatrix, Nothing}: property to store an N×N covariance matrix.

Ideally perform any necessary validation checks at instantiation and setproperty!.

PortfolioOptimiser.RiskMeasureMu — Type

abstract type RiskMeasureMu <: RiskMeasure end

Abstract type for subtyping RiskMeasure which can use a T×1 AbstractWeights vector for computing the weighted mean, and an N×1 expected returns vector.

See also: RiskMeasure, RiskMeasureTarget, MAD, SSD, SVariance, Kurt, SKurt.

Implementation

Concrete subtypes must contain the following properties:

mu::Union{<:AbstractVector, Nothing}: property to store an N×1 expected returns vector.

w::Union{<:AbstractWeights, Nothing}: property to store a T×1 AbstractWeights vector for computing the weighted mean.

Ideally perform any necessary validation checks at instantiation and setproperty!.

PortfolioOptimiser.HCRiskMeasureMu — Type

abstract type HCRiskMeasureMu <: RiskMeasure end

Abstract type for subtyping HCRiskMeasure which can use a T×1 AbstractWeights vector for computing the weighted mean, and an N×1 expected returns vector.

See also: HCRiskMeasure, FTCM.

Implementation

Concrete subtypes must contain the following properties:

mu::Union{<:AbstractVector, Nothing}: property to store an N×1 expected returns vector.

w::Union{<:AbstractWeights, Nothing}: property to store a T×1 AbstractWeights vector for computing the weighted mean.

Ideally perform any necessary validation checks at instantiation and setproperty!.

PortfolioOptimiser.NoOptRiskMeasureMu — Type

abstract type NoOptRiskMeasureMu <: RiskMeasure end

Abstract type for subtyping NoOptRiskMeasure which can use a T×1 AbstractWeights vector for computing the weighted mean, and an N×1 expected returns vector.

See also: NoOptRiskMeasure, TCM.

Implementation

Concrete subtypes must contain the following properties:

w::Union{<:AbstractWeights, Nothing}: property to store a T×1 AbstractWeights vector for computing the weighted mean.

mu::Union{<:AbstractVector, Nothing}: property to store an N×1 expected returns vector.

Ideally perform any necessary validation checks at instantiation and setproperty!.

PortfolioOptimiser.RiskMeasureTarget — Type

abstract type RiskMeasureTarget <: RiskMeasureMu end

Abstract type for subtyping RiskMeasure which can use a scalar or an N×1 vector specifying the minimum return threshold for classifying downside returns, a T×1 AbstractWeights vector for computing the weighted mean, and an N×1 expected returns vector.

See also: RiskMeasure, FLPM, SSD, SVariance.

Implementation

Concrete subtypes must contain the following properties:

target::Union{<:Real, <:AbstractVector{<:Real}, Nothing}: scalar or N×1 minimum return threshold for classifying downside returns. Only returns equal to or below this value are considered in the calculation. Must be in the same frequency as the returns.

w::Union{<:AbstractWeights, Nothing}: property to store a T×1 AbstractWeights vector for computing the weighted mean.

mu::Union{<:AbstractVector, Nothing}: property to store an N×1 expected returns vector.

Ideally perform any necessary validation checks at instantiation and setproperty!.

PortfolioOptimiser.HCRiskMeasureTarget — Type

abstract type HCRiskMeasureTarget <: RiskMeasureMu end

Abstract type for subtyping HCRiskMeasure which can use a scalar or N×1 vector specifying the minimum return threshold for classifying downside returns, a T×1 AbstractWeights vector for computing the weighted mean, and an N×1 expected returns vector.

See also: HCRiskMeasure, TLPM, FTLPM.

Implementation

Concrete subtypes must contain the following properties:

target::Union{<:Real, <:AbstractVector{<:Real}, Nothing}: scalar or N×1 minimum return threshold for classifying downside returns. Only returns equal to or below this value are considered in the calculation. Must be in the same frequency as the returns.

w::Union{<:AbstractWeights, Nothing}: property to store a T×1 AbstractWeights vector for computing the weighted mean.

mu::Union{<:AbstractVector, Nothing}: property to store an N×1 expected returns vector.

Ideally perform any necessary validation checks at instantiation and setproperty!.

PortfolioOptimiser.RiskMeasureSolvers — Type

abstract type RiskMeasureSolvers <: RiskMeasure end

Abstract type for subtyping RiskMeasure for which computing the risk requires solving an optimisation model.

See also: RiskMeasure, EVaR, EDaR, RLVaR, RLDaR.

Implementation

Concrete subtypes must contain the following properties:

solvers::Union{Nothing, PortOptSolver, <:AbstractVector{PortOptSolver}}: (optional) instance or AbstractVector of PortOptSolver.

Ideally perform any necessary validation checks at instantiation and setproperty!.

PortfolioOptimiser.HCRiskMeasureSolvers — Type

abstract type HCRiskMeasureSolvers <: HCRiskMeasure end

Abstract type for subtyping HCRiskMeasure for which computing the risk requires solving an optimisation model.

See also: HCRiskMeasure, EDaR_r, RLDaR_r.

Implementation

Concrete subtypes must contain the following properties:

solvers::Union{Nothing, PortOptSolver, <:AbstractVector{PortOptSolver}}: (optional) instance or AbstractVector of PortOptSolver.

Ideally perform any necessary validation checks at instantiation and setproperty!.

PortfolioOptimiser.RiskMeasureOWA — Type

abstract type RiskMeasureOWA <: RiskMeasure end

Abstract type for subtyping RiskMeasure which is implemented via an Ordered Weight Array formulation.

See also: RiskMeasure, GMD, TG, TGRG, OWA.

Implementation

Concrete subtypes must contain the following properties:

formulation::OWAFormulation: property to store the formulation dispatch type.

Ideally perform any necessary validation checks at instantiation and setproperty!.

PortfolioOptimiser.RiskMeasureSkew — Type

abstract type RiskMeasureSkew <: RiskMeasure end

Abstract type for subtyping RiskMeasure which can use an N×N² coskew matrix and N×N matrix that stores the sum of the symmetric negative spectral slices of the coskewness.

See also: RiskMeasure, NQSkew, NQSSkew.

Implementation

Concrete subtypes must contain the following properties:

skew::Union{<:AbstractMatrix, Nothing}: property to store an N×N² coskew matrix.

V::Union{<:AbstractMatrix, Nothing}: property to store an N×N matrix that stores the sum of the symmetric negative spectral slices of the coskewness.

Ideally perform any necessary validation checks at instantiation and setproperty!.

Constants

It is useful to group risk measures by their properties as well as their types. By defining constants which group various risk measure types with certain characteristics, we can facilitate their use in the many different optimisation types offered by PortfolioOptimiser. This makes adding new features easier and simplifies the definition of customised behaviour.

PortfolioOptimiser.RMSolvers — Type

const RMSolvers = Union{RiskMeasureSolvers, HCRiskMeasureSolvers}

Risk measures that use solvers.

See also: RiskMeasureSolvers, HCRiskMeasureSolvers.

PortfolioOptimiser.RMSigma — Type

const RMSigma = Union{RiskMeasureSigma}

Risk measures that use sigma.

See also: RiskMeasureSigma.

PortfolioOptimiser.RMSkew — Type

const RMSkew = Union{RiskMeasureSkew}

Risk measures that use skew and V.

See also: RiskMeasureSkew.

PortfolioOptimiser.RMOWA — Type

const RMOWA = Union{RiskMeasureOWA}

Risk measures that use Ordered Weight Array formulations.

See also: RiskMeasureOWA.

PortfolioOptimiser.RMMu — Type

const RMMu = Union{RiskMeasureMu, HCRiskMeasureMu, NoOptRiskMeasureMu}

Risk measures that use mu and w.

See also: RiskMeasureMu, HCRiskMeasureMu, NoOptRiskMeasureMu, calc_ret_mu.

PortfolioOptimiser.RMTarget — Type

const RMTarget = Union{RiskMeasureTarget, HCRiskMeasureTarget}

Risk measures that use target, mu and w.

See also: RiskMeasureTarget, HCRiskMeasureTarget, calc_target_ret_mu.

Settings

Portfolio Optimisation has typically limited itself to using a single risk measure. However, some risk measures can/must use internal parameters in order to compute the risk. There is nothing preventing us from using multiple measures simultaneously, or even multiple instances of the same risk measure with different hyperparameters. This can be achieved via multiple objective optimisation (not currently supported, but may in the future), or a scalarisation procedure PortfolioOptimiser.AbstractScalarisation. Each instance's contribution to the overall risk expression is tunable via a weight parameter (called scale for disambiguation purposes).

For risk measures that can be used in optimisations using JuMP models, we must also provide a way to define their upper bound, and whether or not they should be included in the risk expression or used only for setting a risk upper bound.

For these purposes, we have two special structures that need to be included in risk measures that are to be used in optimisations. The choice of which one to use depends on what optimisation types the risk measure is meant to be compatible with.

PortfolioOptimiser.AbstractRMSettings — Type

abstract type AbstractRMSettings end

Abstract type for subtyping risk measure settings.

See also: RMSettings, HCRMSettings.

PortfolioOptimiser.RMSettings — Type

mutable struct RMSettings{T1 <: Real, T2 <: Real} <: AbstractRMSettings

Configuration settings for concrete subtypes of RiskMeasure. Having this property makes it possible for a risk measure to be used in any optimisation types that take risk measures as parameters.

See also: RiskMeasure, AbstractScalarisation, expected_risk.

Keyword Arguments

In optimisations which take risk measures and use JuMP models

flag::Bool = true:

If true: it is included in the optimisation's risk vector.

If false: it is not included in the optimisation's risk vector, used when you want to constrain the upper bound of a risk measure without having that risk measure appear in the MinRisk, Utility, or Sharpe objective measures.

scale::T1 = 1.0: weight parameter of the risk measure in the AbstractScalarisation method being used.

ub::T2 = Inf: upper bound risk constraint.

In optimisations which take risk measures and do not use JuMP models

flag::Bool = true: no effect, the risk cannot be bounded in these optimisations.

scale::T1 = 1.0: weight parameter of the risk measure in the AbstractScalarisation method being used.

ub::T2 = Inf: no effect, the risk cannot be bounded in these optimisations.

Examples

PortfolioOptimiser.HCRMSettings — Type

mutable struct HCRMSettings{T1 <: Real} <: AbstractRMSettings

Configuration settings for concrete subtypes of HCRiskMeasure.

See also: HCRiskMeasure, AbstractScalarisation, expected_risk.

Keyword Arguments

scale::T1 = 1.0: weight parameter of the risk measure in the AbstractScalarisation method being used.

Examples