Specifics

This section details types, structures, and constants that play a supporting role in defining specific risk measures.

RMOWA

Certain risk measures make use of Ordered Weight Array formulations in optimisations which use JuMP models. PortfolioOptimiser implements two formulations.

An exact but expensive formulation.

An approximate, tunable one based on PowerCone constraints.

We therefore provide a structure for dispatching the exact method, and a structure for tuning and dispatching the approximate one.

PortfolioOptimiser.OWAFormulation — Type

abstract type OWAFormulation end

Abstract type for subtyping Ordered Weight Array formulations in optimisations which use JuMP models.

See also: OWAExact, OWAApprox, RiskMeasureOWA.

PortfolioOptimiser.OWAExact — Type

struct OWAExact <: OWAFormulation end

Type for dispatching the exact formulation of Ordered Weight Array risk measures.

See also: OWAApprox, RiskMeasureOWA.

PortfolioOptimiser.OWAApprox — Type

mutable struct OWAApprox{T1 <: AbstractVector{<:Real}} <: OWAFormulation

Type for dispatching and tuning the approximate formulation of Ordered Weight Array risk measures.

See also: OWAExact, RiskMeasureOWA.

Keyword Arguments

p::T1 = Float64[2, 3, 4, 10, 50]: vector of the p-norm orders to be used in the approximation.

Behaviour

Uses PowerCone constraints.

Examples

Variance & SVariance

PortfolioOptimiser.VarianceFormulation — Type

abstract type VarianceFormulation end

Abstract type for implementing various formulations of the Variance and SVariance in optimisations which use JuMP models.

Variance

If either network_adj or cluster_adj property of the Portfolio instance is SDP, the formulation has no effect because this constraint type requires a PSDCone formulation of the variance.

See also: Variance, Quad, SOC, Portfolio, SDP.

PortfolioOptimiser.Quad — Type

struct Quad <: VarianceFormulation end

Explicit quadratic formulation for the Variance and SVariance as an optimisation model.

See also: VarianceFormulation, SOC, RSOC, Variance, SVariance.

General Behaviour

No additional variables or constraints introduced.

Performance may degrade for large portfolios.

Variance

\[\begin{align} \underset{\bm{w}}{\mathrm{opt}} &\qquad \bm{w}^\intercal \mathbf{\Sigma} \bm{w}\\ \end{align}\]

Where:

$\bm{w}$: is the N×1 vector of asset weights.

$\mathbf{\Sigma}$: is the N×N asset covariance matrix.

Behaviour

Produces a QuadExpr risk expression variance_risk = dot(w, sigma, w).

Examples

SVariance

\[\begin{align} \underset{\bm{w}}{\mathrm{opt}} &\qquad \dfrac{\bm{y} \cdot \bm{y}}{T-1}\nonumber\\ \textrm{s.t.} &\qquad \bm{y} \geq \bm{0}\nonumber\\ &\qquad (\mathbf{X} - \bm{\mu}^\intercal) \bm{w} \geq -\bm{y} \end{align}\]

Where:

$\bm{w}$: is the N×1 vector of asset weights.

$\mathbf{X}$: is the T×N matrix of asset returns.

$\bm{\mu}$: is the N×1 vector of expected returns.

$\bm{w}$: is the N×1 vector of asset weights.

$\bm{y}$: is the T×1 vector of deviations from the expected portfolio return that meets the minimum return threshold.

Behaviour

Produces a QuadExpr risk expression svariance_risk = dot(svariance, svariance)/(T-1).

PortfolioOptimiser.SOC — Type

struct SOC <: VarianceFormulation end

Second-Order Cone (SOC) formulation for the Variance and SVariance. Formulates the quadratic variance/semi variance expression using a SecondOrderCone cone constraint.

See also: VarianceFormulation, Quad, RSOC, Variance, SVariance.

General Behaviour

Uses SecondOrderCone constraints.

Often more numerically stable than direct quadratic formulation.

Better scaling properties for large portfolios.

Compatible with specialised conic solvers.

Particularly effective for large-scale problems.

Variance

\[\begin{align} \underset{\bm{w}}{\mathrm{opt}} &\qquad \sigma^2\nonumber\\ \textrm{s.t.} &\qquad \left\lVert \mathbf{G} \bm{w} \right\rVert_{2} \leq \sigma\,. \end{align}\]

Where:

$\bm{w}$: is the N×1 vector of asset weights.

$\mathbf{G}$: is a suitable factorisation of the N×N covariance matrix, such as the square root matrix, or the Cholesky factorisation.

$\sigma^2$: is the portfolio variance.

$\lVert \cdot \rVert_{2}$: is the L-2 norm, which is modelled as an SecondOrderCone.

Behaviour

Produces a QuadExpr risk expression variance_risk = dev^2.

Examples

SVariance

\[\begin{align} \underset{\bm{w}}{\mathrm{opt}} &\qquad \dfrac{v^{2}}{T-1}\nonumber\\ \textrm{s.t.} &\qquad \bm{y} \geq \bm{0}\nonumber\\ &\qquad (\mathbf{X} - \bm{\mu}^\intercal) \bm{w} \geq -\bm{y}\nonumber\\ &\qquad (v,\, \bm{y}) \in \mathcal{K}_{\textrm{soc}} \end{align}\]

Where:

$\bm{w}$: is the N×1 vector of asset weights.

$\mathbf{X}$: is the T×N matrix of asset returns.

$\bm{\mu}$: is the N×1 vector of expected returns.

$\bm{w}$: is the N×1 vector of asset weights.

$\bm{y}$: is the T×1 vector of deviations from the expected portfolio return that meets the minimum return threshold.

Behaviour

Produces a QuadExpr risk expression svariance_risk = sdev^2/(T-1).

PortfolioOptimiser.RSOC — Type

struct RSOC <: VarianceFormulation end

Rotated Second-Order Cone (RSOC) formulation for the Variance and SVariance. Formulates the quadratic variance/semi variance expression using a RotatedSecondOrderCone constraint.

See also: VarianceFormulation, Quad, SOC, Variance, SVariance.

General Behaviour

Uses SecondOrderCone constraints.

Uses RotatedSecondOrderCone constraints.

Compatible with specialised conic solvers.

Performance may degrade for large portfolios.

Variance

\[\begin{align} \underset{\bm{w}}{\mathrm{opt}} &\qquad \dfrac{v}{T-1}\nonumber\\ \textrm{s.t.} &\qquad \bm{y} = (\mathbf{X} - \bm{\mu}^\intercal) \bm{w}\nonumber\\ &\qquad (v,\, 0.5,\, \bm{y}) \in \mathcal{K}_{\textrm{rsoc}} \end{align}\]

Where:

$\bm{w}$: is the N×1 vector of asset weights.

$\bm{v}$: is the portfolio variance.

$T$: is the number of observations.

$\bm{y}$: is the T×1 vector of deviations from the expected portfolio return.

$\bm{\mu}$: is the N×1 vector of expected asset returns.

$\mathcal{K}_{\textrm{rsoc}}$: is the RotatedSecondOrderCone. Which in this formulation represents the sum of squares of a vector.

Behaviour

Produces a QuadExpr risk expression svariance_risk = tvariance/(T-1).

Warning

Using w to compute the Variance risk of a portfolio optimised via an optimisation which uses JuMP, you have to ensure that the value of mu used by the optimisation is consistent with the value of w–-i.e. it was computed with MuSimple using w. Otherwise, the calculation will be inconsistent with the value of :svariance_risk. Alternatively, use the value of mu in both.

Examples

SVariance

\[\begin{align} \underset{\bm{w}}{\mathrm{opt}} &\qquad \dfrac{v}{T-1}\nonumber\\ \textrm{s.t.} &\qquad \bm{y} \geq \bm{0}\nonumber\\ &\qquad (\mathbf{X} - \bm{\mu}^\intercal) \bm{w} \geq -\bm{y}\nonumber\\ &\qquad (v,\, 0.5,\, \bm{y}) \in \mathcal{K}_{\textrm{rsoc}} \end{align}\]

Where:

$\bm{w}$: is the N×1 vector of asset weights.

$\mathbf{X}$: is the T×N matrix of asset returns.

$\bm{\mu}$: is the N×1 vector of expected returns.

$\bm{w}$: is the N×1 vector of asset weights.

$\bm{y}$: is the T×1 vector of deviations from the expected portfolio return that meets the minimum return threshold.

Behaviour

Produces a QuadExpr risk expression svariance_risk = tsvariance/(T-1).

RMMu

PortfolioOptimiser.calc_ret_mu — Function

calc_ret_mu(x::AbstractVector, w::AbstractVector, rm::RMMu)

Computes the mean portfolio return for rm.

If isnothing(rm.mu) || isempty(rm.mu), computes the mean return from x.

If isnothing(rm.w): computes the unweighted mean.

Else: computes the weighted mean.

Else: computes the mean return as dot(rm.mu, w).

See also: RMMu.

Positional Arguments

x: T×1 vector of portfolio returns.

w: N×1 vector of asset weights.

rm: RMMu risk measure.

Returns

mu::Real: portfolio mean return.

RMTarget

PortfolioOptimiser.calc_target_ret_mu — Function

calc_target_ret_mu(x::AbstractVector, w::AbstractVector, rm::RMTarget)

Computes the minimum acceptable portfolio return target for rm. Only returns equal to or below this value are accounted for in the calculation.

If isnothing(rm.target) || isa(rm.target, AbstractVector) && isempty(rm.target), computes the mean acceptable return from x via calc_ret_mu.

Else: returns rm.target.

See also: RMTarget.

Positional Arguments

x: T×1 vector of portfolio returns.

w: N×1 vector of asset weights.

rm: RMTarget risk measure.

Returns

target::Real: minimum acceptable return target.

RMSolvers

PortfolioOptimiser.ERM — Function

ERM(x::AbstractVector{<:Real}, z::Real = 1.0, α::Real = 0.05)

Computes the Entropic Risk Measure.

\[\begin{align} \mathrm{ERM}(\bm{X},\, z, \,\alpha) &= z \ln \left(\dfrac{M_{\bm{X}}\left(z^{-1}\right)}{\alpha} \right)\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

$z$: is the entropic moment.

$M_{\bm{X}}\left(t\right)$: is the moment generating function of $\bm{X}$.

$\alpha$: is the significance parameter, $\alpha \in (0,\,1)$.

Positional Arguments

x::AbstractVector: T×1 vector of portfolio returns.

z::Real = 1.0: entropic moment.

α::Real = 0.05: significance level, α ∈ (0, 1).

Outputs

er::Real: entropic risk.

Examples

ERM(x::AbstractVector{<:Real},
         solvers::Union{PortOptSolver, <:AbstractVector{PortOptSolver}},
         alpha::Real = 0.05)

Computes the Entropic Risk Measure.

\[\begin{align} \mathrm{ERM}(\bm{X}, \,\alpha) &= \begin{cases} \underset{z,\, t,\, u}{\inf} & t + z \ln\left(\dfrac{1}{\alpha T}\right)\\ \mathrm{s.t.} & z \geq \sum\limits_{i=1}^{T} u_{i}\nonumber\\ & (-X_{i}-t,\, z,\, u_{i}) \in \mathcal{K}_{\exp} \, \forall \, i=1,\,\dots{},\, T \end{cases}\\ &= \underset{z>0}{\inf}\left\{ z \ln \left(\dfrac{M_{\bm{X}}\left(z^{-1}\right)}{\alpha} \right)\right\}\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance parameter, $\alpha \in (0,\,1)$.

$z$: is the entropic moment.

$t$ and $\bm{u}$: are auxiliary variables for modelling the moment generating function.

$T$: is the number of observations.

$X_{i}$: is the i-th entry of the portfolio returns vector.

$M_{\bm{X}}\left(t\right)$: is the moment generating function of $\bm{X}$.

$\mathcal{K}_{\mathrm{exp}}$: is the ExponentialCone.

See also: EVaR, EDaR, EDaR_r, expected_risk.

Inputs

x::AbstractVector: T×1 vector of portfolio returns.

solvers::Union{PortOptSolver, <:AbstractVector{PortOptSolver}}: instance or AbstractVector of PortOptSolver which supports ExponentialCone constraints.

α::Real = 0.05: significance level, α ∈ (0, 1).

Behaviour

If no valid solution is found returns NaN.

Warning

α is not validated because this is an internal function. It should have been validated by EVaR, EDaR, or EDaR_r.

Outputs

er::Real: entropic risk.

Examples

PortfolioOptimiser.RRM — Function

RRM(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05, κ::Real = 0.3)

Computes the Relativistic Risk Measure.

\[\begin{align} \mathrm{RRM}(X,\,\alpha,\,\kappa) &= \left\{ \begin{align} &\underset{z,\, t,\, \psi,\, \theta,\, \varepsilon,\, \omega}{\text{inf}} && t + z \ln_{\kappa} \left(\dfrac{1}{\alpha T}\right) + \sum\limits_{i=1}^T \left(\psi_{i} + \theta_{i} \right) \nonumber\\ &\mathrm{s.t.} && -X - t + \varepsilon + \omega \leq 0 \nonumber\\ &&&z \geq 0 \\ &&&\left( z\left(\dfrac{1+\kappa}{2\kappa}\right),\, \psi_{i}\left(\dfrac{1+\kappa}{\kappa}\right),\, \varepsilon_{i} \right) \in \mathcal{P}_{3}^{1/(1+\kappa),\, \kappa/(1+\kappa)}\nonumber\\ &&&\left( \omega_{i}\left(\dfrac{1}{1-\kappa}\right),\, \theta_{i}\left(\dfrac{1}{\kappa}\right),\, -z \left(\dfrac{1}{2\kappa}\right) \right) \in \mathcal{P}_{3}^{1-\kappa,\, \kappa}\nonumber\\ &&&\forall \, i=1,\,\dots{},\, T \nonumber \end{align} \right.\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance parameter, $\alpha \in (0,\,1)$.

$\kappa$: is the relativistic deformation parameter, $\kappa \in (0,\,1)$.

$z$: is the relativistic moment.

$t$, $\psi$, $\theta$, $\varepsilon$, and $\omega$: are auxiliary variables for modelling the relativistic deformation of the moment generating function.

$T$: is the number of observations.

$\mathcal{P}_3^{c,\, 1-c}$: is the PowerCone parametrised by $c \in (0,\,1)$.

See also: RLVaR, RLDaR, RLDaR_r, expected_risk.

Inputs

x::AbstractVector: T×1 vector of portfolio returns.

solvers::Union{Nothing, PortOptSolver, <:AbstractVector{PortOptSolver}}: instance or AbstractVector of PortOptSolver which supports ExponentialCone constraints.

α::Real = 0.05: significance level, α ∈ (0, 1).

κ::Real = 0.3: relativistic deformation parameter, κ ∈ (0, 1).

Behaviour

If no valid solution is found returns NaN.

Warning

α and κ are not validated because this is an internal function. They should have been validated by RLVaR, RLDaR, or RLDaR_r.

Outputs

rlr::Real: relativistic risk.

Examples