RiskMeasure

This contains the documentation for all available RiskMeasure subtypes.

Dispersion

These measure the spread of the returns distribution.

PortfolioOptimiser.VarianceFormulation — Type

abstract type VarianceFormulation end

Abstract type for implementing various formulations of the Variance in optimisations which use JuMP models.

If either network_adj or cluster_adj property of the Portfolio instance is SDP, the formulation has no effect because this constraint type requires a PSDCone formulation of the variance.

See also: Variance, Quad, SOC, Portfolio, SDP.

PortfolioOptimiser.Quad — Type

struct Quad <: VarianceFormulation end

Explicit quadratic formulation for the Variance as an optimisation model.

\[\begin{align} \underset{\bm{w}}{\mathrm{opt}} &\qquad \bm{w}^\intercal \mathbf{\Sigma} \bm{w}\\ \end{align}\]

Where:

$\bm{w}$: is the N×1 vector of asset weights.

$\mathbf{\Sigma}$: is the N×N asset covariance matrix.

See also: VarianceFormulation, SOC, RSOC, Variance.

Behaviour

Produces a QuadExpr risk expression variance_risk = dot(w, sigma, w).

Not compatible with NOC optimisations because QuadExpr are not strictly convex.

No additional variables or constraints introduced.

Requires a solver capable of handling quadratic expressions.

Performance may degrade for large portfolios.

Examples

PortfolioOptimiser.SOC — Type

struct SOC <: VarianceFormulation end

Second-Order Cone (SOC) formulation for the Variance. Reformulates the quadratic variance expression using a SecondOrderCone cone constraint.

\[\begin{align} \underset{\bm{w}}{\mathrm{opt}} &\qquad \sigma^2\nonumber\\ \textrm{s.t.} &\qquad \left\lVert \mathbf{G} \bm{w} \right\rVert_{2} \leq \sigma\,. \end{align}\]

Where:

$\bm{w}$: is the N×1 vector of asset weights.

$\mathbf{G}$: is a suitable factorisation of the N×N covariance matrix, such as the square root matrix, or the Cholesky factorisation.

$\sigma^2$: is the portfolio variance.

$\lVert \cdot \rVert_{2}$: is the L-2 norm, which is modelled as an SecondOrderCone.

See also: VarianceFormulation, Quad, RSOC, Variance.

Behaviour

Uses SecondOrderCone constraints.

Defines a standard deviation variable dev.

Produces a QuadExpr risk expression variance_risk = dev^2.

Not compatible with NOC (Near Optimal Centering) optimisations because QuadExpr are not strictly convex.

Requires a solver capable of handling quadratic expressions.

Often more numerically stable than direct quadratic formulation.

Better scaling properties for large portfolios.

Compatible with specialised conic solvers.

May introduce more variables but often leads to better solution times.

Particularly effective for large-scale problems.

Examples

PortfolioOptimiser.RSOC — Type

struct RSOC <: VarianceFormulation end

Rotated Second-Order Cone (RSOC) formulation for the Variance. Reformulates the quadratic variance expression using a RotatedSecondOrderCone constraint.

\[\begin{align} \underset{\bm{w}}{\mathrm{opt}} &\qquad \dfrac{v}{T-1}\nonumber\\ \textrm{s.t.} &\qquad \bm{y} = (\mathbf{X} - \bm{\mu}^\intercal) \bm{w}\nonumber\\ &\qquad (v,\, 0.5,\, \bm{y}) \in \mathcal{K}_{\textrm{rsoc}} \end{align}\]

Where:

$\bm{w}$: is the N×1 vector of asset weights.

$\bm{v}$: is the portfolio variance.

$T$: is the number of observations.

$\bm{y}$: is the T×1 vector of deviations from the expected portfolio return.

$\bm{\mu}$: is the N×1 vector of expected asset returns.

$\mathcal{K}_{\textrm{rsoc}}$: is the RotatedSecondOrderCone. Which in this formulation represents the sum of squares of a vector.

See also: VarianceFormulation, Quad, SOC, Variance.

Behaviour

Uses SecondOrderCone constraints.

Uses RotatedSecondOrderCone constraints.

Defines a standard deviation variable dev.

Produces a QuadExpr risk expression variance_risk = tvariance.

Not compatible with NOC (Near Optimal Centering) optimisations because QuadExpr are not strictly convex.

Requires a solver capable of handling quadratic expressions.

Performance may degrade for large portfolios.

PortfolioOptimiser.Variance — Type

mutable struct Variance{T1 <: Union{<:AbstractMatrix, Nothing}} <: RiskMeasureSigma

Measures and computes the portfolio Variance (Variance).

\[\begin{align} \mathrm{Variance}(\bm{w},\, \mathbf{\Sigma}) &= \bm{w}^\intercal \, \mathbf{\Sigma}\, \bm{w}\,. \end{align}\]

Where:

$\bm{w}$: is the N×1 vector of asset weights.

$\mathbf{\Sigma}$: is the N×N asset covariance matrix.

See also: RiskMeasureSigma, RMSettings, SD, PortClass, OptimType, NOC, NoAdj, IP, SDP, Portfolio, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

sigma::Union{<:AbstractMatrix, Nothing} = nothing: (optional) N×N covariance matrix.

If nothing: takes its value from the instance Portfolio, the specific property depends on the PortClass parameter of the OptimType used.

Behaviour in optimisations which take risk measures and use JuMP models

The Variance risk is defined as the key :variance_risk.

NoAdj, IP network and cluster constraints.

Requires a solver that supports SecondOrderCone constraints.

Defines the variance risk, :variance_risk, as a QuadExpr.

Incompatible with NOC (Near Optimal Centering) optimisations because QuadExpr are strictly not convex.

If it exists, the upper bound is defined via the portfolio standard deviation with key, :dev_ub.

SDP network and/or cluster constraints.

Requires a solver that supports PSDCone constraints.

Defines the variance risk, :variance_risk, as an AffExpr.

Compatible with NOC (Near Optimal Centering) optimisations because AffExpr are strictly convex.

If it exists, the upper bound is defined via the portfolio variance with key, :variance_risk_ub.

Functor

(variance::Variance)(w::AbstractVector): computes the Variance of an N×1 vector of asset weights.

Examples

PortfolioOptimiser.SD — Type

mutable struct SD <: RiskMeasureSigma

Measures and computes the portfolio Standard Deviation (SD).

\[\begin{align} \mathrm{SD}(\bm{w},\, \mathbf{\Sigma}) &= \left(\bm{w}^\intercal \, \mathbf{\Sigma} \, \bm{w}\right)^{1/2}\,. \end{align}\]

Where:

$\bm{w}$: is the N×1 vector of asset weights.

$\mathbf{\Sigma}$: is the N×N asset covariance matrix.

See also: RiskMeasureSigma, RMSettings, Variance, Portfolio, PortClass, OptimType, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

sigma::Union{<:AbstractMatrix, Nothing} = nothing: (optional) N×N covariance matrix.

If nothing: takes its value from the instance Portfolio, the specific property depends on the PortClass parameter of the OptimType used.

Behaviour in optimisations which take risk measures and use JuMP models

The SD risk is defined as a VariableRef with the key, :sd_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :sd_risk_ub.

Requires a solver that supports SecondOrderCone constraints.

Functor

(sd::SD)(w::AbstractVector): computes the SD risk of an N×1 vector of asset weights.

Examples

PortfolioOptimiser.MAD — Type

mutable struct MAD <: RiskMeasureMu

Measures and computes the portfolio Mean Absolute Deviation (MAD). In other words, it is the expected value of the absolute deviation from the expected value of the returns vector. This is a generalisation to accomodate the use of weighted means.

\[\begin{align} \mathrm{MAD}(\bm{X}) &= \mathbb{E}\left(\left\lvert \bm{X} - \mathbb{E}\left(\bm{X}\right) \right\rvert\right)\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 portfolio returns vector.

$\lvert \cdot \rvert$: is the absolute value.

$\mathbb{E}(\cdot)$: is the expected value.

See also: RiskMeasureMu, RMSettings, Portfolio, PortClass, OptimType, calc_risk, optimise!, set_rm, calc_ret_mu.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

w::Union{<:AbstractWeights, Nothing} = nothing: (optional, functor-exclusive) T×1 vector of weights for computing the expected value of the returns vector (internal expected value) via calc_ret_mu.

we::Union{<:AbstractWeights, Nothing} = nothing: (optional) T×1 vector of weights for computing the expected value of the absolute deviation (external expected value).

If isnothing(we): computes the unweighted mean.

Else: computes the weighted mean.

mu::Union{<:AbstractVector{<:Real}, Nothing} = nothing: (optional) N×1 vector of weights for computing the expected value of the returns vector (internal expected value) via calc_ret_mu or set_rm.

Warning

Using w to compute the MAD risk of a portfolio optimised via an optimisation which uses JuMP, you have to ensure that the value of mu used by the optimisation is consistent with the value of w–-i.e. it was computed with MuSimple using w. Otherwise, the calculation will be inconsistent with the value of :mad_risk. Alternatively, use the value of mu in both.

Behaviour in optimisations which take risk measures and use JuMP models

The MAD risk is defined as an AffExpr with the key, :mad_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :mad_risk_ub.

Functor

(mad::MAD)(X::AbstractMatrix, w::AbstractVector, fees = 0.0): computes the MAD risk of a T×N returns matrix, a N×1 vector of asset weights w, and fees fees.

fees: must be consistent with the returns frequency.

Examples

PortfolioOptimiser.Kurt — Type

mutable struct Kurt <: RiskMeasureMu

Measures and computes the portfolio Square Root Kurtosis (Kurt).

\[\begin{align} \mathrm{Kurt}(\bm{X}) &= \left(\dfrac{1}{T} \sum\limits_{t=1}^{T} \left(X_{t} - \mathbb{E}(\bm{X}) \right)^{4} \right)^{1/2}\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

$T$: is the number of observations.

$X_{t}$: is the t-th value of the portfolio returns vector.

$\mathbb{E}(\cdot)$: is the expected value.

See also: RiskMeasureMu, RMSettings, SKurt, Kurtosis, SKurtosis, calc_risk, Portfolio, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

w::Union{<:AbstractWeights, Nothing} = nothing: (optional, functor-exclusive) T×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu.

mu::Union{<:AbstractVector{<:Real}, Nothing} = nothing: (optional, functor-exclusive) N×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu.

kt::Union{<:AbstractMatrix, Nothing} = nothing: (optional, set_rm exclusive) N²×N² cokurtosis matrix.

If nothing: takes its value from the kurt instance Portfolio.

kt has no effect in optimisations that don't use JuMP models, because the kurtosis is computed from the returns vector.

Warning

Using w to compute the Kurt risk of a portfolio optimised via an optimisation which uses JuMP, you have to ensure that the value of kurt in the instance of Portfolio used by the optimisation is consistent with the value of w–-i.e. it was computed with MuSimple using w. Otherwise, the calculation will be inconsistent with the value of :kurt_risk.

Behaviour in optimisations which take risk measures and use JuMP models

Uses PSDCone constraints.

The Kurt risk is defined as an AffExpr with the key, :kurt_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :kurt_risk_ub.

Functor

(kurt::Kurt)(x::AbstractVector): computes the Square Root Kurtosis of a T×1 vector of portfolio returns x.

Examples

PortfolioOptimiser.RG — Type

mutable struct RG{T1 <: Real} <: RiskMeasure

Measures and computes the portfolio Range (RG). It is the difference between the best and worst return.

\[\begin{align} \mathrm{RG}(\bm{X}) &= \max(\bm{X}) - \min(\bm{X})\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

See also: RiskMeasure, RMSettings, Portfolio, WR, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

Behaviour in optimisations which take risk measures and use JuMP models

The RG risk is defined as an AffExpr with the key, :rg_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :rg_risk_ub.

Functor

(rg::RG)(x::AbstractVector): computes the RG risk of a T×1 vector of portfolio returns x.

Examples

PortfolioOptimiser.CVaRRG — Type

mutable struct CVaRRG{T1 <: Real, T2 <: Real} <: RiskMeasure

Measures and computes the portfolio Conditional Value at Risk Range (CVaRRG). It measures the difference between the weighted average of all head gains up to the Value at Risk for gains, which is the threshold above or equal to which are the best beta % of portfolio returns. And the weighted average of all tail losses up to the Value at Risk, which is the threshold below or equal to which are the worst alpha % of portfolio returns.

\[\begin{align} \mathrm{CVaRRG}(\bm{X},\, \alpha,\,\beta) &= \mathrm{CVaR}(\bm{X},\, \alpha) - \mathrm{CVaR}(-\bm{X},\, \beta) \end{align}\]

Where:

$\mathrm{CVaR}(\bm{X},\, \alpha)$: is the Conditional Value at Risk as defined in CVaR.

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance level of losses.

$\beta$: is the significance level of gains.

See also: See also: RiskMeasure, RMSettings, Portfolio, CVaR, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

alpha::T1 = 0.05: significance level of losses, alpha ∈ (0, 1).

beta::T1 = 0.05: significance level of gains, beta ∈ (0, 1).

Behaviour in optimisations which take risk measures and use JuMP models

The CVaRRG is defined as an AffExpr with the key, :cvarrg_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :cvarrg_risk_ub.

Functor

(cvarrg::CVaRRG)(x::AbstractVector): computes the CVaRRG of a T×1 vector of portfolio returns x.

Examples

PortfolioOptimiser.TGRG — Type

mutable struct TGRG{T1 <: Real, T2 <: Real, T3 <: Integer, 
                    T4 <: Real, T5 <: Real, T6 <: Integer} <: RiskMeasureOWA

Measures and computes the portfolio Tail Gini Range (TGRG). It is the difference between the Head Gini and Tail Gini.

\[\begin{align} \mathrm{TGRG}(\bm{X},\, \alpha,\, \beta) &= \mathrm{TG}(\bm{X},\,\alpha) - \mathrm{TG}(-\bm{X},\,\beta)\,. \end{align}\]

Where:

$\mathrm{TG}$: is the Tail Gini defined in TG.

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance level of losses.

$\beta$: is the significance level of gains.

See also: See also: RiskMeasureOWA, RMSettings, Portfolio, owa_tgrg, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

formulation::OWAFormulation = OWAApprox(): OWA risk measure settings.

alpha_i::T1 = 0.0001: start value of the significance level of CVaR losses, 0 < alpha_i < alpha < 1.

alpha::T2 = 0.05: end value of the significance level of CVaR losses, 0 < alpha_i < alpha < 1.

a_sim::T3 = 100: number of CVaRs to approximate the Tail Gini losses, a_sim > 0.

beta_i::T4 = 0.0001: start value of the significance level of CVaR gains, 0 < beta_i < beta < 1.

beta::T5 = 0.05: end value of the significance level of CVaR gains, 0 < beta_i < beta < 1.

b_sim::T6 = 100: number of CVaRs to approximate the Tail Gini gains, b_sim > 0.

Behaviour in optimisations which take risk measures and use JuMP models

The TGRG risk is defined as an AffExpr with the key, :tgrg_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :tgrg_risk_ub.

Examples

PortfolioOptimiser.GMD — Type

mutable struct GMD <: RiskMeasureOWA

Measures and computes the portfolio Gini Mean Difference (GMD).

See also: See also: RiskMeasureOWA, RMSettings, Portfolio, owa_gmd, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

formulation::OWAFormulation = OWAApprox(): OWA risk measure settings.

Behaviour in optimisations which take risk measures and use JuMP models

The GMD risk is defined as an AffExpr with the key, :gmd_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :gmd_risk_ub.

Examples

PortfolioOptimiser.Skew — Type

mutable struct Skew <: RiskMeasureSkew

Measures and computes the portfolio Quadratic Skewness (Skew).

\[\begin{align} \nu &= \left(\bm{w}^{\intercal} \mathbf{V} \bm{w}\right)^{1/2}\\ \end{align}\]

Where:

$\bm{w}$: is the vector of asset weights.

$\mathbf{V}$: is the sum of the symmetric negative spectral slices of the coskewness.

See also: See also: RiskMeasureSkew, RMSettings, Portfolio, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

skew::Union{<:AbstractMatrix, Nothing}: (optional) N×N² custom coskewness matrix.

If isnothing(skew): takes its value from the skew property of the instance of Portfolio.

V::Union{Nothing, <:AbstractMatrix}: (optional) Na×Na custom sum of the symmetric negative spectral slices of the coskewness.

If isnothing(V): takes its value from the V property of the instance of Portfolio.

Behaviour in optimisations which take risk measures and use JuMP models

The Skew risk is defined as an AffExpr with the key, :skew_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :skew_risk_ub.

Examples

PortfolioOptimiser.BDVariance — Type

struct BDVariance <: RiskMeasure

Measures and computes the portfolio Brownian Distance Variance (BDVariance). It captures linear and non-linear codependence between assets.

\[\begin{align} \mathrm{BDVariance}(\bm{X}) &= \mathrm{BDCov}(\bm{X},\, \bm{X}) = \dfrac{1}{T^{2}} \sum\limits_{i=1}^{T}\sum\limits_{j=1}^{T} A_{i,\,j}^2\\ \mathrm{BDCov}(\bm{X},\, \bm{Y}) &= \dfrac{1}{T^{2}} \sum\limits_{i=1}^{T} \sum\limits_{j=1}^{T} A_{i,\,j} B_{i,\,j}\\ A_{i,\,j} &= a_{i,\,j} - \bar{a}_{i\,.} - \bar{a}_{.\,j} + \bar{a}_{.\,.}\\ B_{i,\,j} &= b_{i,\,j} - \bar{b}_{i\,.} - \bar{b}_{.\,j} + \bar{b}_{.\,.}\\ a_{i,\,j} &= \lVert X_{i} - X_{j} \rVert_{2}, \quad \forall i,\, j = 1,\, \ldots ,\, T\\ b_{i,\,j} &= \lVert Y_{i} - Y_{j} \rVert_{2}, \quad \forall i,\, j = 1,\, \ldots ,\, T\,. \end{align}\]

Where:

$\bm{X}$ and $\bm{Y}$ : are random variables, they are equal in this case as they are the portfolio returns.

$a_{i,\,j}$ and $b_{i,\,j}$ : are entries of a distance matrix where $i$ and $j$ are points in time. Each entry is defined as the Euclidean distance $\lVert \cdot \rVert_{2}$ between the value of the random variable at time $i$ and its value at time $j$.

$\bar{a}_{i\,\cdot}$ and $\bar{b}_{i\,\cdot}$ : are the $i$-th row means of their respective matrices.

$\bar{a}_{\cdot\,j}$ and $\bar{b}_{\cdot\,j}$ : are the $j$-th column means of their respective matrices.

$\bar{a}_{\cdot\,\cdot}$ and $\bar{b}_{\cdot\,\cdot}$ : are the grand means of their respective matrices.

$A_{i,\,j}$ and $B_{i,\,j}$ : are doubly centered distances.

See also: See also: RiskMeasure, RMSettings, Portfolio, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

Behaviour in optimisations which take risk measures and use JuMP models

The BDVariance risk is defined as a QuadExpr with the key, :bdvariance_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :bdvariance_risk_ub.

Examples

PortfolioOptimiser.WCVariance — Type

mutable struct WCVariance{T1} <: RiskMeasureSigma

Downside dispersion

These measure how far the returns deviate from the mean in the negative direction.

PortfolioOptimiser.SSD — Type

mutable struct SSD <: RiskMeasureMu

Measures and computes the portfolio Semi Standard Deviation (SD) equal to or below the mean return.

\[\begin{align} \mathrm{SSD}(\bm{X}) &= \left(\dfrac{1}{T-1} \sum\limits_{t=1}^{T}\min\left(X_{t} - \mathbb{E}(\bm{X}),\, 0\right)^{2}\right)^{1/2}\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

$T$: is the number of observations.

$X_{t}$: is the t-th value of the portfolio returns vector.

$\mathbb{E}(\cdot)$: is the expected value.

See also: RiskMeasureMu, RMSettings, Portfolio, SD, Variance, SVariance, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

w::Union{<:AbstractWeights, Nothing} = nothing: (optional, functor-exclusive) T×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu.

mu::Union{<:AbstractVector{<:Real}, Nothing} = nothing: (optional) N×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu or set_rm.

Warning

Using w to compute the SSD risk of a portfolio optimised via an optimisation which uses JuMP, you have to ensure that the value of mu used by the optimisation is consistent with the value of w–-i.e. it was computed with MuSimple using w. Otherwise, the calculation will be inconsistent with the value of :ssd_risk. Alternatively, use the value of mu in both.

Behaviour in optimisations which take risk measures and use JuMP models

The SSD risk is defined as an AffExpr with the key, :ssd_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :ssd_risk_ub.

Functor

(ssd::SSD)(X::AbstractMatrix, w::AbstractVector, fees = 0.0): computes the SSD risk of a T×N returns matrix, a N×1 vector of asset weights w, and fees fees.

fees: must be consistent with the returns frequency.

Examples

PortfolioOptimiser.FLPM — Type

mutable struct FLPM{T1 <: Real} <: RiskMeasureTarget

Measures and computes the portfolio First Lower Partial Moment (FLPM). Measures the dispersion equal to or below the target return threshold. The risk-adjusted return ratio of this risk measure is commonly known as the Omega ratio.

\[\begin{align} \mathrm{FLPM}(\bm{X},\, r) &= \dfrac{1}{T} \sum\limits_{t=1}^{T}\max\left(r - X_{t},\, 0\right)\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

$r$: is the minimum acceptable return.

$T$: is the number of observations.

$X_{t}$: is the t-th value of the portfolio returns vector.

See also: RiskMeasureTarget, RMSettings, Portfolio, SLPM, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

target::T1 = 0.0: minimum return threshold for classifying downside returns. Only returns equal to or below this value are considered in the calculation. Its value can be computed via calc_target_ret_mu or calc_rm_target.

w::Union{<:AbstractWeights, Nothing} = nothing: (optional, functor-exclusive) T×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu.

mu::Union{<:AbstractVector{<:Real}, Nothing} = nothing: (optional) N×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu or set_rm.

Warning

Using w to compute the FLPM risk of a portfolio optimised via an optimisation which uses JuMP, you have to ensure that the value of mu used by the optimisation is consistent with the value of w–-i.e. it was computed with MuSimple using w. Otherwise, the calculation will be inconsistent with the value of :flpm_risk. Alternatively, use the value of mu in both.

Behaviour in optimisations which take risk measures and use JuMP models

The FLPM risk is defined as an AffExpr with the key, :flpm_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :flpm_risk_ub.

Functor

(flpm::FLPM)(X::AbstractMatrix, w::AbstractVector, fees = 0.0): computes the FLPM risk of a T×N returns matrix, a N×1 vector of asset weights w, and fees fees.

fees: must be consistent with the returns frequency.

Examples

PortfolioOptimiser.SLPM — Type

mutable struct SLPM{T1 <: Real} <: RiskMeasureTarget

Measures and computes the portfolio Second Lower Partial Moment (SLPM). Measures the dispersion equal to or below the target return threshold. The risk-adjusted return ratio of this risk measure is commonly known as the Sortino ratio.

\[\begin{align} \mathrm{SLPM}(\bm{X},\, r) &= \left(\dfrac{1}{T-1} \sum\limits_{t=1}^{T}\max\left(r - X_{t},\, 0\right)^{2}\right)^{1/2}\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

$r$: is the minimum acceptable return.

$T$: is the number of observations.

$X_{t}$: is the t-th value of the portfolio returns vector.

See also: RiskMeasureTarget, RMSettings, Portfolio, FLPM, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

target::T1 = 0.0: minimum return threshold for classifying downside returns. Only returns equal to or below this value are considered in the calculation. Its value can be computed via calc_target_ret_mu or calc_rm_target.

w::Union{<:AbstractWeights, Nothing} = nothing: (optional, functor-exclusive) T×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu.

mu::Union{<:AbstractVector{<:Real}, Nothing} = nothing: (optional) N×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu or set_rm.

Warning

Using w to compute the SLPM risk of a portfolio optimised via an optimisation which uses JuMP, you have to ensure that the value of mu used by the optimisation is consistent with the value of w–-i.e. it was computed with MuSimple using w. Otherwise, the calculation will be inconsistent with the value of :slpm_risk. Alternatively, use the value of mu in both.

Behaviour in optimisations which take risk measures and use JuMP models

The SLPM risk is defined as an AffExpr with the key, :slpm_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :slpm_risk_ub.

Functor

(slpm::SLPM)(X::AbstractMatrix, w::AbstractVector, fees = 0.0): computes the SLPM risk of a T×N returns matrix, a N×1 vector of asset weights w, and fees fees.

fees: must be consistent with the returns frequency.

Examples

PortfolioOptimiser.SKurt — Type

mutable struct SKurt{T1 <: Real} <: RiskMeasureMu

Measures and computes the portfolio Square Root Semi Kurtosis (SKurt). Measures the kurtosis equal to or below the mean portfolio returns.

\[\begin{align} \mathrm{SKurt}(\bm{X}) &= \left(\dfrac{1}{T} \sum\limits_{t=1}^{T} \min\left(X_{t} - \mathbb{E}(\bm{X}),\, 0 \right)^{4} \right)^{1/2}\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

$T$: is the number of observations.

$X_{t}$: is the t-th value of the portfolio returns vector.

$\mathbb{E}(\cdot)$: is the expected value.

See also: RiskMeasureMu, RMSettings, Kurt, Kurtosis, SKurtosis, Portfolio, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

w::Union{<:AbstractWeights, Nothing} = nothing: (optional, functor-exclusive) T×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu.

mu::Union{<:AbstractVector{<:Real}, Nothing} = nothing: (optional, functor-exclusive) N×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu.

kt::Union{<:AbstractMatrix, Nothing} = nothing: (optional, set_rm exclusive) N²×N² cokurtosis matrix.

If nothing: takes its value from the skurt instance Portfolio.

kt has no effect in optimisations that don't use JuMP models, because the kurtosis is computed from the returns vector.

Warning

Using w to compute the Kurt risk of a portfolio optimised via an optimisation which uses JuMP, you have to ensure that the value of skurt in the instance of Portfolio used by the optimisation is consistent with the value of w–-i.e. it was computed with MuSimple using w. Otherwise, the calculation will be inconsistent with the value of :skurt_risk.

Behaviour in optimisations which take risk measures and use JuMP models

Uses PSDCone constraints.

The Kurt risk is defined as an AffExpr with the key, :skurt_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :skurt_risk_ub.

Functor

(skurt::SKurt)(x::AbstractVector): computes the Semi Square Root Kurtosis of a T×1 vector of portfolio returns x.

Examples

PortfolioOptimiser.SSkew — Type

mutable struct SSkew <: RiskMeasureSkew

Measures and computes the portfolio Quadratic Semi Skewness (SSkew).

\[\begin{align} \nu &= \left(\bm{w}^{\intercal} \mathbf{V} \bm{w}\right)^{1/2}\\ \end{align}\]

Where:

$\bm{w}$: is the vector of asset weights.

$\mathbf{V}$: is the sum of the symmetric negative spectral slices of the semi coskewness.

See also: See also: RiskMeasureSkew, RMSettings, Portfolio, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

skew::Union{<:AbstractMatrix, Nothing}: (optional) N×N² custom semi coskewness matrix.

If isnothing(skew): takes its value from the sskew property of the instance of Portfolio.

V::Union{Nothing, <:AbstractMatrix}: (optional) Na×Na custom sum of the symmetric negative spectral slices of the semi coskewness.

If isnothing(V): takes its value from the SV property of the instance of Portfolio.

Behaviour in optimisations which take risk measures and use JuMP models

The SSkew risk is defined as an AffExpr with the key, :sskew_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :sskew_risk_ub.

Examples

Downside

These measure different aspects of the tail (negative side) of the returns distribution.

PortfolioOptimiser.WR — Type

struct WR <: RiskMeasure

Measures and computes the portfolio Worst Realization/Return (WR). It is the absolute value of the most extreme loss for the period. Best used in combination with other risk measures.

$\mathrm{VaR}(\bm{X},\, \alpha) \leq \mathrm{CVaR}(\bm{X},\, \alpha) \leq \mathrm{EVaR}(\bm{X},\, \alpha) \leq \mathrm{RLVaR}(\bm{X},\, \alpha,\, \kappa) \leq \mathrm{WR}(\bm{X})$.

\[\begin{align} \mathrm{WR}(\bm{X}) &= -\min(\bm{X})\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

See also: RiskMeasure, RMSettings, Portfolio, VaR, CVaR, EVaR, RLVaR, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

Behaviour in optimisations which take risk measures and use JuMP models

The WR risk is defined as an AffExpr with the key, :wr_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :wr_risk_ub.

Functor

(wr::WR)(x::AbstractVector): computes the WR risk of a T×1 vector of portfolio returns x.

Examples

PortfolioOptimiser.CVaR — Type

mutable struct CVaR{T1 <: Real} <: RiskMeasure

Measures and computes the portfolio Conditional Value at Risk (CVaR). Also known as the Expected Shortfall, it is the weighted average of all tail losses up to the Value at Risk, which is the threshold below or equal to which are the worst alpha % of portfolio returns.

$\mathrm{VaR}(\bm{X},\, \alpha) \leq \mathrm{CVaR}(\bm{X},\, \alpha) \leq \mathrm{EVaR}(\bm{X},\, \alpha) \leq \mathrm{RLVaR}(\bm{X},\, \alpha,\, \kappa) \leq \mathrm{WR}(\bm{X})$.

\[\begin{align} \mathrm{CVaR}(\bm{X},\, \alpha) &= \mathrm{VaR}(\bm{X},\, \alpha) + \dfrac{1}{\alpha T} \sum\limits_{t=1}^{T} \max\left(-X_{t} - \mathrm{VaR}(\bm{X},\, \alpha),\, 0\right)\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance level.

$\mathrm{VaR}(\bm{X},\, \alpha)$: is the Value at Risk as defined in VaR.

$T$: is the number of observations.

$X_{t}$: is the t-th value of the portfolio returns vector.

See also: RiskMeasure, RMSettings, Portfolio, VaR, WR, EVaR, RLVaR, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

alpha::T1 = 0.05: significance level, alpha ∈ (0, 1).

Behaviour in optimisations which take risk measures and use JuMP models

The CVaR is defined as an AffExpr with the key, :cvar_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :cvar_risk_ub.

Functor

(cvar::CVaR)(x::AbstractVector): computes the CVaR of a T×1 vector of portfolio returns x.

Examples

PortfolioOptimiser.DRCVaR — Type

PortfolioOptimiser.EVaR — Type

mutable struct EVaR{T1 <: Real} <: RiskMeasureSolvers

Measures and computes the portfolio Entropic Value at Risk (EVaR). It is the upper bound of the Chernoff inequality for the VaR and CVaR.

Uses ExponentialCone constraints.

$\mathrm{VaR}(\bm{X},\, \alpha) \leq \mathrm{CVaR}(\bm{X},\, \alpha) \leq \mathrm{EVaR}(\bm{X},\, \alpha) \leq \mathrm{RLVaR}(\bm{X},\, \alpha,\, \kappa) \leq \mathrm{WR}(\bm{X})$.

\[\begin{align} \mathrm{EVaR}(\bm{X},\alpha) &= \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\bm{X},\, z, \,\alpha)\right\}\,. \end{align}\]

Where:

$\mathrm{ERM}(\bm{X},\, z, \,\alpha)$ is the entropic risk measure as defined in ERM.

$z$: is the entropic value at risk.

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance level.

See also: RiskMeasureSolvers, RMSettings, Portfolio, VaR, WR, CVaR, RLVaR, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

alpha::T1 = 0.05: significance level, alpha ∈ (0, 1).

solvers::Union{Nothing, PortOptSolver, <:AbstractVector{PortOptSolver}} = nothing: optional abstract dictionary containing the solvers, their settings, solution criteria, and other arguments.

If isnothing(solvers): it takes its value from the solvers property of the instance of Portfolio.

Behaviour in optimisations which take risk measures and use JuMP models

The EVaR is defined as an AffExpr with the key, :evar_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :evar_risk_ub.

Functor

(evar::EVaR)(x::AbstractVector): computes the EVaR of a T×1 vector of portfolio returns x.

Examples

PortfolioOptimiser.RLVaR — Type

mutable struct RLVaR{T1 <: Real, T2 <: Real} <: RiskMeasureSolvers

Measures and computes the portfolio Relativistic Value at Risk (RLVaR). It is a generalisation of the EVaR.

Uses PowerCone constraints.

$\mathrm{VaR}(\bm{X},\, \alpha) \leq \mathrm{CVaR}(\bm{X},\, \alpha) \leq \mathrm{EVaR}(\bm{X},\, \alpha) \leq \mathrm{RLVaR}(\bm{X},\, \alpha,\, \kappa) \leq \mathrm{WR}(\bm{X})$.

$\lim\limits_{\kappa \to 0} \mathrm{RLVaR}(\bm{X},\, \alpha,\, \kappa) \approx \mathrm{EVaR}(\bm{X},\, \alpha)$

$\lim\limits_{\kappa \to 1} \mathrm{RLVaR}(\bm{X},\, \alpha,\, \kappa) \approx \mathrm{WR}(\bm{X})$

\[\begin{align} \mathrm{RLVaR}(\bm{X},\, \alpha,\, \kappa) &= \mathrm{RRM}(\bm{X},\, \alpha,\, \kappa)\,. \end{align}\]

Where:

$\mathrm{RRM}(\bm{X},\, \alpha,\, \kappa)$ is the Relativistic Risk Measure as defined in RRM.

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance level.

$\kappa$: is the relativistic deformation parameter.

See also: RiskMeasureSolvers, RMSettings, Portfolio, VaR, WR, CVaR, EVaR, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

alpha::T1 = 0.05: significance level, alpha ∈ (0, 1).

kappa::T1 = 0.3: relativistic deformation level, kappa ∈ (0, 1).

solvers::Union{Nothing, PortOptSolver, <:AbstractVector{PortOptSolver}} = nothing: optional abstract dictionary containing the solvers, their settings, solution criteria, and other arguments. In order to solve the problem, a solver must be compatible with ExponentialCone.

If isnothing(solvers): it takes its value from the solvers property of the instance of Portfolio.

Behaviour in optimisations which take risk measures and use JuMP models

The RLVaR is defined as an AffExpr with the key, :rlvar_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :rlvar_risk_ub.

Functor

(rlvar::RLVaR)(x::AbstractVector): computes the RLVaR of a T×1 vector of portfolio returns x.

Examples

PortfolioOptimiser.TG — Type

mutable struct TG{T1 <: Real, T2 <: Real, T3 <: Integer} <: RiskMeasureOWA

Measures and computes the portfolio Tail Gini (TG).

See also: See also: RiskMeasureOWA, RMSettings, Portfolio, owa_tg, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

formulation::OWAFormulation = OWAApprox(): OWA risk measure settings.

alpha_i::T1 = 0.0001: start value of the significance level of CVaR losses, 0 < alpha_i < alpha < 1.

alpha::T2 = 0.05: end value of the significance level of CVaR losses, 0 < alpha_i < alpha < 1.

a_sim::T3 = 100: number of CVaRs to approximate the Tail Gini losses, a_sim > 0.

Behaviour in optimisations which take risk measures and use JuMP models

The TG risk is defined as an AffExpr with the key, :tg_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :tg_risk_ub.

Examples

Drawdown

These measure the drops in portfolio value from local maxima to subsequent local minima.

PortfolioOptimiser.MDD — Type

struct MDD <: RiskMeasure

Measures and computes the portfolio Maximum Drawdown of uncompounded cumulative returns (MDD). It measures the largest peak-to-trough decline. Best used in combination with other risk measures. The risk-adjusted return ratio of this risk measure is commonly known as the Calmar ratio.

$\mathrm{DaR}(\bm{X},\, \alpha) \leq \mathrm{CDaR}(\bm{X},\, \alpha) \leq \mathrm{EDaR}(\bm{X},\, \alpha) \leq \mathrm{RLDaR}(\bm{X},\, \alpha,\, \kappa) \leq \mathrm{MDD}(\bm{X})$.

\[\begin{align} \mathrm{MDD_{a}}(\bm{X}) &= \max\mathrm{DD_{a}}(\bm{X})\,. \end{align}\]

Where:

$\mathrm{DD_{a}}(\bm{X})$ is the Drawdown of uncompounded cumulative returns as defined in DaR.

$\bm{X}$: is the T×1 vector of portfolio returns.

See also: RiskMeasure, RMSettings, Portfolio, DaR, CDaR, EDaR, RLDaR, DaR_r, CDaR_r, EDaR_r, RLDaR_r, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

Behaviour in optimisations which take risk measures and use JuMP models

The MDD is defined as an AffExpr with the key, :mdd_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :mdd_risk_ub.

Functor

(mdd::MDD)(x::AbstractVector): computes the MDD of a T×1 vector of portfolio returns x.

Examples

PortfolioOptimiser.ADD — Type

struct ADD <: RiskMeasure

Measures and computes the portfolio Average (Expected) Drawdown of uncompounded cumulative returns (ADD). This is a generalisation to accomodate the use of weighted means.

\[\begin{align} \mathrm{ADD_{a}}(\bm{X}) &= \mathbb{E}\lleft(\mathrm{DD_{a}}(\bm{X}, j)\,\forall j=1\,\ldots\,T\right)\,. \end{align}\]

Where:

$\mathrm{DD_{a}}(\bm{X}, j)$ is the Drawdown of uncompounded cumulative returns at time $j$ as defined in DaR.

$T$: is the number of observations.

$\bm{X}$: is the T×1 vector of portfolio returns.

See also: RiskMeasure, RMSettings, Portfolio, DaR, calc_risk, optimise!, set_rm.

Keyword Arguments

w::Union{<:AbstractWeights, Nothing} = nothing: optional T×1 vector of weights for computing the expected value of the returns vector.

Behaviour in optimisations which take risk measures and use JuMP models

The ADD is defined as an AffExpr with the key, :add_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :add_risk_ub.

Functor

(add::ADD)(x::AbstractVector): computes the ADD of a T×1 vector of portfolio returns x.

Examples

PortfolioOptimiser.UCI — Type

mutable struct UCI <: RiskMeasure

Measures and computes the portfolio Ulcer Index of uncompounded cumulative returns (UCI). It is the normalised L2-norm of the portfolio drawdowns of uncompounded cumulative returns.

\[\begin{align} \mathrm{UCI_{a}}(\bm{X}) &= \left(\dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{a}}(\bm{X},\, j)^{2}\right)^{1/2}\,. \end{align}\]

Where:

$\mathrm{DD_{a}}(\bm{X},\, j)$ is the Drawdown of uncompounded cumulative returns at time $j$ as defined in DaR.

$\bm{X}$: is the T×1 vector of portfolio returns.

See also: RiskMeasure, RMSettings, Portfolio, DaR, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

Behaviour in optimisations which take risk measures and use JuMP models

The UCI is defined as an AffExpr with the key, :uci_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :uci_risk_ub.

Functor

(uci::UCI)(x::AbstractVector): computes the UCI of a T×1 vector of portfolio returns x.

Examples

PortfolioOptimiser.CDaR — Type

mutable struct CDaR{T1 <: Real} <: RiskMeasure

Measures and computes the portfolio Conditional Drawdown at Risk of uncompounded cumulative returns (CDaR). It is the weighted average of all drawdowns up to the Drawdown at Risk, which is the threshold below or equal to which are the worst alpha % of portfolio drawdowns.

$\mathrm{DaR}(\bm{X},\, \alpha) \leq \mathrm{CDaR}(\bm{X},\, \alpha) \leq \mathrm{EDaR}(\bm{X},\, \alpha) \leq \mathrm{RLDaR}(\bm{X},\, \alpha,\, \kappa) \leq \mathrm{MDD}(\bm{X})$.

\[\begin{align} \mathrm{CDaR_{a}}(\bm{X},\, \alpha) &= \mathrm{DaR_{a}}(\bm{X},\, \alpha) + \dfrac{1}{\alpha T} \sum\limits_{j=0}^{T} \max\left(\mathrm{DD_{a}}(\bm{X},\, j) - \mathrm{DaR_{a}}(\bm{X},\, \alpha),\, 0 \right)\,. \end{align}\]

Where:

$\mathrm{DD_{a}}(\bm{X},\, j)$ is the Drawdown of uncompounded cumulative returns at time $j$ as defined in DaR.

$\mathrm{DaR_{a}}(\bm{X},\, \alpha)$ is the Drawdown at Risk of uncompounded cumulative returns as defined in DaR.

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance level.

$T$: is the number of observations.

See also: RiskMeasure, RMSettings, Portfolio, DaR, MDD, EDaR, RLDaR, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

alpha::T1 = 0.05: significance level, alpha ∈ (0, 1).

Behaviour in optimisations which take risk measures and use JuMP models

The CDaR is defined as an AffExpr with the key, :cdar_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :cdar_risk_ub.

Functor

(cdar::CDaR)(x::AbstractVector): computes the CDaR of a T×1 vector of portfolio returns x.

Examples

PortfolioOptimiser.EDaR — Type

mutable struct EDaR{T1 <: Real} <: RiskMeasureSolvers

Measures and computes the portfolio Entropic Drawdown at Risk of uncompounded cumulative returns (EDaR). It is the upper bound of the Chernoff inequality for the DaR and CDaR.

Uses ExponentialCone constraints.

$\mathrm{DaR}(\bm{X},\, \alpha) \leq \mathrm{CDaR}(\bm{X},\, \alpha) \leq \mathrm{EDaR}(\bm{X},\, \alpha) \leq \mathrm{RLDaR}(\bm{X},\, \alpha,\, \kappa) \leq \mathrm{MDD}(\bm{X})$.

\[\begin{align} \mathrm{EDaR_{a}}(\bm{X},\alpha) &= \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\mathrm{DD_{a}}(\bm{X}),\, z, \,\alpha)\right\}\,. \end{align}\]

Where:

$\mathrm{ERM}(\mathrm{DD_{a}}(\bm{X}),\, z, \,\alpha)$ is the Entropic Risk Measure as defined in ERM, using the Drawdown of uncompounded cumulative returns as defined in DaR.

$z$: is the entropic drawdown at risk.

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance level.

See also: RiskMeasureSolvers, RMSettings, Portfolio, DaR, MDD, CDaR, RLDaR, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

alpha::T1 = 0.05: significance level, alpha ∈ (0, 1).

solvers::Union{Nothing, PortOptSolver, <:AbstractVector{PortOptSolver}} = nothing: optional abstract dictionary containing the solvers, their settings, solution criteria, and other arguments.

If isnothing(solvers): it takes its value from the solvers property of the instance of Portfolio.

Behaviour in optimisations which take risk measures and use JuMP models

The EDaR is defined as an AffExpr with the key, :edar_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :edar_risk_ub.

Functor

(edar::EDaR)(x::AbstractVector): computes the EDaR of a T×1 vector of portfolio returns x.

Examples

PortfolioOptimiser.RLDaR — Type

mutable struct RLDaR{T1 <: Real, T2 <: Real} <: RiskMeasureSolvers

Measures and computes the portfolio Relativistic Drawdown at Risk of uncompounded cumulative returns (RLDaR). It is a generalisation of the EDaR.

Uses PowerCone constraints.

$\mathrm{DaR}(\bm{X},\, \alpha) \leq \mathrm{CDaR}(\bm{X},\, \alpha) \leq \mathrm{EDaR}(\bm{X},\, \alpha) \leq \mathrm{RLDaR}(\bm{X},\, \alpha,\, \kappa) \leq \mathrm{MDD}(\bm{X})$.

$\lim\limits_{\kappa \to 0} \mathrm{RLDaR}(\bm{X},\, \alpha,\, \kappa) \approx \mathrm{EDaR}(\bm{X},\, \alpha)$

$\lim\limits_{\kappa \to 1} \mathrm{RLDaR}(\bm{X},\, \alpha,\, \kappa) \approx \mathrm{MDD}(\bm{X})$

\[\begin{align} \mathrm{RLDaR_{a}}(\bm{X},\, \alpha,\, \kappa) &= \mathrm{RRM}(\mathrm{DD_{a}}(\bm{X}),\, \alpha,\, \kappa)\,. \end{align}\]

Where:

$\mathrm{RRM}(\mathrm{DD_{a}}(\bm{X}),\, \alpha,\, \kappa)$ is the Relativistic Risk Measure as defined in RRM, using the Drawdown of uncompounded cumulative returns as defined in DaR.

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance level.

$\kappa$: is the relativistic deformation parameter.

See also: RiskMeasureSolvers, RMSettings, Portfolio, VaR, WR, CVaR, EVaR, calc_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

alpha::T1 = 0.05: significance level, alpha ∈ (0, 1).

kappa::T1 = 0.3: relativistic deformation level, kappa ∈ (0, 1).

solvers::Union{Nothing, PortOptSolver, <:AbstractVector{PortOptSolver}} = nothing: optional abstract dictionary containing the solvers, their settings, solution criteria, and other arguments.

If isnothing(solvers): it takes its value from the solvers property of the instance of Portfolio.

Behaviour in optimisations which take risk measures and use JuMP models

The RLDaR is defined as an AffExpr with the key, :rldar_risk.

If it exists, the upper bound is defined via the portfolio variance with key, :rldar_risk_ub.

Functor

(rldar::RLDaR)(x::AbstractVector): computes the RLDaR of a T×1 vector of portfolio returns x.

Examples

Linear moments

These measure different combinations of linear moments (L-moments) of the returns distribution.

PortfolioOptimiser.OWA — Type

mutable struct OWA <: RiskMeasureOWA

Description

Defines the generic Ordered Weight Array.

Uses a vector of ordered weights generated by owa_l_moment or owa_l_moment_crm for arbitrary L-moment optimisations.

See also: RiskMeasureOWA, RMSettings, Portfolio, optimise!, set_rm, calc_risk(::OWA, ::AbstractVector), owa_l_moment, owa_l_moment_crm.

Properties

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

formulation::OWAFormulation = OWAApprox(): OWA risk measure settings.

w::Union{<:AbstractWeights, Nothing} = nothing: T×1 ordered weight vector of arbitrary L-moments generated by owa_l_moment or owa_l_moment_crm.

Examples

Tracking and Turnover

These measure how far a portfolio deviates from a benchmark.

PortfolioOptimiser.TrackingRM — Type

PortfolioOptimiser.TurnoverRM — Type