Dispersion

These measure the spread of the returns distribution.

Variance

PortfolioOptimiser.Variance — Type

mutable struct Variance <: RiskMeasureSigma

Measures and computes the portfolio Variance (Variance).

\[\begin{align} \mathrm{Variance}(\bm{w},\, \mathbf{\Sigma}) &= \bm{w}^\intercal \, \mathbf{\Sigma}\, \bm{w}\,. \end{align}\]

Where:

$\bm{w}$: is the N×1 vector of asset weights.

$\mathbf{\Sigma}$: is the N×N asset covariance matrix.

See also: RiskMeasureSigma, RMSettings, SD, PortClass, OptimType, NOC, NoAdj, IP, SDP, Portfolio, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

sigma::Union{<:AbstractMatrix, Nothing} = nothing: (optional) N×N covariance matrix.

If nothing: takes its value from the instance Portfolio, the specific property depends on the PortClass parameter of the OptimType used.

Behaviour in optimisations which take risk measures and use JuMP models

NoAdj, IP network and cluster constraints

Requires a solver that supports SecondOrderCone constraints.

Requires a solver capable of handling quadratic expressions.

Defines the variance risk, :variance_risk, as a QuadExpr.

If it exists, the upper bound is defined via the portfolio standard deviation with the key, :dev_ub.

Incompatible with NOC (Near Optimal Centering) optimisations because QuadExpr are not strictly convex.

SDP network and/or cluster constraints

Requires a solver that supports PSDCone constraints.

Defines the variance risk, :variance_risk, as an AffExpr.

Compatible with NOC (Near Optimal Centering) optimisations because AffExpr are strictly convex.

If it exists, the upper bound is defined via the portfolio variance with the key, :variance_risk_ub.

Functor

(variance::Variance)(w::AbstractVector): computes the Variance of an N×1 vector of asset weights.

Examples

Standard Deviation

PortfolioOptimiser.SD — Type

mutable struct SD <: RiskMeasureSigma

Measures and computes the portfolio Standard Deviation (SD).

\[\begin{align} \mathrm{SD}(\bm{w},\, \mathbf{\Sigma}) &= \left(\bm{w}^\intercal \, \mathbf{\Sigma} \, \bm{w}\right)^{1/2}\,. \end{align}\]

Where:

$\bm{w}$: is the N×1 vector of asset weights.

$\mathbf{\Sigma}$: is the N×N asset covariance matrix.

See also: RiskMeasureSigma, RMSettings, Variance, Portfolio, PortClass, OptimType, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

sigma::Union{<:AbstractMatrix, Nothing} = nothing: (optional) N×N covariance matrix.

If nothing: takes its value from the instance Portfolio, the specific property depends on the PortClass parameter of the OptimType used.

Behaviour in optimisations which take risk measures and use JuMP models

The SD risk is defined as a VariableRef with the key, :sd_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :sd_risk_ub.

Requires a solver that supports SecondOrderCone constraints.

Functor

(sd::SD)(w::AbstractVector): computes the SD risk of an N×1 vector of asset weights.

Examples

Mean Absolute Deviation

PortfolioOptimiser.MAD — Type

mutable struct MAD <: RiskMeasureTarget

Measures and computes the portfolio Mean Absolute Deviation (MAD). In other words, it is the expected value of the absolute deviation from a target return of the returns vector. This is a generalisation to accomodate the use of weighted means, and predefined targets rather than forcing the MAD to be computed from the expected returns vector.

\[\begin{align} \mathrm{MAD}(\bm{X}) &= \mathbb{E}\left(\left\lvert \bm{X} - r, \bm{w_{e}}\right\rvert\right)\,. \end{align}\]

Where:

$\mathbb{E}(\cdot, \bm{w})$: is the optionally weighted expected value.

If $\bm{w}$ is not provided it is the expected value.

$\bm{X}$: is the T×1 portfolio returns vector.

$\bm{w_{e}}$: is an optional weights vector for computing the MAD.

$\lvert \cdot \rvert$: is the absolute value.

$r$: is the target return value. Defaults to $\mathbb{E}(\bm{X},\, \bm{w})$.

$\bm{w}$: is an optional weights vector used for computing the expected value.

See also: RiskMeasureTarget, RMSettings, Portfolio, PortClass, OptimType, expected_risk, optimise!, calc_target_ret_mu, calc_ret_mu, calc_rm_target, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

target::target::Union{<:Real, <:AbstractVector{<:Real}, Nothing} = nothing:

If isnothing(target): Its value can be computed via calc_target_ret_mu or calc_rm_target.

else: directly provides the target return for computing the deviation.

w::Union{<:AbstractWeights, Nothing} = nothing: (optional, functor-exclusive) T×1 vector of weights for computing the expected value of the returns vector (internal expected value) via calc_target_ret_mu.

we::Union{<:AbstractWeights, Nothing} = nothing: (optional) T×1 vector of weights for computing the expected value of the absolute deviation (external expected value).

If isnothing(we): computes the unweighted mean.

Else: computes the weighted mean.

mu::Union{<:AbstractVector{<:Real}, Nothing} = nothing: (optional) N×1 vector of weights for computing the expected value of the returns vector (internal expected value) via calc_ret_mu or set_rm.

Warning

Using w to compute the MAD risk of a portfolio optimised via an optimisation which uses JuMP, you have to ensure that the value of mu used by the optimisation is consistent with the value of w–-i.e. it was computed with MuSimple using w. Otherwise, the calculation will be inconsistent with the value of :mad_risk. Alternatively, use the value of mu in both.

Behaviour in optimisations which take risk measures and use JuMP models

The MAD risk is defined as an AffExpr with the key, :mad_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :mad_risk_ub.

Functor

(mad::MAD)(X::AbstractMatrix, w::AbstractVector, fees::Real = 0.0): computes the MAD risk of a T×N returns matrix, a N×1 vector of asset weights w, and fees fees.

fees: must be consistent with the returns frequency.

Examples

Square Root Kurtosis

PortfolioOptimiser.Kurt — Type

mutable struct Kurt <: RiskMeasureMu

Measures and computes the portfolio Square Root Kurtosis (Kurt).

\[\begin{align} \mathrm{Kurt}(\bm{X}) &= \left(\dfrac{1}{T} \sum\limits_{t=1}^{T} \left(X_{t} - \mathbb{E}(\bm{X}) \right)^{4} \right)^{1/2}\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

$T$: is the number of observations.

$X_{t}$: is the t-th value of the portfolio returns vector.

$\mathbb{E}(\cdot)$: is the expected value.

See also: RiskMeasureMu, RMSettings, SKurt, Kurtosis, SKurtosis, expected_risk, Portfolio, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

w::Union{<:AbstractWeights, Nothing} = nothing: (optional, functor-exclusive) T×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu.

mu::Union{<:AbstractVector{<:Real}, Nothing} = nothing: (optional, functor-exclusive) N×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu.

kt::Union{<:AbstractMatrix, Nothing} = nothing: (optional, set_rm exclusive) N²×N² cokurtosis matrix.

If nothing: takes its value from the kurt instance Portfolio.

kt has no effect in optimisations that don't use JuMP models, because the kurtosis is computed from the returns vector.

Warning

Using w to compute the Kurt risk of a portfolio optimised via an optimisation which uses JuMP, you have to ensure that the value of kurt in the instance of Portfolio used by the optimisation is consistent with the value of w–-i.e. it was computed with MuSimple using w. Otherwise, the calculation will be inconsistent with the value of :kurt_risk.

Behaviour in optimisations which take risk measures and use JuMP models

Uses PSDCone constraints.

The Kurt risk is defined as an AffExpr with the key, :kurt_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :kurt_risk_ub.

Functor

(kurt::Kurt)(x::AbstractVector): computes the Kurt risk of a T×1 vector of portfolio returns x.

Examples

Range

PortfolioOptimiser.RG — Type

mutable struct RG <: RiskMeasure

Measures and computes the portfolio Range (RG). It is the difference between the best and worst return.

\[\begin{align} \mathrm{RG}(\bm{X}) &= \max(\bm{X}) - \min(\bm{X})\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

See also: RiskMeasure, RMSettings, Portfolio, WR, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

Behaviour in optimisations which take risk measures and use JuMP models

The RG risk is defined as an AffExpr with the key, :rg_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :rg_risk_ub.

Functor

(rg::RG)(x::AbstractVector): computes the RG risk of a T×1 vector of portfolio returns x.

Examples

Conditional Value at Risk Range

PortfolioOptimiser.CVaRRG — Type

mutable struct CVaRRG{T1 <: Real, T2 <: Real} <: RiskMeasure

Measures and computes the portfolio Conditional Value at Risk Range (CVaRRG). It measures the difference between the weighted average of all head gains up to the Value at Risk for gains, which is the threshold above or equal to which are the best beta % of portfolio returns. And the weighted average of all tail losses up to the Value at Risk, which is the threshold below or equal to which are the worst alpha % of portfolio returns.

\[\begin{align} \mathrm{CVaRRG}(\bm{X},\, \alpha,\,\beta) &= \mathrm{CVaR}(\bm{X},\, \alpha) - \mathrm{CVaR}(-\bm{X},\, \beta) \end{align}\]

Where:

$\mathrm{CVaR}(\bm{X},\, \alpha)$: is the Conditional Value at Risk as defined in CVaR.

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance level of losses.

$\beta$: is the significance level of gains.

See also: See also: RiskMeasure, RMSettings, Portfolio, CVaR, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

alpha::T1 = 0.05: significance level of losses, alpha ∈ (0, 1).

beta::T1 = 0.05: significance level of gains, beta ∈ (0, 1).

Behaviour in optimisations which take risk measures and use JuMP models

The CVaRRG is defined as an AffExpr with the key, :cvarrg_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :cvarrg_risk_ub.

Functor

(cvarrg::CVaRRG)(x::AbstractVector): computes the CVaRRG of a T×1 vector of portfolio returns x.

Examples

Tail Gini Range

PortfolioOptimiser.TGRG — Type

mutable struct TGRG{T1 <: Real, T2 <: Real, T3 <: Integer, 
                    T4 <: Real, T5 <: Real, T6 <: Integer} <: RiskMeasureOWA

Measures and computes the portfolio Tail Gini Range (TGRG). It is the difference between the Head Gini and Tail Gini.

\[\begin{align} \mathrm{TGRG}(\bm{X},\, \alpha,\, \beta) &= \mathrm{TG}(\bm{X},\,\alpha) - \mathrm{TG}(-\bm{X},\,\beta)\,. \end{align}\]

Where:

$\mathrm{TG}$: is the Tail Gini defined in TG.

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance level of losses.

$\beta$: is the significance level of gains.

See also: See also: RiskMeasureOWA, RMSettings, Portfolio, owa_tgrg, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

formulation::OWAFormulation = OWAApprox(): OWA risk measure settings.

alpha_i::T1 = 0.0001: start value of the significance level of CVaR losses, 0 < alpha_i < alpha < 1.

alpha::T2 = 0.05: end value of the significance level of CVaR losses, 0 < alpha_i < alpha < 1.

a_sim::T3 = 100: number of CVaRs to approximate the Tail Gini losses, a_sim > 0.

beta_i::T4 = 0.0001: start value of the significance level of CVaR gains, 0 < beta_i < beta < 1.

beta::T5 = 0.05: end value of the significance level of CVaR gains, 0 < beta_i < beta < 1.

b_sim::T6 = 100: number of CVaRs to approximate the Tail Gini gains, b_sim > 0.

Behaviour in optimisations which take risk measures and use JuMP models

The TGRG risk is defined as an AffExpr with the key, :tgrg_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :tgrg_risk_ub.

Functor

(tgrg::TGRG)(x::AbstractVector): computes the TGRG risk of a T×1 vector of portfolio returns x.

Examples

Gini Mean Difference

PortfolioOptimiser.GMD — Type

mutable struct GMD <: RiskMeasureOWA

Measures and computes the portfolio Gini Mean Difference (GMD).

See also: See also: RiskMeasureOWA, RMSettings, Portfolio, owa_gmd, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

formulation::OWAFormulation = OWAApprox(): OWA risk measure settings.

Behaviour in optimisations which take risk measures and use JuMP models

The GMD risk is defined as an AffExpr with the key, :gmd_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :gmd_risk_ub.

Functor

(gmd::GMD)(x::AbstractVector): computes the GMD risk of a T×1 vector of portfolio returns x.

Examples

Negative Quadratic Skewness

PortfolioOptimiser.NQSkew — Type

mutable struct NQSkew <: RiskMeasureSkew

Measures and computes the portfolio Negative Quadratic Skewness (NQSkew).

\[\begin{align} \nu &= \bm{w}^{\intercal} \mathbf{V} \bm{w}\\ \end{align}\]

Where:

$\bm{w}$: is the vector of asset weights.

$\mathbf{V}$: is the sum of the symmetric negative spectral slices of the coskewness.

See also: See also: RiskMeasureSkew, RMSettings, Portfolio, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

skew::Union{<:AbstractMatrix, Nothing}: (optional) N×N² custom coskewness matrix.

If isnothing(skew): takes its value from the skew property of the instance of Portfolio.

V::Union{Nothing, <:AbstractMatrix}: (optional) Na×Na custom sum of the symmetric negative spectral slices of the coskewness.

If isnothing(V): takes its value from the V property of the instance of Portfolio.

Behaviour in optimisations which take risk measures and use JuMP models

The NQSkew risk is defined as an QuadExpr with the key, :nqskew_risk = t_nqskew^2.

If it exists, the upper bound is defined via the portfolio variance with the key, :t_nqskew_ub.

Incompatible with NOC (Near Optimal Centering) optimisations because QuadExpr are not strictly convex.

Functor

(nqskew::NQSkew)(w::AbstractVector): computes the NQSkew risk of an N×1 vector of asset weights.

Examples

Negative Skewness

PortfolioOptimiser.NSkew — Type

mutable struct NSkew <: RiskMeasureSkew

Measures and computes the portfolio Negative Skewness (NSkew).

\[\begin{align} \nu &= \left(\bm{w}^{\intercal} \mathbf{V} \bm{w}\right)^{1/2}\\ \end{align}\]

Where:

$\bm{w}$: is the vector of asset weights.

$\mathbf{V}$: is the sum of the symmetric negative spectral slices of the coskewness.

See also: See also: RiskMeasureSkew, RMSettings, Portfolio, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

skew::Union{<:AbstractMatrix, Nothing}: (optional) N×N² custom coskewness matrix.

If isnothing(skew): takes its value from the skew property of the instance of Portfolio.

V::Union{Nothing, <:AbstractMatrix}: (optional) Na×Na custom sum of the symmetric negative spectral slices of the coskewness.

If isnothing(V): takes its value from the V property of the instance of Portfolio.

Behaviour in optimisations which take risk measures and use JuMP models

The NQSkew risk is defined as an AffExpr with the key, :nskew_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :nskew_risk_ub.

Functor

(nskew::NSkew)(w::AbstractVector): computes the NSkew risk of an N×1 vector of asset weights.

Examples

Brownian Distance Variance

PortfolioOptimiser.BDVariance — Type

struct BDVariance <: RiskMeasure

Measures and computes the portfolio Brownian Distance Variance (BDVariance). It captures linear and non-linear codependence between assets.

\[\begin{align} \mathrm{BDVariance}(\bm{X}) &= \mathrm{BDCov}(\bm{X},\, \bm{X}) = \dfrac{1}{T^{2}} \sum\limits_{i=1}^{T}\sum\limits_{j=1}^{T} A_{i,\,j}^2\\ \mathrm{BDCov}(\bm{X},\, \bm{Y}) &= \dfrac{1}{T^{2}} \sum\limits_{i=1}^{T} \sum\limits_{j=1}^{T} A_{i,\,j} B_{i,\,j}\\ A_{i,\,j} &= a_{i,\,j} - \bar{a}_{i\,.} - \bar{a}_{.\,j} + \bar{a}_{.\,.}\\ B_{i,\,j} &= b_{i,\,j} - \bar{b}_{i\,.} - \bar{b}_{.\,j} + \bar{b}_{.\,.}\\ a_{i,\,j} &= \lVert X_{i} - X_{j} \rVert_{2}, \quad \forall i,\, j = 1,\, \ldots ,\, T\\ b_{i,\,j} &= \lVert Y_{i} - Y_{j} \rVert_{2}, \quad \forall i,\, j = 1,\, \ldots ,\, T\,. \end{align}\]

Where:

$\bm{X}$ and $\bm{Y}$ : are random variables, they are equal in this case as they are the portfolio returns.

$a_{i,\,j}$ and $b_{i,\,j}$ : are entries of a distance matrix where $i$ and $j$ are points in time. Each entry is defined as the Euclidean distance $\lVert \cdot \rVert_{2}$ between the value of the random variable at time $i$ and its value at time $j$.

$\bar{a}_{i\,\cdot}$ and $\bar{b}_{i\,\cdot}$ : are the $i$-th row means of their respective matrices.

$\bar{a}_{\cdot\,j}$ and $\bar{b}_{\cdot\,j}$ : are the $j$-th column means of their respective matrices.

$\bar{a}_{\cdot\,\cdot}$ and $\bar{b}_{\cdot\,\cdot}$ : are the grand means of their respective matrices.

$A_{i,\,j}$ and $B_{i,\,j}$ : are doubly centered distances.

See also: See also: RiskMeasure, RMSettings, Portfolio, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

Behaviour in optimisations which take risk measures and use JuMP models

The BDVariance risk is defined as a QuadExpr with the key, :bdvariance_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :bdvariance_risk_ub.

Functor

(bdvariance::BDVariance)(x::AbstractVector): computes the BDVariance risk of a T×1 vector of portfolio returns x.

Examples

Worst Case Variance

PortfolioOptimiser.WCVariance — Type

mutable struct WCVariance{T1 <: Real} <: RiskMeasureSigma

Measures and computes the portfolio Variance with uncertainty sets.

Only has an effect on optimisations which take risk measures and use JuMP models.

It is not affected by VarianceFormulation.

It otherwise acts the same as the Variance except for the fact that it always gets its value from the cov property of the Portfolio instance.

\[\begin{align} \mathrm{Variance}(\bm{w},\, \mathbf{\Sigma}) &= \bm{w}^\intercal \, \mathbf{\Sigma}\, \bm{w}\,. \end{align}\]

Where:

$\bm{w}$: is the N×1 vector of asset weights.

$\mathbf{\Sigma}$: is the N×N asset covariance matrix.

See also: RiskMeasureSigma, RMSettings, WCType, WCSetMuSigma, Box, Ellipse, Variance, VarianceFormulation, Portfolio, wc_statistics!, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

wc_set::WCSetMuSigma = Box(): uncertainty set apply to the variance in the optimisation.

If isa(wc_set, Box): uses cov_l and cov_u, either from the risk measure, or their corresponding instance of Portfolio.

If isa(wc_set, Ellipse): uses cov_sigma and k_sigma, either from the risk measure, or their corresponding instance of Portfolio.

sigma::Union{<:AbstractMatrix, Nothing} = nothing: (optional) N×N covariance matrix.

If nothing: takes its value from the cov property of the instance Portfolio.

cov_l::Union{AbstractMatrix{<:Real}, Nothing} = nothing: N×N lower percentile of the covariance matrix of the Box uncertainty set.

If nothing: takes its value from the cov_l property of the instance of Portfolio.

cov_u::Union{AbstractMatrix{<:Real}, Nothing} = nothing: N×N upper percentile of the covariance matrix of the Box uncertainty set.

If nothing: takes its value from the cov_u property of the instance of Portfolio.

cov_sigma::Union{AbstractMatrix{<:Real}, Nothing} = nothing: N²×N² covariance matrix of the estimation errors of the covariance of the Ellipse uncertainty set.

If nothing: takes its value from the cov_sigma property of the instance of Portfolio.

k_sigma::T1 = Inf: square root of the size of the Ellipse uncertainty set.

If nothing: takes its value from the k_sigma property of the instance of Portfolio.

Behaviour in optimisations which take risk measures and use JuMP models

Uses PSDCone constraints.

The WCVariance risk is defined as an AffExpr with the key, :wcvariance_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :wcvariance_risk_ub.

Functor

(variance::WCVariance)(w::AbstractVector): computes the WCVariance (which is equal to the variance) of an N×1 vector of asset weights.

Examples