Downside dispersion

These measure how far the returns deviate from the mean in the negative direction.

Semi Variance

PortfolioOptimiser.SVariance — Type

mutable struct SVariance <: RiskMeasureTarget

Measures and computes the portfolio Semi Variance (SVariance). Measures the variance equal to or below the target return.

\[\begin{align} \mathrm{SVariance}(\bm{X}) &= \dfrac{1}{T-1} \sum\limits_{t=1}^{T}\min\left(X_{t} - \mathbb{E}(\bm{X}),\, 0\right)^{2}\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

$T$: is the number of observations.

$X_{t}$: is the t-th value of the portfolio returns vector.

$\mathbb{E}(\cdot)$: is the expected value.

See also: RiskMeasureMu, RMSettings, Portfolio, SD, Variance, SSD, NOC, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

target::target::Union{<:Real, <:AbstractVector{<:Real}, Nothing} = nothing: minimum return threshold for classifying downside returns. Only returns equal to or below this value are considered in the calculation. Its value can be computed via calc_target_ret_mu or calc_rm_target.

w::Union{<:AbstractWeights, Nothing} = nothing: (optional, functor-exclusive) T×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu.

mu::Union{<:AbstractVector{<:Real}, Nothing} = nothing: (optional) N×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu or set_rm.

Warning

Using w to compute the SVariance risk of a portfolio optimised via an optimisation which uses JuMP, you have to ensure that the value of mu used by the optimisation is consistent with the value of w–-i.e. it was computed with MuSimple using w. Otherwise, the calculation will be inconsistent with the value of :svariance_risk. Alternatively, use the value of mu in both.

Behaviour in optimisations which take risk measures and use JuMP models

The SVariance risk is defined as a QuadExpr with the key, :svariance_risk.

If it exists, the upper bound is defined via the portfolio semi standard deviation with the key, :sdev_ub.

Incompatible with NOC (Near Optimal Centering) optimisations because QuadExpr are not strictly convex.

Functor

(svariance::SVariance)(X::AbstractMatrix, w::AbstractVector, fees::Real = 0.0): computes the SVariance risk of a T×N returns matrix, a N×1 vector of asset weights w, and fees fees.

fees: must be consistent with the returns frequency.

Examples

Semi Standard Deviation

PortfolioOptimiser.SSD — Type

mutable struct SSD <: RiskMeasureTarget

Measures and computes the portfolio Semi Standard Deviation (SSD). Measures the standard deviation equal to or below the target return. If the target return is the risk-free rate, the risk-adjusted return ratio of this risk measure is commonly known as the Sortino ratio.

\[\begin{align} \mathrm{SSD}(\bm{X}) &= \left(\dfrac{1}{T-1} \sum\limits_{t=1}^{T}\min\left(X_{t} - r,\, 0\right)^{2}\right)^{1/2}\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

$T$: is the number of observations.

$r$: is the minimum acceptable return.

$X_{t}$: is the t-th value of the portfolio returns vector.

$\mathbb{E}(\cdot)$: is the expected value.

See also: RiskMeasureMu, RMSettings, Portfolio, SD, Variance, SVariance, expected_risk, optimise!, calc_target_ret_mu, calc_ret_mu, calc_rm_target, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

target::target::Union{<:Real, <:AbstractVector{<:Real}, Nothing} = nothing: minimum return threshold for classifying downside returns. Only returns equal to or below this value are considered in the calculation. Its value can be computed via calc_target_ret_mu or calc_rm_target.

w::Union{<:AbstractWeights, Nothing} = nothing: (optional, functor-exclusive) T×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu.

mu::Union{<:AbstractVector{<:Real}, Nothing} = nothing: (optional) N×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu or set_rm.

Warning

Using w to compute the SSD risk of a portfolio optimised via an optimisation which uses JuMP, you have to ensure that the value of mu used by the optimisation is consistent with the value of w–-i.e. it was computed with MuSimple using w. Otherwise, the calculation will be inconsistent with the value of :ssd_risk. Alternatively, use the value of mu in both.

Behaviour in optimisations which take risk measures and use JuMP models

The SSD risk is defined as an AffExpr with the key, :ssd_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :ssd_risk_ub.

Functor

(ssd::SSD)(X::AbstractMatrix, w::AbstractVector, fees::Real = 0.0): computes the SSD risk of a T×N returns matrix, a N×1 vector of asset weights w, and fees fees.

fees: must be consistent with the returns frequency.

Examples

First Lower Partial Moment

PortfolioOptimiser.FLPM — Type

mutable struct FLPM <: RiskMeasureTarget

Measures and computes the portfolio First Lower Partial Moment (FLPM). Measures the dispersion equal to or below the target return threshold. The risk-adjusted return ratio of this risk measure is commonly known as the Omega ratio.

\[\begin{align} \mathrm{FLPM}(\bm{X},\, r) &= \dfrac{1}{T} \sum\limits_{t=1}^{T}\max\left(r - X_{t},\, 0\right)\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

$r$: is the minimum acceptable return.

$T$: is the number of observations.

$X_{t}$: is the t-th value of the portfolio returns vector.

See also: RiskMeasureTarget, RMSettings, Portfolio, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

target::T1 = 0.0: minimum return threshold for classifying downside returns. Only returns equal to or below this value are considered in the calculation. Its value can be computed via calc_target_ret_mu or calc_rm_target.

w::Union{<:AbstractWeights, Nothing} = nothing: (optional, functor-exclusive) T×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu.

mu::Union{<:AbstractVector{<:Real}, Nothing} = nothing: (optional) N×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu or set_rm.

Warning

Using w to compute the FLPM risk of a portfolio optimised via an optimisation which uses JuMP, you have to ensure that the value of mu used by the optimisation is consistent with the value of w–-i.e. it was computed with MuSimple using w. Otherwise, the calculation will be inconsistent with the value of :flpm_risk. Alternatively, use the value of mu in both.

Behaviour in optimisations which take risk measures and use JuMP models

The FLPM risk is defined as an AffExpr with the key, :flpm_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :flpm_risk_ub.

Functor

(flpm::FLPM)(X::AbstractMatrix, w::AbstractVector, fees::Real = 0.0): computes the FLPM risk of a T×N returns matrix, a N×1 vector of asset weights w, and fees fees.

fees: must be consistent with the returns frequency.

Examples

Square Root Semi Kurtosis

PortfolioOptimiser.SKurt — Type

mutable struct SKurt <: RiskMeasureMu

Measures and computes the portfolio Square Root Semi Kurtosis (SKurt). Measures the kurtosis equal to or below the mean portfolio returns.

\[\begin{align} \mathrm{SKurt}(\bm{X}) &= \left(\dfrac{1}{T} \sum\limits_{t=1}^{T} \min\left(X_{t} - \mathbb{E}(\bm{X}),\, 0 \right)^{4} \right)^{1/2}\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

$T$: is the number of observations.

$X_{t}$: is the t-th value of the portfolio returns vector.

$\mathbb{E}(\cdot)$: is the expected value.

See also: RiskMeasureMu, RMSettings, Kurt, Kurtosis, SKurtosis, Portfolio, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

w::Union{<:AbstractWeights, Nothing} = nothing: (optional, functor-exclusive) T×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu.

mu::Union{<:AbstractVector{<:Real}, Nothing} = nothing: (optional, functor-exclusive) N×1 vector of weights for computing the expected value of the returns vector via calc_ret_mu.

kt::Union{<:AbstractMatrix, Nothing} = nothing: (optional, set_rm exclusive) N²×N² cokurtosis matrix.

If nothing: takes its value from the skurt instance Portfolio.

kt has no effect in optimisations that don't use JuMP models, because the kurtosis is computed from the returns vector.

Warning

Using w to compute the Kurt risk of a portfolio optimised via an optimisation which uses JuMP, you have to ensure that the value of skurt in the instance of Portfolio used by the optimisation is consistent with the value of w–-i.e. it was computed with MuSimple using w. Otherwise, the calculation will be inconsistent with the value of :skurt_risk.

Behaviour in optimisations which take risk measures and use JuMP models

Uses PSDCone constraints.

The Kurt risk is defined as an AffExpr with the key, :skurt_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :skurt_risk_ub.

Functor

(skurt::SKurt)(x::AbstractVector): computes the SKurt risk of a T×1 vector of portfolio returns x.

Examples

Negative Quadratic Semi Skewness

PortfolioOptimiser.NQSSkew — Type

mutable struct NQSSkew <: RiskMeasureSkew

Measures and computes the portfolio Negative Quadratic Semi Skewness (NQSSkew).

\[\begin{align} \nu &= \bm{w}^{\intercal} \mathbf{V} \bm{w}\\ \end{align}\]

Where:

$\bm{w}$: is the vector of asset weights.

$\mathbf{V}$: is the sum of the symmetric negative spectral slices of the semi coskewness.

See also: See also: RiskMeasureSkew, RMSettings, Portfolio, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

skew::Union{<:AbstractMatrix, Nothing}: (optional) N×N² custom semi coskewness matrix.

If isnothing(skew): takes its value from the sskew property of the instance of Portfolio.

V::Union{Nothing, <:AbstractMatrix}: (optional) Na×Na custom sum of the symmetric negative spectral slices of the semi coskewness.

If isnothing(V): takes its value from the SV property of the instance of Portfolio.

Behaviour in optimisations which take risk measures and use JuMP models

The NQSSkew risk is defined as an AffExpr with the key, :nqsskew_risk = t_nqsskew^2.

If it exists, the upper bound is defined via the portfolio variance with the key, :t_nqsskew_ub.

Incompatible with NOC (Near Optimal Centering) optimisations because QuadExpr are not strictly convex.

Functor

(nqsskew::NQSSkew)(w::AbstractVector): computes the NQSSkew risk of an N×1 vector of asset weights.

Examples

Negative Semi Skewness

PortfolioOptimiser.NSSkew — Type

mutable struct NSSkew <: RiskMeasureSkew

Measures and computes the portfolio Negative Semi Skewness (NSSkew).

\[\begin{align} \nu &= \left(\bm{w}^{\intercal} \mathbf{V} \bm{w}\right)^{1/2}\\ \end{align}\]

Where:

$\bm{w}$: is the vector of asset weights.

$\mathbf{V}$: is the sum of the symmetric negative spectral slices of the semi coskewness.

See also: See also: RiskMeasureSkew, RMSettings, Portfolio, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

skew::Union{<:AbstractMatrix, Nothing}: (optional) N×N² custom semi coskewness matrix.

If isnothing(skew): takes its value from the sskew property of the instance of Portfolio.

V::Union{Nothing, <:AbstractMatrix}: (optional) Na×Na custom sum of the symmetric negative spectral slices of the semi coskewness.

If isnothing(V): takes its value from the SV property of the instance of Portfolio.

Behaviour in optimisations which take risk measures and use JuMP models

The NSSkew risk is defined as an AffExpr with the key, :nsskew_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :nsskew_risk_ub.

Functor

(nsskew::NSSkew)(w::AbstractVector): computes the NSSkew risk of an N×1 vector of asset weights.

Examples