Downside

These measure different aspects of the tail (negative side) of the returns distribution.

Worst Realisation

PortfolioOptimiser.WR — Type

struct WR <: RiskMeasure

Measures and computes the portfolio Worst Realization/Return (WR). It is the absolute value of the most extreme loss for the period. Best used in combination with other risk measures.

$\mathrm{VaR}(\bm{X},\, \alpha) \leq \mathrm{CVaR}(\bm{X},\, \alpha) \leq \mathrm{EVaR}(\bm{X},\, \alpha) \leq \mathrm{RLVaR}(\bm{X},\, \alpha,\, \kappa) \leq \mathrm{WR}(\bm{X})$.

\[\begin{align} \mathrm{WR}(\bm{X}) &= -\min(\bm{X})\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

See also: RiskMeasure, RMSettings, Portfolio, VaR, CVaR, EVaR, RLVaR, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

Behaviour in optimisations which take risk measures and use JuMP models

The WR risk is defined as an AffExpr with the key, :wr_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :wr_risk_ub.

Functor

(wr::WR)(x::AbstractVector): computes the WR risk of a T×1 vector of portfolio returns x.

Examples

Conditional Value at Risk

PortfolioOptimiser.CVaR — Type

mutable struct CVaR{T1 <: Real} <: RiskMeasure

Measures and computes the portfolio Conditional Value at Risk (CVaR). Also known as the Expected Shortfall, it is the weighted average of all tail losses up to the Value at Risk, which is the threshold below or equal to which are the worst alpha % of portfolio returns.

$\mathrm{VaR}(\bm{X},\, \alpha) \leq \mathrm{CVaR}(\bm{X},\, \alpha) \leq \mathrm{EVaR}(\bm{X},\, \alpha) \leq \mathrm{RLVaR}(\bm{X},\, \alpha,\, \kappa) \leq \mathrm{WR}(\bm{X})$.

\[\begin{align} \mathrm{CVaR}(\bm{X},\, \alpha) &= \mathrm{VaR}(\bm{X},\, \alpha) + \dfrac{1}{\alpha T} \sum\limits_{t=1}^{T} \max\left(-X_{t} - \mathrm{VaR}(\bm{X},\, \alpha),\, 0\right)\,. \end{align}\]

Where:

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance level.

$\mathrm{VaR}(\bm{X},\, \alpha)$: is the Value at Risk as defined in VaR.

$T$: is the number of observations.

$X_{t}$: is the t-th value of the portfolio returns vector.

See also: RiskMeasure, RMSettings, Portfolio, VaR, WR, EVaR, RLVaR, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

alpha::T1 = 0.05: significance level, alpha ∈ (0, 1).

Behaviour in optimisations which take risk measures and use JuMP models

The CVaR is defined as an AffExpr with the key, :cvar_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :cvar_risk_ub.

Functor

(cvar::CVaR)(x::AbstractVector): computes the CVaR of a T×1 vector of portfolio returns x.

Examples

Distributionally Robust Conditional Value at Risk

PortfolioOptimiser.DRCVaR — Type

Entropic Value at Risk

PortfolioOptimiser.EVaR — Type

mutable struct EVaR{T1 <: Real} <: RiskMeasureSolvers

Measures and computes the portfolio Entropic Value at Risk (EVaR). It is the upper bound of the Chernoff inequality for the VaR and CVaR.

Uses ExponentialCone constraints.

$\mathrm{VaR}(\bm{X},\, \alpha) \leq \mathrm{CVaR}(\bm{X},\, \alpha) \leq \mathrm{EVaR}(\bm{X},\, \alpha) \leq \mathrm{RLVaR}(\bm{X},\, \alpha,\, \kappa) \leq \mathrm{WR}(\bm{X})$.

\[\begin{align} \mathrm{EVaR}(\bm{X},\alpha) &= \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\bm{X},\, z, \,\alpha)\right\}\,. \end{align}\]

Where:

$\mathrm{ERM}(\bm{X},\, z, \,\alpha)$ is the entropic risk measure as defined in ERM.

$z$: is the entropic value at risk.

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance level.

See also: RiskMeasureSolvers, RMSettings, Portfolio, VaR, WR, CVaR, RLVaR, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

alpha::T1 = 0.05: significance level, alpha ∈ (0, 1).

solvers::Union{Nothing, PortOptSolver, <:AbstractVector{PortOptSolver}} = nothing: optional abstract dictionary containing the solvers, their settings, solution criteria, and other arguments.

If isnothing(solvers): it takes its value from the solvers property of the instance of Portfolio.

Behaviour in optimisations which take risk measures and use JuMP models

The EVaR is defined as an AffExpr with the key, :evar_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :evar_risk_ub.

Functor

(evar::EVaR)(x::AbstractVector): computes the EVaR of a T×1 vector of portfolio returns x.

Examples

Relativistic Value at Risk

PortfolioOptimiser.RLVaR — Type

mutable struct RLVaR{T1 <: Real, T2 <: Real} <: RiskMeasureSolvers

Measures and computes the portfolio Relativistic Value at Risk (RLVaR). It is a generalisation of the EVaR.

Uses PowerCone constraints.

$\mathrm{VaR}(\bm{X},\, \alpha) \leq \mathrm{CVaR}(\bm{X},\, \alpha) \leq \mathrm{EVaR}(\bm{X},\, \alpha) \leq \mathrm{RLVaR}(\bm{X},\, \alpha,\, \kappa) \leq \mathrm{WR}(\bm{X})$.

$\lim\limits_{\kappa \to 0} \mathrm{RLVaR}(\bm{X},\, \alpha,\, \kappa) \approx \mathrm{EVaR}(\bm{X},\, \alpha)$

$\lim\limits_{\kappa \to 1} \mathrm{RLVaR}(\bm{X},\, \alpha,\, \kappa) \approx \mathrm{WR}(\bm{X})$

\[\begin{align} \mathrm{RLVaR}(\bm{X},\, \alpha,\, \kappa) &= \mathrm{RRM}(\bm{X},\, \alpha,\, \kappa)\,. \end{align}\]

Where:

$\mathrm{RRM}(\bm{X},\, \alpha,\, \kappa)$ is the Relativistic Risk Measure as defined in RRM.

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance level.

$\kappa$: is the relativistic deformation parameter.

See also: RiskMeasureSolvers, RMSettings, Portfolio, VaR, WR, CVaR, EVaR, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

alpha::T1 = 0.05: significance level, alpha ∈ (0, 1).

kappa::T1 = 0.3: relativistic deformation level, kappa ∈ (0, 1).

solvers::Union{Nothing, PortOptSolver, <:AbstractVector{PortOptSolver}} = nothing: optional abstract dictionary containing the solvers, their settings, solution criteria, and other arguments. In order to solve the problem, a solver must be compatible with ExponentialCone.

If isnothing(solvers): it takes its value from the solvers property of the instance of Portfolio.

Behaviour in optimisations which take risk measures and use JuMP models

The RLVaR is defined as an AffExpr with the key, :rlvar_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :rlvar_risk_ub.

Functor

(rlvar::RLVaR)(x::AbstractVector): computes the RLVaR of a T×1 vector of portfolio returns x.

Examples

Tail Gini

PortfolioOptimiser.TG — Type

mutable struct TG{T1 <: Real, T2 <: Real, T3 <: Integer} <: RiskMeasureOWA

Measures and computes the portfolio Tail Gini (TG).

See also: See also: RiskMeasureOWA, RMSettings, Portfolio, owa_tg, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

formulation::OWAFormulation = OWAApprox(): OWA risk measure settings.

alpha_i::T1 = 0.0001: start value of the significance level of CVaR losses, 0 < alpha_i < alpha < 1.

alpha::T2 = 0.05: end value of the significance level of CVaR losses, 0 < alpha_i < alpha < 1.

a_sim::T3 = 100: number of CVaRs to approximate the Tail Gini losses, a_sim > 0.

Behaviour in optimisations which take risk measures and use JuMP models

The TG risk is defined as an AffExpr with the key, :tg_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :tg_risk_ub.

Functor

(tg::TG)(x::AbstractVector): computes the TG risk of a T×1 vector of portfolio returns x.

Examples