Drawdown

These measure the drops in portfolio value from local maxima to subsequent local minima.

Maximum Drawdown of Uncompounded Cumulative Returns

PortfolioOptimiser.MDD — Type

struct MDD <: RiskMeasure

Measures and computes the portfolio Maximum Drawdown of uncompounded cumulative returns (MDD). It measures the largest peak-to-trough decline. Best used in combination with other risk measures. The risk-adjusted return ratio of this risk measure is commonly known as the Calmar ratio.

$\mathrm{DaR}(\bm{X},\, \alpha) \leq \mathrm{CDaR}(\bm{X},\, \alpha) \leq \mathrm{EDaR}(\bm{X},\, \alpha) \leq \mathrm{RLDaR}(\bm{X},\, \alpha,\, \kappa) \leq \mathrm{MDD}(\bm{X})$.

\[\begin{align} \mathrm{MDD_{a}}(\bm{X}) &= \max\mathrm{DD_{a}}(\bm{X})\,. \end{align}\]

Where:

$\mathrm{DD_{a}}(\bm{X})$ is the Drawdown of uncompounded cumulative returns as defined in DaR.

$\bm{X}$: is the T×1 vector of portfolio returns.

See also: RiskMeasure, RMSettings, Portfolio, DaR, CDaR, EDaR, RLDaR, DaR_r, CDaR_r, EDaR_r, RLDaR_r, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): risk measure configuration settings.

Behaviour in optimisations which take risk measures and use JuMP models

The MDD is defined as an AffExpr with the key, :mdd_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :mdd_risk_ub.

Functor

(mdd::MDD)(x::AbstractVector): computes the MDD of a T×1 vector of portfolio returns x.

Examples

Average Drawdown of Uncompounded Cumulative Returns

PortfolioOptimiser.ADD — Type

struct ADD <: RiskMeasure

Measures and computes the portfolio Average (Expected) Drawdown of uncompounded cumulative returns (ADD). This is a generalisation to accomodate the use of weighted means.

\[\begin{align} \mathrm{ADD_{a}}(\bm{X}) &= \mathbb{E}\lleft(\mathrm{DD_{a}}(\bm{X}, j)\,\forall j=1\,\ldots\,T\right)\,. \end{align}\]

Where:

$\mathrm{DD_{a}}(\bm{X}, j)$ is the Drawdown of uncompounded cumulative returns at time $j$ as defined in DaR.

$T$: is the number of observations.

$\bm{X}$: is the T×1 vector of portfolio returns.

See also: RiskMeasure, RMSettings, Portfolio, DaR, expected_risk, optimise!, set_rm.

Keyword Arguments

w::Union{<:AbstractWeights, Nothing} = nothing: optional T×1 vector of weights for computing the expected value of the returns vector.

Behaviour in optimisations which take risk measures and use JuMP models

The ADD is defined as an AffExpr with the key, :add_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :add_risk_ub.

Functor

(add::ADD)(x::AbstractVector): computes the ADD of a T×1 vector of portfolio returns x.

Examples

Ulcer Index of Uncompounded Cumulative Returns

PortfolioOptimiser.UCI — Type

mutable struct UCI <: RiskMeasure

Measures and computes the portfolio Ulcer Index of uncompounded cumulative returns (UCI). It is the normalised L2-norm of the portfolio drawdowns of uncompounded cumulative returns.

\[\begin{align} \mathrm{UCI_{a}}(\bm{X}) &= \left(\dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{a}}(\bm{X},\, j)^{2}\right)^{1/2}\,. \end{align}\]

Where:

$\mathrm{DD_{a}}(\bm{X},\, j)$ is the Drawdown of uncompounded cumulative returns at time $j$ as defined in DaR.

$\bm{X}$: is the T×1 vector of portfolio returns.

See also: RiskMeasure, RMSettings, Portfolio, DaR, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

Behaviour in optimisations which take risk measures and use JuMP models

The UCI is defined as an AffExpr with the key, :uci_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :uci_risk_ub.

Functor

(uci::UCI)(x::AbstractVector): computes the UCI of a T×1 vector of portfolio returns x.

Examples

Conditional Drawdown at Risk of Uncompounded Cumulative Returns

PortfolioOptimiser.CDaR — Type

mutable struct CDaR{T1 <: Real} <: RiskMeasure

Measures and computes the portfolio Conditional Drawdown at Risk of uncompounded cumulative returns (CDaR). It is the weighted average of all drawdowns up to the Drawdown at Risk, which is the threshold below or equal to which are the worst alpha % of portfolio drawdowns.

$\mathrm{DaR}(\bm{X},\, \alpha) \leq \mathrm{CDaR}(\bm{X},\, \alpha) \leq \mathrm{EDaR}(\bm{X},\, \alpha) \leq \mathrm{RLDaR}(\bm{X},\, \alpha,\, \kappa) \leq \mathrm{MDD}(\bm{X})$.

\[\begin{align} \mathrm{CDaR_{a}}(\bm{X},\, \alpha) &= \mathrm{DaR_{a}}(\bm{X},\, \alpha) + \dfrac{1}{\alpha T} \sum\limits_{j=0}^{T} \max\left(\mathrm{DD_{a}}(\bm{X},\, j) - \mathrm{DaR_{a}}(\bm{X},\, \alpha),\, 0 \right)\,. \end{align}\]

Where:

$\mathrm{DD_{a}}(\bm{X},\, j)$ is the Drawdown of uncompounded cumulative returns at time $j$ as defined in DaR.

$\mathrm{DaR_{a}}(\bm{X},\, \alpha)$ is the Drawdown at Risk of uncompounded cumulative returns as defined in DaR.

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance level.

$T$: is the number of observations.

See also: RiskMeasure, RMSettings, Portfolio, DaR, MDD, EDaR, RLDaR, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

alpha::T1 = 0.05: significance level, alpha ∈ (0, 1).

Behaviour in optimisations which take risk measures and use JuMP models

The CDaR is defined as an AffExpr with the key, :cdar_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :cdar_risk_ub.

Functor

(cdar::CDaR)(x::AbstractVector): computes the CDaR of a T×1 vector of portfolio returns x.

Examples

Entropic Drawdown at Risk of Uncompounded Cumulative Returns

PortfolioOptimiser.EDaR — Type

mutable struct EDaR{T1 <: Real} <: RiskMeasureSolvers

Measures and computes the portfolio Entropic Drawdown at Risk of uncompounded cumulative returns (EDaR). It is the upper bound of the Chernoff inequality for the DaR and CDaR.

Uses ExponentialCone constraints.

$\mathrm{DaR}(\bm{X},\, \alpha) \leq \mathrm{CDaR}(\bm{X},\, \alpha) \leq \mathrm{EDaR}(\bm{X},\, \alpha) \leq \mathrm{RLDaR}(\bm{X},\, \alpha,\, \kappa) \leq \mathrm{MDD}(\bm{X})$.

\[\begin{align} \mathrm{EDaR_{a}}(\bm{X},\alpha) &= \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\mathrm{DD_{a}}(\bm{X}),\, z, \,\alpha)\right\}\,. \end{align}\]

Where:

$\mathrm{ERM}(\mathrm{DD_{a}}(\bm{X}),\, z, \,\alpha)$ is the Entropic Risk Measure as defined in ERM, using the Drawdown of uncompounded cumulative returns as defined in DaR.

$z$: is the entropic drawdown at risk.

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance level.

See also: RiskMeasureSolvers, RMSettings, Portfolio, DaR, MDD, CDaR, RLDaR, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

alpha::T1 = 0.05: significance level, alpha ∈ (0, 1).

solvers::Union{Nothing, PortOptSolver, <:AbstractVector{PortOptSolver}} = nothing: optional abstract dictionary containing the solvers, their settings, solution criteria, and other arguments.

If isnothing(solvers): it takes its value from the solvers property of the instance of Portfolio.

Behaviour in optimisations which take risk measures and use JuMP models

The EDaR is defined as an AffExpr with the key, :edar_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :edar_risk_ub.

Functor

(edar::EDaR)(x::AbstractVector): computes the EDaR of a T×1 vector of portfolio returns x.

Examples

Relativistic Drawdown at Risk of Uncompounded Cumulative Returns

PortfolioOptimiser.RLDaR — Type

mutable struct RLDaR{T1 <: Real, T2 <: Real} <: RiskMeasureSolvers

Measures and computes the portfolio Relativistic Drawdown at Risk of uncompounded cumulative returns (RLDaR). It is a generalisation of the EDaR.

Uses PowerCone constraints.

$\mathrm{DaR}(\bm{X},\, \alpha) \leq \mathrm{CDaR}(\bm{X},\, \alpha) \leq \mathrm{EDaR}(\bm{X},\, \alpha) \leq \mathrm{RLDaR}(\bm{X},\, \alpha,\, \kappa) \leq \mathrm{MDD}(\bm{X})$.

$\lim\limits_{\kappa \to 0} \mathrm{RLDaR}(\bm{X},\, \alpha,\, \kappa) \approx \mathrm{EDaR}(\bm{X},\, \alpha)$

$\lim\limits_{\kappa \to 1} \mathrm{RLDaR}(\bm{X},\, \alpha,\, \kappa) \approx \mathrm{MDD}(\bm{X})$

\[\begin{align} \mathrm{RLDaR_{a}}(\bm{X},\, \alpha,\, \kappa) &= \mathrm{RRM}(\mathrm{DD_{a}}(\bm{X}),\, \alpha,\, \kappa)\,. \end{align}\]

Where:

$\mathrm{RRM}(\mathrm{DD_{a}}(\bm{X}),\, \alpha,\, \kappa)$ is the Relativistic Risk Measure as defined in RRM, using the Drawdown of uncompounded cumulative returns as defined in DaR.

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance level.

$\kappa$: is the relativistic deformation parameter.

See also: RiskMeasureSolvers, RMSettings, Portfolio, VaR, WR, CVaR, EVaR, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

alpha::T1 = 0.05: significance level, alpha ∈ (0, 1).

kappa::T1 = 0.3: relativistic deformation level, kappa ∈ (0, 1).

solvers::Union{Nothing, PortOptSolver, <:AbstractVector{PortOptSolver}} = nothing: optional abstract dictionary containing the solvers, their settings, solution criteria, and other arguments.

If isnothing(solvers): it takes its value from the solvers property of the instance of Portfolio.

Behaviour in optimisations which take risk measures and use JuMP models

The RLDaR is defined as an AffExpr with the key, :rldar_risk.

If it exists, the upper bound is defined via the portfolio variance with the key, :rldar_risk_ub.

Functor

(rldar::RLDaR)(x::AbstractVector): computes the RLDaR of a T×1 vector of portfolio returns x.

Examples