Dispersion

These measure the spread of the returns distribution.

Value at Risk Range

PortfolioOptimiser.VaRRG — Type

mutable struct VaRRG{T1 <: Real, T2 <: Real} <: HCRiskMeasure

Measures and computes the portfolio Value at Risk Range (VaRRG). It measures the difference between the threshold above or equal to which are the best beta % of portfolio returns. And the threshold below or equal to which are the worst alpha % of portfolio returns.

\[\begin{align} \mathrm{VaRRG}(\bm{X},\, \alpha,\,\beta) &= \mathrm{VaR}(\bm{X},\, \alpha) - \mathrm{VaR}(-\bm{X},\, \beta) \end{align}\]

Where:

$\mathrm{VaR}(\bm{X},\, \alpha)$: is the Value at Risk as defined in VaR.

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance level of losses.

$\beta$: is the significance level of gains.

See also: See also: RiskMeasure, RMSettings, Portfolio, VaR, expected_risk, optimise!, set_rm.

Keyword Arguments

settings::RMSettings = RMSettings(): configuration settings for the risk measure.

alpha::T1 = 0.05: significance level of losses, alpha ∈ (0, 1).

beta::T1 = 0.05: significance level of gains, beta ∈ (0, 1).

Functor

(varrg::VaRRG)(x::AbstractVector): computes the VaRRG of a T×1 vector of portfolio returns x.

Examples

Fourth Central Moment

PortfolioOptimiser.FTCM — Type