Downside

These measure different aspects of the tail (negative side) of the returns distribution.

Value at Risk

PortfolioOptimiser.VaR — Type

mutable struct VaR{T1 <: Real} <: HCRiskMeasure

\[\begin{align} \mathrm{CVaRRG}(\bm{X},\, \alpha,\,\beta) &= \mathrm{CVaR}(\bm{X},\, \alpha) - \mathrm{CVaR}(-\bm{X},\, \beta) \end{align}\]

Where:

$\mathrm{CVaR}(\bm{X},\, \alpha)$: is the Conditional Value at Risk as defined in CVaR.

$\bm{X}$: is the T×1 vector of portfolio returns.

$\alpha$: is the significance level of losses.

$\beta$: is the significance level of gains.

Description

Defines the Value at Risk.

Measures lower bound of the losses in the worst alpha % of cases.

$\mathrm{VaR}(\bm{X},\, \alpha) \leq \mathrm{CVaR}(\bm{X},\, \alpha) \leq \mathrm{EVaR}(\bm{X},\, \alpha) \leq \mathrm{RLVaR}(\bm{X},\, \alpha,\, \kappa) \leq \mathrm{WR}(\bm{X})$.

\[\begin{align} \mathrm{VaR}(\bm{X},\, \alpha) &= -\underset{t \in (0,\, T)}{\inf} \left\{ X_{t} \in \mathbb{R} \, | \, F_{\bm{X}}(X_{t}) > \alpha \right\}\,. \end{align}\]

See also: HCRiskMeasure, HCRMSettings, optimise!, expected_risk(::VaR, ::AbstractVector), CVaR, EVaR, RLVaR WR, DaR, DaR_r, CDaR, CDaR_r, EDaR, EDaR_r, RLDaR, RLDaR_r, MDD, MDD_r.

Properties

settings::HCRMSettings = HCRMSettings(): configuration settings for the risk measure.

alpha::T1 = 0.05: significance level, alpha ∈ (0, 1).

Behaviour

Validation

When setting alpha at construction or runtime, alpha ∈ (0, 1).

Examples

# Default settings
var = VaR()

# Custom significance level
var = VaR(; settings = HCRMSettings(; scale = 1.0), alpha = 0.01)