Risk value

Despite the fact that HCRiskMeasure are only compatible with , it is possible to compute the value of every risk measure for a given returns matrix and vector of weights.

There are similarly named higher level functions that operate at the level of PortfolioOptimiser.AbstractPortfolio.

PortfolioOptimiser.calc_risk — Function

calc_risk(::Equal, w::AbstractVector; delta::Real = 0, kwargs...)

Description

Compute the risk as the inverse of the length of w.

Inputs

Positional

equal::Equal: risk measure.

w::AbstractVector: N×1 vector of asset weights.

Named

delta::Real: is a displacement, used in risk_contribution and factor_risk_contribution.

Outputs

equal::Real: Equal Risk.

Examples

# Sample weights vector
w = [0.3, 0.5, 0.2]

# Calculate the equal risk measure
equal_rm = Equal()
equal_risk = calc_risk(equal_rm, w)

calc_risk(port::AbstractPortfolio; X::AbstractMatrix = port.returns,
          type::Symbol = isa(port, Portfolio) || isa(port, Portfolio) ? :Trad : :HRP,
          rm::AbstractRiskMeasure = SD())

Compute the risk for an AbstractRiskMeasure for a portfolio.

Inputs

port: portfolio.

X: T×N returns matrix.

type: optimisation type used to retrieve the weights vector from port.optimal[type].

rm: rism measure.

Outputs

r: risk.

PortfolioOptimiser.ERM — Function

ERM(x::AbstractVector, z::Real = 1.0, α::Real = 0.05)

Description

Compute the Entropic Risk Measure.

\[\begin{align} \mathrm{ERM}(\bm{X},\, z, \,\alpha) &= z \ln \left(\dfrac{M_{\bm{X}}\left(z^{-1}\right)}{\alpha} \right)\,. \end{align}\]

Where:

$M_{\bm{X}}\left(t\right)$ is the moment generating function of $\bm{X}$.

$\alpha \in (0,\,1)$ is the significance parameter.

Inputs

x::AbstractVector: T×1 returns vector.

z::Real = 1.0: entropic moment, can be obtained from get_z_from_model and get_z after optimising a Portfolio.

α::Real = 0.05: significance level, α ∈ (0, 1).

Outputs

er::Real: entropic risk.

Examples

# Sample returns vector
returns = [0.05, -0.03, 0.02, -0.01, 0.04]

# Calculate the entropic risk measure with default parameters
er1 = ERM(returns, 2.3, 0.03)

# Calculate with a 2.3 entropic moment and 3 % significance parameter
er2 = ERM(returns, 2.3, 0.03)

ERM(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05)

Description

Compute the Entropic Risk Measure.

\[\begin{align} \mathrm{ERM}(\bm{X},\, z, \,\alpha) &= \begin{cases} \underset{z,\, t,\, u}{\inf} & t + z \ln\left(\dfrac{1}{\alpha T}\right)\\ \mathrm{s.t.} & z \geq \sum\limits_{i=1}^{T} u_{i}\nonumber\\ & (-X_{i}-t,\, z,\, u_{i}) \in \mathcal{K}_{\exp} \, \forall \, i=1,\,\dots{},\, T \end{cases}\\ &= \underset{z>0}{\inf}\left\{ z \ln \left(\dfrac{M_{\bm{X}}\left(z^{-1}\right)}{\alpha} \right)\right\}\,. \end{align}\]

Where:

$M_{\bm{X}}\left(t\right)$ is the moment generating function of $\bm{X}$.

$\mathcal{K}_{\mathrm{exp}}$ is the exponential cone.

$\alpha \in (0,\,1)$ is the significance parameter.

See also: EVaR, calc_risk(::EVaR, ::AbstractVector), EVaR, EDaR, calc_risk(::EDaR, ::AbstractVector), EDaR, EDaR_r, calc_risk(::EDaR_r, ::AbstractVector), EDaR_r.

Inputs

x::AbstractVector: T×1 returns vector.

solvers::AbstractDict: JuMP-compatible solvers for exponential cone problems.

α::Real = 0.05: significance level, α ∈ (0, 1).

Behaviour

If no valid solution is found returns NaN.

Warning

α is not validated because this is an internal function. It should have been validated by EVaR, EDaR, or EDaR_r.

Outputs

er::Real: entropic risk.

Examples

# Sample returns vector
returns = [0.05, -0.03, 0.02, -0.01, 0.04]

# Calculate the entropic risk measure with default parameters
er1 = ERM(returns, Dict("solver" => my_solver))

# Calculate with a 3 % significance parameter
er2 = ERM(returns, Dict("solver" => my_solver), 0.03)

PortfolioOptimiser.RRM — Function

RRM(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05, κ::Real = 0.3)

Description

Compute the Relativistic Risk Measure. Used in RLVaR, RLDaR and RLDaR_r.

\[\begin{align} \mathrm{RRM}(X,\,\alpha,\,\kappa) &= \left\{ \begin{align} &\underset{z,\, t,\, \psi,\, \theta,\, \varepsilon,\, \omega}{\text{inf}} && t + z \ln_{\kappa} \left(\dfrac{1}{\alpha T}\right) + \sum\limits_{i=1}^T \left(\psi_{i} + \theta_{i} \right) \nonumber\\ &\mathrm{s.t.} && -X - t + \varepsilon + \omega \leq 0 \nonumber\\ &&&z \geq 0 \\ &&&\left( z\left(\dfrac{1+\kappa}{2\kappa}\right),\, \psi_{i}\left(\dfrac{1+\kappa}{\kappa}\right),\, \varepsilon_{i} \right) \in \mathcal{P}_{3}^{1/(1+\kappa),\, \kappa/(1+\kappa)}\nonumber\\ &&&\left( \omega_{i}\left(\dfrac{1}{1-\kappa}\right),\, \theta_{i}\left(\dfrac{1}{\kappa}\right),\, -z \left(\dfrac{1}{2\kappa}\right) \right) \in \mathcal{P}_{3}^{1-\kappa,\, \kappa}\nonumber\\ &&&\forall \, i=1,\,\dots{},\, T \nonumber \end{align} \right.\,. \end{align}\]

Where:

$\mathcal{P}_3^{\alpha,\, 1-\alpha}$ is the 3D power cone.

$\alpha \in (0,\,1)$ is the significance parameter.

$\kappa \in (0,\,1)$ is the relativistic deformation parameter.

See also: RLVaR, calc_risk(::RLVaR, ::AbstractVector), RLVaR, RLDaR, calc_risk(::RLDaR, ::AbstractVector), RLDaR, RLDaR_r, calc_risk(::RLDaR_r, ::AbstractVector), RLDaR_r.

Inputs

x::AbstractVector: T×1 returns vector.

solvers::AbstractDict: JuMP-compatible solvers for 3D power cone problems.

α::Real = 0.05: significance level, α ∈ (0, 1).

κ::Real = 0.3: relativistic deformation parameter, κ ∈ (0, 1).

Behaviour

If no valid solution is found returns NaN.

Warning

α and κ are not validated because this is an internal function. They should have been validated by RLVaR, RLDaR, or RLDaR_r.

Outputs

rlr::Real: relativistic risk.

Examples

# Sample returns vector
returns = [0.05, -0.03, 0.02, -0.01, 0.04]

# Calculate the relativistic risk with default parameters
rlr1 = RRM(returns, Dict("solver" => my_solver))

# Calculate with a 3 % significance parameter and 80 % deformation parameter
rlr2 = RRM(returns, Dict("solver" => my_solver), 0.03, 0.8)