Risk Functions

PortfolioOptimiser.Variance — Function

Variance(w::AbstractVector, Σ::AbstractMatrix)

Compute the Variance of a portfolio with weights w and covariance Σ. Square of SD.

\[\mathrm{Variance}(\bm{w},\, \mathbf{\Sigma}) = \bm{w}^\intercal \, \mathbf{\Sigma}\, \bm{w}\,.\]

PortfolioOptimiser.SD — Function

SD(w::AbstractVector, Σ::AbstractMatrix)

Compute the Standard Deviation of a portfolio with weights w and covariance Σ. Square root of Variance.

\[\mathrm{SD}(\bm{w},\, \mathbf{\Sigma}) = \left[\bm{w}^\intercal \, \mathbf{\Sigma} \, \bm{w}\right]^{1/2}\,.\]

PortfolioOptimiser.MAD — Function

MAD(x::AbstractVector)

Compute the Mean Absolute Deviation of a vector x of portfolio returns.

\[\mathrm{MAD}(\bm{x}) = \dfrac{1}{T} \sum\limits_{t=1}^T \left\lvert x_t - \mathbb{E}(\bm{x}) \right\rvert\,.\]

PortfolioOptimiser.Semi_SD — Function

Semi_SD(x::AbstractVector)

Compute the mean Semi-Standard Deviation of a vector x of portfolio returns.

\[\mathrm{SemiSD}(\bm{x}) = \left[\dfrac{1}{T-1} \sum\limits_{t=1}^{T}\min\left( \bm{x}_{t} - \mathbb{E}(\bm{x}),\, 0\right)^{2}\right]^{1/2}\,.\]

PortfolioOptimiser.FLPM — Function

FLPM(x::AbstractVector, r::Real = 0.0)

Compute the First Lower Partial Moment (omega ratio) of a vector x of portfolio returns with a minimum return target of r.

\[\mathrm{FLPM}(\bm{x},\, r) = \dfrac{1}{T} \sum\limits_{t=1}^{T}\max\left(r - \bm{x}_{t},\, 0\right)\,.\]

PortfolioOptimiser.SLPM — Function

SLPM(x::AbstractVector, r::Real = 0.0)

Compute the Second Lower Partial Moment (Omega Ratio) of a vector x of portfolio returns with a minimum return target of r.`

\[\mathrm{SLPM}(\bm{x},\, r) = \left[\dfrac{1}{T-1} \sum\limits_{t=1}^{T}\max\left(r - \bm{x}_{t},\, 0\right)^{2}\right]^{1/2}\,\]

PortfolioOptimiser.WR — Function

WR(x::AbstractVector)

Compute the Worst Realisation or Worst Case Scenario of a returns vector x.

\[\begin{align*} \mathrm{WR}(\bm{x}) &= -\min(\bm{x})\\ &= \max(-\bm{x})\,. \end{align*}\]

PortfolioOptimiser.VaR — Function

VaR(x::AbstractVector, α::Real = 0.05)

Compute the Value at Risk of a returns vector x at a significance level of α.

\[\mathrm{VaR}(\bm{x},\, \alpha) = -\inf_{t \in (0,\, T)} \left\{ x_{t} \in \mathbb{R} : F_{\bm{x}}(x_{t}) > \alpha \right\}\,,\]

PortfolioOptimiser.CVaR — Function

CVaR(x::AbstractVector, α::Real = 0.05)

Compute the Conditional Value at Risk of a returns vector x at a significance level of α.

\[\begin{align*} \mathrm{CVaR}(\bm{x},\, \alpha) &= \mathrm{VaR}(\bm{x},\, \alpha) - \dfrac{1}{\alpha T} \sum\limits_{t=1}^{T} \min\left( x_t + \mathrm{VaR}(\bm{x},\, \alpha),\, 0\right)\\ &= \mathrm{VaR}(\bm{x},\, \alpha) + \dfrac{1}{\alpha T} \sum\limits_{t=1}^{T} \max\left( -x_t - \mathrm{VaR}(\bm{x},\, \alpha),\, 0\right)\,. \end{align*}\]

PortfolioOptimiser.ERM — Function

ERM(x::AbstractVector, z::Real = 1.0, α::Real = 0.05)

Compute the Entropic Risk Measure of a vector x at a significance level of α for a given value of z.

ERM(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05)

Compute the Entropic Risk Measure at a significance level of α by minimising with respect to z, using a dictionary of JuMP-supported solvers. This is because in general we don't know the value of z that minimises the function.

\[\mathrm{ERM}(\bm{x},\, z, \,\alpha) = z \ln \left( \dfrac{M_{\bm{x}}\left(z^{-1}\right)}{\alpha} \right)\,,\]

where $M_{\bm{x}}\left(z^{-1}\right)$ is the moment generating function of $\bm{x}$.

PortfolioOptimiser.EVaR — Function

EVaR(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05)

Compute the Entropic Value at Risk of a returns vector x at a significance level of α.

\[\mathrm{EVaR}(\bm{x},\alpha) = \inf_{z > 0} \left\{\mathrm{ERM}(\bm{x},\, z, \,\alpha)\right\}\,.\]