Risk Measures

Public

PortfolioOptimiser.ADD_abs — Method

ADD_abs(x::AbstractVector)

Compute the Average Drawdown of uncompounded cumulative returns.

\[\mathrm{ADD_{a}}(\bm{x}) = \dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{a}}(\bm{x},\, j)\,,``` where ``\mathrm{DD_{a}}(\bm{x},\, j)`` is the Drawdown of uncompounded cumulative returns as defined in [`DaR_abs`](@ref). # Inputs - `x`: vector of portfolio returns.\]

PortfolioOptimiser.ADD_rel — Method

ADD_rel(x::AbstractVector)

Compute the Average Drawdown of compounded cumulative returns.

\[\mathrm{ADD_{r}}(\bm{r}) = \dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{r}}(\bm{x},\, j)\,,``` where ``\mathrm{DD_{a}}(\bm{x},\, j)`` is the Drawdown of compounded cumulative returns as defined in [`DaR_rel`](@ref). # Inputs - `x`: vector of portfolio returns.\]

PortfolioOptimiser.CDaR_abs — Function

CDaR_abs(x::AbstractVector; alpha::Real = 0.05)

Compute the Conditional Drawdown at Risk of uncompounded cumulative returns.

\[\mathrm{CDaR_{a}}(\bm{x},\, \alpha) = \mathrm{DaR_{a}}(\bm{x},\, \alpha) + \dfrac{1}{\alpha T} \sum\limits_{j=0}^{T} \max\left[\mathrm{DD_{a}}(\bm{x},\, j) - \mathrm{DaR_{a}}(\bm{x},\, \alpha),\, 0 \right] \,,``` where ``\mathrm{DD_{a}}(\bm{x},\, j)`` is the Drawdown of uncompounded cumulative returns as defined in [`DaR_abs`](@ref), and ``\mathrm{DaR_{a}}(\bm{x},\, \alpha)`` the Drawdown at Risk of uncompounded cumulative returns as defined in [`DaR_abs`](@ref). # Inputs - `x`: vector of portfolio returns. - `alpha`: significance level, alpha in (0, 1).\]

PortfolioOptimiser.CDaR_rel — Function

CDaR_rel(x::AbstractVector; alpha::Real = 0.05)

Compute the Conditional Drawdown at Risk of compounded cumulative returns.

\[\mathrm{CDaR_{r}}(\bm{x},\, \alpha) = \mathrm{DaR_{r}}(\bm{x},\, \alpha) + \dfrac{1}{\alpha T} \sum\limits_{j=0}^{T} \max\left[\mathrm{DD_{r}}(\bm{x},\, j) - \mathrm{DaR_{r}}(\bm{x},\, \alpha),\, 0 \right] \,,``` where ``\mathrm{DD_{r}}(\bm{x},\, j)`` is the Drawdown of compounded cumulative returns as defined in [`DaR_rel`](@ref), and ``\mathrm{DaR_{r}}(\bm{x},\, \alpha)`` the Drawdown at Risk of compounded cumulative returns as defined in [`DaR_rel`](@ref). # Inputs - `x`: vector of portfolio returns. - `alpha`: significance level, alpha in (0, 1).\]

PortfolioOptimiser.CVaR — Function

CVaR(x::AbstractVector, alpha::Real = 0.05)

Compute the Conditional Value at Risk.

\[\mathrm{CVaR}(\bm{x},\, \alpha) = \mathrm{VaR}(\bm{x},\, \alpha) - \dfrac{1}{\alpha T} \sum\limits_{t=1}^{T} \min\left( x_t + \mathrm{VaR}(\bm{x},\, \alpha),\, 0\right)\,,``` where ``\mathrm{VaR}(\bm{x},\, \alpha)`` is the value at risk as defined in [`VaR`](@ref). # Inputs - `x`: vector of portfolio returns. - `alpha`: significance level, alpha in (0, 1).\]

PortfolioOptimiser.DaR_abs — Function

DaR_abs(x::AbstractArray; alpha::Real = 0.05)

Compute the Drawdown at Risk of uncompounded cumulative returns.

\[\begin{align*} \mathrm{DaR_{a}}(\bm{x},\, \alpha) &= \underset{j \in (0,\, T)}{\max} \left\{ \mathrm{DD_{a}}(\bm{x},\, j) \in \mathbb{R} : F_{\mathrm{DD}}\left(\mathrm{DD_{a}}(\bm{x},\, j)\right) < 1 - \alpha \right\}\\ \mathrm{DD_{a}}(\bm{x},\, j) &= \underset{t \in (0,\, j)}{\max}\left( \sum\limits_{i=0}^{t} x_{i} \right) - \sum\limits_{i=0}^{j} x_{i} \end{align*}\,.\]

Inputs

x: vector of portfolio returns.

alpha: significance level, alpha in (0, 1).

PortfolioOptimiser.DaR_rel — Function

DaR_rel(x::AbstractArray; alpha::Real = 0.05)

Compute the Drawdown at Risk of compounded cumulative returns.

\[\begin{align*} \mathrm{DaR_{r}}(\bm{x},\, \alpha) &= \underset{j \in (0,\, T)}{\max} \left\{ \mathrm{DD_{r}}(\bm{x},\, j) \in \mathbb{R} : F_{\mathrm{DD}}\left(\mathrm{DD_{r}}(\bm{x},\, j)\right) < 1 - \alpha \right\}\\ \mathrm{DD_{r}}(\bm{x},\, j) &= \underset{t \in (0,\, j)}{\max}\left( \prod\limits_{i=0}^{t} \left(1+x_{i}\right) \right) - \prod\limits_{i=0}^{j} \left(1+x_{i}\right) \end{align*}\,.\]

Inputs

x: vector of portfolio returns.

alpha: significance level, alpha in (0, 1).

PortfolioOptimiser.EDaR_abs — Function

EDaR_abs(x::AbstractVector, solvers::AbstractDict, alpha::Real = 0.05)

Compute the Entropic Drawdown at Risk of uncompounded cumulative returns.

\[\begin{align*} \mathrm{EDaR_{a}}(\bm{x},\alpha) &= \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\mathrm{DD_{a}}(\bm{x}),\, z, \,\alpha)\right\}\\ \mathrm{DD_{a}}(\bm{x}) &= \left\{j \in (0,\, T) : \mathrm{DD_{a}}(\bm{x},\, j) \right\}\,,\end{align*}\]

where $\mathrm{ERM}(\bm{x},\, z, \,\alpha)$ is the entropic risk measure as defined in ERM and $\mathrm{DD_{a}}(\bm{x},\, j)$ the drawdown of uncompounded cumulative returns as defined in DaR_abs.

Inputs

x: vector of portfolio returns.

solvers: Provides the solvers and corresponding parameters for solving the JuMP model. There can be two key => value pairs.

:solver => value: value is a JuMP optimiser. The optimiser can be declared alongside its attributes by using JuMP.solver_with_attributes. Solver must support MOI.ExponentialCone, or JuMP must be able to transform it/them into a supported form.

:params => value: (optional) value must be a Dict or NamedTuple whose key => value pairs are the solver-specific settings.

alpha: significance level, alpha in (0, 1).

PortfolioOptimiser.EDaR_rel — Function

EDaR_rel(x::AbstractVector, solvers::AbstractDict, alpha::Real = 0.05)

Compute the Entropic Drawdown at Risk of compounded cumulative returns.

\[\begin{align*} \mathrm{EDaR_{r}}(\bm{x},\alpha) &= \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\mathrm{DD_{r}}(\bm{x}),\, z, \,\alpha)\right\}\\ \mathrm{DD_{r}}(\bm{x}) &= \left\{j \in (0,\, T) : \mathrm{DD_{r}}(\bm{x},\, j) \right\}\,,\end{align*}\]

where $\mathrm{ERM}(\bm{x},\, z, \,\alpha)$ is the entropic risk measure as defined in ERM and $\mathrm{DD_{r}}(\bm{x},\, j)$ the drawdown of compounded cumulative returns as defined in DaR_rel.

Inputs

x: vector of portfolio returns.

solvers: Provides the solvers and corresponding parameters for solving the JuMP model. There can be two key => value pairs.

:solver => value: value is a JuMP optimiser. The optimiser can be declared alongside its attributes by using JuMP.solver_with_attributes. Solver must support MOI.ExponentialCone, or JuMP must be able to transform it/them into a supported form.

:params => value: (optional) value must be a Dict or NamedTuple whose key => value pairs are the solver-specific settings.

alpha: significance level, alpha in (0, 1).

κ: relativistic deformation parameter.

PortfolioOptimiser.ERM — Function

ERM(x::AbstractVector, z::Real = 1.0, alpha::Real = 0.05)

Compute the Entropic Risk Measure.

\[\mathrm{ERM}(\bm{x},\, z, \,\alpha) = z \ln \left( \dfrac{M_{\bm{x}}\left(z^{-1}\right)}{\alpha} \right)\,,``` where ``M_{\bm{x}}\left(z^{-1}\right)`` is the moment generating function of ``\bm{x}``. # Inputs - `x`: vector. - `alpha`: significance level, alpha in (0, 1). - `z`: free parameter.\]

julia ERM(x::AbstractVector, solvers::AbstractDict, alpha::Real = 0.05)

Compute the Entropic Risk Measure by minimising the function with respect to `z`. Used in [`EVaR`](@ref), [`EDaR_abs`](@ref) and [`EDaR_rel`](@ref).

math \mathrm{ERM} = \begin{cases} \underset{z,\, t,\, u}{\min} & t + z \ln(\dfrac{1}{\alpha T})\ \mathrm{s.t.} & z \geq \sum\limits{i=1}^{T} u{i}\ & (-x{i}-t,\, z,\, u{i}) \in \mathcal{K}{\exp} \, \forall \, i=1,\,\dots{},\, T \end{cases}\,,`$where$\mathcal{K}{\exp}`` is the exponential cone.

Inputs

x: vector.

solvers: Provides the solvers and corresponding parameters for solving the JuMP model. There can be two key => value pairs.

:solver => value: value is a JuMP optimiser. The optimiser can be declared alongside its attributes by using JuMP.solver_with_attributes. Solver must support MOI.ExponentialCone, or JuMP must be able to transform it/them into a supported form.

:params => value: (optional) value must be a Dict or NamedTuple whose key => value pairs are the solver-specific settings.

alpha: significance level, alpha in (0, 1).

PortfolioOptimiser.EVaR — Function

EVaR(x::AbstractVector, solvers::AbstractDict, alpha::Real = 0.05)

Compute the Entropic Value at Risk. math \mathrm{EVaR}(\bm{x},\alpha) = \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\bm{x},\, z, \,\alpha)\right\}\,, where $\mathrm{ERM}(\bm{x},\, z, \,\alpha)$ is the entropic risk measure as defined in ERM.

Inputs

x: vector of portfolio returns.

solvers: Provides the solvers and corresponding parameters for solving the JuMP model. There can be two key => value pairs.

:solver => value: value is a JuMP optimiser. The optimiser can be declared alongside its attributes by using JuMP.solver_with_attributes. Solver must support MOI.ExponentialCone, or JuMP must be able to transform it/them into a supported form.

:params => value: (optional) value must be a Dict or NamedTuple whose key => value pairs are the solver-specific settings.

alpha: significance level, alpha in (0, 1).

PortfolioOptimiser.FLPM — Function

FLPM(x::AbstractVector; r::Real = 0.0)

Compute the First Lower Partial Moment (Omega ratio).

\[\mathrm{FLPM}(\bm{x},\, r) = \dfrac{1}{T} \sum\limits_{t=1}^{T}\max\left(r - \bm{x}_{t},\, 0\right)\,.\]

Inputs

x: vector of portfolio returns.

r: minimum return target.

PortfolioOptimiser.GMD — Method

GMD(x::AbstractVector)

Compute the Gini Mean Difference.

Inputs

x: vector of portfolio returns.

PortfolioOptimiser.Kurt — Method

Kurt(x::AbstractVector)

Compute the square root kurtosis.

\[\mathrm{Kurt}(\bm{x}) = \left[\dfrac{1}{T} \sum\limits_{t=1}^{T} \left( x_{t} - \mathbb{E}(\bm{x}) \right)^{4} \right]^{1/2}\,.\]

Inputs

x: vector of portfolio returns.

PortfolioOptimiser.MAD — Method

MAD(x::AbstractVector)

Compute the Mean Absolute Deviation.

\[\mathrm{MAD}(\bm{x}) = \dfrac{1}{T} \sum\limits_{t=1}^T \left\lvert x_t - \mathbb{E}(\bm{x}) \right\rvert\,.\]

Inputs

x: vector of portfolio returns.

PortfolioOptimiser.MDD_abs — Method

MDD_abs(x::AbstractVector)

Compute the Maximum Drawdown of uncompounded cumulative returns.

\[\mathrm{MDD_{a}}(\bm{x}) = \underset{j \in (0,\, T)}{\max} \mathrm{DD_{a}}(\bm{x},\, j)\,,``` where ``\mathrm{DD_{a}}(\bm{x},\, j)`` is the Drawdown of uncompounded cumulative returns as defined in [`DaR_abs`](@ref). # Inputs - `x`: vector of portfolio returns.\]

PortfolioOptimiser.MDD_rel — Method

MDD_rel(x::AbstractVector)

Compute the Maximum Drawdown of compounded cumulative returns.

\[\mathrm{MDD_{r}}(\bm{x}) = \underset{j \in (0,\, T)}{\max} \mathrm{DD_{r}}(\bm{x},\, j)\,,``` where ``\mathrm{DD_{a}}(\bm{x},\, j)`` is the Drawdown of compounded cumulative returns as defined in [`DaR_rel`](@ref). # Inputs - `x`: vector of portfolio returns.\]

PortfolioOptimiser.OWA — Method

OWA(x::AbstractVector, w::AbstractVector)

Compute the Ordered Weight Array risk measure.

Inputs

w: vector of asset weights.

x: vector of portfolio returns.

PortfolioOptimiser.RCVaR — Method

RCVaR(x::AbstractVector; alpha::Real = 0.05, beta::Real = alpha)

Compute the CVaR Range.

Inputs

x: vector of portfolio returns.

alpha: significance level of CVaR losses, alpha in (0, 1).

beta: significance level of CVaR gains, beta in (0, 1).

PortfolioOptimiser.RDaR_abs — Function

RDaR_abs(
    x::AbstractVector,    solvers::AbstractDict,    alpha::Real = 0.05,    kappa::Real = 0.3)

Compute the Relativistic Drawdown at Risk of uncompounded cumulative returns. math \mathrm{RDaR_{a}}(\bm{x},\, \alpha,\, \kappa) = \mathrm{RRM}(\mathrm{DD_{a}}(\bm{x}),\, \alpha,\, \kappa)\,, where $\mathrm{RRM}(\mathrm{DD_{a}}(\bm{x}),\, \alpha,\, \kappa)$ is the relativistic risk measure as defined in RRM, and $\mathrm{DD_{a}}(\bm{x})$ the drawdown of uncompounded cumulative returns as defined in DaR_abs.

Inputs

x: vector of portfolio returns.

solvers: Provides the solvers and corresponding parameters for solving the JuMP model. There can be two key => value pairs.

:solver => value: value is a JuMP optimiser. The optimiser can be declared alongside its attributes by using JuMP.solver_with_attributes. Solver must support MOI.PowerCone, or JuMP must be able to transform it/them into a supported form.

:params => value: (optional) value must be a Dict or NamedTuple whose key => value pairs are the solver-specific settings.

alpha: significance level, alpha in (0, 1).

κ: relativistic deformation parameter.

PortfolioOptimiser.RDaR_rel — Function

RDaR_rel(
    x::AbstractVector,    solvers::AbstractDict,    alpha::Real = 0.05,    kappa::Real = 0.3)

Compute the Relativistic Drawdown at Risk of compounded cumulative returns. math \mathrm{RDaR_{r}}(\bm{x},\, \alpha,\, \kappa) = \mathrm{RRM}(\mathrm{DD_{r}}(\bm{x}),\, \alpha,\, \kappa)\,, where $\mathrm{RRM}(\mathrm{DD_{r}}(\bm{x}),\, \alpha,\, \kappa)$ is the Relativistic Risk Measure as defined in RRM where the returns vector, and $\mathrm{DD_{r}}(\bm{x})$ the drawdown of compounded cumulative returns as defined in DaR_rel.

Inputs

x: vector of portfolio returns.

solvers: Provides the solvers and corresponding parameters for solving the JuMP model. There can be two key => value pairs.

:solver => value: value is a JuMP optimiser. The optimiser can be declared alongside its attributes by using JuMP.solver_with_attributes. Solver must support MOI.PowerCone, or JuMP must be able to transform it/them into a supported form.

:params => value: (optional) value must be a Dict or NamedTuple whose key => value pairs are the solver-specific settings.

alpha: significance level, alpha in (0, 1).

κ: relativistic deformation parameter.

PortfolioOptimiser.RG — Method

RG(x::AbstractVector)

Compute the Range.

Inputs

x: vector of portfolio returns.

PortfolioOptimiser.RRM — Function

RRM(
    x::AbstractVector,    solvers::AbstractDict,    alpha::Real = 0.05,    κ::Real = 0.3)

Compute the Relativistic Risk Measure. ```math \mathrm{RRM}(\bm{x},\, \alpha,\, \kappa) = \begin{cases} \underset{z,\, t,\, \psi,\, \theta,\, \varepsilon,\, \omega}{\min} & t + z \ln{\kappa}\left(\dfrac{1}{\alpha T}\right) + \sum\limits{i=1}^{T} \left(\psi{i} + \theta{i}\right) \ \mathrm{s.t.} & z \geq 0 \ & -x{i} - t + \varepsilon{i} + \omega{i} \leq 0 \ & \left(z\left(\dfrac{1+\kappa}{2\kappa}\right),\, \psi{i}\left(\dfrac{1+\kappa}{\kappa}\right),\, \varepsilon{i} \right) \in \mathcal{P}{3}^{1/(1+\kappa),\, \kappa/(1+\kappa)} \ & \left(\omega{i}\left(\dfrac{1}{1-\kappa}\right),\, \theta{i}\left(\dfrac{1}{\kappa}\right),\, -z\left(\dfrac{1}{2\kappa}\right) \right) \in \mathcal{P}{3}^{1-\kappa,\, \kappa} \ & \forall \, i = 1,\,\ldots{},\, T \end{cases}\,,`$where$\ln{\kappa}(x) = \dfrac{x^{\kappa} - x^{-\kappa}}{2 \kappa}$and$\mathcal{P}_3^{\alpha,\, 1-\alpha}`` is the 3D Power Cone.

Inputs

x: vector.

solvers: Provides the solvers and corresponding parameters for solving the JuMP model. There can be two key => value pairs.

:solver => value: value is a JuMP optimiser. The optimiser can be declared alongside its attributes by using JuMP.solver_with_attributes. Solver must support MOI.PowerCone, or JuMP must be able to transform it/them into a supported form.

:params => value: (optional) value must be a Dict or NamedTuple whose key => value pairs are the solver-specific settings.

alpha: significance level, alpha in (0, 1).

κ: relativistic deformation parameter.

PortfolioOptimiser.RTG — Method

RTG(x::AbstractVector; alpha_i::Real = 0.0001, alpha::Real = 0.05, a_sim::Real = 100,
    beta_i::Real = alpha_i, beta::Real = alpha, b_sim::Integer = a_sim)

Compute the Tail Gini Range.

Inputs

x: vector of portfolio returns.

alpha_i: start value of the significance level of CVaR losses, 0 <alpha_i < alpha < 1.

alpha: end value of the significance level of CVaR losses, alpha in (0, 1).

a_sim: number of CVaRs to approximate the Tail Gini losses, a_sim > 0.

beta_i: start value of the significance level of CVaR gains, 0 < beta_i < beta < 1.

beta: end value of the significance level of CVaR gains, beta in (0, 1).

b_sim: number of CVaRs to approximate the Tail Gini gains, b_sim > 0.

PortfolioOptimiser.RVaR — Function

RVaR(
    x::AbstractVector,    solvers::AbstractDict,    alpha::Real = 0.05,    κ::Real = 0.3)

Compute the Relativistic Value at Risk. math \mathrm{RVaR}(\bm{x},\, \alpha,\, \kappa) = \mathrm{RRM}(\bm{x},\, \alpha,\, \kappa)\,, where $\mathrm{RRM}(\bm{x},\, \alpha,\, \kappa)$ is the Relativistic Risk Measure as defined in RRM.

Inputs

x: vector of portfolio returns.

solvers: Provides the solvers and corresponding parameters for solving the JuMP model. There can be two key => value pairs.

:solver => value: value is a JuMP optimiser. The optimiser can be declared alongside its attributes by using JuMP.solver_with_attributes. Solver must support MOI.PowerCone, or JuMP must be able to transform it/them into a supported form.

:params => value: (optional) value must be a Dict or NamedTuple whose key => value pairs are the solver-specific settings.

alpha: significance level, alpha in (0, 1).

κ: relativistic deformation parameter.

PortfolioOptimiser.SD — Method

SD(w::AbstractVector, Σ::AbstractMatrix)

Compute the Standard Deviation. Square root of Variance.

\[\mathrm{SD}(\bm{w},\, \mathbf{\Sigma}) = \left[\bm{w}^\intercal \, \mathbf{\Sigma} \, \bm{w}\right]^{1/2}\,.\]

Inputs

w: vector of asset weights.

Σ: covariance matrix of asset returns.

PortfolioOptimiser.SKurt — Method

SKurt(x::AbstractVector)

Compute the square root semi-kurtosis.

\[\mathrm{SKurt}(\bm{x}) = \left[\dfrac{1}{T} \sum\limits_{t=1}^{T} \min\left( x_{t} - \mathbb{E}(\bm{x}),\, 0 \right)^{4} \right]^{1/2}\,.\]

Inputs

x: vector of portfolio returns.

PortfolioOptimiser.SLPM — Function

SLPM(x::AbstractVector; r::Real = 0.0)

Compute the Second Lower Partial Moment (Sortino Ratio).

\[\mathrm{SLPM}(\bm{x},\, r) = \left[\dfrac{1}{T-1} \sum\limits_{t=1}^{T}\max\left(r - \bm{x}_{t},\, 0\right)^{2}\right]^{1/2}\,``` # Inputs - `x`: vector of portfolio returns. - `r`: minimum return target.\]

PortfolioOptimiser.SSD — Method

SSD(x::AbstractVector)

Compute the mean Semi-Standard Deviation.

\[\mathrm{SSD}(\bm{x}) = \left[\dfrac{1}{T-1} \sum\limits_{t=1}^{T}\min\left(\bm{x}_{t} - \mathbb{E}(\bm{x}),\, 0\right)^{2}\right]^{1/2}\,.\]

Inputs

x: vector of portfolio returns.

PortfolioOptimiser.TG — Method

TG(x::AbstractVector; alpha_i::Real = 0.0001, alpha::Real = 0.05, a_sim::Int = 100)

Compute the Tail Gini.

Inputs

x: vector of portfolio returns.

alpha_i: start value of the significance level of CVaR losses, 0 <alpha_i < alpha < 1.

alpha: end value of the significance level of CVaR losses, alpha in (0, 1).

a_sim: number of CVaRs to approximate the Tail Gini losses, a_sim > 0.

PortfolioOptimiser.UCI_abs — Method

UCI_abs(x::AbstractVector)

Compute the Ulcer Index of uncompounded cumulative returns.

\[\mathrm{UCI_{a}}(\bm{x}) = \left[\dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{a}}(\bm{x},\, j)^{2}\right]^{1/2}\,,``` where ``\mathrm{DD_{a}}(\bm{x},\, j)`` is the Drawdown of uncompounded cumulative returns as defined in [`DaR_abs`](@ref). # Inputs - `x`: vector of portfolio returns.\]

PortfolioOptimiser.UCI_rel — Method

UCI_rel(x::AbstractVector)

Compute the Ulcer Index of compounded cumulative returns.

\[\mathrm{UCI_{r}}(\bm{x}) = \left[\dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{r}}(\bm{x},\, j)^{2}\right]^{1/2}\,,``` where ``\mathrm{DD_{r}}(\bm{x},\, j)`` is the Drawdown of compounded cumulative returns as defined in [`DaR_rel`](@ref). # Inputs - `x`: vector of portfolio returns.\]

PortfolioOptimiser.VaR — Function

VaR(x::AbstractVector; alpha::Real = 0.05)

Compute the Value at Risk, used in CVaR.

\[\mathrm{VaR}(\bm{x},\, \alpha) = -\underset{t \in (0,\, T)}{\inf} \left\{ x_{t} \in \mathbb{R} : F_{\bm{x}}(x_{t}) > \alpha \right\}\,,``` # Inputs - `x`: vector of portfolio returns. - `alpha`: significance level, alpha in (0, 1).\]

PortfolioOptimiser.Variance — Method

Variance(w::AbstractVector, Σ::AbstractMatrix)

Compute the Variance. Square of SD.

\[\mathrm{Variance}(\bm{w},\, \mathbf{\Sigma}) = \bm{w}^\intercal \, \mathbf{\Sigma}\, \bm{w}\,.\]

Inputs

w: vector of asset weights.

Σ: covariance matrix of asset returns.

PortfolioOptimiser.WR — Method

WR(x::AbstractVector)

Compute the Worst Realisation or Worst Case Scenario.

\[\mathrm{WR}(\bm{x}) = -\min(\bm{x})\,.\]

Inputs

x: vector of portfolio returns.

PortfolioOptimiser.calc_risk — Method

calc_risk(
    w::AbstractVector,    returns::AbstractMatrix;    rm::Symbol = :SD,    rf::Real = 0.0,    sigma::AbstractMatrix = Matrix{Float64}(undef, 0, 0),    alpha_i::Real = 0.0001,    alpha::Real = 0.05,    a_sim::Int = 100,    beta_i::Real = alpha_i,    beta::Real = alpha,    b_sim::Integer = a_sim,    kappa::Real = 0.3,    owa_w::AbstractVector{<:Real} = Vector{Float64}(undef, 0),    solvers::Union{<:AbstractDict, Nothing} = nothing)

Compute the value of a risk measure given a vector of asset weights and returns.

Inputs

w: vector of asset weights.

returns: matrix of asset returns where columns are assets and rows are timesteps.

rm: risk measure from RiskMeasures and HCRiskMeasures.

rf: risk-free rate at the frequency of returns, used as the minimum return target, r, in FLPM and SLPM.

sigma: covariance matrix of asset returns.

alpha_i: start value of the significance level of CVaR losses, 0 <alpha_i < alpha < 1.

alpha: alpha in (0, 1)

if rm is not an OWA range measure: significance level;

if rm is an OWA range measure: end value of the significance level of CVaR losses.

a_sim: number of CVaRs to approximate the Tail Gini losses, a_sim > 0.

beta_i: start value of the significance level of CVaR gains, 0 < beta_i < beta < 1.

beta: end value of the significance level of CVaR gains, beta in (0, 1).

b_sim: number of CVaRs to approximate the Tail Gini gains, b_sim > 0.

solvers: Provides the solvers and corresponding parameters for solving the JuMP model. There can be two key => value pairs.

:solver => value: value is a JuMP optimiser. The optimiser can be declared alongside its attributes by using JuMP.solver_with_attributes. Solver must support MOI.ExponentialCone and MOI.PowerCone respectively, or JuMP must be able to transform it/them into a supported form.

:params => value: (optional) value must be a Dict or NamedTuple whose key => value pairs are the solver-specific settings.

calc_risk(
    portfolio::AbstractPortfolio;    type::Symbol = isa(portfolio, Portfolio) ? :Trad : :HRP,    rm::Symbol = :SD,    rf::Real = 0.0)

Compute the value of a risk measure given a portfolio.

Inputs

portfolio: optimised portfolio.

type: type of portfolio from PortTypes or HCPortTypes.

rm: risk measure from RiskMeasures and HCRiskMeasures.

rf: risk-free rate at the frequency of portfolio.returns, used as the minimum return target, r, in FLPM and SLPM.

Private