Statistics Functions

Public

PortfolioOptimiser.asset_statistics! — Method

asset_statistics!(portfolio::AbstractPortfolio; calc_cov::Bool = true, calc_mu::Bool = true,
                  calc_kurt::Bool = true, calc_cor::Bool = true,
                  cov_opt::CovOpt = CovOpt(;), mu_opt::MuOpt = MuOpt(;),
                  kurt_opt::KurtOpt = KurtOpt(;), cor_opt::CorOpt = CorOpt(;))

Compute the asset statistics for a given portfolio in-place. See covar_mtx, mean_vec, cokurt_mtx, cor_dist_mtx.

Depending on the conditions, modifies:

portfolio.mu

portfolio.cov

portfolio.kurt

portfolio.skurt

portfolio.cor

portfolio.dist

mu_opt.method

Inputs

portfolio: instance of Portfolio or HCPortfolio.

Flags

calc_cor:

isa(portfolio, HCPortfolio):

true: compute and set the correlation and distance matrices.

calc_cov:

true: compute and set the covariance matrix.

calc_mu:

true: compute and set the expected returns vector.

mu_opt.method ∈ (:JS, :BS, :BOP, :CAPM): require the covariance matrix, so it will be computed and mu_opt.sigma will be set to the covariance matrix. portfolio.cov will only be modified when calc_cov == true.

calc_kurt:

true: compute and set the cokurtosis and semi cokurtosis matrices.

Options

cov_opt: instance of CovOpt, defines the parameters for computing the covariance matrix.

mu_opt: instance of MuOpt, defines the parameters for computing the expected returns vector.

kurt_opt: instance of KurtOpt, defines how the cokurtosis and semi cokurtoes are computed.

cor_opt: instance of CorOpt, defines how the correlation and distance matrices are computed.

PortfolioOptimiser.augmented_black_litterman — Method

augmented_black_litterman(returns::AbstractMatrix, w::AbstractVector;
                          F::Union{AbstractMatrix, Nothing}   = nothing,
                          B::Union{AbstractMatrix, Nothing}   = nothing,
                          P::Union{AbstractMatrix, Nothing}   = nothing,
                          P_f::Union{AbstractMatrix, Nothing} = nothing,
                          Q::Union{AbstractVector, Nothing}   = nothing,
                          Q_f::Union{AbstractVector, Nothing} = nothing,
                          cov_opt::CovOpt                     = CovOpt(;),
                          mu_opt::MuOpt                       = MuOpt(;),
                          f_cov_opt::CovOpt                   = CovOpt(;),
                          f_mu_opt::MuOpt                     = MuOpt(;),
                          bl_opt::BLOpt                       = BLOpt(;))

PortfolioOptimiser.bayesian_black_litterman — Method

bayesian_black_litterman(returns::AbstractMatrix, F::AbstractMatrix, B::AbstractMatrix,
                         P_f::AbstractMatrix, Q_f::AbstractVector;
                         cov_opt::CovOpt = CovOpt(;), mu_opt::MuOpt = MuOpt(;),
                         bl_opt::BLOpt = BLOpt(;))

PortfolioOptimiser.black_litterman — Method

black_litterman(returns::AbstractMatrix, P::AbstractMatrix, Q::AbstractVector,
                w::AbstractVector; cov_opt::CovOpt = CovOpt(;), mu_opt::MuOpt = MuOpt(;),
                bl_opt::BLOpt = BLOpt(;))

Estimates the expected returns vector and covariance matrix based on the Black Litterman model [54, 55]. See covar_mtx_mean_vec, covar_mtx, and mean_vec.

\[\begin{align*} \bm{\Pi} &= \begin{cases} \delta \mathbf{\Sigma} \bm{w} &\quad \mathrm{if~ eq = true}\\ \bm{\mu} - r &\quad \mathrm{if~ eq = false} \end{cases}\\ \mathbf{\Omega} &= \tau \mathrm{Diagonal}\left(\mathbf{P} \mathbf{\Sigma} \mathbf{P}^{\intercal}\right)\\ \mathbf{M} &= \left[ \left(\tau \mathbf{\Sigma} \right)^{-1} + \mathbf{P}^{\intercal} \mathbf{\Omega}^{-1} \mathbf{P}\right]^{-1}\\ \bm{\Pi}_{\mathrm{BL}} &= \mathbf{M} \left[\left(\tau \mathbf{\Sigma}\right)^{-1} \bm{\Pi} + \mathbf{P}^{\intercal} \mathbf{\Omega}^{-1} \mathbf{Q} \right]\\ \tau &= \dfrac{1}{T}\\ \bm{\mu}_{\mathrm{BL}} &= \bm{\Pi}_{\mathrm{BL}} + r\\ \mathbf{\Sigma}_{\mathrm{BL}} &= \mathbf{\Sigma} + \mathbf{M} \end{align*}\]

Where:

$\bm{\Pi}$: is the equilibrium excess returns.

$\delta$: is the risk aversion parameter.

$\mathbf{\Sigma}$: is the asset covariance matrix.

$\bm{w}$: is the vector of benchmark asset weights.

$\mathbf{P}$: is the asset views matrix.

$\bm{Q}$: is the asset views returns vector.

$\mathbf{\Omega}$: is the covariance matrix of the errors of the asset views.

$\mathbf{M}$: is an intermediate covariance matrix.

$\bm{\Pi}_{\mathbf{BL}}$: is the equilibrium excess returns after being adjusted by the views.

$T$: is the number of returns observations.

$\bm{\mu}_{\mathbf{BL}}$: is the vector of asset expected returns obtained via the Black Litterman model.

$\mathbf{\Sigma}_{\mathrm{BL}}$: is the asset covariance matrix obtained via the Black Litterman model.

Inputs

returns: T×N matrix of returns, where T is the number of returns observations, and N is the number of assets or factors.

P: Nv×Na analyst's views matrix, can be relative or absolute, where Nv is the number of views, and Na the number of assets.

Q: Nv×1 analyst's expected returns vector, where Nv is the number of views.

w: Na×1 benchmark weights vector, where Na is the number of assets.

Options

cov_opt: instance of CovOpt, defines the parameters for computing the covariance matrix.

mu_opt: instance of MuOpt, defines the parameters for computing the expected returns vector.

bl_opt: instance of BLOpt, defines the parameters for computing the Black Litterman model's statistics.

Outputs

mu: Na×1 Black Litterman adjusted expected returns vector, where Na is the number of assets.

cov_mtx: Na×Na Black Litterman adjusted covariance matrix, where Na is the number of assets.

w: Na×1 Black Litterman adjusted asset weights vector, where Na is the number of assets.

Note

Note that both bl_opt and mu_opt have rf fields for the risk-free rate (see MuOpt and BLOpt). This gives users more granular control over the model.

PortfolioOptimiser.black_litterman_factor_satistics! — Function

black_litterman_factor_satistics!(portfolio::Portfolio,
                                  w::AbstractVector                   = Vector{Float64}(undef, 0);
                                  B::Union{DataFrame, Nothing}        = nothing,
                                  P::Union{AbstractMatrix, Nothing}   = nothing,
                                  P_f::Union{AbstractMatrix, Nothing} = nothing,
                                  Q::Union{AbstractVector, Nothing}   = nothing,
                                  Q_f::Union{AbstractVector, Nothing} = nothing,
                                  loadings_opt::LoadingsOpt           = LoadingsOpt(;),
                                  cov_opt::CovOpt                     = CovOpt(;),
                                  mu_opt::MuOpt                       = MuOpt(;),
                                  f_cov_opt::CovOpt                   = CovOpt(;),
                                  f_mu_opt::MuOpt                     = MuOpt(;),
                                  bl_opt::BLOpt                       = BLOpt(;))

PortfolioOptimiser.black_litterman_statistics! — Function

black_litterman_statistics!(portfolio::Portfolio, P::AbstractMatrix, Q::AbstractVector;
                            w::AbstractVector = Vector{Float64}(undef, 0),
                            cov_opt::CovOpt = CovOpt(;), mu_opt::MuOpt = MuOpt(;),
                            bl_opt::BLOpt = BLOpt(;))

Estimates the Black Litterman statistics for a given portfolio in-place. See black_litterman, and covar_mtx_mean_vec.

Modifies:

portfolio.bl_mu

portfolio.bl_cov

Inputs

portfolio: instance of Portfolio.

P: Nv×Na analyst's views matrix, can be relative or absolute, where Nv is the number of views, and Na the number of assets.

Q: Nv×1 analyst's expected returns vector, where Nv is the number of views.

w: Na×1 benchmark weights vector, sets portfolio.bl_bench_weights, where Na is the number of assets.

isempty(w):

isempty(portfolio.bl_bench_weights): every entry is in portfolio.bl_bench_weights is assumed to be equal to 1/Na, w is set to be equal to portfolio.bl_bench_weights.

!isempty(portfolio.bl_bench_weights): w is set to be equal to portfolio.bl_bench_weights.

!isempty(w): portfolio.bl_bench_weights is set to be equal to w.

Options

cov_opt: instance of CovOpt, defines the parameters for computing the covariance matrix.

mu_opt: instance of MuOpt, defines the parameters for computing the expected returns vector.

bl_opt: instance of BLOpt, defines the parameters for computing the Black Litterman model's statistics.

isnothing(bl_opt.delta): modifies bl_opt by setting bl_opt.delta to:

\[\delta = \dfrac{\bm{\mu} \cdot \bm{w} - r}{\bm{w}^{\intercal} \mathbf{\Sigma} \bm{w}}\,.\]

Where:

$\delta$: is bl_opt.delta.

$\bm{\mu}$: is portfolio.mu.

$\bm{w}$: is portfolio.bl_bench_weights.

$r$: is bl_opt.rf.

$\mathbf{\Sigma}$: is portfolio.cov.

Note

Note that both bl_opt and mu_opt have rf fields for the risk-free rate (see MuOpt and BLOpt). This gives users more granular control over the model.

PortfolioOptimiser.cluster_assets — Function

cluster_assets

PortfolioOptimiser.cluster_assets! — Function

cluster_assets!

PortfolioOptimiser.cluster_assets — Method

cluster_assets

PortfolioOptimiser.cluster_assets — Method

cluster_assets

PortfolioOptimiser.cokurt_mtx — Function

cokurt_mtx

PortfolioOptimiser.cor_dist_mtx — Function

cor_dist_mtx(returns::AbstractMatrix, opt::CorOpt = CorOpt(;))

PortfolioOptimiser.covar_mtx — Function

covar_mtx(returns::AbstractMatrix, opt::CovOpt = CovOpt(;))

Compute the covariance matrix. See gerber0, gerber1, gerber2, sb0, sb1, gerbersb0, gerbersb1, _denoise_logo_mtx, posdef_fix!.

Inputs

returns: T×N matrix of returns, where T is the number of returns observations, and N is the number of assets or factors.

cov_opt: instance of CovOpt, defines the parameters for computing the covariance matrix.

Outputs

sigma: N×N covariance matrix, where N is the number of assets or factors.

PortfolioOptimiser.factor_statistics! — Method

factor_statistics!(portfolio::Portfolio; cov_f_opt::CovOpt = CovOpt(;),
                   mu_f_opt::MuOpt = MuOpt(;), cov_fm_opt::CovOpt = CovOpt(;),
                   mu_fm_opt::MuOpt = MuOpt(;), factor_opt::FactorOpt = FactorOpt(;))

Compute the factor and factor adjusted statistics for a given portfolio in-place. See covar_mtx_mean_vec, and risk_factors.

Modifies:

portfolio.f_cov

portfolio.f_mu

portfolio.fm_mu

portfolio.fm_cov

portfolio.fm_returns

portfolio.loadings

portfolio.loadings_opt

Inputs

portfolio: instance of Portfolio.

Options

cov_f_opt: instance of CovOpt, defines the parameters for computing the factor covariance matrix.

mu_f_opt: instance of MuOpt, defines the parameters for computing the factor expected returns vector.

cov_fm_opt: instance of CovOpt, defines the parameters for computing the factor adjusted covariance matrix.

mu_fm_opt: instance of MuOpt, defines the parameters for computing the factor adjusted expected returns vector.

factor_opt: instance of FactorOpt, defines how the factor statistics are computed.

PortfolioOptimiser.gerber0 — Function

gerber0(x::AbstractMatrix, opt::GerberOpt = GerberOpt(;))

PortfolioOptimiser.gerber1 — Function

gerber1(x::AbstractMatrix, opt::GerberOpt = GerberOpt(;))

PortfolioOptimiser.gerber2 — Function

gerber2(x::AbstractMatrix, opt::GerberOpt = GerberOpt(;))

PortfolioOptimiser.gerbersb0 — Function

gerbersb0(x::AbstractMatrix, gerberopt::GerberOpt = GerberOpt(;), sbopt::SBOpt = SBOpt(;))

PortfolioOptimiser.gerbersb1 — Function

gerbersb1(x::AbstractMatrix, gerberopt::GerberOpt = GerberOpt(;), sbopt::SBOpt = SBOpt

PortfolioOptimiser.loadings_matrix — Function

loadings_matrix(x::DataFrame, y::DataFrame, opt::LoadingsOpt = LoadingsOpt(;))

PortfolioOptimiser.mean_vec — Function

mean_vec(returns::AbstractMatrix, opt::MuOpt = MuOpt(;))

Compute the expected returns vector for a returns series.

Inputs

returns: T×N matrix of returns, where T is the number of returns observations, and N is the number of assets or factors.

mu_opt: instance of MuOpt, defines the parameters for computing the expected returns vector.

Outputs

mu: N×1 vector of expected returns, where N is the number of assets or factors.

PortfolioOptimiser.mu_estimator — Function

mu_estimator

PortfolioOptimiser.mut_var_info_mtx — Function

mut_var_info_mtx(x::AbstractMatrix{<:Real}, bins_info::Union{Symbol, <:Integer} = :KN,
                 normed::Bool = true)

Compute the mutual information and variation of information matrices.

Inputs

x: T×N array containing the returns series of the assets. T is the number of observations and N the number of assets.

bins_info: selection criterion for computing the number of bins used to calculate the mutual and variation of information statistics. Can take on an integer value or the following values:

An integer value explicitly defines the number of bins.

A choice of optimal bin width selection algorithms from BinMethods.

:KN: Knuth's choice.

:FD: Freedman-Diaconis' choice.

:SC: Schotts' choice.

:HGR: Hacine-Gharbi and Ravier's choice.

PortfolioOptimiser.posdef_fix! — Function

posdef_fix!(mtx::AbstractMatrix, opt::PosdefFixOpt = PosdefFixOpt(;), msg::String = "")

PortfolioOptimiser.risk_factors — Method

risk_factors(x::DataFrame, y::DataFrame; factor_opt::FactorOpt = FactorOpt(;),
             cov_opt::CovOpt = CovOpt(;), mu_opt::MuOpt = MuOpt(;))

Estimates the expected returns vector and covariance matrix of the assets based on the risk factor model [56, 57]. See loadings_matrix, covar_mtx_mean_vec, and posdef_fix!.

\[\begin{align*} \mathbf{X}_{\mathrm{FM}} &= \mathbf{F} \mathbf{B}^{\intercal}\\ \bm{\mu}_{\mathrm{FM}} &= \mathbf{B} \bm{\mu}_{F}\\ \bm{\Sigma}_{\mathrm{FM}} &= \mathbf{B} \mathbf{\Sigma}_{F} \mathbf{B}^{\intercal} + \mathbf{\Sigma}_{\epsilon}\\ \mathbf{\Sigma}_{\epsilon} &= \begin{cases}\mathrm{Diagonal}\left(\mathrm{var}\left(\mathbf{X} - \mathbf{X}_{\mathrm{FM}},\, \mathrm{dims} = 1\right)\right) &\quad \mathrm{if~ error = true}\\ \mathbf{0} &\quad \mathrm{if~ error = false} \end{cases} \end{align*}\]

Where:

$\mathbf{X}_{\mathrm{FM}}$: is the T×Na matrix of factor adjusted asset returns, where T is the number of returns observations, and Na the number of assets.

$\mathbf{F}$: is the T×Nf matrix of factor returns, where T is the number of returns observations, and Nf the number of assets.

$\mathbf{B}$: is the Na×Nf loadings matrix, where Nf is the number of assets.

$\bm{\mu}_{\mathrm{FM}}$: is the Na×1 adjusted expected returns vector of the assets, where Na is the number of assets.

$\bm{\mu}_{F}$: is the Nf×1 expected returns vector of the factors, where Nf is the number of factors.

$\mathbf{\Sigma}_{\mathrm{FM}}$: is the Na×Na factor adjusted covariance matrix of the assets, where Na is the number of assets.

PortfolioOptimiser.sb0 — Function

sb0(x::AbstractMatrix, gerberopt::GerberOpt = GerberOpt(;), sbopt::SBOpt = SBOpt(;))

PortfolioOptimiser.sb1 — Function

sb1(x::AbstractMatrix, gerberopt::GerberOpt = GerberOpt(;), sbopt::SBOpt = SBOpt(;))

PortfolioOptimiser.wc_statistics! — Function

wc_statistics!(portfolio::Portfolio, opt::WCOpt = WCOpt(;))

Worst case optimisation statistics.

Private

PortfolioOptimiser._denoise_logo_mtx — Function

_denoise_logo_mtx(T::Integer, N::Integer, mtx::AbstractMatrix,
                  opt::Union{CovOpt, CorOpt, KurtOpt, BLOpt}, mtx_name::Symbol = :cov)

PortfolioOptimiser._hierarchical_clustering — Function

_hierarchical_clustering

PortfolioOptimiser.covar_mtx_mean_vec — Method

covar_mtx_mean_vec(returns::AbstractMatrix, cov_opt::CovOpt = CovOpt(;),
                   mu_opt::MuOpt = MuOpt(;))

Compute the expected returns vector and covariance matrix. See covar_mtx and mean_vec.

Inputs

returns: T×N matrix of returns, where T is the number of returns observations, and N is the number of assets or factors.

cov_opt: instance of CovOpt, defines the parameters for computing the covariance matrix.

mu_opt: instance of MuOpt, defines the parameters for computing the expected returns vector.

Outputs

sigma: N×N covariance matrix, where N is the number of assets or factors.

mu: N×1 vector of expected returns, where N is the number of assets or factors.

PortfolioOptimiser.gen_bootstrap — Function

gen_bootstrap(returns::AbstractMatrix, kind::Symbol = :Stationary, n_sim::Integer = 3_000,
              block_size::Integer = 3, seed::Union{<:Integer, Nothing} = nothing,
              rng = Random.default_rng())