API

Objective functions

PortfolioOptimiser.AbstractPortfolioOptimiser — Type

Abstract type for subtyping concrete portfolio optimiser structures.

PortfolioOptimiser.port_variance — Function

port_variance(w, cov_mtx)

Computes the overall portfolio variance according to the weights w, and covariance matrix cov_mtx.

The portfolio variance $\sigma^2$, is defined as:

\[\sigma^2 = \bm{w}^T\, \Sigma\, \bm{w}\,,\]

where $\sigma$ is the overall portfolio standard deviation, $^T$ the transpose operator, $w$ the weights, and $\Sigma$ the covariance matrix.

PortfolioOptimiser.port_return — Function

port_return(w, mean_ret)

Computes the overall portfolio return according to the weights w, and mean returns mean_ret.

The portfolio return $\mu$, is defined as:

\[\mu = \bm{w} \cdot \bm{E},\]

where $\bm{w}$ are the weights, and $\bm{E}$ the mean returns.

PortfolioOptimiser.sharpe_ratio — Function

sharpe_ratio(w, mean_ret, cov_mtx, rf::Real = 0.02)

Computes the portfolio's Sharpe ratio given the weights w, mean returns mean_ret, covariance matrix cov_mtx, and risk free rate rf.

sharpe_ratio(μ, σ, rf::Real = 0.02)

Computes the Sharpe ratio given the return μ, standard deviation σ, and risk free rate rf.

The Sharpe ratio $s$, is defined as:

\[s = \dfrac{\mu - r}{\sigma},\]

where $\mu$ the portfolio's return (see port_return), and $\sigma$ the portfolio's standard deviation (see port_variance). Generally speaking, the greater the Sharpe ratio the better the portfolio.

Note

The Sharpe ratio penalises large swings in both directions, so assets that tend to have large increases in value are disproportionally penalised by this measure. The Sortino ratio has the same formula but uses an adjusted covariance matrix that accounts only for the negative fluctuations in value. SemiVar() and ExpSemiVar() in AbstractRiskModel implement this adjustment on the covariance matrix. The Mean-Semivariance portfolios in AbstractMeanSemiVar use it in their optimisations.

PortfolioOptimiser.L2_reg — Function

L2_reg(w, γ = 1)

L2 regularisation. Minimising this reduces the number of negligible weights w, with a tuning parameter γ. Increasing γ will decrease the number of negligible weights.

The $L_2$ regularisation is defined as:

\[L_2 = \gamma \Vert \bm{w} \Vert\,,\]

where $\gamma$ is the tuning parameter, and $w$ the weights.

PortfolioOptimiser.quadratic_utility — Function

quadratic_utility(w, mean_ret, cov_mtx, risk_aversion = 1)

Calculates the quadratic utility given the weights w, mean returns mean_ret, covariance matrix cov_mtx and risk aversion risk_aversion. Increasing risk_aversion decreases risk.

The quadratic utility $Q$, is defined as:

\[Q = \mu - \dfrac{1}{2} \delta \bm{w}^T\, \Sigma\, \bm{w}\,,\]

where $\mu$ is the overall portfolio expected return (see port_return), $\delta$ the risk aversion, $w$ the weights, and $\Sigma$ the covariance matrix.

PortfolioOptimiser.semi_ret — Function

semi_ret(returns, benchmark = 0)

Compute the historical semi-returns from historical returns, for a given benchmark. Return values greater than benchmark are "upside" returns, values lower than benchmark are "downside" returns. Each column of returns corresponds to an asset.

Historical semi returns $\mathrm{R_b}$, are defined as:

\[\mathrm{R_b} = \dfrac{\mathrm{R} - b}{\sqrt{N}}\,\]

where $\mathrm{R}$ are historical returns, $b$ is the benchmark, and $N$ is the number of historical entries (not the number of assets).

Note

The value of benchmark should correspond to the frequency of historical returns, i.e. if using daily returns, benchmark should be a reasonable value for daily returns.

PortfolioOptimiser.transaction_cost — Function

transaction_cost(w, w_prev, k = 0.001)

Simple transaction cost model. Sums all the absolute changes in weight and multiplies by a constant k. Simulates a fixed percentage commision from a broker.

The transacion cost $C$, is defined as:

\[C = k \lvert \bm{w} - \bm{w}_{\mathrm{prev}} \rvert\]

Warning

As of JuMP 1.0 there is no support for norm in objective functions. A model wishing to add this to their objective function should instead add it as a Vector Cone constraint, eg:

n = length(tickers)
prev_weights = fill(1 / n), n))
k = 0.001

extra_vars = [:(z[1:($n)])]
extra_constraints =
    [:([$k * (model[:w] - $prev_weights); model[:z]] in MOI.NormOneCone($(n + n)))]

ef = EfficientFrontier(tickers, mu, S;
        extra_vars = extra_vars,
        extra_constraints = extra_constraints,
    )

Similarly L2-norms must be turned into constraints of type MOI.NormTwoCone. More information on how to do this can be found in JuMP Vector Cones.

PortfolioOptimiser.ex_ante_tracking_error — Function

ex_ante_tracking_error(w, cov_mtx, w_bmk)

Compute the square of the ex-ante tracking error from the weights w, covariance matrix cov_mtx, and benchamark weights w_bmk.

The ex-ante tracking error $\mathrm{Err}$, is defined as:

\[\mathrm{Err} = (\bm{w} - \bm{w}_b)^T \, \Sigma \, (\bm{w} - \bm{w}_b)\,,\]

where $\bm{w}$ are the weights, $\bm{w}_b$ the benchmark weights, and $\Sigma$ the covariance matrix.

PortfolioOptimiser.ex_post_tracking_error — Function

ex_post_tracking_error(w, returns, ret_bmk)

Compute the square of the ex-post tracking error from the weights w, historical returns returns, and benchmark returns ret_bmk.

The ex-post tracking error $\mathrm{Err}$, is defined as:

\[\begin{aligned} \bm{x} &= \mathrm{R} \bm{w} - \bm{R}_b \\ \mu &= \dfrac{1}{N} \sum\limits_{i=1}^N x_i \\ \mathrm{Err} &= (\bm{x} - \mu) \cdot (\bm{x} - \mu) = Var(r - r_b)\,, \end{aligned}\]

where $w$ are the weights, $\mathrm{R}$ historical returns, $\bm{R}_b$ benchmark historical returns. In other words the ex-post tracking error is the variance of the historical returns $r$, minus the benchmark returns $r_b$, of the portfolio.

PortfolioOptimiser.logarithmic_barrier — Function

logarithmic_barrier(w, cov_mtx, k = 0.1)

Compute the logarithmic barrier adjusted by k, from the weights w, and covariance matrix cov_mtx.

The logarithmic barrier $L$, is defined as:

\[L = \bm{w}^T\, \Sigma\, \bm{w} - k \sum\limits_i \ln( w_i )\,,\]

where $\sigma^2 = \bm{w}^T\, \Sigma\, \bm{w}$ (see port_variance), $k$ the adjustment constant, and $w_i$ the weight of the $i$'th asset.

PortfolioOptimiser.kelly_objective — Function

kelly_objective(w, mean_ret, cov_mtx, k = 3)

Compute a Kelly objective from adjusted by k, from the weights w, mean returns mean_ret, and covariance matrix cov_mtx.

The Kelly objective $K$, is defined as:

\[K = \bm{w}^T\, \Sigma\, \bm{w} - \dfrac{1}{2} k \bm{w} \cdot \bm{E} \,,\]

where $\sigma^2 = \bm{w}^T\, \Sigma\, \bm{w}$ (see port_variance), $k$ is the adjustment constant, $\bm{w}$ the weights, and $\bm{E}$ the mean returns.

Expected returns

Risk models

Efficient Frontier

Mean-var

Mean-semivar

Critical Value at Risk (CVaR)

Critical Drawdown at Risk (CDaR)

Hierarchical portfolio

Black-Litterman

Critical line algorithm

Asset allocation

PortfolioOptimiser.AbstractAllocation — Type

abstract type AbstractAllocation end
struct Lazy <: AbstractAllocation end
struct Greedy <: AbstractAllocation end
struct LP <: AbstractAllocation end

Types for allocation dispatch.

Lazy: Lazy allocation algorithm. Can do fractional shares.

Greedy: Greedy allocation algorithm. Can do fractional shares.

LP: Mixed linear integer programming algorithm. Can only handle full shares.

PortfolioOptimiser.Allocation — Type

struct Allocation{T1,T2,T3}
    tickers::T1
    weights::T2
    shares::T3
end

Concrete allocation created by Allocation.

tickers: list of tickers.

weights: list of corresponding weights according to share allocation and latest prices.

shares: list of corresponding shares according to their optimal allocation.

PortfolioOptimiser.Allocation — Method

Allocation(
    type::AbstractAllocation,
    portfolio::AbstractPortfolioOptimiser,
    latest_prices::AbstractVector;
    investment = 1e4,
    rounding = 1,
    reinvest = false,
    short_ratio = nothing,
    optimiser = HiGHS.Optimizer,
    silent = true,
)

Returns a tuple of an Allocation structure and remaining money after allocating assets.

Note

Short positions have negative weights and shares.

type: Allocation algorithm to use, can be Lazy(), Greedy() or LP() (see AbstractAllocation).

portfolio: Portfolio AbstractPortfolioOptimiser.

latest_prices: Vector of latest prices, entries should be in the same order as portfolio.tickers.

investment: value to be invested in the portfolio.

rounding: Greedy() and Lazy() support fractional shares, rounds shares down to the nearest multiple of rounding. This has no effect on an LP() allocation.

reinvest: if true, reinvests the money earned from short positions of the portfolio into buying more long positions.

short_ratio: long to short portfolio ratio, 0.2 corresponds to 120/20 long to short shares. If nothing defaults to portfolio weights.

optimiser: only used for the LP() allocation. Needs to support mixed-interger linear programming.

silent: if true the optimiser is silent.

Example

ef = EfficientFrontier(
    tickers,
    mean_ret,
    cov_mtx;
    market_neutral = true,
    weight_bounds = (-1, 1),
)
max_quadratic_utility!(ef)

# Mixed integer, programming, no rounding available.
allocLP, leftoverLP = Allocation(LP(), ef, latest_prices; investment = 3590)

# Greedy allocation up to quarter shares.
allocGreedy, leftoverGreedy = Allocation(Greedy(), ef, latest_prices; rounding = 0.25)

# Lazy allocation with a 110/10 long-short ratio and reinvesting 
# earnings from selling the short positions into buying long positions.
allocLazy, leftoverLazy = Allocation(
                            Lazy(), ef, latest_prices; 
                            short_ratio = 0.1, reinvest = true
                        )

PortfolioOptimiser.roundmult — Function

roundmult(val, prec [, args...] [; kwargs...])

Round a number to a multiple of prec. Uses the same defaults and has the same args and kwargs of the built-in Base.round.

Equivalent to:

round(div(val, prec) * prec, args...; kwargs...)