Asset allocation

PortfolioOptimiser.AbstractAllocation — Type

abstract type AbstractAllocation end
struct Greedy <: AbstractAllocation end
struct LP <: AbstractAllocation end

Types for allocation dispatch.

Greedy: Greedy allocation algorithm. Can do fractional shares.

LP: Mixed linear integer programming algorithm. Can only handle full shares.

PortfolioOptimiser.Allocation — Type

struct Allocation{T1,T2,T3}
    tickers::T1
    shares::T2
    weights::T3
end

Concrete allocation created by Allocation.

tickers: list of tickers.

shares: list of corresponding shares according to their optimal allocation.

weights: list of corresponding weights according to share allocation and latest prices.

PortfolioOptimiser.Allocation — Method

Allocation(
    type::AbstractAllocation,
    portfolio::AbstractPortfolioOptimiser,
    latest_prices::AbstractVector;
    investment = 1e4,
    rounding = 1,
    reinvest = false,
    short_ratio = nothing,
    optimiser = HiGHS.Optimizer,
    silent = true,
)

Returns a tuple of an Allocation structure and remaining money after allocating assets.

Note

Short positions have negative weights and shares.

type: Allocation algorithm to use, can be Greedy() or LP() (see AbstractAllocation).

portfolio: Portfolio AbstractPortfolioOptimiser.

latest_prices: Vector of latest prices, entries should be in the same order as portfolio.tickers.

investment: value to be invested in the portfolio.

rounding: Greedy() supports fractional shares, rounds shares down to the nearest multiple of rounding. This has no effect on an LP() allocation.

reinvest: if true, reinvests the money earned from short positions of the portfolio into buying more long positions.

short_ratio: long to short portfolio ratio, 0.2 corresponds to 120/20 long to short shares. If nothing defaults to portfolio weights.

optimiser: only used for the LP() allocation. Needs to support mixed-interger linear programming.

silent: if true the optimiser is silent.

Dispatches one of the following depending on type:

Greedy allocation

Allocation(
    type::Greedy,
    tickers::AbstractArray,
    weights::AbstractArray,
    latest_prices::AbstractVector,
    investment::Real = 1e4,
    rounding::Real = 1,
    reinvest::Bool = false,
    short_ratio::Union{Real, Nothing} = nothing,
    optimiser = nothing,
    silent = nothing,
)

Greedy algorithm that allocates assets according to their optimised weights. It priorites assets according to the ones with the greatest differnce between their weight in the portfolio and ideal weight. It allocates assets by progressively adding rounding to each share until the cost exceeds the available funds.

Note

optimiser and silent are not used but are in the function signature for dispatch purposes.

LP allocation

Allocation(
    type::LP,
    tickers::AbstractArray,
    weights::AbstractArray,
    latest_prices::AbstractVector,
    investment::Real = 1e4,
    rounding = nothing,
    reinvest::Bool = false,
    short_ratio::Union{Real, Nothing} = nothing,
    optimiser = HiGHS.Optimizer,
    silent = true,
)

Linear mixed-integer optimisation of assets.

Note

rounding has no effect, as it can only allocate integer numbers of shares.

Example

ef = EffMeanVar(
    tickers,
    mean_ret,
    cov_mtx;
    market_neutral = true,
    weight_bounds = (-1, 1),
)
max_utility!(ef)

# Mixed integer, programming, no rounding available.
allocLP, leftoverLP = Allocation(LP(), ef, latest_prices; investment = 3590)

# Greedy allocation up to quarter shares.
allocGreedy, leftoverGreedy = Allocation(Greedy(), ef, latest_prices; rounding = 0.25)

PortfolioOptimiser.roundmult — Function

roundmult(val, prec [, args...] [; kwargs...])

Round a number to a multiple of prec. Uses the same defaults and has the same args and kwargs of the built-in Base.round.

Equivalent to:

round(div(val, prec) * prec, args...; kwargs...)

PortfolioOptimiser._short_allocation — Function

_short_allocation(
    type,
    tickers,
    weights,
    latest_prices,
    investment,
    rounding,
    reinvest,
    short_ratio,
    optimiser,
    silent
)

Helper function for allocating short-long portfolios. Ensures the shares and weights of the shorts are negative.

PortfolioOptimiser._sub_allocation — Function

_sub_allocation(
    type,
    tickers,
    weights,
    latest_prices,
    investment,
    rounding,
    idx,
    reinvest,
    short_ratio,
    optimiser,
    silent,
)

Helper function for creating sub allocations. It calls Allocation again but only for the idx provided.

PortfolioOptimiser._clean_zero_shares — Function

_clean_zero_shares(shares_bought, tickers, latest_prices)

Helper function for removing tickers with zero shares.