Base Optimiser

PortfolioOptimiser.AbstractPortfolioOptimiser — Type

abstract type AbstractPortfolioOptimiser end

Abstract type for subtyping concrete portfolio optimiser structures.

PortfolioOptimiser.custom_optimiser! — Function

custom_optimiser!(
    portfolio::AbstractPortfolioOptimiser,
    obj,
    obj_params = ();
    initial_guess = nothing,
    optimiser = Ipopt.Optimizer,
    silent = true,
    optimiser_attributes = (),
)

Minimises user-provided objectives that are supported by JuMP.@objective.

portfolio: any concrete subtype of AbstractPortfolioOptimiser.

obj: objective function supported by JuMP.@objective.

obj_params: vector or tuple of arguments of obj.

initial_guess: initial guess for optimiser, if nothing defaults to uniform weights.

optimiser: optimiser for solving optimisation problem.

silent: if false, the optimiser prints to console.

Example

This example defines a new function, identical to kelly_objective, but you can use the predefined functions in PortfolioOptimiser as long as they are supported by JuMP.@objective.

function kelly_objective2(w, mean_ret, cov_mtx, k = 3)
    variance = dot(w, cov_mtx, w)
    objective = variance * 0.5 * k - dot(w, mean_ret)
    return objective
end

ef = MeanVar(tickers, expected_returns, cov_mtx)
obj_params = (ef.mean_ret, ef.cov_mtx, 1000)
custom_optimiser!(ef, kelly_objective2, obj_params)

Unlike, custom_nloptimiser!, functions defined by PortfolioOptimiser don't have to be prepended with PortfolioOptimiser, they can be used as normal. Here we use kelly_objective, which is provided by the package and as previously mentioned is identical to kelly_objective2 of the previous example.

ef = MeanVar(tickers, expected_returns, cov_mtx)
obj_params = (ef.mean_ret, ef.cov_mtx, 1000)
custom_optimiser!(ef, kelly_objective, obj_params)

Note

obj_params can be any variable that can be splatted. It is also optional, so objectives with no parameters are valid too.

Warning

This minimises obj, so if you want to maximise a function, make it negative–-for example, the portfolio return. Furthermore, this does not add extra objective terms, so they must be added to the definition of obj. We illustrate how we can use this to maximise the return subject to L2 regularisation (don't do this, 'tis a silly idea).

function max_ret_l2_reg(w, mean_ret, γ = 1)
    μ = port_return(w, mean_ret)
    l2 = L2_reg(w, γ)

    # Minimise L2 regularisation, maximise the return by minimising its negative.
    return l2 - μ
end

ef = MeanVar(tickers, mean_ret, cov_mtx)

γ = 1
obj_params = (mean_ret, γ)

custom_optimiser!(ef, max_ret_l2_reg, obj_params)

PortfolioOptimiser.custom_nloptimiser! — Function

custom_nloptimiser!(
    portfolio::AbstractPortfolioOptimiser,
    obj,
    obj_params = (),
    extra_vars = ();
    initial_guess = nothing,
    optimiser = Ipopt.Optimizer,
    silent = true,
    optimiser_attributes = (),
)

Minimises user-provided nonlinear objectives that are supported by JuMP.@NLobjective.

portfolio: any concrete subtype of AbstractPortfolioOptimiser.

obj: objective function supported by JuMP.@NLobjective.

obj_params: vector or tuple of arguments of obj.

extra_vars: collection of tuples of extra variables for the nonlinear optimiser and their starting values, each element takes the form of (variable, value). If !isnothing(value), it sets/overrides the start value of variable, else it takes the default. This is important because the optimiser can fail if the start value causes a discontinuity in the obj function, or one of its derivatives. Furthermore, variable must be a variable registered to the model.

initial_guess: initial guess for optimiser, if nothing defaults to unifrom weights.

optimiser: optimiser for solving optimisation problem.

silent: if false, the optimiser prints to console.

Warning

As of JuMP 1.0, JuMP.@NLobjective only supports scalar arguments, as such, it's important to define the nonlinear objective to take in scalar arguments only. The example shows how this can be done.

Example

In this first example we define a logarithmic barrier function identical to logarithmic_barrier, then we have to define the scalar-only function that JuMP.@NLobjective can use.

# Import logarithmic_barrier so we can add the scalar method.
import PortfolioOptimiser: logarithmic_barrier

function logarithmic_barrier2(w, cov_mtx, k = 0.1)
    # Add eps() to avoid log(0) divergence.
    log_sum = sum(log.(w .+ eps()))
    var = dot(w, cov_mtx, w)
    return var - k * log_sum
end

function logarithmic_barrier(w::T...) where {T}
    cov_mtx = obj_params[1]
    k = obj_params[2]
    w = [i for i in w]
    logarithmic_barrier2(w, cov_mtx, k)
end

ef = MeanVar(tickers, mean_ret, cov_mtx)
obj_params = (ef.cov_mtx, 0.001)
custom_nloptimiser!(ef, logarithmic_barrier, obj_params)

We can also use objective functions that are already defined by PortfolioOptimiser. However, given how JuMP.@NLobjective requires user-defined functions to be registered, we need to prepend them with PortfolioOptimiser so that it can be recognised and registered by JuMP into the model.

# Import logarithmic_barrier so we can add the scalar method.
import PortfolioOptimiser: logarithmic_barrier

function logarithmic_barrier(w::T...) where {T}
    cov_mtx = obj_params[1]
    k = obj_params[2]
    w = [i for i in w]
    PortfolioOptimiser.logarithmic_barrier(w, cov_mtx, k)
end

ef = MeanVar(tickers, mean_ret, cov_mtx)
obj_params = [ef.cov_mtx, 0.001]
custom_nloptimiser!(ef, logarithmic_barrier, obj_params)

# This definition will error in custom_nloptimiser!
function logarithmic_barrier(w::T...) where {T}
    cov_mtx = obj_params[1]
    k = obj_params[2]
    w = [i for i in w]
    logarithmic_barrier(w, cov_mtx, k)
end

ef = MeanVar(tickers, mean_ret, cov_mtx)
obj_params = [ef.cov_mtx, 0.001]
custom_nloptimiser!(ef, logarithmic_barrier, obj_params)

Warning

Note how in both cases, the parameters in the scalar-only function are not declared in the function signature. As such, they must be defined in terms of obj_params, or as literals, as the variable must be known to custom_nloptimiser!. If a variable name other than obj_params is used, the model will error becuase the variable will be unkown to the custom_nloptimiser!. For example, the following custom nonlinear objective function will not work because obj_param and obj_parameter do not exist inside custom_nloptimiser!.

# Import logarithmic_barrier so we can add the scalar method.
import PortfolioOptimiser: logarithmic_barrier

function logarithmic_barrier(w::T...) where {T}
    cov_mtx = obj_param[1]  # Unkown variable, change to obj_params[1]
    k = obj_parameter[2]    # Unkown variable, change to obj_params[2]
    w = [i for i in w]
    PortfolioOptimiser.logarithmic_barrier(w, cov_mtx, k)
end

ef = MeanVar(tickers, mean_ret, cov_mtx)
obj_params = [ef.cov_mtx, 0.001]
custom_nloptimiser!(ef, logarithmic_barrier, obj_params)

Note

obj_params can be any variable that can be splatted. It is also optional, so nonlinear objectives with no parameters are valid too.

Warning

This minimises obj, so if you want to maximise a function, make it negative–-for example, the sharpe ratio. Furthermore, this does not add extra objective terms, so they must be added to the definition of obj. We illustrate how we can use this to maximise the return subject to L2 regularisation (don't do this, 'tis a silly idea).

function sharpe_l2_reg(w::T...) where {T}
    mean_ret = obj_params[1]
    cov_mtx = obj_params[2]
    rf = obj_params[3]
    γ = obj_params[4]

    w = [i for i in w]
    sr = PortfolioOptimiser.sharpe_ratio(w, mean_ret, cov_mtx, rf)
    l2 = PortfolioOptimiser.L2_reg(w, γ)

    # L2 reg has to be minimised. To maximise the sharpe ratio we minimise its negative.
    return l2 - sr
end

ef = MeanVar(tickers, mean_ret, cov_mtx)

γ = 1
obj_params = (mean_ret, cov_mtx, ef.rf, γ)

custom_nloptimiser!(ef, sharpe_l2_reg, obj_params)

PortfolioOptimiser.add_sector_constraint! — Function

add_sector_constraint!(
    portfolio::AbstractPortfolioOptimiser,
    sector_map,
    sector_lower,
    sector_upper,
)

Constrains the sum of the weights of the tickers belonging to a sector to be between that sector's lower and upper bounds. For example, if we have two technology tickers, and we define the lower and upper bounds of the technology sector to be (0.1, 0.3), the sum of the weights of those tickers will be between (0.1, 0.3).

sector_map: a dictionary mapping tickers to sectors where they key is the ticker and the value the sector, Dict(<ticker> => <sector>).

sector_lower: a dictionary mapping a sector to its lower bound, Dict(<sector> => sector_lower_bound).

sector_upper: a dictionary mapping a sector to its upper bound, Dict(<sector> => sector_upper_bound).

PortfolioOptimiser._refresh_model! — Function

_refresh_model!(default_keys, model)

Helper function for cleaning up JuMP.Model() objects.

default_keys: collection of keys that are to be kept in the model.

model: JuMP.Model() object to be cleaned up.

PortfolioOptimiser._create_weight_bounds — Function

_create_weight_bounds(num_tickers, bounds)

Create the upper and lower bounds for asset weights.

num_tickers: The number of tickers.

bounds: weight bounds.

If bounds is a collection of length two, it assumes the first entry bounds[1] are the lower bounds for all weights, and bounds[2] is the upper bounds for all weights.

If bounds is an abstract vector of length equal to num_tickers, and each element of the vector is a tuple of length 2. Then each element corresponds to an asset, the first element of the tuple is the lower bound, the second the upper bound for said asset.

Else throw an error.

It returns a tuple of:

(lower_bounds, upper_bounds): where the length of lower_bounds and upper_bounds is equal to num_tickers. These are then used by AbstractEfficient portfolios to set the bounds constraints for their JuMP.Model().

PortfolioOptimiser._make_weight_sum_constraint! — Function

_make_weight_sum_constraint!(model, market_neutral)

Create the weight sum constraints for a JuMP.Model().

model: a JuMP.Model() of a portfolio optimiser.

market_neutral: if true the portfolio is market neutral and the weights will be constrained to sum to 0. If false, the portfolio is not market neutral and the weights will be constrained to sum to 1.

If market_neutral is true and all lower_bounds constraints are greater than or equal to zero, then a portfolio cannot be market neutral. If this is the case, the lower_bounds constraint for all assets will be changed to be greater than or equal to -1.

PortfolioOptimiser._add_var_to_model! — Function

_add_var_to_model!(model, var, args...)

Function to add a JuMP.@variable to a JuMP.Model().

model: a JuMP.Model().

var: symbol or expression representing a JuMP.@variable. Extra array variables have to be expressions. Extra scalar variables can be symbols or expressions.

args...: expression representing the args... to a JuMP.@variable.

Variables can also be directly added to a model using JuMP.@variable. This is used internally by AbstractEfficient portfolios, so that when a portfolio optimiser is refreshed, the extra variables can be added back automatically.

Example

We can use this to add a vector variable c of length 3, a scalar variable k, and a 3 x 5 x 4 multidimensional array variable m.

extra_vars = [ :( c[1:3] ), :k, :( m[1:3, 1:5, 1:4] ) ]
ef = MeanVar(tickers, mean_ret, cov_mtx)
_add_var_to_model!.(ef.model, extra_vars)

Alternatively, we can make sure they get registered in the optimiser. If the portfolio optimiser is refreshed, the extra variables are added back to the model automatically. They'd have to be added back manually otherwise.

extra_vars = [ :( c[1:3] ), :k, :( m[1:3, 1:5, 1:4] ) ]
ef = MeanVar(tickers, mean_ret, cov_mtx; extra_vars = extra_vars)

Note

When adding variables to the constructor, they must be a collection (Tuple or vector), even if it's of length 1.

PortfolioOptimiser._add_constraint_to_model! — Function

_add_constraint_to_model!(model, key, constraint)

Function to add a JuMP.@constraint to a JuMP.Model().

model: a JuMP.Model().

key: the key under which the constraint will be registered in the JuMP.Model().

constraint: expression representing the constraint.

Constraints can also be directly added to a model using JuMP.@constraint. This is used internally by AbstractEfficient portfolios, so that when a portfolio optimiser is refreshed, the extra constraints can be added back automatically.

Warning

Variables used within the constraint must be registered in the JuMP.Model() object (hence _add_var_to_model!), and referred to as model[<key>], where <key> is the key of the variable. For example, when adding a constraint such that the first weight is less than or equal to 0.2, :(model[:w][1] <= 0.2).

Warning

Constraints must be expressions :(), not quotes quote ... end, because quotes can contain be multiple expressions.

Note

This uses JuMP.@constraint internally, so any constraint given must be compatible with JuMP.@constraint.

Example

We can use this to add a constraint, :const1 such that the weight of the second asset is greater than or equal to 2, and :const2 such that the sum of all weights is less than or equal to 5. These are nonsensical constraints, but they illustrate the point. A constraint will work as long as it is a valid JuMP.@constraint. For a more complex example of a constraint see transaction_cost.

extra_constraints = [:(model[:w][2] >= 2), :(sum(model[:w]) <= 5)]
ef = MeanVar(tickers, mean_ret, cov_mtx)
_add_constraint_to_model!.(ef.model, [:const1, :const2], extra_constraints)

We can also use variable interpolation into the expressions.

ef = MeanVar(tickers, mean_ret, cov_mtx)
w = ef.model[:w]
extra_constraints = [:($(w[2]) >= 2), :(sum($w) <= 5)]
_add_constraint_to_model!.(ef.model, [:const1, :const2], extra_constraints)

If we want the extra constraints to be added back automatically when refreshing a portfolio optimiser, we can add them to the constructor.

extra_constraints = [:(model[:w][2] >= 2), :(sum(model[:w]) <= 5)]
ef = MeanVar(tickers, mean_ret, cov_mtx; extra_constraints = extra_constraints)

Note

When adding constraints to the constructor, they must be a collection (Tuple or vector), even if it's of length 1.

PortfolioOptimiser._add_to_objective! — Function

_add_to_objective!(model, expr)

Function to add a term to a JuMP.@objective to a JuMP.Model().

model: a JuMP.Model().

expr: the expression to add to the objective function.

Warning

Variables used within the extra objective must be registered in the JuMP.Model() object (hence _add_var_to_model!), and referred to as model[<key>], where <key> is the key of the variable. For example, quote model[:w][1] * 5 - sum(model[:w]) end or :( model[:w][1] * 5 - sum(model[:w]) ). Extra objective terms can be quotes or expressions. Using quotes is recommended because extra objective terms can be functions, as shown in the example.

Note

Can only be used for JuMP.@objective. For problems which require a non-linear objective, use custom_nloptimiser! and add the extra terms to the custom function.

Example

We can add extra objective terms to be minimised in the following way. We can use functions defined by PortfolioOptimiser, user defined functions, or anything else that is a valid JuMP.@objective. You can also use variable interpolation, but we don't use it in this example.

extra_obj_terms = [
    quote
        model[:w][1] * 5 - sum(model[:w])
    end,
    quote
       L2_reg(model[:w], 0.05)
    end
]
ef = MeanVar(tickers, mean_ret, cov_mtx)
_add_to_objective!.(ef.model, extra_obj_terms)

They can also be added to the constructor so they are added back automatically when the portfolio optimiser is refreshed.

extra_obj_terms = [
    quote
        model[:w][1] * 5 - sum(model[:w])
    end,
    quote
       L2_reg(model[:w], 0.05)
    end
]
ef = MeanVar(tickers, mean_ret, cov_mtx; extra_obj_terms = extra_obj_terms)

Note

When adding elements to an objective in the constructor, they must be a collection (Tuple or vector), even if it's of length 1.