Efficient Frontier

PortfolioOptimiser.AbstractEfficient — Type

abstract type AbstractEfficient <: AbstractPortfolioOptimiser end

Abstract type for subtyping efficient frontier optimisers.

PortfolioOptimiser._refresh_add_var_and_constraints — Function

_refresh_add_var_and_constraints(default_keys, portfolio)

Helper function for refreshing portfolios and re-adding extra variables and constraints used to build the model via the constructor.

PortfolioOptimiser._function_vs_portfolio_val_warn — Function

_function_vs_portfolio_val_warn(fval, pval, name)

Helper function for throwing generic warnings about inconsistent values between calls to optimiser functions and the values registered in the portfolio.

PortfolioOptimiser._val_compare_benchmark — Function

_val_compare_benchmark(val, op, benchmark, correction, name)

Helper function for throwing generic warnings about values out of the domain and ammending those values.

Mean-variance

PortfolioOptimiser.AbstractEfficientMeanVar — Type

abstract type AbstractEfficientMeanVar <: AbstractEfficient end

Abstract type for subtyping efficient mean variance optimisers.

PortfolioOptimiser.EfficientMeanVar — Type

struct EfficientMeanVar{T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13} <:
       AbstractEfficientMeanVar
    tickers::T1
    mean_ret::T2
    weights::T3
    cov_mtx::T4
    rf::T5
    market_neutral::T6
    risk_aversion::T7
    target_volatility::T8
    target_ret::T9
    extra_vars::T10
    extra_constraints::T11
    extra_obj_terms::T12
    model::T13
end

Structure for a mean-variance portfolio.

tickers: list of tickers.

mean_ret: mean returns, don't need it to optimise for minimum variance.

weights: weight of each ticker in the portfolio.

cov_mtx: covariance matrix.

rf: risk free return.

market_neutral: whether a portfolio is market neutral or not. Used in max_quadratic_utility!, efficient_risk!, efficient_return!.

risk_aversion: risk aversion parameter. Used in max_quadratic_utility!.

target_volatility: target volatility parameter. Used in efficient_risk!.

target_ret: target return parameter. Used in efficient_return!.

extra_vars: extra variables for the model.

extra_constraints: extra constraints for the model.

extra_obj_terms: extra objective terms for the model.

model: model for optimising portfolio.

PortfolioOptimiser.EfficientMeanVar — Method

EfficientMeanVar(
    tickers,
    mean_ret,
    cov_mtx;
    weight_bounds = (0, 1),
    rf = 0.02,
    market_neutral = false,
    risk_aversion = 1.0,
    target_volatility = rank(cov_mtx) < size(cov_mtx, 1) ? 1 / sum(diag(cov_mtx)) :
                        sqrt(1 / sum(inv(cov_mtx))),
    target_ret = mean(mean_ret),
    extra_vars = [],
    extra_constraints = [],
    extra_obj_terms = [],
)

Function for creating a JuMP model for creating an efficient mean-var portfolio to be optimised.

tickers: list of tickers.

mean_ret: mean returns, don't need it to optimise for minimum variance.

weights: weight of each ticker in the portfolio.

cov_mtx: covariance matrix.

rf: risk free return.

market_neutral: whether a portfolio is market neutral or not. Used in max_quadratic_utility!, efficient_risk!, efficient_return!.

risk_aversion: risk aversion parameter. Used in max_quadratic_utility!.

target_volatility: target volatility parameter. Used in efficient_risk!.

target_ret: target return parameter. Used in efficient_return!.

extra_vars: extra variables for the model. See _add_var_to_model! for details on how to use this.

extra_constraints: extra constraints for the model. See _add_constraint_to_model! for details on how to use this.

extra_obj_terms: extra objective terms for the model. See _add_to_objective! for details on how to use this.

Mean-semivariance

PortfolioOptimiser.EfficientSemiVar — Type

struct EfficientSemiVar{
    T1,
    T2,
    T3,
    T4,
    T5,
    T6,
    T7,
    T8,
    T9,
    T10,
    T11,
    T12,
    T13,
    T14,
    T15,
} <: AbstractEfficientSemiVar
    tickers::T1
    mean_ret::T2
    weights::T3
    returns::T4
    freq::T5
    benchmark::T6
    rf::T7
    market_neutral::T8
    risk_aversion::T9
    target_semidev::T10
    target_ret::T11
    extra_vars::T12
    extra_constraints::T13
    extra_obj_terms::T14
    model::T15
end

Critical Value at Risk (CVaR)

Critical Drawdown at Risk (CDaR)