Expected Returns

PortfolioOptimiser.AbstractReturnModel — Type

abstract type AbstractReturnModel end
struct MRet <: AbstractReturnModel end
struct EMRet <: AbstractReturnModel end
struct CAPMRet <: AbstractReturnModel end
struct ECAPMRet <: AbstractReturnModel end

Return types, used to dispatch on ret_model.

MRet: Mean returns.

EMRet: Exponentially weighted mean returns.

CAPMRet: Capital Asset Pricing Model (CAPM) returns.

ECAPMRet: Exponentially weighted Capital Asset Pricing Model (ECAPM) returns.

PortfolioOptimiser.ret_model — Function

Compute the mean returns given a concrete type of AbstractReturnModel, and returns array where each column is an asset and each row is an entry.

For all methods:

compound: if true the compute compound mean returns, if false compute uncompounded mean returns.

Mean

Arithmetic mean

ret_model(::MRet, returns; compound = true)

Compute the mean returns.

Exponentially weighted arithmetic mean

ret_model(::EMRet, returns; compound = true, span = Int(ceil(4*size(returns, 1) / log(size(returns, 1) + 2))))

Compute the exponentially weighted areithmetic mean returns. More recent returns are weighted more heavily.

span: defines the weighing parameter, α = 2 / (span + 1).

Capital Asset Price Model (CAPM) returns

Compute the Capital Asset Price Model (CAPM) returns.

If the market returns are not provided, it estimates them by averaging across all assets and uses those as the market returns. These methods require computing the covariance between asset and market returns. These methods give a fairer value to a stock returns according to how they relate to the market.

cov_type: a concrete type of AbstractRiskModel for computing the covariance of the assets to the market.

CAPM with arithmetic mean

ret_model(
    ::CAPMRet,
    returns,
    market_returns = nothing;
    rf = 1.02^(1 / 252) - 1,
    compound = true,
    cov_type::AbstractRiskModel = Cov(),
    target = 1.02^(1 / 252) - 1,
    fix_method::AbstractFixPosDef = SFix(),
    span = Int(ceil(4*size(returns, 1) / log(size(returns, 1) + 2))),
    custom_cov_estimator = nothing,
    custom_cov_args = (),
    custom_cov_kwargs = (),
    )

Capital Asset Pricing Model (CAPM) returns. CAPMRet() uses the normal mean to estimate the mean market returns.

target, fix_method, and span correspond to the same keyword arguments of cov for computing the covariance specified by cov_type.

CAPM with exponentially weighted arithmetic mean

ret_model(
    ::ECAPMRet,
    returns,
    market_returns = nothing;
    rf = 1.02^(1 / 252) - 1,
    compound = true,
    rspan = Int(ceil(4*size(returns, 1) / log(size(returns, 1) + 2))),
    cov_type::AbstractRiskModel = ECov(),
    target = 1.02^(1 / 252) - 1,
    fix_method::AbstractFixPosDef = SFix(),
    cspan = Int(ceil(4*size(returns, 1) / log(size(returns, 1) + 2))),
    custom_cov_estimator = nothing,
    custom_cov_args = (),
    custom_cov_kwargs = (),
)

Exponentially weighted Capital Asset Pricing Model (ECAPM) returns. ECAPMRet() uses the exponentially weighted mean to estimate the mean market returns.

rspan: span for the exponentially weighted mean returns, same as span for EMRet() above.

cspan: span for the exponentially weighted covariance, same as span for cov.

PortfolioOptimiser.returns_from_prices — Function

returns_from_prices(prices, log_ret = false)

Compute the returns from prices.

prices: array of prices, each column is an asset, each row an entry.

log_ret: if false compute the normal returns, if true compute the logarithmic returns.

PortfolioOptimiser.prices_from_returns — Function

prices_from_returns(returns, log_ret = false)

Compute the prices from returns.

returns: array of returns, each column is an asset, each row an entry.

log_ret: if false asume normal returns, if true asume logarithmic returns.