Objective Functions

PortfolioOptimiser.port_variance — Function

port_variance(w, cov_mtx)

Computes the overall portfolio variance according to the weights w, and covariance matrix cov_mtx.

The portfolio variance $\sigma^2$, is defined as:

\[\sigma^2 = \bm{w}^T\, \Sigma\, \bm{w}\,,\]

where $\sigma$ is the overall portfolio standard deviation, $^T$ the transpose operator, $w$ the weights, and $\Sigma$ the covariance matrix.

PortfolioOptimiser.port_return — Function

port_return(w, mean_ret)

Computes the overall portfolio return according to the weights w, and mean returns mean_ret.

The portfolio return $\mu$, is defined as:

\[\mu = \bm{w} \cdot \bm{E},\]

where $\bm{w}$ are the weights, and $\bm{E}$ the mean returns.

PortfolioOptimiser.sharpe_ratio — Function

sharpe_ratio(w, mean_ret, cov_mtx, rf::Real = 0.02)

Computes the portfolio's Sharpe ratio given the weights w, mean returns mean_ret, covariance matrix cov_mtx, and risk free rate rf.

sharpe_ratio(μ, σ, rf::Real = 0.02)

Computes the Sharpe ratio given the return μ, standard deviation σ, and risk free rate rf.

The Sharpe ratio $s$, is defined as:

\[s = \dfrac{\mu - r}{\sigma},\]

where $\mu$ the portfolio's return (see port_return), and $\sigma$ the portfolio's standard deviation (see port_variance). Generally speaking, the greater the Sharpe ratio the better the portfolio.

Note

The Sharpe ratio penalises large swings in both directions, so assets that tend to have large increases in value are disproportionally penalised by this measure. The Sortino ratio has the same formula but uses an adjusted covariance matrix that accounts only for the negative fluctuations in value. The semicovariance is implemented by cov when given SCov() or ESCov() as its first argument. The Mean-Semivariance optimisations EffMeanSemivar make the adjustment too.

PortfolioOptimiser.port_semivar — Function

port_semivar(w, returns, target = 0, freq = 252)

Compute the semivariance from the weights w, historical returns returns, according to the target target, and frequency of returns freq.

The semivariance is defined as:

\[\begin{aligned} \bm{r} &= \mathrm{R} \bm{w}\,,\\ \sigma_s^2 &= \dfrac{f}{N} \sum\limits_{i = 1,\, b < r_i}^{N} (r_i - b)^2\,, \end{aligned}\]

where $\bm{r}$ are the portfolio historical returns with where the subscript $i$ describes a specific point in time (entry), $\mathrm{R}$ the asset historical returns, $\bm{w}$ the asset weights, $\sigma_s$ the semideviation, $f$ the frequency of the historical returns, $N$ the number of entries in historical returns (not the number of assets), and $b$ the target for splitting "upside" and "downside" returns.

The condition $b < r_i$ ensures we only consider entries for which the historical portfolio return fell below the target.

PortfolioOptimiser.L2_reg — Function

L2_reg(w, γ = 1)

L2 regularisation. Minimising this reduces the number of negligible weights w, with a tuning parameter γ. Increasing γ will decrease the number of negligible weights.

The $L_2$ regularisation is defined as:

\[L_2 = \gamma (\bm{w} \cdot \bm{w}) = \gamma \lVert \bm{w} \rVert ^2\,,\]

where $\gamma$ is the tuning parameter, and $\bm{w}$ the weights.

PortfolioOptimiser.quadratic_utility — Function

quadratic_utility(w, mean_ret, cov_mtx, risk_aversion = 1)

Calculates the quadratic utility given the weights w, mean returns mean_ret, covariance matrix cov_mtx and risk aversion risk_aversion. Increasing risk_aversion decreases risk.

The quadratic utility $Q$, is defined as:

\[Q = \mu - \dfrac{1}{2} \delta \bm{w}^T\, \Sigma\, \bm{w}\,,\]

where $\mu$ is the overall portfolio expected return (see port_return), $\delta$ the risk aversion, $\bm{w}$ the weights, and $\Sigma$ the covariance matrix.

PortfolioOptimiser.transaction_cost — Function

transaction_cost(w, w_prev, k = 0.001)

Simple transaction cost model. Sums all the absolute changes in weight and multiplies by a constant k. Simulates a fixed percentage commision from a broker.

The transacion cost $C$, is defined as:

\[C = k \lvert \bm{w} - \bm{w}_{\mathrm{prev}} \rvert\,,\]

where $k$ is the fixed percentage commision, $\bm{w}$ the asset weights, and $\bm{w}_{\mathrm{prev}}$ the previous weights of the assets.

Warning

JuMP doesn't yet support norm in the objective (v1.0), so we need to turn them into a variable subject to a MOI.NormOneCone constraint. We can follow the example in the JuMP tutorial to do this.

n = length(tickers)
prev_weights = fill(1 / n, n)
k = 0.001
ef = EffMeanVar(
    tickers,
    mu,
    S;
    # Add the variable that will contain the value of the l1 regularisation.
    extra_vars = [:(0 <= l1)],
    # We constrain it to be within a MOI.NormOneCone.
    extra_constraints = [
        :([model[:l1]; (model[:w] - $prev_weights)] in MOI.NormOneCone($(n + 1))),
    ],
    # We add the variable to the objective and multiply it by the adjustment parameter.
    extra_obj_terms = [quote
        $k * model[:l1]
    end],
)

Similarly L2-norms must be turned into constraints of type MOI.NormTwoCone. More information on how to do this can be found in JuMP Vector Cones.

PortfolioOptimiser.ex_ante_tracking_error — Function

ex_ante_tracking_error(w, cov_mtx, w_bmk)

Compute the square of the ex-ante tracking error from the weights w, covariance matrix cov_mtx, and benchamark weights w_bmk.

The ex-ante tracking error $\mathrm{Err}$, is defined as:

\[\mathrm{Err} = (\bm{w} - \bm{w}_b)^T \, \Sigma \, (\bm{w} - \bm{w}_b)\,,\]

where $\bm{w}$ are the weights, $\bm{w}_b$ the target weights, and $\Sigma$ the covariance matrix.

PortfolioOptimiser.ex_post_tracking_error — Function

ex_post_tracking_error(w, returns, ret_bmk)

Compute the square of the ex-post tracking error from the weights w, historical returns returns, and target returns ret_bmk.

The ex-post tracking error $\mathrm{Err}$, is defined as:

\[\begin{aligned} \bm{x} &= \mathrm{R} \bm{w} - \bm{R}_b \,,\\ \mu &= \dfrac{1}{N} \sum\limits_{i=1}^N x_i \,,\\ \mathrm{Err} &= (\bm{x} - \mu) \cdot (\bm{x} - \mu) = Var(r - r_b)\,, \end{aligned}\]

where $w$ are the weights, $\mathrm{R}$ historical returns, $\bm{R}_b$ target historical returns. In other words the ex-post tracking error is the variance of the historical returns $r$, minus the target historical returns $r_b$, of the portfolio.

PortfolioOptimiser.logarithmic_barrier — Function

logarithmic_barrier(w, cov_mtx, k = 0.1)

Compute the logarithmic barrier adjusted by k, from the weights w, and covariance matrix cov_mtx.

The logarithmic barrier $L$, is defined as:

\[L = \bm{w}^T\, \Sigma\, \bm{w} - k \sum\limits_i \ln( w_i )\,,\]

where $\sigma^2 = \bm{w}^T\, \Sigma\, \bm{w}$ (see port_variance), $k$ the adjustment constant, and $w_i$ the weight of the $i$'th asset.

PortfolioOptimiser.kelly_objective — Function

kelly_objective(w, mean_ret, cov_mtx, k = 3)

Compute a Kelly objective from adjusted by k, from the weights w, mean returns mean_ret, and covariance matrix cov_mtx.

The Kelly objective $K$, is defined as:

\[K = \bm{w}^T\, \Sigma\, \bm{w} - \dfrac{1}{2} k \bm{w} \cdot \bm{E} \,,\]

where $\sigma^2 = \bm{w}^T\, \Sigma\, \bm{w}$ (see port_variance), $k$ is the adjustment constant, $\bm{w}$ the weights, and $\bm{E}$ the mean returns.