Risk Models

PortfolioOptimiser.AbstractFixPosDef — Type

abstract type AbstractFixPosDef end
struct SFix <: AbstractFixPosDef end
struct DFix <: AbstractFixPosDef end

Types for fixing non-positive definite matrices in risk_matrix.

SFix: fix via the spectral method (eigenvalue decomposition).

DFix: fix by adding a multiple of the identity matrix (damping factor).

PortfolioOptimiser.AbstractRiskModel — Type

abstract type AbstractRiskModel end
struct Cov <: AbstractRiskModel end
struct SCov <: AbstractRiskModel end
struct ECov <: AbstractRiskModel end
struct ESCov <: AbstractRiskModel end

Risk models for dispatch on risk_matrix.

Cov: Mean returns.

SCov: Exponentially weighted mean returns.

ECov: Capital Asset Pricing Model (CAPM) returns.

ESCov: Exponentially weighted Capital Asset Pricing Model (ECAPM) returns.

PortfolioOptimiser.risk_matrix — Function

Calculate covariance matrices.

Wrapper function

risk_matrix(
    type::AbstractRiskModel,
    returns;
    target = 1.02^(1 / 252) - 1,
    fix_method::Union{SFix, DFix} = SFix(),
    freq = 252,
    span = Int(ceil(freq / 1.4)),
)

Dispatches specific covariance matrix calculation methods according to type.

returns: array of asset returns, each column is an asset, each row an entry.

target: target for distinguishing "upside" and "downside" returns. Defaults to the daily risk-free rate.

fix_method: the fixing method used in case the matrix is non-positive definite.

freq: frequency at which returns are recorded, defaults to the average number of trading days in a year.

span: span parmeter for calculating the exponential weights, α = 2 / (span + 1).

Normal covariance

risk_matrix(::Cov, returns; fix_method::Union{SFix, DFix} = SFix(), freq = 252)

Semi-covariance

The semi-covariance sets a target for distinguishing "upside" and "downside" risk. It only considers fluctuations below target which heavily biases losses. Lets users minimise negative risk.

risk_matrix(
    ::SCov,
    returns;
    target = 1.02^(1 / 252) - 1,
    fix_method::Union{SFix, DFix} = SFix(),
    freq = 252,
)

Exponentially weighted covariance

The exponentially weighted covariance uses exponentially weighted expected returns rather than normally weighted expected returns. More recent returns are weighted more heavily.

risk_matrix(
    ::ECov,
    returns;
    fix_method::Union{SFix, DFix} = SFix(),
    freq = 252,
    span = Int(ceil(freq / 1.4)),
)

Exponentially weighted semi-covariance

Combines the semi-covariance and exponentially weighted covariance.

risk_matrix(
    ::ESCov,
    returns;
    target = 1.02^(1 / 252) - 1, # Daily risk free rate.
    fix_method::Union{SFix, DFix} = SFix(),
    freq = 252,
    span = Int(ceil(freq / 1.4)),
)

PortfolioOptimiser.make_pos_def — Function

Utility function for fixing non-positive definite matrices.

Spectral method

make_pos_def(::SFix, matrix)

Uses eigenvalue decomposition to make the negative eigenvalues equal to zero and reconstruct the matrix with new eigenvalues.

Diagonal method

make_pos_def(::DFix, matrix, scale = 1.1)

Uses a Levenberg-Marquardt factor, c = scale * min(eigenvalues), which is used to subtract a scaled identity matrix from matrix.

StatsBase.cov2cor — Function

cov2or(cov_mtx)

Wraps StatsBase.cov2cor to provide the standard deviations vector from the diagonal of cov_mtx.