PortfolioOptimiser

Documentation for PortfolioOptimiser.

PortfolioOptimiser.AbstractAllocation

PortfolioOptimiser.AbstractPortfolioOptimiser

PortfolioOptimiser.Allocation

PortfolioOptimiser.Allocation

PortfolioOptimiser.L2_reg

PortfolioOptimiser.L2_reg

PortfolioOptimiser._Allocation

PortfolioOptimiser._Allocation

PortfolioOptimiser._Allocation

PortfolioOptimiser._add_constraint_to_model!

PortfolioOptimiser._add_nlparameter_to_model!

PortfolioOptimiser._add_to_objective!

PortfolioOptimiser._add_var_to_model!

PortfolioOptimiser._clean_zero_shares

PortfolioOptimiser._create_weight_bounds

PortfolioOptimiser._short_allocation

PortfolioOptimiser._sub_allocation

PortfolioOptimiser.add_sector_constraint!

PortfolioOptimiser.ex_ante_tracking_error

PortfolioOptimiser.ex_ante_tracking_error

PortfolioOptimiser.ex_post_tracking_error

PortfolioOptimiser.ex_post_tracking_error

PortfolioOptimiser.kelly_objective

PortfolioOptimiser.kelly_objective

PortfolioOptimiser.logarithmic_barrier_objective

PortfolioOptimiser.logarithmic_barrier_objective

PortfolioOptimiser.max_return

PortfolioOptimiser.max_sharpe!

PortfolioOptimiser.min_volatility!

PortfolioOptimiser.port_return

PortfolioOptimiser.port_return

PortfolioOptimiser.port_variance

PortfolioOptimiser.port_variance

PortfolioOptimiser.quadratic_utility

PortfolioOptimiser.quadratic_utility

PortfolioOptimiser.roundmultiple

PortfolioOptimiser.roundmultiple

PortfolioOptimiser.semi_ret

PortfolioOptimiser.semi_ret

PortfolioOptimiser.sharpe_ratio

PortfolioOptimiser.sharpe_ratio

PortfolioOptimiser.transaction_cost

PortfolioOptimiser.transaction_cost

StatsBase.cov2cor

PortfolioOptimiser.L2_reg — Function

L2_reg(w, γ = 1)

L2 regularisation. Minimising this reduces the number of negligible weights w, with a tuning parameter γ. Increasing γ will decrease the number of negligible weights.

$L_2$ regularisation is defined as:

\[L_2 = \gamma \Vert \bm{w} \Vert\,,\]

where $\gamma$ is the tuning parameter, and $w$ the asset weights.

PortfolioOptimiser._Allocation — Function

_Allocation(
    type::Lazy,
    tickers::AbstractArray,
    weights::AbstractArray,
    latest_prices::AbstractVector,
    investment::Real = 1e4,
    rounding::Real = 1,
    reinvest::Bool = false,
    short_ratio::Union{Real, Nothing} = nothing,
    optimiser = nothing,
    silent = nothing,
)

Lazy asset allocation. Simply finds how many shares up to the smallest multiple of rounding you can buy for each ticker.

Note

optimiser and silent are not used but are in the function signature for dispatch purposes.

PortfolioOptimiser._Allocation — Function

_Allocation(
    type::LP,
    tickers::AbstractArray,
    weights::AbstractArray,
    latest_prices::AbstractVector,
    investment::Real = 1e4,
    rounding = nothing,
    reinvest::Bool = false,
    short_ratio::Union{Real, Nothing} = nothing,
    optimiser = HiGHS.Optimizer,
    silent = true,
)

Linear mixed-integer optimisation of assets.

Note

rounding has no effect, as it can only allocate integer numbers of shares.

PortfolioOptimiser._Allocation — Function

_Allocation(
    type::Greedy,
    tickers::AbstractArray,
    weights::AbstractArray,
    latest_prices::AbstractVector,
    investment::Real = 1e4,
    rounding::Real = 1,
    reinvest::Bool = false,
    short_ratio::Union{Real, Nothing} = nothing,
    optimiser = nothing,
    silent = nothing,
)

Greedy algorithm that allocates assets according to their optimised weights. It priorites assets according to the ones with the greatest differnce between their weight in the portfolio and ideal weight. It allocates assets by progressively adding rounding to each share until the cost exceeds the available funds.

Note

optimiser and silent are not used but are in the function signature for dispatch purposes.

PortfolioOptimiser._add_constraint_to_model! — Method

extra_constraints = [:(model[:w][2] >= 2), :(sum(model[:w]) <= 5)]
maxSharpe = EfficientFrontier(MaxSharpe(), tickers, mean_ret, cov_mtx, weight_bounds)
_add_constraint_to_model!.(maxSharpe.model, [:const1, :const2], extra_constraints)

Extra constraints have to be expressions. Variables have to be registered in the model and they have to be explicitly written as model[<variable_key>]. This is because the variables in the constraints must be within scope when the code is automatically generated.

PortfolioOptimiser._add_nlparameter_to_model! — Method

# Now try with a nonconvex objective from  Kolm et al (2014)
objective = quote
    function deviation_risk_parity(w::T...) where {T}
        w = [i for i in w]
        tmp = w .* (value.(cov_mtx) * w)
        diff = tmp .- tmp'
        return sum(diff .* diff)
    end
end

obj_args = [:w]
obj_params =
    [(:(cov_mtx[i = 1:length(model[:w]), j = 1:length(model[:w])]), :(portfolio.cov_mtx))]

objargs are the arguments of the objective function objparams array of tuples, first entry in tuple is the parametr, the second is the value of the parameter

deviationriskparity the arguments are objargs, and the part where it says value.(covmtx) is the parameter

PortfolioOptimiser._add_to_objective! — Method

extra_obj_terms = [
    quote
        model[:w][1] * 5 - sum(model[:w])
    end, 
    quote
        model[:k]
    end
]
maxSharpe = EfficientFrontier(MaxSharpe(), tickers, mean_ret, cov_mtx, weight_bounds)
_add_to_objective!.(maxSharpe.model, extra_obj_terms)

Extra objective function terms have to be quotes because they are automatically spliced into code. Again the variables used have to be registered in the model, and model[<variable_key>] has to be explicitly typed to ensure the variables are within scope when the code is automatically generated. The terms have to be compliant with the type of objective function.

PortfolioOptimiser._add_var_to_model! — Method

extra_vars = [:(c[1:3]), :k]
maxSharpe = EfficientFrontier(MaxSharpe(), tickers, mean_ret, cov_mtx, weight_bounds)
_add_var_to_model!.(maxSharpe.model, extra_vars)

Extra array variables have to be expressions. Extra scalar variables can be symbols or expressions.

PortfolioOptimiser._clean_zero_shares — Method

_clean_zero_shares(shares_bought, tickers, latest_prices)

Helper function for removing tickers with zero shares.

PortfolioOptimiser._create_weight_bounds — Method

_map_bounds_to_constraints(bounds) 

Convert the bounds to JuMP constraints.

PortfolioOptimiser._short_allocation — Method

function _short_allocation(
    type,
    tickers,
    weights,
    latest_prices,
    investment,
    rounding,
    reinvest,
    short_ratio,
    optimiser,
    silent
)

Helper function for allocating short-long portfolios. Ensures the shares and weights of the shorts are negative.

PortfolioOptimiser._sub_allocation — Method

_sub_allocation(type, tickers, weights, latest_prices, investment, rounding, idx)

Helper function for creating sub allocations. It calls Allocationagain but only for theidx` provided.

PortfolioOptimiser.add_sector_constraint! — Method

function add_sector_constraint!(opt::PortTypes, sector_map, sector_lower, sector_upper)

Adds constraints on the sum of the weights of different tickers in the same group/sector. The portfolio will only be exposed to <= x % of each sector according to the weights.

sector_map = Dict(<ticker> => <sector>)
sector_lower = Dict(<sector> => sector_lower_bound)
sector_upper = Dict(<sector> => sector_upper_bound)

PortfolioOptimiser.ex_ante_tracking_error — Method

ex_ante_tracking_error(w, cov_mtx, w_bmk)

Compute the square of the ex-ante tracking error from the weights w, covariance matrix cov_mtx, and benchamark weights w_bmk.

The ex-ante tracking error, $\mathrm{Err}$ is defined as:

\[\mathrm{Err} = (\bm{w} - \bm{w}_b)^T \, \Sigma \, (\bm{w} - \bm{w}_b)\,,\]

where $\bm{w}$ are the asset weights, $\bm{w}_b$ the benchmark weights, and $\Sigma$ the covariance matrix.

PortfolioOptimiser.ex_post_tracking_error — Method

ex_post_tracking_error(w, returns, ret_bmk)

Compute the square of the ex-post tracking error from the given weights w, historical returns returns, and benchmark returns ret_bmk.

The ex-post tracking error, $\mathrm{Err}$ is defined as:

\[\begin{aligned} x &= \mathrm{R} \bm{w} - \bm{R}_b \\ \mu &= \dfrac{1}{N} \sum\limits^N x \\ \mathrm{Err} &= (x - \mu) \cdot (x - \mu) = Var(r - r_b)\,, \end{aligned}\]

where $w$ are the asset weights, $\mathrm{R}$ historical returns, $\bm{R}_b$ benchmark historical returns. In other words the ex-post tracking error is the variance of the historical returns $r$, minus the benchmark returns $r_b$, of the portfolio.

PortfolioOptimiser.kelly_objective — Function

kelly_objective(w, mean_ret, cov_mtx, k = 3)

Compute a Kelly objective from adjusted by k, from the given weights w, mean returns mean_ret, and covariance matrix cov_mtx.

The Kelly objective, $K$ is defined as:

\[K = \bm{w}^T\, \Sigma\, \bm{w} - \dfrac{k}{2} \bm{w} \cdot \bm{E} \,,\]

where $\sigma^2 = \bm{w}^T\, \Sigma\, \bm{w}$ (see port_variance), $k$ is the adjustment constant, $\bm{w}$ the asset weights, and $\bm{E}$ the asset mean returns.

PortfolioOptimiser.logarithmic_barrier_objective — Function

logarithmic_barrier_objective(w, cov_mtx, k = 0.1)

Compute the logarithmic barrier adjusted by k, from the given weights w, covariance matrix cov_mtx.

The logarithmic barrier, $L$ is defined as:

\[L = \bm{w}^T\, \Sigma\, \bm{w} - k \sum\limits_i \ln( w_i )\,,\]

where $\sigma^2 = \bm{w}^T\, \Sigma\, \bm{w}$ (see port_variance), $k$ the adjustment constant, and $w_i$ the weight of the $i$'th asset.

PortfolioOptimiser.max_return — Function

optimise(portfolio::EfficientFrontier)

Warning

This should not be used for optimising portfolios. It's used internally for some optimisations. It yields portfolios with large volatilities.

PortfolioOptimiser.max_sharpe! — Function

optimise!(
    portfolio::EfficientFrontier{MaxSharpe},
    optimiser = Ipopt.Optimizer,
    silent = true,
)

Maximise portfolio for maximum sharpe ratio.

PortfolioOptimiser.min_volatility! — Function

function optimise!(
    portfolio::EfficientFrontier{MinVolatility},
    optimiser = Ipopt.Optimizer,
    silent = true,
)

Optimise portfolio for minimum volatility.

PortfolioOptimiser.port_return — Method

port_return(w, mean_ret)

Computes the overall portfolio return according to the weights w, and mean returns mean_ret:

\[\mu = \bm{w} \cdot \bm{E},\]

where $\mu$ is the overall portfolio return, $\bm{w}$ the weights, and $\bm{E}$ the asset mean returns.

PortfolioOptimiser.port_variance — Method

port_variance(w, cov_mtx)

Computes the overall portfolio variance according to the weights w, and covariance matrix cov_mtx:

\[\sigma^2 = \bm{w}^T\, \Sigma\, \bm{w}\,,\]

where $\sigma$ is the overall portfolio standard deviation, $^T$ the transpose operator, $w$ the weights, and $\Sigma$ the covariance matrix.

PortfolioOptimiser.quadratic_utility — Function

quadratic_utility(w, mean_ret, cov_mtx, risk_aversion = 1)

Calculates the quadratic utility given weights w, mean returns mean_ret, covariance matrix cov_mtx and risk aversion risk_aversion. Increasing risk_aversion decreases risk.

Quadratic utility $Q$, is defined as:

\[Q = \mu - \dfrac{1}{2} \delta \bm{w}^T\, \Sigma\, \bm{w}\,,\]

where $\mu$ is the overall portfolio expected return (see port_return), $\delta$ the risk aversion, $w$ the asset weights, and $\Sigma$ the covariance matrix.

PortfolioOptimiser.roundmultiple — Method

roundmultiple(val, prec [, args...] [; kwargs...])

Round a number to a multiple of prec. Uses args and kwargs of the built-in round.

PortfolioOptimiser.semi_ret — Function

semi_ret(returns, benchmark = 0)

Compute the historical semi-returns from historical returns, for a given benchmark. Return values greater than benchmark are "upside" returns, values lower than benchmark are "downside" returns. Each column of returns corresponds to an asset.

Historical semi returns $\mathrm{R_b}$, are defined as:

\[\mathrm{R_b} = \dfrac{\mathrm{R} - b}{\sqrt{N}}\,\]

where $\mathrm{R}$ are historical returns, $b$ is the benchmark, and $N$ is the number of historical entries (not the number of assets). The value of benchmark should correspond to the frequency of historical returns, i.e. if using daily returns benchmark should be a reasonable value for daily returns.

PortfolioOptimiser.sharpe_ratio — Function

sharpe_ratio(w, mean_ret, cov_mtx, rf::Real = 0.02)

Computes the portfolio's Sharpe ratio given the weights w, mean returns mean_ret, covariance matrix cov_mtx, and risk free rate rf.

sharpe_ratio(μ, σ, rf::Real = 0.02)

Computes the Sharpe ratio given the return μ, standard deviation σ, and risk free rate rf.

The Sharpe ratio is defined as:

\[s = \dfrac{\mu - r}{\sigma},\]

where $s$ is the Sharpe ratio, $\mu$ the portfolio's return (see port_return), and $\sigma$ the portfolio's standard deviation (see port_variance). Generally speaking, the greater the Sharpe ratio the better the portfolio.

Note

The Sharpe ratio penalises large swings in both directions, so assets that tend to have large increases in value are disproportionally penalised by this measure. The Sortino ratio has the same formula but uses an adjusted covariance matrix that accounts only for the negative fluctuations in value. SemiVar() and ExpSemiVar() in AbstractRiskModel](@ref) implement this adjustment, and AbstractMeanSemiVar portfolios use it in their optimisations.

PortfolioOptimiser.transaction_cost — Function

transaction_cost(w, w_prev, k = 0.001)

Simple transaction cost model. Sums all the weight absolute changes and multiplies by a fraction k. Simulates a fixed percentage commision from a broker.

It is defined as:

\[C = k \lvert \bm{w} - \bm{w}_{\mathrm{prev}} \rvert\]

Warning

As of JuMP 1.0 there is no support for norm in objective functions. A model wishing to add this to their objective function should instead it as a constraint Cone, eg:

n = length(tickers)
prev_weights = fill(1 / n), n))
k = 0.001

extra_vars = [:(z[1:($n)])]
extra_constraints =
    [:([$k * (model[:w] - $prev_weights); model[:z]] in MOI.NormOneCone($(n + n)))]

ef = EfficientFrontier(tickers, mu, S;
        extra_vars = extra_vars,
        extra_constraints = extra_constraints,
    )

StatsBase.cov2cor — Method

cov2or(cov_mtx)

Uses the fact that the diagonal elements of the covariance matrix are the variances. Puts the covariances in a diagonal matrix, takes their square root and inverts it. It then left and right multiplies the covariance matrix. Basically performs all the necessary divisions by the individual standard deviations.