PortfolioOptimiser

PortfolioOptimiser.PortfolioOptimiser — Module

PortfolioOptimiser

Category	Badge

Docs

Examples

CI

Coverage

Description

PortfolioOptimiser is a library for portfolio optimisation. It offers a broad range of functionality, and is designed with ease of use, composability, extensibility, maintainability in mind. It does so by leveraging Julia's type system, multiple dispatch, and the principle of separation of concerns.

PortfolioOptimiser takes a hands-off approach when it comes to solvers and solver settings. The built-in forumations are mostly conic in nature as these tend to yield more accurate solutions that are easier to solve. This means the choice of solver and its parameters are left to the user.

The library currently focues purely on optimisation and parameter estimation. There is currently no plan for backtesting, model selection, validation, or returns series generation. These may be implemented in the future as separate packages.

Quick Start

The following example shows how one can download the data and perform a simple optimisation.

using PortfolioOptimiser, TimeSeries, DataFrames, Clarabel, HiGHS, YFinance, Dates, JuMP

function stock_price_to_time_array(x)
    # Only get the keys that are not ticker or datetime.
    coln = collect(keys(x))[3:end]
    # Convert the dictionary into a matrix.
    m = hcat([x[k] for k ∈ coln]...)
    return TimeArray(x["timestamp"], m, Symbol.(coln), x["ticker"])
end

# Tickers of the assets we want to download.
assets = sort!(["SOUN", "RIVN", "GME", "AMC", "SOFI", "ENVX", "ANVS", "LUNR", "EOSE", "SMR",
                "NVAX", "NKLA", "ACHR", "RKLB", "MARA"])

# Prices date range.
Date_0 = "2024-01-01"
Date_1 = "2025-01-01"

# Download the price data using YFinance.
prices = get_prices.(assets; startdt = Date_0, enddt = Date_1)

# Convert vector of ordered dicts into a TimeArray.
prices = stock_price_to_time_array.(prices)
prices = hcat(prices...)

# Select only the adjusted close prices.
cidx = colnames(prices)[occursin.(r"adj", string.(colnames(prices)))]
prices = prices[cidx]

# Rename the columns to the asset tickers.
TimeSeries.rename!(prices, Symbol.(assets))

# Generate the portfolio.
portfolio = Portfolio(; prices = prices,
                      # Continuous solvers.
                      solvers = PortOptSolver(
                                              # Key-value pair for the solver, solution acceptance
                                              # criteria, model bridge argument, and solver attributes.
                                              ; name = :Clarabel,
                                              # Solver we wish to use.
                                              solver = Clarabel.Optimizer,
                                              # (Optional) Solution acceptance criteria.
                                              check_sol = (allow_local = true,
                                                           allow_almost = true),
                                              # (Optional) Solver-specific attributes.
                                              params = Dict("verbose" => false,
                                                            "max_step_fraction" => 0.75,
                                                            "tol_gap_abs" => 1e-9,
                                                            "tol_gap_rel" => 1e-9)),
                      # Discrete solvers (for discrete allocation).
                      alloc_solvers = PortOptSolver(
                                                    # Key-value pair for the solver, solution acceptance
                                                    # criteria, model bridge argument, and solver attributes.
                                                    ; name = :HiGHS,
                                                    # Solver we wish to use.
                                                    solver = optimizer_with_attributes(HiGHS.Optimizer,
                                                                                       MOI.Silent() => true),
                                                    # (Optional) Solution acceptance criteria.
                                                    check_sol = (allow_local = true,
                                                                 allow_almost = true)))

# Compute the asset statistics.
asset_statistics!(portfolio)

# Optimise the portfolio using the traditional Markowitz model with a conic formulation.
w = optimise!(portfolio, Trad())

# Discretely allocate 69420 dollars in shares of the assets.
wd = allocate!(portfolio; investment = 69420)

# Show discrete allocation with optimal allocation.
wd.optimal_weights = w.weights
DataFrames.rename!(wd, :weights => :allocated_weights)
show(wd)

15×6 DataFrame
 Row │ tickers  shares  price    cost      allocated_weights  optimal_weights 
     │ String   Int64   Float64  Float64   Float64            Float64         
─────┼────────────────────────────────────────────────────────────────────────
   1 │ ACHR        220     9.75   2145.0           0.0308991       0.0308762
   2 │ AMC         709     3.98   2821.82          0.0406487       0.0406629
   3 │ ANVS          0     5.03      0.0           0.0             4.03909e-8
   4 │ ENVX          0    10.87      0.0           0.0             3.05551e-7
   5 │ EOSE       1256     4.86   6104.16          0.0879314       0.0879854
   6 │ GME           0    31.34      0.0           0.0             4.34053e-8
   7 │ LUNR        125    18.16   2270.0           0.0326997       0.0328007
   8 │ MARA          0    16.77      0.0           0.0             1.04058e-7
   9 │ NKLA       4682     1.19   5571.58          0.0802595       0.0802858
  10 │ NVAX        424     8.04   3408.96          0.0491066       0.0491266
  11 │ RIVN        710    13.3    9443.0           0.136028        0.135939
  12 │ RKLB        407    25.47  10366.3           0.149328        0.149194
  13 │ SMR           0    17.93      0.0           0.0             6.62625e-7
  14 │ SOFI       1772    15.4   27288.8           0.393099        0.393129
  15 │ SOUN          0    19.84      0.0           0.0             4.11381e-8

Functionality

Non-hierarchical optimisation models

Traditional, Trad.

Risk Budgeting, RB.

Relaxed Risk Budgetting (Variance only), RRB.

Near Optimal Centering, NOC.

Hierarchical optimisation models

Hierarchical Risk Parity, HRP.

Hierarchical Risk Parity Schur Complement (Variance only), SchurHRP.

Hierarchical Equal Risk Parity, HERC.

Nested Clustered Optimisation, NCO.

Expected returns estimators

Arithmetic (weighted and unweighted), MuSimple.

Equilibrium, MuEquil.

Shringage with Grand Mean GM, Volatility Weighted VW, and Mean Square Error MSE targets:

James-Stein, MuJS.

Bayes-Stein, MuBS.

Bodnar-Okhrin-Parolya, MuBOP.

Covariance estimators

These utilise StatsBase's API to define covariance estimators. Which means PortfolioOptimiser is compatible with CovarianceEstimation.

PortfolioOptimiser, PortCovCor.

Full, CovFull.

Semi, CovSemi.

Mutual Information, CovMutualInfo.

Distance, CovDistance.

Lower Tail Dependence, CovLTD.

Gerber, CovGerber0, CovGerber1, CovGerber2.

Smyth-Broby, CovSB0, CovSB1, CovSB2.

Smyth-Broby-Gerber, CovGerberSB0, CovGerberSB1, CovGerberSB2.

Correlation estimators

These utilise StatsBase's API to define covariance estimators. Which means PortfolioOptimiser is compatible with CovarianceEstimation.

All covariance estimators.

Spearman, CorSpearman.

Kendall, CorKendall.

Distance estimators

We provide distance and distance of distances estimators. Distance estimators have the prefix Dist, distance of distances have the prefix DistDist.

Canonical, DistCanonical, DistDistCanonical.

Marcos López de Prado, DistMLP, DistDistMLP.

Generalised Marcos López de Prado, GenDistMLP, GenDistDistMLP.

Log, DistLog, DistDistLog.

Distance correlation, DistCor, DistDistCor.

Variation of information, DistVarInfo, DistDistVarInfo.

Cokurtosis estimators

Full, KurtFull.

Semi, KurtSemi.

Coskewness estimators

Full, SkewFull.

Semi, SkewSemi.

Square matrix post-processing

Fixing non-positive definite matrices, NoPosdef, PosdefNearest.

Matrix denoising, NoDenoise, DenoiseFixed, DenoiseSpectral, DenoiseShrink.

Matrix detoning, NoDetone, Detone.

Local-global sparsification of the matrix inverse, NoLoGo, LoGo.

Prior estimators

Empirical, asset_statistics!.

Worst-case uncertainty sets, wc_statistics!.

Factor models, factor_statistics!.

Black-Litterman, black_litterman_statistics!.

Black-Litterman factor models, black_litterman_factor_statistics!.

Regression models

Forward and Backward Regression, FReg, BReg, with criteria.

p-value threshold, PVal.

Akaike's Information Criterion, AIC.

Corrected Akaike's Information Criterion for small sample sizes, AICC.

Bayesian Information Criterion, BIC.

R² of a linear model, RSq.

adjusted R² for a linear model, AdjRSq.

Principal Component-Based Regression, PCAReg, with targets.

PCA target, PCATarget.

Probabilistic PCA target, PPCATarget.

Worst-case uncertainty set estimators

These are only for the covariance and expected returns.

Box and Elliptical sets, Box, Ellipse.

Autoregressive Conditional Heteroskedasticity models, ArchWC, with bootstraps.

Stationary, StationaryBS.

Circular, CircularBS.

Moving, MovingBS.

Normal, NormalWC.

Box sets only.

Delta, DeltaWC.

Elliptical set constraint error size estimation.

Normal, KNormalWC.

General, KGeneralWC

Clustering

Direct Bubble Hierarchy Trees, DBHT.

Hierarchical clustering, HAC.

Optimal number of clusters.

Two-difference gap statistic, TwoDiff.

Standardised silhouette scores, StdSilhouette.

Networks

Triangular maximally filtered graphs (TMFG), TMFG.

Minimum spanning trees (MST), MST.

Kruskal, KruskalTree.

Boruvka, BoruvkaTree.

Prim, PrimTree.

Centrality measures.

Betweenness, BetweennessCentrality.

Closeness, ClosenessCentrality.

Degree, DegreeCentrality.

Eigenvector, EigenvectorCentrality.

Katz, KatzCentrality.

Pagerank, Pagerank.

Radiality, RadialityCentrality.

Stress, StressCentrality.

UP TO HERE

Black Litterman models

Black Litterman, BLType.

Black Litterman factor models

Augmented Black Litterman, ABLType.

Bayesian Black Litterman, BBLType.

Linear moments (L-moments)

Normalised constant relative risk aversion coefficients, CRRA.

Maximum entropy, MaxEntropy.

Minimum Sum of Squares, MinSumSq.

Minimum Square Distance, MinSqDist.

Portfolio optimisation

These types of optimisations act on instances of Portfolio.

Traditional, Trad

This type of optimisation is the traditional efficient frontier optimisation.

Classes, PortClass

Classic, Classic.

Factor model, FM.

Black Litterman, BL.

Black Litterman Factor model, BLFM.

Expected returns

Arithmetic returns, NoKelly.

Approximate logarithmic mean returns, AKelly.

Exact logarithmic mean returns, EKelly.

Objective functions

Minimum risk, MinRisk.

Maximum utility, Utility.

Maximum risk adjusted return ratio, Sharpe.

Maximum return, MaxRet.

Constraints

Maximum expected risk constraints.

Minimum expected return constraint.

Linear weight constraints.

Connected asset centrality constraints.

Asset network constraints.

Leverage constraints.

Maximum number of assets constraint.

Minimum number of effective assets constraint.

Tracking error (weights or returns) constraint.

Turnover constraint.

Rebalancing penalty.

Risk measures

Dispersion

Full dispersion.

Standard deviation, SD.

Mean absolute deviation (MAD), MAD.

Square root kurtosis, Kurt.

Range, RG.

Conditional value at risk range (CVaR range), CVaRRG.

Tail Gini range, TGRG.

Gini mean difference (GMD), GMD.

Quadratic negative skewness, NQSkew.

Brownian distance variance (BDVariance), BDVariance.

Downside dispersion.

Semi standard deviation (Sortino ratio), SSD.

First lower partial moment (Omega ratio), FLPM.

Square root semi kurtosis, SKurt.

Quadratic negative semi skewness, NQSSkew.

Downside

Worst case realisation (Minimax), WR.

Conditional value at risk (CVaR), CVaR.

Entropic value at risk (EVaR), EVaR.

Relativistic value at risk (RLVaR), RLVaR.

Tail Gini, TG.

Drawdown

Maximum drawdown (Calmar ratio) for uncompounded cumulative returns, MDD.

Average drawdown for uncompounded cumulative returns, ADD.

Ulcer index for uncompounded cumulative returns, UCI.

Conditional drawdown at risk for uncompounded cumulative returns (CDaR), CDaR.

Entropic drawdown at risk for uncompounded cumulative returns (EDaR), EDaR.

Relativistic drawdown at risk for uncompounded cumulative returns (RLDaR), RLDaR.

Linear moments (L-moments)

L-moment ordered weight array, OWA.

Worst case mean variance

This type of optimisation requires worst case sets for the covariance and expected returns. The optimisation uses these sets to perform a mean variance optimisation.

Constraints

Maximum expected worst case standard deviation constraint.

Minimum expected worst case return constraint.

Linear weight constraints.

Connected asset centrality constraints.

Asset network constraints.

Leverage constraints.

Maximum number of assets constraint.

Minimum number of effective assets constraint.

Tracking error (weights or returns) constraint.

Turnover constraint.

Rebalancing penalty.

Risk parity, RB

This type of optimisation requires a risk budget per asset or factor. The optimisation attempts to minimise the difference between the risk budget and risk contribution of the asset or factor in the optimised portfolio.

Classes, PortClass

Classic, Classic.

Factor model, FM.

Factor risk contribution, FC.

Constraints

Minimum expected return constraint.

Linear weight constraints.

Risk measures

Dispersion

Full dispersion.

Standard deviation, SD.

Mean absolute deviation (MAD), MAD.

Square root kurtosis, Kurt.

Range, RG.

Conditional value at risk range (CVaR range), CVaRRG.

Tail Gini range, TGRG.

Gini mean difference (GMD), GMD.

Quadratic negative skewness, NQSkew.

Brownian distance variance (BDVariance), BDVariance.

Downside dispersion.

Semi standard deviation (Sortino ratio), SSD.

First lower partial moment (Omega ratio), FLPM.

Square root semi kurtosis, SKurt.

Quadratic negative semi skewness, NQSSkew.

Downside

Worst case realisation (Minimax), WR.

Conditional value at risk (CVaR), CVaR.

Entropic value at risk (EVaR), EVaR.

Relativistic value at risk (RLVaR), RLVaR.

Tail Gini, TG.

Drawdown

Maximum drawdown (Calmar ratio) for uncompounded cumulative returns, MDD.

Average drawdown for uncompounded cumulative returns, ADD.

Ulcer index for uncompounded cumulative returns, UCI.

Conditional drawdown at risk for uncompounded cumulative returns (CDaR), CDaR.

Entropic drawdown at risk for uncompounded cumulative returns (EDaR), EDaR.

Relativistic drawdown at risk for uncompounded cumulative returns (RLDaR), RLDaR.

Linear moments (L-moments)

L-moment ordered weight array, OWA.

Relaxed risk parity mean variance, RRB

This type of optimisation requires a risk budget per asset. The optimisation attempts to minimise the difference between the risk budget and relaxed formulation of the standard deviation risk measure.

Classes, PortClass

Classic, Classic.

Factor model, FM.

Constraints

Minimum expected return constraint.

Linear weight constraints.

Near Optimal Centering, NOC

Near optimal centering utilise the weights of an optimised portfolio. It computes a region of near optimality using the bounds of the efficient frontier, the expected risk and return of the optimal portfolio, and a user-provided parameter. It then optimises for a portfolio that best describes the region. It provides more diversification and robustness than Trad and smooths out the weight transitions as the efficient frontier is traversed.

Classes, PortClass

Classic, Classic.

Factor model, FM.

Black Litterman, BL.

Black Litterman Factor model, BLFM.

Constraints

Minimum expected return constraint.

Linear weight constraints.

Classes, PortClass

Classic, Classic.

Factor model, FM.

Black Litterman, BL.

Black Litterman Factor model, BLFM.

Expected returns

Arithmetic returns, NoKelly.

Approximate logarithmic mean returns, AKelly.

Exact logarithmic mean returns, EKelly.

Constraints

Maximum expected risk constraints.

Minimum expected return constraint.

Risk measures

Dispersion

Full dispersion.

Standard deviation, SD.

Mean absolute deviation (MAD), MAD.

Square root kurtosis, Kurt.

Range, RG.

Conditional value at risk range (CVaR range), CVaRRG.

Tail Gini range, TGRG.

Gini mean difference (GMD), GMD.

Quadratic negative skewness, NQSkew.

Brownian distance variance (BDVariance), BDVariance.

Downside dispersion.

Semi standard deviation (Sortino ratio), SSD.

First lower partial moment (Omega ratio), FLPM.

Square root semi kurtosis, SKurt.

Quadratic negative semi skewness, NQSSkew.

Downside

Worst case realisation (Minimax), WR.

Conditional value at risk (CVaR), CVaR.

Entropic value at risk (EVaR), EVaR.

Relativistic value at risk (RLVaR), RLVaR.

Tail Gini, TG.

Drawdown

Maximum drawdown (Calmar ratio) for uncompounded cumulative returns, MDD.

Average drawdown for uncompounded cumulative returns, ADD.

Ulcer index for uncompounded cumulative returns, UCI.

Conditional drawdown at risk for uncompounded cumulative returns (CDaR), CDaR.

Entropic drawdown at risk for uncompounded cumulative returns (EDaR), EDaR.

Relativistic drawdown at risk for uncompounded cumulative returns (RLDaR), RLDaR.

Linear moments (L-moments)

L-moment ordered weight array, OWA.

Hierarchical portfolio optimisation

These types of optimisations act on instances of .

Hierarchical risk parity, HRP, and hierarchical equal risk parity, HERC

Hierarchical risk parity, HRP

Hierarchical risk parity optimisations use the hierarchical clustering of assets to assign risk contributions by iteratively splitting the dendrogram in half and assigning weights to each half according to the relative risk each half represents with respect to the other. It does this until it splits the dendrogram all the way down to single leaves.

Hierarchical equal risk parity, HERC

Hierarchical equal risk parity optimisations use the hierarchical clustering relationships between assets to assign risk contributions by splitting the dendrogram into k clusters. It starts with the full dendrogram and progressively cuts it into k-1 levels (since the comparison for each side belongs to the k-th level). At each step, it loops through the clusters and checks to which side of the sub-dendrogram the cluster belongs. It accumulates the risk of that cluster to the risk of the side it belongs to. The weights for the assets on each side of the dendrogram are assigned based on the relative (with respect to the other side) aggregate risk from all clusters belonging to it, these are the inter-cluster weights. It then computes the risk for each cluster, assigning weights to each asset according to the relative risk it represents with respect to other assets, these are the intra-cluster weights. It then elementwise multiplies both weights to get the final asset weights.

HERC can make use of two risk measure arguments, one for the intra-cluster and one for the inter-cluster risk calculation. They can take linear combinations of risk measures.

Constraints

Minimum and maximum weights per asset.

Risk measures

Dispersion

Full dispersion.

Variance, Variance.

Standard deviation, SD.

Mean absolute deviation (MAD), MAD.

Square root kurtosis, Kurt.

Range, RG.

Conditional value at risk range (CVaR range), CVaRRG.

Tail Gini range, TGRG.

Gini mean difference (GMD), GMD.

Quadratic negative skewness, NQSkew.

Brownian distance variance (BDVariance), BDVariance.

Downside dispersion.

Semi Variance, Variance.

Semi standard deviation (Sortino ratio), SSD.

First lower partial moment (Omega ratio), FLPM.

Square root semi kurtosis, SKurt.

Quadratic negative semi skewness, NQSSkew.

Downside

Worst case realisation (Minimax), WR.

Value at risk (VaR), VaR.

Conditional value at risk (CVaR), CVaR.

Entropic value at risk (EVaR), EVaR.

Relativistic value at risk (RLVaR), RLVaR.

Tail Gini, TG.

Drawdown

Maximum drawdown (Calmar ratio) for uncompounded cumulative returns, MDD.

Average drawdown for uncompounded cumulative returns, ADD.

Ulcer index for uncompounded cumulative returns, UCI.

Drawdown at for uncompounded cumulative returns risk (DaR), DaR.

Conditional drawdown at risk for uncompounded cumulative returns (CDaR), CDaR.

Entropic drawdown at risk for uncompounded cumulative returns (EDaR), EDaR.

Relativistic drawdown at risk for uncompounded cumulative returns (RLDaR), RLDaR.

Maximum drawdown (Calmar ratio) for compounded cumulative returns, MDD_r.

Average drawdown for compounded cumulative returns, ADD_r.

Ulcer index for compounded cumulative returns, UCI_r.

Drawdown at for compounded cumulative returns risk (DaR), DaR_r.

Conditional drawdown at risk for compounded cumulative returns (CDaR), CDaR_r.

Entropic drawdown at risk for compounded cumulative returns (EDaR), EDaR_r.

Relativistic drawdown at risk for compounded cumulative returns (RLDaR), RLDaR_r.

Linear moments (L-moments)

L-moment ordered weight array, OWA.

Equal Risk Contribution

Equal risk contribution, Equal.

Nested clustered optimisation, NCO

Nested clustered optimisation combines the ideas of hierarchical equal risk parity optimisations and portfolio optimisations. They use the hierarchical clustering relationships between assets and splitting the dendrogram into k clusters. It then treats each cluster as its own isntance of Portfolio which is optimised in the usual way. The weights of each cluster are saved in a matrix, these are the intra-cluster weights. Then each cluster as a whole is treated as a synthetic asset, it statistics are internally computed from the fields in the NCO type. An instance of Portfolio is created from these synthetic assets and then optimised, these are the inter-cluster weights. The inter-cluster and intra-cluster weights are multiplied to give the asset weights.

NCO can make use of two risk measure arguments, one for the intra-cluster and one for the inter-cluster risk calculation. They can take linear combinations of risk measures.

Sub-types

NCO can take keyword arguments that define the supported by optimisations of Portfolio. Since there are intra- and inter-cluster optimisations, it can take individual arguments for both. This means it can perform any combination of Portfolio optimisations.

NCO-Trad, NCO-NOC-Trad

Objective functions

Minimum risk, MinRisk.

Maximum utility, Utility.

Maximum risk adjusted return ratio, Sharpe.

Maximum return, MaxRet.

Constraints

When applied to the intra-cluster optimisation the same constraint will be applied to all every cluster.

Maximum expected risk constraints.

Minimum expected return constraint.

Leverage constraints.

NCO-RB, NCO-RRB, NCO-NOC-RB, NCO-NOC-RRB

Constraints

When applied to the intra-cluster optimisation the same constraint will be applied to all every cluster.

Minimum expected return constraint.