Definitions

PortfolioOptimiser.AbstractPortfolio — Type

AbstractPortfolio

Abstract type for portfolios. Concrete portfolios subtype this see Portfolio and HCPortfolio.

PortfolioOptimiser.RiskMeasures — Constant

RiskMeasures = (
    :SD,
    :MAD,
    :SSD,
    :FLPM,
    :SLPM,
    :WR,
    :CVaR,
    :EVaR,
    :RVaR,
    :MDD,
    :ADD,
    :CDaR,
    :UCI,
    :EDaR,
    :RDaR,
    :Kurt,
    :SKurt,
    :GMD,
    :RG,
    :RCVaR,
    :TG,
    :RTG,
    :OWA,
)

Available risk measures for :trad and :rp type (see PortTypes) of Portfolio.

:SD = standard deviation (SD);

:MAD = max absolute deviation (MAD);

:SSD = semi standard deviation (SSD);

:FLPM = first lower partial moment (omega ratio) (FLPM);

:SLPM = second lower partial moment (sortino ratio) (SLPM);

:WR = worst realisation (WR);

:CVaR = conditional value at risk (CVaR);

:EVaR = entropic value at risk (EVaR);

:RVaR = relativistic value at risk (RVaR);

:MDD = maximum drawdown of uncompounded cumulative returns (MDD_abs);

:ADD = average drawdown of uncompounded cumulative returns (ADD_abs);

:CDaR = conditional drawdown at risk of uncompounded cumulative returns (CDaR_abs);

:UCI = ulcer index of uncompounded cumulative returns (UCI_abs);

:EDaR = entropic drawdown at risk of uncompounded cumulative returns (EDaR_abs);

:RDaR = relativistic drawdown at risk of uncompounded cumulative returns (RDaR_abs);

:Kurt = square root kurtosis (Kurt);

:SKurt = square root semi-kurtosis (SKurt);

:GMD = gini mean difference (GMD);

:RG = range of returns (RG);

:RCVaR = range of conditional value at risk (RCVaR);

:TG = tail gini (TG);

:RTG = range of tail gini (RTG);

:OWA = ordered weight array (generic OWA weights) (OWA).

PortfolioOptimiser.KellyRet — Constant

KellyRet = (:none, :approx, :exact)

Available types of Kelly returns for Portfolio.

:none: arithmetic mean return, $R(\bm{w}) = \bm{\mu} \cdot \bm{w}$;

:approx: first moment approximation of the logarithmic returns, $R(\bm{w}) = \bm{\mu} \cdot \bm{w} - \dfrac{1}{2} \bm{w}^{\intercal} \mathbf{\Sigma} \bm{w}$;

:exact: exact logarithmic returns, $R(\bm{w}) = \dfrac{1}{T}\sum\limits_{t=1}^{T}\ln\left(1 + \bm{x}_t \cdot \bm{w}\right)$.

Where:

$\mathbf{\Sigma}$ is the covariance matrix of the asset returns;

$\bm{x}_t$ is the vector of asset returns at timestep $t$;

$\bm{\mu}$ is the vector of expected returns for each asset over the timespan;

and $\bm{w}$ is the asset weights vector.

PortfolioOptimiser.TrackingErrKinds — Constant

TrackingErrKinds = (:weights, :returns)

Available kinds of tracking errors for Portfolio.

:weights: provide a vector of asset weights which is used to compute the vector of benchmark returns;

:returns: directly provide the vector of benchmark returns.

The benchmark is then used as a reference to optimise a portfolio that tracks it up to a given error.

PortfolioOptimiser.ObjFuncs — Constant

ObjFuncs = (:min_risk, :utility, :sharpe, :max_ret)

Available objective functions for the :trad type (see PortTypes) of Portfolio.

:min_risk: minimum risk portfolio,

\[\begin{align*} \underset{\bm{w}}{\min} &\qquad \phi_{j}(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\ &\qquad R(\bm{w}) \geq \hat{\mu} \end{align*}\,.\]

:utility: maximum utility portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad R(\bm{w}) - \lambda \phi_{j}(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\ &\qquad R(\bm{w}) \geq \hat{\mu} \end{align*}\,.\]

:sharpe: maximum risk-adjusted return ratio portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad \dfrac{R(\bm{w}) - r}{\phi_{j}(\bm{w})} \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\ &\qquad R(\bm{w}) \geq \hat{\mu} \end{align*}\,.\]

:max_ret: maximum return portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad R(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\ &\qquad R(\bm{w}) \geq \hat{\mu} \end{align*}\,.\]

Where:

$\bm{w}$ are the asset weights;

$\phi_{i}$ is risk measure $i$ from the set of available risk measures $\left\{\Phi\right\}$ (see RiskMeasures);

$\mathbf{A} \bm{w} \geq \mathbf{B}$ are the asset linear constraints;

$c_{i}$ is the maximum acceptable value for risk measure $\phi_{i}$ of the optimised portfolio;

$R(\bm{w})$ is the return function from KellyRet;

$\hat{\mu}$ is the minimum acceptable return of the optimised portfolio;

$\lambda$ is the risk aversion coefficient;

and $r$ is the risk free rate.

PortfolioOptimiser.ValidTermination — Constant

ValidTermination =
(MOI.OPTIMAL, MOI.ALMOST_OPTIMAL, MOI.LOCALLY_SOLVED, MOI.ALMOST_LOCALLY_SOLVED)

Valid JuMP termination codes after optimising an instance of Portfolio. If the termination code is different to these, then the failures are logged in the .fail field of HCPortfolio and Portfolio.

PortfolioOptimiser.PortClasses — Constant

PortClasses = (:classic,)

Available classes for Portfolio.

PortfolioOptimiser.PortTypes — Constant

PortTypes = (:trad, :rp, :rrp, :wc)

Available optimisation types for Portfolio.

:trad – Traditional

Optimisations for which ObjFuncs and RiskMeasures apply.

:rp – Risk Parity

Optimisations for which RiskMeasures apply. Optimises portfolios based on a vector of risk contributions per asset.

\[\begin{align*} \underset{\bm{w}}{\min} &\qquad \phi(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \bm{b} \cdot \ln(\bm{w}) \geq c \\ &\qquad R(\bm{w}) \geq \hat{\mu}\\ &\qquad \bm{w} \geq \bm{0} \end{align*}\,.\]

Where:

$\bm{w}$ are the asset weights;

$\phi$ a risk measure from the set of available risk measures (see RiskMeasures);

$\mathbf{A} \bm{w} \geq \mathbf{B}$ are the asset linear constraints;

$\bm{b}$ is the vector of maximum allowable risk contribution per asset to the optimised portfolio;

$c$ is an auxiliary variable;

$R(\bm{w})$ is the return function from KellyRet;

and $\hat{\mu}$ is the minimum acceptable return of the optimised portfolio.

:rrp – Relaxed Risk Parity

Defines the risk measure internally, which is derived from the portfolio covariance. Optimises portfolios based on a vector of risk contributions per asset.

\[\begin{align*} \underset{\bm{w}}{\min} &\qquad \psi - \gamma \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \bm{w}^\intercal \, \mathbf{\Sigma}\, \bm{w} \leq (\psi^{2} - \rho^{2})\\ &\qquad \lambda \bm{w}^\intercal \, \mathbf{\Theta}\, \bm{w} \leq \rho^{2}\\ &\qquad \bm{\zeta} = \mathbf{\Sigma} \bm{w}\\ &\qquad w_{i} \zeta_{i} \geq \gamma^{2} b_{i} \qquad \forall \, i = 1,\, \ldots{},\, N\\ &\qquad R(\bm{w}) \geq \hat{\mu}\\ &\qquad \sum\limits_{i=1}^{N} w_{i} = 1\\ &\qquad \bm{w} \geq \bm{0}\\ &\qquad \psi,\, \gamma,\, \rho \geq 0 \end{align*}\,.\]

Where:

$\bm{w}$ are the asset weights;

$\psi$ is the average risk of the portfolio;

$\gamma$ is the lower bound of the risk contribution for each asset;

$\mathbf{A} \bm{w} \geq \mathbf{B}$ are the asset linear constraints;

$\mathbf{\Sigma}$ is the portfolio covariance;

$\rho$ is a regularisation variable;

$\mathbf{\Theta} = \mathrm{diag}\left(\mathbf{\Sigma}\right)$ ;

$\lambda$ is a penalty parameter for $\rho$;

$\bm{\zeta}$ is the vector of marginal risk for each asset;

$b_{i}$ is the maximum allowable risk contribution for asset $i$;

$N$ is the number of assets;

$R(\bm{w})$ is the return function from KellyRet;

and $\hat{\mu}$ is the minimum acceptable return of the optimised portfolio.

PortfolioOptimiser.RRPVersions — Constant

RRPVersions = (:none, :reg, :reg_pen)

Available versions of relaxed risk parity optimisations of Portfolio.

PortfolioOptimiser.RPConstraintTypes — Constant

RPConstraintTypes = (:assets, :classes)

Types of risk parity constraints for rp_constraints.

PortfolioOptimiser.UncertaintyTypes — Constant

UncertaintyTypes = (:none, :box, :ellipse)

Types of uncertainty sets for worst case optimisations of Portfolio.

PortfolioOptimiser.KindBootstrap — Constant

Bootstrap for worst case optimisations.

KindBootstrap = (:stationary, :circular, :moving)

PortfolioOptimiser.EllipseTypes — Constant

Ellipse and box types for worst case optimisations.

EllipseTypes = (:stationary, :circular, :moving, :normal)

PortfolioOptimiser.HRRiskMeasures — Constant

HRRiskMeasures = (
    :SD,
    :MAD,
    :SSD,
    :FLPM,
    :SLPM,
    :WR,
    :CVaR,
    :EVaR,
    :RVaR,
    :MDD,
    :ADD,
    :CDaR,
    :UCI,
    :EDaR,
    :RDaR,
    :Kurt,
    :SKurt,
    :GMD,
    :RG,
    :RCVaR,
    :TG,
    :RTG,
    :OWA,
    :Variance,
    :Equal,
    :VaR,
    :DaR,
    :DaR_r,
    :MDD_r,
    :ADD_r,
    :CDaR_r,
    :EDaR_r,
    :RDaR_r,
)

Available risk measures for optimisations of HCPortfolio.

:SD = standard deviation (SD);

:MAD = max absolute deviation (MAD);

:SSD = semi standard deviation (SSD);

:FLPM = first lower partial moment (Omega ratio) (FLPM);

:SLPM = second lower partial moment (Sortino ratio) (SLPM);

:WR = worst realisation (WR);

:CVaR = conditional value at risk (CVaR);

:EVaR = entropic value at risk (EVaR);

:RVaR = relativistic value at risk (RVaR);

:MDD = maximum drawdown of uncompounded cumulative returns (Calmar ratio) (MDD_abs);

:ADD = average drawdown of uncompounded cumulative returns (ADD_abs);

:CDaR = conditional drawdown at risk of uncompounded cumulative returns (CDaR_abs);

:UCI = ulcer index of uncompounded cumulative returns (UCI_abs);

:EDaR = entropic drawdown at risk of uncompounded cumulative returns (EDaR_abs);

:RDaR = relativistic drawdown at risk of uncompounded cumulative returns (RDaR_abs);

:Kurt = square root kurtosis (Kurt);

:SKurt = square root semi-kurtosis (SKurt);

:GMD = gini mean difference (GMD);

:RG = range of returns (RG);

:RCVaR = range of conditional value at risk (RCVaR);

:TG = tail gini (TG);

:RTG = range of tail gini (RTG);

:OWA = ordered weight array (generic OWA weights) (OWA);

:Variance = variance (Variance);

:Equal = equal risk contribution, 1/N where N is the number of assets;

:VaR = value at risk (VaR);

:DaR = drawdown at risk of uncompounded cumulative returns (DaR_abs);

:DaR_r = drawdown at risk of compounded cumulative returns (DaR_rel);

:MDDr = maximum drawdown of compounded cumulative returns ([`MDDrel`](@ref));

:ADDr = average drawdown of compounded cumulative returns ([`ADDrel`](@ref));

:CDaR_r = conditional drawdown at risk of compounded cumulative returns (CDaR_rel);

:UCIr = ulcer index of compounded cumulative returns ([`UCIrel`](@ref));

:EDaR_r = entropic drawdown at risk of compounded cumulative returns (EDaR_rel);

:RDaR_r = relativistic drawdown at risk of compounded cumulative returns (RDaR_rel).

```