Definitions

PortfolioOptimiser.AbstractPortfolio — Type

AbstractPortfolio

Abstract type for portfolios. Concrete portfolios subtype this see Portfolio and HCPortfolio.

PortfolioOptimiser.RiskMeasures — Constant

RiskMeasures = (
    :SD,
    :MAD,
    :SSD,
    :FLPM,
    :SLPM,
    :WR,
    :CVaR,
    :EVaR,
    :RVaR,
    :MDD,
    :ADD,
    :CDaR,
    :UCI,
    :EDaR,
    :RDaR,
    :Kurt,
    :SKurt,
    :GMD,
    :RG,
    :RCVaR,
    :TG,
    :RTG,
    :OWA,
)

Available risk measures for type = :trad optimisations of Portfolio.

:SD = standard deviation (SD);

:MAD = max absolute deviation (MAD);

:SSD = semi standard deviation (SSD);

:FLPM = first lower partial moment (omega ratio) (FLPM);

:SLPM = second lower partial moment (sortino ratio) (SLPM);

:WR = worst realisation (WR);

:CVaR = conditional value at risk (CVaR);

:EVaR = entropic value at risk (EVaR);

:RVaR = relativistic value at risk (RVaR);

:MDD = maximum drawdown for uncompounded cummulative returns (MDD_abs);

:ADD = average drawdown for uncompounded cummulative returns (ADD_abs);

:CDaR = conditional drawdown at risk for uncompounded cummulative returns (CDaR_abs);

:UCI = ulcer index for uncompounded cummulative returns (UCI_abs);

:EDaR = entropic drawdown at risk for uncompounded cummulative returns (EDaR_abs);

:RDaR = relativistic drawdown at risk for uncompounded cummulative returns (RDaR_abs);

:Kurt = square root kurtosis (Kurt);

:SKurt = square root semi-kurtosis (SKurt);

:GMD = gini mean difference (GMD);

:RG = range of returns (RG);

:RCVaR = range of conditional value at risk (RCVaR);

:TG = tail gini (TG);

:RTG = range of tail gini (RTG);

:OWA = ordered weight array (generic OWA weights) (OWA).

PortfolioOptimiser.KellyRet — Constant

const KellyRet = (:none, :approx, :exact)

Available types of Kelly returns for Portfolio.

:none: arithmetic mean return, $\bm{\mu} \cdot \bm{w}$;

:approx: first moment approximation of the logarithmic returns, $\bm{\mu} \cdot \bm{w} - \dfrac{1}{2} \bm{w}^{\intercal} \mathbf{\Sigma} \bm{w}$;

:exact: exact logarithmic returns, $\dfrac{1}{T}\sum\limits_{t=1}^{T}\ln\left(1 + \bm{x}_t \cdot \bm{w}\right)$.

Where $\mathbf{\Sigma}$ is the covariance matrix of the asset returns; $\bm{x}_t$ is the vector of asset returns at timestep $t$; $\bm{\mu}$ is the vector of expected returns for each asset over the timespan; and $\bm{w}$ is the asset weights vector.

PortfolioOptimiser.TrackingErrKinds — Constant

const TrackingErrKinds = (:weights, :returns)

Available kinds of tracking errors for Portfolio.

:weights: provide a vector of asset weights which is used to compute the vector of benchmark returns;

:returns: directly provide the vector of benchmark returns.

The benchmark is then used as a reference to optimise a portfolio that tracks it up to a given error.

PortfolioOptimiser.ObjFuncs — Constant

const ObjFuncs = (:min_risk, :utility, :sharpe, :max_ret)

Available objective functions for Portfolio.

:min_risk: minimum risk portfolio,

\[\begin{align*} \underset{\bm{w}}{\min} &\qquad \phi_{j}(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \\ &\qquad R(\bm{w}) \geq \hat{\mu} \end{align*}\,.\]

:utility: maximum utility portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad R(\bm{w}) - \lambda \phi_{j}(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \\ &\qquad R(\bm{w}) \geq \hat{\mu} \end{align*}\,.\]

:sharpe: maximum risk-adjusted return ratio portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad \dfrac{R(\bm{w}) - r}{\phi_{j}(\bm{w})} \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \\ &\qquad R(\bm{w}) \geq \hat{\mu} \end{align*}\,.\]

:max_ret: maximum return portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad R(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\ &\qquad R(\bm{w}) \geq \hat{\mu} \end{align*}\,.\]

Where $\phi_{i}$ is a risk function from the set of available risk functions $\left\{\Phi\right\}$ from RiskMeasures; $c_{i}$ is the maximum acceptable value for the risk function $\phi_{i}$ of the optimised portfolio; $R(\bm{w})$ is the return function from KellyRet; $\hat{\mu}$ is the minimum acceptable value for the return of the optimised portfolio; $r$ is the risk free rate; $\lambda$ is the risk aversion coefficient; and $\mathbf{A} \bm{w} \geq \mathbf{B}$ are the asset linear constraints.

PortfolioOptimiser.ValidTermination — Constant

const ValidTermination =
(MOI.OPTIMAL, MOI.ALMOST_OPTIMAL, MOI.LOCALLY_SOLVED, MOI.ALMOST_LOCALLY_SOLVED)

Valid JuMP terminations after optimising an instance of Portfolio.

PortfolioOptimiser.PortClasses — Constant

const PortClasses = (:classic,)

Available classes for Portfolio.

PortfolioOptimiser.PortTypes — Constant

const PortTypes = (:trad, :rp, :rrp, :OWA, :wc)

Available optimisation types for Portfolio.

PortfolioOptimiser.RRPVersions — Constant

const RRPVersions = (:none, :reg, :reg_pen)

Available versions of relaxed risk parity optimisations of Portfolio.

PortfolioOptimiser.RPConstraintTypes — Constant

const RPConstraintTypes = (:assets, :classes)

Types of risk parity constraints for rp_constraints.

PortfolioOptimiser.UncertaintyTypes — Constant

const UncertaintyTypes = (:none, :box, :ellipse)

Types of uncertainty sets for worst case optimisations of Portfolio.

PortfolioOptimiser.KindBootstrap — Constant

Bootstrap for worst case optimisations.

const KindBootstrap = (:stationary, :circular, :moving)

PortfolioOptimiser.EllipseTypes — Constant

Ellipse and box types for worst case optimisations.

const EllipseTypes = (:stationary, :circular, :moving, :normal)
const BoxTypes = (EllipseTypes..., :delta)