Constant Definitions

Many functions only accept specific methods and/or require certain arguments to be within appropriate bounds. These constants ensure such requirements are met.

PortfolioOptimiser.KellyRet — Constant

KellyRet = (:None, :Approx, :Exact)

Available types of Kelly returns for Portfolio.

:None: arithmetic mean return, $R(\bm{w}) = \bm{\mu} \cdot \bm{w}$.

:Approx: first moment approximation of the logarithmic returns, $R(\bm{w}) = \bm{\mu} \cdot \bm{w} - \dfrac{1}{2} \bm{w}^{\intercal} \mathbf{\Sigma} \bm{w}$.

:Exact: exact logarithmic returns, $R(\bm{w}) = \dfrac{1}{T}\sum\limits_{t=1}^{T}\ln\left(1 + \bm{x}_t \cdot \bm{w}\right)$. Where:

$\mathbf{\Sigma}$ is the covariance matrix of the asset returns.

$\bm{x}_t$ is the vector of asset returns at timestep $t$.

$\bm{\mu}$ is the vector of expected returns for each asset.

and $\bm{w}$ is the asset weights vector.

PortfolioOptimiser.ObjFuncs — Constant

ObjFuncs = (:Min_Risk, :Utility, :Sharpe, :Max_Ret)

Objective functions available for use in :Trad and :WC optimisations of Portfolio (see PortTypes).

PortfolioOptimiser.ValidTermination — Constant

ValidTermination = (MOI.OPTIMAL, MOI.ALMOST_OPTIMAL, MOI.LOCALLY_SOLVED,
                    MOI.ALMOST_LOCALLY_SOLVED, MOI.SOLUTION_LIMIT, MOI.OBJECTIVE_LIMIT)

Valid JuMP termination codes after optimising an instance of Portfolio. If the termination code is different to these, then the failures are logged in the .fail field of HCPortfolio and Portfolio.

PortfolioOptimiser.PortClasses — Constant

PortClasses = (:Classic, :FM, :BL, :BLFM)

Available choicees of summary parameters $\bm{\mu}$ and $\bm{\Sigma}$ Portfolio.

:Classic: when optimising with this option, $\bm{\mu}$ and $\bm{\Sigma}$ take their values from historical estimates computed by asset_statistics!.

:FM: when optimising with this option, $\bm{\mu}$ and $\bm{\Sigma}$ take their values from the factor model computed by factor_statistics!.

:BL: when optimising with this option, $\bm{\mu}$ and $\bm{\Sigma}$ take their values from the Black-Litterman model computed by black_litterman_statistics!.

:BLFM: when optimising with this option, $\bm{\mu}$ and $\bm{\Sigma}$ take their values from the factor Black-Litterman model computed by black_litterman_factor_satistics!. This model has two versions defined in BLFMMethods.

PortfolioOptimiser.PortTypes — Constant

PortTypes = (:Trad, :RP, :RRP, :WC)

Available optimisation types for Portfolio.

:Trad – Traditional Optimisations

Available objective functions for :Trad optimisations. We can chose any of the objective functions in ObjFuncs and risk measures in RiskMeasures.

:Min_Risk: minimum risk portfolio,```math \begin{align} \underset{\bm{w}}{\min} &\qquad \phi{j}(\bm{w}) \ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B} \ &\qquad \phi{i}(\bm{w}) \leq c{i} \, \forall \, \phi{i} \in \left{\Phi\right} \ &\qquad R(\bm{w}) \geq \overline{\mu} \end{align}\,.

- `:Utility`: maximum utility portfolio,```math
\begin{align*}
\underset{\bm{w}}{\max} &\qquad R(\bm{w}) - \lambda \phi_{j}(\bm{w}) \\
\mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B} \\
&\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\
&\qquad R(\bm{w}) \geq \overline{\mu}
\end{align*}\,.

:Sharpe: maximum risk-adjusted return ratio portfolio,```math \begin{align} \underset{\bm{w}}{\max} &\qquad \dfrac{R(\bm{w}) - r}{\phi{j}(\bm{w})} \ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B} \ &\qquad \phi{i}(\bm{w}) \leq c{i} \, \forall \, \phi{i} \in \left{\Phi\right} \ &\qquad R(\bm{w}) \geq \overline{\mu} \end{align}\,.

- `:Max_Ret`: maximum return portfolio,```math
\begin{align*}
\underset{\bm{w}}{\max} &\qquad R(\bm{w}) \\
\mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B} \\
&\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\
&\qquad R(\bm{w}) \geq \overline{\mu}
\end{align*}\,.

Where:

$\bm{w}$ are the asset weights.

$\phi_{i}$ is risk measure $i$ from the set of available risk measures $\left\{\Phi\right\}$ (see RiskMeasures).

$\mathbf{A} \bm{w} \geq \bm{B}$ is a set of linear constraints.

$c_{i}$ is the maximum acceptable value for risk measure $\phi_{i}$ of the optimised portfolio.

$R(\bm{w})$ is the return function from KellyRet.

$\overline{\mu}$ is the minimum acceptable return of the optimised portfolio.

$\lambda$ is the risk aversion coefficient.

and $r$ is the risk-free rate.

:RP – Risk Parity Optimisations

Optimises portfolios based on a vector of risk contributions per asset. We can chose any of the risk measures in RiskMeasures.

\[\begin{align*} \underset{\bm{w}}{\min} &\qquad \phi(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B} \\ &\qquad \bm{b} \cdot \ln(\bm{w}) \geq c \\ &\qquad R(\bm{w}) \geq \overline{\mu}\\ &\qquad \bm{w} \geq \bm{0} \end{align*}\,.\]

Where:

$\bm{w}$ are the asset weights.

$\phi$ a risk measure from the set of available risk measures (see RiskMeasures).

$\mathbf{A} \bm{w} \geq \bm{B}$ is a set of linear constraints.

$\bm{b}$ is the vector of maximum allowable risk contribution per asset to the optimised portfolio.

$c$ is an auxiliary variable.

$R(\bm{w})$ is the return function from KellyRet.

and $\overline{\mu}$ is the minimum acceptable return of the optimised portfolio.

:RRP – Relaxed Risk Parity Optimisations

Optimises portfolios based on a vector of risk contributions per asset. Defines its own risk measure using the portfolio returns covariance.

\[\begin{align*} \underset{\bm{w}}{\min} &\qquad \psi - \gamma \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B} \\ &\qquad \bm{w}^\intercal \, \mathbf{\Sigma}\, \bm{w} \leq (\psi^{2} - \rho^{2})\\ &\qquad \lambda \bm{w}^\intercal \, \mathbf{\Theta}\, \bm{w} \leq \rho^{2}\\ &\qquad \bm{\zeta} = \mathbf{\Sigma} \bm{w}\\ &\qquad w_{i} \zeta_{i} \geq \gamma^{2} b_{i} \qquad \forall \, i = 1,\, \ldots{},\, N\\ &\qquad R(\bm{w}) \geq \overline{\mu}\\ &\qquad \sum\limits_{i=1}^{N} w_{i} = 1\\ &\qquad \bm{w} \geq \bm{0}\\ &\qquad \psi,\, \gamma,\, \rho \geq 0 \end{align*}\,.\]

Where:

$\bm{w}$ are the asset weights.

$\psi$ is the average risk of the portfolio.

$\gamma$ is the lower bound of the risk contribution for each asset.

$\mathbf{A} \bm{w} \geq \bm{B}$ is a set of linear constraints.

$\mathbf{\Sigma}$ is the portfolio covariance.

$\rho$ is a regularisation variable.

$\mathbf{\Theta} = \mathrm{diag}\left(\mathbf{\Sigma}\right)$ .

$\lambda$ is a penalty parameter for $\rho$, taken from the available choices in RRPVersions.

$\bm{\zeta}$ is the vector of marginal risk for each asset.

$b_{i}$ is the maximum allowable risk contribution for asset $i$.

$N$ is the number of assets.

$R(\bm{w})$ is the return function from KellyRet.

and $\overline{\mu}$ is the minimum acceptable return of the optimised portfolio.

:WC – Worst Case Mean Variance Optimisations

Computes the worst case mean variance portfolio according to user-selected uncertainty sets (see UncertaintyTypes) for the portfolio return and covariance. We can chose any of the objective functions in ObjFuncs.

:Min_Risk: worst case minimum risk mean-variance portfolio,```math \begin{align} \underset{\bm{w}}{\max} &\qquad \underset{\mathbf{\Sigma}\, \in\, U_{\mathbf{\Sigma}}}{\max}\, \bm{w}^{\intercal}\, \mathbf{\Sigma}\, \bm{w}\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B}\,. \end{align}

- `:Utility`: worst case maximum utility mean-variance portfolio,```math
\begin{align*}
\underset{\bm{w}}{\max} &\qquad \underset{\bm{\mu}\, \in\, U_{\bm{\mu}}}{\min} R(\bm{w})\, -\, \underset{\mathbf{\Sigma}\, \in\, U_{\mathbf{\Sigma}}}{\max}\, \lambda \bm{w}^{\intercal}\, \mathbf{\Sigma}\, \bm{w}\\
\mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B}\,.
\end{align*}

:Sharpe: worst case maximum risk-adjusted return ratio mean-variance portfolio,```math \begin{align} \underset{\bm{w}}{\max} &\qquad \dfrac{\underset{\bm{\mu}\, \in\, U{\bm{\mu}}}{\min} R(\bm{w}) - r}{\underset{\mathbf{\Sigma}\, \in\, U{\mathbf{\Sigma}}}{\max}\, \left(\bm{w}^{\intercal}\, \mathbf{\Sigma}\, \bm{w}\right)^{1/2}} \ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B}\,. \end{align}

- `:Max_Ret`: worst case maximum return mean-variance portfolio,```math
\begin{align*}
\underset{\bm{w}}{\max} &\qquad \underset{\bm{\mu}\, \in\, U_{\bm{\mu}}}{\min} R(\bm{w})\\
\mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \bm{B}\,.
\end{align*}

Where:

$\bm{w}$ are the asset weights.

$\mathbf{\Sigma}$ is the covariance matrix of asset returns.

$U_{\Sigma}$ is the uncertainty set for the covariance matrix, they can be:

\[\begin{align*} U_{\Sigma}^{\mathrm{box}} &= \left\{\mathbf{\Sigma}\, \vert\, \mathbf{\Sigma}_{l} \leq \mathbf{\Sigma} \leq \mathbf{\Sigma}_{u},\, \mathbf{\Sigma} \succeq 0\right\}\\ U_{\Sigma}^{\mathrm{ellipse}} &= \left\{\mathbf{\Sigma}\, \vert\, \left[\mathrm{vec}\left(\mathbf{\Sigma}\right) - \mathrm{vec}\left(\hat{\mathbf{\Sigma}}\right)\right] \mathbf{\Sigma}_{\mathbf{\Sigma}}^{-1} \left[\mathrm{vec}\left(\mathbf{\Sigma}\right) - \mathrm{vec}\left(\hat{\mathbf{\Sigma}}\right)\right]^{\intercal} \leq k_{\mathbf{\Sigma}}^2 ,\, \mathbf{\Sigma} \succeq 0\right\}\,. \end{align*}\]

Where the following variables are estimated by assuming that the portfolio's asset return covariance can be generated by some matrix distribution. The distribution is sampled, and the estimates are calculated from them. Available choices can be found in BoxMethods and EllipseMethods for the box and ellipse sets respectively:

the $\mathrm{l}$ and $\mathrm{u}$ subscripts denote lower and upper bounds for the covariance matrix given the samples.

$\mathbf{\Sigma}_{\mathbf{\Sigma}}$ is the covariance of the samples.

$\hat{\mathbf{\Sigma}}$ the expected covariance given the samples.

and $k_{\mathbf{\Sigma}}$ is a significance parameter of the matrix distribution.

$\mathbf{A} \bm{w} \geq \bm{B}$ is a set of linear constraints.

$\bm{\mu}$ is the vector of expected returns for each asset.

$R(\bm{w})$ is the return function from KellyRet.

$U_{\bm{\mu}}$ is the uncertainty set for the asset returns, they can be:

\[\begin{align*} U_{\bm{\mu}}^{\mathrm{box}} &= \left\{\bm{\mu}\, \vert\, \vert \bm{\mu} - \bm{\hat{\mu}} \vert \leq \delta \right\}\\ U_{\bm{\mu}}^{\mathrm{ellipse}} &= \left\{\bm{\mu}\, \vert\, \left(\bm{\mu} - \bm{\hat{\mu}}\right) \mathbf{\Sigma}_{\bm{\mu}}^{-1} \left(\bm{\mu} - \bm{\hat{\mu}}\right)^{\intercal} \leq k_{\bm{\mu}}^{2}\right\}\,. \end{align*}\]

Where the following variables are estimated by assuming that the portfolio's asset mean returns can be generated by some distribution. The distribution is sampled, and the estimates are calculated from them. Available choices can be found in BoxMethods and EllipseMethods for the box and ellipse sets respectively:

$\hat{\bm{\mu}}$ the expected portfolio asset mean returns given the samples.

$\mathbf{\Sigma}_{\bm{\mu}}$ is the covariance of the samples.

and $k_{\bm{\mu}}$ is a significance parameter of the distribution.

$\lambda$ is the risk aversion coefficient.

and $r$ is the risk-free rate.

The worst case uncertainty sets are computed by wc_statistics!.

PortfolioOptimiser.HCPortTypes — Constant

HCPortTypes = (:HRP, :HERC, :NCO)

Available optimisation types for HCPortfolio.

:HRP: Hierarchical Risk Parity.

:HERC: Hierarhical Equal Risk Contribution.

:NCO: Nested Clustered Optimisation. Both :HERC and :NCO split their optimisations into two parts:

inter-cluster optimisation.

intra-cluster optimisation. Threfore they can make use of extra parameters:

:HERC: accepts an extra risk measure rm_i and OWA weights owa_w_i parameters, which are used for the intra-cluster optimisations. They default to the same value as their external counterparts.

:NCO: accepts an extra objective function obj_i kelly return kelly_i risk aversion parameter l_i risk measure rm_i and OWA weights owa_w_i parameters. They default to the same value as their external counterparts.

PortfolioOptimiser.LinkageTypes — Constant

LinkageTypes = (:single, :complete, :average, :ward, :ward_presquared, :DBHT)

Linkage types available when optimising a HCPortfolio.

:DBHT: is Direct Bubble Hierarchical Tree clustering.

The rest are linkage types supported by Clustering.hclust.

PortfolioOptimiser.BranchOrderTypes — Constant

BranchOrderTypes = (:optimal, :barjoseph, :r, :default)

Choice of algorithm for ordering hierarchical clustering dendrogram leaves and branches.

:default: if linkage is :DBHT, the leaves and branches remain as the algorithm orders them. If any other linkage is used, they fall back to :r as that is their default according to Clustering.hclust.

All other branch orders are as defined by Clustering.hclust.

PortfolioOptimiser.HCObjFuncs — Constant

HCObjFuncs = (:Min_Risk, :Utility, :Sharpe, :Max_Ret, :Equal)

Objective funcions for :NCO HCPortTypes of HCPortfolio.

:Min_Risk, :Utility, :Sharpe and :Max_Ret: optimise the sub-portfolios as :Trad PortTypes of Portfolio according to their respective definitions in ObjFuncs.

:Equal: optimises the sub-portfolios as :RP optimisations with equal risk contribution per asset/cluster. We can't offer customiseable risk contributions because the size and composition of the clusters is initially unknown and depends on the chosen linkage method.

PortfolioOptimiser.BLHist — Constant

BLHist = (1, 2, 3)

Choice of what estimate of $\mathbf{\Sigma}$ to use. Factor models, :FM and :BLFM, can also use a factor model estimate of the returns matrix $\mathbf{X}$.

1:

:FM: uses the factor model estimates of $\mathbf{\Sigma}$ and $\mathbf{X}$.

:BL: uses the Black Litterman estimate of $\mathbf{\Sigma}$.

:BLFM: uses the Black Litterman factor model estimate of $\mathbf{\Sigma}$ and factor model estimate of $\mathbf{X}$.

2:

:FM: uses the standard estimates of $\mathbf{\Sigma}$ and $\mathbf{X}$.

:BL: uses the standard estimate of $\mathbf{\Sigma}$.

:BLFM: uses the standard estimates of $\mathbf{\Sigma}$ and $\mathbf{X}$.

3

:FM and :BL do not support this option.

:BLFM: uses the factor model estimates of $\mathbf{\Sigma}$ and $\mathbf{X}$.