Definitions

PortfolioOptimiser.AbstractPortfolio — Type

AbstractPortfolio

Abstract type for portfolios. Concrete portfolios subtype this see Portfolio and HCPortfolio.

PortfolioOptimiser.RiskMeasures — Constant

RiskMeasures = (
    :SD,
    :MAD,
    :SSD,
    :FLPM,
    :SLPM,
    :WR,
    :CVaR,
    :EVaR,
    :RVaR,
    :MDD,
    :ADD,
    :CDaR,
    :UCI,
    :EDaR,
    :RDaR,
    :Kurt,
    :SKurt,
    :GMD,
    :RG,
    :RCVaR,
    :TG,
    :RTG,
    :OWA,
)

Available risk measures for :Trad and :RP type (see PortTypes) of Portfolio.

:SD = standard deviation (SD);

:MAD = max absolute deviation (MAD);

:SSD = semi standard deviation (SSD);

:FLPM = first lower partial moment (omega ratio) (FLPM);

:SLPM = second lower partial moment (sortino ratio) (SLPM);

:WR = worst realisation (WR);

:CVaR = conditional value at risk (CVaR);

:EVaR = entropic value at risk (EVaR);

:RVaR = relativistic value at risk (RVaR);

:MDD = maximum drawdown of uncompounded cumulative returns (MDD_abs);

:ADD = average drawdown of uncompounded cumulative returns (ADD_abs);

:CDaR = conditional drawdown at risk of uncompounded cumulative returns (CDaR_abs);

:UCI = ulcer index of uncompounded cumulative returns (UCI_abs);

:EDaR = entropic drawdown at risk of uncompounded cumulative returns (EDaR_abs);

:RDaR = relativistic drawdown at risk of uncompounded cumulative returns (RDaR_abs);

:Kurt = square root kurtosis (Kurt);

:SKurt = square root semi-kurtosis (SKurt);

:GMD = gini mean difference (GMD);

:RG = range of returns (RG);

:RCVaR = range of conditional value at risk (RCVaR);

:TG = tail gini (TG);

:RTG = range of tail gini (RTG);

:OWA = ordered weight array (generic OWA weights) (OWA).

PortfolioOptimiser.KellyRet — Constant

KellyRet = (:None, :Approx, :Exact)

Available types of Kelly returns for Portfolio.

:None: arithmetic mean return, $R(\bm{w}) = \bm{\mu} \cdot \bm{w}$;

:Approx: first moment approximation of the logarithmic returns, $R(\bm{w}) = \bm{\mu} \cdot \bm{w} - \dfrac{1}{2} \bm{w}^{\intercal} \mathbf{\Sigma} \bm{w}$;

:Exact: exact logarithmic returns, $R(\bm{w}) = \dfrac{1}{T}\sum\limits_{t=1}^{T}\ln\left(1 + \bm{x}_t \cdot \bm{w}\right)$.

Where:

$\mathbf{\Sigma}$ is the covariance matrix of the asset returns;

$\bm{x}_t$ is the vector of asset returns at timestep $t$;

$\bm{\mu}$ is the vector of expected returns for each asset;

and $\bm{w}$ is the asset weights vector.

PortfolioOptimiser.TrackingErrKinds — Constant

TrackingErrKinds = (:Weights, :Returns)

Available kinds of tracking errors for Portfolio.

:Weights: provide a vector of asset weights which is used to compute the vector of benchmark returns;

:Returns: directly provide the vector of benchmark returns.

The benchmark is then used as a reference to optimise a portfolio that tracks it up to a given error.

PortfolioOptimiser.ObjFuncs — Constant

ObjFuncs = (:Min_Risk, :Utility, :Sharpe, :Max_Ret)

Objective functions available for use in :Trad and :WC optimisations of Portfolio (see PortTypes).

PortfolioOptimiser.ValidTermination — Constant

ValidTermination =
(MOI.OPTIMAL, MOI.ALMOST_OPTIMAL, MOI.LOCALLY_SOLVED, MOI.ALMOST_LOCALLY_SOLVED)

Valid JuMP termination codes after optimising an instance of Portfolio. If the termination code is different to these, then the failures are logged in the .fail field of HCPortfolio and Portfolio.

PortfolioOptimiser.PortClasses — Constant

PortClasses = (:Classic,)

Available classes for Portfolio.

PortfolioOptimiser.PortTypes — Constant

PortTypes = (:Trad, :RP, :RRP, :WC)

Available optimisation types for Portfolio.

:Trad – Traditional Optimisations

Available objective functions for :Trad optimisations. We can chose any of the objective functions in ObjFuncs and risk measures in RiskMeasures.

:Min_Risk: minimum risk portfolio,

\[\begin{align*} \underset{\bm{w}}{\min} &\qquad \phi_{j}(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\ &\qquad R(\bm{w}) \geq \overline{\mu} \end{align*}\,.\]

:Utility: maximum utility portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad R(\bm{w}) - \lambda \phi_{j}(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\ &\qquad R(\bm{w}) \geq \overline{\mu} \end{align*}\,.\]

:Sharpe: maximum risk-adjusted return ratio portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad \dfrac{R(\bm{w}) - r}{\phi_{j}(\bm{w})} \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\ &\qquad R(\bm{w}) \geq \overline{\mu} \end{align*}\,.\]

:Max_Ret: maximum return portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad R(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\ &\qquad R(\bm{w}) \geq \overline{\mu} \end{align*}\,.\]

Where:

$\bm{w}$ are the asset weights;

$\phi_{i}$ is risk measure $i$ from the set of available risk measures $\left\{\Phi\right\}$ (see RiskMeasures);

$\mathbf{A} \bm{w} \geq \mathbf{B}$ is a set of linear constraints;

$c_{i}$ is the maximum acceptable value for risk measure $\phi_{i}$ of the optimised portfolio;

$R(\bm{w})$ is the return function from KellyRet;

$\overline{\mu}$ is the minimum acceptable return of the optimised portfolio;

$\lambda$ is the risk aversion coefficient;

and $r$ is the risk-free rate.

:RP – Risk Parity Optimisations

Optimises portfolios based on a vector of risk contributions per asset. We can chose any of the risk measures in RiskMeasures.

\[\begin{align*} \underset{\bm{w}}{\min} &\qquad \phi(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \bm{b} \cdot \ln(\bm{w}) \geq c \\ &\qquad R(\bm{w}) \geq \overline{\mu}\\ &\qquad \bm{w} \geq \bm{0} \end{align*}\,.\]

Where:

$\bm{w}$ are the asset weights;

$\phi$ a risk measure from the set of available risk measures (see RiskMeasures);

$\mathbf{A} \bm{w} \geq \mathbf{B}$ is a set of linear constraints;

$\bm{b}$ is the vector of maximum allowable risk contribution per asset to the optimised portfolio;

$c$ is an auxiliary variable;

$R(\bm{w})$ is the return function from KellyRet;

and $\overline{\mu}$ is the minimum acceptable return of the optimised portfolio.

:RRP – Relaxed Risk Parity Optimisations

Optimises portfolios based on a vector of risk contributions per asset. Defines its own risk measure using the portfolio returns covariance.

\[\begin{align*} \underset{\bm{w}}{\min} &\qquad \psi - \gamma \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \bm{w}^\intercal \, \mathbf{\Sigma}\, \bm{w} \leq (\psi^{2} - \rho^{2})\\ &\qquad \lambda \bm{w}^\intercal \, \mathbf{\Theta}\, \bm{w} \leq \rho^{2}\\ &\qquad \bm{\zeta} = \mathbf{\Sigma} \bm{w}\\ &\qquad w_{i} \zeta_{i} \geq \gamma^{2} b_{i} \qquad \forall \, i = 1,\, \ldots{},\, N\\ &\qquad R(\bm{w}) \geq \overline{\mu}\\ &\qquad \sum\limits_{i=1}^{N} w_{i} = 1\\ &\qquad \bm{w} \geq \bm{0}\\ &\qquad \psi,\, \gamma,\, \rho \geq 0 \end{align*}\,.\]

Where:

$\bm{w}$ are the asset weights;

$\psi$ is the average risk of the portfolio;

$\gamma$ is the lower bound of the risk contribution for each asset;

$\mathbf{A} \bm{w} \geq \mathbf{B}$ is a set of linear constraints;

$\mathbf{\Sigma}$ is the portfolio covariance;

$\rho$ is a regularisation variable;

$\mathbf{\Theta} = \mathrm{diag}\left(\mathbf{\Sigma}\right)$ ;

$\lambda$ is a penalty parameter for $\rho$, taken from the available choices in RRPVersions;

$\bm{\zeta}$ is the vector of marginal risk for each asset;

$b_{i}$ is the maximum allowable risk contribution for asset $i$;

$N$ is the number of assets;

$R(\bm{w})$ is the return function from KellyRet;

and $\overline{\mu}$ is the minimum acceptable return of the optimised portfolio.

:WC – Worst Case Mean Variance Optimisations

Computes the worst case mean variance portfolio according to user-selected uncertainty sets (see UncertaintyTypes) for the portfolio return and covariance. We can chose any of the objective functions in ObjFuncs.

:Min_Risk: worst case minimum risk mean-variance portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad \underset{\mathbf{\Sigma}\, \in\, U_{\mathbf{\Sigma}}}{\max}\, \bm{w}^{\intercal}\, \mathbf{\Sigma}\, \bm{w}\\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B}\,. \end{align*}\]

:Utility: worst case maximum utility mean-variance portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad \underset{\bm{\mu}\, \in\, U_{\bm{\mu}}}{\min} R(\bm{w})\, -\, \underset{\mathbf{\Sigma}\, \in\, U_{\mathbf{\Sigma}}}{\max}\, \lambda \bm{w}^{\intercal}\, \mathbf{\Sigma}\, \bm{w}\\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B}\,. \end{align*}\]

:Sharpe: worst case maximum risk-adjusted return ratio mean-variance portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad \dfrac{\underset{\bm{\mu}\, \in\, U_{\bm{\mu}}}{\min} R(\bm{w}) - r}{\underset{\mathbf{\Sigma}\, \in\, U_{\mathbf{\Sigma}}}{\max}\, \left(\bm{w}^{\intercal}\, \mathbf{\Sigma}\, \bm{w}\right)^{1/2}} \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B}\,. \end{align*}\]

:Max_Ret: worst case maximum return mean-variance portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad \underset{\bm{\mu}\, \in\, U_{\bm{\mu}}}{\min} R(\bm{w})\\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B}\,. \end{align*}\]

Where:

$\bm{w}$ are the asset weights;

$\mathbf{\Sigma}$ is the covariance matrix of asset returns;

$U_{\Sigma}$ is the uncertainty set for the covariance matrix, they can be:

\[\begin{align*} U_{\Sigma}^{\mathrm{box}} &= \left\{\mathbf{\Sigma}\, \vert\, \mathbf{\Sigma}_{l} \leq \mathbf{\Sigma} \leq \mathbf{\Sigma}_{u},\, \mathbf{\Sigma} \succeq 0\right\}\\ U_{\Sigma}^{\mathrm{ellipse}} &= \left\{\mathbf{\Sigma}\, \vert\, \left[\mathrm{vec}\left(\mathbf{\Sigma}\right) - \mathrm{vec}\left(\hat{\mathbf{\Sigma}}\right)\right] \mathbf{\Sigma}_{\mathbf{\Sigma}}^{-1} \left[\mathrm{vec}\left(\mathbf{\Sigma}\right) - \mathrm{vec}\left(\hat{\mathbf{\Sigma}}\right)\right]^{\intercal} \leq k_{\mathbf{\Sigma}}^2 ,\, \mathbf{\Sigma} \succeq 0\right\}\,. \end{align*}\]

Where the following variables are estimated by assuming that the portfolio's asset return covariance can be generated by some matrix distribution. The distribution is sampled, and the estimates are calculated from them. Available choices can be found in BoxTypes and EllipseTypes for the box and ellipse sets respectively:

the $\mathrm{l}$ and $\mathrm{u}$ subscripts denote lower and upper bounds for the covariance matrix given the samples;

$\mathbf{\Sigma}_{\mathbf{\Sigma}}$ is the covariance of the samples;

$\hat{\mathbf{\Sigma}}$ the expected covariance given the samples;

and $k_{\mathbf{\Sigma}}$ is a significance parameter of the matrix distribution.

$\mathbf{A} \bm{w} \geq \mathbf{B}$ is a set of linear constraints;

$\bm{\mu}$ is the vector of expected returns for each asset;

$R(\bm{w})$ is the return function from KellyRet;

$U_{\bm{\mu}}$ is the uncertainty set for the asset returns, they can be:

\[\begin{align*} U_{\bm{\mu}}^{\mathrm{box}} &= \left\{\bm{\mu}\, \vert\, \vert \bm{\mu} - \bm{\hat{\mu}} \vert \leq \delta \right\}\\ U_{\bm{\mu}}^{\mathrm{ellipse}} &= \left\{\bm{\mu}\, \vert\, \left(\bm{\mu} - \bm{\hat{\mu}}\right) \mathbf{\Sigma}_{\bm{\mu}}^{-1} \left(\bm{\mu} - \bm{\hat{\mu}}\right)^{\intercal} \leq k_{\bm{\mu}}^{2}\right\}\,. \end{align*}\]

Where the following variables are estimated by assuming that the portfolio's asset mean returns can be generated by some distribution. The distribution is sampled, and the estimates are calculated from them. Available choices can be found in BoxTypes and EllipseTypes for the box and ellipse sets respectively:

$\hat{\bm{\mu}}$ the expected portfolio asset mean returns given the samples;

$\mathbf{\Sigma}_{\bm{\mu}}$ is the covariance of the samples;

and $k_{\bm{\mu}}$ is a significance parameter of the distribution.

$\lambda$ is the risk aversion coefficient;

and $r$ is the risk-free rate.

PortfolioOptimiser.RRPVersions — Constant

RRPVersions = (:None, :Reg, :Reg_Pen)

Available versions of Relaxed Risk Parity Optimisations (see PortTypes). Where:

:None: no penalty;

:Reg: regularisation constraint, $\rho$;

:Reg_Pen: regularisation and penalisation constraints, $\lambda$ and $\rho$.

PortfolioOptimiser.RPConstraintTypes — Constant

RPConstraintTypes = (:Assets, :Classes)

Types of risk parity constraints for building the set of linear constraints via rp_constraints.

PortfolioOptimiser.UncertaintyTypes — Constant

UncertaintyTypes = (:None, :Box, :Ellipse)

Types of uncertainty sets for Worst Case Optimisations of Portfolio (see PortTypes).

PortfolioOptimiser.KindBootstrap — Constant

KindBootstrap = (:Stationary, :Circular, :Moving)

Kind of bootstrap for computing the uncertainty sets with wc_statistics!, which are used by Worst Case Mean Variance Optimisations (see PortTypes).

PortfolioOptimiser.EllipseTypes — Constant

EllipseTypes = (:Stationary, :Circular, :Moving, :Normal)

Available types of elliptical sets that can be computed with wc_statistics!, which are used by Worst Case Mean Variance Optimisations (see PortTypes).

PortfolioOptimiser.BoxTypes — Constant

BoxTypes = (:Stationary, :Circular, :Moving, :Normal, :Delta)

Available types of box sets that can be computed with wc_statistics!, which are used by Worst Case Mean Variance Optimisations (see PortTypes).

PortfolioOptimiser.DBHTRootMethods — Constant

DBHTRootMethods = (:Unique, :Equal)

Methods for finding the root of a Direct Bubble Hierarchical Clustering Tree in DBHTs.

PortfolioOptimiser.OWAMethods — Constant

OWAMethods = (:CRRA, :ME, :MSS, :SD)

Methods for computing the higher order L-moments of Ordered Weight Arrays in owa_l_moment_crm.

PortfolioOptimiser.BinTypes — Constant

BinTypes = (:KN, :FD, :SC, :HGR)

Bin type choices for calculating bin widths for mutual and variational information matrices computed by mut_var_info_mtx.

PortfolioOptimiser.InfoTypes — Constant

InfoTypes = (:Mutual, :Variation)

Type of information matrix to compute when choosing :Mutual_Info from CodepTypes in asset_statistics!.

PortfolioOptimiser.HRRiskMeasures — Constant

HRRiskMeasures = (
    :SD,
    :MAD,
    :SSD,
    :FLPM,
    :SLPM,
    :WR,
    :CVaR,
    :EVaR,
    :RVaR,
    :MDD,
    :ADD,
    :CDaR,
    :UCI,
    :EDaR,
    :RDaR,
    :Kurt,
    :SKurt,
    :GMD,
    :RG,
    :RCVaR,
    :TG,
    :RTG,
    :OWA,
    :Variance,
    :Equal,
    :VaR,
    :DaR,
    :DaR_r,
    :MDD_r,
    :ADD_r,
    :CDaR_r,
    :EDaR_r,
    :RDaR_r,
)

Available risk measures for optimisations of HCPortfolio.

:SD = standard deviation (SD);

:MAD = max absolute deviation (MAD);

:SSD = semi standard deviation (SSD);

:FLPM = first lower partial moment (Omega ratio) (FLPM);

:SLPM = second lower partial moment (Sortino ratio) (SLPM);

:WR = worst realisation (WR);

:CVaR = conditional value at risk (CVaR);

:EVaR = entropic value at risk (EVaR);

:RVaR = relativistic value at risk (RVaR);

:MDD = maximum drawdown of uncompounded cumulative returns (Calmar ratio) (MDD_abs);

:ADD = average drawdown of uncompounded cumulative returns (ADD_abs);

:CDaR = conditional drawdown at risk of uncompounded cumulative returns (CDaR_abs);

:UCI = ulcer index of uncompounded cumulative returns (UCI_abs);

:EDaR = entropic drawdown at risk of uncompounded cumulative returns (EDaR_abs);

:RDaR = relativistic drawdown at risk of uncompounded cumulative returns (RDaR_abs);

:Kurt = square root kurtosis (Kurt);

:SKurt = square root semi-kurtosis (SKurt);

:GMD = gini mean difference (GMD);

:RG = range of returns (RG);

:RCVaR = range of conditional value at risk (RCVaR);

:TG = tail gini (TG);

:RTG = range of tail gini (RTG);

:OWA = ordered weight array (generic OWA weights) (OWA);

:Variance = variance (Variance);

:Equal = equal risk contribution, 1/N where N is the number of assets;

:VaR = value at risk (VaR);

:DaR = drawdown at risk of uncompounded cumulative returns (DaR_abs);

:DaR_r = drawdown at risk of compounded cumulative returns (DaR_rel);

:MDD_r = maximum drawdown of compounded cumulative returns (MDD_rel);

:ADD_r = average drawdown of compounded cumulative returns (ADD_rel);

:CDaR_r = conditional drawdown at risk of compounded cumulative returns (CDaR_rel);

:UCI_r = ulcer index of compounded cumulative returns (UCI_rel);

:EDaR_r = entropic drawdown at risk of compounded cumulative returns (EDaR_rel);

:RDaR_r = relativistic drawdown at risk of compounded cumulative returns (RDaR_rel).

PortfolioOptimiser.HCPortTypes — Constant

HCPortTypes = (:HRP, :HERC, :HERC2, :NCO)

PortfolioOptimiser.CodepTypes — Constant

CodepTypes = (
    :Pearson,
    :Spearman,
    :Kendall,
    :Gerber1,
    :Gerber2,
    :Abs_Pearson,
    :Abs_Spearman,
    :Abs_Kendall,
    :Distance,
    :Mutual_Info,
    :Tail,
    :Cov_to_Cor,
    :Custom_Func,
    :Custom_Val,
)

Available types of codependency measures computed in asset_statistics! when the portfolio is a HCPortfolio.

PortfolioOptimiser.LinkageTypes — Constant

LinkageTypes = (:single, :complete, :average, :ward, :ward_presquared, :DBHT)

Linkage types available when optimising a HCPortfolio. Where:

:DBHT: is Direct Bubble Hierarchical Tree clustering.

The rest are linkage types supported by Clustering.hclust.

PortfolioOptimiser.BranchOrderTypes — Constant

BranchOrderTypes = (:optimal, :barjoseph, :r, :default)

Algorithm to order leaves and branches.

:default: if linkage is :DBHT, the leaves and branches remain as the algorithm orders them. If any other linkage is used, they fall back to :r as that is their default according to Clustering.hclust.

All other branch orders are as defined by Clustering.hclust.

PortfolioOptimiser.HRObjFuncs — Constant

HRObjFuncs = (:Min_Risk, :Utility, :Sharpe, :Max_Ret, :Equal)

Objective funcions for :NCO HCPortTypes of HCPortfolio.

:Min_Risk, :Utility, :Sharpe and :Max_Ret optimise the sub-portfolios as :Trad PortTypes of Portfolio according to their respective definitions in ObjFuncs.

:Equal optimises the sub-portfolios as :RP PortTypes of Portfolio with equal risk contribution per asset. We can't offer customiseable risk contributions because the size and composition of the clusters is initially unknown and depends on the chosen linkage method.