Constant Definitions

Many functions only accept specific methods and/or require certain arguments to be within appropriate bounds. These constants ensure such requirements are met.

PortfolioOptimiser.RiskMeasures — Constant

RiskMeasures = (
    :SD,
    :MAD,
    :SSD,
    :FLPM,
    :SLPM,
    :WR,
    :CVaR,
    :EVaR,
    :RVaR,
    :MDD,
    :ADD,
    :CDaR,
    :UCI,
    :EDaR,
    :RDaR,
    :Kurt,
    :SKurt,
    :GMD,
    :RG,
    :RCVaR,
    :TG,
    :RTG,
    :OWA,
)

Available risk measures for :Trad and :RP type (see PortTypes) of Portfolio.

:SD = standard deviation (SD).

:MAD = max absolute deviation (MAD).

:SSD = semi standard deviation (SSD).

:FLPM = first lower partial moment (omega ratio) (FLPM).

:SLPM = second lower partial moment (sortino ratio) (SLPM).

:WR = worst realisation (WR).

:CVaR = conditional value at risk (CVaR).

:EVaR = entropic value at risk (EVaR).

:RVaR = relativistic value at risk (RVaR).

:MDD = maximum drawdown of uncompounded cumulative returns (MDD_abs).

:ADD = average drawdown of uncompounded cumulative returns (ADD_abs).

:CDaR = conditional drawdown at risk of uncompounded cumulative returns (CDaR_abs).

:UCI = ulcer index of uncompounded cumulative returns (UCI_abs).

:EDaR = entropic drawdown at risk of uncompounded cumulative returns (EDaR_abs).

:RDaR = relativistic drawdown at risk of uncompounded cumulative returns (RDaR_abs).

:Kurt = square root kurtosis (Kurt).

:SKurt = square root semi-kurtosis (SKurt).

:GMD = gini mean difference (GMD).

:RG = range of returns (RG).

:RCVaR = range of conditional value at risk (RCVaR).

:TG = tail gini (TG).

:RTG = range of tail gini (RTG).

:OWA = ordered weight array (generic OWA weights) (OWA).

PortfolioOptimiser.KellyRet — Constant

KellyRet = (:None, :Approx, :Exact)

Available types of Kelly returns for Portfolio.

:None: arithmetic mean return, $R(\bm{w}) = \bm{\mu} \cdot \bm{w}$.

:Approx: first moment approximation of the logarithmic returns, $R(\bm{w}) = \bm{\mu} \cdot \bm{w} - \dfrac{1}{2} \bm{w}^{\intercal} \mathbf{\Sigma} \bm{w}$.

:Exact: exact logarithmic returns, $R(\bm{w}) = \dfrac{1}{T}\sum\limits_{t=1}^{T}\ln\left(1 + \bm{x}_t \cdot \bm{w}\right)$.

Where:

$\mathbf{\Sigma}$ is the covariance matrix of the asset returns.

$\bm{x}_t$ is the vector of asset returns at timestep $t$.

$\bm{\mu}$ is the vector of expected returns for each asset.

and $\bm{w}$ is the asset weights vector.

PortfolioOptimiser.TrackingErrKinds — Constant

TrackingErrKinds = (:Weights, :Returns)

Available kinds of tracking errors for Portfolio.

:Weights: provide a vector of asset weights which is used to compute the vector of benchmark returns,

$\bm{b} = \mathbf{X} \bm{w}$,

where $\bm{b}$ is the benchmark returns vector, $\mathbf{X}$ the $T \times{} N$ asset returns matrix, and $\bm{w}$ the asset weights vector.

:Returns: directly provide the vector of benchmark returns.

The benchmark is then used as a reference to optimise a portfolio that tracks it up to a given error.

PortfolioOptimiser.ObjFuncs — Constant

ObjFuncs = (:Min_Risk, :Utility, :Sharpe, :Max_Ret)

Objective functions available for use in :Trad and :WC optimisations of Portfolio (see PortTypes).

PortfolioOptimiser.ValidTermination — Constant

ValidTermination = (
    MOI.OPTIMAL,
    MOI.ALMOST_OPTIMAL,
    MOI.LOCALLY_SOLVED,
    MOI.ALMOST_LOCALLY_SOLVED,
    MOI.SOLUTION_LIMIT,
    MOI.OBJECTIVE_LIMIT,
)

Valid JuMP termination codes after optimising an instance of Portfolio. If the termination code is different to these, then the failures are logged in the .fail field of HCPortfolio and Portfolio.

PortfolioOptimiser.PortClasses — Constant

PortClasses = (:Classic, :FM, :BL, :BLFM)

Available choicees of summary parameters $\bm{\mu}$ and $\bm{\Sigma}$ Portfolio.

:Classic: when optimising with this option, $\bm{\mu}$ and $\bm{\Sigma}$ take their values from historical estimates computed by asset_statistics!.

:FM: when optimising with this option, $\bm{\mu}$ and $\bm{\Sigma}$ take their values from the factor model computed by factor_statistics!.

:BL: when optimising with this option, $\bm{\mu}$ and $\bm{\Sigma}$ take their values from the Black-Litterman model computed by black_litterman_statistics!.

:BLFM: when optimising with this option, $\bm{\mu}$ and $\bm{\Sigma}$ take their values from the factor Black-Litterman model computed by black_litterman_factor_satistics!. This model has two versions defined in BLFMType.

PortfolioOptimiser.PortTypes — Constant

PortTypes = (:Trad, :RP, :RRP, :WC)

Available optimisation types for Portfolio.

:Trad – Traditional Optimisations

Available objective functions for :Trad optimisations. We can chose any of the objective functions in ObjFuncs and risk measures in RiskMeasures.

:Min_Risk: minimum risk portfolio,

\[\begin{align*} \underset{\bm{w}}{\min} &\qquad \phi_{j}(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\ &\qquad R(\bm{w}) \geq \overline{\mu} \end{align*}\,.\]

:Utility: maximum utility portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad R(\bm{w}) - \lambda \phi_{j}(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\ &\qquad R(\bm{w}) \geq \overline{\mu} \end{align*}\,.\]

:Sharpe: maximum risk-adjusted return ratio portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad \dfrac{R(\bm{w}) - r}{\phi_{j}(\bm{w})} \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\ &\qquad R(\bm{w}) \geq \overline{\mu} \end{align*}\,.\]

:Max_Ret: maximum return portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad R(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \phi_{i}(\bm{w}) \leq c_{i} \, \forall \, \phi_{i} \in \left\{\Phi\right\} \\ &\qquad R(\bm{w}) \geq \overline{\mu} \end{align*}\,.\]

Where:

$\bm{w}$ are the asset weights.

$\phi_{i}$ is risk measure $i$ from the set of available risk measures $\left\{\Phi\right\}$ (see RiskMeasures).

$\mathbf{A} \bm{w} \geq \mathbf{B}$ is a set of linear constraints.

$c_{i}$ is the maximum acceptable value for risk measure $\phi_{i}$ of the optimised portfolio.

$R(\bm{w})$ is the return function from KellyRet.

$\overline{\mu}$ is the minimum acceptable return of the optimised portfolio.

$\lambda$ is the risk aversion coefficient.

and $r$ is the risk-free rate.

:RP – Risk Parity Optimisations

Optimises portfolios based on a vector of risk contributions per asset. We can chose any of the risk measures in RiskMeasures.

\[\begin{align*} \underset{\bm{w}}{\min} &\qquad \phi(\bm{w}) \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \bm{b} \cdot \ln(\bm{w}) \geq c \\ &\qquad R(\bm{w}) \geq \overline{\mu}\\ &\qquad \bm{w} \geq \bm{0} \end{align*}\,.\]

Where:

$\bm{w}$ are the asset weights.

$\phi$ a risk measure from the set of available risk measures (see RiskMeasures).

$\mathbf{A} \bm{w} \geq \mathbf{B}$ is a set of linear constraints.

$\bm{b}$ is the vector of maximum allowable risk contribution per asset to the optimised portfolio.

$c$ is an auxiliary variable.

$R(\bm{w})$ is the return function from KellyRet.

and $\overline{\mu}$ is the minimum acceptable return of the optimised portfolio.

:RRP – Relaxed Risk Parity Optimisations

Optimises portfolios based on a vector of risk contributions per asset. Defines its own risk measure using the portfolio returns covariance.

\[\begin{align*} \underset{\bm{w}}{\min} &\qquad \psi - \gamma \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B} \\ &\qquad \bm{w}^\intercal \, \mathbf{\Sigma}\, \bm{w} \leq (\psi^{2} - \rho^{2})\\ &\qquad \lambda \bm{w}^\intercal \, \mathbf{\Theta}\, \bm{w} \leq \rho^{2}\\ &\qquad \bm{\zeta} = \mathbf{\Sigma} \bm{w}\\ &\qquad w_{i} \zeta_{i} \geq \gamma^{2} b_{i} \qquad \forall \, i = 1,\, \ldots{},\, N\\ &\qquad R(\bm{w}) \geq \overline{\mu}\\ &\qquad \sum\limits_{i=1}^{N} w_{i} = 1\\ &\qquad \bm{w} \geq \bm{0}\\ &\qquad \psi,\, \gamma,\, \rho \geq 0 \end{align*}\,.\]

Where:

$\bm{w}$ are the asset weights.

$\psi$ is the average risk of the portfolio.

$\gamma$ is the lower bound of the risk contribution for each asset.

$\mathbf{A} \bm{w} \geq \mathbf{B}$ is a set of linear constraints.

$\mathbf{\Sigma}$ is the portfolio covariance.

$\rho$ is a regularisation variable.

$\mathbf{\Theta} = \mathrm{diag}\left(\mathbf{\Sigma}\right)$ .

$\lambda$ is a penalty parameter for $\rho$, taken from the available choices in RRPVersions.

$\bm{\zeta}$ is the vector of marginal risk for each asset.

$b_{i}$ is the maximum allowable risk contribution for asset $i$.

$N$ is the number of assets.

$R(\bm{w})$ is the return function from KellyRet.

and $\overline{\mu}$ is the minimum acceptable return of the optimised portfolio.

:WC – Worst Case Mean Variance Optimisations

Computes the worst case mean variance portfolio according to user-selected uncertainty sets (see UncertaintyTypes) for the portfolio return and covariance. We can chose any of the objective functions in ObjFuncs.

:Min_Risk: worst case minimum risk mean-variance portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad \underset{\mathbf{\Sigma}\, \in\, U_{\mathbf{\Sigma}}}{\max}\, \bm{w}^{\intercal}\, \mathbf{\Sigma}\, \bm{w}\\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B}\,. \end{align*}\]

:Utility: worst case maximum utility mean-variance portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad \underset{\bm{\mu}\, \in\, U_{\bm{\mu}}}{\min} R(\bm{w})\, -\, \underset{\mathbf{\Sigma}\, \in\, U_{\mathbf{\Sigma}}}{\max}\, \lambda \bm{w}^{\intercal}\, \mathbf{\Sigma}\, \bm{w}\\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B}\,. \end{align*}\]

:Sharpe: worst case maximum risk-adjusted return ratio mean-variance portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad \dfrac{\underset{\bm{\mu}\, \in\, U_{\bm{\mu}}}{\min} R(\bm{w}) - r}{\underset{\mathbf{\Sigma}\, \in\, U_{\mathbf{\Sigma}}}{\max}\, \left(\bm{w}^{\intercal}\, \mathbf{\Sigma}\, \bm{w}\right)^{1/2}} \\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B}\,. \end{align*}\]

:Max_Ret: worst case maximum return mean-variance portfolio,

\[\begin{align*} \underset{\bm{w}}{\max} &\qquad \underset{\bm{\mu}\, \in\, U_{\bm{\mu}}}{\min} R(\bm{w})\\ \mathrm{s.t.} &\qquad \mathbf{A} \bm{w} \geq \mathbf{B}\,. \end{align*}\]

Where:

$\bm{w}$ are the asset weights.

$\mathbf{\Sigma}$ is the covariance matrix of asset returns.

$U_{\Sigma}$ is the uncertainty set for the covariance matrix, they can be:

\[\begin{align*} U_{\Sigma}^{\mathrm{box}} &= \left\{\mathbf{\Sigma}\, \vert\, \mathbf{\Sigma}_{l} \leq \mathbf{\Sigma} \leq \mathbf{\Sigma}_{u},\, \mathbf{\Sigma} \succeq 0\right\}\\ U_{\Sigma}^{\mathrm{ellipse}} &= \left\{\mathbf{\Sigma}\, \vert\, \left[\mathrm{vec}\left(\mathbf{\Sigma}\right) - \mathrm{vec}\left(\hat{\mathbf{\Sigma}}\right)\right] \mathbf{\Sigma}_{\mathbf{\Sigma}}^{-1} \left[\mathrm{vec}\left(\mathbf{\Sigma}\right) - \mathrm{vec}\left(\hat{\mathbf{\Sigma}}\right)\right]^{\intercal} \leq k_{\mathbf{\Sigma}}^2 ,\, \mathbf{\Sigma} \succeq 0\right\}\,. \end{align*}\]

Where the following variables are estimated by assuming that the portfolio's asset return covariance can be generated by some matrix distribution. The distribution is sampled, and the estimates are calculated from them. Available choices can be found in BoxTypes and EllipseTypes for the box and ellipse sets respectively:

the $\mathrm{l}$ and $\mathrm{u}$ subscripts denote lower and upper bounds for the covariance matrix given the samples.

$\mathbf{\Sigma}_{\mathbf{\Sigma}}$ is the covariance of the samples.

$\hat{\mathbf{\Sigma}}$ the expected covariance given the samples.

and $k_{\mathbf{\Sigma}}$ is a significance parameter of the matrix distribution.

$\mathbf{A} \bm{w} \geq \mathbf{B}$ is a set of linear constraints.

$\bm{\mu}$ is the vector of expected returns for each asset.

$R(\bm{w})$ is the return function from KellyRet.

$U_{\bm{\mu}}$ is the uncertainty set for the asset returns, they can be:

\[\begin{align*} U_{\bm{\mu}}^{\mathrm{box}} &= \left\{\bm{\mu}\, \vert\, \vert \bm{\mu} - \bm{\hat{\mu}} \vert \leq \delta \right\}\\ U_{\bm{\mu}}^{\mathrm{ellipse}} &= \left\{\bm{\mu}\, \vert\, \left(\bm{\mu} - \bm{\hat{\mu}}\right) \mathbf{\Sigma}_{\bm{\mu}}^{-1} \left(\bm{\mu} - \bm{\hat{\mu}}\right)^{\intercal} \leq k_{\bm{\mu}}^{2}\right\}\,. \end{align*}\]

Where the following variables are estimated by assuming that the portfolio's asset mean returns can be generated by some distribution. The distribution is sampled, and the estimates are calculated from them. Available choices can be found in BoxTypes and EllipseTypes for the box and ellipse sets respectively:

$\hat{\bm{\mu}}$ the expected portfolio asset mean returns given the samples.

$\mathbf{\Sigma}_{\bm{\mu}}$ is the covariance of the samples.

and $k_{\bm{\mu}}$ is a significance parameter of the distribution.

$\lambda$ is the risk aversion coefficient.

and $r$ is the risk-free rate.

The worst case uncertainty sets are computed by wc_statistics!.

PortfolioOptimiser.RRPVersions — Constant

RRPVersions = (:None, :Reg, :Reg_Pen)

Available versions of Relaxed Risk Parity Optimisations (see PortTypes).

:None: no penalty.

:Reg: regularisation constraint, $\rho$.

:Reg_Pen: regularisation and penalisation constraints, $\lambda$ and $\rho$.

PortfolioOptimiser.UncertaintyTypes — Constant

UncertaintyTypes = (:None, :Box, :Ellipse)

Available types of uncertainty sets that can be computed with wc_statistics!, which are used by Worst Case Mean Variance Optimisations (see PortTypes, EllipseTypes, and BoxTypes).

:Box: are box uncertainty sets, ie the sets are full matrices.

:Ellipse: are elliptical uncertainty sets, ie the sets are diagonal matrices.

PortfolioOptimiser.EllipseTypes — Constant

EllipseTypes = (:Stationary, :Circular, :Moving, :Normal)

Available types of elliptical sets that can be computed with wc_statistics!, which are used by Worst Case Mean Variance Optimisations (see PortTypes).

:Stationary: stationary bootstrapping method.

:Circular: circular block bootstrapping method.

:Moving: moving block bootstrapping method.

:Normal: sets generated by assuming returns are normally distributed.

PortfolioOptimiser.BoxTypes — Constant

BoxTypes = (:Stationary, :Circular, :Moving, :Normal, :Delta)

Available types of box sets that can be computed with wc_statistics!, which are used by Worst Case Mean Variance Optimisations (see PortTypes).

:Stationary: stationary bootstrapping method.

:Circular: circular block bootstrapping method.

:Moving: moving block bootstrapping method.

:Normal: normally distributed covariance and mean samples.

:Delta: box sets are assumed to fall in the extrema of a well-defined interval.

PortfolioOptimiser.KindBootstrap — Constant

KindBootstrap = (:Stationary, :Circular, :Moving)

Kind of bootstrap for computing the uncertainty sets with wc_statistics!, which are used by Worst Case Mean Variance Optimisations (see PortTypes).

:Stationary: stationary bootstrapping method.

:Circular: circular block bootstrapping method.

:Moving: moving block bootstrapping method.

PortfolioOptimiser.BLFMType — Constant

BLFMType = (:A, :B)

Versions of the factor Black-Litterman Model.

:B: Bayesian Black-Litterman, which uses the factors to generate the Black-Litterman estimates.

:A: Augmented Black-Litterman, which uses the factors to adjust the Black-Litterman views.

PortfolioOptimiser.RPConstraintTypes — Constant

RPConstraintTypes = (:Assets, :Classes)

Types of risk parity constraints for building the set of linear constraints via rp_constraints.

:Assets: restrict the assets that meet the criteria.

:Classes: restrict the assets in that class.

PortfolioOptimiser.OWAMethods — Constant

OWAMethods = (:CRRA, :E, :SS, :SD)

Methods for computing the weights used to combine higher order L-moments in owa_l_moment_crm.

CRRA: Normalised Constant Relative Risk Aversion Coefficients.

:E: Maximum Entropy.

:SS: Minimum Sum of Squares.

:SD: Minimum Square Distance.

PortfolioOptimiser.BinTypes — Constant

BinTypes = (:KN, :FD, :SC, :HGR)

Methods for calculating bin widths for mutual and variational information matrices computed by mut_var_info_mtx.

:KN: Knuth's choice.

:FD: Freedman-Diaconis' choice.

:SC: Schotts' choice.

:HGR: Hacine-Gharbi and Ravier's choice.

PortfolioOptimiser.HRRiskMeasures — Constant

HRRiskMeasures = (
    :SD,
    :MAD,
    :SSD,
    :FLPM,
    :SLPM,
    :WR,
    :CVaR,
    :EVaR,
    :RVaR,
    :MDD,
    :ADD,
    :CDaR,
    :UCI,
    :EDaR,
    :RDaR,
    :Kurt,
    :SKurt,
    :GMD,
    :RG,
    :RCVaR,
    :TG,
    :RTG,
    :OWA,
    :Variance,
    :Equal,
    :VaR,
    :DaR,
    :DaR_r,
    :MDD_r,
    :ADD_r,
    :CDaR_r,
    :EDaR_r,
    :RDaR_r,
)

Available risk measures for optimisations of HCPortfolio.

:SD = standard deviation (SD).

:MAD = max absolute deviation (MAD).

:SSD = semi standard deviation (SSD).

:FLPM = first lower partial moment (Omega ratio) (FLPM).

:SLPM = second lower partial moment (Sortino ratio) (SLPM).

:WR = worst realisation (WR).

:CVaR = conditional value at risk (CVaR).

:EVaR = entropic value at risk (EVaR).

:RVaR = relativistic value at risk (RVaR).

:MDD = maximum drawdown of uncompounded cumulative returns (Calmar ratio) (MDD_abs).

:ADD = average drawdown of uncompounded cumulative returns (ADD_abs).

:CDaR = conditional drawdown at risk of uncompounded cumulative returns (CDaR_abs).

:UCI = ulcer index of uncompounded cumulative returns (UCI_abs).

:EDaR = entropic drawdown at risk of uncompounded cumulative returns (EDaR_abs).

:RDaR = relativistic drawdown at risk of uncompounded cumulative returns (RDaR_abs).

:Kurt = square root kurtosis (Kurt).

:SKurt = square root semi-kurtosis (SKurt).

:GMD = gini mean difference (GMD).

:RG = range of returns (RG).

:RCVaR = range of conditional value at risk (RCVaR).

:TG = tail gini (TG).

:RTG = range of tail gini (RTG).

:OWA = ordered weight array (generic OWA weights) (OWA).

:Variance = variance (Variance).

:Equal = equal risk contribution, 1/N where N is the number of assets.

:VaR = value at risk (VaR).

:DaR = drawdown at risk of uncompounded cumulative returns (DaR_abs).

:DaR_r = drawdown at risk of compounded cumulative returns (DaR_rel).

:MDD_r = maximum drawdown of compounded cumulative returns (MDD_rel).

:ADD_r = average drawdown of compounded cumulative returns (ADD_rel).

:CDaR_r = conditional drawdown at risk of compounded cumulative returns (CDaR_rel).

:UCI_r = ulcer index of compounded cumulative returns (UCI_rel).

:EDaR_r = entropic drawdown at risk of compounded cumulative returns (EDaR_rel).

:RDaR_r = relativistic drawdown at risk of compounded cumulative returns (RDaR_rel).

PortfolioOptimiser.HCPortTypes — Constant

HCPortTypes = (:HRP, :HERC, :NCO)

Available optimisation types for HCPortfolio.

:HRP: Hierarchical Risk Parity.

:HERC: Hierarhical Equal Risk Contribution.

:NCO: Nested Clustered Optimisation.

Both :HERC and :NCO split their optimisations into two parts:

inter-cluster optimisation.

intra-cluster optimisation.

Threfore they can make use of extra parameters:

:HERC: accepts an extra risk measure rm_i and OWA weights owa_w_i parameters, which are used for the intra-cluster optimisations. They default to the same value as their external counterparts.

:NCO: accepts an extra objective function obj_i kelly return kelly_i risk aversion parameter l_i risk measure rm_i and OWA weights owa_w_i parameters. They default to the same value as their external counterparts.

PortfolioOptimiser.CodepTypes — Constant

CodepTypes = (
    :Pearson,
    :Spearman,
    :Kendall,
    :Gerber0,
    :Gerber1,
    :Gerber2,
    :Abs_Pearson,
    :Abs_Spearman,
    :Abs_Kendall,
    :Distance,
    :Mutual_Info,
    :Tail,
    :Cov_to_Cor,
    :Custom_Func,
    :Custom_Val,
)

Methods for estimating the codependence (correlation) matrix $\mathbf{P}$, and distance matrix $\mathbf{D}$.

:Pearson: Pearson correlation, $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{P}_{i,\,j} \right)}$.

:Spearman: Spearman correlation, $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{P}_{i,\,j} \right)}$.

:Kendall: Kendall correlation, $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{P}_{i,\,j} \right)}$.

:Gerber0: Gerber statistic 0, $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{P}_{i,\,j} \right)}$.

:Gerber1: Gerber statistic 1, $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{P}_{i,\,j} \right)}$.

:Gerber2: Gerber statistic 2, $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{P}_{i,\,j} \right)}$.

:Abs_Pearson: absolute value of the Pearson correlation, $\mathbf{D}_{i,\,j} = \sqrt{1 - \left\lvert\mathbf{P}_{i,\,j}\right\rvert}$.

:Abs_Spearman: absolute value of the Spearman correlation, $\mathbf{D}_{i,\,j} = \sqrt{1 - \left\lvert\mathbf{P}_{i,\,j}\right\rvert}$.

:Abs_Kendall: absolute value of the Kendall correlation, $\mathbf{D}_{i,\,j} = \sqrt{1 - \left\lvert\mathbf{P}_{i,\,j}\right\rvert}$.

:Distance: distance correlation matrix, , $\mathbf{D}_{i,\,j} = \sqrt{1 - \mathbf{P}_{i,\,j}}$

:Mutual_Info: mutual information matrix, $\mathbf{D}_{i,\,j}$ is the variation information matrix.

:Tail: lower tail dependence index matrix, $\mathbf{D}_{i,\,j} = -\log\left(\mathbf{P}_{i,\,j}\right)$

:Cov_to_Cor: the covariance matrix is converted to a correlation matrix, the distance matrix is computed by a distance function which defaults to $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{P}_{i,\,j} \right)}$.

:Custom_Func: custom function provided, the distance matrix is computed by a distance function which defaults to $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{P}_{i,\,j} \right)}$.

:Custom_Val: custom value provided, the distance matrix is computed by a distance function which defaults to $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{P}_{i,\,j} \right)}$.

PortfolioOptimiser.LinkageTypes — Constant

LinkageTypes = (:single, :complete, :average, :ward, :ward_presquared, :DBHT)

Linkage types available when optimising a HCPortfolio.

:DBHT: is Direct Bubble Hierarchical Tree clustering.

The rest are linkage types supported by Clustering.hclust.

PortfolioOptimiser.BranchOrderTypes — Constant

BranchOrderTypes = (:optimal, :barjoseph, :r, :default)

Choice of algorithm for ordering hierarchical clustering dendrogram leaves and branches.

:default: if linkage is :DBHT, the leaves and branches remain as the algorithm orders them. If any other linkage is used, they fall back to :r as that is their default according to Clustering.hclust.

All other branch orders are as defined by Clustering.hclust.

PortfolioOptimiser.HRObjFuncs — Constant

HRObjFuncs = (:Min_Risk, :Utility, :Sharpe, :Max_Ret, :Equal)

Objective funcions for :NCO HCPortTypes of HCPortfolio.

:Min_Risk, :Utility, :Sharpe and :Max_Ret: optimise the sub-portfolios as :Trad PortTypes of Portfolio according to their respective definitions in ObjFuncs.

:Equal: optimises the sub-portfolios as :RP optimisations with equal risk contribution per asset/cluster. We can't offer customiseable risk contributions because the size and composition of the clusters is initially unknown and depends on the chosen linkage method.

PortfolioOptimiser.MuTypes — Constant

MuTypes = (:Default, :JS, :BS, :BOP, :CAPM, :Custom_Func, :Custom_Val)

Methods for estimating the mean returns vector $\bm{\mu}$ in mean_vec.

:Default: is the standard historical.

:JS: James-Stein.

:BS: Bayes-Stein.

:BOP: Bodnar-Okhrin-Parolya.

:CAPM: Capital Asset Pricing Model.

:Custom_Func: custom function provided.

:Custom_Val: custom value provided.

PortfolioOptimiser.MuTargets — Constant

MuTargets = (:GM, :VW, :SE)

Targets for the :JS, :BS and :BOP estimators in mean_vec and mu_estimator.

:GM: grand mean.

:VW: volatility-weighted grand mean.

:SE: mean square error of sample mean.

PortfolioOptimiser.CovTypes — Constant

CovTypes = (:Full, :Semi, :Gerber0, :Gerber1, :Gerber2, :Custom_Func, :Custom_Val)

Methods for estimating the covariance matrix $\mathbf{\Sigma}$.

:Full: full covariance matrix.

:Semi: semi-covariance matrix.

:Gerber0: Gerber statistic 0.

:Gerber1: Gerber statistic 1.

:Gerber2: Gerber statistic 2.

:Custom_Func: custom function provided.

:Custom_Val: custom value provided.

PortfolioOptimiser.PosdefFixes — Constant

PosdefFixes = (:None, :Nearest, :Custom_Func)

Methods for fixing non-positive definite matrices.

:None: no fix is applied.

:Nearest: nearest correlation matrix.

Custom_Func: custom function provided.

PortfolioOptimiser.RegCriteria — Constant

RegCriteria = (:pval, GLM.aic, GLM.aicc, GLM.bic, GLM.r2, GLM.adjr2)

Criteria for feature selection in regression functions.

:pval: p-value feature selection.

The rest are methods applied to a fitted General Linear Model from GLM.jl.

PortfolioOptimiser.FSType — Constant

FSType = (:FReg, :BReg, :PCR)

Methods for feature selection when creating the loadings matrix.

:FReg: forward regression;

:Breg: backward regression;

:PCR: Principal Component Regression using PCA.

PortfolioOptimiser.BLHist — Constant

BLHist = (1, 2, 3)

Choice of what estimate of $\mathbf{\Sigma}$ to use. Factor models, :FM and :BLFM, can also use a factor model estimate of the returns matrix $\mathbf{X}$.

1:

:FM: uses the factor model estimates of $\mathbf{\Sigma}$ and $\mathbf{X}$.

:BL: uses the Black Litterman estimate of $\mathbf{\Sigma}$.

:BLFM: uses the Black Litterman factor model estimate of $\mathbf{\Sigma}$ and factor model estimate of $\mathbf{X}$.

2:

:FM: uses the standard estimates of $\mathbf{\Sigma}$ and $\mathbf{X}$.

:BL: uses the standard estimate of $\mathbf{\Sigma}$.

:BLFM: uses the standard estimates of $\mathbf{\Sigma}$ and $\mathbf{X}$.

3

:FM and :BL do not support this option.

:BLFM: uses the factor model estimates of $\mathbf{\Sigma}$ and $\mathbf{X}$.