Ordered Weight Arrays

Public

PortfolioOptimiser.owa_cvar — Function

owa_cvar(T::Integer, alpha::Real = 0.05)

Calculate the OWA weights corresponding to the Critical Value at Risk (CVaR) of a returns series [OWA].

Inputs

T: number of returns observations.

alpha: significance level of CVaR, alpha in (0, 1).

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_gmd — Method

owa_gmd(T::Integer)

Computes the Gini Mean Difference (GMD) of a returns series [OWA].

Inputs

T: number of returns observations.

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_l_moment — Function

owa_l_moment(T::Integer, k::Integer = 2)

Calculates the OWA weights of the k'th linear moment (L-moment) of a returns series [30].

Inputs

T: number of returns observations.

k: order of the L-moment.

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_l_moment_crm — Method

owa_l_moment_crm(
    T::Integer;    k::Integer = 2,    method::Symbol = :SD,    g::Real = 0.5,    max_phi::Real = 0.5,    solvers = Dict())

Compute the OWA weights for the convex risk measure considering higher order L-moments [30].

Inputs

T: number of returns observations.

k: order of the L-moment, k ≥ 2.

method: method for computing the weights used to combine L-moments higher than 2, used in OWAMethods.

:CRRA: Normalised Constant Relative Risk Aversion Coefficients.

:E: Maximum Entropy. Solver must support MOI.RelativeEntropyCone and MOI.NormOneCone, or JuMP must be able to transform it/them into a supported form.

:SS: Minimum Sum of Squares. Solver must support MOI.SecondOrderCone, or JuMP must be able to transform it/them into a supported form.

:SD: Minimum Square Distance. Solver must support MOI.SecondOrderCone, or JuMP must be able to transform it/them into a supported form.

g: the risk aversion coefficient.

max_phi: maximum weight constraint of the L-moments.

solvers: Provides the solvers and corresponding parameters for solving the OWA L-Moment JuMP model. There can be two key => value pairs.

:solver => value: value is a JuMP optimiser. The optimiser can be declared alongside its attributes by using JuMP.solver_with_attributes.

:params => value: (optional) value must be a Dict or NamedTuple whose key => value pairs are the solver-specific settings.

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_rcvar — Method

owa_rcvar(T::Integer; alpha::Real = 0.05, beta::Real = alpha)

Compute the OWA weights for the CVaR Range of a returns series [OWA].

Inputs

T: number of returns observations.

alpha: significance level of CVaR losses, alpha in (0, 1).

beta: significance level of CVaR gains, beta in (0, 1).

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_rg — Method

owa_rg(T::Integer)

Compute the OWA weights for the Range of a returns series [OWA].

Inputs

T: number of returns observations.

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_rtg — Method

owa_rtg(
    T::Integer;    alpha_i::Real = 0.0001,    alpha::Real = 0.05,    a_sim::Integer = 100,    beta_i::Real = alpha_i,    beta::Real = alpha,    b_sim::Integer = a_sim)

Compute the OWA weights for the Tail Gini Range of a returns series [OWA].

Inputs

T: number of returns observations.

alpha_i: initial significance level of Tail Gini losses, 0 < alpha_i < alpha < 1.

a_sim: number of CVaRs to approximate the Tail Gini losses, a_sim > 0.

alpha: significance level of Tail Gini losses, alpha in (0, 1).

beta_i: initial significance level of Tail Gini gains, 0 < beta_i < beta < 1.

b_sim: number of CVaRs to approximate the Tail Gini gains, b_sim > 0.

beta: significance level of Tail Gini gains, beta in (0, 1).

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_rwcvar — Function

owa_rwcvar(
    T::Integer,    alphas::AbstractVector{<:Real},    weights_a::AbstractVector{<:Real},    betas::AbstractVector{<:Real} = alphas,    weights_b::AbstractVector{<:Real} = weights_b)

Compute the OWA weights for the Weighted Conditional Value at Risk (WCVaR) of a returns series [OWA].

Inputs

T: number of returns observations.

alphas: N×1 vector of significance levels of the losses for each CVaR model, where N is the number of losses models, each alpha in (0, 1).

weights_a: N×1 vector of weights of the losses for each CVaR model, where N is the number of losses models.

betas: M×1 vector of significance levels of the gains for each CVaR model, where M is the number of gains models, each beta in (0, 1).

weights_b: M×1 vector of weights of the gains for each CVaR model, where M is the number of gains models.

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_tg — Method

owa_tg(
    T::Integer;    alpha_i::Real = 0.0001,    alpha::Real = 0.05,    a_sim::Integer = 100)

Compute the OWA weights for the Tail Gini of a returns series [OWA].

Inputs

T: number of returns observations.

alpha_i: initial significance level of Tail Gini losses, 0 < alpha_i < alpha < 1.

a_sim: number of CVaRs to approximate the Tail Gini losses, a_sim > 0.

alpha: significance level of Tail Gini losses, alpha in (0, 1).

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_wcvar — Method

owa_wcvar(
    T::Integer,    alphas::AbstractVector{<:Real},    weights::AbstractVector{<:Real})

Compute the OWA weights for the Weighted Conditional Value at Risk (WCVaR) of a returns series [OWA].

Inputs

T: number of returns observations.

alphas: N×1 vector of significance levels of each CVaR model, where N is the number of models, each alpha in (0, 1).

weights: N×1 vector of weights of each CVaR model, where N is the number of models.

Outputs

w: T×1 ordered weight vector.

PortfolioOptimiser.owa_wr — Method

owa_wr(T::Integer)

Compute the OWA weights for the Worst Realisation (WR) of a returns series [OWA].

Inputs

T: number of returns observations.

Outputs

w: T×1 ordered weight vector.

Private

PortfolioOptimiser._crra_method — Method

_crra_method(weights::AbstractMatrix{<:Real}, k::Integer, g::Real)

Internal function for computing the Normalized Constant Relative Risk Aversion coefficients.

Inputs

weights: T×(k-1) matrix where T is the number of observations and k the order of the L-moments to combine, the i'th column contains the weights for the (i+1)'th L-moment.

k: the maximum order of the L-moments.

g: the risk aversion coefficient.

Outputs

w: T×1 ordered weight vector of the combined L-moments.

PortfolioOptimiser._optimise_owa — Method

_optimise_owa(model, solvers)

Internal function to optimise an OWA JuMP model.

Inputs

model: JuMP model.

solvers: Provides the solvers and corresponding parameters for solving the OWA L-Moment JuMP model. There can be two key => value pairs.

:solver => value: value is a JuMP optimiser. The optimiser can be declared alongside its attributes by using JuMP.solver_with_attributes.

:params => value: (optional) value must be a Dict or NamedTuple whose key => value pairs are the solver-specific settings.

Outputs

term_status: JuMP termination status.

solvers_tried: Dictionary that contains a dictionary of failed optimisations. Dict(key => Dict(...)), where key is the solver key used for the iteration of solver that failed.

If an MOI call fails on a model:

Dict(:jump_error => jump_error): JuMP error code.

If the optimiser fails to optimise the model satisfactorily:

Dict(:objective_val => JuMP.objective_value(model), :term_status => term_status, :params => haskey(val, :params) ? val[:params] : missing), where val is the value of the dictionary corresponding to key.

OWACajas, Dany, OWA Portfolio Optimization: A Disciplined Convex Programming Framework (December 18, 2021). Available at SSRN: https://ssrn.com/abstract=3988927 or http://dx.doi.org/10.2139/ssrn.3988927