Parameter Estimation

Public

PortfolioOptimiser.black_litterman — Method

black_litterman(bl::BLType, X::AbstractMatrix, P::AbstractMatrix,
                         Q::AbstractVector, w::AbstractVector;
                         cov_type::PortfolioOptimiserCovCor = PortCovCor(),
                         mu_type::MeanEstimator = MuSimple())

Private

PortfolioOptimiser._bl_mu_cov_w — Method

_mu_cov_w(tau, omega, P, Pi, Q, rf, sigma, delta, T, N, opt, cov_type, cov_flag = true)

Internal function for computing the Black Litterman statistics as defined in black_litterman. See .

Inputs

tau: variable of the same name in the Black-Litterman model.

omega: variable of the same name in the Black-Litterman model.

P: variable of the same name in the Black-Litterman model.

Pi: variable of the same name in the Black-Litterman model.

Q: variable of the same name in the Black-Litterman model.

rf: variable of the same name in the Black-Litterman model.

sigma: variable of the same name in the Black-Litterman model.

delta: variable of the same name in the Black-Litterman model.

T: variable of the same name in the Black-Litterman model.

N: variable of the same name in the Black-Litterman model.

opt: any valid instance of opt for .

cov_type: any valid value from .

cov_flag: whether the matrix is a covariance matrix or not.

Outputs

mu: asset expected returns vector obtained via the Black-Litterman model.

cov_mtx: asset covariance matrix obtained via the Black-Litterman model.

w: asset weights obtained via the Black-Litterman model.

Pi_: equilibrium excess returns after being adjusted by the views.

PortfolioOptimiser._posdef_fix! — Method

_posdef_fix!(method::PosdefNearest, X::AbstractMatrix)

Overload this for other posdef fix methods.