Black Litterman models

Public

PortfolioOptimiser.ABLType — Type

@kwdef mutable struct ABLType{T1 <: Real} <: BlackLittermanFactor
    constant::Bool = true
    eq::Bool = true
    delta::Union{<:Real, Nothing} = 1.0
    rf::T1 = 0.0
    posdef::PosdefFix = PosdefNearest()
    denoise::Denoise = NoDenoise()
    logo::AbstractLoGo = NoLoGo()
end

PortfolioOptimiser.BBLType — Type

@kwdef mutable struct BBLType{T1 <: Real} <: BlackLittermanFactor
    constant::Bool = true
    error::Bool = true
    delta::Union{<:Real, Nothing} = 1.0
    rf::T1 = 0.0
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    var_w::Union{<:AbstractWeights, Nothing} = nothing
end

PortfolioOptimiser.BLType — Type

@kwdef mutable struct BLType{T1 <: Real} <: BlackLitterman
    eq::Bool = true
    delta::Union{<:Real, Nothing} = 1.0
    rf::T1 = 0.0
    posdef::PosdefFix = PosdefNearest()
    denoise::Denoise = NoDenoise()
    logo::AbstractLoGo = NoLoGo()
end

PortfolioOptimiser.black_litterman — Method

black_litterman(bl::BLType, X::AbstractMatrix, P::AbstractMatrix,
                         Q::AbstractVector, w::AbstractVector;
                         cov_type::PortfolioOptimiserCovCor = PortCovCor(),
                         mu_type::MeanEstimator = MuSimple())

Private

PortfolioOptimiser.BlackLittermanFactor — Type

abstract type BlackLittermanFactor <: BlackLitterman end

PortfolioOptimiser._bl_mu_cov_w — Method

_mu_cov_w(tau, omega, P, Pi, Q, rf, sigma, delta, T, N, opt, cov_type, cov_flag = true)

Internal function for computing the Black Litterman statistics as defined in black_litterman. See .

Inputs

tau: variable of the same name in the Black-Litterman model.

omega: variable of the same name in the Black-Litterman model.

P: variable of the same name in the Black-Litterman model.

Pi: variable of the same name in the Black-Litterman model.

Q: variable of the same name in the Black-Litterman model.

rf: variable of the same name in the Black-Litterman model.

sigma: variable of the same name in the Black-Litterman model.

delta: variable of the same name in the Black-Litterman model.

T: variable of the same name in the Black-Litterman model.

N: variable of the same name in the Black-Litterman model.

opt: any valid instance of opt for .

cov_type: any valid value from .

cov_flag: whether the matrix is a covariance matrix or not.

Outputs

mu: asset expected returns vector obtained via the Black-Litterman model.

cov_mtx: asset covariance matrix obtained via the Black-Litterman model.

w: asset weights obtained via the Black-Litterman model.

Pi_: equilibrium excess returns after being adjusted by the views.