Covariance, correlation, cokurtosis and coskewness

Public

PortfolioOptimiser.CorDistance — Type

mutable struct CorDistance <: PortfolioOptimiserCovCor
    distance::Distances.UnionMetric
    dist_args::Tuple
    dist_kwargs::NamedTuple
    mean_w1::Union{<:AbstractWeights, Nothing}
    mean_w2::Union{<:AbstractWeights, Nothing}
    mean_w3::Union{<:AbstractWeights, Nothing}
end

PortfolioOptimiser.CorGerber0 — Type

mutable struct CorGerber0{T1, T2, T3, T4, T5} <: CorSB
    normalise::Bool
    threshold::T1
    ve::StatsBase.CovarianceEstimator
    std_w::Union{<:AbstractWeights, Nothing}
    mean_w::Union{<:AbstractWeights, Nothing}
    posdef::PosdefFix
end

PortfolioOptimiser.CorGerber1 — Type

mutable struct CorGerber1{T1, T2, T3, T4, T5} <: CorSB
    normalise::Bool
    threshold::T1
    ve::StatsBase.CovarianceEstimator
    std_w::Union{<:AbstractWeights, Nothing}
    mean_w::Union{<:AbstractWeights, Nothing}
    posdef::PosdefFix
end

PortfolioOptimiser.CorGerber2 — Type

mutable struct CorGerber2{T1, T2, T3, T4, T5} <: CorSB
    normalise::Bool
    threshold::T1
    ve::StatsBase.CovarianceEstimator
    std_w::Union{<:AbstractWeights, Nothing}
    mean_w::Union{<:AbstractWeights, Nothing}
    posdef::PosdefFix
end

PortfolioOptimiser.CorGerberSB0 — Type

mutable struct CorGerberSB0{T1, T2, T3, T4, T5} <: CorSB
    normalise::Bool
    threshold::T1
    c1::T2
    c2::T3
    c3::T4
    n::T5
    ve::StatsBase.CovarianceEstimator
    std_w::Union{<:AbstractWeights, Nothing}
    mean_w::Union{<:AbstractWeights, Nothing}
    posdef::PosdefFix
end

PortfolioOptimiser.CorGerberSB1 — Type

mutable struct CorGerberSB1{T1, T2, T3, T4, T5} <: CorSB
    normalise::Bool
    threshold::T1
    c1::T2
    c2::T3
    c3::T4
    n::T5
    ve::StatsBase.CovarianceEstimator
    std_w::Union{<:AbstractWeights, Nothing}
    mean_w::Union{<:AbstractWeights, Nothing}
    posdef::PosdefFix
end

PortfolioOptimiser.CorKendall — Type

@kwdef mutable struct CorKendall <: CorRank
    absolute::Bool = false
end

Kendall type correlation estimator.

Parameters

absolute:

if flag: compute an absolute correlation, abs.(corkendall(X)).

PortfolioOptimiser.CorLTD — Type

mutable struct CorLTD <: PortfolioOptimiserCovCor
    alpha::Real
    ve::StatsBase.CovarianceEstimator
    w::Union{<:AbstractWeights, Nothing}
end

PortfolioOptimiser.CorMutualInfo — Type

@kwdef mutable struct CorMutualInfo <: PortfolioOptimiserCovCor
    bins::Union{<:Integer, <:AbstractBins} = HGR()
    normalise::Bool = true
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    w::Union{<:AbstractWeights, Nothing} = nothing
end

Define the mutual information correlation matrix.

Parameters

bins: defines the bin function, or bin width directly and if so bins > 0.

normalise: whether or not to normalise the mutual information.

ve: variance estimator.

w: variance weights, if nothing

PortfolioOptimiser.CorSB0 — Type

mutable struct CorSB0{T1, T2, T3, T4, T5} <: CorSB
    normalise::Bool
    threshold::T1
    c1::T2
    c2::T3
    c3::T4
    n::T5
    ve::StatsBase.CovarianceEstimator
    std_w::Union{<:AbstractWeights, Nothing}
    mean_w::Union{<:AbstractWeights, Nothing}
    posdef::PosdefFix
end

PortfolioOptimiser.CorSB1 — Type

mutable struct CorSB1{T1, T2, T3, T4, T5} <: CorSB
    normalise::Bool
    threshold::T1
    c1::T2
    c2::T3
    c3::T4
    n::T5
    ve::StatsBase.CovarianceEstimator
    std_w::Union{<:AbstractWeights, Nothing}
    mean_w::Union{<:AbstractWeights, Nothing}
    posdef::PosdefFix
end

PortfolioOptimiser.CorSpearman — Type

@kwdef mutable struct CorSpearman <: CorRank
    absolute::Bool = false
end

Spearman type correlation estimator.

Parameters

absolute:

if true: compute an absolute correlation, abs.(corspearman(X)).

PortfolioOptimiser.CovFull — Type

@kwdef mutable struct CovFull <: CorPearson
    absolute::Bool = false
    ce::StatsBase.CovarianceEstimator = StatsBase.SimpleCovariance(; corrected = true)
    w::Union{<:AbstractWeights, Nothing} = nothing
end

Full Pearson-type covariance and correlation estimator.

Parameters

absolute:

if true: compute an absolute correlation, abs.(cor(X)).

ce: covariance estimator.

w: weights for computing the covariance, if nothing apply no weights.

PortfolioOptimiser.CovSemi — Type

@kwdef mutable struct CovSemi <: CorPearson
    absolute::Bool = false
    ce::StatsBase.CovarianceEstimator = StatsBase.SimpleCovariance(; corrected = true)
    target::Union{<:Real, AbstractVector{<:Real}} = 0.0
    w::Union{<:AbstractWeights, Nothing} = nothing
end

Semi Pearson-type covariance and correlation estimator.

Parameters

absolute:

if true: compute an absolute correlation, abs.(cor(X)).

ce: covariance estimator.

target: minimum acceptable return target.

if isa(target, Real): apply the same target to all assets.

if isa(target, AbstractVector): apply individual target to each asset.

w: weights for computing the covariance, if nothing apply no weights.

PortfolioOptimiser.KurtFull — Type

@kwdef mutable struct KurtFull <: KurtEstimator
    posdef::PosdefFix = PosdefNearest(;)
    denoise::Denoise = NoDenoise(;)
    logo::AbstractLoGo = NoLoGo(;)
end

PortfolioOptimiser.KurtSemi — Type

@kwdef mutable struct KurtSemi <: KurtEstimator
    target::Union{<:Real, AbstractVector{<:Real}} = 0.0
    posdef::PosdefFix = PosdefNearest(;)
    denoise::Denoise = NoDenoise(;)
    logo::AbstractLoGo = NoLoGo(;)
end

PortfolioOptimiser.LoGo — Type

@kwdef mutable struct LoGo <: AbstractLoGo
    distance::DistanceMethod = DistanceMLP()
    similarity::DBHTSimilarity = DBHTMaxDist()
end

PortfolioOptimiser.NoLoGo — Type

struct NoLoGo <: AbstractLoGo end

PortfolioOptimiser.PortCovCor — Type

@kwdef mutable struct PortCovCor <: PortfolioOptimiserCovCor
    ce::CovarianceEstimator = CovFull(;)
    posdef::PosdefFix = PosdefNearest(;)
    denoise::Denoise = NoDenoise(;)
    logo::AbstractLoGo = NoLoGo(;)
end

PortfolioOptimiser.SimpleVariance — Type

@kwdef mutable struct SimpleVariance <: StatsBase.CovarianceEstimator
    corrected::Bool = true
end

Simple variance estimator.

Parameters

corrected: if true correct the bias by dividing by N-1, if false the bias is not corrected and the division is by N.

PortfolioOptimiser.SkewFull — Type

struct SkewFull <: SkewEstimator end

PortfolioOptimiser.SkewSemi — Type

@kwdef mutable struct SkewSemi <: SkewEstimator
    target::Union{<:Real, AbstractVector{<:Real}} = 0.0
end

PortfolioOptimiser.coskew — Method

coskew(::SkewFull, X::AbstractMatrix, mu::AbstractVector)

PortfolioOptimiser.coskew — Method

coskew(se::SkewSemi, X::AbstractMatrix, mu::AbstractVector)

Private

PortfolioOptimiser.AbstractLoGo — Type

abstract type AbstractLoGo end

PortfolioOptimiser.CorPearson — Type

abstract type CorPearson <: PortfolioOptimiserCovCor end

Abstract type for subtyping Pearson type covariance estimators.

PortfolioOptimiser.CorRank — Type

abstract type CorRank <: PortfolioOptimiserCovCor end

Abstract type for subtyping rank based covariance estimators.

PortfolioOptimiser.PortfolioOptimiserCovCor — Type

abstract type PortfolioOptimiserCovCor <: StatsBase.CovarianceEstimator end

Abstract type for subtyping portfolio optimiser covaraince and correlation estimators.