Covariance, correlation, cokurtosis and coskewness

Public

PortfolioOptimiser.CorKendall — Type

@kwdef mutable struct CorKendall <: CorRank
    absolute::Bool = false
end

Kendall type correlation estimator.

Parameters

absolute:

if true: compute an absolute correlation, abs.(corkendall(X)).

PortfolioOptimiser.CorSpearman — Type

@kwdef mutable struct CorSpearman <: CorRank
    absolute::Bool = false
end

Spearman type correlation estimator.

Parameters

absolute:

if true: compute an absolute correlation, abs.(corspearman(X)).

PortfolioOptimiser.CovDistance — Type

mutable struct CovDistance <: PortfolioOptimiserCovCor
    distance::Distances.Metric
    dist_args::Tuple
    dist_kwargs::NamedTuple
    mean_w1::Union{<:AbstractWeights, Nothing}
    mean_w2::Union{<:AbstractWeights, Nothing}
    mean_w3::Union{<:AbstractWeights, Nothing}
end

Distance covariance and correlation matrix estimator.

Parameters

distance: distance metric from Distances.jl.

dist_args: args for the Distances.pairwise function of Distances.jl.

dist_kwargs: kwargs for the Distances.pairwise function of Distances.jl.

mean_w1: optional T×1 vector of weights for computing the mean of the pairwise distance matrices along its rows (dims = 1).

mean_w2: optional T×1 vector of weights for computing the mean of the pairwise distance matrices along its columns (dims = 2).

mean_w3: optional T×1 vector of weights for computing the mean of the entirety of the pairwise distance matrices.

PortfolioOptimiser.CovFull — Type

@kwdef mutable struct CovFull <: CorPearson
    absolute::Bool = false
    ce::StatsBase.CovarianceEstimator = StatsBase.SimpleCovariance(; corrected = true)
    w::Union{<:AbstractWeights, Nothing} = nothing
end

Full Pearson-type covariance and correlation estimator.

Parameters

absolute:

if true: compute an absolute correlation, abs.(cor(X)).

ce: covariance estimator.

w: optional T×1 vector of weights for computing the covariance.

PortfolioOptimiser.CovGerber0 — Type

@kwdef mutable struct CovGerber0{T1 <: Real} <: CovGerberBasic
    normalise::Bool = false
    threshold::T1 = 0.5
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    std_w::Union{<:AbstractWeights, Nothing} = nothing
    mean_w::Union{<:AbstractWeights, Nothing} = nothing
    posdef::AbstractPosdefFix = PosdefNearest()
end

Gerber type 0 covariance and correlation matrices.

Parameters

normalise:

if true: Z-normalise the data before applying the Gerber criteria.

threshold: Gerber significance threshold, threshold ∈ (0, 1).

ve: variance estimator StatsBase.CovarianceEstimator.

std_w: optional T×1 vector of weights for computing the variance.

Only used when normalise == true:

mean_w: optional T×1 vector of weights for computing the mean.

posdef: type for fixing the Gerber type 0 correaltion matrix AbstractPosdefFix.

PortfolioOptimiser.CovGerber1 — Type

@kwdef mutable struct CovGerber1{T1 <: Real} <: CovGerberBasic
    normalise::Bool = false
    threshold::T1 = 0.5
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    std_w::Union{<:AbstractWeights, Nothing} = nothing
    mean_w::Union{<:AbstractWeights, Nothing} = nothing
    posdef::AbstractPosdefFix = PosdefNearest()
end

Gerber type 1 covariance and correlation matrices.

Parameters

normalise:

if true: Z-normalise the data before applying the Gerber criteria.

threshold: Gerber significance threshold, threshold ∈ (0, 1).

ve: variance estimator StatsBase.CovarianceEstimator.

std_w: optional T×1 vector of weights for computing the variance.

Only used when normalise == true:

mean_w: optional T×1 vector of weights for computing the mean.

posdef: type for fixing the Gerber type 1 correaltion matrix AbstractPosdefFix.

PortfolioOptimiser.CovGerber2 — Type

@kwdef mutable struct CovGerber2{T1 <: Real} <: CovGerberBasic
    normalise::Bool = false
    threshold::T1 = 0.5
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    std_w::Union{<:AbstractWeights, Nothing} = nothing
    mean_w::Union{<:AbstractWeights, Nothing} = nothing
    posdef::AbstractPosdefFix = PosdefNearest()
end

Gerber type 2 covariance and correlation matrices.

Parameters

normalise:

if true: Z-normalise the data before applying the Gerber criteria.

threshold: Gerber significance threshold, threshold ∈ (0, 1).

ve: variance estimator StatsBase.CovarianceEstimator.

std_w: optional T×1 vector of weights for computing the variance.

Only used when normalise == true:

mean_w: optional T×1 vector of weights for computing the mean.

posdef: type for fixing the Gerber type 2 correaltion matrix AbstractPosdefFix.

PortfolioOptimiser.CovGerberSB0 — Type

@kwdef mutable struct CovGerberSB0{T1, T2, T3, T4, T5} <: CovSB
    normalise::Bool = false
    threshold::T1 = 0.5
    c1::T2 = 0.5
    c2::T3 = 0.5
    c3::T4 = 4.0
    n::T5 = 2.0
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    std_w::Union{<:AbstractWeights, Nothing} = nothing
    mean_w::Union{<:AbstractWeights, Nothing} = nothing
    posdef::AbstractPosdefFix = PosdefNearest()
end

Smyth-Broby modification with vote counting of the Gerber type 0 covariance and correlation matrices.

Parameters

normalise:

if true: Z-normalise the data before applying the Gerber criteria.

threshold: Gerber significance threshold, threshold ∈ (0, 1).

c1: confusion zone threshold ($c_1$ in the paper), c1 ∈ (0, 1].

c2: indecision zone threshold ($c_2$ in the paper), c2 ∈ (0, 1].

c3: large co-movement threshold (4 in the paper).

n: exponent of the regularisation term ($n = 2$ in the paper).

ve: variance estimator StatsBase.CovarianceEstimator.

std_w: optional T×1 vector of weights for computing the variance.

mean_w: optional T×1 vector of weights for computing the mean.

posdef: type for fixing the Smyth-Broby modification with vote counting of the Gerber type 0 correaltion matrix AbstractPosdefFix.

PortfolioOptimiser.CovGerberSB1 — Type

@kwdef mutable struct CovGerberSB1{T1, T2, T3, T4, T5} <: CovSB
    normalise::Bool = false
    threshold::T1 = 0.5
    c1::T2 = 0.5
    c2::T3 = 0.5
    c3::T4 = 4.0
    n::T5 = 2.0
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    std_w::Union{<:AbstractWeights, Nothing} = nothing
    mean_w::Union{<:AbstractWeights, Nothing} = nothing
    posdef::AbstractPosdefFix = PosdefNearest()
end

Smyth-Broby modification with vote counting of the Gerber type 1 covariance and correlation matrices.

Parameters

normalise:

if true: Z-normalise the data before applying the Gerber criteria.

threshold: Gerber significance threshold, threshold ∈ (0, 1).

c1: confusion zone threshold ($c_1$ in the paper), c1 ∈ (0, 1].

c2: indecision zone threshold ($c_2$ in the paper), c2 ∈ (0, 1].

c3: large co-movement threshold (4 in the paper).

n: exponent of the regularisation term ($n = 2$ in the paper).

ve: variance estimator StatsBase.CovarianceEstimator.

std_w: optional T×1 vector of weights for computing the variance.

mean_w: optional T×1 vector of weights for computing the mean.

posdef: type for fixing the Smyth-Broby modification with vote counting of the Gerber type 1 correaltion matrix AbstractPosdefFix.

PortfolioOptimiser.CovGerberSB2 — Type

mutable struct CovGerberSB2{T1, T2, T3, T4, T5} <: CovSB

PortfolioOptimiser.CovLTD — Type

mutable struct CovLTD <: PortfolioOptimiserCovCor
    alpha::Real
    ve::StatsBase.CovarianceEstimator
    w::Union{<:AbstractWeights, Nothing}
end

Lower tail dependence correlation and covariance matrix estimator.

Parameters

alpha: significance level of the lower tail dependence, alpha ∈ (0, 1).

Only used when computing covariance matrices:

ve: variance estimator StatsBase.CovarianceEstimator.

w: optional T×1 vector of weights for computing the variance.

PortfolioOptimiser.CovMutualInfo — Type

@kwdef mutable struct CovMutualInfo <: PortfolioOptimiserCovCor
    bins::Union{<:Integer, <:AbstractBins} = HGR()
    normalise::Bool = true
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    w::Union{<:AbstractWeights, Nothing} = nothing
end

Mutual information correlation matrix estimator.

Parameters

bins:

if isa(bins, AbstractBins): defines the function for computing bin widths.

if isa(bins, Integer) and bins > 0: directly provide the number of bins.

normalise:

if true: normalise the mutual information.

Only used when computing covariance matrices:

ve: variance estimator StatsBase.CovarianceEstimator.

w: optional T×1 vector of weights for computing the variance.

PortfolioOptimiser.CovSB0 — Type

@kwdef mutable struct CovSB0{T1, T2, T3, T4, T5} <: CovSB
    normalise::Bool = false
    threshold::T1 = 0.5
    c1::T2 = 0.5
    c2::T3 = 0.5
    c3::T4 = 4.0
    n::T5 = 2.0
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    std_w::Union{<:AbstractWeights, Nothing} = nothing
    mean_w::Union{<:AbstractWeights, Nothing} = nothing
    posdef::AbstractPosdefFix = PosdefNearest()
end

Smyth-Broby modification of the Gerber type 0 covariance and correlation matrices.

Parameters

normalise:

if true: Z-normalise the data before applying the Gerber criteria.

threshold: Gerber significance threshold, threshold ∈ (0, 1).

c1: confusion zone threshold ($c_1$ in the paper), c1 ∈ (0, 1].

c2: indecision zone threshold ($c_2$ in the paper), c2 ∈ (0, 1].

c3: large co-movement threshold (4 in the paper).

n: exponent of the regularisation term ($n = 2$ in the paper).

ve: variance estimator StatsBase.CovarianceEstimator.

std_w: optional T×1 vector of weights for computing the variance.

mean_w: optional T×1 vector of weights for computing the mean.

posdef: type for fixing the Smyth-Broby modification of the Gerber type 0 correaltion matrix AbstractPosdefFix.

PortfolioOptimiser.CovSB1 — Type

@kwdef mutable struct CovSB1{T1, T2, T3, T4, T5} <: CovSB
    normalise::Bool = false
    threshold::T1 = 0.5
    c1::T2 = 0.5
    c2::T3 = 0.5
    c3::T4 = 4.0
    n::T5 = 2.0
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    std_w::Union{<:AbstractWeights, Nothing} = nothing
    mean_w::Union{<:AbstractWeights, Nothing} = nothing
    posdef::AbstractPosdefFix = PosdefNearest()
end

Smyth-Broby modification of the Gerber type 1 covariance and correlation matrices.

Parameters

normalise:

if true: Z-normalise the data before applying the Gerber criteria.

threshold: Gerber significance threshold, threshold ∈ (0, 1).

c1: confusion zone threshold ($c_1$ in the paper), c1 ∈ (0, 1].

c2: indecision zone threshold ($c_2$ in the paper), c2 ∈ (0, 1].

c3: large co-movement threshold (4 in the paper).

n: exponent of the regularisation term ($n = 2$ in the paper).

ve: variance estimator StatsBase.CovarianceEstimator.

std_w: optional T×1 vector of weights for computing the variance.

mean_w: optional T×1 vector of weights for computing the mean.

posdef: type for fixing the Smyth-Broby modification of the Gerber type 1 correaltion matrix AbstractPosdefFix.

PortfolioOptimiser.CovSB2 — Type

CovSB2{T1, T2, T3, T4, T5} <: CovSB

PortfolioOptimiser.CovSemi — Type

@kwdef mutable struct CovSemi <: CorPearson
    absolute::Bool = false
    ce::StatsBase.CovarianceEstimator = StatsBase.SimpleCovariance(; corrected = true)
    target::Union{<:Real, AbstractVector{<:Real}} = 0.0
    w::Union{<:AbstractWeights, Nothing} = nothing
end

Semi Pearson-type covariance and correlation estimator.

Parameters

absolute:

if true: compute an absolute correlation, abs.(cor(X)).

ce: covariance estimator.

target: minimum return threshold for classifying downside returns.

if isa(target, Real): apply the same target to all assets.

if isa(target, AbstractVector): apply individual target to each asset.

w: optional T×1 vector of weights for computing the covariance.

PortfolioOptimiser.KurtFull — Type

@kwdef mutable struct KurtFull <: KurtEstimator
    posdef::AbstractPosdefFix = PosdefNearest(;)
    denoise::AbstractDenoise = NoDenoise(;)
    logo::AbstractLoGo = NoLoGo(;)
end

Full cokurtosis estimator.

Parameters

posdef: type for fixing non a positive definite cokurtosis matrix AbstractPosdefFix.

denoise: type for denoising the cokurtosis matrix AbstractDenoise.

logo: type for computing the LoGo cokurtosis matrix AbstractLoGo.

PortfolioOptimiser.KurtSemi — Type

@kwdef mutable struct KurtSemi <: KurtEstimator
    target::Union{<:Real, AbstractVector{<:Real}} = 0.0
    posdef::AbstractPosdefFix = PosdefNearest(;)
    denoise::AbstractDenoise = NoDenoise(;)
    logo::AbstractLoGo = NoLoGo(;)
end

Semi cokurtosis estimator.

Parameters

target: minimum return threshold for classifying downside returns.

if isa(target, Real): apply the same target to all assets.

if isa(target, AbstractVector): apply individual target to each asset.

posdef: type for fixing non a positive definite semi cokurtosis matrix AbstractPosdefFix.

denoise: type for denoising the semi cokurtosis matrix AbstractDenoise.

logo: type for computing the LoGo semi cokurtosis matrix AbstractLoGo.

PortfolioOptimiser.PortCovCor — Type

@kwdef mutable struct PortCovCor <: PortfolioOptimiserCovCor
    ce::CovarianceEstimator = CovFull(;)
    posdef::AbstractPosdefFix = PosdefNearest(;)
    denoise::AbstractDenoise = NoDenoise(;)
    logo::AbstractLoGo = NoLoGo(;)
end

PortfolioOptimiser covariance and correlation estimator.

Parameters

ce: covariance estimator.

posdef: type for fixing the portfolio covariance or correlation matrix AbstractPosdefFix.

denoise: type for denoising the portfolio covariance or correlation matrix AbstractDenoise.

logo: type for computing the LoGo portfolio covariance or correlation matrix AbstractLoGo.

PortfolioOptimiser.SimpleVariance — Type

@kwdef mutable struct SimpleVariance <: StatsBase.CovarianceEstimator
    corrected::Bool = true
end

Simple variance estimator.

Parameters

corrected:

if true: correct the bias dividing by N-1 instead of N.

PortfolioOptimiser.SkewFull — Type

struct SkewFull <: SkewEstimator end

Full cokurtosis estimator.

PortfolioOptimiser.SkewSemi — Type

@kwdef mutable struct SkewSemi <: SkewEstimator
    target::Union{<:Real, AbstractVector{<:Real}} = 0.0
end

Semi cokurtosis estimator.

Parameters

target: minimum return threshold for classifying downside returns.

if isa(target, Real): apply the same target to all assets.

if isa(target, AbstractVector): apply individual target to each asset.

PortfolioOptimiser.coskew — Method

coskew(::SkewFull, X::AbstractMatrix, mu::AbstractVector)

PortfolioOptimiser.coskew — Method

coskew(se::SkewSemi, X::AbstractMatrix, mu::AbstractVector)

Private

PortfolioOptimiser.PosdefFixCovCor — Type

const PosdefFixCovCor = Union{<:CovGerber, PortCovCor}

Covariance and correlation estimators that support positive definite fixes.

PortfolioOptimiser.CorPearson — Type

abstract type CorPearson <: PortfolioOptimiserCovCor end

Abstract type for subtyping Pearson type covariance estimators.

PortfolioOptimiser.CorRank — Type

abstract type CorRank <: PortfolioOptimiserCovCor end

Abstract type for subtyping rank based covariance estimators.

PortfolioOptimiser.CovGerber — Type

abstract type CovGerber <: PortfolioOptimiserCovCor end

Abstract type for subtyping Gerber type covariance and correlation estimators.

PortfolioOptimiser.CovGerberBasic — Type

abstract type CovGerberBasic <: CovGerber end

Abstract type for subtyping the original Gerber type covariance and correlation estimators.

PortfolioOptimiser.CovGerberSB — Type

abstract type CovSB <: CovGerber end

Abstract type for subtyping the Smyth-Broby modifications with vote counting of Gerber type covariance and correlation estimators.

PortfolioOptimiser.CovSB — Type

abstract type CovSB <: CovGerber end

Abstract type for subtyping the Smyth-Broby modifications of Gerber type covariance and correlation estimators.

PortfolioOptimiser.KurtEstimator — Type

abstract type KurtEstimator end

Abstract type for subtyping cokurtosis estimators.

PortfolioOptimiser.PortfolioOptimiserCovCor — Type

abstract type PortfolioOptimiserCovCor <: StatsBase.CovarianceEstimator end

Abstract type for subtyping portfolio covariance and correlation estimators.

PortfolioOptimiser.SkewEstimator — Type

abstract type SkewEstimator end

Abstract type for subtyping coskew estimators.