Distance matrices

Public

PortfolioOptimiser.DistanceCanonical — Type

struct DistanceCanonical <: DistanceMethod end

Struct for computing the canonical distance for a given correlation subtype of PortfolioOptimiserCovCor.

PortfolioOptimiser.DistanceLog — Type

struct DistanceLog <: DistanceMethod end

Defines the log-distance matrix from the correlation matrix.

\[D_{i,\,j} = -\log\left(C_{i,\,j}\right)\,.\]

Where:

$D_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N log-distance matrix.

$C_{i,\,j}$: is the $(i,\,j)$-th entry of an absolute correlation matrix.

PortfolioOptimiser.DistanceMLP — Type

@kwdef mutable struct DistanceMLP <: DistanceMethod
    absolute::Bool = false
end

Defines the distance matrix from a correlation matrix [2] in dist.

\[D_{i,\,j} = \begin{cases} \sqrt{\dfrac{1}{2} \left(1 - C_{i,\,j}\right)} &\quad \mathrm{if~ absolute = false}\\ \sqrt{1 - \lvert C_{i,\,j} \rvert} &\quad \mathrm{if~ absolute = true}\nonumber\,. \end{cases}\]

Where:

$D_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N distance matrix $\mathbf{C}$.

$C_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N correlation matrix $\mathbf{D}$.

absolute:

if true: the correlation being used is absolute.

Parameters

absolute:

if true: the correlation being used is absolute.

PortfolioOptimiser.DistanceSqMLP — Type

@kwdef mutable struct DistanceSqMLP <: DistanceMethod
    absolute::Bool = false
    distance::Distances.UnionMetric = Distances.Euclidean()
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the distance of distances matrix from a correlation matrix [2] in dist.

\[\tilde{D}_{i,\,j} = f_{m}\left(\bm{D}_{i},\, \bm{D}_j\right)\,.\]

Where:

$\bm{D}_{i}$: is the $i$-th column/row of the N×N distance matrix defined in DistanceMLP.

$f_{m}$: is the pairwise distance function for metric $m$. We use the Distances.pairwise function which computes the entire matrix at once, the output is a vector so it gets reshaped into an N×N matrix.

$\tilde{D}_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N distances of distances matrix.

absolute:

if true: the correlation being used is absolute.

Parameters

absolute:

if true: the correlation being used is absolute.

distance: distance metric from Distances.jl.

args: args for the Distances.pairwise function.

kwargs: key word args for the Distances.pairwise function.

PortfolioOptimiser.DistanceVarInfo — Type

@kwdef mutable struct DistanceVarInfo <: DistanceMethod
    bins::Union{<:Integer, <:AbstractBins} = HGR()
    normalise::Bool = true
end

Defines the variation of information distance matrix.

Parameters

bins: defines the bin function, or bin width directly and if so bins > 0.

normalise: whether or not to normalise the variation of information.

PortfolioOptimiser.Freedman — Type

struct Freedman <: AstroBins end

Freedman's bin width algorithm from astropy.

PortfolioOptimiser.HGR — Type

struct HGR <: AbstractBins end

Hacine-Gharbi and Ravier's bin width algorithm [3].

PortfolioOptimiser.Knuth — Type

struct Knuth <: AstroBins end

Knuth's bin width algorithm from astropy.

PortfolioOptimiser.Scott — Type

struct Scott <: AstroBins end

Scott's bin width algorithm from astropy.

PortfolioOptimiser.dist — Method

dist(de::DistanceMethod, X, Y)

Private

PortfolioOptimiser.AbstractBins — Type

abstract type AbstractBins end

Abstract type for defining bin width estimation functions when computing DistanceVarInfo and CorMutualInfo distance and correlation matrices respectively.

PortfolioOptimiser.AstroBins — Type

abstract type AstroBins <: AbstractBins end

Abstract type for defining which bin width function to use from astropy.

PortfolioOptimiser.DistanceMethod — Type

abstract type DistanceMethod end

Abstract type for subtyping methods for computing distance matrices from correlation ones.