Worst case mean variance sets

Public

PortfolioOptimiser.ArchWC — Type

@kwdef mutable struct ArchWC{T1 <: Integer, T2 <: Integer, T3 <: Real} <: WorstCaseType
    bootstrap::WorstCaseArchType = StationaryBS()
    n_sim::T1 = 3_000
    block_size::T2 = 3
    q::T3 = 0.05
    seed::Union{<:Integer, Nothing} = nothing
end

PortfolioOptimiser.CircularBS — Type

struct CircularBS <: WorstCaseArchType end

Circular block bootstrap.

PortfolioOptimiser.DeltaWC — Type

@kwdef mutable struct DeltaWC{T1 <: Real, T2 <: Real} <: WorstCaseType
    dcov::T1 = 0.1
    dmu::T2 = 0.1
end

PortfolioOptimiser.KGeneralWC — Type

struct KGeneralWC <: WorstCaseKType end

PortfolioOptimiser.KNormalWC — Type

struct KNormalWC <: WorstCaseKType end

PortfolioOptimiser.MovingBS — Type

struct MovingBS <: WorstCaseArchType end

Moving block bootstrap.

PortfolioOptimiser.NormalWC — Type

@kwdef mutable struct NormalWC{T1 <: Integer, T2 <: Real} <: WorstCaseType
    n_sim::T1 = 3_000
    q::T2 = 0.05
    rng::AbstractRNG = Random.default_rng()
    seed::Union{<:Integer, Nothing} = nothing
end

PortfolioOptimiser.StationaryBS — Type

struct StationaryBS <: WorstCaseArchType end

Stationary bootstrap.

PortfolioOptimiser.WCType — Type

@kwdef mutable struct WCType
    cov_type::PortfolioOptimiserCovCor = PortCovCor(;)
    mu_type::MeanEstimator = MuSimple(;)
    box::WorstCaseType = NormalWC(;)
    ellipse::WorstCaseType = NormalWC(;)
    k_sigma::Union{<:Real, WorstCaseKType} = KNormalWC(;)
    k_mu::Union{<:Real, WorstCaseKType} = KNormalWC(;)
    posdef::AbstractPosdefFix = PosdefNearest(;)
    diagonal::Bool = false
end

PortfolioOptimiser.WorstCaseSet — Type

abstract type WorstCaseSet end

Abstract type for subtyping worst case mean variance set types.

PortfolioOptimiser.Box — Type

struct Box <: WCSetMuSigma end

Box sets for worst case mean variance optimisation.

PortfolioOptimiser.Ellipse — Type

struct Ellipse <: WCSetMuSigma end

Elliptical sets for worst case mean variance optimisation.

PortfolioOptimiser.NoWC — Type

@kwdef mutable struct NoWC <: WorstCaseSet
    formulation::VarianceFormulation = SOC()
end

Use no set for worst case mean variance optimisation.

Parameters

formulation: quadratic expression formulation of SD risk measure to use VarianceFormulation.

Private

PortfolioOptimiser.WCSetMuSigma — Type

PortfolioOptimiser.WCSetMu — Type

PortfolioOptimiser.WorstCaseArchType — Type

abstract type WorstCaseArchType <: WorstCaseType end

Abstract type for subtyping types for computing worst case mean variance for bootstrapping with arch.

PortfolioOptimiser.WorstCaseKType — Type

abstract type WorstCaseKType end

PortfolioOptimiser.WorstCaseType — Type

abstract type WorstCaseType end

Abstract type for subtyping types for computing worst case mean variance sets.