Parameter Estimation Types

Public

PortfolioOptimiser.CorDistance — Type

mutable struct CorDistance <: PortfolioOptimiserCovCor
    distance::Distances.UnionMetric
    dist_args::Tuple
    dist_kwargs::NamedTuple
    mean_w1::Union{<:AbstractWeights, Nothing}
    mean_w2::Union{<:AbstractWeights, Nothing}
    mean_w3::Union{<:AbstractWeights, Nothing}
end

PortfolioOptimiser.CorGerberSB1 — Type

mutable struct CorGerberSB1{T1, T2, T3, T4, T5} <: CorSB
    normalise::Bool
    threshold::T1
    c1::T2
    c2::T3
    c3::T4
    n::T5
    ve::StatsBase.CovarianceEstimator
    std_w::Union{<:AbstractWeights, Nothing}
    mean_w::Union{<:AbstractWeights, Nothing}
    posdef::PosdefFix
end

PortfolioOptimiser.CorLTD — Type

mutable struct CorLTD <: PortfolioOptimiserCovCor
    alpha::Real
    ve::StatsBase.CovarianceEstimator
    std_w::Union{<:AbstractWeights, Nothing}
end

PortfolioOptimiser.CorMutualInfo — Type

mutable struct CorMutualInfo <: PortfolioOptimiserCovCor
    bins::Union{<:Integer, <:AbstractBins}
    normalise::Bool
    ve::StatsBase.CovarianceEstimator
    std_w::Union{<:AbstractWeights, Nothing}
end

PortfolioOptimiser.DistanceMLP — Type

@kwdef mutable struct DistanceMLP <: DistanceMethod
    absolute::Bool = false
end

Defines the distance matrix from a correlation matrix [1] in dist:

\[\begin{align} \mathbf{D} &= \begin{cases} \sqrt{\dfrac{1}{2} \left(\mathbf{1} - \mathbf{C}\right)} &\quad \mathrm{if~ absolute = false}\\ \\ \sqrt{\mathbf{1} - \lvert\mathbf{C}\rvert} &\quad \mathrm{if~ absolute = true} \end{cases} \end{align}\]

Where:

$\mathbf{D}$: is the N×N distance matrix.

$\mathbf{C}$: is the N×N correlation matrix.

absolute: is a flag whether the correlation is absolute.

Parameters

absolute: flag for stating whether or not an absolute correlation is being used.

PortfolioOptimiser.DistanceSqMLP — Type

@kwdef mutable struct DistanceSqMLP <: DistanceMethod
    absolute::Bool = false
    distance::Distances.UnionMetric = Distances.Euclidean()
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the distance of distances matrix from a correlation matrix [1] in dist:

\[\begin{align} \mathbf{d} &= \begin{cases} \sqrt{\dfrac{1}{2} \left(\mathbf{1} - \mathbf{C}\right)} &\quad \mathrm{if~ absolute = false}\\ \\ \sqrt{\mathbf{1} - \lvert\mathbf{C}\rvert} &\quad \mathrm{if~ absolute = true} \end{cases}\\ \mathbf{D} &= \mathrm{pairwise}(m,\, \mathbf{d}) \end{align}\]

Where:

$\mathbf{d}$: is the N×N distance matrix.

$\mathbf{C}$: is the N×N correlation matrix.

$\mathrm{pairwise}$: is Distances.pairwise, with its output reshaped to the appropriate shape.

$m$: is a concrete subtype of Distances.UnionMetric from Distances.jl.

$\mathbf{D}$: is the N×N distance of distances matrix.

absolute: is a flag whether the correlation is absolute.

Parameters

absolute: flag for stating whether or not an absolute correlation is being used.

distance: distance metric from Distances.jl.

args: args for the Distances.pairwise function.

kwargs: key word args for the Distances.pairwise function.

PortfolioOptimiser.Fixed — Type

@kwdef mutable struct Fixed{T1, T2, T3, T4} <: Denoise
    detone::Bool = false
    mkt_comp::Integer = 1
    kernel = AverageShiftedHistograms.Kernels.gaussian
    m::Integer = 10
    n::Integer = 1000
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the parameters for using the fixed method in denoise! [2, Chapter 2].

Parameters

detone:

true: remove the largest mkt_comp eigenvalues from the correlation matrix.

Warning

Removing eigenvalues from the matrix may make it singular.

mkt_comp: the number of largest eigenvalues to remove from the correlation matrix.

kernel: kernel for fitting the average shifted histograms from AverageShiftedHistograms.jl Kernel Functions.

m: number of adjacent histograms to smooth over AverageShiftedHistograms.jl Usage.

n: number of points used when creating the range of values to which the average shifted histogram is to be fitted AverageShiftedHistograms.jl Usage.

args: arguments for Optim.optimize

kwargs: keyword arguments for Optim.optimize

PortfolioOptimiser.GM — Type

struct GM <: MeanTarget end

PortfolioOptimiser.MuBOP — Type

@kwdef mutable struct MuBOP{T1} <: MeanEstimator
    target::MeanTarget = GM()
    w::Union{<:AbstractWeights, Nothing} = nothing
    sigma::T1 = Matrix{Float64}(undef, 0, 0)
end

PortfolioOptimiser.MuBS — Type

@kwdef mutable struct MuBS{T1} <: MeanEstimator
    target::MeanTarget = GM()
    w::Union{<:AbstractWeights, Nothing} = nothing
    sigma::T1 = Matrix{Float64}(undef, 0, 0)
end

PortfolioOptimiser.MuJS — Type

@kwdef mutable struct MuJS{T1} <: MeanEstimator
    target::MeanTarget = GM()
    w::Union{<:AbstractWeights, Nothing} = nothing
    sigma::T1 = Matrix{Float64}(undef, 0, 0)
end

PortfolioOptimiser.MuSimple — Type

@kwdef mutable struct MuSimple <: MeanEstimator
    w::Union{<:AbstractWeights, Nothing} = nothing
end

PortfolioOptimiser.NoDenoise — Type

struct NoDenoise <: Denoise end

No denoising is performed in denoise!.

PortfolioOptimiser.NoPosdef — Type

struct NoPosdef <: PosdefFix end

Non positive definite matrices will not be fixed.

PortfolioOptimiser.PortCovCor — Type

@kwdef mutable struct PortCovCor <: PortfolioOptimiserCovCor
    ce::CovarianceEstimator = CovFull(;)
    posdef::PosdefFix = PosdefNearest(;)
    denoise::Denoise = NoDenoise(;)
    jlogo::AbstractJLoGo = NoJLoGo(;)
end

PortfolioOptimiser.PosdefNearest — Type

@kwdef mutable struct PosdefNearest <: PosdefFix
    method::NearestCorrelationMatrix.NCMAlgorithm = NearestCorrelationMatrix.Newton(;
                                                                                    tau = 1e-12)
end

Defines which method from NearestCorrelationMatrix to use in posdef_fix!.

PortfolioOptimiser.SE — Type

struct SE <: MeanTarget end

PortfolioOptimiser.Shrink — Type

mutable struct Shrink{T1, T2, T3, T4, T5} <: Denoise
    detone::Bool = false
    mkt_comp::Integer = 1
    kernel = AverageShiftedHistograms.Kernels.gaussian
    m::Integer = 10
    n::Integer = 1000
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the parameters for using the shrink method in denoise! [2, Chapter 2].

Parameters

detone:

true: take only the largest mkt_comp eigenvalues from the correlation matrix.

Warning

Removing eigenvalues from the matrix may make it singular.

mkt_comp: the number of largest eigenvalues to keep from the correlation matrix.

kernel: kernel for fitting the average shifted histograms from AverageShiftedHistograms.jl Kernel Functions.

m: number of adjacent histograms to smooth over AverageShiftedHistograms.jl Usage.

n: number of points used when creating the range of values to which the average shifted histogram is to be fitted AverageShiftedHistograms.jl Usage.

args: arguments for Optim.optimize

kwargs: keyword arguments for Optim.optimize

PortfolioOptimiser.Spectral — Type

@kwdef mutable struct Spectral{T1, T2, T3, T4} <: Denoise
    detone::Bool = false
    mkt_comp::Integer = 1
    kernel = AverageShiftedHistograms.Kernels.gaussian
    m::Integer = 10
    n::Integer = 1000
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the parameters for using the spectral method in denoise! [2, Chapter 2].

Parameters

detone:

true: take only the largest mkt_comp eigenvalues from the correlation matrix.

Warning

Removing eigenvalues from the matrix may make it singular.

mkt_comp: the number of largest eigenvalues to keep from the correlation matrix.

kernel: kernel for fitting the average shifted histograms from AverageShiftedHistograms.jl Kernel Functions.

m: number of adjacent histograms to smooth over AverageShiftedHistograms.jl Usage.

n: number of points used when creating the range of values to which the average shifted histogram is to be fitted AverageShiftedHistograms.jl Usage.

args: arguments for Optim.optimize

kwargs: keyword arguments for Optim.optimize

PortfolioOptimiser.VW — Type

struct VW <: MeanTarget end

Private

PortfolioOptimiser.Denoise — Type

abstract type Denoise end

Abstract type for subtyping denoising methods.

PortfolioOptimiser.DistanceMethod — Type

abstract type DistanceMethod end

Abstract type for subtyping methods for computing distance matrices from correlation ones.

PortfolioOptimiser.MeanEstimator — Type

abstract type MeanEstimator end

PortfolioOptimiser.MeanTarget — Type

abstract type MeanTarget end

PortfolioOptimiser.OWAMethods — Type

OWAMethods = (:CRRA, :E, :SS, :SD)

Methods for computing the weights used to combine L-moments higher than 2 [3].

:CRRA: Normalised Constant Relative Risk Aversion Coefficients.

:E: Maximum Entropy. Solver must support MOI.RelativeEntropyCone and MOI.NormOneCone.

:SS: Minimum Sum of Squares. Solver must support MOI.SecondOrderCone.

:SD: Minimum Square Distance. Solver must support MOI.SecondOrderCone.

PortfolioOptimiser.PosdefFix — Type

abstract type PosdefFix end

Abstract type for subtyping methods for fixing non positive definite matrices.