Parameter Estimation Types

Public

PortfolioOptimiser.ClusterNode — Type

struct ClusterNode{tid, tl, tr, td, tcnt}
    id::tid
    left::tl
    right::tr
    height::td
    level::tcnt

    function ClusterNode(id, left::Union{ClusterNode, Nothing} = nothing,
                         right::Union{ClusterNode, Nothing} = nothing, height::Real = 0.0,
                         level::Int = 1)
        ilevel = isnothing(left) ? level : (left.level + right.level)

        return new{typeof(id), typeof(left), typeof(right), typeof(height), typeof(level)}(id,
                                                                                         left,
                                                                                         right,
                                                                                         height,
                                                                                         ilevel)
    end
end

Structure for definining a cluster node. This is used for turning a clustering result into a tree, is_leaf, pre_order, to_tree.

Parameters

id: node ID.

left: node to the left.

right: node to the right.

height: node height.

level: node level.

PortfolioOptimiser.CorDistance — Type

mutable struct CorDistance <: PortfolioOptimiserCovCor
    distance::Distances.UnionMetric
    dist_args::Tuple
    dist_kwargs::NamedTuple
    mean_w1::Union{<:AbstractWeights, Nothing}
    mean_w2::Union{<:AbstractWeights, Nothing}
    mean_w3::Union{<:AbstractWeights, Nothing}
end

PortfolioOptimiser.CorGerberSB1 — Type

mutable struct CorGerberSB1{T1, T2, T3, T4, T5} <: CorSB
    normalise::Bool
    threshold::T1
    c1::T2
    c2::T3
    c3::T4
    n::T5
    ve::StatsBase.CovarianceEstimator
    std_w::Union{<:AbstractWeights, Nothing}
    mean_w::Union{<:AbstractWeights, Nothing}
    posdef::PosdefFix
end

PortfolioOptimiser.CorKendall — Type

@kwdef mutable struct CorKendall <: CorRank
    absolute::Bool = false
end

Kendall type correlation estimator.

Parameters

absolute: whether or not to compute an absolute correlation, abs.(corkendall(X)).

PortfolioOptimiser.CorLTD — Type

mutable struct CorLTD <: PortfolioOptimiserCovCor
    alpha::Real
    ve::StatsBase.CovarianceEstimator
    w::Union{<:AbstractWeights, Nothing}
end

PortfolioOptimiser.CorMutualInfo — Type

@kwdef mutable struct CorMutualInfo <: PortfolioOptimiserCovCor
    bins::Union{<:Integer, <:AbstractBins} = HGR()
    normalise::Bool = true
    ve::StatsBase.CovarianceEstimator = SimpleVariance()
    w::Union{<:AbstractWeights, Nothing} = nothing
end

Define the mutual information correlation matrix.

Parameters

bins: defines the bin function, or bin width directly and if so bins > 0.

normalise: whether or not to normalise the mutual information.

ve: variance estimator.

w: variance weights, if nothing

PortfolioOptimiser.CorSpearman — Type

@kwdef mutable struct CorSpearman <: CorRank
    absolute::Bool = false
end

Spearman type correlation estimator.

Parameters

absolute: whether or not to compute an absolute correlation, abs.(corspearman(X)).

PortfolioOptimiser.CovFull — Type

@kwdef mutable struct CovFull <: CorPearson
    absolute::Bool = false
    ce::StatsBase.CovarianceEstimator = StatsBase.SimpleCovariance(; corrected = true)
    w::Union{<:AbstractWeights, Nothing} = nothing
end

Full Pearson-type covariance and correlation estimator.

Parameters

absolute: whether or not to compute an absolute correlation.

ce: covariance estimator.

w: weights for computing the covariance, if nothing apply no weights.

PortfolioOptimiser.CovSemi — Type

@kwdef mutable struct CovSemi <: CorPearson
    absolute::Bool = false
    ce::StatsBase.CovarianceEstimator = StatsBase.SimpleCovariance(; corrected = true)
    target::Union{<:Real, AbstractVector{<:Real}} = 0.0
    w::Union{<:AbstractWeights, Nothing} = nothing
end

Semi Pearson-type covariance and correlation estimator.

Parameters

absolute: whether or not to compute an absolute correlation, abs.(cor(X)).

ce: covariance estimator.

target: minimum acceptable return target.

if isa(target, Real): apply the same target to all assets.

if isa(target, AbstractVector): apply individual target to each asset.

w: weights for computing the covariance, if nothing apply no weights.

PortfolioOptimiser.DBHT — Type

mutable struct DBHT <: HClustAlg
    distance::DistanceMethod
    similarity::DBHTSimilarity
    root_method::DBHTRootMethod
end

Defines the parameters for computing DBHTs [2].

PortfolioOptimiser.DBHTExp — Type

struct DBHTExp <: DBHTSimilarity end

Defines the similarity matrix for use in PMFG_T2s as the element-wise exponential decay of the dissimilarity matrix in dbht_similarity.

PortfolioOptimiser.DBHTMaxDist — Type

struct DBHTMaxDist <: DBHTSimilarity end

Defines the similarity matrix for use in PMFG_T2s as the element-wise distance from the maximum value of the dissimilarity matrix dbht_similarity.

PortfolioOptimiser.DenoiseFixed — Type

@kwdef mutable struct DenoiseFixed{T1, T2, T3, T4} <: Denoise
    detone::Bool = false
    mkt_comp::Integer = 1
    kernel = AverageShiftedHistograms.Kernels.gaussian
    m::Integer = 10
    n::Integer = 1000
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the parameters for using the fixed method in denoise! [1, Chapter 2].

Parameters

detone:

true: remove the largest mkt_comp eigenvalues from the correlation matrix.

Warning

Removing eigenvalues from the matrix may make it singular.

mkt_comp: the number of largest eigenvalues to remove from the correlation matrix.

kernel: kernel for fitting the average shifted histograms from AverageShiftedHistograms.jl Kernel Functions.

m: number of adjacent histograms to smooth over AverageShiftedHistograms.ash.

n: number of points used when creating the range of values to which the average shifted histogram is to be fitted AverageShiftedHistograms.ash.

args: arguments for Optim.optimize

kwargs: keyword arguments for Optim.optimize

PortfolioOptimiser.DenoiseShrink — Type

@kwdef mutable struct DenoiseShrink{T1, T2, T3, T4, T5} <: Denoise
    alpha::Real = 0.0
    detone::Bool = false
    mkt_comp::Integer = 1
    kernel = AverageShiftedHistograms.Kernels.gaussian
    m::Integer = 10
    n::Integer = 1000
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the parameters for using the shrink method in denoise! [1, Chapter 2].

Parameters

alpha: tuning parameter for how much the matrix should be shrunk, alpha ∈ [0, 1].

detone:

true: take only the largest mkt_comp eigenvalues from the correlation matrix.

Warning

Removing eigenvalues from the matrix may make it singular.

mkt_comp: the number of largest eigenvalues to keep from the correlation matrix.

kernel: kernel for fitting the average shifted histograms from AverageShiftedHistograms.jl Kernel Functions.

m: number of adjacent histograms to smooth over AverageShiftedHistograms.ash.

n: number of points used when creating the range of values to which the average shifted histogram is to be fitted AverageShiftedHistograms.ash.

args: arguments for Optim.optimize

kwargs: keyword arguments for Optim.optimize

PortfolioOptimiser.DenoiseSpectral — Type

@kwdef mutable struct DenoiseSpectral{T1, T2, T3, T4} <: Denoise
    detone::Bool = false
    mkt_comp::Integer = 1
    kernel = AverageShiftedHistograms.Kernels.gaussian
    m::Integer = 10
    n::Integer = 1000
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the parameters for using the spectral method in denoise! [1, Chapter 2].

Parameters

detone:

true: take only the largest mkt_comp eigenvalues from the correlation matrix.

Warning

Removing eigenvalues from the matrix may make it singular.

mkt_comp: the number of largest eigenvalues to keep from the correlation matrix.

kernel: kernel for fitting the average shifted histograms from AverageShiftedHistograms.jl Kernel Functions.

m: number of adjacent histograms to smooth over AverageShiftedHistograms.ash.

n: number of points used when creating the range of values to which the average shifted histogram is to be fitted AverageShiftedHistograms.ash.

args: arguments for Optim.optimize

kwargs: keyword arguments for Optim.optimize

PortfolioOptimiser.DistanceCanonical — Type

struct DistanceCanonical <: DistanceMethod end

Struct for computing the canonical distance for a given correlation subtype of PortfolioOptimiserCovCor.

PortfolioOptimiser.DistanceLog — Type

struct DistanceLog <: DistanceMethod end

Defines the log-distance matrix from the correlation matrix.

\[D_{i,\,j} = -\log\left(C_{i,\,j}\right)\,.\]

Where:

$D_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N log-distance matrix.

$C_{i,\,j}$: is the $(i,\,j)$-th entry of an absolute correlation matrix.

PortfolioOptimiser.DistanceMLP — Type

@kwdef mutable struct DistanceMLP <: DistanceMethod
    absolute::Bool = false
end

Defines the distance matrix from a correlation matrix [6] in dist.

\[D_{i,\,j} = \begin{cases} \sqrt{\dfrac{1}{2} \left(\mathbf{1} - C_{i,\,j}\right)} &\quad \mathrm{if~ absolute = false}\\ \sqrt{1 - \lvert C_{i,\,j} \rvert} &\quad \mathrm{if~ absolute = true}\nonumber\,. \end{cases}\]

Where:

$D_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N distance matrix $\mathbf{C}$.

$C_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N correlation matrix $\mathbf{D}$.

absolute: flag for stating whether or not an absolute correlation is being used.

Parameters

absolute: flag for stating whether or not an absolute correlation is being used.

PortfolioOptimiser.DistanceSqMLP — Type

@kwdef mutable struct DistanceSqMLP <: DistanceMethod
    absolute::Bool = false
    distance::Distances.UnionMetric = Distances.Euclidean()
    args::Tuple = ()
    kwargs::NamedTuple = (;)
end

Defines the distance of distances matrix from a correlation matrix [6] in dist.

\[\tilde{D}_{i,\,j} = f_{m}\left(\bm{D}_{i},\, \bm{D}_j\right)\,.\]

Where:

$\bm{D}_{i}$: is the $i$-th column/row of the N×N distance matrix defined in DistanceMLP.

$f_{m}$: is the pairwise distance function for metric $m$. We use the Distances.pairwise function which computes the entire matrix at once, the output is a vector so it gets reshaped into an N×N matrix.

$\tilde{D}_{i,\,j}$: is the $(i,\,j)$-th entry of the N×N distances of distances matrix.

absolute: is a flag whether the correlation is absolute.

Parameters

absolute: flag for stating whether or not an absolute correlation is being used.

distance: distance metric from Distances.jl.

args: args for the Distances.pairwise function.

kwargs: key word args for the Distances.pairwise function.

PortfolioOptimiser.DistanceVarInfo — Type

@kwdef mutable struct DistanceVarInfo <: DistanceMethod
    bins::Union{<:Integer, <:AbstractBins} = HGR()
    normalise::Bool = true
end

Defines the variation of information distance matrix.

Parameters

bins: defines the bin function, or bin width directly and if so bins > 0.

normalise: whether or not to normalise the variation of information.

PortfolioOptimiser.EqualDBHT — Type

struct EqualDBHT <: DBHTRootMethod end

Create a clique's root from its adjacency tree in CliqueRoot.

PortfolioOptimiser.Freedman — Type

struct Freedman <: AstroBins end

Freedman's bin width algorithm from astropy.

PortfolioOptimiser.GM — Type

struct GM <: MeanTarget end

PortfolioOptimiser.HAC — Type

@kwdef mutable struct HAC <: HClustAlg
    linkage::Symbol = :ward
end

Use a hierarchical clustering algorithm from Clustering.jl.

Parameters

linkage: linkage type supported by hclust.

PortfolioOptimiser.HCOpt — Type

@kwdef mutable struct HCOpt{T1 <: Integer, T2 <: Integer}
    branchorder::Symbol = :optimal
    k_method::NumClusterMethod = TwoDiff()
    k::T1 = 0
    max_k::T2 = 0
end

Defines the options for processing clustering results in an instance of Clustering.Hclust.

Parameters

branchorder: parameter for ordering a dendrogram's branches accepted by Clustering.jl.

k_method: method subtyping NumClusterMethod for computing the number of clusters.

k: directly provide the number of clusters, if 0 use k_method for computing the number of clusters.

max_k: maximum number of clusters, if 0 defaults to ⌈sqrt(N)⌉, where N is the number of assets.

PortfolioOptimiser.HGR — Type

struct HGR <: AbstractBins end

Hacine-Gharbi and Ravier's bin width algorithm [7].

PortfolioOptimiser.Knuth — Type

struct Knuth <: AstroBins end

Knuth's bin width algorithm from astropy.

PortfolioOptimiser.LoGo — Type

@kwdef mutable struct LoGo <: AbstractLoGo
    distance::DistanceMethod = DistanceMLP()
    similarity::DBHTSimilarity = DBHTMaxDist()
end

PortfolioOptimiser.MuBOP — Type

@kwdef mutable struct MuBOP{T1} <: MeanEstimator
    target::MeanTarget = GM()
    w::Union{<:AbstractWeights, Nothing} = nothing
    sigma::T1 = Matrix{Float64}(undef, 0, 0)
end

PortfolioOptimiser.MuBS — Type

@kwdef mutable struct MuBS{T1} <: MeanEstimator
    target::MeanTarget = GM()
    w::Union{<:AbstractWeights, Nothing} = nothing
    sigma::T1 = Matrix{Float64}(undef, 0, 0)
end

PortfolioOptimiser.MuJS — Type

@kwdef mutable struct MuJS{T1} <: MeanEstimator
    target::MeanTarget = GM()
    w::Union{<:AbstractWeights, Nothing} = nothing
    sigma::T1 = Matrix{Float64}(undef, 0, 0)
end

PortfolioOptimiser.MuSimple — Type

@kwdef mutable struct MuSimple <: MeanEstimator
    w::Union{<:AbstractWeights, Nothing} = nothing
end

PortfolioOptimiser.NoDenoise — Type

struct NoDenoise <: Denoise end

No denoising is performed in denoise!.

PortfolioOptimiser.NoPosdef — Type

struct NoPosdef <: PosdefFix end

Non positive definite matrices will not be fixed in in posdef_fix!.

PortfolioOptimiser.PortCovCor — Type

@kwdef mutable struct PortCovCor <: PortfolioOptimiserCovCor
    ce::CovarianceEstimator = CovFull(;)
    posdef::PosdefFix = PosdefNearest(;)
    denoise::Denoise = NoDenoise(;)
    logo::AbstractLoGo = NoLoGo(;)
end

PortfolioOptimiser.PosdefNearest — Type

@kwdef mutable struct PosdefNearest <: PosdefFix
    method::NearestCorrelationMatrix.NCMAlgorithm = NearestCorrelationMatrix.Newton(;
                                                                                    tau = 1e-12)
end

Defines which method from NearestCorrelationMatrix to use in posdef_fix!.

PortfolioOptimiser.SE — Type

struct SE <: MeanTarget end

PortfolioOptimiser.Scott — Type

struct Scott <: AstroBins end

Scott's bin width algorithm from astropy.

PortfolioOptimiser.SimpleVariance — Type

@kwdef mutable struct SimpleVariance <: StatsBase.CovarianceEstimator
    corrected::Bool = true
end

Simple variance estimator.

Parameters

corrected: if true correct the bias by dividing by N-1, if false the bias is not corrected and the division is by N.

PortfolioOptimiser.StdSilhouette — Type

@kwdef mutable struct StdSilhouette <: NumClusterMethod
    metric::Union{Distances.SemiMetric, Nothing} = nothing
end

Use the standardised silhouette score for computing the number of clusters in calc_k_clusters.

PortfolioOptimiser.TwoDiff — Type

struct TwoDiff <: NumClusterMethod end

Use the two difference gap statistic for computing the number of clusters in calc_k_clusters.

PortfolioOptimiser.UniqueDBHT — Type

struct UniqueDBHT <: DBHTRootMethod end

Create a unique root for a clique in CliqueRoot [5].

PortfolioOptimiser.VW — Type

struct VW <: MeanTarget end

Private

PortfolioOptimiser.AbstractBins — Type

abstract type AbstractBins end

Abstract type for defining bin width estimation functions when computing DistanceVarInfo and CorMutualInfo distance and correlation matrices respectively.

PortfolioOptimiser.AstroBins — Type

abstract type AstroBins <: AbstractBins end

Abstract type for defining which bin width function to use from astropy.

PortfolioOptimiser.CorPearson — Type

abstract type CorPearson <: PortfolioOptimiserCovCor end

Abstract type for subtyping Pearson type covariance estimators.

PortfolioOptimiser.CorRank — Type

abstract type CorRank <: PortfolioOptimiserCovCor end

Abstract type for subtyping rank based covariance estimators.

PortfolioOptimiser.DBHTRootMethod — Type

abstract type DBHTRootMethod end

Abstract type for subtyping methods creating roots of cliques in CliqueRoot [5].

PortfolioOptimiser.DBHTSimilarity — Type

abstract type DBHTSimilarity end

Abstract type for subtyping methods for defining functions for computing similarity matrices used in DBHT clustering PMFG_T2s [2, 4].

PortfolioOptimiser.Denoise — Type

abstract type Denoise end

Abstract type for subtyping denoising methods.

PortfolioOptimiser.DistanceMethod — Type

abstract type DistanceMethod end

Abstract type for subtyping methods for computing distance matrices from correlation ones.

PortfolioOptimiser.HClustAlg — Type

abstract type HClustAlg end

Abstract type for subtyping hierarchical clustering methods.

PortfolioOptimiser.MeanEstimator — Type

abstract type MeanEstimator end

PortfolioOptimiser.MeanTarget — Type

abstract type MeanTarget end

PortfolioOptimiser.NumClusterMethod — Type

abstract type NumClusterMethod end

Abstract type for subtyping methods for determining the number of clusters in a Clustering.Hclust when calling calc_k_clusters.

PortfolioOptimiser.OWAMethods — Type

OWAMethods = (:CRRA, :E, :SS, :SD)

Methods for computing the weights used to combine L-moments higher than 2 [8].

:CRRA: Normalised Constant Relative Risk Aversion Coefficients.

:E: Maximum Entropy. Solver must support MOI.RelativeEntropyCone and MOI.NormOneCone.

:SS: Minimum Sum of Squares. Solver must support MOI.SecondOrderCone.

:SD: Minimum Square Distance. Solver must support MOI.SecondOrderCone.

PortfolioOptimiser.PortfolioOptimiserCovCor — Type

abstract type PortfolioOptimiserCovCor <: StatsBase.CovarianceEstimator end

Abstract type for subtyping portfolio optimiser covaraince and correlation estimators.

PortfolioOptimiser.PosdefFix — Type

abstract type PosdefFix end

Abstract type for subtyping methods for fixing non positive definite matrices.