Portfolio and HCPortfolio

Public

PortfolioOptimiser.HCPortfolio — Type

mutable struct HCPortfolio{ast, dat, r, tmu, tcov, tkurt, tskurt, tl2, ts2, tskew, tv,
                           tsskew, tsv, wmi, wma, tco, tdist, tcl, tk, topt, tsolv, tf, tlp,
                           taopt, talo, tasolv, taf, tamod} <: AbstractPortfolio
    assets::ast
    timestamps::dat
    returns::r
    mu::tmu
    cov::tcov
    kurt::tkurt
    skurt::tskurt
    L_2::tl2
    S_2::ts2
    skew::tskew
    V::tv
    sskew::tsskew
    SV::tsv
    w_min::wmi
    w_max::wma
    cor::tco
    dist::tdist
    clusters::tcl
    k::tk
    optimal::topt
    solvers::tsolv
    fail::tf
    latest_prices::tlp
    alloc_optimal::taopt
    alloc_leftover::talo
    alloc_solvers::tasolv
    alloc_fail::taf
    alloc_model::tamod
end

Structure for defining a hierarchical clustering portfolio.

Parameters

assets: N×1 vector of asset names.

timestamps: T×1 vector of asset returns timestamps.

returns: T×N matrix of asset returns.

mu: N×1 vector of asset expected returns.

cov: N×N asset covariance matrix.

kurt: N^2×N^2 cokurtosis matrix.

skurt: N^2×N^2 semi cokurtosis matrix.

L_2: (N^2)×((N^2 + N)/2) elimination matrix.

S_2: ((N^2 + N)/2)×(N^2) summation matrix.

skew: N×N^2 coskew matrix.

V: N×N sum of the symmetric negative spectral slices of coskewness.

sskew: N×N^2 semi coskew matrix.

SV: N×N sum of the symmetric negative spectral slices of semi coskewness.

w_min: minimum allowable asset weights.

if isa vector: N×1 vector of minimum allowable weight per asset.

if isa scalar: minimum asset weight for all assets.

w_max: maximum allowable asset weights.

if isa vector: N×1 vector of maximum allowable weight per asset.

if isa scalar: maximum asset weight for all assets.

cor: N×N asset correlation matrix.

dist: N×N asset distance matrix.

clusters: Clustering.Hclust of asset clusters.

k: number of asset clusters.

if is zero: compute the number of clusters via one of the cluster number methods NumClusterMethod.

if is not zero: use this value directly.

optimal: collection capable of storing key value pairs for storing optimal portfolios.

solvers: collection capable of storing key value pairs for storing JuMP-supported solvers. They must have the following structure.

solvers = Dict(
               # Key-value pair for the solver, solution acceptance 
               # criteria, and solver attributes.
               :Clarabel => Dict(
                                 # Solver we wish to use.
                                 :solver => Clarabel.Optimizer,
                                 # (Optional) Solution acceptance criteria.
                                 :check_sol => (allow_local = true, allow_almost = true),
                                 # (Optional) Solver-specific attributes.
                                 :params => Dict("verbose" => false)))

The dictionary contains a key value pair for each solver (plus optional solution acceptance criteria and optional attributes) we want to use.

:solver: defines the solver to use. One can also use JuMP.optimizer_with_attributes to direcly provide a solver with attributes already attached.

:check_sol: (optional) defines the keyword arguments passed on to JuMP.is_solved_and_feasible for accepting/rejecting solutions.

:params: (optional) defines solver-specific parameters.

Users are also able to provide multiple solvers by adding additional key-value pairs to the top-level dictionary as in the following snippet.

using JuMP
solvers = Dict(:Clarabel => Dict(:solver => Clarabel.Optimizer,
                                 :check_sol => (allow_local = true, allow_almost = true),
                                 :params => Dict("verbose" => false)),
               # Provide solver with pre-attached attributes and no arguments 
               # for the `JuMP.is_solved_and_feasible` function.
               :COSMO => Dict(:solver => JuMP.optimizer_with_attributes(COSMO.Optimizer,
                                                                        "maxiter" => 5000)))

optimise! will iterate over the solvers until it finds the first one to successfully solve the problem.

fail: collection capable of storing key value pairs for storing failed optimisation attempts.

latest_prices: Na×1 vector of latest asset prices.

alloc_optimal: collection capable of storing key value pairs for storing optimal discretely allocated portfolios.

alloc_leftover: collection capable of storing key value pairs for containing points in the leftover investment after allocating.

alloc_solvers: collection capable of storing key value pairs for storing JuMP-supported solvers that support Mixed-Integer Programming, only used in the LP allocation.

alloc_fail: collection capable of storing key value pairs for storing failed discrete asset allocation attempts.

alloc_model: JuMP.Model which defines the discrete asset allocation model.

PortfolioOptimiser.HCPortfolio — Method

HCPortfolio(; prices::TimeArray = TimeArray(TimeType[], []),
              returns::DataFrame = DataFrame(),
              ret::Matrix{<:Real} = Matrix{Float64}(undef, 0, 0),
              timestamps::Vector{<:Dates.AbstractTime} = Vector{Date}(undef, 0),
              assets::AbstractVector = Vector{String}(undef, 0),
              mu::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
              cov::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
              kurt::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
              skurt::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
              skew::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
              V = Matrix{eltype(returns)}(undef, 0, 0),
              sskew::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
              SV = Matrix{eltype(returns)}(undef, 0, 0),
              w_min::Union{<:Real, <:AbstractVector{<:Real}} = 0.0,
              w_max::Union{<:Real, <:AbstractVector{<:Real}} = 1.0,
              cor::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
              dist::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
              clusters::Clustering.Hclust = Hclust{Float64}(Matrix{Int64}(undef, 0, 2),
                                                          Float64[], Int64[],
                                                          :nothing),
              k::Integer = 0, optimal::AbstractDict = Dict(),
              solvers::AbstractDict = Dict(), fail::AbstractDict = Dict(),
              latest_prices::AbstractVector = Vector{Float64}(undef, 0),
              alloc_optimal::AbstractDict = Dict(),
              alloc_leftover::AbstractDict = Dict(),
              alloc_solvers::AbstractDict = Dict(),
              alloc_fail::AbstractDict = Dict(),
              alloc_model::JuMP.Model = JuMP.Model())

Constructor for HCPortfolio. Performs data validation checks and automatically extracts the data from prices, returns, f_prices, and f_returns if they are provided.

Inputs

prices: (T+1)×Na TimeArray of asset prices.

If provided: will take precedence over returns, ret, timestamps, assets, and latest_prices because they will be automatically computed from prices.

returns: T×Na DataFrame of asset returns.

If provided: will take precedence over ret, timestamps, and assets because they will be automatically computed from returns.

ret: set returns matrix directly.

assets: sets assets.

timestamps: sets timestamps.

returns: sets returns.

mu: sets mu.

cov: sets cov.

kurt: sets kurt.

skurt: sets skurt.

L_2: sets L_2.

S_2: sets S_2.

skew: sets skew.

V: sets V.

sskew: sets sskew.

SV: sets SV.

w_min: sets w_min.

w_max: sets w_max.

cor: sets cor.

dist: sets dist.

clusters: sets clusters.

k: sets k.

optimal: sets optimal.

solvers: sets solvers.

fail: sets fail.

latest_prices: sets latest_prices.

alloc_optimal: sets alloc_optimal.

alloc_leftover: sets alloc_leftover.

alloc_solvers: sets alloc_solvers.

alloc_fail: sets alloc_fail.

alloc_model: sets alloc_model.

PortfolioOptimiser.Portfolio — Type

mutable struct Portfolio{ast, dat, r, tfa, tfdat, tretf, l, lo, s, us, ul, nal, nau, naus,
                         mnak, mnaks, rb, to, kte, blbw, ami, bvi, rbv, frbv, nm, amc, bvc,
                         ler, tmu, tcov, tkurt, tskurt, tl2, ts2, tskew, tv, tsskew, tsv,
                         tmuf, tcovf, trfm, tmufm, tcovfm, tmubl, tcovbl, tmublf, tcovblf,
                         tcovl, tcovu, tcovmu, tcovs, tdmu, tkmu, tks, topt, tlim, tfront,
                         tsolv, tf, tmod, tlp, taopt, talo, tasolv, taf, tamod} <:
               AbstractPortfolio
    assets::ast
    timestamps::dat
    returns::r
    f_assets::tfa
    f_timestamps::tfdat
    f_returns::tretf
    loadings::l
    regression_type::lo
    short::s
    short_u::us
    long_u::ul
    num_assets_l::nal
    num_assets_u::nau
    num_assets_u_scale::naus
    max_num_assets_kurt::mnak
    max_num_assets_kurt_scale::mnaks
    rebalance::rb
    turnover::to
    tracking_err::kte
    bl_bench_weights::blbw
    a_mtx_ineq::ami
    b_vec_ineq::bvi
    risk_budget::rbv
    f_risk_budget::frbv
    network_method::nm
    a_vec_cent::amc
    b_cent::bvc
    mu_l::ler
    mu::tmu
    cov::tcov
    kurt::tkurt
    skurt::tskurt
    L_2::tl2
    S_2::ts2
    skew::tskew
    V::tv
    sskew::tsskew
    SV::tsv
    f_mu::tmuf
    f_cov::tcovf
    fm_returns::trfm
    fm_mu::tmufm
    fm_cov::tcovfm
    bl_mu::tmubl
    bl_cov::tcovbl
    blfm_mu::tmublf
    blfm_cov::tcovblf
    cov_l::tcovl
    cov_u::tcovu
    cov_mu::tcovmu
    cov_sigma::tcovs
    d_mu::tdmu
    k_mu::tkmu
    k_sigma::tks
    optimal::topt
    limits::tlim
    frontier::tfront
    solvers::tsolv
    fail::tf
    model::tmod
    latest_prices::tlp
    alloc_optimal::taopt
    alloc_leftover::talo
    alloc_solvers::tasolv
    alloc_fail::taf
    alloc_model::tamod
end

Structure for defining a traditional portfolio. Na is the number of assets, and Nf is the number of factors. For details on how some of these parameters are computed see asset_statistics!, wc_statistics!, factor_statistics!, black_litterman_statistics!, and black_litterman_factor_statistics!.

Parameters

assets: Na×1 vector of asset names.

timestamps: T×1 vector of asset returns timestamps.

returns: T×Na matrix of asset returns.

f_assets: Nf×1 vector of factor names.

f_timestamps: T×1 vector of factor returns timestamps.

f_returns: T×Nf matrix of asset returns.

loadings: loadings matrix for working with factor models.

regression_type: RegressionType used for computing the loadings matrix.

short:

if true: shorting is enabled.

else: long-only portfolio.

short_u: upper bound for the absolute value of the sum of the negative weights.

long_u: upper bound for the sum of the positive weights.

num_assets_l: lower bound for the minimum number of significant assets.

if > 0: applies the constraint.

num_assets_u: upper bound for the maximum number of significant assets.

if > 0: applies the constraint.

num_assets_u_scale: scaling factor for the decision variable used for applying the num_assets_u constraint when optimising the Sharpe objective function.

max_num_assets_kurt: maximum number of assets to use the complete kurtosis model.

if > 0: the approximate model will be used if the number of assets in the portfolio exceeds max_number_assets_kurt.

max_num_assets_kurt_scale: multipies Na to find the number of eigenvalues when computing the approximate kurtosis model, must be ∈ [1, Na].

rebalance: AbstractTR for defining the portfolio rebalancing penalty.

\[\begin{align} p_{r} &= \sum\limits_{i=1}^{N} r_{i} \lvert w_{i} - b_{i} \rvert\,. \end{align}\]

Where:

$p_{r}$ is the portfolio rebalancing penalty.

$N$ is the number of assets.

$r_{i}$ is the rebalancing penalty for the $i$-th asset.

$w_{i}$ is the weight of the $i$-th asset.

$b_{i}$ is the benchmark weight of the $i$-th asset.

turnover: AbstractTR for defining the asset turnover constraint.

\[\begin{align} \lvert w_{i} - b_{i} \rvert \leq t_{i}\quad \forall i = 1,\,\ldots,\,N\,. \end{align}\]

Where:

$t_{i}$ is the turnover constraint for the $i$-th asset.

$w_{i}$ is the weight of the $i$-th asset.

$b_{i}$ is the benchmark weight of the $i$-th asset.

$N$ is the number of assets.

tracking_err: TrackingErr for defining the tracking error constraint.

\[\begin{align} \left\lVert \dfrac{\mathbf{X} \bm{w} - \bm{b}}{T - 1} \right\rVert_{2} \leq \epsilon \end{align}\]

Where:

$\lVert \cdot \rVert_{2}$ is the L2 norm.

$\mathbf{X}$ is the $T \times N$ matrix of asset returns.

$T$ is the number of returns observations.

$N$ is the number of assets.

$\bm{w}$ is the $N \times 1$ vector of asset weights.

$\bm{b}$ is the $T \times 1$ vector of benchmark returns.

$\epsilon$ is the tracking error.

bl_bench_weights: benchmark weights for Black-Litterman models BlackLittermanClass.

a_mtx_ineq: C×N matrix of asset weight linear constraints.

if isempty: the constraint is not set.

b_vec_ineq: C×1 vector of asset weight linear constraints.

if isempty: the constraint is not set.

The linear weight constraint is defined as.

\[\begin{align} \mathbf{A} \bm{w} \geq \bm{b}\,. \end{align}\]

Where:

$\mathbf{A}$ is the $C×N$ matrix of asset weight linear constraints.

$\bm{b}$ is the $C×1$ vector of asset weight linear constraints.

$C$ is the number of constraints.

$N$ is the number of assets.

risk_budget: Na×1 vector of asset risk budgets.

f_risk_budget: Nf×1 vector of factor risk budgets.

network_method: NetworkMethods for defining the asset network constraint. This can be defined in two ways, using an exact mixed-integer approach IP or an approximate semi-definite one SDP. See their docs for the constraint definition for each case.

if NoNtwk: the constraint is not set.

a_vec_cent: centrality vector for defining the centrality constraint.

if isempty: the constraint is not set.

b_cent: average centrality of the assets the portfolio.

if isinf: the constraint is not set.

The centrality constraint is defined as.

\[\begin{align} \bm{C} \cdot \bm{w} = \bar{c} \end{align}\]

Where:

$\bm{w}$ is the $N\times 1$ vector of asset weights.

$\bm{C}$ is the $N \times 1$ centrality vector of the asset adjacency matrix.

$\cdot$ is the dot product.

$\bar{c}$ is the desired average centrality measure of the portfolio.

mu_l: lower bound for the expected return of the portfolio.

if is Inf: the constraint is not applied.

mu: Na×1 vector of asset expected returns.

cov: Na×Na asset covariance matrix.

kurt: Na^2×Na^2 cokurtosis matrix.

skurt: Na^2×Na^2 semi cokurtosis matrix.

L_2: (Na^2)×((Na^2 + Na)/2) elimination matrix.

S_2: ((Na^2 + Na)/2)×(Na^2) summation matrix.

skew: Na×Na^2 coskew matrix.

V: Na×Na sum of the symmetric negative spectral slices of coskewness.

sskew: Na×Na^2 semi coskew matrix.

SV: Na×Na sum of the symmetric negative spectral slices of semi coskewness.

f_mu: Nf×1 vector of factor expected returns.

f_cov: Nf×Nf factor covariance matrix.

fm_returns: T×Na factor model adjusted returns matrix.

fm_mu: Na×1 factor model adjusted asset expected returns.

fm_cov: Na×Na factor model adjusted asset covariance matrix.

bl_mu: Na×1 Black Litterman model adjusted asset expected returns.

bl_cov: Na×Na Black Litterman model adjusted asset covariance matrix.

blfm_mu: Na×1 Black Litterman factor model adjusted asset expected returns.

blfm_cov: Na×Na Black Litterman factor model adjusted asset covariance matrix.

cov_l: Na×Na lower bound for the worst case covariance matrix.

cov_u: Na×Na upper bound for the worst case covariance matrix.

cov_mu: Na×Na matrix of the estimation errors of the asset expected returns vector set.

cov_sigma: Na×Na matrix of the estimation errors of the asset covariance matrix set.

d_mu: absolute deviation of the worst case upper and lower asset expected returns vectors.

k_mu: distance parameter of the uncertainty in the asset expected returns vector for the worst case optimisation.

k_sigma: distance parameter of the uncertainty in the asset covariance matrix for the worst case optimisation.

optimal: collection capable of storing key value pairs for storing optimal portfolios.

limits: collection capable of storing key value pairs for storing the minimal and maximal risk portfolios.

frontier: collection capable of storing key value pairs for containing points in the efficient frontier.

solvers: collection capable of storing key value pairs for storing JuMP-supported solvers. They must have the following structure.

solvers = Dict(
               # Key-value pair for the solver, solution acceptance 
               # criteria, and solver attributes.
               :Clarabel => Dict(
                                 # Solver we wish to use.
                                 :solver => Clarabel.Optimizer,
                                 # (Optional) Solution acceptance criteria.
                                 :check_sol => (allow_local = true, allow_almost = true),
                                 # (Optional) Solver-specific attributes.
                                 :params => Dict("verbose" => false)))

The dictionary contains a key value pair for each solver (plus optional solution acceptance criteria and optional attributes) we want to use.

:solver: defines the solver to use. One can also use JuMP.optimizer_with_attributes to direcly provide a solver with attributes already attached.

:check_sol: (optional) defines the keyword arguments passed on to JuMP.is_solved_and_feasible for accepting/rejecting solutions.

:params: (optional) defines solver-specific parameters.

Users are also able to provide multiple solvers by adding additional key-value pairs to the top-level dictionary as in the following snippet.

using JuMP
solvers = Dict(:Clarabel => Dict(:solver => Clarabel.Optimizer,
                                 :check_sol => (allow_local = true, allow_almost = true),
                                 :params => Dict("verbose" => false)),
               # Provide solver with pre-attached attributes and no arguments 
               # for the `JuMP.is_solved_and_feasible` function.
               :COSMO => Dict(:solver => JuMP.optimizer_with_attributes(COSMO.Optimizer,
                                                                        "maxiter" => 5000)))

optimise! will iterate over the solvers until it finds the first one to successfully solve the problem.

fail: collection capable of storing key value pairs for storing failed optimisation attempts.

model: JuMP.Model which defines the optimisation model.

latest_prices: Na×1 vector of latest asset prices.

alloc_optimal: collection capable of storing key value pairs for storing optimal discretely allocated portfolios.

alloc_leftover: collection capable of storing key value pairs for containing points in the leftover investment after allocating.

alloc_solvers: collection capable of storing key value pairs for storing JuMP-supported solvers that support Mixed-Integer Programming, only used in the LP allocation.

alloc_fail: collection capable of storing key value pairs for storing failed discrete asset allocation attempts.

alloc_model: JuMP.Model which defines the discrete asset allocation model.

PortfolioOptimiser.Portfolio — Method

Portfolio(; prices::TimeArray = TimeArray(TimeType[], []),
            returns::DataFrame = DataFrame(),
            ret::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            timestamps::AbstractVector = Vector{Date}(undef, 0),
            assets::AbstractVector = Vector{String}(undef, 0),
            f_prices::TimeArray = TimeArray(TimeType[], []),
            f_returns::DataFrame = DataFrame(),
            f_ret::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            f_timestamps::AbstractVector = Vector{Date}(undef, 0),
            f_assets::AbstractVector = Vector{String}(undef, 0),
            loadings::DataFrame = DataFrame(),
            regression_type::Union{<:RegressionType, Nothing} = nothing,
            short::Bool = false, short_u::Real = 0.2, long_u::Real = 1.0,
            num_assets_l::Integer = 0, num_assets_u::Integer = 0,
            num_assets_u_scale::Real = 100_000.0, max_num_assets_kurt::Integer = 0,
            max_num_assets_kurt_scale::Integer = 2, rebalance::AbstractTR = NoTR(),
            turnover::AbstractTR = NoTR(), tracking_err::TrackingErr = NoTracking(),
            bl_bench_weights::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
            a_mtx_ineq::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            b_vec_ineq::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
            risk_budget::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
            f_risk_budget::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
            network_method::NetworkMethods = NoNtwk(),
            a_vec_cent::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
            b_cent::Real = Inf, mu_l::Real = Inf,
            mu::AbstractVector = Vector{Float64}(undef, 0),
            cov::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            kurt::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            skurt::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            skew::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            V::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            sskew::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            SV::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            f_mu::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
            f_cov::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            fm_returns::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            fm_mu::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
            fm_cov::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            bl_mu::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
            bl_cov::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            blfm_mu::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
            blfm_cov::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            cov_l::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            cov_u::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            cov_mu::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            cov_sigma::AbstractMatrix{<:Real} = Matrix{Float64}(undef, 0, 0),
            d_mu::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
            k_mu::Real = Inf, k_sigma::Real = Inf, optimal::AbstractDict = Dict(),
            limits::AbstractDict = Dict(), frontier::AbstractDict = Dict(),
            solvers::AbstractDict = Dict(), fail::AbstractDict = Dict(),
            model::JuMP.Model = JuMP.Model(),
            latest_prices::AbstractVector{<:Real} = Vector{Float64}(undef, 0),
            alloc_optimal::AbstractDict = Dict(),
            alloc_leftover::AbstractDict = Dict(),
            alloc_solvers::AbstractDict = Dict(), alloc_fail::AbstractDict = Dict(),
            alloc_model::JuMP.Model = JuMP.Model())

Constructor for Portfolio. Performs data validation checks and automatically extracts the data from prices, returns, f_prices, and f_returns if they are provided.

Inputs

prices: (T+1)×Na TimeArray of asset prices.

If provided: will take precedence over returns, ret, timestamps, assets, and latest_prices because they will be automatically computed from prices.

returns: T×Na DataFrame of asset returns.

If provided: will take precedence over ret, timestamps, and assets because they will be automatically computed from returns.

ret: set returns matrix directly.

timestamps: set timestamps.

assets: set assets.

f_prices: (T+1)×Nf TimeArray of factor prices.

If provided: will take precedence over f_returns, f_ret, f_timestamps, and f_assets because they will be automatically computed from f_prices.

f_returns: T×Nf DataFrame of factor returns.

If provided: will take precedence over f_ret, f_timestamps, and f_assets because they will be automatically computed from returns.

f_ret: set f_ret.

f_timestamps: set f_timestamps.

f_assets: set f_assets.

loadings: set loadings.

regression_type: set regression_type.

short: set short.

short_u: set short_u.

long_u: set long_u.

num_assets_l: set numassetsl.

num_assets_u: set numassetsu.

num_assets_u_scale: set numassetsu_scale.

max_num_assets_kurt: set maxnumassets_kurt.

max_num_assets_kurt_scale: set maxnumassetskurtscale.

rebalance: set rebalance.

turnover: set turnover.

tracking_err: set tracking_err.

bl_bench_weights: set blbenchweights.

a_mtx_ineq: set amtxineq.

b_vec_ineq: set bvecineq.

risk_budget: set risk_budget.

f_risk_budget: set friskbudget.

network_method: set network_method.

a_vec_cent: set aveccent.

b_cent: set b_cent.

mu_l: set mu_l.

mu: set mu.

cov: set cov.

kurt: set kurt.

skurt: set skurt.

L_2: set L_2.

S_2: set S_2.

skew: set skew.

V: set V.

sskew: set sskew.

SV: set SV.

f_mu: set f_mu.

f_cov: set f_cov.

fm_returns: set fm_returns.

fm_mu: set fm_mu.

fm_cov: set fm_cov.

bl_mu: set bl_mu.

bl_cov: set bl_cov.

blfm_mu: set blfm_mu.

blfm_cov: set blfm_cov.

cov_l: set cov_l.

cov_u: set cov_u.

cov_mu: set cov_mu.

cov_sigma: set cov_sigma.

d_mu: set d_mu.

k_mu: set k_mu.

k_sigma: set k_sigma.

optimal: set optimal.

limits: set limits.

frontier: set frontier.

solvers: set solvers.

fail: set fail.

model: set model.

latest_prices: set latest_prices.

alloc_optimal: set alloc_optimal.

alloc_leftover: set alloc_leftover.

alloc_solvers: set alloc_solvers.

alloc_fail: set alloc_fail.

alloc_model: set alloc_model.

PortfolioOptimiser.calc_risk — Method

calc_risk(port::AbstractPortfolio; X::AbstractMatrix = port.returns,
          type::Symbol = isa(port, Portfolio) ? :Trad : :HRP,
          rm::AbstractRiskMeasure = SD())

Compute the risk for an AbstractRiskMeasure for a portfolio.

Inputs

port: portfolio.

X: T×N returns matrix.

type: optimisation type used to retrieve the weights vector from port.optimal[type].

rm: rism measure.

PortfolioOptimiser.factor_risk_contribution — Method

factor_risk_contribution(port::AbstractPortfolio; X::AbstractMatrix = port.returns,
                         F::AbstractMatrix = port.f_returns,
                         type::Symbol = isa(port, Portfolio) ? :Trad : :HRP,
                         rm::AbstractRiskMeasure = SD(), delta::Real = 1e-6)

Compute the factor risk contribution for an AbstractRiskMeasure for a portfolio.

Inputs

port: portfolio.

w: Na×1 vector of asset weights.

X: T×Na matrix of asset returns.

F: T×Nf matrix of factor returns.

type: optimisation type used to retrieve the weights vector from port.optimal[type].

rm: risk measure.

delta: small displacement used for computing the marginal risk and equal risk measure Equal.

Outputs

rc_f: Nf×1 vector of risk contribution per factor.

PortfolioOptimiser.risk_contribution — Method

risk_contribution(port::AbstractPortfolio; X::AbstractMatrix = port.returns,
                  type::Symbol = isa(port, Portfolio) ? :Trad : :HRP,
                  rm::AbstractRiskMeasure = SD(), delta::Real = 1e-6,
                  marginal::Bool = false)

Compute the asset risk contribution for an AbstractRiskMeasure for a portfolio.

Inputs

port: portfolio.

X: T×N returns matrix.

type: optimisation type used to retrieve the weights vector from port.optimal[type].

rm: risk measure.

delta: small displacement used for computing the marginal risk and equal risk measure Equal.

marginal:

if true: compute the marginal risk contribution.

else: compute the risk by contribution by multiplying the marginal risk by the asset weight.

Outputs

rc: Na×1 vector of risk contribution per asset.

PortfolioOptimiser.sharpe_ratio — Method

sharpe_ratio(port::AbstractPortfolio; X::AbstractMatrix = port.returns,
             mu::AbstractVector = port.mu,
             type::Symbol = isa(port, Portfolio) ? :Trad : :HRP,
             rm::AbstractRiskMeasure = SD(), delta::Real = 1e-6, rf::Real = 0.0,
             kelly::Bool = false)

Compute the risk-adjusted return ratio for an AbstractRiskMeasure for a portfolio.

Inputs

port: portfolio.

X: T×N matrix of asset returns.

mu: N×1 vector of expected returns.

type: optimisation type used to retrieve the weights vector from port.optimal[type].

rm: risk measure.

delta: small displacement used for computing the Equal risk measure.

rf: risk free rate.

kelly:

if true: use the kelly return.

else: use the arithmetic return.

PortfolioOptimiser.asset_statistics! — Method

asset_statistics!(port::AbstractPortfolio;
                  cov_type::PortfolioOptimiserCovCor = PortCovCor(),
                  set_cov::Bool = true, mu_type::MeanEstimator = MuSimple(),
                  set_mu::Bool = true, kurt_type::KurtFull = KurtFull(),
                  set_kurt::Bool = true, skurt_type::KurtSemi = KurtSemi(),
                  set_skurt::Bool = true, skew_type::SkewFull = SkewFull(),
                  set_skew::Bool = true, sskew_type::SkewSemi = SkewSemi(),
                  set_sskew::Bool = true,
                  cor_type::PortfolioOptimiserCovCor = PortCovCor(),
                  set_cor::Bool = true,
                  dist_type::DistanceMethod = DistanceCanonical(),
                  set_dist::Bool = true)

Compute the asset statistics for a portfolio. See the estimators' docs for details. If a statistic requires another to be computed, the funciton will do so from the relevant estimator.

The set_* variables are flags for deciding whether or not to set the statistic. If a statistic's flag is false the statistic will not be set. Furthermore, if the flag is false and the statistic is not required by another one, it will not be computed.

Inputs

port: portfolio AbstractPortfolio.

cov_type: covariance estimator PortfolioOptimiserCovCor.

set_cov: flag for setting port.cov

mu_type: expected returns estimator MeanEstimator.

set_mu: flag for setting port.mu

kurt_type: cokurtosis matrix estimator KurtFull.

set_kurt: flag for setting port.kurt.

skurt_type: cokurtosis matrix estimator KurtSemi.

set_skurt: flag for setting port.skurt.

skew_type: coskew estimator SkewFull.

set_skew: set port.skew and port.V.

sskew_type: semi coskew estimator SkewSemi.

set_sskew: set port.skew and port.SV.

Only relevant for HCPortfolio.

cor_type: correlation matrix estimator PortfolioOptimiserCovCor.

set_cor: flag for setting port.cor.

dist_type: method for computing the distance matrix DistanceMethod.

set_dist: flag for setting port.dist.

PortfolioOptimiser.black_litterman_factor_statistics! — Method

black_litterman_factor_statistics!(port::Portfolio;
                                   w::AbstractVector = port.bl_bench_weights,
                                   B::Union{DataFrame, Nothing} = port.loadings,
                                   P::Union{AbstractMatrix, Nothing} = nothing,
                                   P_f::Union{AbstractMatrix, Nothing} = nothing,
                                   Q::Union{AbstractVector, Nothing} = nothing,
                                   Q_f::Union{AbstractVector, Nothing} = nothing,
                                   factor_type::FactorType = FactorType(),
                                   cov_type::PortfolioOptimiserCovCor = PortCovCor(),
                                   mu_type::MeanEstimator = MuSimple(),
                                   f_cov_type::PortfolioOptimiserCovCor = PortCovCor(),
                                   f_mu_type::MeanEstimator = MuSimple(),
                                   bl_type::BlackLittermanFactor = BBLType())

Compute the Black Litterman factor model statistics. Na is the number of assets, Nva is the number of asset views, Nf is the number of factors, Nvf is the number of factors views. See the estimators' docs for details.

Inputs

port: portfolio Portfolio.

w: N×1 vector of benchmark weights for the Black-Litterman model.

B: loadings matrix.

if is empty: computes the loadings matrix using factor_type.

P: Nva×Na matrix of asset views.

P_f: Nvf×Nf matrix of factor views.

Q: Nva×1 vector of asset views.

Q_f: Nvf×1 vector of factor views.

factor_type: factor statistics estimator FactorType.

cov_type: asset covariance estimator PortfolioOptimiserCovCor.

mu_type: asset expected returns estimator MeanEstimator.

f_cov_type: factor covariance estimator PortfolioOptimiserCovCor.

f_mu_type: factor expected returns estimator MeanEstimator.

bl_type: Black Litterman factor model estimator BlackLittermanFactor.

PortfolioOptimiser.black_litterman_statistics! — Method

black_litterman_statistics!(port::Portfolio; P::AbstractMatrix, Q::AbstractVector,
                            w::AbstractVector = port.bl_bench_weights,
                            cov_type::PortfolioOptimiserCovCor = PortCovCor(),
                            mu_type::MeanEstimator = MuSimple(),
                            bl_type::BLType = BLType())

Compute the factor statistics. N is the number of assets, Nv is the number of asset views. See the estimators' docs for details.

Inputs

port: portfolio Portfolio.

P: Nv×N matrix of asset views.

Q: Nv×1 vector of asset views.

w: N×1 vector of benchmark weights for the Black-Litterman model.

cov_type: covariance estimator PortfolioOptimiserCovCor.

mu_type: expected returns estimator MeanEstimator.

bl_type: Black Litterman model estimator BLType.

PortfolioOptimiser.factor_statistics! — Method

factor_statistics!(port::Portfolio; factor_type::FactorType = FactorType(),
                   cov_type::PortfolioOptimiserCovCor = PortCovCor(),
                   mu_type::MeanEstimator = MuSimple())

Compute the factor statistics. See the estimators' docs for details.

Inputs

port: portfolio Portfolio.

factor_type: factor statistics estimator FactorType.

cov_type: covariance estimator PortfolioOptimiserCovCor.

mu_type: expected returns estimator MeanEstimator.

PortfolioOptimiser.wc_statistics! — Function

wc_statistics!(port::Portfolio, wc::WCType = WCType(); set_box::Bool = true,
                        set_ellipse::Bool = true)

Compute the worst case mean-variance statistics. Only used in WC optimisations. The set_* variables are used to compute and set the relevant statistics. See the estimators' docs for details.

Inputs

port: portfolio Portfolio.

wc: worst-case mean-variance statistics estimator WCType.

set_box: compute and set the box uncertainty sets port.cov_l, port.cov_u, port.d_mu.

set_ellipse: compute and set the elliptical uncertainty sets and parameters port.cov_mu, port.cov_sigma, port.k_mu, port.k_sigma.

Private

PortfolioOptimiser.AbstractPortfolio — Type

abstract type AbstractPortfolio end

Abstract type for subtyping portfolios.