Portfolio and HCPortfolio

Public

PortfolioOptimiser.HCPortfolio — Type

mutable struct HCPortfolio{ast, dat, r, tmu, tcov, tkurt, tskurt, tl2, ts2, tskew, tv,
                           tsskew, tsv, wmi, wma, ttco, tco, tdist, tcl, tk, topt, tsolv,
                           tf, tlp, taopt, talo, tasolv, taf, tamod} <: AbstractPortfolio
    assets::ast
    timestamps::dat
    returns::r
    mu::tmu
    cov::tcov
    kurt::tkurt
    skurt::tskurt
    L_2::tl2
    S_2::ts2
    skew::tskew
    V::tv
    sskew::tsskew
    SV::tsv
    w_min::wmi
    w_max::wma
    cor_type::ttco
    cor::tco
    dist::tdist
    clusters::tcl
    k::tk
    optimal::topt
    solvers::tsolv
    fail::tf
    latest_prices::tlp
    alloc_optimal::taopt
    alloc_leftover::talo
    alloc_solvers::tasolv
    alloc_fail::taf
    alloc_model::tamod
end

Structure for defining a hierarchical clustering portfolio.

PortfolioOptimiser.Portfolio — Type

mutable struct Portfolio{ast, dat, r, s, us, ul, nal, nau, naus, tfa, tfdat, tretf, l, lo,
                         mnak, mnaks, rb, to, kte, blbw, ami, bvi, rbv, frbv, nm, amc, bvc,
                         ler, tmu, tcov, tkurt, tskurt, tl2, ts2, tskew, tv, tsskew, tsv,
                         tmuf, tcovf, trfm, tmufm, tcovfm, tmubl, tcovbl, tmublf, tcovblf,
                         tcovl, tcovu, tcovmu, tcovs, tdmu, tkmu, tks, topt, tz, tlim,
                         tfront, tsolv, tf, tmod, tlp, taopt, talo, tasolv, taf, tamod} <:
               AbstractPortfolio
    assets::ast
    timestamps::dat
    returns::r
    short::s
    short_u::us
    long_u::ul
    num_assets_l::nal
    num_assets_u::nau
    num_assets_u_scale::naus
    f_assets::tfa
    f_timestamps::tfdat
    f_returns::tretf
    loadings::l
    regression_type::lo
    max_num_assets_kurt::mnak
    max_num_assets_kurt_scale::mnaks
    rebalance::rb
    turnover::to
    tracking_err::kte
    bl_bench_weights::blbw
    a_mtx_ineq::ami
    b_vec_ineq::bvi
    risk_budget::rbv
    f_risk_budget::frbv
    network_method::nm
    a_vec_cent::amc
    b_cent::bvc
    mu_l::ler
    mu::tmu
    cov::tcov
    kurt::tkurt
    skurt::tskurt
    L_2::tl2
    S_2::ts2
    skew::tskew
    V::tv
    sskew::tsskew
    SV::tsv
    f_mu::tmuf
    f_cov::tcovf
    fm_returns::trfm
    fm_mu::tmufm
    fm_cov::tcovfm
    bl_mu::tmubl
    bl_cov::tcovbl
    blfm_mu::tmublf
    blfm_cov::tcovblf
    cov_l::tcovl
    cov_u::tcovu
    cov_mu::tcovmu
    cov_sigma::tcovs
    d_mu::tdmu
    k_mu::tkmu
    k_sigma::tks
    optimal::topt
    z::tz
    limits::tlim
    frontier::tfront
    solvers::tsolv
    fail::tf
    model::tmod
    latest_prices::tlp
    alloc_optimal::taopt
    alloc_leftover::talo
    alloc_solvers::tasolv
    alloc_fail::taf
    alloc_model::tamod
end

Structure for defining a traditional portfolio.

Parameters

assets: Na×1 vector of asset names.

timestamps: T×1 vector of asset returns timestamps.

returns: T×Na matrix of asset returns.

short:

if true: shorting is enabled.

else: long-only portfolio.

short_u: upper bound for the absolute value of the sum of the negative weights.

long_u: upper bound for the sum of the positive weights.

num_assets_l: lower bound for the minimum number of significant assets.

if > 0: applies the constraint.

num_assets_u: upper bound for the maximum number of significant assets.

if > 0: applies the constraint.

num_assets_u_scale: scaling factor for the decision variable used for applying the num_assets_u constraint.

f_assets: Nf×1 vector of factor names.

f_timestamps: T×1 vector of factor returns timestamps.

f_returns: T×Nf matrix of asset returns.

loadings: loadings matrix for working with factor models.

regression_type: RegressionType used for computing the loadings matrix.

max_num_assets_kurt: maximum number of assets to use the complete kurtosis model.

if > 0: the approximate model will be used if the number of assets in the portfolio exceeds max_number_assets_kurt.

max_num_assets_kurt_scale: multipies Na to find the number of eigenvalues when computing the approximate kurtosis model.

rebalance: AbstractTR for defining the portfolio rebalancing penalty.

\[\begin{align} p_{r} &= \sum\limits_{i=1}^{N} r_{i} \lvert w_{i} - b_{i} \rvert\,. \end{align}\]

Where:

$p_{r}$ is the portfolio rebalancing penalty.

$N$ is the number of assets.

$r_{i}$ is the rebalancing penalty for the $i$-th asset.

$w_{i}$ is the weight of the $i$-th asset.

$b_{i}$ is the benchmark weight of the $i$-th asset.

turnover: AbstractTR for defining the asset turnover constraint.

\[\begin{align} \lvert w_{i} - b_{i} \rvert \leq t_{i}\quad \forall i = 1,\,\ldots,\,N\,. \end{align}\]

Where:

$t_{i}$ is the turnover constraint for the $i$-th asset.

$w_{i}$ is the weight of the $i$-th asset.

$b_{i}$ is the benchmark weight of the $i$-th asset.

$N$ is the number of assets.

tracking_err: TrackingErr for defining the tracking error constraint.

\[\begin{align} \left\lVert \dfrac{\mathbf{X} \bm{w} - \bm{b}}{T - 1} \right\rVert_{2} \leq \epsilon \end{align}\]

Where:

$\lVert \cdot \rVert_{2}$ is the L2 norm.

$\mathbf{X}$ is the $T \times N$ matrix of asset returns.

$T$ is the number of returns observations.

$N$ is the number of assets.

$\bm{w}$ is the $N \times 1$ vector of asset weights.

$\bm{b}$ is the $T \times 1$ vector of benchmark returns.

$\epsilon$ is the tracking error.

bl_bench_weights: benchmark weights for Black-Litterman models BlackLittermanClass.

a_mtx_ineq: C×N matrix of asset weight linear constraints.

if isempty: the constraint is not set.

b_vec_ineq: C×1 vector of asset weight linear constraints.

if isempty: the constraint is not set.

The linear weight constraint is defined as.

\[\begin{align} \mathbf{A} \bm{w} \geq \bm{b}\,. \end{align}\]

Where:

$\mathbf{A}$ is the $C×N$ matrix of asset weight linear constraints.

$\bm{b}$ is the $C×1$ vector of asset weight linear constraints.

$C$ is the number of constraints.

$N$ is the number of assets.

risk_budget: Na×1 vector of asset risk budgets.

f_risk_budget: Nf×1 vector of factor risk budgets.

network_method: NetworkMethods for defining the asset network constraint. This can be defined in two ways, using an exact mixed-integer approach IP or an approximate semi-definite one SDP. See their docs for the constraint definition for each case.

if NoNtwk: the constraint is not set.

a_vec_cent: centrality vector for defining the centrality constraint.

if isempty: the constraint is not set.

b_cent: average centrality of the assets the portfolio.

if isinf: the constraint is not set.

The centrality constraint is defined as.

\[\begin{align} \bm{C} \cdot \bm{w} = \bar{c} \end{align}\]

Where:

$\bm{w}$ is the $N\times 1$ vector of asset weights.

$\bm{C}$ is the $N \times 1$ centrality vector of the asset adjacency matrix.

$\cdot$ is the dot product.

$\bar{c}$ is the desired average centrality measure of the portfolio.

mu_l:

mu:

cov:

kurt:

skurt:

L_2:

S_2:

skew:

V:

sskew:

SV:

f_mu:

f_cov:

fm_returns:

fm_mu:

fm_cov:

bl_mu:

bl_cov:

blfm_mu:

blfm_cov:

cov_l:

cov_u:

cov_mu:

cov_sigma:

d_mu:

k_mu:

k_sigma:

optimal:

z:

limits:

frontier:

solvers:

fail:

model:

latest_prices:

alloc_optimal:

alloc_leftover:

alloc_solvers:

alloc_fail:

alloc_model:

PortfolioOptimiser.calc_risk — Method

calc_risk(port::AbstractPortfolio; X::AbstractMatrix = port.returns,
          type::Symbol = isa(port, Portfolio) ? :Trad : :HRP,
          rm::AbstractRiskMeasure = SD())

Compute the risk for an AbstractRiskMeasure for a portfolio.

Inputs

port: portfolio.

X: T×N returns matrix.

type: optimisation type used to retrieve the weights vector from port.optimal[type].

rm: rism measure.

PortfolioOptimiser.factor_risk_contribution — Method

factor_risk_contribution(port::AbstractPortfolio; X::AbstractMatrix = port.returns,
                         F::AbstractMatrix = port.f_returns,
                         type::Symbol = isa(port, Portfolio) ? :Trad : :HRP,
                         rm::AbstractRiskMeasure = SD(), delta::Real = 1e-6)

Compute the factor risk contribution for an AbstractRiskMeasure for a portfolio.

Inputs

port: portfolio.

w: Na×1 vector of asset weights.

X: T×Na matrix of asset returns.

F: T×Nf matrix of factor returns.

type: optimisation type used to retrieve the weights vector from port.optimal[type].

rm: risk measure.

delta: small displacement used for computing the marginal risk and equal risk measure Equal.

Outputs

rc_f: Nf×1 vector of risk contribution per factor.

PortfolioOptimiser.risk_contribution — Method

risk_contribution(port::AbstractPortfolio; X::AbstractMatrix = port.returns,
                  type::Symbol = isa(port, Portfolio) ? :Trad : :HRP,
                  rm::AbstractRiskMeasure = SD(), delta::Real = 1e-6,
                  marginal::Bool = false)

Compute the asset risk contribution for an AbstractRiskMeasure for a portfolio.

Inputs

port: portfolio.

X: T×N returns matrix.

type: optimisation type used to retrieve the weights vector from port.optimal[type].

rm: risk measure.

delta: small displacement used for computing the marginal risk and equal risk measure Equal.

marginal:

if true: compute the marginal risk contribution.

else: compute the risk by contribution by multiplying the marginal risk by the asset weight.

Outputs

rc: Na×1 vector of risk contribution per asset.

PortfolioOptimiser.sharpe_ratio — Method

sharpe_ratio(port::AbstractPortfolio; X::AbstractMatrix = port.returns,
             mu::AbstractVector = port.mu,
             type::Symbol = isa(port, Portfolio) ? :Trad : :HRP,
             rm::AbstractRiskMeasure = SD(), delta::Real = 1e-6, rf::Real = 0.0,
             kelly::Bool = false)

Compute the risk-adjusted return ratio for an AbstractRiskMeasure for a portfolio.

Inputs

port: portfolio.

X: T×N matrix of asset returns.

mu: N×1 vector of expected returns.

type: optimisation type used to retrieve the weights vector from port.optimal[type].

rm: risk measure.

delta: small displacement used for computing the Equal risk measure.

rf: risk free rate.

kelly:

+ if `true`: use the kelly return.
+ else: use the arithmetic return.

PortfolioOptimiser.asset_statistics! — Method

asset_statistics!(port::AbstractPortfolio;
                           cov_type::PortfolioOptimiserCovCor = PortCovCor(),
                           set_cov::Bool = true, mu_type::MeanEstimator = MuSimple(),
                           set_mu::Bool = true, kurt_type::KurtFull = KurtFull(),
                           set_kurt::Bool = true, skurt_type::KurtSemi = KurtSemi(),
                           set_skurt::Bool = true, skew_type::SkewFull = SkewFull(),
                           set_skew::Bool = true, sskew_type::SkewSemi = SkewSemi(),
                           set_sskew::Bool = true,
                           cor_type::PortfolioOptimiserCovCor = PortCovCor(),
                           set_cor::Bool = true,
                           dist_type::DistanceMethod = DistanceCanonical(),
                           set_dist::Bool = true)

PortfolioOptimiser.black_litterman_factor_statistics! — Method

black_litterman_factor_statistics!(port::Portfolio;
                                            w::AbstractVector = port.bl_bench_weights,
                                            B::Union{DataFrame, Nothing} = port.loadings,
                                            P::Union{AbstractMatrix, Nothing} = nothing,
                                            P_f::Union{AbstractMatrix, Nothing} = nothing,
                                            Q::Union{AbstractVector, Nothing} = nothing,
                                            Q_f::Union{AbstractVector, Nothing} = nothing,
                                            factor_type::FactorType = FactorType(),
                                            cov_type::PortfolioOptimiserCovCor = PortCovCor(),
                                            mu_type::MeanEstimator = MuSimple(),
                                            f_cov_type::PortfolioOptimiserCovCor = PortCovCor(),
                                            f_mu_type::MeanEstimator = MuSimple(),
                                            bl_type::BlackLittermanFactor = BBLType())

PortfolioOptimiser.black_litterman_statistics! — Method

black_litterman_statistics!(port::Portfolio; P::AbstractMatrix, Q::AbstractVector,
                                     w::AbstractVector = port.bl_bench_weights,
                                     cov_type::PortfolioOptimiserCovCor = PortCovCor(),
                                     mu_type::MeanEstimator = MuSimple(),
                                     bl_type::BLType = BLType())

PortfolioOptimiser.factor_statistics! — Method

factor_statistics!(port::Portfolio; factor_type::FactorType = FactorType(),
                            cov_type::PortfolioOptimiserCovCor = PortCovCor(),
                            mu_type::MeanEstimator = MuSimple())

PortfolioOptimiser.wc_statistics! — Function

wc_statistics!(port::Portfolio, wc::WCType = WCType(); set_box::Bool = true,
                        set_ellipse::Bool = true)

Private

PortfolioOptimiser.AbstractPortfolio — Type

abstract type AbstractPortfolio end

Abstract type for subtyping portfolios.