Portfolio and HCPortfolio

Public

PortfolioOptimiser.HCPortfolio — Type

mutable struct HCPortfolio{ast, dat, r, tmu, tcov, tkurt, tskurt, tl2, ts2, tskew, tv,
                           tsskew, tsv, wmi, wma, ttco, tco, tdist, tcl, tk, topt, tsolv,
                           tf, tlp, taopt, talo, tasolv, taf, tamod} <: AbstractPortfolio
    assets::ast
    timestamps::dat
    returns::r
    mu::tmu
    cov::tcov
    kurt::tkurt
    skurt::tskurt
    L_2::tl2
    S_2::ts2
    skew::tskew
    V::tv
    sskew::tsskew
    SV::tsv
    w_min::wmi
    w_max::wma
    cor_type::ttco
    cor::tco
    dist::tdist
    clusters::tcl
    k::tk
    optimal::topt
    solvers::tsolv
    fail::tf
    latest_prices::tlp
    alloc_optimal::taopt
    alloc_leftover::talo
    alloc_solvers::tasolv
    alloc_fail::taf
    alloc_model::tamod
end

PortfolioOptimiser.Portfolio — Type

mutable struct Portfolio{ast, dat, r, s, us, ul, nal, nau, naus, tfa, tfdat, tretf, l, lo,
                         mnak, mnaks, rb, to, kte, blbw, ami, bvi, rbv, frbv, nm, amc, bvc,
                         ler, tmu, tcov, tkurt, tskurt, tl2, ts2, tskew, tv, tsskew, tsv,
                         tmuf, tcovf, trfm, tmufm, tcovfm, tmubl, tcovbl, tmublf, tcovblf,
                         tcovl, tcovu, tcovmu, tcovs, tdmu, tkmu, tks, topt, tz, tlim,
                         tfront, tsolv, tf, tmod, tlp, taopt, talo, tasolv, taf, tamod} <:
               AbstractPortfolio
    assets::ast
    timestamps::dat
    returns::r
    short::s
    short_u::us
    long_u::ul
    num_assets_l::nal
    num_assets_u::nau
    num_assets_u_scale::naus
    f_assets::tfa
    f_timestamps::tfdat
    f_returns::tretf
    loadings::l
    loadings_opt::lo
    max_num_assets_kurt::mnak
    max_num_assets_kurt_scale::mnaks
    rebalance::rb
    turnover::to
    tracking_err::kte
    bl_bench_weights::blbw
    a_mtx_ineq::ami
    b_vec_ineq::bvi
    risk_budget::rbv
    f_risk_budget::frbv
    network_method::nm
    a_vec_cent::amc
    b_cent::bvc
    mu_l::ler
    mu::tmu
    cov::tcov
    kurt::tkurt
    skurt::tskurt
    L_2::tl2
    S_2::ts2
    skew::tskew
    V::tv
    sskew::tsskew
    SV::tsv
    f_mu::tmuf
    f_cov::tcovf
    fm_returns::trfm
    fm_mu::tmufm
    fm_cov::tcovfm
    bl_mu::tmubl
    bl_cov::tcovbl
    blfm_mu::tmublf
    blfm_cov::tcovblf
    cov_l::tcovl
    cov_u::tcovu
    cov_mu::tcovmu
    cov_sigma::tcovs
    d_mu::tdmu
    k_mu::tkmu
    k_sigma::tks
    optimal::topt
    z::tz
    limits::tlim
    frontier::tfront
    solvers::tsolv
    fail::tf
    model::tmod
    latest_prices::tlp
    alloc_optimal::taopt
    alloc_leftover::talo
    alloc_solvers::tasolv
    alloc_fail::taf
    alloc_model::tamod
end

PortfolioOptimiser.asset_statistics! — Method

asset_statistics!(port::AbstractPortfolio;
                           cov_type::PortfolioOptimiserCovCor = PortCovCor(),
                           set_cov::Bool = true, mu_type::MeanEstimator = MuSimple(),
                           set_mu::Bool = true, kurt_type::KurtFull = KurtFull(),
                           set_kurt::Bool = true, skurt_type::KurtSemi = KurtSemi(),
                           set_skurt::Bool = true, skew_type::SkewFull = SkewFull(),
                           set_skew::Bool = true, sskew_type::SkewSemi = SkewSemi(),
                           set_sskew::Bool = true,
                           cor_type::PortfolioOptimiserCovCor = PortCovCor(),
                           set_cor::Bool = true,
                           dist_type::DistanceMethod = DistanceCanonical(),
                           set_dist::Bool = true)

PortfolioOptimiser.black_litterman_factor_statistics! — Method

black_litterman_factor_statistics!(port::Portfolio;
                                            w::AbstractVector = port.bl_bench_weights,
                                            B::Union{DataFrame, Nothing} = port.loadings,
                                            P::Union{AbstractMatrix, Nothing} = nothing,
                                            P_f::Union{AbstractMatrix, Nothing} = nothing,
                                            Q::Union{AbstractVector, Nothing} = nothing,
                                            Q_f::Union{AbstractVector, Nothing} = nothing,
                                            factor_type::FactorType = FactorType(),
                                            cov_type::PortfolioOptimiserCovCor = PortCovCor(),
                                            mu_type::MeanEstimator = MuSimple(),
                                            f_cov_type::PortfolioOptimiserCovCor = PortCovCor(),
                                            f_mu_type::MeanEstimator = MuSimple(),
                                            bl_type::BlackLittermanFactor = BBLType())

PortfolioOptimiser.black_litterman_statistics! — Method

black_litterman_statistics!(port::Portfolio; P::AbstractMatrix, Q::AbstractVector,
                                     w::AbstractVector = port.bl_bench_weights,
                                     cov_type::PortfolioOptimiserCovCor = PortCovCor(),
                                     mu_type::MeanEstimator = MuSimple(),
                                     bl_type::BLType = BLType())

PortfolioOptimiser.factor_statistics! — Method

factor_statistics!(port::Portfolio; factor_type::FactorType = FactorType(),
                            cov_type::PortfolioOptimiserCovCor = PortCovCor(),
                            mu_type::MeanEstimator = MuSimple())

PortfolioOptimiser.wc_statistics! — Function

wc_statistics!(port::Portfolio, wc::WCType = WCType(); set_box::Bool = true,
                        set_ellipse::Bool = true)

Private