Risk measaures

Some risk measures require solvers to compute.

solvers = Dict(
               # Key-value pair for the solver, solution acceptance 
               # criteria, and solver attributes.
               :Clarabel => Dict(
                                 # Solver we wish to use.
                                 :solver => Clarabel.Optimizer,
                                 # (Optional) Solution acceptance criteria.
                                 :check_sol => (allow_local = true, allow_almost = true),
                                 # (Optional) Solver-specific attributes.
                                 :params => Dict("verbose" => false)))

The dictionary contains a key value pair for each solver (plus optional solution acceptance criteria and optional attributes) we want to use.

:solver: defines the solver to use. One can also use JuMP.optimizer_with_attributes to direcly provide a solver with attributes already attached.

:check_sol: (optional) defines the keyword arguments passed on to JuMP.is_solved_and_feasible for accepting/rejecting solutions.

:params: (optional) defines solver-specific parameters.

Users are also able to provide multiple solvers by adding additional key-value pairs to the top-level dictionary/tuple as in the following snippet.

using JuMP
solvers = Dict(:Clarabel => Dict(:solver => Clarabel.Optimizer,
                                 :check_sol => (allow_local = true, allow_almost = true),
                                 :params => Dict("verbose" => false)),
               # Provide solver with pre-attached attributes and no arguments 
               # for the `JuMP.is_solved_and_feasible` function.
               :COSMO => Dict(:solver => JuMP.optimizer_with_attributes(COSMO.Optimizer,
                                                                        "maxiter" => 5000)))

PortfolioOptimiser will iterate over the solvers until it finds the first one to successfully solve the problem.

Public

PortfolioOptimiser.ERM — Function

ERM(x::AbstractVector, z::Real = 1.0, α::Real = 0.05)

Compute the Entropic Risk Measure.

\[\begin{align} \mathrm{ERM}(\bm{X},\, z, \,\alpha) &= z \ln \left(\dfrac{M_{\bm{X}}\left(z^{-1}\right)}{\alpha} \right)\,. \end{align}\]

Where:

$M_{\bm{X}}\left(t\right)$ is the moment generating function of $\bm{X}$.

$\alpha \in (0,\,1)$ is the significance parameter.

Inputs

x: T×1 returns vector.

α: significance level, α ∈ (0, 1).

z: entropic moment, can be obtained from get_z_from_model and get_z after optimising a Portfolio.

ERM(x::AbstractVector, solvers:AbstractDict, α::Real = 0.05)

Compute the Entropic Risk Measure. Used in _EVaR, _EDaR and _EDaR_r.

\[\begin{align} \mathrm{ERM}(\bm{X},\, z, \,\alpha) &= \begin{cases} \underset{z,\, t,\, u}{\inf} & t + z \ln\left(\dfrac{1}{\alpha T}\right)\\ \mathrm{s.t.} & z \geq \sum\limits_{i=1}^{T} u_{i}\nonumber\\ & (-X_{i}-t,\, z,\, u_{i}) \in \mathcal{K}_{\exp} \, \forall \, i=1,\,\dots{},\, T \end{cases}\\ &= \underset{z>0}{\inf}\left\{ z \ln \left(\dfrac{M_{\bm{X}}\left(z^{-1}\right)}{\alpha} \right)\right\}\,. \end{align}\]

Where:

$M_{\bm{X}}\left(t\right)$ is the moment generating function of $\bm{X}$.

$\mathcal{K}_{\mathrm{exp}}$ is the exponential cone.

$\alpha \in (0,\,1)$ is the significance parameter.

Inputs

x: T×1 returns vector.

solvers: abstract dict containing the a JuMP-compatible solver capable of solving exponential conic problems.

α: significance level, α ∈ (0, 1).

If no valid solution is found then NaN will be returned.

PortfolioOptimiser.RRM — Function

RRM(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05, κ::Real = 0.3)

Compute the Relativistic Risk Measure. Used in _RLVaR, _RLDaR and _RLDaR_r.

\[\begin{align} \mathrm{RRM}^{\kappa}_{\alpha}(X) &= \left\{ \begin{align} &\underset{z,\, t,\, \psi,\, \theta,\, \varepsilon,\, \omega}{\text{inf}} && t + z \ln_{\kappa} \left(\dfrac{1}{\alpha T}\right) + \sum\limits_{i=1}^T \left(\psi_{i} + \theta_{i} \right) \nonumber\\ &\mathrm{s.t.} && -X - t + \varepsilon + \omega \leq 0 \nonumber\\ &&&z \geq 0 \\ &&&\left( z\left(\dfrac{1+\kappa}{2\kappa}\right),\, \psi_{i}\left(\dfrac{1+\kappa}{\kappa}\right),\, \varepsilon_{i} \right) \in \mathcal{P}_{3}^{1/(1+\kappa),\, \kappa/(1+\kappa)}\nonumber\\ &&&\left( \omega_{i}\left(\dfrac{1}{1-\kappa}\right),\, \theta_{i}\left(\dfrac{1}{\kappa}\right),\, -z \left(\dfrac{1}{2\kappa}\right) \right) \in \mathcal{P}_{3}^{1-\kappa,\, \kappa}\nonumber\\ &&&\forall \, i=1,\,\dots{},\, T \nonumber \end{align} \right.\,. \end{align}\]

Where:

$\mathcal{P}_3^{\alpha,\, 1-\alpha}$ is the power cone 3D.

$\kappa \in (0,\,1)$ is the relativistic deformation parameter.

$\alpha \in (0,\,1)$ is the significance parameter.

Inputs

x: T×1 returns vector.

solvers: abstract dict containing the a JuMP-compatible solver capable of solving 3D power cone problems.

κ is the relativistic deformation parameter, κ ∈ (0, 1).

α: significance level, α ∈ (0, 1).

PortfolioOptimiser.risk_contribution — Method

risk_contribution(rm::RiskMeasure, w::AbstractVector;
                           X::AbstractMatrix = Matrix{Float64}(undef, 0, 0),
                           V::AbstractMatrix = Matrix{Float64}(undef, 0, 0),
                           SV::AbstractMatrix = Matrix{Float64}(undef, 0, 0),
                           delta::Real = 1e-6, marginal::Bool = false, kwargs...)

Private

PortfolioOptimiser._ADD — Method

_ADD(x::AbstractVector)

Compute the Average Drawdown of uncompounded cumulative returns.

\[\begin{align} \mathrm{ADD_{a}}(\bm{X}) = \dfrac{1}{T} \sum\limits_{j=1}^{T} \mathrm{DD_{a}}(\bm{X}, j)\,. \end{align}\]

Where:

$\mathrm{DD_{a}}(\bm{X}, j)$ is the Drawdown of uncompounded cumulative returns at time $j$ as defined in _DaR.

Inputs

x: T×1 returns vector.

PortfolioOptimiser._ADD_r — Method

_ADD_r(x::AbstractVector)

Compute the Average Drawdown of compounded cumulative returns.

\[\begin{align} \mathrm{ADD_{r}}(\bm{X}) = \dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{r}}(\bm{X},\, j)\,. \end{align}\]

Where:

$\mathrm{DD_{r}}(\bm{X},\, j)$ is the Drawdown of compounded cumulative returns at time $j$ as defined in _DaR_r.

Inputs

x: T×1 returns vector.

PortfolioOptimiser._CDaR — Function

_CDaR(x::AbstractVector, α::Real = 0.05)

Compute the Conditional Drawdown at Risk of uncompounded cumulative returns.

\[\begin{align} \mathrm{CDaR_{a}}(\bm{X},\, \alpha) = \mathrm{DaR_{a}}(\bm{X},\, \alpha) + \dfrac{1}{\alpha T} \sum\limits_{j=0}^{T} \max\left(\mathrm{DD_{a}}(\bm{X},\, j) - \mathrm{DaR_{a}}(\bm{X},\, \alpha),\, 0 \right)\,. \end{align}\]

Where:

$\mathrm{DD_{a}}(\bm{X},\, j)$ is the Drawdown of uncompounded cumulative returns at time $j$ as defined in _DaR, and $\mathrm{DaR_{a}}(\bm{X},\, \alpha)$ the Drawdown at Risk of uncompounded cumulative returns as defined in _DaR.

Inputs

x: T×1 returns vector.

α: significance level, α ∈ (0, 1).

Warning

In-place sorts the input vector.

PortfolioOptimiser._CDaR_r — Function

_CDaR_r(x::AbstractVector, α::Real = 0.05)

Compute the Conditional Drawdown at Risk of compounded cumulative returns.

\[\begin{align} \mathrm{CDaR_{r}}(\bm{X},\, \alpha) = \mathrm{DaR_{r}}(\bm{X},\, \alpha) + \dfrac{1}{\alpha T} \sum\limits_{j=0}^{T} \max\left(\mathrm{DD_{r}}(\bm{X},\, j) - \mathrm{DaR_{r}}(\bm{X},\, \alpha),\, 0 \right)\,. \end{align}\]

Where:

$\mathrm{DD_{r}}(\bm{X},\, j)$ is the Drawdown of compounded cumulative returns at time $j$ as defined in _DaR_r.

$\mathrm{DaR_{r}}(\bm{X},\, \alpha)$ the Drawdown at Risk of compounded cumulative returns as defined in _DaR_r.

Inputs

x: T×1 returns vector.

α: significance level, α ∈ (0, 1).

Warning

In-place sorts the input vector.

PortfolioOptimiser._CVaR — Function

_CVaR(x::AbstractVector, α::Real = 0.05)

Compute the Conditional Value at Risk.

\[\begin{align} \mathrm{CVaR}(\bm{X},\, \alpha) = \mathrm{VaR}(\bm{X},\, \alpha) + \dfrac{1}{\alpha T} \sum\limits_{t=1}^{T} \max\left( -X_{t} - \mathrm{VaR}(\bm{X},\, \alpha),\, 0\right)\,. \end{align}\]

Where:

$\mathrm{VaR}(\bm{X},\, \alpha)$ is the Value at Risk as defined in _VaR.

Inputs

x: T×1 returns vector.

α: significance level, α ∈ (0, 1).

Warning

In-place sorts the input vector.

PortfolioOptimiser._CVaRRG — Method

_CVaRRG(x::AbstractVector; alpha::Real = 0.05, beta::Real = alpha)

Compute the _CVaR Range.

Inputs

x: T×1 returns vector.

alpha: significance level of CVaR losses, alpha ∈ (0, 1).

beta: significance level of CVaR gains, beta ∈ (0, 1).

Warning

In-place sorts the input vector.

PortfolioOptimiser._DaR — Function

_DaR(x::AbstractArray, α::Real = 0.05)

Compute the Drawdown at Risk of uncompounded cumulative returns.

\[\begin{align} \mathrm{DaR_{a}}(\bm{X},\, \alpha) &= \underset{j \in (0,\, T)}{\max} \left\{ \mathrm{DD_{a}}(\bm{X},\, j) \in \mathbb{R} : F_{\mathrm{DD}}\left(\mathrm{DD_{a}}(\bm{X},\, j)\right) < 1 - \alpha \right\}\\ \mathrm{DD_{a}}(\bm{X},\, j) &= \underset{t \in (0,\, j)}{\max}\left( \sum\limits_{i=0}^{t} X_{i} \right) - \sum\limits_{i=0}^{j} X_{i}\\ \mathrm{DD_{a}}(\bm{X}) &= \mathrm{DD_{a}}(\bm{X},\, j) \quad \forall j = 1,\,\ldots,\,T\,. \end{align}\]

Where:

$\mathrm{DD_{a}}(\bm{X})$ is the Drawdown of uncompounded cumulative returns.

$\mathrm{DD_{a}}(\bm{X},\, j)$ is the Drawdown of uncompounded cumulative returns at time $j$.

$\mathrm{DaR_{a}}(\bm{X},\, \alpha)$ the Drawdown at Risk of uncompounded cumulative returns.

Inputs

x: T×1 returns vector.

α: significance level, α ∈ (0, 1).

Warning

In-place sorts the input vector.

PortfolioOptimiser._DaR_r — Function

_DaR_r(x::AbstractArray, α::Real = 0.05)

Compute the Drawdown at Risk of compounded cumulative returns.

\[\begin{align} \mathrm{DaR_{r}}(\bm{X},\, \alpha) &= \underset{j \in (0,\, T)}{\max} \left\{ \mathrm{DD_{r}}(\bm{X},\, j) \in \mathbb{R} : F_{\mathrm{DD}}\left(\mathrm{DD_{r}}(\bm{X},\, j)\right) < 1 - \alpha \right\}\\ \mathrm{DD_{r}}(\bm{X},\, j) &= \underset{t \in (0,\, j)}{\max}\left( \prod\limits_{i=0}^{t} \left(1+X_{i}\right) \right) - \prod\limits_{i=0}^{j} \left(1+X_{i}\right)\\ \mathrm{DD_{r}}(\bm{X}) &= \mathrm{DD_{r}}(\bm{X},\, j) \quad \forall j = 1,\,\ldots,\,T\,. \end{align}\]

Where:

$\mathrm{DD_{r}}(\bm{X})$ is the Drawdown of compounded cumulative returns.

$\mathrm{DD_{r}}(\bm{X},\, j)$ is the Drawdown of compounded cumulative returns at time $j$.

$\mathrm{DaR_{r}}(\bm{X},\, \alpha)$ the Drawdown at Risk of compounded cumulative returns.

Inputs

x: T×1 returns vector.

α: significance level, α ∈ (0, 1).

Warning

In-place sorts the input vector.

PortfolioOptimiser._EDaR — Function

_EDaR(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05)

Compute the Entropic Drawdown at Risk of uncompounded cumulative returns.

\[\begin{align} \mathrm{EDaR_{a}}(\bm{X},\alpha) &= \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\mathrm{DD_{a}}(\bm{X}),\, z, \,\alpha)\right\}\,. \end{align}\]

Where:

$\mathrm{ERM}(\mathrm{DD_{a}}(\bm{X}),\, z, \,\alpha)$ is the Entropic Risk Measure as defined in ERM, using the Drawdown of uncompounded cumulative returns as defined in _DaR.

Inputs

x: T×1 returns vector.

solvers: abstract dict containing the a JuMP-compatible solver capable of solving 3D power cone problems.

α: significance level, α ∈ (0, 1).

PortfolioOptimiser._EDaR_r — Function

_EDaR_r(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05)

Compute the Entropic Drawdown at Risk of compounded cumulative returns.

\[\begin{align} \mathrm{EDaR_{r}}(\bm{X},\alpha) &= \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\mathrm{DD_{r}}(\bm{X}),\, z, \,\alpha)\right\}\,. \end{align}\]

Where:

$\mathrm{ERM}(\mathrm{DD_{r}}(\bm{X}),\, z, \,\alpha)$ is the Entropic Risk Measure as defined in ERM, using the Drawdown of compounded cumulative returns as defined in _DaR_r.

Inputs

x: T×1 returns vector.

solvers: abstract dict containing the a JuMP-compatible solver capable of solving 3D power cone problems.

α: significance level, α ∈ (0, 1).

PortfolioOptimiser._EVaR — Function

_EVaR(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05)

Compute the Entropic Value at Risk.

\[\begin{align} \mathrm{EVaR}(\bm{X},\alpha) = \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\bm{X},\, z, \,\alpha)\right\}\,. \end{align}\]

Where:

$\mathrm{ERM}(\bm{X},\, z, \,\alpha)$ is the entropic risk measure as defined in ERM.

Inputs

x: T×1 returns vector.

solvers: abstract dict containing the a JuMP-compatible solver capable of solving 3D power cone problems.

α: significance level, α ∈ (0, 1).

PortfolioOptimiser._FLPM — Function

_FLPM(x::AbstractVector, r::Real = 0.0)

Compute the First Lower Partial Moment (Omega ratio).

\[\begin{align} \mathrm{FLPM}(\bm{X},\, r) = \dfrac{1}{T} \sum\limits_{t=1}^{T}\max\left(r - X_{t},\, 0\right)\,. \end{align}\]

Inputs

x: T×1 returns vector.

r: minimum return target.

PortfolioOptimiser._GMD — Method

_GMD(x::AbstractVector)

Compute the Gini Mean Difference.

Inputs

x: T×1 returns vector.

Warning

In-place sorts the input vector.

PortfolioOptimiser._Kurt — Function

_Kurt(x::AbstractVector, w::Union{AbstractWeights, Nothing} = nothing;
               scale::Bool = false)

Compute the Square Root Kurtosis.

\[\begin{align} \mathrm{Kurt}(\bm{X}) = \left(\dfrac{1}{T} \sum\limits_{t=1}^{T} \left( X_{t} - \mathbb{E}(\bm{X}) \right)^{4} \right)^{1/2}\,. \end{align}\]

Inputs

x: T×1 returns vector.

w: T×1 optional vector of weights for computing the expected return.

scale: if true divides by 2, used in risk_contribution.

PortfolioOptimiser._MAD — Function

_MAD(x::AbstractVector, w::Union{AbstractWeights, Nothing} = nothing)

Compute the Mean Absolute Deviation.

\[\begin{align} \mathrm{MAD}(\bm{X}) = \dfrac{1}{T} \sum\limits_{t=1}^T \left\lvert X_{t} - \mathbb{E}(\bm{X}) \right\rvert\,. \end{align}\]

Inputs

x: T×1 returns vector.

w: T×1 optional vector of weights for computing the expected return.

PortfolioOptimiser._MDD — Method

_MDD(x::AbstractVector)

Compute the Maximum Drawdown of uncompounded cumulative returns.

\[\begin{align} \mathrm{MDD_{a}}(\bm{X}) = \max\mathrm{DD_{a}}(\bm{X})\,. \end{align}\]

Where:

$\mathrm{DD_{a}}(\bm{X})$ is the Drawdown of uncompounded cumulative returns as defined in _DaR.

Inputs

x: T×1 returns vector.

PortfolioOptimiser._MDD_r — Method

_MDD_r(x::AbstractVector)

Compute the Maximum Drawdown of compounded cumulative returns.

\[\begin{align} \mathrm{MDD_{r}}(\bm{X}) = \max \mathrm{DD_{r}}(\bm{X})\,. \end{align}\]

Where:

$\mathrm{DD_{r}}(\bm{X})$ is the Drawdown of compounded cumulative returns as defined in _DaR_r.

Inputs

x: T×1 returns vector.

PortfolioOptimiser._OWA — Method

_OWA(x::AbstractVector, w::AbstractVector)

Compute the Ordered Weight Array risk measure.

Inputs

w: vector of asset weights.

x: T×1 returns vector.

Warning

In-place sorts the input vector.

PortfolioOptimiser._RG — Method

_RG(x::AbstractVector)

Compute the Range.

Inputs

x: T×1 returns vector.

Warning

In-place sorts the input vector.

PortfolioOptimiser._RLDaR — Function

_RLDaR(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05, κ::Real = 0.3)

Compute the Relativistic Drawdown at Risk of uncompounded cumulative returns.

\[\begin{align} \mathrm{RLDaR_{a}}(\bm{X},\, \alpha,\, \kappa) = \mathrm{RRM}(\mathrm{DD_{a}}(\bm{X}),\, \alpha,\, \kappa)\,. \end{align}\]

Where:

$\mathrm{RRM}(\mathrm{DD_{a}}(\bm{X}),\, \alpha,\, \kappa)$ is the Relativistic Risk Measure as defined in RRM, using the Drawdown of uncompounded cumulative returns as defined in _DaR.

Inputs

x: T×1 returns vector.

solvers: abstract dict containing the a JuMP-compatible solver capable of solving 3D power cone problems.

κ is the relativistic deformation parameter, κ ∈ (0, 1).

α: significance level, α ∈ (0, 1).

PortfolioOptimiser._RLDaR_r — Function

_RLDaR_r(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05, κ::Real = 0.3)

Compute the Relativistic Drawdown at Risk of compounded cumulative returns.

\[\begin{align} \mathrm{RLDaR_{r}}(\bm{X},\, \alpha,\, \kappa) = \mathrm{RRM}(\mathrm{DD_{r}}(\bm{X}),\, \alpha,\, \kappa)\,. \end{align}\]

Where:

$\mathrm{RRM}(\mathrm{DD_{r}}(\bm{X}),\, \alpha,\, \kappa)$ is the Relativistic Risk Measure as defined in RRM, using the Drawdown of compounded cumulative returns as defined in _DaR_r.

Inputs

x: T×1 returns vector.

solvers: abstract dict containing the a JuMP-compatible solver capable of solving 3D power cone problems.

κ is the relativistic deformation parameter, κ ∈ (0, 1).

α: significance level, α ∈ (0, 1).

PortfolioOptimiser._RLVaR — Function

_RLVaR(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05, κ::Real = 0.3)

Compute the Relativistic Value at Risk.

\[\begin{align} \mathrm{RLVaR}(\bm{X},\, \alpha,\, \kappa) = \mathrm{RRM}(\bm{X},\, \alpha,\, \kappa)\,. \end{align}\]

Where:

$\mathrm{RRM}(\bm{X},\, \alpha,\, \kappa)$ is the Relativistic Risk Measure as defined in RRM.

Inputs

x: T×1 returns vector.

α: significance level, α ∈ (0, 1).

κ: relativistic deformation parameter.

PortfolioOptimiser._SD — Method

_SD(w::AbstractVector, Σ::AbstractMatrix)

Compute the Standard Deviation. Square root of _Variance.

\[\begin{align} \mathrm{SD}(\bm{w},\, \mathbf{\Sigma}) = \left(\bm{w}^\intercal \, \mathbf{\Sigma} \, \bm{w}\right)^{1/2}\,. \end{align}\]

Inputs

w: vector of asset weights.

Σ: covariance matrix of asset returns.

PortfolioOptimiser._SKurt — Function

_SKurt(x::AbstractVector, r::Real = 0.0,
        w::Union{AbstractWeights, Nothing} = nothing; scale::Bool = false)

Compute the Square Root Semi Kurtosis.

\[\begin{align} \mathrm{SKurt}(\bm{X}) = \left(\dfrac{1}{T} \sum\limits_{t=1}^{T} \min\left( X_{t} - \mathbb{E}(\bm{X}),\, r \right)^{4} \right)^{1/2}\,. \end{align}\]

Inputs

x: T×1 returns vector.

r: minimum return target.

w: T×1 optional vector of weights for computing the expected return.

scale: if true divides by 2, used in risk_contribution.

PortfolioOptimiser._SLPM — Function

_SLPM(x::AbstractVector, r::Real = 0.0)

Compute the Second Lower Partial Moment (Sortino Ratio).

\[\begin{align} \mathrm{SLPM}(\bm{X},\, r) = \left(\dfrac{1}{T-1} \sum\limits_{t=1}^{T}\max\left(r - X_{t},\, 0\right)^{2}\right)^{1/2}\,. \end{align}\]

Inputs

x: T×1 returns vector.

r: minimum return target.

PortfolioOptimiser._SSD — Function

_SSD(x::AbstractVector, r::Real = 0.0, w::Union{AbstractWeights, Nothing} = nothing)

Compute the Semi-Standard Deviation.

\[\begin{align} \mathrm{SSD}(\bm{X}) = \left(\dfrac{1}{T-1} \sum\limits_{t=1}^{T}\min\left(X_{t} - \mathbb{E}(\bm{X}),\, r\right)^{2}\right)^{1/2}\,. \end{align}\]

Inputs

x: T×1 returns vector.

r: minimum return target.

w: T×1 optional vector of weights for computing the expected return.

PortfolioOptimiser._SVariance — Function

_SVariance(x::AbstractVector, r::Real = 0.0, w::Union{AbstractWeights, Nothing} = nothing)

Compute the Semi-Standard Deviation.

\[\begin{align} \mathrm{SVariance}(\bm{X}) = \dfrac{1}{T-1} \sum\limits_{t=1}^{T}\min\left(X_{t} - \mathbb{E}(\bm{X}),\, r\right)^{2}\,. \end{align}\]

Inputs

x: T×1 returns vector.

r: minimum return target.

w: T×1 optional vector of weights for computing the expected return.

PortfolioOptimiser._Skew — Method

_Skew(w::AbstractVector, V::AbstractArray)

PortfolioOptimiser._TG — Method

_TG(x::AbstractVector; alpha_i::Real = 0.0001, alpha::Real = 0.05, a_sim::Int = 100)

Compute the Tail Gini.

Inputs

x: T×1 returns vector.

alpha_i: start value of the significance level of CVaR losses, 0 < alpha_i < alpha < 1.

alpha: end value of the significance level of CVaR losses, alpha ∈ (0, 1).

a_sim: number of CVaRs to approximate the Tail Gini losses, a_sim > 0.

Warning

In-place sorts the input vector.

PortfolioOptimiser._TGRG — Method

_TGRG(x::AbstractVector; alpha_i::Real = 0.0001, alpha::Real = 0.05, a_sim::Real = 100,
     beta_i::Real = alpha_i, beta::Real = alpha, b_sim::Integer = a_sim)

Compute the Tail Gini Range.

Inputs

x: T×1 returns vector.

alpha_i: start value of the significance level of CVaR losses, 0 < alpha_i < alpha < 1.

alpha: end value of the significance level of CVaR losses, alpha ∈ (0, 1).

a_sim: number of CVaRs to approximate the Tail Gini losses, a_sim > 0.

beta_i: start value of the significance level of CVaR gains, 0 < beta_i < beta < 1.

beta: end value of the significance level of CVaR gains, beta ∈ (0, 1).

b_sim: number of CVaRs to approximate the Tail Gini gains, b_sim > 0.

Warning

In-place sorts the input vector.

PortfolioOptimiser._UCI — Method

_UCI(x::AbstractVector)

Compute the Ulcer Index of uncompounded cumulative returns.

\[\begin{align} \mathrm{UCI_{a}}(\bm{X}) = \left(\dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{a}}(\bm{X},\, j)^{2}\right)^{1/2}\,. \end{align}\]

Where:

$\mathrm{DD_{a}}(\bm{X},\, j)$ is the Drawdown of uncompounded cumulative returns at time $j$ as defined in _DaR.

Inputs

x: T×1 returns vector.

PortfolioOptimiser._UCI_r — Method

_UCI_r(x::AbstractVector)

Compute the Ulcer Index of compounded cumulative returns.

\[\begin{align} \mathrm{UCI_{r}}(\bm{X}) = \left(\dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{r}}(\bm{X},\, j)^{2}\right)^{1/2}\,. \end{align}\]

Where:

$\mathrm{DD_{r}}(\bm{X},\, j)$ is the Drawdown of compounded cumulative returns at time $j$ as defined in _DaR_r.

Inputs

x: T×1 returns vector.

PortfolioOptimiser._VaR — Function

_VaR(x::AbstractVector, α::Real = 0.05)

Compute the Value at Risk, used in _CVaR.

\[\begin{align} \mathrm{VaR}(\bm{X},\, \alpha) = -\underset{t \in (0,\, T)}{\inf} \left\{ X_{t} \in \mathbb{R} : F_{\bm{X}}(X_{t}) > \alpha \right\}\,. \end{align}\]

Inputs

x: T×1 returns vector.

α: significance level, α ∈ (0, 1).

Warning

In-place sorts the input vector.

PortfolioOptimiser._Variance — Method

_Variance(w::AbstractVector, Σ::AbstractMatrix)

Compute the Variance. Square of _SD.

\[\begin{align} \mathrm{Variance}(\bm{w},\, \mathbf{\Sigma}) = \bm{w}^\intercal \, \mathbf{\Sigma}\, \bm{w}\,. \end{align}\]

Inputs

w: vector of asset weights.

Σ: covariance matrix of asset returns.

PortfolioOptimiser._WR — Method

_WR(x::AbstractVector)

Compute the Worst Realisation or Worst Case Scenario.

\[\begin{align} \mathrm{WR}(\bm{X}) = -\min(\bm{X})\,. \end{align}\]

Inputs

x: T×1 returns vector.

PortfolioOptimiser._dVar — Method

_dVar(x::AbstractVector)

Compute the Brownian distance variance.

\[\begin{align} \mathrm{dVar}(\bm{X}) &= \mathrm{dCov}(\bm{X},\, \bm{X}) = \dfrac{1}{T^{2}} \sum\limits_{i=1}^{T}\sum\limits_{j=1}^{T} A_{i,\,j}^2\\ \mathrm{dCov}(\bm{X},\, \bm{Y}) &= \dfrac{1}{T^{2}} \sum\limits_{i=1}^{T} \sum\limits_{j=1}^{T} A_{i,\,j} B_{i,\,j}\\ A_{i,\,j} &= a_{i,\,j} - \bar{a}_{i\,.} - \bar{a}_{.\,j} + \bar{a}_{.\,.}\\ B_{i,\,j} &= b_{i,\,j} - \bar{b}_{i\,.} - \bar{b}_{.\,j} + \bar{b}_{.\,.}\\ a_{i,\,j} &= \lVert X_{i} - X_{j} \rVert_{2}, \quad \forall i,\, j = 1,\, \ldots ,\, T\\ b_{i,\,j} &= \lVert Y_{i} - Y_{j} \rVert_{2}, \quad \forall i,\, j = 1,\, \ldots ,\, T\,. \end{align}\]

where:

$\bm{X}$ and $\bm{Y}$ are random variables, they are equal in this case as they are the portfolio returns.

$a_{i,\,j}$ and $b_{i,\,j}$ are entries of a distance matrix where $i$ and $j$ are points in time. Each entry is defined as the Euclidean distance $\lVert \cdot \rVert_{2}$ between the value of the random variable at time $i$ and its value at time $j$.

$\bar{a}_{i,\,\cdot}$ and $\bar{b}_{i,\,\cdot}$ are the $i$-th row means of their respective matrices.

$\bar{a}_{\cdot,\,j}$ and $\bar{b}_{\cdot,\,j}$ are the $j$-th column means of their respective matrices.

$\bar{a}_{\cdot,\,\cdot}$ and $\bar{b}_{\cdot,\,\cdot}$ are the grand means of their respective matrices.

$A_{i,\,j}$ and $B_{i,\,j}$ are doubly centered distances.

Inputs

x: T×1 returns vector.