Risk Functions

PortfolioOptimiser.port_variance — Function

port_variance(w, cov_mtx)

Computes the overall portfolio variance according to the weights w, and covariance matrix cov_mtx.

The portfolio variance $\sigma^2$, is defined as:

\[\sigma^2 = \bm{w}^T\, \Sigma\, \bm{w}\,,\]

where $\sigma$ is the overall portfolio standard deviation, $^T$ the transpose operator, $w$ the weights, and $\Sigma$ the covariance matrix.

port_variance(w::AbstractVector, cov::AbstractMatrix)

Compute the variance of a portfolio with weights w and covariance cov. Square of port_std.

\[\mathrm{var}(\bm{x}, \mathbf{\Sigma}) = \bm{x}^\intercal \mathbf{\Sigma} \bm{x}\,,\]

where $\bm{x}$ are the weights and $\mathbf{\Sigma}$ the covariance matrix.

PortfolioOptimiser.port_std — Function

port_std(w::AbstractVector, cov::AbstractMatrix)

Compute the standard deviation of a portfolio with weights w and covariance cov. Square root of port_variance.

\[\mathrm{std}(\bm{x}, \mathbf{\Sigma}) = \sqrt{\bm{x}^\intercal \mathbf{\Sigma} \bm{x}}\,,\]

where $\bm{x}$ are the weights and $\mathbf{\Sigma}$ the covariance matrix.

PortfolioOptimiser.port_mad — Function

port_mad(x::AbstractVector)

Compute the mean absolute deviation of a vector x of portfolio returns.

\[\mathrm{MAD}(\bm{x}) = \dfrac{1}{T} \sum\limits_{t=1}^T \lvert x_t - \mathbb{E}(\bm{x}) \rvert\,.\]