Risk Functions

PortfolioOptimiser.Variance — Function

Variance(w::AbstractVector, Σ::AbstractMatrix)

Compute the Variance of a portfolio with weights w and covariance Σ. Square of SD.

\[\mathrm{Variance}(\bm{w},\, \mathbf{\Sigma}) = \bm{w}^\intercal \, \mathbf{\Sigma}\, \bm{w}\,.\]

PortfolioOptimiser.SD — Function

SD(w::AbstractVector, Σ::AbstractMatrix)

Compute the Standard Deviation of a portfolio with weights w and covariance Σ. Square root of Variance.

\[\mathrm{SD}(\bm{w},\, \mathbf{\Sigma}) = \left[\bm{w}^\intercal \, \mathbf{\Sigma} \, \bm{w}\right]^{1/2}\,.\]

PortfolioOptimiser.MAD — Function

MAD(x::AbstractVector)

Compute the Mean Absolute Deviation of a vector x of portfolio returns.

\[\mathrm{MAD}(\bm{x}) = \dfrac{1}{T} \sum\limits_{t=1}^T \left\lvert x_t - \mathbb{E}(\bm{x}) \right\rvert\,.\]

PortfolioOptimiser.SSD — Function

SSD(x::AbstractVector)

Compute the mean Semi-Standard Deviation of a vector x of portfolio returns.

\[\mathrm{SSD}(\bm{x}) = \left[\dfrac{1}{T-1} \sum\limits_{t=1}^{T}\min\left(\bm{x}_{t} - \mathbb{E}(\bm{x}),\, 0\right)^{2}\right]^{1/2}\,.\]

PortfolioOptimiser.FLPM — Function

FLPM(x::AbstractVector, r::Real = 0.0)

Compute the First Lower Partial Moment (omega ratio) of a vector x of portfolio returns with a minimum return target of r.

\[\mathrm{FLPM}(\bm{x},\, r) = \dfrac{1}{T} \sum\limits_{t=1}^{T}\max\left(r - \bm{x}_{t},\, 0\right)\,.\]

PortfolioOptimiser.SLPM — Function

SLPM(x::AbstractVector, r::Real = 0.0)

Compute the Second Lower Partial Moment (Omega Ratio) of a vector x of portfolio returns with a minimum return target of r.`

\[\mathrm{SLPM}(\bm{x},\, r) = \left[\dfrac{1}{T-1} \sum\limits_{t=1}^{T}\max\left(r - \bm{x}_{t},\, 0\right)^{2}\right]^{1/2}\,\]

PortfolioOptimiser.WR — Function

WR(x::AbstractVector)

Compute the Worst Realisation or Worst Case Scenario of a returns vector x.

\[\begin{align*} \mathrm{WR}(\bm{x}) &= -\min(\bm{x})\\ &= \max(-\bm{x})\,. \end{align*}\]

PortfolioOptimiser.VaR — Function

VaR(x::AbstractVector, α::Real = 0.05)

Compute the Value at Risk of a returns vector x at a significance level of α ∈ (0, 1).

\[\mathrm{VaR}(\bm{x},\, \alpha) = -\underset{t \in (0,\, T)}{\inf} \left\{ x_{t} \in \mathbb{R} : F_{\bm{x}}(x_{t}) > \alpha \right\}\,,\]

PortfolioOptimiser.CVaR — Function

CVaR(x::AbstractVector, α::Real = 0.05)

Compute the Conditional Value at Risk of a returns vector x at a significance level of α ∈ (0, 1).

\[\begin{align*} \mathrm{CVaR}(\bm{x},\, \alpha) &= \mathrm{VaR}(\bm{x},\, \alpha) - \dfrac{1}{\alpha T} \sum\limits_{t=1}^{T} \min\left( x_t + \mathrm{VaR}(\bm{x},\, \alpha),\, 0\right)\\ &= \mathrm{VaR}(\bm{x},\, \alpha) + \dfrac{1}{\alpha T} \sum\limits_{t=1}^{T} \max\left( -x_t - \mathrm{VaR}(\bm{x},\, \alpha),\, 0\right)\,. \end{align*}\]

PortfolioOptimiser.ERM — Function

ERM(x::AbstractVector, z::Real = 1.0, α::Real = 0.05)

Compute the Entropic Risk Measure of a vector x at a significance level of α ∈ (0, 1) for a given value of z,

\[\mathrm{ERM}(\bm{x},\, z, \,\alpha) = z \ln \left( \dfrac{M_{\bm{x}}\left(z^{-1}\right)}{\alpha} \right)\,,\]

where $M_{\bm{x}}\left(z^{-1}\right)$ is the moment generating function of $\bm{x}$.

ERM(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05)

Compute the Entropic Risk Measure at a significance level of α ∈ (0, 1) by minimising with respect to z, using a dictionary of JuMP-supported solvers with Exponential Cone support. This is because in general we don't know the value of z that minimises the function,

\[\mathrm{ERM} = \begin{cases} \underset{z,\, t,\, u}{\min} & t + z \ln(\dfrac{1}{\alpha T})\\ \mathrm{s.t.} & z \geq \sum\limits_{i=1}^{T} u_{i}\\ & (x_{i}-t,\, z,\, u_{i}) \in \mathcal{K}_{\exp} \, \forall i=1,\,\dots{},\, T \end{cases}\,,\]

where $\mathcal{K}_{xp}$ is the exponential cone.

PortfolioOptimiser.EVaR — Function

EVaR(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05)

Compute the Entropic Value at Risk of a returns vector x at a significance level of α ∈ (0, 1), using a dictionary of JuMP-supported solvers with Exponential Cone support.

\[\mathrm{EVaR}(\bm{x},\alpha) = \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\bm{x},\, z, \,\alpha)\right\}\,,\]

where $\mathrm{ERM}(\bm{x},\, z, \,\alpha)$ is the entropic risk measure at a significance level of α ∈ (0, 1) as defined by ERM.

PortfolioOptimiser.RRM — Function

RRM(
    x::AbstractVector,
    solvers::AbstractDict,
    α::Real = 0.05,
    κ::Real = 0.3,
)

Compute the Relativistic Risk Measure of a vector x at a significance level of α ∈ (0, 1) and deformation parameter of κ ∈ (0, 1), using a dictionary of JuMP-supported solvers with 3D Power Cone support.

\[\mathrm{RRM}(\bm{x},\, \alpha,\, \kappa) = \begin{cases} \underset{z,\, t,\, \psi,\, \theta,\, \varepsilon,\, \omega}{\min} & t + z \ln_{\kappa}\left(\dfrac{1}{\alpha T}\right) + \sum\limits_{i=1}^{T} \left(\psi_{i} + \theta_{i}\right) \\ \mathrm{s.t.} & z \geq 0 \\ & -x_{i} - t + \varepsilon_{i} + \omega_{i} \leq 0 \\ & \left(z\left(\dfrac{1+\kappa}{2\kappa}\right),\, \psi_{i}\left(\dfrac{1+\kappa}{\kappa}\right),\, \varepsilon_{i} \right) \in \mathcal{P}_{3}^{1/(1+\kappa),\, \kappa/(1+\kappa)} \\ & \left(\omega_{i}\left(\dfrac{1}{1-\kappa}\right),\, \theta_{i}\left(\dfrac{1}{\kappa}\right),\, -z\left(\dfrac{1}{2\kappa}\right) \right) \in \mathcal{P}_{3}^{1-\kappa,\, \kappa} \\ & \forall i = 1,\,\ldots{},\, T \end{cases}\,,\]

where $\ln_{\kappa}(x) = \dfrac{x^{\kappa} - x^{-\kappa}}{2 \kappa}$ and $\mathcal{P}_3^{\alpha,\, 1-\alpha}$ is the 3D Power Cone.

PortfolioOptimiser.RVaR — Function

RVaR(
    x::AbstractVector,
    solvers::AbstractDict,
    α::Real = 0.05,
    κ::Real = 0.3,
)

Compute the Relativistic Value at Risk of a vector of returns x at a significance level of α ∈ (0, 1) and deformation parameter of κ ∈ (0, 1), using a dictionary of JuMP-supported solvers with 3D Power Cone support.

\[\mathrm{RVaR}(\bm{x},\, \alpha,\, \kappa) = \mathrm{RRM}(\bm{x},\, \alpha,\, \kappa)\,,\]

where $\mathrm{RRM}(\bm{x},\, \alpha,\, \kappa)$ is the Relativistic Risk Measure as defined in RRM.

PortfolioOptimiser.DaR_abs — Function

DaR_abs(x::AbstractArray, alpha::Real = 0.05)

Compute the Drawdown at Risk forof a returns vector x for a significance level α ∈ (0, 1) using uncompounded cumulative returns.

\[\begin{align*} \mathrm{DaR_{a}}(\bm{x},\, \alpha) &= \underset{j \in (0,\, T)}{\max} \left\{ \mathrm{DD_{a}}(\bm{x},\, j) \in \mathbb{R} : F_{\mathrm{DD}}\left(\mathrm{DD_{a}}(\bm{x},\, j)\right) < 1 - \alpha \right\}\\ \mathrm{DD_{a}}(\bm{x},\, j) &= \underset{t \in (0,\, j)}{\max}\left( \sum\limits_{i=0}^{t} x_{i} \right) - \sum\limits_{i=0}^{j} x_{i} \end{align*}\,.\]

PortfolioOptimiser.MDD_abs — Function

MDD_abs(x::AbstractVector)

Compute the Maximum Drawdown of a returns vector x using uncompounded cumulative returns.

\[\mathrm{MDD_{a}}(\bm{x}) = \underset{j \in (0,\, T)}{\max} \mathrm{DD_{a}}(\bm{x},\, j)\,,\]

where $\mathrm{DD_{a}}(\bm{x},\, j)$ is the uncompounded cumulative Drawdown as defined in DaR_abs.

PortfolioOptimiser.ADD_abs — Function

ADD_abs(x::AbstractVector)

Compute the Average Drawdown of a returns vector x using uncompounded cumulative returns.

\[\mathrm{ADD_{a}}(\bm{x}) = \dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{a}}(\bm{x},\, j)\,,\]

where $\mathrm{DD_{a}}(\bm{x},\, j)$ is the uncompounded cumulative Drawdown as defined in DaR_abs.

PortfolioOptimiser.CDaR_abs — Function

CDaR_abs(x::AbstractVector, alpha::Real = 0.05)

Compute the Conditional Drawdown at Risk of a returns vector x using uncompounded cumulative returns at a significance level of α ∈ (0, 1).

\[\mathrm{CDaR_{a}}(\bm{x},\, \alpha) = \mathrm{DaR_{a}}(\bm{x},\, \alpha) + \dfrac{1}{\alpha T} \sum\limits_{j=0}^{T} \max\left[\mathrm{DD_{a}}(\bm{x},\, j) - \mathrm{DaR_{a}}(\bm{x},\, \alpha),\, 0 \right] \,,\]

where $\mathrm{DD_{a}}(\bm{x},\, j)$ is the uncompounded cumulative Drawdown as defined in DaR_abs, and $\mathrm{DaR_{a}}(\bm{x},\, \alpha)$ the Drawdown at Risk of uncompounded cumulative returns as defined in DaR_abs.

PortfolioOptimiser.UCI_abs — Function

UCI_abs(x::AbstractVector)

Compute the Ulcer Index of a returns vector x using uncompounded cumulative returns.

\[\mathrm{UCI_{a}}(\bm{x}) = \left[\dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{a}}(\bm{x},\, j)^{2}\right]^{1/2}\,,\]

where $\mathrm{DD_{a}}(\bm{x},\, j)$ is the uncompounded cumulative Drawdown as defined in DaR_abs.

PortfolioOptimiser.EDaR_abs — Function

EDaR_abs(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05)

Compute the Entropic Drawdown at Risk of uncompounded cummulative returns of a returns vector x at a significance level of α ∈ (0, 1), using a dictionary of JuMP-supported solvers with Exponential Cone support.

\[\begin{align*} \mathrm{EDaR_{a}}(\bm{x},\alpha) &= \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\mathrm{DD_{a}}(\bm{x}),\, z, \,\alpha)\right\}\\ \mathrm{DD_{a}}(\bm{x}) &= \left\{j \in (0,\, T) : \mathrm{DD_{a}}(\bm{x},\, j) \right\}\,, \end{align*}\]

where $\mathrm{ERM}(\bm{x},\, z, \,\alpha)$ is the entropic risk measure at a significance level of α ∈ (0, 1) as defined by ERM and $\mathrm{DD_{a}}(\bm{x},\, j)$ the drawdown of uncompounded returns as defined in DaR_abs.

PortfolioOptimiser.RDaR_abs — Function

RDaR_abs(
    x::AbstractVector,
    solvers::AbstractDict,
    alpha::Real = 0.05,
    kappa::Real = 0.3,
)

Compute the Relativistic Drawdown at Risk of uncompounded cumulative returns of a vector of returns x at a significance level of α ∈ (0, 1) and deformation parameter of κ ∈ (0, 1), using a dictionary of JuMP-supported solvers with 3D Power Cone support.

\[\mathrm{RDaR_{a}}(\bm{x},\, \alpha,\, \kappa) = \mathrm{RRM}(\mathrm{DD_{a}}(\bm{x}),\, \alpha,\, \kappa)\,,\]

where $\mathrm{RRM}(\mathrm{DD_{a}}(\bm{x}),\, \alpha,\, \kappa)$ is the Relativistic Risk Measure as defined in RRM where the returns vector, $\mathrm{DD_{a}}(\bm{x})$, is the set of drawdowns of uncompounded returns for the timespan $(0,\, T)$ as defined in EDaR_abs.

PortfolioOptimiser.DaR_rel — Function

DaR_ret(x::AbstractArray, alpha::Real = 0.05)

Compute the Drawdown at Risk of a returns vector x for a significance level α ∈ (0, 1) using compounded cumulative returns.

\[\begin{align*} \mathrm{DaR_{r}}(\bm{x},\, \alpha) &= \underset{j \in (0,\, T)}{\max} \left\{ \mathrm{DD_{r}}(\bm{x},\, j) \in \mathbb{R} : F_{\mathrm{DD}}\left(\mathrm{DD_{r}}(\bm{x},\, j)\right) < 1 - \alpha \right\}\\ \mathrm{DD_{r}}(\bm{x},\, j) &= \underset{t \in (0,\, j)}{\max}\left( \prod\limits_{i=0}^{t} \left(1+x_{i}\right) \right) - \prod\limits_{i=0}^{j} \left(1+x_{i}\right) \end{align*}\,.\]

PortfolioOptimiser.MDD_rel — Function

MDD_rel(x::AbstractVector)

Compute the Maximum Drawdown of a returns vector x using compounded cumulative returns.

\[\mathrm{MDD_{r}}(\bm{x}) = \underset{j \in (0,\, T)}{\max} \mathrm{DD_{r}}(\bm{x},\, j)\,,\]

where $\mathrm{DD_{r}}(\bm{x},\, j)$ is the compounded cumulative Drawdown as defined in DaR_rel.

PortfolioOptimiser.ADD_rel — Function

ADD_rel(x::AbstractVector)

Compute the Average Drawdown of a returns vector x using compounded cumulative returns.

\[\mathrm{ADD_{r}}(\bm{r}) = \dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{r}}(\bm{x},\, j)\,,\]

where $\mathrm{DD_{r}}(\bm{x},\, j)$ is the compounded cumulative Drawdown as defined in DaR_rel.

PortfolioOptimiser.CDaR_rel — Function

CDaR_rel(x::AbstractVector, alpha::Real = 0.05)

Compute the Conditional Drawdown at Risk of a returns vector x using compounded cumulative returns at a significance level of α ∈ (0, 1).

\[\mathrm{CDaR_{r}}(\bm{x},\, \alpha) = \mathrm{DaR_{r}}(\bm{x},\, \alpha) + \dfrac{1}{\alpha T} \sum\limits_{j=0}^{T} \max\left[\mathrm{DD_{r}}(\bm{x},\, j) - \mathrm{DaR_{r}}(\bm{x},\, \alpha),\, 0 \right] \,,\]

where $\mathrm{DD_{r}}(\bm{x},\, j)$ is the compounded cumulative Drawdown as defined in DaR_rel, and $\mathrm{DaR_{r}}(\bm{x},\, \alpha)$ the Drawdown at Risk of compounded cumulative returns as defined in DaR_rel.

PortfolioOptimiser.UCI_rel — Function

UCI_rel(x::AbstractVector)

Compute the Ulcer Index of a returns vector x using compounded cumulative returns.

\[\mathrm{UCI_{r}}(\bm{x}) = \left[\dfrac{1}{T} \sum\limits_{j=0}^{T} \mathrm{DD_{r}}(\bm{x},\, j)^{2}\right]^{1/2}\,,\]

where $\mathrm{DD_{r}}(\bm{x},\, j)$ is the compounded cumulative Drawdown as defined in DaR_rel.

PortfolioOptimiser.EDaR_rel — Function

EDaR_rel(x::AbstractVector, solvers::AbstractDict, α::Real = 0.05)

Compute the Entropic Drawdown at Risk of compounded cummulative returns of a returns vector x at a significance level of α ∈ (0, 1), using a dictionary of JuMP-supported solvers with Exponential Cone support.

\[\begin{align*} \mathrm{EDaR_{r}}(\bm{x},\alpha) &= \underset{z > 0}{\inf} \left\{\mathrm{ERM}(\mathrm{DD_{r}}(\bm{x}),\, z, \,\alpha)\right\}\\ \mathrm{DD_{r}}(\bm{x}) &= \left\{j \in (0,\, T) : \mathrm{DD_{r}}(\bm{x},\, j) \right\}\,, \end{align*}\]

where $\mathrm{ERM}(\bm{x},\, z, \,\alpha)$ is the entropic risk measure at a significance level of α ∈ (0, 1) as defined by ERM and $\mathrm{DD_{r}}(\bm{x},\, j)$ the drawdown of compounded returns as defined in DaR_rel.

PortfolioOptimiser.RDaR_rel — Function

RDaR_rel(
    x::AbstractVector,
    solvers::AbstractDict,
    alpha::Real = 0.05,
    kappa::Real = 0.3,
)

Compute the Relativistic Drawdown at Risk of compounded cumulative returns of a vector of returns x at a significance level of α ∈ (0, 1) and deformation parameter of κ ∈ (0, 1), using a dictionary of JuMP-supported solvers with 3D Power Cone support.

\[\mathrm{RDaR_{r}}(\bm{x},\, \alpha,\, \kappa) = \mathrm{RRM}(\mathrm{DD_{r}}(\bm{x}),\, \alpha,\, \kappa)\,,\]

where $\mathrm{RRM}(\mathrm{DD_{r}}(\bm{x}),\, \alpha,\, \kappa)$ is the Relativistic Risk Measure as defined in RRM where the returns vector, $\mathrm{DD_{r}}(\bm{x})$, is the set of drawdowns of compounded returns for the timespan $(0,\, T)$ as defined in EDaR_rel.

PortfolioOptimiser.Kurt — Function

Kurt(x::AbstractVector)

Compute the square root kurtosis of a returns vector x.

\[\mathrm{Kurt} = \left[\dfrac{1}{T} \sum\limits_{t=1}^{T} \left( x_{t} - \mathbb{E}(\bm{x}) \right)^{4} \right]^{1/2}\,.\]

PortfolioOptimiser.SKurt — Function

SKurt(x::AbstractVector)

Compute the square root semi-kurtosis of a returns vector x.

\[\mathrm{SKurt} = \left[\dfrac{1}{T} \sum\limits_{t=1}^{T} \min\left( x_{t} - \mathbb{E}(\bm{x}),\, 0 \right)^{4} \right]^{1/2}\,.\]

PortfolioOptimiser.GMD — Function

GMD(x::AbstractVector)

Compute the Gini Mean Difference of a returns vector x.

PortfolioOptimiser.RG — Function

RG(x::AbstractVector)

Compute the Range of a returns vector x.

PortfolioOptimiser.RCVaR — Function

RCVaR(
    x::AbstractVector;
    α::Real = 0.05,
    β::Union{<:Real, Nothing} = nothing,
)

Compute the CVaR Range of a returns vector x for a significance level of CVaR losses α ∈ (0, 1) and CVaR gains β ∈ (0, 1). If β is nothing, it takes on the value of α.

PortfolioOptimiser.TG — Function

TG(x::AbstractVector; α_i::Real = 0.0001, α::Real = 0.05, α_sim::Int = 100)

Compute the Tail Gini of a returns vector x for CVaR losses starting at α_i through to α, using α_sim steps.

PortfolioOptimiser.RTG — Function

RTG(
    x::AbstractVector;
    α_i::Real = 0.0001,
    α::Real = 0.05,
    α_sim::Real = 100,
    β_i::Union{<:Real, Nothing} = nothing,
    β::Union{<:Real, Nothing} = nothing,
    β_sim::Union{Int, Nothing} = nothing,
)

Compute the Tail Gini Range of a returns vector x for CVaR losses starting at α_i through to α, using α_sim steps, and for CVaR gains starting at β_i through to β, using β_sim steps. If any argument relating to CVaR gains is nothing, it will take the value of its corresponding CVaR loss argument.

PortfolioOptimiser.OWA — Function

OWA(x::AbstractVector, w::AbstractVector)

Compute the Ordered Weight Array risk measure from a returns vector x and am Ordered Weights Array vector w.

PortfolioOptimiser.calc_risk — Function

calc_risk(
    w::AbstractVector,
    returns::AbstractMatrix,
    cov::AbstractMatrix;
    rm::Symbol = :SD,
    rf::Real = 0.0,
    α_i::Real = 0.0001,
    α::Real = 0.05,
    α_sim::Int = 100,
    β_i::Union{<:Real, Nothing} = nothing,
    β::Union{<:Real, Nothing} = nothing,
    β_sim::Union{<:Real, Nothing} = nothing,
    κ::Real = 0.3,
    solvers::Union{<:AbstractDict, Nothing} = nothing,
)

Compute the risk for a given weights vector w, returns matrix returns, covariance matrix cov, risk measure rm, risk free rate rf, CVaR losses starting at α_i through to α, using α_sim steps, and for CVaR gains starting at β_i through to β, using β_sim steps. If any argument relating to CVaR gains is nothing, it will take the value of its corresponding CVaR loss argument. Relativistic risk measure deformation parameter κ and dictionary of solvers with Exponential Cone and/or 3D Power Cone support in case of using an entropic or relativistic risk measure. Some arguments are only used for specific risk measures.

calc_risk(
    portfolio::AbstractPortfolio;
    type::Symbol = :trad,
    rm::Symbol = :SD,
    rf::Real = 0.0,
)

Compute the risk for a given portfolio of type, for risk measure rm, and risk free rate rf.