Statistics Functions

Public

PortfolioOptimiser.asset_statistics! — Method

asset_statistics!(portfolio::AbstractPortfolio; calc_cov::Bool = true, calc_mu::Bool = true,
                  calc_kurt::Bool = true, calc_cor::Bool = true,
                  cov_opt::CovOpt = CovOpt(;), mu_opt::MuOpt = MuOpt(;),
                  kurt_opt::KurtOpt = KurtOpt(;), cor_opt::CorOpt = CorOpt(;))

Compute the asset statistics for a given portfolio in-place. See covar_mtx_mean_vec, covar_mtx, mean_vec and cokurt_mtx.

Inputs

portfolio: instance of Portfolio or HCPortfolio.

Flags

calc_cor:

isa(portfolio, HCPortfolio):

true: compute the correlation matrix, setting the value of portfolio.cor.

calc_cov:

true: compute the covariance matrix, setting the value of portfolio.cov.

calc_mu:

true: compute the expected returns matrix, setting the value of portfolio.mu.

mu_opt.method ∈ (:JS, :BS, :BOP, :CAPM): use the covariance matrix. Therefore, the covariance matrix is computed whenever calc_mu == true, but portfolio.cov is only set when calc_cov == true.

calc_kurt:

true: compute the cokurtosis and semi cokurtosis matrices, setting portfolio.kurt and portfolio.skurt.

Options

cov_opt: instance of CovOpt, defines the parameters for computing the covariance matrix.

mu_opt: instance of MuOpt, defines the parameters for computing the expected returns vector.

kurt_opt: instance of KurtOpt, defines how the cokurtosis and semi cokurtoes are computed.

cor_opt: instance of CorOpt, defines how the correlation matrix is computed.

PortfolioOptimiser.augmented_black_litterman — Method

augmented_black_litterman(returns::AbstractMatrix, w::AbstractVector;
                          F::Union{AbstractMatrix, Nothing}   = nothing,
                          B::Union{AbstractMatrix, Nothing}   = nothing,
                          P::Union{AbstractMatrix, Nothing}   = nothing,
                          P_f::Union{AbstractMatrix, Nothing} = nothing,
                          Q::Union{AbstractVector, Nothing}   = nothing,
                          Q_f::Union{AbstractVector, Nothing} = nothing,
                          cov_opt::CovOpt                     = CovOpt(;),
                          mu_opt::MuOpt                       = MuOpt(;),
                          f_cov_opt::CovOpt                   = CovOpt(;),
                          f_mu_opt::MuOpt                     = MuOpt(;),
                          bl_opt::BLOpt                       = BLOpt(;))

PortfolioOptimiser.bayesian_black_litterman — Method

bayesian_black_litterman(returns::AbstractMatrix, F::AbstractMatrix, B::AbstractMatrix,
                         P_f::AbstractMatrix, Q_f::AbstractVector;
                         cov_opt::CovOpt = CovOpt(;), mu_opt::MuOpt = MuOpt(;),
                         bl_opt::BLOpt = BLOpt(;))

PortfolioOptimiser.black_litterman — Method

black_litterman(returns::AbstractMatrix, P::AbstractMatrix, Q::AbstractVector,
                w::AbstractVector; cov_opt::CovOpt = CovOpt(;), mu_opt::MuOpt = MuOpt(;),
                bl_opt::BLOpt = BLOpt(;))

PortfolioOptimiser.black_litterman_factor_satistics! — Function

black_litterman_factor_satistics!(portfolio::AbstractPortfolio;
                                  w::AbstractVector                   = Vector{Float64}(undef, 0),
                                  B::Union{DataFrame, Nothing}        = nothing,
                                  P::Union{AbstractMatrix, Nothing}   = nothing,
                                  P_f::Union{AbstractMatrix, Nothing} = nothing,
                                  Q::Union{AbstractVector, Nothing}   = nothing,
                                  Q_f::Union{AbstractVector, Nothing} = nothing,
                                  loadings_opt::LoadingsOpt           = LoadingsOpt(;),
                                  cov_opt::CovOpt                     = CovOpt(;),
                                  mu_opt::MuOpt                       = MuOpt(;),
                                  f_cov_opt::CovOpt                   = CovOpt(;),
                                  f_mu_opt::MuOpt                     = MuOpt(;),
                                  bl_opt::BLOpt                       = BLOpt(;))

PortfolioOptimiser.black_litterman_statistics! — Function

black_litterman_statistics!(portfolio::AbstractPortfolio, P::AbstractMatrix,
                            Q::AbstractVector;
                            w::AbstractVector = Vector{Float64}(undef, 0),
                            cov_opt::CovOpt = CovOpt(;), mu_opt::MuOpt = MuOpt(;),
                            bl_opt::BLOpt = BLOpt(;))

Compute the Black Litterman statistics, portfolio.bl_mu and portfolio.bl_cov, for a given portfolio in-place. See black_litterman.

Inputs

portfolio: instance of Portfolio or HCPortfolio.

P: Nv×Na analyst's views matrix, can be relative or absolute, where Nv is the number of views, and Na the number of assets.

Q: Nv×1 analyst's expected returns vector, where Nv is the number of views.

w: Na×1 benchmark weights vector, sets portfolio.bl_bench_weights.

isempty(w):

isempty(portfolio.bl_bench_weights): every entry is in portfolio.bl_bench_weights is assumed to be equal to 1/Na, and w is set to be equal to portfolio.bl_bench_weights.

!isempty(portfolio.bl_bench_weights): w is set to be equal to portfolio.bl_bench_weights.

!isempty(w): portfolio.bl_bench_weights is set to be equal to w.

Options

cov_opt: instance of CovOpt, defines the parameters for computing the covariance matrix.

mu_opt: instance of MuOpt, defines the parameters for computing the expected returns vector.

bl_opt: instance of BLOpt, defines the parameters for computing the Black Litterman model's statistics.

isnothing(bl_opt.delta): it is defined internally as:

\[\delta = \dfrac{\bm{\mu} \cdot \bm{w} - r}{\bm{w}^{\intercal} \mathbf{\Sigma} \bm{w}}\,.\]

Where:

$\delta$: is bl_opt.delta.

$\bm{\mu}$: is portfolio.mu.

$\bm{w}$: is portfolio.bl_bench_weights.

$r$: is bl_opt.rf.

$\mathbf{\Sigma}$: is portfolio.cov.

Note

Note that both bl_opt and mu_opt have rf fields for the risk-free rate (see MuOpt and BLOpt). This gives users more granular control over the model.

PortfolioOptimiser.cluster_assets — Function

cluster_assets

PortfolioOptimiser.cluster_assets! — Function

cluster_assets!

PortfolioOptimiser.cluster_assets — Method

cluster_assets

PortfolioOptimiser.cluster_assets — Method

cluster_assets

PortfolioOptimiser.cokurt_mtx — Function

cokurt_mtx

PortfolioOptimiser.cor_dist_mtx — Function

cor_dist_mtx(
    returns::AbstractMatrix;    alpha_tail::Real = 0.05,    bins_info::Union{Symbol, Integer} = :KN,    cor_method::Symbol = :Pearson,    cor_args::Tuple = (),    cor_func::Function = cor,    cor_kwargs::NamedTuple = (;),    custom_cor::Union{AbstractMatrix, Nothing} = nothing,    dist_args::Tuple = (),    dist_func::Function = x -> sqrt.(clamp!((1 .- x) / 2, 0, 1)),    dist_kwargs::NamedTuple = (;),    gs_threshold::Real = 0.5,    posdef_args::Tuple = (),    posdef_fix::Symbol = :Nearest,    posdef_func::Function = x -> x,    posdef_kwargs::NamedTuple = (;),    sigma::Union{AbstractMatrix, Nothing} = nothing,    std_args::Tuple = (),    std_func::Function = std,    std_kwargs::NamedTuple = (;),    uplo::Symbol = :L)

PortfolioOptimiser.covar_mtx — Function

covar_mtx(returns::AbstractMatrix; opt::CovOpt = CovOpt(;))

Compute the covariance matrix.

Inputs

returns: T×N matrix of returns, where T is the number of returns observations, and N is the number of assets or factors.

cov_opt: instance of CovOpt, defines the parameters for computing the covariance matrix.

Outputs

sigma: N×N covariance matrix, where N is the number of assets or factors.

PortfolioOptimiser.factor_statistics! — Method

factor_statistics!(portfolio::AbstractPortfolio;    # cov_mtx
                   alpha::Real = 0.0, cov_args::Tuple = (),
                   cov_est::CovarianceEstimator = StatsBase.SimpleCovariance(;
                                                                             corrected = true),
                   cov_func::Function = cov, cov_kwargs::NamedTuple = (;),
                   cov_method::Symbol = :Full,
                   cov_weights::Union{AbstractWeights, Nothing} = nothing,
                   custom_cov::Union{AbstractMatrix, Nothing} = nothing,
                   denoise::Bool = false, detone::Bool = false,
                   gs_threshold::Real = portfolio.gs_threshold, jlogo::Bool = false,
                   kernel = ASH.Kernels.gaussian, m::Integer = 10, method::Symbol = :Fixed,
                   mkt_comp::Integer = 1, n::Integer = 1000, opt_args = (),
                   opt_kwargs = (;), posdef_args::Tuple = (), posdef_fix::Symbol = :Nearest,
                   posdef_func::Function = x -> x, posdef_kwargs::NamedTuple = (;),
                   std_args::Tuple = (), std_func::Function = std,
                   std_kwargs::NamedTuple = (;),
                   target_ret::Union{Real, AbstractVector{<:Real}} = 0.0,    # mean_vec
                   custom_mu::Union{AbstractVector, Nothing} = nothing,
                   mean_args::Tuple = (), mean_func::Function = mean,
                   mean_kwargs::NamedTuple = (;),
                   mkt_ret::Union{AbstractVector, Nothing} = nothing,
                   mu_target::Symbol = :GM, mu_method::Symbol = :Default,
                   mu_weights::Union{AbstractWeights, Nothing} = nothing, rf = 0.0,    # Loadings matrix
                   B::Union{DataFrame, Nothing} = nothing, criterion::Symbol = :pval,
                   error::Bool = true, pca_kwargs::NamedTuple = (;),
                   pca_std_kwargs::NamedTuple = (;), pca_std_type = ZScoreTransform,
                   reg_method::Symbol = :FReg, threshold::Real = 0.05,
                   var_func::Function = var, var_args::Tuple = (),
                   var_kwargs::NamedTuple = (;))

PortfolioOptimiser.loadings_matrix — Function

loadings_matrix(x::DataFrame, y::DataFrame; opt::LoadingsOpt = LoadingsOpt(;))

PortfolioOptimiser.mean_vec — Function

mean_vec(returns::AbstractMatrix; opt::MuOpt = MuOpt(;))

Compute the expected returns vector for a returns series.

Inputs

returns: T×N matrix of returns, where T is the number of returns observations, and N is the number of assets or factors.

mu_opt: instance of MuOpt, defines the parameters for computing the expected returns vector.

Outputs

mu: N×1 vector of expected returns, where N is the number of assets or factors.

PortfolioOptimiser.mu_estimator — Function

mu_estimator

PortfolioOptimiser.mut_var_info_mtx — Function

mut_var_info_mtx(x::AbstractMatrix{<:Real}; bins_info::Union{Symbol, <:Integer} = :KN,
                 normed::Bool = true)

Compute the mutual information and variation of information matrices.

Inputs

x: T×N array containing the returns series of the assets. T is the number of observations and N the number of assets.

bins_info: selection criterion for computing the number of bins used to calculate the mutual and variation of information statistics. Can take on an integer value or the following values:

An integer value explicitly defines the number of bins.

A choice of optimal bin width selection algorithms from BinMethods.

:KN: Knuth's choice.

:FD: Freedman-Diaconis' choice.

:SC: Schotts' choice.

:HGR: Hacine-Gharbi and Ravier's choice.

PortfolioOptimiser.risk_factors — Method

risk_factors(x::DataFrame, y::DataFrame; factor_opt::FactorOpt = FactorOpt(;),
             cov_opt::CovOpt = CovOpt(;), mu_opt::MuOpt = MuOpt(;))

PortfolioOptimiser.wc_statistics! — Function

wc_statistics!(portfolio; box = :Stationary, ellipse = :Stationary, calc_box = true,
               calc_ellipse = true, q = 0.05, n_sim = 3_000, block_size = 3, dmu = 0.1,
               dcov = 0.1, seed = nothing, rng = Random.default_rng(),
               fix_cov_args::Tuple = (), fix_cov_kwargs::NamedTuple = (;))

Worst case optimisation statistics.

Private

PortfolioOptimiser._hierarchical_clustering — Function

_hierarchical_clustering

PortfolioOptimiser.covar_mtx_mean_vec — Method

covar_mtx_mean_vec(returns::AbstractMatrix; cov_opt::CovOpt = CovOpt(;),
                   mu_opt::MuOpt = MuOpt(;))

Compute the expected returns vector and covariance matrix. See covar_mtx and mean_vec.

Inputs

returns: T×N matrix of returns, where T is the number of returns observations, and N is the number of assets or factors.

cov_opt: instance of CovOpt, defines the parameters for computing the covariance matrix.

mu_opt: instance of MuOpt, defines the parameters for computing the expected returns vector.

Outputs

sigma: N×N covariance matrix, where N is the number of assets or factors.

mu: N×1 vector of expected returns, where N is the number of assets or factors.

PortfolioOptimiser.gen_bootstrap — Function

gen_bootstrap(returns::AbstractMatrix; kind::Symbol = :Stationary, n_sim::Integer = 3_000,
              block_size::Integer = 3, seed::Union{<:Integer, Nothing} = nothing,
              rng = Random.default_rng())