Statistics Functions

Public

PortfolioOptimiser.asset_statistics! — Method

asset_statistics!(portfolio::AbstractPortfolio; calc_cov::Bool = true, calc_mu::Bool = true,
                  calc_kurt::Bool = true, calc_skew::Bool=true, calc_cor::Bool = true,
                  cov_opt::CovOpt = CovOpt(;), mu_opt::MuOpt = MuOpt(;),
                  kurt_opt::KurtOpt = KurtOpt(;), cor_opt::CorOpt = CorOpt(;))

Compute the asset statistics for a given portfolio in-place. See covar_mtx, mean_vec, cokurt_mtx, cor_dist_mtx.

Depending on conditions, modifies:

portfolio.mu

portfolio.cov

portfolio.kurt

portfolio.skurt

portfolio.cor

portfolio.dist

mu_opt.method

Inputs

portfolio: instance of Portfolio or HCPortfolio.

Flags

calc_cov:

true: compute and set the covariance matrix.

calc_mu:

true: compute and set the expected returns vector.

mu_opt.method ∈ (:JS, :BS, :BOP, :CAPM): require the covariance matrix, so it will be computed and mu_opt.sigma will be set to the covariance matrix. portfolio.cov will only be modified when calc_cov == true.

calc_kurt:

true: compute and set the cokurtosis and semi cokurtosis matrices.

calc_skew:

true: compute and set the coskewness and semi coskewness matrices.

calc_cor:

isa(portfolio, HCPortfolio):

true: compute and set the correlation and distance matrices.

Options

cov_opt: instance of CovOpt, defines the parameters for computing the covariance matrix.

mu_opt: instance of MuOpt, defines the parameters for computing the expected returns vector.

kurt_opt: instance of KurtOpt, defines how the cokurtosis and semi cokurtoes are computed.

cor_opt: instance of CorOpt, defines how the correlation and distance matrices are computed.

PortfolioOptimiser.black_litterman_factor_statistics! — Function

black_litterman_factor_statistics!(portfolio::Portfolio,
                                  w::AbstractVector                   = portfolio.bl_bench_weights;
                                  B::Union{DataFrame, Nothing}        = portfolio.loadings,
                                  P::Union{AbstractMatrix, Nothing}   = nothing,
                                  P_f::Union{AbstractMatrix, Nothing} = nothing,
                                  Q::Union{AbstractVector, Nothing}   = nothing,
                                  Q_f::Union{AbstractVector, Nothing} = nothing,
                                  loadings_opt::LoadingsOpt           = LoadingsOpt(;),
                                  cov_opt::CovOpt                     = CovOpt(;),
                                  mu_opt::MuOpt                       = MuOpt(;),
                                  f_cov_opt::CovOpt                   = CovOpt(;),
                                  f_mu_opt::MuOpt                     = MuOpt(;),
                                  bl_opt::BLOpt                       = BLOpt(;))

Estimates the factor Black-Litterman statistics in-place according to BLFMMethods. See augmented_black_litterman, bayesian_black_litterman.

Modifies:

portfolio.blfm_mu

portfolio.blfm_cov

portfolio.bl_bench_weights

portfolio.loadings

bl_opt.constant

Depending on conditions, modifies:

bl_opt.delta

portfolio.loadings_opt

Inputs

portfolio: instance of Portfolio.

w: Na×1 benchmark weights vector, sets portfolio.bl_bench_weights, where Na is the number of assets.

isempty(w): every entry is assumed to be 1/Na.

B: is the T×(Nf+c) loadings matrix in Dataframe form, sets portfolio.loadings and bl_opt.constant = "const" ∈ names(B), where T is the number of returns observations, Nf the number of factors, and c ∈ (0, 1, 2) represents the two optional columns as described in FactorOpt.

(isnothing(B) || isempty(B)) && isempty(portfolio.loadings): internally computes B using loadings_matrix and sets portfolio.loadings_opt = loadings_opt.

P: Nva×Na analyst's asset views matrix, can be relative or absolute, where Nva is the number of asset views, and Na the number of assets.

P_F: Nvf×Nf analyst's factor views matrix, can be relative or absolute, where Nvf is the number of factor views, and Nf the number of factors.

Q: Nva×1 analyst's asset views returns vector, where Nva is the number of asset views.

Q_f: Nvf×1 analyst's factor views returns vector, where Nvf is the number of factor views.

Options

loadings_opt: instance of LoadingsOpt, defines the parameters for computing the loadings matrix.

cov_opt: instance of CovOpt, defines the parameters for computing the vanilla asset covariance matrix.

mu_opt: instance of MuOpt, defines the parameters for computing the vanilla asset expected returns vector.

f_cov_opt: instance of CovOpt, defines the parameters for computing the vanilla factor covariance matrix.

f_mu_opt: instance of MuOpt, defines the parameters for computing the vanilla factor expected returns vector.

bl_opt: instance of BLOpt, defines the parameters for computing the factor Black-Litterman model's statistics.

isnothing(bl_opt.delta): sets bl_opt.delta to:

\[\delta = \dfrac{\bm{\mu} \cdot \bm{w} - r}{\bm{w}^{\intercal} \mathbf{\Sigma} \bm{w}}\,.\]

Where:

$\delta$: is bl_opt.delta.

$\bm{\mu}$: is portfolio.mu.

$\bm{w}$: is portfolio.bl_bench_weights.

$r$: is bl_opt.rf.

$\mathbf{\Sigma}$: is portfolio.cov.

Note

Note that both bl_opt, f_mu_opt, and mu_opt have rf fields for the risk-free rate (see MuOpt and BLOpt). This gives users more granular control over the model.

PortfolioOptimiser.black_litterman_statistics! — Function

black_litterman_statistics!(portfolio::Portfolio, P::AbstractMatrix, Q::AbstractVector;
                            w::AbstractVector = portfolio.bl_bench_weights,
                            cov_opt::CovOpt = CovOpt(;), mu_opt::MuOpt = MuOpt(;),
                            bl_opt::BLOpt = BLOpt(;))

Estimates the Black-Litterman statistics for a given portfolio in-place. See black_litterman.

Modifies:

portfolio.bl_mu

portfolio.bl_cov

portfolio.bl_bench_weights

Depending on conditions, modifies:

bl_opt.delta

Inputs

portfolio: instance of Portfolio.

P: Nv×N analyst's views matrix, can be relative or absolute, where Nv is the number of views, and N the number of assets.

Q: Nv×1 analyst's expected returns vector, where Nv is the number of views.

w: N×1 benchmark weights vector, sets portfolio.bl_bench_weights, where N is the number of assets.

isempty(w): every entry is assumed to be 1/N.

Options

cov_opt: instance of CovOpt, defines the parameters for computing the vanilla covariance matrix.

mu_opt: instance of MuOpt, defines the parameters for computing the vanilla expected returns vector.

bl_opt: instance of BLOpt, defines the parameters for computing the Black-Litterman model's statistics.

isnothing(bl_opt.delta): sets bl_opt.delta to:

\[\delta = \dfrac{\bm{\mu} \cdot \bm{w} - r}{\bm{w}^{\intercal} \mathbf{\Sigma} \bm{w}}\,.\]

Where:

$\delta$: is bl_opt.delta.

$\bm{\mu}$: is portfolio.mu.

$\bm{w}$: is portfolio.bl_bench_weights.

$r$: is bl_opt.rf.

$\mathbf{\Sigma}$: is portfolio.cov.

Note

Note that both bl_opt, and mu_opt have rf fields for the risk-free rate (see MuOpt and BLOpt). This gives users more granular control over the model.

PortfolioOptimiser.cluster_assets — Function

cluster_assets

PortfolioOptimiser.cluster_assets! — Function

cluster_assets!

PortfolioOptimiser.cluster_assets — Method

cluster_assets

PortfolioOptimiser.cluster_assets — Method

cluster_assets

PortfolioOptimiser.factor_statistics! — Method

factor_statistics!(portfolio::Portfolio; cov_opt::CovOpt = CovOpt(;),
                   mu_opt::MuOpt = MuOpt(;), factor_opt::FactorOpt = FactorOpt(;))

Compute the factor and factor adjusted statistics for a given portfolio in-place. See covar_mtx_mean_vec, and risk_factors.

Modifies:

portfolio.f_mu

portfolio.f_cov

portfolio.fm_mu

portfolio.fm_cov

portfolio.fm_returns

portfolio.loadings

portfolio.loadings_opt

Inputs

portfolio: instance of Portfolio.

Options

cov_opt: instance of CovOpt, defines the parameters for computing the factor covariance matrix.

mu_opt: instance of MuOpt, defines the parameters for computing the factor expected returns vector.

factor_opt: instance of FactorOpt, defines how the factor statistics are computed.

PortfolioOptimiser.loadings_matrix — Function

loadings_matrix(x::DataFrame, y::DataFrame, opt::LoadingsOpt = LoadingsOpt(;))

Estimate the loadings matrix using regression. See stepwise_regression, and mv_regression.

Inputs

x: is the T×Nf Dataframe of factor returns, where the column names are the factor names, T is the number of returns observations, and Nf the number of factors.

y: is the T×Na Dataframe of asset returns, where the column names are the factor names, T is the number of returns observations, and Na the number of assets.

Options

loadings_opt: instance of LoadingsOpt, defines the parameters for computing the loadings matrix.

Outputs

B: is the (Na+2)×Nf loadings matrix as a Dataframe, where Na is the number of assets, Nf the number of factors. The two extra columns are the optional columns for B as described in FactorOpt.

PortfolioOptimiser.wc_statistics! — Function

wc_statistics!(portfolio::Portfolio, opt::WCOpt = WCOpt(;))

Compute worst case statistics for a given portfolio in-place. See gen_bootstrap, vec_of_vecs_to_mtx, posdef_fix!, and commutation_matrix.

Depending on conditions, modifies:

portfolio.cov_l

portfolio.cov_u

portfolio.d_mu

portfolio.cov_mu

portfolio.cov_sigma

portfolio.k_mu

portfolio.k_sigma

Inputs

portfolio: instance of Portfolio or HCPortfolio.

Options

opt: instance of WCOpt, defines the parameters for computing the worst case statistics.

Private

PortfolioOptimiser._Pi — Method

_Pi(eq, delta, sigma, w, mu, rf)

PortfolioOptimiser._denoise_logo_mtx — Function

_denoise_logo_mtx(T::Integer, N::Integer, mtx::AbstractMatrix,
                  opt::Union{CovOpt, CorOpt, KurtOpt, BLOpt}, mtx_name::Symbol = :cov)

PortfolioOptimiser._hierarchical_clustering — Function

_hierarchical_clustering

PortfolioOptimiser._mu_cov_w — Method

_mu_cov_w(tau, omega, P, Pi, Q, rf, sigma, delta, T, N, opt, cov_type, cov_flag = true)

Internal function for computing the Black Litterman statistics as defined in black_litterman. See _denoise_logo_mtx.

Inputs

tau: variable of the same name in the Black-Litterman model.

omega: variable of the same name in the Black-Litterman model.

P: variable of the same name in the Black-Litterman model.

Pi: variable of the same name in the Black-Litterman model.

Q: variable of the same name in the Black-Litterman model.

rf: variable of the same name in the Black-Litterman model.

sigma: variable of the same name in the Black-Litterman model.

delta: variable of the same name in the Black-Litterman model.

T: variable of the same name in the Black-Litterman model.

N: variable of the same name in the Black-Litterman model.

opt: any valid instance of opt for _denoise_logo_mtx.

cov_type: any valid value from DenoiseLoGoNames.

cov_flag: whether the matrix is a covariance matrix or not.

Outputs

mu: asset expected returns vector obtained via the Black-Litterman model.

cov_mtx: asset covariance matrix obtained via the Black-Litterman model.

w: asset weights obtained via the Black-Litterman model.

Pi_: equilibrium excess returns after being adjusted by the views.

PortfolioOptimiser._omega — Method

_omega(P, tau_sigma)

PortfolioOptimiser.augmented_black_litterman — Method

augmented_black_litterman(returns::AbstractMatrix, w::AbstractVector;
                          F::Union{AbstractMatrix, Nothing}   = nothing,
                          B::Union{AbstractMatrix, Nothing}   = nothing,
                          P::Union{AbstractMatrix, Nothing}   = nothing,
                          P_f::Union{AbstractMatrix, Nothing} = nothing,
                          Q::Union{AbstractVector, Nothing}   = nothing,
                          Q_f::Union{AbstractVector, Nothing} = nothing,
                          cov_opt::CovOpt                     = CovOpt(;),
                          mu_opt::MuOpt                       = MuOpt(;),
                          f_cov_opt::CovOpt                   = CovOpt(;),
                          f_mu_opt::MuOpt                     = MuOpt(;),
                          bl_opt::BLOpt                       = BLOpt(;))

Estimates the Augmented Black-Litterman statistics to BLFMMethods. See covar_mtx_mean_vec, _omega, _Pi, _mu_cov_w, and _denoise_logo_mtx.

Inputs

returns: T×Na matrix of asset returns, where T is the number of returns observations, and Na the number of assets.

w: Na×1 benchmark weights vector, where Na is the number of assets.

F: T×Nf matrix of factor returns, where T is the number of returns observations, and Nf the number of factors.

B: is the T×(Nf+c) loadings matrix, where T is the number of returns observations, Nf the number of factors, and c ∈ (0, 1) represents the optional column for the constant term as described in FactorOpt.

P: Nva×Na analyst's asset views matrix, can be relative or absolute, where Nva is the number of asset views, and Na the number of assets.

P_F: Nvf×Nf analyst's factor views matrix, can be relative or absolute, where Nvf is the number of factor views, and Nf the number of factors.

Q: Nva×1 analyst's asset views returns vector, where Nva is the number of asset views.

Q_f: Nvf×1 analyst's factor views returns vector, where Nvf is the number of factor views.

The model can be adjusted depending on which views matrices and views expected returns vectors are provided.

If any of P or Q are provided, both must be provided.

If any of P_f or Q_f are provided, both must be provided, as well as both B and F must be provided.

If both P and Q are provided, but P_f and Q_f are not, the augmented model reduces to the traditional Black-Litterman model in black_litterman. In other words, all augmented variables defined in BLFMMethods only include the asset statistics.

If P_f and Q_f are provided, but P and Q are not, the augmneted model computes its statistics based on the factor model only. In other words, all augmented variables defined in BLFMMethods only include the estimated asset statistics using the factor model.

If P, Q, P_f and Q_f are all provided, the full augmented model as defined in BLFMMethods is used.

Options

cov_opt: instance of CovOpt, defines the parameters for computing the vanilla asset covariance matrix.

mu_opt: instance of MuOpt, defines the parameters for computing the vanilla asset expected returns vector.

f_cov_opt: instance of CovOpt, defines the parameters for computing the vanilla factor covariance matrix.

f_mu_opt: instance of MuOpt, defines the parameters for computing the vanilla factor expected returns vector.

bl_opt: instance of BLOpt, defines the parameters for computing the factor Black-Litterman model's statistics.

Outputs

mu_a: estimated asset expected returns vector via the Augmented Black-Litterman model.

cov_mtx_a: estimated asset covariance via the Augmented Black-Litterman model.

w_a: estimated benchmark weights via the Augmented Black-Litterman model.

Note

Note that both bl_opt, f_mu_opt, and mu_opt have rf fields for the risk-free rate (see MuOpt and BLOpt). This gives users more granular control over the model.

PortfolioOptimiser.backward_regression — Function

backward_regression(x::DataFrame, y::AbstractVector, criterion::Symbol = :pval,
                    threshold::Real = 0.05)

Select the factors that best estimate the model based on backward regression. This tends to be more robust than forward_regression.

Inputs

x: is the T×N Dataframe of factor returns, where the column names are the factor names, T is the number of returns observations, and N the number of factors.

y: is the T×1 vector of returns for an asset.

criterion: one of RegCriteria that decides what criterion to use when selecting the most significant factors.

threshold:

criterion == :pval: a factor is considered significant if its p-value is lower than threshold. If no factor has a p-value lower than threshold, selects the factor with the lowest p-value.

Outputs

features: is the C×1 vector of significant factors, where C is the number of significant factors.

PortfolioOptimiser.bayesian_black_litterman — Method

bayesian_black_litterman(returns::AbstractMatrix, F::AbstractMatrix, B::AbstractMatrix,
                         P_f::AbstractMatrix, Q_f::AbstractVector;
                         cov_opt::CovOpt = CovOpt(;), mu_opt::MuOpt = MuOpt(;),
                         bl_opt::BLOpt = BLOpt(;))

Estimates the Bayesian Black-Litterman statistics according to BLFMMethods. See covar_mtx_mean_vec, and _omega.

Inputs

returns: T×Na matrix of asset returns, where T is the number of returns observations, and Na the number of assets.

F: T×Nf matrix of factor returns, where T is the number of returns observations, and Nf the number of factors.

B: is the T×(Nf+c) loadings matrix, where T is the number of returns observations, Nf the number of factors, and c ∈ (0, 1) represents the optional column for the constant term as described in FactorOpt.

P_F: Nvf×Nf analyst's factor views matrix, can be relative or absolute, where Nvf is the number of factor views, and Nf the number of factors.

Q_f: Nvf×1 analyst's factor views returns vector, where Nvf is the number of factor views.

Options

cov_opt: instance of CovOpt, defines the parameters for computing the vanilla factor covariance matrix.

mu_opt: instance of MuOpt, defines the parameters for computing the vanilla factor expected returns vector.

bl_opt: instance of BLOpt, defines the parameters for computing the factor Black-Litterman model's statistics.

Outputs

mu_b: estimated asset expected returns vector via the Bayesian Black-Litterman model.

cov_mtx_b: estimated asset covariance via the Bayesian Black-Litterman model.

w_b: estimated benchmark weights via the Bayesian Black-Litterman model.

Note

Note that both bl_opt, and mu_opt have rf fields for the risk-free rate (see MuOpt and BLOpt). This gives users more granular control over the model.

PortfolioOptimiser.black_litterman — Method

black_litterman(returns::AbstractMatrix, P::AbstractMatrix, Q::AbstractVector,
                w::AbstractVector; cov_opt::CovOpt = CovOpt(;), mu_opt::MuOpt = MuOpt(;),
                bl_opt::BLOpt = BLOpt(;))

Estimates the expected returns vector and covariance matrix based on the Black-Litterman model [56, 57]. See covar_mtx_mean_vec, and _mu_cov_w.

\[\begin{align*} \bm{\Pi} &= \begin{cases} \delta \mathbf{\Sigma} \bm{w} &\quad \mathrm{if~ eq = true}\\ \bm{\mu} - r &\quad \mathrm{if~ eq = false} \end{cases}\\ \mathbf{\Omega} &= \tau \mathrm{Diagonal}\left(\mathbf{P} \mathbf{\Sigma} \mathbf{P}^{\intercal}\right)\\ \mathbf{M} &= \left[ \left(\tau \mathbf{\Sigma} \right)^{-1} + \mathbf{P}^{\intercal} \mathbf{\Omega}^{-1} \mathbf{P}\right]^{-1}\\ \bm{\Pi}_{\mathrm{BL}} &= \mathbf{M} \left[\left(\tau \mathbf{\Sigma}\right)^{-1} \bm{\Pi} + \mathbf{P}^{\intercal} \mathbf{\Omega}^{-1} \mathbf{Q} \right]\\ \tau &= \dfrac{1}{T}\\ \bm{\mu}_{\mathrm{BL}} &= \bm{\Pi}_{\mathrm{BL}} + r\\ \mathbf{\Sigma}_{\mathrm{BL}} &= \mathbf{\Sigma} + \mathbf{M} \end{align*}\]

Where:

$\bm{\Pi}$: is N×1 the equilibrium excess returns, where N is the number of assets.

$\delta$: is the risk aversion parameter.

$\mathbf{\Sigma}$: is the N×N asset covariance matrix, where N is the number of assets.

$\bm{w}$: is the N×1 vector of benchmark asset weights, where N is the number of assets.

$\mathbf{P}$: is the Nv×N asset views matrix, where Nv is the number of asset views, and N the number of assets.

$\bm{Q}$: is the Nv×1 asset views returns vector, where Nv is the number of asset views.

$\mathbf{\Omega}$: is the Nv×Nv covariance matrix of the errors of the asset views, where Nv is the number of asset views.

$\mathbf{M}$: is an N×N intermediate covariance matrix, where N is the number of assets, and M the number of assets.

$\bm{\Pi}_{\mathbf{BL}}$: is the N×1 equilibrium excess returns after being adjusted by the views, where N is the number of assets.

$T$: is the number of returns observations.

$\bm{\mu}_{\mathbf{BL}}$: is the N×1 vector of asset expected returns obtained via the Black-Litterman model, where N is the number of assets.

$\mathbf{\Sigma}_{\mathrm{BL}}$: is the N×N asset covariance matrix obtained via the Black-Litterman model, where N is the number of assets.

Inputs

returns: T×N matrix of returns, where T is the number of returns observations, and N is the number of assets or factors.

P: Nv×N analyst's asset views matrix, can be relative or absolute, where Nv is the number of asset views, and N the number of assets.

Q: Nv×1 analyst's asset viewsd returns vector, where Nv is the number of asset views.

w: N×1 benchmark weights vector, where N is the number of assets.

Options

cov_opt: instance of CovOpt, defines the parameters for computing the vanilla covariance matrix.

mu_opt: instance of MuOpt, defines the parameters for computing the vanilla expected returns vector.

bl_opt: instance of BLOpt, defines the parameters for computing the Black-Litterman model's statistics.

Outputs

mu: N×1 Black-Litterman adjusted expected returns vector, where N is the number of assets.

sigma: N×N Black-Litterman adjusted covariance matrix, where N is the number of assets.

w: N×1 Black-Litterman adjusted asset weights vector, where N is the number of assets.

Note

Note that both bl_opt, and mu_opt have rf fields for the risk-free rate (see MuOpt and BLOpt). This gives users more granular control over the model.

PortfolioOptimiser.cokurt_mtx — Function

cokurt_mtx

PortfolioOptimiser.commutation_matrix — Method

commutation_matrix(x::AbstractMatrix)

Generates the commutation matrix for x.

Inputs

x: matrix.

Outputs

y: commutation matrix.

PortfolioOptimiser.cor_dist_mtx — Function

cor_dist_mtx(returns::AbstractMatrix, opt::CorOpt = CorOpt(;))

PortfolioOptimiser.covar_mtx — Function

covar_mtx(returns::AbstractMatrix, opt::CovOpt = CovOpt(;))

Compute the covariance matrix. See gerber0, gerber1, gerber2, sb0, sb1, gerbersb0, gerbersb1, _denoise_logo_mtx, posdef_fix!.

Inputs

returns: T×N matrix of returns, where T is the number of returns observations, and N is the number of assets or factors.

cov_opt: instance of CovOpt, defines the parameters for computing the covariance matrix.

Outputs

sigma: N×N covariance matrix, where N is the number of assets or factors.

PortfolioOptimiser.covar_mtx_mean_vec — Method

covar_mtx_mean_vec(returns::AbstractMatrix, cov_opt::CovOpt = CovOpt(;),
                   mu_opt::MuOpt = MuOpt(;))

Compute the expected returns vector and covariance matrix. See covar_mtx and mean_vec.

Inputs

returns: T×N matrix of returns, where T is the number of returns observations, and N is the number of assets or factors.

cov_opt: instance of CovOpt, defines the parameters for computing the covariance matrix.

mu_opt: instance of MuOpt, defines the parameters for computing the expected returns vector.

Outputs

sigma: N×N covariance matrix, where N is the number of assets or factors.

mu: N×1 vector of expected returns, where N is the number of assets or factors.

PortfolioOptimiser.forward_regression — Function

forward_regression(x::DataFrame, y::AbstractVector, criterion::Symbol = :pval,
                   threshold::Real = 0.05)

Select the factors that best estimate the model based on forward regression.

Inputs

x: is the T×N Dataframe of factor returns, where the column names are the factor names, T is the number of returns observations, and N the number of factors.

y: is the T×1 vector of returns for an asset.

criterion: one of RegCriteria that decides what criterion to use when selecting the most significant factors.

threshold:

criterion == :pval: a factor is considered significant if its p-value is lower than threshold. If no factor has a p-value lower than threshold, selects the factor with the lowest p-value.

Outputs

features: is the C×1 vector of significant factors, where C is the number of significant factors.

PortfolioOptimiser.gen_bootstrap — Function

gen_bootstrap(returns::AbstractMatrix, cov_opt::CovOpt, mu_opt::MuOpt,
              method::Symbol = :Stationary, n_sim::Integer = 3_000, block_size::Integer = 3,
              seed::Union{<:Integer, Nothing} = nothing)

Simulate returns series using bootstrapping with arch.

Inputs

returns: T×N matrix of returns, where T is the number of returns observations, and N the number of assets.

cov_opt: instance of CovOpt, defines the parameters for computing the covariance matrices of the bootstrapping simulations, as well as fixing non-positive definite matrices.

mu_opt: instance of MuOpt, defines the parameters for computing the expected returns vectors of the bootstrapping simulations.

kind: bootstrapping method from BootstrapMethods.

n_sim: number of simulations.

block_size: average block size to use.

seed: random number generator seed for bootstrapping.

PortfolioOptimiser.gerber0 — Function

gerber0(x::AbstractMatrix, opt::GerberOpt = GerberOpt(;))

PortfolioOptimiser.gerber1 — Function

gerber1(x::AbstractMatrix, opt::GerberOpt = GerberOpt(;))

PortfolioOptimiser.gerber2 — Function

gerber2(x::AbstractMatrix, opt::GerberOpt = GerberOpt(;))

PortfolioOptimiser.gerbersb0 — Function

gerbersb0(x::AbstractMatrix, gerberopt::GerberOpt = GerberOpt(;), sbopt::SBOpt = SBOpt(;))

PortfolioOptimiser.gerbersb1 — Function

gerbersb1(x::AbstractMatrix, gerberopt::GerberOpt = GerberOpt(;), sbopt::SBOpt = SBOpt

PortfolioOptimiser.mean_vec — Function

mean_vec(returns::AbstractMatrix, opt::MuOpt = MuOpt(;))

Compute the expected returns vector for a returns series.

Inputs

returns: T×N matrix of returns, where T is the number of returns observations, and N is the number of assets or factors.

mu_opt: instance of MuOpt, defines the parameters for computing the expected returns vector.

Outputs

mu: N×1 vector of expected returns, where N is the number of assets or factors.

PortfolioOptimiser.mu_estimator — Function

mu_estimator

PortfolioOptimiser.mut_var_info_mtx — Function

mut_var_info_mtx(x::AbstractMatrix{<:Real}, bins_info::Union{Symbol, <:Integer} = :KN,
                 normalise::Bool = true)

Compute the mutual information and variation of information matrices.

PortfolioOptimiser.mv_regression — Function

mv_regression(x::DataFrame, y::DataFrame, opt::MVROpt = MVROpt(;))

Select the factors that best estimate the model based on multivariate regression. This tends to be more robust than backward_regression.

Inputs

x: is the T×N Dataframe of factor returns, where the column names are the factor names, T is the number of returns observations, and N the number of factors.

y: is the T×N Dataframe of asset returns.

opt: instance of MVROpt for defining the regression and its parameters.

Outputs

B: is the (Na+2)×Nf loadings matrix as a Dataframe, where Na is the number of assets, Nf the number of factors. The two extra columns are the optional columns for B as described in FactorOpt.

PortfolioOptimiser.posdef_fix! — Function

posdef_fix!(mtx::AbstractMatrix, opt::PosdefFixOpt = PosdefFixOpt(;), msg::String = "")

PortfolioOptimiser.risk_factors — Method

risk_factors(x::DataFrame, y::DataFrame; factor_opt::FactorOpt = FactorOpt(;),
             cov_opt::CovOpt = CovOpt(;), mu_opt::MuOpt = MuOpt(;))

Estimates the returns matrix, expected returns vector, and covariance matrix of the assets based on the risk factor model [58, 59]. See loadings_matrix, covar_mtx_mean_vec, and posdef_fix!.

\[\begin{align*} \mathbf{X}_{\mathrm{FM}} &= \mathbf{F} \mathbf{B}^{\intercal}\\ \bm{\mu}_{\mathrm{FM}} &= \mathbf{B} \bm{\mu}_{F}\\ \bm{\Sigma}_{\mathrm{FM}} &= \mathbf{B} \mathbf{\Sigma}_{F} \mathbf{B}^{\intercal} + \mathbf{\Sigma}_{\epsilon}\\ \mathbf{\Sigma}_{\epsilon} &= \begin{cases}\mathrm{Diagonal}\left(\mathrm{var}\left(\mathbf{X} - \mathbf{X}_{\mathrm{FM}},\, \mathrm{dims} = 1\right)\right) &\quad \mathrm{if~ error = true}\\ \mathbf{0} &\quad \mathrm{if~ error = false} \end{cases} \end{align*}\]

Where:

$\mathbf{X}_{\mathrm{FM}}$: is the T×Na matrix of factor adjusted asset returns, where T is the number of returns observations, and Na the number of assets.

$\mathbf{F}$: is the T×Nf matrix of factor returns, where T is the number of returns observations, and Nf the number of assets.

$\mathbf{B}$: is the Na×Nf loadings matrix, where Na is the number of assets, and Nf the number of factors.

$\bm{\mu}_{\mathrm{FM}}$: is the Na×1 estimated expected asset returns vector computed using the factor model, where Na is the number of assets.

$\bm{\mu}_{F}$: is the Nf×1 expected factor returns vector, where Nf is the number of factors.

$\mathbf{\Sigma}_{\epsilon}$: is an Na×Na diagonal matrix constructed from the variances of the errors between the asset and estimated asset returns using the factor model, where Na is the number of assets. The variance is taken over all T timestamps of Na assets.

$\mathbf{\Sigma}_{\mathrm{FM}}$: is the Na×Na estimated asset covariance matrix computed using the factor model, where Na is the number of assets.

Inputs

x: is the T×Nf Dataframe of factor returns, where the column names are the factor names, T is the number of returns observations, and Nf the number of factors.

y: is the T×Na Dataframe of asset returns, where the column names are the factor names, T is the number of returns observations, and Na the number of assets.

Options

factor_opt: instance of FactorOpt, defines the parameters for computing the factors and loadings matrix.

isnothing(factor_opt.B): the loadings matris is computed internally using factor_opt.loadings_opt.

cov_opt: instance of CovOpt, defines the parameters for computing the factor covariance matrix, as well as fixing non-positive definite matrices.

If the loadings matrix contains the const column cov_opt.posdef.method is set to :None before computing the factor covariance matrix because the constant term makes it impossible to turn it into a positive definite matrix. cov_opt.posdef.method is then reset to its original value so it can be used to fix the estimated asset covariance matrix if needed.

mu_opt: instance of MuOpt, defines the parameters for computing the expected factor returns vector.

Outputs

mu: is the Na×1 estimated expected asset returns vector computed using the factor model, where Na is the number of assets.

sigma: is the Na×Na estimated asset covariance matrix computed using the factor model, where Na is the number of assets.

returns: is the T×Na matrix of factor adjusted asset returns, where T is the number of returns observations, and Na the number of assets.

B: is the (Na+c)×Nf loadings matrix as a Dataframe, where Na is the number of assets, Nf the number of factors, and c ∈ (0, 1, 2) represents the two optional columns as described in FactorOpt.

PortfolioOptimiser.sb0 — Function

sb0(x::AbstractMatrix, gerberopt::GerberOpt = GerberOpt(;), sbopt::SBOpt = SBOpt(;))

PortfolioOptimiser.sb1 — Function

sb1(x::AbstractMatrix, gerberopt::GerberOpt = GerberOpt(;), sbopt::SBOpt = SBOpt(;))

PortfolioOptimiser.stepwise_regression — Function

stepwise_regression(x::DataFrame, y::DataFrame, opt::LoadingsOpt = LoadingsOpt(;))

Perform forward or backward stepwise regression. See forward_regression, and backward_regression.

Inputs

x: is the T×Nf Dataframe of factor returns, where the column names are the factor names, T is the number of returns observations, and Nf the number of factors.

y: is the T×Na Dataframe of asset returns, where the column names are the factor names, T is the number of returns observations, and Na the number of assets.

Options

loadings_opt: instance of LoadingsOpt, defines the parameters for computing the loadings matrix.

Outputs

B: is the (Na+2)×Nf loadings matrix as a Dataframe, where Na is the number of assets, Nf the number of factors. The two extra columns are the optional columns for B as described in FactorOpt.

PortfolioOptimiser.vec_of_vecs_to_mtx — Method

vec_of_vecs_to_mtx(x::AbstractVector{<:AbstractArray})

Turns a vector of arrays into a matrix.

Inputs

x: vector of arrays.

Outputs

y: matrix.