Asset Statistics

Public API

PortfolioOptimiser.asset_statistics! — Function

asset_statistics!(
    portfolio::AbstractPortfolio;
    target_ret::AbstractFloat = 0.0,
    mean_func::Function = mean,
    cov_func::Function = cov,
    cor_func::Function = cor,
    std_func = std,
    dist_func::Function = x -> sqrt.(clamp!((1 .- x) / 2, 0, 1)),
    codep_type::Symbol = isa(portfolio, HCPortfolio) ? portfolio.codep_type : :Pearson,
    custom_mu = nothing,
    custom_cov = nothing,
    custom_kurt = nothing,
    custom_skurt = nothing,
    mean_args::Tuple = (),
    cov_args::Tuple = (),
    cor_args::Tuple = (),
    dist_args::Tuple = (),
    std_args::Tuple = (),
    calc_kurt = true,
    mean_kwargs = (; dims = 1),
    cov_kwargs::NamedTuple = (;),
    cor_kwargs::NamedTuple = (;),
    dist_kwargs::NamedTuple = (;),
    std_kwargs::NamedTuple = (;),
    uplo = :L,
)

PortfolioOptimiser.wc_statistics! — Function

wc_statistics!(
    portfolio;
    box = :Stationary,
    ellipse = :Stationary,
    calc_box = true,
    calc_ellipse = true,
    q = 0.05,
    n_sim = 3_000,
    window = 3,
    dmu = 0.1,
    dcov = 0.1,
    n_samples = 10_000,
    seed = nothing,
    rng = Random.default_rng(),
    fix_cov_args::Tuple = (),
    fix_cov_kwargs::NamedTuple = (;),
)

Worst case optimisation statistics.

PortfolioOptimiser.factor_statistics! — Function

factor_statistics!(
    portfolio::AbstractPortfolio;
    # cov_mtx
    alpha::Real = 0.0,
    cov_args::Tuple = (),
    cov_est::CovarianceEstimator = StatsBase.SimpleCovariance(; corrected = true),
    cov_func::Function = cov,
    cov_kwargs::NamedTuple = (;),
    cov_type::Symbol = :Full,
    cov_weights::Union{AbstractWeights, Nothing} = nothing,
    custom_cov::Union{AbstractMatrix, Nothing} = nothing,
    denoise::Bool = false,
    detone::Bool = false,
    gs_threshold::Real = portfolio.gs_threshold,
    jlogo::Bool = false,
    kernel = ASH.Kernels.gaussian,
    m::Integer = 10,
    method::Symbol = :Fixed,
    mkt_comp::Integer = 0,
    n::Integer = 1000,
    opt_args = (),
    opt_kwargs = (;),
    posdef_args::Tuple = (),
    posdef_fix::Symbol = :Nearest,
    posdef_func::Function = x -> x,
    posdef_kwargs::NamedTuple = (;),
    std_args::Tuple = (),
    std_func::Function = std,
    std_kwargs::NamedTuple = (;),
    target_ret::Union{Real, AbstractVector{<:Real}} = 0.0,
    # mean_vec
    custom_mu::Union{AbstractVector, Nothing} = nothing,
    mean_args::Tuple = (),
    mean_func::Function = mean,
    mean_kwargs::NamedTuple = (;),
    mkt_ret::Union{AbstractVector, Nothing} = nothing,
    mu_target::Symbol = :GM,
    mu_type::Symbol = :Default,
    mu_weights::Union{AbstractWeights, Nothing} = nothing,
    rf = 0.0,
    # Loadings matrix
    B::Union{DataFrame, Nothing} = nothing,
    criterion::Union{Symbol, Function} = :pval,
    error::Bool = true,
    pca_kwargs::NamedTuple = (;),
    pca_std_kwargs::NamedTuple = (;),
    pca_std_type = ZScoreTransform,
    reg_type::Symbol = :FReg,
    threshold::Real = 0.05,
    var_func::Function = var,
    var_args::Tuple = (),
    var_kwargs::NamedTuple = (;),
)

PortfolioOptimiser.black_litterman_statistics! — Function

black_litterman_statistics!(
    portfolio::AbstractPortfolio,
    P::AbstractMatrix,
    Q::AbstractVector,
    w::AbstractVector = Vector{Float64}(undef, 0);
    # cov_mtx
    cov_args::Tuple = (),
    cov_est::CovarianceEstimator = StatsBase.SimpleCovariance(; corrected = true),
    cov_func::Function = cov,
    cov_kwargs::NamedTuple = (;),
    cov_type::Symbol = :Full,
    cov_weights::Union{AbstractWeights, Nothing} = nothing,
    custom_cov::Union{AbstractMatrix, Nothing} = nothing,
    gs_threshold::Real = portfolio.gs_threshold,
    jlogo::Bool = false,
    posdef_args::Tuple = (),
    posdef_fix::Symbol = :Nearest,
    posdef_func::Function = x -> x,
    posdef_kwargs::NamedTuple = (;),
    std_args::Tuple = (),
    std_func::Function = std,
    std_kwargs::NamedTuple = (;),
    target_ret::Union{Real, AbstractVector{<:Real}} = 0.0,
    # mean_vec
    custom_mu::Union{AbstractVector, Nothing} = nothing,
    mean_args::Tuple = (),
    mean_func::Function = mean,
    mean_kwargs::NamedTuple = (;),
    mkt_ret::Union{AbstractVector, Nothing} = nothing,
    mu_target::Symbol = :GM,
    mu_type::Symbol = :Default,
    mu_weights::Union{AbstractWeights, Nothing} = nothing,
    # Black Litterman
    delta::Union{Real, Nothing} = nothing,
    eq::Bool = true,
    rf::Real = 0.0,
)

PortfolioOptimiser.black_litterman_factor_satistics! — Function

black_litterman_factor_satistics!(
    portfolio::AbstractPortfolio,
    w::AbstractVector = Vector{Float64}(undef, 0);
    # cov_mtx
    cov_args::Tuple = (),
    cov_est::CovarianceEstimator = StatsBase.SimpleCovariance(; corrected = true),
    cov_func::Function = cov,
    cov_kwargs::NamedTuple = (;),
    cov_type::Symbol = :Full,
    cov_weights::Union{AbstractWeights, Nothing} = nothing,
    custom_cov::Union{AbstractMatrix, Nothing} = nothing,
    gs_threshold::Real = portfolio.gs_threshold,
    jlogo::Bool = false,
    posdef_args::Tuple = (),
    posdef_fix::Symbol = :Nearest,
    posdef_func::Function = x -> x,
    posdef_kwargs::NamedTuple = (;),
    std_args::Tuple = (),
    std_func::Function = std,
    std_kwargs::NamedTuple = (;),
    target_ret::Union{Real, AbstractVector{<:Real}} = 0.0,
    # mean_vec
    custom_mu::Union{AbstractVector, Nothing} = nothing,
    mean_args::Tuple = (),
    mean_func::Function = mean,
    mean_kwargs::NamedTuple = (;),
    mkt_ret::Union{AbstractVector, Nothing} = nothing,
    mu_target::Symbol = :GM,
    mu_type::Symbol = :Default,
    mu_weights::Union{AbstractWeights, Nothing} = nothing,
    # Black Litterman
    B::Union{DataFrame, Nothing} = nothing,
    P::Union{AbstractMatrix, Nothing} = nothing,
    P_f::Union{AbstractMatrix, Nothing} = nothing,
    Q::Union{AbstractVector, Nothing} = nothing,
    Q_f::Union{AbstractVector, Nothing} = nothing,
    bl_type::Symbol = :B,
    delta::Real = 1.0,
    diagonal::Bool = true,
    eq::Bool = true,
    rf::Real = 0.0,
    var_args::Tuple = (),
    var_func::Function = var,
    var_kwargs::NamedTuple = (;),
    # Loadings matrix
    criterion::Union{Symbol, Function} = :pval,
    pca_kwargs::NamedTuple = (;),
    pca_std_kwargs::NamedTuple = (;),
    pca_std_type = ZScoreTransform,
    reg_type::Symbol = :FReg,
    threshold::Real = 0.05,
)

PortfolioOptimiser.mean_vec — Function

mean_vec(
    returns::AbstractMatrix;
    custom_mu::Union{AbstractVector, Nothing} = nothing,
    mean_args::Tuple = (),
    mean_func::Function = mean,
    mean_kwargs::NamedTuple = (;),
    mu_target::Symbol = :GM,
    mu_type::Symbol = :Default,
    mu_weights::Union{AbstractWeights, Nothing} = nothing,
    rf::Real = 0.0,
    sigma::Union{AbstractMatrix, Nothing} = nothing,
)

PortfolioOptimiser.mu_estimator — Function

mu_estimator

PortfolioOptimiser.covar_mtx — Function

covar_mtx

PortfolioOptimiser.mut_var_info_mtx — Function

mut_var_info_mtx(
    x::AbstractMatrix{<:Real},
    bins_info::Union{Symbol, <:Integer} = :KN,
    normed::Bool = true,
)

Compute the mutual information and variation of information matrices.

Inputs

x: T×N array containing the returns series of the assets. T is the number of observations and N the number of assets.

bins_info: selection criterion for computing the number of bins used to calculate the mutual and variation of information statistics. Can take on an integer value or the following values:

An integer value explicitly defines the number of bins.

A choice of optimal bin width selection algorithms from BinTypes.

:KN: Knuth's choice.

:FD: Freedman-Diaconis' choice.

:SC: Schotts' choice.

:HGR: Hacine-Gharbi and Ravier's choice.

PortfolioOptimiser.cokurt_mtx — Function

cokurt_mtx

Private API

PortfolioOptimiser.BinTypes — Constant

BinTypes = (:KN, :FD, :SC, :HGR)

Methods for calculating optimal bin widths for the mutual and variational information matrices computed by mut_var_info_mtx.

:KN: Knuth's choice.

:FD: Freedman-Diaconis' choice.

:SC: Schotts' choice.

:HGR: Hacine-Gharbi and Ravier's choice.