Asset Statistics

Public API

PortfolioOptimiser.asset_statistics! — Function

asset_statistics!(portfolio::AbstractPortfolio; target_ret::AbstractFloat = 0.0,
                  mean_func::Function = mean, cov_func::Function = cov,
                  cor_func::Function = cor, std_func = std,
                  dist_func::Function = x -> sqrt.(clamp!((1 .- x) / 2, 0, 1)),
                  cor_method::Symbol = isa(portfolio, HCPortfolio) ? portfolio.cor_method :                           # flags
                                       :Pearson, custom_mu = nothing, custom_cov = nothing,
                  custom_kurt = nothing, custom_skurt = nothing, mean_args::Tuple = (),                           # cov_mtx
                  cov_args::Tuple = (), cor_args::Tuple = (), dist_args::Tuple = (),
                  std_args::Tuple = (), calc_kurt = true, mean_kwargs = (; dims = 1),
                  cov_kwargs::NamedTuple = (;), cor_kwargs::NamedTuple = (;),
                  dist_kwargs::NamedTuple = (;), std_kwargs::NamedTuple = (;), uplo = :L,)

PortfolioOptimiser.wc_statistics! — Function

wc_statistics!(portfolio; box = :Stationary, ellipse = :Stationary, calc_box = true,
               calc_ellipse = true, q = 0.05, n_sim = 3_000, window = 3, dmu = 0.1,
               dcov = 0.1, seed = nothing, rng = Random.default_rng(),
               fix_cov_args::Tuple = (), fix_cov_kwargs::NamedTuple = (;),)

Worst case optimisation statistics.

PortfolioOptimiser.factor_statistics! — Function

factor_statistics!(portfolio::AbstractPortfolio;    # cov_mtx
                   alpha::Real = 0.0, cov_args::Tuple = (),
                   cov_est::CovarianceEstimator = StatsBase.SimpleCovariance(;
                                                                             corrected = true),
                   cov_func::Function = cov, cov_kwargs::NamedTuple = (;),
                   cov_method::Symbol = :Full,
                   cov_weights::Union{AbstractWeights, Nothing} = nothing,
                   custom_cov::Union{AbstractMatrix, Nothing} = nothing,
                   denoise::Bool = false, detone::Bool = false,
                   gs_threshold::Real = portfolio.gs_threshold, jlogo::Bool = false,
                   kernel = ASH.Kernels.gaussian, m::Integer = 10, method::Symbol = :Fixed,
                   mkt_comp::Integer = 1, n::Integer = 1000, opt_args = (),
                   opt_kwargs = (;), posdef_args::Tuple = (), posdef_fix::Symbol = :Nearest,
                   posdef_func::Function = x -> x, posdef_kwargs::NamedTuple = (;),
                   std_args::Tuple = (), std_func::Function = std,
                   std_kwargs::NamedTuple = (;),
                   target_ret::Union{Real, AbstractVector{<:Real}} = 0.0,    # mean_vec
                   custom_mu::Union{AbstractVector, Nothing} = nothing,
                   mean_args::Tuple = (), mean_func::Function = mean,
                   mean_kwargs::NamedTuple = (;),
                   mkt_ret::Union{AbstractVector, Nothing} = nothing,
                   mu_target::Symbol = :GM, mu_method::Symbol = :Default,
                   mu_weights::Union{AbstractWeights, Nothing} = nothing, rf = 0.0,    # Loadings matrix
                   B::Union{DataFrame, Nothing} = nothing,
                   criterion::Union{Symbol, Function} = :pval, error::Bool = true,
                   pca_kwargs::NamedTuple = (;), pca_std_kwargs::NamedTuple = (;),
                   pca_std_type = ZScoreTransform, reg_method::Symbol = :FReg,
                   threshold::Real = 0.05, var_func::Function = var, var_args::Tuple = (),
                   var_kwargs::NamedTuple = (;),)

PortfolioOptimiser.black_litterman_statistics! — Function

black_litterman_statistics!(portfolio::AbstractPortfolio, P::AbstractMatrix,
                            Q::AbstractVector;
                            w::AbstractVector = Vector{Float64}(undef, 0),    # cov_mtx
                            cov_args::Tuple = (),
                            cov_est::CovarianceEstimator = StatsBase.SimpleCovariance(;
                                                                                      corrected = true),
                            cov_func::Function = cov, cov_kwargs::NamedTuple = (;),
                            cov_method::Symbol = :Full,
                            cov_weights::Union{AbstractWeights, Nothing} = nothing,
                            custom_cov::Union{AbstractMatrix, Nothing} = nothing,
                            gs_threshold::Real = portfolio.gs_threshold,
                            jlogo::Bool = false, posdef_args::Tuple = (),
                            posdef_fix::Symbol = :Nearest, posdef_func::Function = x -> x,
                            posdef_kwargs::NamedTuple = (;), std_args::Tuple = (),
                            std_func::Function = std, std_kwargs::NamedTuple = (;),
                            target_ret::Union{Real, AbstractVector{<:Real}} = 0.0,    # mean_vec
                            custom_mu::Union{AbstractVector, Nothing} = nothing,
                            mean_args::Tuple = (), mean_func::Function = mean,
                            mean_kwargs::NamedTuple = (;),
                            mkt_ret::Union{AbstractVector, Nothing} = nothing,
                            mu_target::Symbol = :GM, mu_method::Symbol = :Default,
                            mu_weights::Union{AbstractWeights, Nothing} = nothing,    # Black Litterman
                            delta::Union{Real, Nothing} = nothing, eq::Bool = true,
                            rf::Real = 0.0,)

PortfolioOptimiser.black_litterman_factor_satistics! — Function

black_litterman_factor_satistics!(portfolio::AbstractPortfolio;
                                  w::AbstractVector = Vector{Float64}(undef, 0),    # cov_mtx
                                  cov_args::Tuple = (),
                                  cov_est::CovarianceEstimator = StatsBase.SimpleCovariance(;
                                                                                            corrected = true),
                                  cov_func::Function = cov, cov_kwargs::NamedTuple = (;),
                                  cov_method::Symbol = :Full,
                                  cov_weights::Union{AbstractWeights, Nothing} = nothing,
                                  custom_cov::Union{AbstractMatrix, Nothing} = nothing,
                                  gs_threshold::Real = portfolio.gs_threshold,
                                  jlogo::Bool = false, posdef_args::Tuple = (),                                           # Black Litterman
                                  posdef_fix::Symbol = :Nearest,
                                  posdef_func::Function = x -> x,
                                  posdef_kwargs::NamedTuple = (;), std_args::Tuple = (),
                                  std_func::Function = std, std_kwargs::NamedTuple = (;),
                                  target_ret::Union{Real, AbstractVector{<:Real}} = 0.0,    # mean_vec                                           # Settings
                                  custom_mu::Union{AbstractVector, Nothing} = nothing,
                                  mean_args::Tuple = (), mean_func::Function = mean,
                                  mean_kwargs::NamedTuple = (;),
                                  mkt_ret::Union{AbstractVector, Nothing} = nothing,
                                  mu_target::Symbol = :GM, mu_method::Symbol = :Default,
                                  mu_weights::Union{AbstractWeights, Nothing} = nothing,    # Black Litterman
                                  B::Union{DataFrame, Nothing} = nothing,
                                  P::Union{AbstractMatrix, Nothing} = nothing,
                                  P_f::Union{AbstractMatrix, Nothing} = nothing,
                                  Q::Union{AbstractVector, Nothing} = nothing,
                                  Q_f::Union{AbstractVector, Nothing} = nothing,
                                  bl_method::Symbol = :B, delta::Real = 1.0,
                                  diagonal::Bool = true, eq::Bool = true, rf::Real = 0.0,
                                  var_args::Tuple = (), var_func::Function = var,
                                  var_kwargs::NamedTuple = (;),    # Loadings matrix
                                  criterion::Union{Symbol, Function} = :pval,
                                  pca_kwargs::NamedTuple = (;),
                                  pca_std_kwargs::NamedTuple = (;),
                                  pca_std_type = ZScoreTransform,
                                  reg_method::Symbol = :FReg, threshold::Real = 0.05,)

PortfolioOptimiser.mean_vec — Function

mean_vec(returns::AbstractMatrix; custom_mu::Union{AbstractVector, Nothing} = nothing,
         mean_args::Tuple = (), mean_func::Function = mean, mean_kwargs::NamedTuple = (;),
         mu_target::Symbol = :GM, mu_method::Symbol = :Default,
         mu_weights::Union{AbstractWeights, Nothing} = nothing, rf::Real = 0.0,
         sigma::Union{AbstractMatrix, Nothing} = nothing,)

PortfolioOptimiser.mu_estimator — Function

mu_estimator

PortfolioOptimiser.covar_mtx — Function

covar_mtx

PortfolioOptimiser.mut_var_info_mtx — Function

mut_var_info_mtx(x::AbstractMatrix{<:Real}; bins_info::Union{Symbol, <:Integer} = :KN,
                 normed::Bool = true)

Compute the mutual information and variation of information matrices.

Inputs

x: T×N array containing the returns series of the assets. T is the number of observations and N the number of assets.

bins_info: selection criterion for computing the number of bins used to calculate the mutual and variation of information statistics. Can take on an integer value or the following values:

An integer value explicitly defines the number of bins.

A choice of optimal bin width selection algorithms from BinMethods.

:KN: Knuth's choice.

:FD: Freedman-Diaconis' choice.

:SC: Schotts' choice.

:HGR: Hacine-Gharbi and Ravier's choice.

PortfolioOptimiser.cokurt_mtx — Function

cokurt_mtx

PortfolioOptimiser.cor_dist_mtx — Function

cor_dist_mtx(
    returns::AbstractMatrix;    alpha_tail::Real = 0.05,    bins_info::Union{Symbol, Integer} = :KN,    cor_method::Symbol = :Pearson,    cor_args::Tuple = (),    cor_func::Function = cor,    cor_kwargs::NamedTuple = (;),    custom_cor::Union{AbstractMatrix, Nothing} = nothing,    dist_args::Tuple = (),    dist_func::Function = x -> sqrt.(clamp!((1 .- x) / 2, 0, 1)),    dist_kwargs::NamedTuple = (;),    gs_threshold::Real = 0.5,    posdef_args::Tuple = (),    posdef_fix::Symbol = :Nearest,    posdef_func::Function = x -> x,    posdef_kwargs::NamedTuple = (;),    sigma::Union{AbstractMatrix, Nothing} = nothing,    std_args::Tuple = (),    std_func::Function = std,    std_kwargs::NamedTuple = (;),    uplo::Symbol = :L,)

Private API

PortfolioOptimiser.BinMethods — Constant

BinMethods = (:KN, :FD, :SC, :HGR)

Methods for calculating optimal bin widths for the mutual and variational information matrices computed by mut_var_info_mtx.

:KN: Knuth's choice.

:FD: Freedman-Diaconis' choice.

:SC: Schotts' choice.

:HGR: Hacine-Gharbi and Ravier's choice.

PortfolioOptimiser.BLFMMethods — Constant

BLFMMethods = (:A, :B)

Versions of the factor Black-Litterman Model.

:B: Bayesian Black-Litterman, which uses the factors to generate the Black-Litterman estimates.

:A: Augmented Black-Litterman, which uses the factors to adjust the Black-Litterman views.

PortfolioOptimiser.EllipseMethods — Constant

EllipseMethods = (:Stationary, :Circular, :Moving, :Normal)

Available types of elliptical sets that can be computed with wc_statistics!, which are used by Worst Case Mean Variance Optimisations (see PortTypes).

:Stationary: stationary bootstrapping method.

:Circular: circular block bootstrapping method.

:Moving: moving block bootstrapping method.

:Normal: sets generated by assuming returns are normally distributed.

PortfolioOptimiser.BoxMethods — Constant

BoxMethods = (:Stationary, :Circular, :Moving, :Normal, :Delta)

Available types of box sets that can be computed with wc_statistics!, which are used by Worst Case Mean Variance Optimisations (see PortTypes).

:Stationary: stationary bootstrapping method.

:Circular: circular block bootstrapping method.

:Moving: moving block bootstrapping method.

:Normal: normally distributed covariance and mean samples.

:Delta: box sets are assumed to fall in the extrema of a well-defined interval.

PortfolioOptimiser.BootstrapMethods — Constant

BootstrapMethods = (:Stationary, :Circular, :Moving)

Kind of bootstrap for computing the uncertainty sets with wc_statistics!, which are used by Worst Case Mean Variance Optimisations (see PortTypes).

:Stationary: stationary bootstrapping method.

:Circular: circular block bootstrapping method.

:Moving: moving block bootstrapping method.

PortfolioOptimiser.MuMethods — Constant

MuMethods = (:Default, :JS, :BS, :BOP, :CAPM, :Custom_Func, :Custom_Val)

Methods for estimating the mean returns vector $\bm{\mu}$ in mean_vec.

:Default: is the standard historical.

:JS: James-Stein.

:BS: Bayes-Stein.

:BOP: Bodnar-Okhrin-Parolya.

:CAPM: Capital Asset Pricing Model.

:Custom_Func: custom function provided.

:Custom_Val: custom value provided.

PortfolioOptimiser.MuTargets — Constant

MuTargets = (:GM, :VW, :SE)

Targets for the :JS, :BS and :BOP estimators in mean_vec and mu_estimator.

:GM: grand mean.

:VW: volatility-weighted grand mean.

:SE: mean square error of sample mean.

PortfolioOptimiser.CovMethods — Constant

CovMethods = (:Full, :Semi, :Gerber0, :Gerber1, :Gerber2, :Custom_Func, :Custom_Val)

Methods for estimating the covariance matrix $\mathbf{\Sigma}$.

:Full: full covariance matrix.

:Semi: semi-covariance matrix.

:Gerber0: Gerber statistic 0.

:Gerber1: Gerber statistic 1.

:Gerber2: Gerber statistic 2.

:Custom_Func: custom function provided.

:Custom_Val: custom value provided.

PortfolioOptimiser.PosdefFixMethods — Constant

PosdefFixMethods = (:None, :Nearest, :Custom_Func)

Methods for fixing non-positive definite matrices.

:None: no fix is applied.

:Nearest: nearest correlation matrix.

Custom_Func: custom function provided.

PortfolioOptimiser.DenoiseMethods — Constant

DenoiseMethods = (:Fixed, :Spectral, :Shrink)

Methods for matrix denoising.

:Fixed: fixed.

:Spectral: spectral.

:Shrink: shrink.

PortfolioOptimiser.RegCriteria — Constant

RegCriteria = (:pval, GLM.aic, GLM.aicc, GLM.bic, GLM.r2, GLM.adjr2)

Criteria for feature selection in regression functions.

:pval: p-value feature selection.

The rest are methods applied to a fitted General Linear Model from GLM.jl.

PortfolioOptimiser.FSMethods — Constant

FSMethods = (:FReg, :BReg, :PCR)

Methods for feature selection when creating the loadings matrix.

:FReg: forward regression;- :Breg: backward regression;- :PCR: Principal Component Regression using PCA.

PortfolioOptimiser.CorMethods — Constant

CorMethods = (:Pearson, :Spearman, :Kendall, :Gerber0, :Gerber1, :Gerber2, :Abs_Pearson,
              :Abs_Spearman, :Abs_Kendall, :Distance, :Mutual_Info, :Tail, :Cov_to_Cor,
              :Custom_Func, :Custom_Val)

Methods for estimating the codependence (correlation) matrix $\mathbf{P}$, and distance matrix $\mathbf{D}$.

:Pearson: Pearson correlation, $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{P}_{i,\,j} \right)}$.

:Spearman: Spearman correlation, $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{P}_{i,\,j} \right)}$.

:Kendall: Kendall correlation, $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{P}_{i,\,j} \right)}$.

:Gerber0: Gerber statistic 0, $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{P}_{i,\,j} \right)}$.

:Gerber1: Gerber statistic 1, $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{P}_{i,\,j} \right)}$.

:Gerber2: Gerber statistic 2, $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{P}_{i,\,j} \right)}$.

:Abs_Pearson: absolute value of the Pearson correlation, $\mathbf{D}_{i,\,j} = \sqrt{1 - \left\lvert\mathbf{P}_{i,\,j}\right\rvert}$.

:Abs_Spearman: absolute value of the Spearman correlation, $\mathbf{D}_{i,\,j} = \sqrt{1 - \left\lvert\mathbf{P}_{i,\,j}\right\rvert}$.

:Abs_Kendall: absolute value of the Kendall correlation, $\mathbf{D}_{i,\,j} = \sqrt{1 - \left\lvert\mathbf{P}_{i,\,j}\right\rvert}$.

:Distance: distance correlation matrix, $\mathbf{D}_{i,\,j} = \sqrt{1 - \mathbf{P}_{i,\,j}}$

:Mutual_Info: mutual information matrix, $\mathbf{D}_{i,\,j}$ is the variation information matrix.

:Tail: lower tail dependence index matrix, $\mathbf{D}_{i,\,j} = -\log\left(\mathbf{P}_{i,\,j}\right)$

:Cov_to_Cor: the covariance matrix is converted to a correlation matrix, the distance matrix is computed by a distance function which defaults to $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{P}_{i,\,j} \right)}$.

:Custom_Func: custom function provided, the distance matrix is computed by a distance function which defaults to $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{P}_{i,\,j} \right)}$.

:Custom_Val: custom value provided, the distance matrix is computed by a distance function which defaults to $\mathbf{D}_{i,\,j} = \sqrt{\dfrac{1}{2} \left(1- \mathbf{P}_{i,\,j} \right)}$.

PortfolioOptimiser.RRPVersions — Constant

RRPVersions = (:None, :Reg, :Reg_Pen)

Available versions of Relaxed Risk Parity Optimisations (see PortTypes).

:None: no penalty.

:Reg: regularisation constraint, $\rho$.

:Reg_Pen: regularisation and penalisation constraints, $\lambda$ and $\rho$.

PortfolioOptimiser.UncertaintyTypes — Constant

UncertaintyTypes = (:None, :Box, :Ellipse)

Available types of uncertainty sets that can be computed with wc_statistics!, which are used by Worst Case Mean Variance Optimisations (see PortTypes, EllipseMethods, and BoxMethods).

:Box: are box uncertainty sets, ie the sets are full matrices.

:Ellipse: are elliptical uncertainty sets, ie the sets are diagonal matrices.