API

Objective Functions

PortfolioOptimiser.port_variance — Function

port_variance(w, cov_mtx)

Computes the overall portfolio variance according to the weights w, and covariance matrix cov_mtx.

The portfolio variance $\sigma^2$, is defined as:

\[\sigma^2 = \bm{w}^T\, \Sigma\, \bm{w}\,,\]

where $\sigma$ is the overall portfolio standard deviation, $^T$ the transpose operator, $w$ the weights, and $\Sigma$ the covariance matrix.

PortfolioOptimiser.port_return — Function

port_return(w, mean_ret)

Computes the overall portfolio return according to the weights w, and mean returns mean_ret.

The portfolio return $\mu$, is defined as:

\[\mu = \bm{w} \cdot \bm{E},\]

where $\bm{w}$ are the weights, and $\bm{E}$ the mean returns.

PortfolioOptimiser.sharpe_ratio — Function

sharpe_ratio(w, mean_ret, cov_mtx, rf::Real = 0.02)

Computes the portfolio's Sharpe ratio given the weights w, mean returns mean_ret, covariance matrix cov_mtx, and risk free rate rf.

sharpe_ratio(μ, σ, rf::Real = 0.02)

Computes the Sharpe ratio given the return μ, standard deviation σ, and risk free rate rf.

The Sharpe ratio $s$, is defined as:

\[s = \dfrac{\mu - r}{\sigma},\]

where $\mu$ the portfolio's return (see port_return), and $\sigma$ the portfolio's standard deviation (see port_variance). Generally speaking, the greater the Sharpe ratio the better the portfolio.

Note

The Sharpe ratio penalises large swings in both directions, so assets that tend to have large increases in value are disproportionally penalised by this measure. The Sortino ratio has the same formula but uses an adjusted covariance matrix that accounts only for the negative fluctuations in value. The semicovariance is implemented by risk_matrix when given SCov() or ESCov() as its first argument. The Mean-Semivariance optimisations EfficientSemiVar make the adjustment too.

PortfolioOptimiser.L2_reg — Function

L2_reg(w, γ = 1)

L2 regularisation. Minimising this reduces the number of negligible weights w, with a tuning parameter γ. Increasing γ will decrease the number of negligible weights.

The $L_2$ regularisation is defined as:

\[L_2 = \gamma \Vert \bm{w} \Vert\,,\]

where $\gamma$ is the tuning parameter, and $w$ the weights.

PortfolioOptimiser.quadratic_utility — Function

quadratic_utility(w, mean_ret, cov_mtx, risk_aversion = 1)

Calculates the quadratic utility given the weights w, mean returns mean_ret, covariance matrix cov_mtx and risk aversion risk_aversion. Increasing risk_aversion decreases risk.

The quadratic utility $Q$, is defined as:

\[Q = \mu - \dfrac{1}{2} \delta \bm{w}^T\, \Sigma\, \bm{w}\,,\]

where $\mu$ is the overall portfolio expected return (see port_return), $\delta$ the risk aversion, $w$ the weights, and $\Sigma$ the covariance matrix.

PortfolioOptimiser.semi_ret — Function

semi_ret(returns, benchmark = 0)

Compute the historical semi-returns from historical returns, for a given benchmark. Return values greater than benchmark are "upside" returns, values lower than benchmark are "downside" returns. Each column of returns corresponds to an asset.

Historical semi returns $\mathrm{R_b}$, are defined as:

\[\mathrm{R_b} = \dfrac{\mathrm{R} - b}{\sqrt{N}}\,\]

where $\mathrm{R}$ are historical returns, $b$ is the benchmark, and $N$ is the number of historical entries (not the number of assets).

Note

The value of benchmark should correspond to the frequency of historical returns, i.e. if using daily returns, benchmark should be a reasonable value for daily returns.

PortfolioOptimiser.transaction_cost — Function

transaction_cost(w, w_prev, k = 0.001)

Simple transaction cost model. Sums all the absolute changes in weight and multiplies by a constant k. Simulates a fixed percentage commision from a broker.

The transacion cost $C$, is defined as:

\[C = k \lvert \bm{w} - \bm{w}_{\mathrm{prev}} \rvert\]

Warning

As of JuMP 1.0 there is no support for norm in objective functions. A model wishing to add this to their objective function should instead add it as a Vector Cone constraint, eg:

n = length(tickers)
prev_weights = fill(1 / n, n)
k = 0.001

extra_vars = [:(z[1:($n)])]
extra_constraints =
    [:([$k * (model[:w] - $prev_weights); model[:z]] in MOI.NormOneCone($(n + n)))]

ef = EfficientFrontier(tickers, mu, S;
        extra_vars = extra_vars,
        extra_constraints = extra_constraints,
    )

Similarly L2-norms must be turned into constraints of type MOI.NormTwoCone. More information on how to do this can be found in JuMP Vector Cones.

PortfolioOptimiser.ex_ante_tracking_error — Function

ex_ante_tracking_error(w, cov_mtx, w_bmk)

Compute the square of the ex-ante tracking error from the weights w, covariance matrix cov_mtx, and benchamark weights w_bmk.

The ex-ante tracking error $\mathrm{Err}$, is defined as:

\[\mathrm{Err} = (\bm{w} - \bm{w}_b)^T \, \Sigma \, (\bm{w} - \bm{w}_b)\,,\]

where $\bm{w}$ are the weights, $\bm{w}_b$ the benchmark weights, and $\Sigma$ the covariance matrix.

PortfolioOptimiser.ex_post_tracking_error — Function

ex_post_tracking_error(w, returns, ret_bmk)

Compute the square of the ex-post tracking error from the weights w, historical returns returns, and benchmark returns ret_bmk.

The ex-post tracking error $\mathrm{Err}$, is defined as:

\[\begin{aligned} \bm{x} &= \mathrm{R} \bm{w} - \bm{R}_b \\ \mu &= \dfrac{1}{N} \sum\limits_{i=1}^N x_i \\ \mathrm{Err} &= (\bm{x} - \mu) \cdot (\bm{x} - \mu) = Var(r - r_b)\,, \end{aligned}\]

where $w$ are the weights, $\mathrm{R}$ historical returns, $\bm{R}_b$ benchmark historical returns. In other words the ex-post tracking error is the variance of the historical returns $r$, minus the benchmark returns $r_b$, of the portfolio.

PortfolioOptimiser.logarithmic_barrier — Function

logarithmic_barrier(w, cov_mtx, k = 0.1)

Compute the logarithmic barrier adjusted by k, from the weights w, and covariance matrix cov_mtx.

The logarithmic barrier $L$, is defined as:

\[L = \bm{w}^T\, \Sigma\, \bm{w} - k \sum\limits_i \ln( w_i )\,,\]

where $\sigma^2 = \bm{w}^T\, \Sigma\, \bm{w}$ (see port_variance), $k$ the adjustment constant, and $w_i$ the weight of the $i$'th asset.

PortfolioOptimiser.kelly_objective — Function

kelly_objective(w, mean_ret, cov_mtx, k = 3)

Compute a Kelly objective from adjusted by k, from the weights w, mean returns mean_ret, and covariance matrix cov_mtx.

The Kelly objective $K$, is defined as:

\[K = \bm{w}^T\, \Sigma\, \bm{w} - \dfrac{1}{2} k \bm{w} \cdot \bm{E} \,,\]

where $\sigma^2 = \bm{w}^T\, \Sigma\, \bm{w}$ (see port_variance), $k$ is the adjustment constant, $\bm{w}$ the weights, and $\bm{E}$ the mean returns.

Expected Returns

PortfolioOptimiser.AbstractReturnModel — Type

abstract type AbstractReturnModel end
struct MRet <: AbstractReturnModel end
struct EMRet <: AbstractReturnModel end
struct CAPMRet <: AbstractReturnModel end
struct ECAPMRet <: AbstractReturnModel end

Return types, used to dispatch on ret_model.

MRet: Mean returns.

EMRet: Exponentially weighted mean returns.

CAPMRet: Capital Asset Pricing Model (CAPM) returns.

ECAPMRet: Exponentially weighted Capital Asset Pricing Model (ECAPM) returns.

PortfolioOptimiser.ret_model — Function

Compute the mean returns given a concrete type of AbstractReturnModel, and returns array where each column is an asset and each row is an entry.

For all methods:

compound: if true the compute compound mean returns, if false compute uncompounded mean returns.

freq: frequency at which the returns are logged, defaults to the average number of days in a trading year.

Mean

Arithmetic mean

ret_model(::MRet, returns; compound = true, freq = 252)

Compute the mean returns.

Exponentially weighted arithmetic mean

ret_model(::EMRet, returns; compound = true, freq = 252, span = Int(ceil(freq / 1.4)))

Compute the exponentially weighted areithmetic mean returns. More recent returns are weighted more heavily.

span: defines the weighing parameter, α = 2 / (span + 1).

Capital Asset Price Model (CAPM) returns

Compute the Capital Asset Price Model (CAPM) returns.

If the market returns are not provided, it estimates them by averaging across all assets and uses those as the market returns. These methods require computing the covariance between asset and market returns. These methods give a fairer value to a stock returns according to how they relate to the market.

cov_type: a concrete type of AbstractRiskModel for computing the covariance of the assets to the market.

CAPM with arithmetic mean

ret_model(
    ::CAPMRet,
    returns,
    market_returns = nothing;
    rf = 0.02,
    compound = true,
    freq = 252,
    cov_type::AbstractRiskModel = Cov(),
    target = 1.02^(1 / 252) - 1,
    fix_method::Union{SFix, DFix} = SFix(),
    span = Int(ceil(freq / 1.4)),
    )

Capital Asset Pricing Model (CAPM) returns. CAPMRet() uses the normal mean to estimate the mean market returns.

target, fix_method, and span correspond to the same keyword arguments of risk_matrix for computing the covariance specified by cov_type.

CAPM with exponentially weighted arithmetic mean

ret_model(
    ::ECAPMRet,
    returns,
    market_returns = nothing;
    rf = 0.02,
    compound = true,
    freq = 252,
    rspan = Int(ceil(freq / 1.4)),
    cov_type::AbstractRiskModel = Cov(),
    target = 1.02^(1 / 252) - 1,
    fix_method::Union{SFix, DFix} = SFix(),
    cspan = Int(ceil(freq / 1.4)),
)

Exponentially weighted Capital Asset Pricing Model (ECAPM) returns. ECAPMRet() uses the exponentially weighted mean to estimate the mean market returns.

rspan: span for the exponentially weighted mean returns, same as span for EMRet() above.

cspan: span for the exponentially weighted covariance, same as span for risk_matrix.

PortfolioOptimiser.returns_from_prices — Function

returns_from_prices(prices, log_ret = false)

Compute the returns from prices.

prices: array of prices, each column is an asset, each row an entry.

log_ret: if false compute the normal returns, if true compute the logarithmic returns.

PortfolioOptimiser.prices_from_returns — Function

prices_from_returns(returns, log_ret = false)

Compute the prices from returns.

returns: array of returns, each column is an asset, each row an entry.

log_ret: if false asume normal returns, if true asume logarithmic returns.

Risk Models

PortfolioOptimiser.AbstractFixPosDef — Type

abstract type AbstractFixPosDef end
struct SFix <: AbstractFixPosDef end
struct DFix <: AbstractFixPosDef end

Types for fixing non-positive definite matrices in risk_matrix.

SFix: fix via the spectral method (eigenvalue decomposition).

DFix: fix by adding a multiple of the identity matrix (damping factor).

PortfolioOptimiser.AbstractRiskModel — Type

abstract type AbstractRiskModel end
struct Cov <: AbstractRiskModel end
struct SCov <: AbstractRiskModel end
struct ECov <: AbstractRiskModel end
struct ESCov <: AbstractRiskModel end

Risk models for dispatch on risk_matrix.

Cov: Mean returns.

SCov: Exponentially weighted mean returns.

ECov: Capital Asset Pricing Model (CAPM) returns.

ESCov: Exponentially weighted Capital Asset Pricing Model (ECAPM) returns.

PortfolioOptimiser.risk_matrix — Function

Calculate covariance matrices.

Wrapper function

risk_matrix(
    type::AbstractRiskModel,
    returns;
    target = 1.02^(1 / 252) - 1,
    fix_method::Union{SFix, DFix} = SFix(),
    freq = 252,
    span = Int(ceil(freq / 1.4)),
)

Dispatches specific covariance matrix calculation methods according to type.

returns: array of asset returns, each column is an asset, each row an entry.

target: target for distinguishing "upside" and "downside" returns. Defaults to the daily risk-free rate.

fix_method: the fixing method used in case the matrix is non-positive definite.

freq: frequency at which returns are recorded, defaults to the average number of trading days in a year.

span: span parmeter for calculating the exponential weights, α = 2 / (span + 1).

Normal covariance

risk_matrix(::Cov, returns; fix_method::Union{SFix, DFix} = SFix(), freq = 252)

Semi-covariance

The semi-covariance sets a target for distinguishing "upside" and "downside" risk. It only considers fluctuations below target which heavily biases losses. Lets users minimise negative risk.

risk_matrix(
    ::SCov,
    returns;
    target = 1.02^(1 / 252) - 1,
    fix_method::Union{SFix, DFix} = SFix(),
    freq = 252,
)

Exponentially weighted covariance

The exponentially weighted covariance uses exponentially weighted expected returns rather than normally weighted expected returns. More recent returns are weighted more heavily.

risk_matrix(
    ::ECov,
    returns;
    fix_method::Union{SFix, DFix} = SFix(),
    freq = 252,
    span = Int(ceil(freq / 1.4)),
)

Exponentially weighted semi-covariance

Combines the semi-covariance and exponentially weighted covariance.

risk_matrix(
    ::ESCov,
    returns;
    target = 1.02^(1 / 252) - 1, # Daily risk free rate.
    fix_method::Union{SFix, DFix} = SFix(),
    freq = 252,
    span = Int(ceil(freq / 1.4)),
)

PortfolioOptimiser.make_pos_def — Function

Utility function for fixing non-positive definite matrices.

Spectral method

make_pos_def(::SFix, matrix)

Uses eigenvalue decomposition to make the negative eigenvalues equal to zero and reconstruct the matrix with new eigenvalues.

Diagonal method

make_pos_def(::DFix, matrix, scale = 1.1)

Uses a Levenberg-Marquardt factor, c = scale * min(eigenvalues), which is used to subtract a scaled identity matrix from matrix.

StatsBase.cov2cor — Function

cov2or(cov_mtx)

Wraps StatsBase.cov2cor to provide the standard deviations vector from the diagonal of cov_mtx.

Portfolio Optimisation

PortfolioOptimiser.AbstractPortfolioOptimiser — Type

Abstract type for subtyping concrete portfolio optimiser structures.

PortfolioOptimiser.custom_optimiser! — Function

custom_optimiser!(
    portfolio::AbstractPortfolioOptimiser,
    obj,
    obj_params = (),
    initial_guess = nothing,
    optimiser = Ipopt.Optimizer,
    silent = true,
)

Users provide their own objectives that are supported by JuMP.@objective.

portfolio: any concrete subtype of AbstractPortfolioOptimiser.

obj: objective function supported by JuMP.@objective.

obj_params: vector or tuple of arguments of obj.

initial_guess: initial guess for optimiser, if nothing lets the optimiser decide.

optimiser: optimiser for solving optimisation problem.

silent: if false, the optimiser prints to console.

Example

This example defines a new function, identical to kelly_objective, but you can use the predefined functions in PortfolioOptimiser as long as they are supported by JuMP.@objective.

function kelly_objective2(w, mean_ret, cov_mtx, k = 3)
    variance = dot(w, cov_mtx, w)
    objective = variance * 0.5 * k - dot(w, mean_ret)
    return objective
end

ef = EfficientFrontier(tickers, expected_returns, cov_mtx)
obj_params = (ef.mean_ret, ef.cov_mtx, 1000)
custom_optimiser!(ef, kelly_objective2, obj_params)

Unlike, custom_nloptimiser!, functions defined by PortfolioOptimiser don't have to be prepended with PortfolioOptimiser, they can be used as normal. Here we use kelly_objective, which is provided by the package and as previously mentioned is identical to kelly_objective2 of the previous example.

ef = EfficientFrontier(tickers, expected_returns, cov_mtx)
obj_params = (ef.mean_ret, ef.cov_mtx, 1000)
custom_optimiser!(ef, kelly_objective, obj_params)

Note

obj_params can be any variable that can be spatted. It is also optional, so objectives with no parameters are valid too.

PortfolioOptimiser.custom_nloptimiser! — Function

custom_nloptimiser!(
    portfolio::AbstractPortfolioOptimiser,
    obj,
    obj_params = (),
    initial_guess = nothing,
    optimiser = Ipopt.Optimizer,
    silent = true,
)

Users provide their own objectives that are supported by JuMP.@NLobjective.

portfolio: any concrete subtype of AbstractPortfolioOptimiser.

obj: objective function supported by JuMP.@NLobjective.

obj_params: vector or tuple of arguments of obj.

initial_guess: initial guess for optimiser, if nothing lets the optimiser decide.

optimiser: optimiser for solving optimisation problem.

silent: if false, the optimiser prints to console.

Warning

As of JuMP 1.0, JuMP.@NLobjective only supports scalar arguments, as such, it's important to define the nonlinear objective to take in scalar arguments only. The example shows how this can be done.

Example

In this first example we define a logarithmic barrier function identical to logarithmic_barrier, then we have to define the scalar-only function that JuMP.@NLobjective can use.

function logarithmic_barrier2(w, cov_mtx, k = 0.1)
    # Add eps() to avoid log(0) divergence.
    log_sum = sum(log.(w .+ eps()))
    var = dot(w, cov_mtx, w)
    return var - k * log_sum
end

function logarithmic_barrier(w::T...) where {T}
    cov_mtx = obj_params[1]
    k = obj_params[2]
    w = [i for i in w]
    logarithmic_barrier2(w, cov_mtx, k)
end

ef = EfficientFrontier(names(df)[2:end], mu, S)
obj_params = (ef.cov_mtx, 0.001)
custom_nloptimiser!(ef, logarithmic_barrier, obj_params)

We can also use objective functions that are already defined by PortfolioOptimiser. However, given how JuMP.@NLobjective requires user-defined functions to be registered, we need to prepend them with PortfolioOptimiser so that it can be recognised and registered by JuMP into the model.

function logarithmic_barrier(w::T...) where {T}
    cov_mtx = obj_params[1]
    k = obj_params[2]
    w = [i for i in w]
    PortfolioOptimiser.logarithmic_barrier(w, cov_mtx, k)
end

ef = EfficientFrontier(names(df)[2:end], mu, S)
obj_params = [ef.cov_mtx, 0.001]
custom_nloptimiser!(ef, logarithmic_barrier, obj_params)

# This definition will error in custom_nloptimiser!
function logarithmic_barrier(w::T...) where {T}
    cov_mtx = obj_params[1]
    k = obj_params[2]
    w = [i for i in w]
    logarithmic_barrier(w, cov_mtx, k)
end

ef = EfficientFrontier(names(df)[2:end], mu, S)
obj_params = [ef.cov_mtx, 0.001]
custom_nloptimiser!(ef, logarithmic_barrier, obj_params)

Warning

Note how in both cases, the parameters in the scalar-only function are not declared in the function signature. As such, they must be defined in terms of obj_params, or as literals, as the variable must be known to custom_nloptimiser!. If a variable name other than obj_params is used, the model will error becuase the variable will be unkown to the custom_nloptimiser!. For example, the following custom nonlinear objective function will not work because obj_param and obj_parameter do not exist inside custom_nloptimiser!.

function logarithmic_barrier(w::T...) where {T}
    cov_mtx = obj_param[1]  # Unkown variable, change to obj_params[1]
    k = obj_parameter[2]    # Unkown variable, change to obj_params[2]
    w = [i for i in w]
    PortfolioOptimiser.logarithmic_barrier(w, cov_mtx, k)
end

ef = EfficientFrontier(names(df)[2:end], mu, S)
obj_params = [ef.cov_mtx, 0.001]
custom_nloptimiser!(ef, logarithmic_barrier, obj_params)

Note

obj_params can be any variable that can be spatted. It is also optional, so nonlinear objectives with no parameters are valid too.

PortfolioOptimiser.add_sector_constraint! — Function

add_sector_constraint!(
    portfolio::AbstractPortfolioOptimiser,
    sector_map,
    sector_lower,
    sector_upper,
)

Constrains the sum of the weights of the tickers belonging to a sector to be between that sector's lower and upper bounds. For example, if we have two technology tickers, and we define the lower and upper bounds of the technology sector to be (0.1, 0.3), the sum of the weights of those tickers will be between (0.1, 0.3).

sector_map: a dictionary mapping tickers to sectors where they key is the ticker and the value the sector, Dict(<ticker> => <sector>).

sector_lower: a dictionary mapping a sector to its lower bound, Dict(<sector> => sector_lower_bound).

sector_upper: a dictionary mapping a sector to its upper bound, Dict(<sector> => sector_upper_bound).

PortfolioOptimiser._refresh_model! — Function

_refresh_model!(default_keys, model)

Helper function for cleaning up JuMP.Model() objects.

default_keys: collection of keys that are to be kept in the model.

model: JuMP.Model() object to be cleaned up.

PortfolioOptimiser._create_weight_bounds — Function

_create_weight_bounds(num_tickers, bounds)

Create the upper and lower bounds for asset weights.

num_tickers: The number of tickers.

bounds: weight bounds.

If bounds is a collection of length two, it assumes the first entry bounds[1] are the lower bounds for all weights, and bounds[2] is the upper bounds for all weights.

If bounds is an abstract vector of length equal to num_tickers, and each element of the vector is a tuple of length 2. Then each element corresponds to an asset, the first element of the tuple is the lower bound, the second the upper bound for said asset.

Else throw an error.

It returns a tuple of:

(lower_bounds, upper_bounds): where the length of lower_bounds and upper_bounds is equal to num_tickers. These are then used by AbstractEfficient portfolios to set the bounds constraints for their JuMP.Model().

PortfolioOptimiser._make_weight_sum_constraint! — Function

_make_weight_sum_constraint!(model, market_neutral)

Create the weight sum constraints for a JuMP.Model().

model: a JuMP.Model() of a portfolio optimiser.

market_neutral: if true the portfolio is market neutral and the weights will be constrained to sum to 0. If false, the portfolio is not market neutral and the weights will be constrained to sum to 1.

If market_neutral is true and all lower_bounds constraints are greater than or equal to zero, then a portfolio cannot be market neutral. If this is the case, the lower_bounds constraint for all assets will be changed to be greater than or equal to -1.

PortfolioOptimiser._add_var_to_model! — Function

_add_var_to_model!(model, var, args...)

Function to add a JuMP.@variable to a JuMP.Model().

model: a JuMP.Model().

var: symbol or expression representing a JuMP.@variable. Extra array variables have to be expressions. Extra scalar variables can be symbols or expressions.

args...: expression representing the args... to a JuMP.@variable.

Variables can also be directly added to a model using JuMP.@variable. This is used internally by AbstractEfficient portfolios, so that when a portfolio optimiser is refreshed, the extra variables can be added back automatically.

Example

We can use this to add a vector variable c of length 3, a scalar variable k, and a 3 x 5 x 4 multidimensional array variable m.

extra_vars = [ :( c[1:3] ), :k, :( m[1:3, 1:5, 1:4] ) ]
ef = EfficientFrontier(tickers, mean_ret, cov_mtx)
_add_var_to_model!.(ef.model, extra_vars)

Alternatively, we can make sure they get registered in the optimiser. If the portfolio optimiser is refreshed, the extra variables are added back to the model automatically. They'd have to be added back manually otherwise.

extra_vars = [ :( c[1:3] ), :k, :( m[1:3, 1:5, 1:4] ) ]
ef = EfficientFrontier(tickers, mean_ret, cov_mtx; extra_vars = extra_vars)

Note

When adding variables to the constructor, they must be a collection (Tuple or vector), even if it's of length 1.

PortfolioOptimiser._add_constraint_to_model! — Function

_add_constraint_to_model!(model, key, constraint)

Function to add a JuMP.@constraint to a JuMP.Model().

model: a JuMP.Model().

key: the key under which the constraint will be registered in the JuMP.Model().

constraint: expression representing the constraint.

Constraints can also be directly added to a model using JuMP.@constraint. This is used internally by AbstractEfficient portfolios, so that when a portfolio optimiser is refreshed, the extra constraints can be added back automatically.

Warning

Variables used within the constraint must be registered in the JuMP.Model() object (hence _add_var_to_model!), and referred to as model[<key>], where <key> is the key of the variable. For example, when adding a constraint such that the first weight is less than or equal to 0.2, :(model[:w][1] <= 0.2).

Warning

Constraints must be expressions :(), not quotes quote ... end, because quotes can contain be multiple expressions.

Note

This uses JuMP.@constraint internally, so any constraint given must be compatible with JuMP.@constraint.

Example

We can use this to add a constraint, :const1 such that the weight of the second asset is greater than or equal to 2, and :const2 such that the sum of all weights is less than or equal to 5. These are nonsensical constraints, but they illustrate the point. A constraint will work as long as it is a valid JuMP.@constraint. For a more complex example of a constraint see transaction_cost.

extra_constraints = [:(model[:w][2] >= 2), :(sum(model[:w]) <= 5)]
ef = EfficientFrontier(tickers, mean_ret, cov_mtx)
_add_constraint_to_model!.(ef.model, [:const1, :const2], extra_constraints)

We can also use variable interpolation into the expressions.

ef = EfficientFrontier(tickers, mean_ret, cov_mtx)
w = ef.model[:w]
extra_constraints = [:($(w[2]) >= 2), :(sum($w) <= 5)]
_add_constraint_to_model!.(ef.model, [:const1, :const2], extra_constraints)

If we want the extra constraints to be added back automatically when refreshing a portfolio optimiser, we can add them to the constructor.

extra_constraints = [:(model[:w][2] >= 2), :(sum(model[:w]) <= 5)]
ef = EfficientFrontier(tickers, mean_ret, cov_mtx; extra_constraints = extra_constraints)

Note

When adding constraints to the constructor, they must be a collection (Tuple or vector), even if it's of length 1.

PortfolioOptimiser._add_to_objective! — Function

_add_to_objective!(model, expr)

Function to add a term to a JuMP.@objective to a JuMP.Model().

model: a JuMP.Model().

expr: the expression to add to the objective function.

Warning

Variables used within the extra objective must be registered in the JuMP.Model() object (hence _add_var_to_model!), and referred to as model[<key>], where <key> is the key of the variable. For example, quote model[:w][1] * 5 - sum(model[:w]) end or :( model[:w][1] * 5 - sum(model[:w]) ). Extra objective terms can be quotes or expressions. Using quotes is recommended because extra objective terms can be functions, as shown in the example.

Note

Can only be used for JuMP.@objective. For problems which require a non-linear objective, use custom_nloptimiser! and add the extra terms to the custom function.

Example

We can add extra objective terms to be minimised in the following way. We can use functions defined by PortfolioOptimiser, user defined functions, or anything else that is a valid JuMP.@objective. You can also use variable interpolation, but we don't use it in this example.

extra_obj_terms = [
    quote
        model[:w][1] * 5 - sum(model[:w])
    end,
    quote
       L2_reg(model[:w], 0.05)
    end
]
ef = EfficientFrontier(tickers, mean_ret, cov_mtx)
_add_to_objective!.(ef.model, extra_obj_terms)

They can also be added to the constructor so they are added back automatically when the portfolio optimiser is refreshed.

extra_obj_terms = [
    quote
        model[:w][1] * 5 - sum(model[:w])
    end,
    quote
       L2_reg(model[:w], 0.05)
    end
]
ef = EfficientFrontier(tickers, mean_ret, cov_mtx; extra_obj_terms = extra_obj_terms)

Note

When adding elements to an objective in the constructor, they must be a collection (Tuple or vector), even if it's of length 1.

Efficient Frontier

PortfolioOptimiser.AbstractEfficient — Type

abstract type AbstractEfficient <: AbstractPortfolioOptimiser end

Abstract type for subtyping efficient frontier optimisers.

Mean-variance

Mean-semivariance

PortfolioOptimiser.EfficientSemiVar — Type

struct EfficientSemiVar{
    T1,
    T2,
    T3,
    T4,
    T5,
    T6,
    T7,
    T8,
    T9,
    T10,
    T11,
    T12,
    T13,
    T14,
    T15,
    T16,
} <: AbstractEfficientSemiVar
    tickers::T1
    mean_ret::T2
    weights::T3
    returns::T4
    freq::T5
    benchmark::T6
    rf::T7
    market_neutral::T8
    risk_aversion::T9
    target_semidev::T10
    target_ret::T11
    extra_vars::T12
    extra_constraints::T13
    extra_obj_terms::T14
    extra_obj_func::T15
    model::T16
end

Critical Value at Risk (CVaR)

Critical Drawdown at Risk (CDaR)

Hierarchical Risk Parity

Black-Litterman

Critical line algorithm

Asset allocation

PortfolioOptimiser.AbstractAllocation — Type

abstract type AbstractAllocation end
struct Lazy <: AbstractAllocation end
struct Greedy <: AbstractAllocation end
struct LP <: AbstractAllocation end

Types for allocation dispatch.

Lazy: Lazy allocation algorithm. Can do fractional shares.

Greedy: Greedy allocation algorithm. Can do fractional shares.

LP: Mixed linear integer programming algorithm. Can only handle full shares.

PortfolioOptimiser.Allocation — Type

struct Allocation{T1,T2,T3}
    tickers::T1
    weights::T2
    shares::T3
end

Concrete allocation created by Allocation.

tickers: list of tickers.

weights: list of corresponding weights according to share allocation and latest prices.

shares: list of corresponding shares according to their optimal allocation.

PortfolioOptimiser.Allocation — Method

Allocation(
    type::AbstractAllocation,
    portfolio::AbstractPortfolioOptimiser,
    latest_prices::AbstractVector;
    investment = 1e4,
    rounding = 1,
    reinvest = false,
    short_ratio = nothing,
    optimiser = HiGHS.Optimizer,
    silent = true,
)

Returns a tuple of an Allocation structure and remaining money after allocating assets.

Note

Short positions have negative weights and shares.

type: Allocation algorithm to use, can be Lazy(), Greedy() or LP() (see AbstractAllocation).

portfolio: Portfolio AbstractPortfolioOptimiser.

latest_prices: Vector of latest prices, entries should be in the same order as portfolio.tickers.

investment: value to be invested in the portfolio.

rounding: Greedy() and Lazy() support fractional shares, rounds shares down to the nearest multiple of rounding. This has no effect on an LP() allocation.

reinvest: if true, reinvests the money earned from short positions of the portfolio into buying more long positions.

short_ratio: long to short portfolio ratio, 0.2 corresponds to 120/20 long to short shares. If nothing defaults to portfolio weights.

optimiser: only used for the LP() allocation. Needs to support mixed-interger linear programming.

silent: if true the optimiser is silent.

Example

ef = EfficientFrontier(
    tickers,
    mean_ret,
    cov_mtx;
    market_neutral = true,
    weight_bounds = (-1, 1),
)
max_quadratic_utility!(ef)

# Mixed integer, programming, no rounding available.
allocLP, leftoverLP = Allocation(LP(), ef, latest_prices; investment = 3590)

# Greedy allocation up to quarter shares.
allocGreedy, leftoverGreedy = Allocation(Greedy(), ef, latest_prices; rounding = 0.25)

# Lazy allocation with a 110/10 long-short ratio and reinvesting 
# earnings from selling the short positions into buying long positions.
allocLazy, leftoverLazy = Allocation(
                            Lazy(), ef, latest_prices; 
                            short_ratio = 0.1, reinvest = true
                        )

PortfolioOptimiser.roundmult — Function

roundmult(val, prec [, args...] [; kwargs...])

Round a number to a multiple of prec. Uses the same defaults and has the same args and kwargs of the built-in Base.round.

Equivalent to:

round(div(val, prec) * prec, args...; kwargs...)