Efficient Frontier

PortfolioOptimiser.AbstractEfficient — Type

abstract type AbstractEfficient <: AbstractPortfolioOptimiser end

Abstract type for subtyping efficient frontier optimisers.

PortfolioOptimiser._refresh_add_var_and_constraints — Function

_refresh_add_var_and_constraints(default_keys, portfolio)

Helper function for refreshing portfolios and re-adding extra variables and constraints used to build the model via the constructor.

PortfolioOptimiser._function_vs_portfolio_val_warn — Function

_function_vs_portfolio_val_warn(fval, pval, name)

Helper function for throwing generic warnings about inconsistent values between calls to optimiser functions and the values registered in the portfolio.

PortfolioOptimiser._val_compare_benchmark — Function

_val_compare_benchmark(val, op, benchmark, correction, name)

Helper function for throwing generic warnings about values out of the domain and ammending those values.

Mean-variance

PortfolioOptimiser.AbstractMeanVar — Type

abstract type AbstractMeanVar <: AbstractEfficient end

Abstract type for subtyping efficient mean variance optimisers.

PortfolioOptimiser.MeanVar — Type

struct MeanVar{T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13} <:
       AbstractMeanVar
    tickers::T1
    mean_ret::T2
    weights::T3
    cov_mtx::T4
    rf::T5
    market_neutral::T6
    risk_aversion::T7
    target_volatility::T8
    target_ret::T9
    extra_vars::T10
    extra_constraints::T11
    extra_obj_terms::T12
    model::T13
end

Structure for a mean-variance portfolio.

tickers: list of tickers.

mean_ret: mean returns, don't need it to optimise for minimum variance.

weights: weight of each ticker in the portfolio.

cov_mtx: covariance matrix.

rf: risk free rate.

market_neutral: whether a portfolio is market neutral or not. Used in max_quadratic_utility!, efficient_risk!, efficient_return!.

risk_aversion: risk aversion parameter. Used in max_quadratic_utility!.

target_volatility: target volatility parameter. Used in efficient_risk!.

target_ret: target return parameter. Used in efficient_return!.

extra_vars: extra variables for the model.

extra_constraints: extra constraints for the model.

extra_obj_terms: extra objective terms for the model.

model: model for optimising portfolio.

PortfolioOptimiser.MeanVar — Method

MeanVar(
    tickers,
    mean_ret,
    cov_mtx;
    weight_bounds = (0, 1),
    rf = 0.02,
    market_neutral = false,
    risk_aversion = 1.0,
    target_volatility = rank(cov_mtx) < size(cov_mtx, 1) ? 1 / sum(diag(cov_mtx)) :
                        sqrt(1 / sum(inv(cov_mtx))),
    target_ret = mean(mean_ret),
    extra_vars = [],
    extra_constraints = [],
    extra_obj_terms = [],
)

Create an MeanVar structure to be optimised via JuMP.

tickers: list of tickers.

mean_ret: mean returns, don't need it to optimise for minimum variance.

cov_mtx: covariance matrix.

weight_bounds: weight bounds for tickers. If it's a Tuple of length 2, the first entry will be the lower bound for all weights, the second entry will be the upper bound for all weights. If it's a vector, its length must be equal to that of tickers, each element must be a tuple of length 2. In that case, each tuple corresponds to the lower and upper bounds for the corresponding ticker. See _create_weight_bounds for further details.

rf: risk free rate. Must be consistent with the frequency at which mean_ret and cov_mtx were calculated. The default value assumes daily returns.

market_neutral: whether a portfolio is market neutral or not. If it is market neutral, the sum of the weights will be equal to 0, else the sum will be equal to 1. Used in max_quadratic_utility!, efficient_risk!, efficient_return!.

risk_aversion: risk aversion parameter, the larger it is, the lower the risk. Used in max_quadratic_utility!.

target_volatility: target volatility parameter. Used in efficient_risk!.

target_ret: target return parameter. Used in efficient_return!.

extra_vars: extra variables for the model. See _add_var_to_model! for details on how to use this.

extra_constraints: extra constraints for the model. See _add_constraint_to_model! for details on how to use this.

extra_obj_terms: extra objective terms for the model. See _add_to_objective! for details on how to use this.

PortfolioOptimiser.min_volatility! — Method

min_volatility!(
    portfolio::MeanVar;
    optimiser = Ipopt.Optimizer,
    silent = true,
    optimiser_attributes = (),
)

Minimise the volatility (port_variance) of a MeanVar portfolio.

portfolio: MeanVar structure.

optimiser: JuMP-supported optimiser, must support quadratic objectives.

silent: if true the optimiser will not print to console, if false the optimiser will print to console.

PortfolioOptimiser.max_return — Method

max_return(portfolio::MeanVar; optimiser = Ipopt.Optimizer, silent = true, optimiser_attributes = ())

Maximise the return (port_return) of a MeanVar portfolio. Internally minimises the negative of the portfolio return.

portfolio: MeanVar structure.

optimiser: JuMP-supported optimiser, must support quadratic objectives.

silent: if true the optimiser will not print to console, if false the optimiser will print to console.

Warning

This should not be used for optimising portfolios. It's used by efficient_return! to validate the target return. This yields portfolios with large volatilities.

PortfolioOptimiser.max_sharpe! — Function

max_sharpe!(
    portfolio::MeanVar,
    rf = portfolio.rf;
    optimiser = Ipopt.Optimizer,
    silent = true,
    optimiser_attributes = (),
)

Maximise the sharpe ratio (sharpe_ratio) of a MeanVar portfolio.

Uses a variable transformation to turn the nonlinear objective that is the sharpe ratio, into a convex quadratic minimisation problem. See Cornuejols and Tutuncu (2006) page 158 for more details.

portfolio: MeanVar structure.

rf: risk free rate.

optimiser: JuMP-supported optimiser, must support quadratic objectives.

silent: if true the optimiser will not print to console, if false the optimiser will print to console.

Warning

The variable transformation also modifies every constraint in portfolio, including all extra constraints. Therefore, one should not call other optimsiations on the same instance after optimising the sharpe ratio, create a fresh intance instead.

Warning

The variable transformation means any extra terms in the objective function may not work as intended. If you need to add extra objective terms, use custom_nloptimiser! (see the example) and add the extra objective terms in the objective function.

PortfolioOptimiser.max_quadratic_utility! — Function

max_quadratic_utility!(
    portfolio::MeanVar,
    risk_aversion = portfolio.risk_aversion;
    optimiser = Ipopt.Optimizer,
    silent = true,
    optimiser_attributes = (),
)

Maximise the quadratic_utility of a MeanVar portfolio. Internally minimises the negative of the quadratic utility.

portfolio: MeanVar structure.

risk_aversion: the risk aversion parameter, the larger it is the lower the risk.

optimiser: JuMP-supported optimiser, must support quadratic objectives.

silent: if true the optimiser will not print to console, if false the optimiser will print to console.

PortfolioOptimiser.efficient_return! — Function

efficient_return!(
    portfolio::MeanVar,
    target_ret = portfolio.target_ret;
    optimiser = Ipopt.Optimizer,
    silent = true,
    optimiser_attributes = (),
)

Minimise the port_variance of a MeanVar portfolio subject to the constraint for the return to be greater than or equal to target_ret. The portfolio is guaranteed to have a return at least equal to target_ret.

portfolio: MeanVar structure.

target_ret: the target return of the portfolio.

optimiser: JuMP-supported optimiser, must support quadratic objectives.

silent: if true the optimiser will not print to console, if false the optimiser will print to console.

PortfolioOptimiser.efficient_risk! — Function

efficient_risk!(
    portfolio::MeanVar,
    target_volatility = portfolio.target_volatility;
    optimiser = Ipopt.Optimizer,
    silent = true,
    optimiser_attributes = (),
)

Maximise the port_return of a MeanVar portfolio subject to the constraint for the volatility to be less than or equal to target_volatility. The portfolio is guaranteed to have a volatility of at most target_volatility.

portfolio: MeanVar structure.

target_ret: the target return of the portfolio.

optimiser: JuMP-supported optimiser, must support quadratic objectives.

silent: if true the optimiser will not print to console, if false the optimiser will print to console.

PortfolioOptimiser.refresh_model! — Method

refresh_model!(portfolio::AbstractMeanVar)

Refreshes an AbstractMeanVar model.

PortfolioOptimiser.portfolio_performance — Method

portfolio_performance(portfolio::MeanVar; rf = portfolio.rf, verbose = false)

Computes the portfolio return (port_return), volatility (square root of port_variance), and sharpe ratio (sharpe_ratio) for a given risk free rate, rf.

Returns a tuple of:

(return, volatility, sharpe ratio)

If verbose == true, it prints out this information.

Mean-semivariance

PortfolioOptimiser.AbstractMeanSemivar — Type

abstract type AbstractMeanSemivar <: AbstractEfficient end

Abstract type for subtyping efficient mean semivariance optimisers.

PortfolioOptimiser.MeanSemivar — Type

struct MeanSemivar{T1, T2, T3, T4, T5, T6, T7, T8, T9, T10, T11, T12, T13, T14, T15} <:
       AbstractMeanSemivar
    tickers::T1
    mean_ret::T2
    weights::T3
    returns::T4
    freq::T5
    benchmark::T6
    rf::T7
    market_neutral::T8
    risk_aversion::T9
    target_semidev::T10
    target_ret::T11
    extra_vars::T12
    extra_constraints::T13
    extra_obj_terms::T14
    model::T15
end

Structure for a mean-semivariance portfolio.

tickers: list of tickers.

mean_ret: mean returns, don't need it to optimise for minimum variance.

weights: weight of each ticker in the portfolio.

returns: asset historical returns.

benchmark: returns benchmark, to differentiate between "downside" (less than benchmark) and "upside" (greater than benchmark) returns.

freq: frequency of returns.

rf: risk free rate.

market_neutral: whether a portfolio is market neutral or not. Used in max_quadratic_utility!, efficient_risk!, efficient_return!.

risk_aversion: risk aversion parameter. Used in max_quadratic_utility!.

target_volatility: target volatility parameter. Used in efficient_risk!.

target_ret: target return parameter. Used in efficient_return!.

extra_vars: extra variables for the model.

extra_constraints: extra constraints for the model.

extra_obj_terms: extra objective terms for the model.

model: model for optimising portfolio.

PortfolioOptimiser.MeanSemivar — Method

MeanSemivar(
    tickers,
    mean_ret,
    returns;
    weight_bounds = (0.0, 1.0),
    freq = 252,
    benchmark = 0,
    rf = 0.02,
    market_neutral = false,
    risk_aversion = 1.0,
    target_semidev = std(returns),
    target_ret = mean(mean_ret),
    extra_vars = [],
    extra_constraints = [],
    extra_obj_terms = [],
)

Create an MeanSemivar structure to be optimised via JuMP.

tickers: list of tickers.

mean_ret: mean returns, don't need it to optimise for minimum variance.

returns: asset historical returns.

weight_bounds: weight bounds for tickers. If it's a Tuple of length 2, the first entry will be the lower bound for all weights, the second entry will be the upper bound for all weights. If it's a vector, its length must be equal to that of tickers, each element must be a tuple of length 2. In that case, each tuple corresponds to the lower and upper bounds for the corresponding ticker. See _create_weight_bounds for further details.

benchmark: returns benchmark, to differentiate between "downside" (less than benchmark) and "upside" (greater than benchmark) returns.

freq: frequency of returns.

rf: risk free rate. Must be consistent with freq. The default value assumes daily returns.

market_neutral: whether a portfolio is market neutral or not. If it is market neutral, the sum of the weights will be equal to 0, else the sum will be equal to 1. Used in max_quadratic_utility!, efficient_risk!, efficient_return!.

risk_aversion: risk aversion parameter, the larger it is, the lower the risk. Used in max_quadratic_utility!.

target_volatility: target volatility parameter. Used in efficient_risk!.

target_ret: target return parameter. Used in efficient_return!.

extra_vars: extra variables for the model. See _add_var_to_model! for details on how to use this.

extra_constraints: extra constraints for the model. See _add_constraint_to_model! for details on how to use this.

extra_obj_terms: extra objective terms for the model. See _add_to_objective! for details on how to use this.

PortfolioOptimiser.refresh_model! — Method

refresh_model!(portfolio::AbstractMeanSemivar)

Refreshes an AbstractMeanSemivar model.

PortfolioOptimiser.portfolio_performance — Method

portfolio_performance(portfolio::MeanSemivar; rf = portfolio.rf, verbose = false)

Computes the portfolio return (port_return), semideviation (square root of port_semivar), and sortino ratio (sharpe_ratio adjusted to the semideviation) for a given risk free rate, rf.

Returns a tuple of:

(return, semideviation, sortino ratio)

If verbose == true, it prints out this information.

Critical Value at Risk (CVaR)

Critical Drawdown at Risk (CDaR)